Crate rusoto

Rusoto

Rusoto is an SDK for Amazon Web Services. It follows best practices for AWS Credentials.

The library is designed with use on AWS EC2 instances in mind but can be used as an AWS CLI replacement.

Credentials

Credentials are sourced from environment variables, AWS credentials profile and IAM instance profiles, in that order. IAM instance profile credentials are refreshed automatically as needed.

Supported services

SQS

S3 (partial implementation)

Requests and request signing

Rusoto uses AWS Signature 4 to sign requests.

Modules

credentials
AWS credentials

error
Errors during AWS communication or parsing

params
Parameters for talking to SQS

regions
AWS Regions and helper functions

request
A request to AWS, pre-signed

s3
S3 bindings for Rust

signature
Tools for signing a request to AWS

sqs
SQS bindings for Rust

xmlutil
Tools for handling XML from AWS with helper functions for testing