Module rusoto::acm

AWS Certificate Manager

Structs

AcmClient
A client for the ACM API.

AddTagsToCertificateRequest

CertificateDetail
This structure is returned in the response object of the DescribeCertificate action.

CertificateSummary
This structure is returned in the response object of ListCertificates action.

DeleteCertificateRequest

DescribeCertificateRequest

DescribeCertificateResponse

DomainValidation
Structure that contains the domain name, the base validation domain to which validation email is sent, and the email addresses used to validate the domain identity.

DomainValidationOption
This structure is used in the request object of the RequestCertificate action.

GetCertificateRequest

GetCertificateResponse

ListCertificatesRequest

ListCertificatesResponse

ListTagsForCertificateRequest

ListTagsForCertificateResponse

RemoveTagsFromCertificateRequest

RequestCertificateRequest

RequestCertificateResponse

ResendValidationEmailRequest

Tag
A key-value pair that identifies or specifies metadata about an ACM resource.

Enums

AddTagsToCertificateError

DeleteCertificateError

DescribeCertificateError

GetCertificateError

ListCertificatesError

ListTagsForCertificateError

RemoveTagsFromCertificateError

RequestCertificateError

ResendValidationEmailError

Type Definitions

Arn

CertificateBody

CertificateChain

CertificateStatus

CertificateStatuses

CertificateSummaryList

DomainList

DomainNameString

DomainValidationList

DomainValidationOptionList

IdempotencyToken

InUseList

KeyAlgorithm

MaxItems

NextToken

RevocationReason

TStamp

TagKey

TagList

TagValue

ValidationEmailList