Struct rusoto::acm::CertificateDetail

pub struct CertificateDetail {
    pub certificate_arn: Option<Arn>,
    pub created_at: Option<TStamp>,
    pub domain_name: Option<DomainNameString>,
    pub domain_validation_options: Option<DomainValidationList>,
    pub failure_reason: Option<FailureReason>,
    pub in_use_by: Option<InUseList>,
    pub issued_at: Option<TStamp>,
    pub issuer: Option<String>,
    pub key_algorithm: Option<KeyAlgorithm>,
    pub not_after: Option<TStamp>,
    pub not_before: Option<TStamp>,
    pub revocation_reason: Option<RevocationReason>,
    pub revoked_at: Option<TStamp>,
    pub serial: Option<String>,
    pub signature_algorithm: Option<String>,
    pub status: Option<CertificateStatus>,
    pub subject: Option<String>,
    pub subject_alternative_names: Option<DomainList>,
}

Contains detailed metadata about an ACM Certificate. This structure is returned in the response to a DescribeCertificate request.

Fields
certificate_arn: Option<Arn>

The Amazon Resource Name (ARN) of the certificate. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

created_at: Option<TStamp>

The time at which the certificate was requested.

domain_name: Option<DomainNameString>

The fully qualified domain name (FQDN) for the certificate, such as www.example.com or example.com.

domain_validation_options: Option<DomainValidationList>

Contains information about the email address or addresses used for domain validation.

failure_reason: Option<FailureReason>

The reason the certificate request failed. This value exists only when the structure's Status is FAILED. For more information, see Certificate Request Failed in the AWS Certificate Manager User Guide.

in_use_by: Option<InUseList>

A list of ARNs for the resources that are using the certificate. An ACM Certificate can be used by multiple AWS resources.

issued_at: Option<TStamp>

The time at which the certificate was issued.

issuer: Option<String>

The X.500 distinguished name of the CA that issued and signed the certificate.

key_algorithm: Option<KeyAlgorithm>

The algorithm used to generate the key pair (the public and private key). Currently the only supported value is RSA_2048.

not_after: Option<TStamp>

The time after which the certificate is not valid.

not_before: Option<TStamp>

The time before which the certificate is not valid.

revocation_reason: Option<RevocationReason>

The reason the certificate was revoked. This value exists only when the certificate status is REVOKED.

revoked_at: Option<TStamp>

The time at which the certificate was revoked. This value exists only when the certificate status is REVOKED.

serial: Option<String>

The serial number of the certificate.

signature_algorithm: Option<String>

The algorithm used to generate a signature. Currently the only supported value is SHA256WITHRSA.

status: Option<CertificateStatus>

The status of the certificate.

subject: Option<String>

The X.500 distinguished name of the entity associated with the public key contained in the certificate.

subject_alternative_names: Option<DomainList>

One or more domain names (subject alternative names) included in the certificate request. After the certificate is issued, this list includes the domain names bound to the public key contained in the certificate. The subject alternative names include the canonical domain name (CN) of the certificate and additional domain names that can be used to connect to the website.