Struct rusoto::acm::DomainValidationOption

pub struct DomainValidationOption {
    pub domain_name: DomainNameString,
    pub validation_domain: DomainNameString,
}

This structure is used in the request object of the RequestCertificate action.

Fields
domain_name: DomainNameString

Fully Qualified Domain Name (FQDN) of the certificate being requested.

validation_domain: DomainNameString

The domain to which validation email is sent. This is the base validation domain that will act as the suffix of the email addresses. This must be the same as the DomainName value or a superdomain of the DomainName value. For example, if you requested a certificate for site.subdomain.example.com and specify a ValidationDomain of subdomain.example.com, ACM sends email to the domain registrant, technical contact, and administrative contact in WHOIS for the base domain and the following five addresses:

admin@subdomain.example.com

administrator@subdomain.example.com

hostmaster@subdomain.example.com

postmaster@subdomain.example.com

webmaster@subdomain.example.com