Struct rusoto::acm::ListCertificatesRequest

pub struct ListCertificatesRequest {
    pub certificate_statuses: Option<CertificateStatuses>,
    pub max_items: Option<MaxItems>,
    pub next_token: Option<NextToken>,
}

Fields
certificate_statuses: Option<CertificateStatuses>

Identifies the statuses of the ACM Certificates for which you want to retrieve the ARNs. This can be one or more of the following values:

PENDING_VALIDATION

ISSUED

INACTIVE

EXPIRED

VALIDATION_TIMED_OUT

REVOKED

FAILED

max_items: Option<MaxItems>

Specify this parameter when paginating results to indicate the maximum number of ACM Certificates that you want to display for each response. If there are additional certificates beyond the maximum you specify, use the NextToken value in your next call to the ListCertificates action.

next_token: Option<NextToken>

String that contains an opaque marker of the next ACM Certificate ARN to be displayed. Use this parameter when paginating results, and only in a subsequent request after you've received a response where the results have been truncated. Set it to an empty string the first time you call this action, and set it to the value of the NextToken element you receive in the response object for subsequent calls.