Struct rusoto::autoscaling::AutoscalingClient

pub struct AutoscalingClient<P, D> where P: ProvideAwsCredentials,
        D: DispatchSignedRequest { /* fields omitted */ }

A client for the Auto Scaling API.

Methods

impl<P, D> AutoscalingClient<P, D> where P: ProvideAwsCredentials,
D: DispatchSignedRequest

fn new(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn attach_instances(&self,
input: &AttachInstancesQuery)
-> Result<(), AttachInstancesError>

Attaches one or more EC2 instances to the specified Auto Scaling group.

When you attach instances, Auto Scaling increases the desired capacity of the group by the number of instances being attached. If the number of instances being attached plus the desired capacity of the group exceeds the maximum size of the group, the operation fails.

If there is a Classic load balancer attached to your Auto Scaling group, the instances are also registered with the load balancer. If there are target groups attached to your Auto Scaling group, the instances are also registered with the target groups.

For more information, see Attach EC2 Instances to Your Auto Scaling Group in the Auto Scaling User Guide.

fn attach_load_balancer_target_groups(&self,
input: &AttachLoadBalancerTargetGroupsType)
-> Result<AttachLoadBalancerTargetGroupsResultType, AttachLoadBalancerTargetGroupsError>

Attaches one or more target groups to the specified Auto Scaling group.

To describe the target groups for an Auto Scaling group, use DescribeLoadBalancerTargetGroups. To detach the target group from the Auto Scaling group, use DetachLoadBalancerTargetGroups.

For more information, see Attach a Load Balancer to Your Auto Scaling Group in the Auto Scaling User Guide.

fn attach_load_balancers(&self,
input: &AttachLoadBalancersType)
-> Result<AttachLoadBalancersResultType, AttachLoadBalancersError>

Attaches one or more Classic load balancers to the specified Auto Scaling group.

To attach an Application load balancer instead, see AttachLoadBalancerTargetGroups.

To describe the load balancers for an Auto Scaling group, use DescribeLoadBalancers. To detach the load balancer from the Auto Scaling group, use DetachLoadBalancers.

For more information, see Attach a Load Balancer to Your Auto Scaling Group in the Auto Scaling User Guide.

fn complete_lifecycle_action(&self,
input: &CompleteLifecycleActionType)
-> Result<CompleteLifecycleActionAnswer, CompleteLifecycleActionError>

Completes the lifecycle action for the specified token or instance with the specified result.

This step is a part of the procedure for adding a lifecycle hook to an Auto Scaling group:

(Optional) Create a Lambda function and a rule that allows CloudWatch Events to invoke your Lambda function when Auto Scaling launches or terminates instances.

(Optional) Create a notification target and an IAM role. The target can be either an Amazon SQS queue or an Amazon SNS topic. The role allows Auto Scaling to publish lifecycle notifications to the target.

Create the lifecycle hook. Specify whether the hook is used when the instances launch or terminate.

If you need more time, record the lifecycle action heartbeat to keep the instance in a pending state.

If you finish before the timeout period ends, complete the lifecycle action.

For more information, see Auto Scaling Lifecycle in the Auto Scaling User Guide.

fn create_auto_scaling_group(&self,
input: &CreateAutoScalingGroupType)
-> Result<(), CreateAutoScalingGroupError>

Creates an Auto Scaling group with the specified name and attributes.

If you exceed your maximum limit of Auto Scaling groups, which by default is 20 per region, the call fails. For information about viewing and updating this limit, see DescribeAccountLimits.

For more information, see Auto Scaling Groups in the Auto Scaling User Guide.

fn create_launch_configuration(&self,
input: &CreateLaunchConfigurationType)
-> Result<(), CreateLaunchConfigurationError>

Creates a launch configuration.

If you exceed your maximum limit of launch configurations, which by default is 100 per region, the call fails. For information about viewing and updating this limit, see DescribeAccountLimits.

For more information, see Launch Configurations in the Auto Scaling User Guide.

fn create_or_update_tags(&self,
input: &CreateOrUpdateTagsType)
-> Result<(), CreateOrUpdateTagsError>

Creates or updates tags for the specified Auto Scaling group.

When you specify a tag with a key that already exists, the operation overwrites the previous tag definition, and you do not get an error message.

For more information, see Tagging Auto Scaling Groups and Instances in the Auto Scaling User Guide.

fn delete_auto_scaling_group(&self,
input: &DeleteAutoScalingGroupType)
-> Result<(), DeleteAutoScalingGroupError>

Deletes the specified Auto Scaling group.

If the group has instances or scaling activities in progress, you must specify the option to force the deletion in order for it to succeed.

If the group has policies, deleting the group deletes the policies, the underlying alarm actions, and any alarm that no longer has an associated action.

To remove instances from the Auto Scaling group before deleting it, call DetachInstances with the list of instances and the option to decrement the desired capacity so that Auto Scaling does not launch replacement instances.

To terminate all instances before deleting the Auto Scaling group, call UpdateAutoScalingGroup and set the minimum size and desired capacity of the Auto Scaling group to zero.

fn delete_launch_configuration(&self,
input: &LaunchConfigurationNameType)
-> Result<(), DeleteLaunchConfigurationError>

Deletes the specified launch configuration.

The launch configuration must not be attached to an Auto Scaling group. When this call completes, the launch configuration is no longer available for use.

fn delete_lifecycle_hook(&self,
input: &DeleteLifecycleHookType)
-> Result<DeleteLifecycleHookAnswer, DeleteLifecycleHookError>

Deletes the specified lifecycle hook.

If there are any outstanding lifecycle actions, they are completed first (ABANDON for launching instances, CONTINUE for terminating instances).

fn delete_notification_configuration(&self,
input: &DeleteNotificationConfigurationType)
-> Result<(), DeleteNotificationConfigurationError>

Deletes the specified notification.

fn delete_policy(&self,
input: &DeletePolicyType)
-> Result<(), DeletePolicyError>

Deletes the specified Auto Scaling policy.

Deleting a policy deletes the underlying alarm action, but does not delete the alarm, even if it no longer has an associated action.

fn delete_scheduled_action(&self,
input: &DeleteScheduledActionType)
-> Result<(), DeleteScheduledActionError>

Deletes the specified scheduled action.

fn delete_tags(&self, input: &DeleteTagsType) -> Result<(), DeleteTagsError>

Deletes the specified tags.

fn describe_account_limits(&self)
-> Result<DescribeAccountLimitsAnswer, DescribeAccountLimitsError>

Describes the current Auto Scaling resource limits for your AWS account.

For information about requesting an increase in these limits, see AWS Service Limits in the Amazon Web Services General Reference.

fn describe_adjustment_types(&self)
-> Result<DescribeAdjustmentTypesAnswer, DescribeAdjustmentTypesError>

Describes the policy adjustment types for use with PutScalingPolicy.

fn describe_auto_scaling_groups(&self,
input: &AutoScalingGroupNamesType)
-> Result<AutoScalingGroupsType, DescribeAutoScalingGroupsError>

Describes one or more Auto Scaling groups.

fn describe_auto_scaling_instances(&self,
input: &DescribeAutoScalingInstancesType)
-> Result<AutoScalingInstancesType, DescribeAutoScalingInstancesError>

Describes one or more Auto Scaling instances.

fn describe_auto_scaling_notification_types(&self)
-> Result<DescribeAutoScalingNotificationTypesAnswer, DescribeAutoScalingNotificationTypesError>

Describes the notification types that are supported by Auto Scaling.

fn describe_launch_configurations(&self,
input: &LaunchConfigurationNamesType)
-> Result<LaunchConfigurationsType, DescribeLaunchConfigurationsError>

Describes one or more launch configurations.

fn describe_lifecycle_hook_types(&self)
-> Result<DescribeLifecycleHookTypesAnswer, DescribeLifecycleHookTypesError>

Describes the available types of lifecycle hooks.

fn describe_lifecycle_hooks(&self,
input: &DescribeLifecycleHooksType)
-> Result<DescribeLifecycleHooksAnswer, DescribeLifecycleHooksError>

Describes the lifecycle hooks for the specified Auto Scaling group.

fn describe_load_balancer_target_groups(&self,
input: &DescribeLoadBalancerTargetGroupsRequest)
-> Result<DescribeLoadBalancerTargetGroupsResponse, DescribeLoadBalancerTargetGroupsError>

Describes the target groups for the specified Auto Scaling group.

fn describe_load_balancers(&self,
input: &DescribeLoadBalancersRequest)
-> Result<DescribeLoadBalancersResponse, DescribeLoadBalancersError>

Describes the load balancers for the specified Auto Scaling group.

Note that this operation describes only Classic load balancers. If you have Application load balancers, use DescribeLoadBalancerTargetGroups instead.

fn describe_metric_collection_types(&self)
-> Result<DescribeMetricCollectionTypesAnswer, DescribeMetricCollectionTypesError>

Describes the available CloudWatch metrics for Auto Scaling.

Note that the GroupStandbyInstances metric is not returned by default. You must explicitly request this metric when calling EnableMetricsCollection.

fn describe_notification_configurations(&self,
input: &DescribeNotificationConfigurationsType)
-> Result<DescribeNotificationConfigurationsAnswer, DescribeNotificationConfigurationsError>

Describes the notification actions associated with the specified Auto Scaling group.

fn describe_policies(&self,
input: &DescribePoliciesType)
-> Result<PoliciesType, DescribePoliciesError>

Describes the policies for the specified Auto Scaling group.

fn describe_scaling_activities(&self,
input: &DescribeScalingActivitiesType)
-> Result<ActivitiesType, DescribeScalingActivitiesError>

Describes one or more scaling activities for the specified Auto Scaling group.

fn describe_scaling_process_types(&self)
-> Result<ProcessesType, DescribeScalingProcessTypesError>

Describes the scaling process types for use with ResumeProcesses and SuspendProcesses.

fn describe_scheduled_actions(&self,
input: &DescribeScheduledActionsType)
-> Result<ScheduledActionsType, DescribeScheduledActionsError>

Describes the actions scheduled for your Auto Scaling group that haven't run. To describe the actions that have already run, use DescribeScalingActivities.

fn describe_tags(&self,
input: &DescribeTagsType)
-> Result<TagsType, DescribeTagsError>

Describes the specified tags.

You can use filters to limit the results. For example, you can query for the tags for a specific Auto Scaling group. You can specify multiple values for a filter. A tag must match at least one of the specified values for it to be included in the results.

You can also specify multiple filters. The result includes information for a particular tag only if it matches all the filters. If there's no match, no special message is returned.

fn describe_termination_policy_types(&self)
-> Result<DescribeTerminationPolicyTypesAnswer, DescribeTerminationPolicyTypesError>

Describes the termination policies supported by Auto Scaling.

fn detach_instances(&self,
input: &DetachInstancesQuery)
-> Result<DetachInstancesAnswer, DetachInstancesError>

Removes one or more instances from the specified Auto Scaling group.

After the instances are detached, you can manage them independently from the rest of the Auto Scaling group.

If you do not specify the option to decrement the desired capacity, Auto Scaling launches instances to replace the ones that are detached.

If there is a Classic load balancer attached to the Auto Scaling group, the instances are deregistered from the load balancer. If there are target groups attached to the Auto Scaling group, the instances are deregistered from the target groups.

For more information, see Detach EC2 Instances from Your Auto Scaling Group in the Auto Scaling User Guide.

fn detach_load_balancer_target_groups(&self,
input: &DetachLoadBalancerTargetGroupsType)
-> Result<DetachLoadBalancerTargetGroupsResultType, DetachLoadBalancerTargetGroupsError>

Detaches one or more target groups from the specified Auto Scaling group.

fn detach_load_balancers(&self,
input: &DetachLoadBalancersType)
-> Result<DetachLoadBalancersResultType, DetachLoadBalancersError>

Detaches one or more Classic load balancers from the specified Auto Scaling group.

Note that this operation detaches only Classic load balancers. If you have Application load balancers, use DetachLoadBalancerTargetGroups instead.

When you detach a load balancer, it enters the Removing state while deregistering the instances in the group. When all instances are deregistered, then you can no longer describe the load balancer using DescribeLoadBalancers. Note that the instances remain running.

fn disable_metrics_collection(&self,
input: &DisableMetricsCollectionQuery)
-> Result<(), DisableMetricsCollectionError>

Disables group metrics for the specified Auto Scaling group.

fn enable_metrics_collection(&self,
input: &EnableMetricsCollectionQuery)
-> Result<(), EnableMetricsCollectionError>

Enables group metrics for the specified Auto Scaling group. For more information, see Monitoring Your Auto Scaling Groups and Instances in the Auto Scaling User Guide.

fn enter_standby(&self,
input: &EnterStandbyQuery)
-> Result<EnterStandbyAnswer, EnterStandbyError>

Moves the specified instances into Standby mode.

For more information, see Auto Scaling Lifecycle in the Auto Scaling User Guide.

fn execute_policy(&self,
input: &ExecutePolicyType)
-> Result<(), ExecutePolicyError>

Executes the specified policy.

fn exit_standby(&self,
input: &ExitStandbyQuery)
-> Result<ExitStandbyAnswer, ExitStandbyError>

Moves the specified instances out of Standby mode.

For more information, see Auto Scaling Lifecycle in the Auto Scaling User Guide.

fn put_lifecycle_hook(&self,
input: &PutLifecycleHookType)
-> Result<PutLifecycleHookAnswer, PutLifecycleHookError>

Creates or updates a lifecycle hook for the specified Auto Scaling Group.

A lifecycle hook tells Auto Scaling that you want to perform an action on an instance that is not actively in service; for example, either when the instance launches or before the instance terminates.

This step is a part of the procedure for adding a lifecycle hook to an Auto Scaling group:

(Optional) Create a Lambda function and a rule that allows CloudWatch Events to invoke your Lambda function when Auto Scaling launches or terminates instances.

(Optional) Create a notification target and an IAM role. The target can be either an Amazon SQS queue or an Amazon SNS topic. The role allows Auto Scaling to publish lifecycle notifications to the target.

Create the lifecycle hook. Specify whether the hook is used when the instances launch or terminate.

If you need more time, record the lifecycle action heartbeat to keep the instance in a pending state.

If you finish before the timeout period ends, complete the lifecycle action.

For more information, see Auto Scaling Lifecycle Hooks in the Auto Scaling User Guide.

If you exceed your maximum limit of lifecycle hooks, which by default is 50 per Auto Scaling group, the call fails. For information about updating this limit, see AWS Service Limits in the Amazon Web Services General Reference.

fn put_notification_configuration(&self,
input: &PutNotificationConfigurationType)
-> Result<(), PutNotificationConfigurationError>

Configures an Auto Scaling group to send notifications when specified events take place. Subscribers to the specified topic can have messages delivered to an endpoint such as a web server or an email address.

This configuration overwrites any existing configuration.

For more information see Getting SNS Notifications When Your Auto Scaling Group Scales in the Auto Scaling User Guide.

fn put_scaling_policy(&self,
input: &PutScalingPolicyType)
-> Result<PolicyARNType, PutScalingPolicyError>

Creates or updates a policy for an Auto Scaling group. To update an existing policy, use the existing policy name and set the parameters you want to change. Any existing parameter not changed in an update to an existing policy is not changed in this update request.

If you exceed your maximum limit of step adjustments, which by default is 20 per region, the call fails. For information about updating this limit, see AWS Service Limits in the Amazon Web Services General Reference.

fn put_scheduled_update_group_action(&self,
input: &PutScheduledUpdateGroupActionType)
-> Result<(), PutScheduledUpdateGroupActionError>

Creates or updates a scheduled scaling action for an Auto Scaling group. When updating a scheduled scaling action, if you leave a parameter unspecified, the corresponding value remains unchanged.

For more information, see Scheduled Scaling in the Auto Scaling User Guide.

fn record_lifecycle_action_heartbeat(&self,
input: &RecordLifecycleActionHeartbeatType)
-> Result<RecordLifecycleActionHeartbeatAnswer, RecordLifecycleActionHeartbeatError>

Records a heartbeat for the lifecycle action associated with the specified token or instance. This extends the timeout by the length of time defined using PutLifecycleHook.

This step is a part of the procedure for adding a lifecycle hook to an Auto Scaling group:

(Optional) Create a Lambda function and a rule that allows CloudWatch Events to invoke your Lambda function when Auto Scaling launches or terminates instances.

(Optional) Create a notification target and an IAM role. The target can be either an Amazon SQS queue or an Amazon SNS topic. The role allows Auto Scaling to publish lifecycle notifications to the target.

Create the lifecycle hook. Specify whether the hook is used when the instances launch or terminate.

If you need more time, record the lifecycle action heartbeat to keep the instance in a pending state.

If you finish before the timeout period ends, complete the lifecycle action.

For more information, see Auto Scaling Lifecycle in the Auto Scaling User Guide.

fn resume_processes(&self,
input: &ScalingProcessQuery)
-> Result<(), ResumeProcessesError>

Resumes the specified suspended Auto Scaling processes, or all suspended process, for the specified Auto Scaling group.

For more information, see Suspending and Resuming Auto Scaling Processes in the Auto Scaling User Guide.

fn set_desired_capacity(&self,
input: &SetDesiredCapacityType)
-> Result<(), SetDesiredCapacityError>

Sets the size of the specified Auto Scaling group.

For more information about desired capacity, see What Is Auto Scaling? in the Auto Scaling User Guide.

fn set_instance_health(&self,
input: &SetInstanceHealthQuery)
-> Result<(), SetInstanceHealthError>

Sets the health status of the specified instance.

For more information, see Health Checks in the Auto Scaling User Guide.

fn set_instance_protection(&self,
input: &SetInstanceProtectionQuery)
-> Result<SetInstanceProtectionAnswer, SetInstanceProtectionError>

Updates the instance protection settings of the specified instances.

For more information, see Instance Protection in the Auto Scaling User Guide.

fn suspend_processes(&self,
input: &ScalingProcessQuery)
-> Result<(), SuspendProcessesError>

Suspends the specified Auto Scaling processes, or all processes, for the specified Auto Scaling group.

Note that if you suspend either the Launch or Terminate process types, it can prevent other process types from functioning properly.

To resume processes that have been suspended, use ResumeProcesses.

For more information, see Suspending and Resuming Auto Scaling Processes in the Auto Scaling User Guide.

fn terminate_instance_in_auto_scaling_group(&self,
input: &TerminateInstanceInAutoScalingGroupType)
-> Result<ActivityType, TerminateInstanceInAutoScalingGroupError>

Terminates the specified instance and optionally adjusts the desired group size.

This call simply makes a termination request. The instance is not terminated immediately.

fn update_auto_scaling_group(&self,
input: &UpdateAutoScalingGroupType)
-> Result<(), UpdateAutoScalingGroupError>

Updates the configuration for the specified Auto Scaling group.

To update an Auto Scaling group with a launch configuration with InstanceMonitoring set to False, you must first disable the collection of group metrics. Otherwise, you will get an error. If you have previously enabled the collection of group metrics, you can disable it using DisableMetricsCollection.

The new settings are registered upon the completion of this call. Any launch configuration settings take effect on any triggers after this call returns. Scaling activities that are currently in progress aren't affected.

Note the following:

If you specify a new value for MinSize without specifying a value for DesiredCapacity, and the new MinSize is larger than the current size of the group, we implicitly call SetDesiredCapacity to set the size of the group to the new value of MinSize.

If you specify a new value for MaxSize without specifying a value for DesiredCapacity, and the new MaxSize is smaller than the current size of the group, we implicitly call SetDesiredCapacity to set the size of the group to the new value of MaxSize.

All other optional parameters are left unchanged if not specified.