Struct rusoto::cloudformation::CloudFormationClient

pub struct CloudFormationClient<P, D> where P: ProvideAwsCredentials,
        D: DispatchSignedRequest { /* fields omitted */ }

A client for the AWS CloudFormation API.

Methods

impl<P, D> CloudFormationClient<P, D> where P: ProvideAwsCredentials,
D: DispatchSignedRequest

fn new(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn cancel_update_stack(&self,
input: &CancelUpdateStackInput)
-> Result<(), CancelUpdateStackError>

Cancels an update on the specified stack. If the call completes successfully, the stack rolls back the update and reverts to the previous stack configuration.

You can cancel only stacks that are in the UPDATE_IN_PROGRESS state.

fn continue_update_rollback(&self,
input: &ContinueUpdateRollbackInput)
-> Result<ContinueUpdateRollbackOutput, ContinueUpdateRollbackError>

For a specified stack that is in the UPDATE_ROLLBACK_FAILED state, continues rolling it back to the UPDATE_ROLLBACK_COMPLETE state. Depending on the cause of the failure, you can manually fix the error and continue the rollback. By continuing the rollback, you can return your stack to a working state (the UPDATE_ROLLBACK_COMPLETE state), and then try to update the stack again.

A stack goes into the UPDATE_ROLLBACK_FAILED state when AWS CloudFormation cannot roll back all changes after a failed stack update. For example, you might have a stack that is rolling back to an old database instance that was deleted outside of AWS CloudFormation. Because AWS CloudFormation doesn't know the database was deleted, it assumes that the database instance still exists and attempts to roll back to it, causing the update rollback to fail.

fn create_change_set(&self,
input: &CreateChangeSetInput)
-> Result<CreateChangeSetOutput, CreateChangeSetError>

Creates a list of changes for a stack. AWS CloudFormation generates the change set by comparing the template's information with the information that you submit. A change set can help you understand which resources AWS CloudFormation will change, and how it will change them, before you update your stack. Change sets allow you to check before making a change to avoid deleting or replacing critical resources.

AWS CloudFormation doesn't make any changes to the stack when you create a change set. To make the specified changes, you must execute the change set by using the ExecuteChangeSet action.

After the call successfully completes, AWS CloudFormation starts creating the change set. To check the status of the change set, use the DescribeChangeSet action.

fn create_stack(&self,
input: &CreateStackInput)
-> Result<CreateStackOutput, CreateStackError>

Creates a stack as specified in the template. After the call completes successfully, the stack creation starts. You can check the status of the stack via the DescribeStacks API.

fn delete_change_set(&self,
input: &DeleteChangeSetInput)
-> Result<DeleteChangeSetOutput, DeleteChangeSetError>

Deletes the specified change set. Deleting change sets ensures that no one executes the wrong change set.

If the call successfully completes, AWS CloudFormation successfully deleted the change set.

fn delete_stack(&self, input: &DeleteStackInput) -> Result<(), DeleteStackError>

Deletes a specified stack. Once the call completes successfully, stack deletion starts. Deleted stacks do not show up in the DescribeStacks API if the deletion has been completed successfully.

fn describe_account_limits(&self,
input: &DescribeAccountLimitsInput)
-> Result<DescribeAccountLimitsOutput, DescribeAccountLimitsError>

Retrieves your account's AWS CloudFormation limits, such as the maximum number of stacks that you can create in your account.

fn describe_change_set(&self,
input: &DescribeChangeSetInput)
-> Result<DescribeChangeSetOutput, DescribeChangeSetError>

Returns the inputs for the change set and a list of changes that AWS CloudFormation will make if you execute the change set. For more information, see Updating Stacks Using Change Sets in the AWS CloudFormation User Guide.

fn describe_stack_events(&self,
input: &DescribeStackEventsInput)
-> Result<DescribeStackEventsOutput, DescribeStackEventsError>

Returns all stack related events for a specified stack in reverse chronological order. For more information about a stack's event history, go to Stacks in the AWS CloudFormation User Guide.

You can list events for stacks that have failed to create or have been deleted by specifying the unique stack identifier (stack ID).

fn describe_stack_resource(&self,
input: &DescribeStackResourceInput)
-> Result<DescribeStackResourceOutput, DescribeStackResourceError>

Returns a description of the specified resource in the specified stack.

For deleted stacks, DescribeStackResource returns resource information for up to 90 days after the stack has been deleted.

fn describe_stack_resources(&self,
input: &DescribeStackResourcesInput)
-> Result<DescribeStackResourcesOutput, DescribeStackResourcesError>

Returns AWS resource descriptions for running and deleted stacks. If StackName is specified, all the associated resources that are part of the stack are returned. If PhysicalResourceId is specified, the associated resources of the stack that the resource belongs to are returned.

Only the first 100 resources will be returned. If your stack has more resources than this, you should use ListStackResources instead.

For deleted stacks, DescribeStackResources returns resource information for up to 90 days after the stack has been deleted.

You must specify either StackName or PhysicalResourceId, but not both. In addition, you can specify LogicalResourceId to filter the returned result. For more information about resources, the LogicalResourceId and PhysicalResourceId, go to the AWS CloudFormation User Guide.

A ValidationError is returned if you specify both StackName and PhysicalResourceId in the same request.

fn describe_stacks(&self,
input: &DescribeStacksInput)
-> Result<DescribeStacksOutput, DescribeStacksError>

Returns the description for the specified stack; if no stack name was specified, then it returns the description for all the stacks created.

If the stack does not exist, an AmazonCloudFormationException is returned.

fn estimate_template_cost(&self,
input: &EstimateTemplateCostInput)
-> Result<EstimateTemplateCostOutput, EstimateTemplateCostError>

Returns the estimated monthly cost of a template. The return value is an AWS Simple Monthly Calculator URL with a query string that describes the resources required to run the template.

fn execute_change_set(&self,
input: &ExecuteChangeSetInput)
-> Result<ExecuteChangeSetOutput, ExecuteChangeSetError>

Updates a stack using the input information that was provided when the specified change set was created. After the call successfully completes, AWS CloudFormation starts updating the stack. Use the DescribeStacks action to view the status of the update.

When you execute a change set, AWS CloudFormation deletes all other change sets associated with the stack because they aren't valid for the updated stack.

If a stack policy is associated with the stack, AWS CloudFormation enforces the policy during the update. You can't specify a temporary stack policy that overrides the current policy.

fn get_stack_policy(&self,
input: &GetStackPolicyInput)
-> Result<GetStackPolicyOutput, GetStackPolicyError>

Returns the stack policy for a specified stack. If a stack doesn't have a policy, a null value is returned.

fn get_template(&self,
input: &GetTemplateInput)
-> Result<GetTemplateOutput, GetTemplateError>

Returns the template body for a specified stack. You can get the template for running or deleted stacks.

For deleted stacks, GetTemplate returns the template for up to 90 days after the stack has been deleted.

If the template does not exist, a ValidationError is returned.

fn get_template_summary(&self,
input: &GetTemplateSummaryInput)
-> Result<GetTemplateSummaryOutput, GetTemplateSummaryError>

Returns information about a new or existing template. The GetTemplateSummary action is useful for viewing parameter information, such as default parameter values and parameter types, before you create or update a stack.

You can use the GetTemplateSummary action when you submit a template, or you can get template information for a running or deleted stack.

For deleted stacks, GetTemplateSummary returns the template information for up to 90 days after the stack has been deleted. If the template does not exist, a ValidationError is returned.

fn list_change_sets(&self,
input: &ListChangeSetsInput)
-> Result<ListChangeSetsOutput, ListChangeSetsError>

Returns the ID and status of each active change set for a stack. For example, AWS CloudFormation lists change sets that are in the CREATE_IN_PROGRESS or CREATE_PENDING state.

fn list_exports(&self,
input: &ListExportsInput)
-> Result<ListExportsOutput, ListExportsError>

Lists all exported output values in the account and region in which you call this action. Use this action to see the exported output values that you can import into other stacks. To import values, use the Fn::ImportValue function.

For more information, see AWS CloudFormation Export Stack Output Values.

fn list_imports(&self,
input: &ListImportsInput)
-> Result<ListImportsOutput, ListImportsError>

Lists all stacks that are importing an exported output value. To modify or remove an exported output value, first use this action to see which stacks are using it. To see the exported output values in your account, see ListExports.

For more information about importing an exported output value, see the Fn::ImportValue function.

fn list_stack_resources(&self,
input: &ListStackResourcesInput)
-> Result<ListStackResourcesOutput, ListStackResourcesError>

Returns descriptions of all resources of the specified stack.

For deleted stacks, ListStackResources returns resource information for up to 90 days after the stack has been deleted.

fn list_stacks(&self,
input: &ListStacksInput)
-> Result<ListStacksOutput, ListStacksError>

Returns the summary information for stacks whose status matches the specified StackStatusFilter. Summary information for stacks that have been deleted is kept for 90 days after the stack is deleted. If no StackStatusFilter is specified, summary information for all stacks is returned (including existing stacks and stacks that have been deleted).

fn set_stack_policy(&self,
input: &SetStackPolicyInput)
-> Result<(), SetStackPolicyError>

Sets a stack policy for a specified stack.

fn signal_resource(&self,
input: &SignalResourceInput)
-> Result<(), SignalResourceError>

Sends a signal to the specified resource with a success or failure status. You can use the SignalResource API in conjunction with a creation policy or update policy. AWS CloudFormation doesn't proceed with a stack creation or update until resources receive the required number of signals or the timeout period is exceeded. The SignalResource API is useful in cases where you want to send signals from anywhere other than an Amazon EC2 instance.

fn update_stack(&self,
input: &UpdateStackInput)
-> Result<UpdateStackOutput, UpdateStackError>

Updates a stack as specified in the template. After the call completes successfully, the stack update starts. You can check the status of the stack via the DescribeStacks action.

To get a copy of the template for an existing stack, you can use the GetTemplate action.

For more information about creating an update template, updating a stack, and monitoring the progress of the update, see Updating a Stack.

fn validate_template(&self,
input: &ValidateTemplateInput)
-> Result<ValidateTemplateOutput, ValidateTemplateError>

Validates a specified template. AWS CloudFormation first checks if the template is valid JSON. If it isn't, AWS CloudFormation checks if the template is valid YAML. If both these checks fail, AWS CloudFormation returns a template validation error.