Module rusoto::cloudfront

The CloudFront API.

Structs

ActiveTrustedSigners
A complex type that lists the AWS accounts, if any, that you included in the TrustedSigners complex type for the default cache behavior or for any of the other cache behaviors for this distribution. These are accounts that you want to allow to create signed URLs for private content.

ActiveTrustedSignersSerializer

AliasListSerializer

Aliases
A complex type that contains information about CNAMEs (alternate domain names), if any, for this distribution.

AliasesSerializer

AllowedMethods
A complex type that controls which HTTP methods CloudFront processes and forwards to your Amazon S3 bucket or your custom origin. There are three choices: - CloudFront forwards only GET and HEAD requests. - CloudFront forwards only GET, HEAD and OPTIONS requests. - CloudFront forwards GET, HEAD, OPTIONS, PUT, PATCH, POST, and DELETE requests. If you pick the third choice, you may need to restrict access to your Amazon S3 bucket or to your custom origin so users can't perform operations that you don't want them to. For example, you may not want users to have permission to delete objects from your origin.

AllowedMethodsSerializer

AwsAccountNumberListSerializer

BooleanSerializer

CacheBehavior
A complex type that describes how CloudFront processes requests. You can create up to 10 cache behaviors.You must create at least as many cache behaviors (including the default cache behavior) as you have origins if you want CloudFront to distribute objects from all of the origins. Each cache behavior specifies the one origin from which you want CloudFront to get objects. If you have two origins and only the default cache behavior, the default cache behavior will cause CloudFront to get objects from one of the origins, but the other origin will never be used. If you don't want to specify any cache behaviors, include only an empty CacheBehaviors element. Don't include an empty CacheBehavior element, or CloudFront returns a MalformedXML error. To delete all cache behaviors in an existing distribution, update the distribution configuration and include only an empty CacheBehaviors element. To add, change, or remove one or more cache behaviors, update the distribution configuration and specify all of the cache behaviors that you want to include in the updated distribution.

CacheBehaviorListSerializer

CacheBehaviorSerializer

CacheBehaviors
A complex type that contains zero or more CacheBehavior elements.

CacheBehaviorsSerializer

CachedMethods
A complex type that controls whether CloudFront caches the response to requests using the specified HTTP methods. There are two choices: - CloudFront caches responses to GET and HEAD requests. - CloudFront caches responses to GET, HEAD, and OPTIONS requests. If you pick the second choice for your S3 Origin, you may need to forward Access-Control-Request-Method, Access-Control-Request-Headers and Origin headers for the responses to be cached correctly.

CachedMethodsSerializer

CertificateSourceSerializer

CloudFrontClient
A client for the CloudFront API.

CloudFrontOriginAccessIdentity
CloudFront origin access identity.

CloudFrontOriginAccessIdentityConfig
Origin access identity configuration.

CloudFrontOriginAccessIdentityConfigSerializer

CloudFrontOriginAccessIdentityList
The CloudFrontOriginAccessIdentityList type.

CloudFrontOriginAccessIdentityListSerializer

CloudFrontOriginAccessIdentitySerializer

CloudFrontOriginAccessIdentitySummary
Summary of the information about a CloudFront origin access identity.

CloudFrontOriginAccessIdentitySummaryListSerializer

CloudFrontOriginAccessIdentitySummarySerializer

CookieNameListSerializer

CookieNames
A complex type that specifies the whitelisted cookies, if any, that you want CloudFront to forward to your origin that is associated with this cache behavior.

CookieNamesSerializer

CookiePreference
A complex type that specifies the cookie preferences associated with this cache behavior.

CookiePreferenceSerializer

CreateCloudFrontOriginAccessIdentityRequest
The request to create a new origin access identity.

CreateCloudFrontOriginAccessIdentityResult
The returned result of the corresponding request.

CreateCloudFrontOriginAccessIdentityResultSerializer

CreateDistributionRequest
The request to create a new distribution.

CreateDistributionResult
The returned result of the corresponding request.

CreateDistributionResultSerializer

CreateDistributionWithTagsRequest
The request to create a new distribution with tags

CreateDistributionWithTagsResult
The returned result of the corresponding request.

CreateDistributionWithTagsResultSerializer

CreateInvalidationRequest
The request to create an invalidation.

CreateInvalidationResult
The returned result of the corresponding request.

CreateInvalidationResultSerializer

CreateStreamingDistributionRequest
The request to create a new streaming distribution.

CreateStreamingDistributionResult
The returned result of the corresponding request.

CreateStreamingDistributionResultSerializer

CreateStreamingDistributionWithTagsRequest
The request to create a new streaming distribution with tags.

CreateStreamingDistributionWithTagsResult
The returned result of the corresponding request.

CreateStreamingDistributionWithTagsResultSerializer

CustomErrorResponse
A complex type that describes how you'd prefer CloudFront to respond to requests that result in either a 4xx or 5xx response. You can control whether a custom error page should be displayed, what the desired response code should be for this error page and how long should the error response be cached by CloudFront. If you don't want to specify any custom error responses, include only an empty CustomErrorResponses element. To delete all custom error responses in an existing distribution, update the distribution configuration and include only an empty CustomErrorResponses element. To add, change, or remove one or more custom error responses, update the distribution configuration and specify all of the custom error responses that you want to include in the updated distribution.

CustomErrorResponseListSerializer

CustomErrorResponseSerializer

CustomErrorResponses
A complex type that contains zero or more CustomErrorResponse elements.

CustomErrorResponsesSerializer

CustomHeaders
A complex type that contains the list of Custom Headers for each origin.

CustomHeadersSerializer

CustomOriginConfig
A customer origin.

CustomOriginConfigSerializer

DefaultCacheBehavior
A complex type that describes the default cache behavior if you do not specify a CacheBehavior element or if files don't match any of the values of PathPattern in CacheBehavior elements.You must create exactly one default cache behavior.

DefaultCacheBehaviorSerializer

DeleteCloudFrontOriginAccessIdentityRequest
The request to delete a origin access identity.

DeleteDistributionRequest
The request to delete a distribution.

DeleteStreamingDistributionRequest
The request to delete a streaming distribution.

Distribution
A distribution.

DistributionConfig
A distribution Configuration.

DistributionConfigSerializer

DistributionConfigWithTags
A distribution Configuration and a list of tags to be associated with the distribution.

DistributionConfigWithTagsSerializer

DistributionList
A distribution list.

DistributionListSerializer

DistributionSerializer

DistributionSummary
A summary of the information for an Amazon CloudFront distribution.

DistributionSummaryListSerializer

DistributionSummarySerializer

ForwardedValues
A complex type that specifies how CloudFront handles query strings, cookies and headers.

ForwardedValuesSerializer

GeoRestriction
A complex type that controls the countries in which your content is distributed. For more information about geo restriction, go to Customizing Error Responses in the Amazon CloudFront Developer Guide. CloudFront determines the location of your users using MaxMind GeoIP databases. For information about the accuracy of these databases, see How accurate are your GeoIP databases? on the MaxMind website.

GeoRestrictionSerializer

GeoRestrictionTypeSerializer

GetCloudFrontOriginAccessIdentityConfigRequest
The request to get an origin access identity's configuration.

GetCloudFrontOriginAccessIdentityConfigResult
The returned result of the corresponding request.

GetCloudFrontOriginAccessIdentityConfigResultSerializer

GetCloudFrontOriginAccessIdentityRequest
The request to get an origin access identity's information.

GetCloudFrontOriginAccessIdentityResult
The returned result of the corresponding request.

GetCloudFrontOriginAccessIdentityResultSerializer

GetDistributionConfigRequest
The request to get a distribution configuration.

GetDistributionConfigResult
The returned result of the corresponding request.

GetDistributionConfigResultSerializer

GetDistributionRequest
The request to get a distribution's information.

GetDistributionResult
The returned result of the corresponding request.

GetDistributionResultSerializer

GetInvalidationRequest
The request to get an invalidation's information.

GetInvalidationResult
The returned result of the corresponding request.

GetInvalidationResultSerializer

GetStreamingDistributionConfigRequest
To request to get a streaming distribution configuration.

GetStreamingDistributionConfigResult
The returned result of the corresponding request.

GetStreamingDistributionConfigResultSerializer

GetStreamingDistributionRequest
The request to get a streaming distribution's information.

GetStreamingDistributionResult
The returned result of the corresponding request.

GetStreamingDistributionResultSerializer

HeaderListSerializer

Headers
A complex type that specifies the headers that you want CloudFront to forward to the origin for this cache behavior. For the headers that you specify, CloudFront also caches separate versions of a given object based on the header values in viewer requests; this is known as varying on headers. For example, suppose viewer requests for logo.jpg contain a custom Product header that has a value of either Acme or Apex, and you configure CloudFront to vary on the Product header. CloudFront forwards the Product header to the origin and caches the response from the origin once for each header value.

HeadersSerializer

HttpVersionSerializer

IntegerSerializer

Invalidation
An invalidation.

InvalidationBatch
An invalidation batch.

InvalidationBatchSerializer

InvalidationList
An invalidation list.

InvalidationListSerializer

InvalidationSerializer

InvalidationSummary
Summary of an invalidation request.

InvalidationSummaryListSerializer

InvalidationSummarySerializer

ItemSelectionSerializer

KeyPairIdListSerializer

KeyPairIds
A complex type that lists the active CloudFront key pairs, if any, that are associated with AwsAccountNumber.

KeyPairIdsSerializer

ListCloudFrontOriginAccessIdentitiesRequest
The request to list origin access identities.

ListCloudFrontOriginAccessIdentitiesResult
The returned result of the corresponding request.

ListCloudFrontOriginAccessIdentitiesResultSerializer

ListDistributionsByWebACLIdRequest
The request to list distributions that are associated with a specified AWS WAF web ACL.

ListDistributionsByWebACLIdResult
The response to a request to list the distributions that are associated with a specified AWS WAF web ACL.

ListDistributionsByWebACLIdResultSerializer

ListDistributionsRequest
The request to list your distributions.

ListDistributionsResult
The returned result of the corresponding request.

ListDistributionsResultSerializer

ListInvalidationsRequest
The request to list invalidations.

ListInvalidationsResult
The returned result of the corresponding request.

ListInvalidationsResultSerializer

ListStreamingDistributionsRequest
The request to list your streaming distributions.

ListStreamingDistributionsResult
The returned result of the corresponding request.

ListStreamingDistributionsResultSerializer

ListTagsForResourceRequest
The request to list tags for a CloudFront resource.

ListTagsForResourceResult
The returned result of the corresponding request.

ListTagsForResourceResultSerializer

LocationListSerializer

LoggingConfig
A complex type that controls whether access logs are written for the distribution.

LoggingConfigSerializer

LongSerializer

MethodSerializer

MethodsListSerializer

MinimumProtocolVersionSerializer

Origin
A complex type that describes the Amazon S3 bucket or the HTTP server (for example, a web server) from which CloudFront gets your files.You must create at least one origin.

OriginCustomHeader
A complex type that contains information related to a Header

OriginCustomHeaderSerializer

OriginCustomHeadersListSerializer

OriginListSerializer

OriginProtocolPolicySerializer

OriginSerializer

OriginSslProtocols
A complex type that contains the list of SSL/TLS protocols that you want CloudFront to use when communicating with your origin over HTTPS.

OriginSslProtocolsSerializer

Origins
A complex type that contains information about origins for this distribution.

OriginsSerializer

PathListSerializer

Paths
A complex type that contains information about the objects that you want to invalidate.

PathsSerializer

PriceClassSerializer

QueryStringCacheKeys

QueryStringCacheKeysListSerializer

QueryStringCacheKeysSerializer

ResourceARNSerializer

Restrictions
A complex type that identifies ways in which you want to restrict distribution of your content.

RestrictionsSerializer

S3Origin
A complex type that contains information about the Amazon S3 bucket from which you want CloudFront to get your media files for distribution.

S3OriginConfig
A complex type that contains information about the Amazon S3 origin. If the origin is a custom origin, use the CustomOriginConfig element instead.

S3OriginConfigSerializer

S3OriginSerializer

SSLSupportMethodSerializer

Signer
A complex type that lists the AWS accounts that were included in the TrustedSigners complex type, as well as their active CloudFront key pair IDs, if any.

SignerListSerializer

SignerSerializer

SslProtocolSerializer

SslProtocolsListSerializer

StreamingDistribution
A streaming distribution.

StreamingDistributionConfig
The configuration for the streaming distribution.

StreamingDistributionConfigSerializer

StreamingDistributionConfigWithTags
A streaming distribution Configuration and a list of tags to be associated with the streaming distribution.

StreamingDistributionConfigWithTagsSerializer

StreamingDistributionList
A streaming distribution list.

StreamingDistributionListSerializer

StreamingDistributionSerializer

StreamingDistributionSummary
A summary of the information for an Amazon CloudFront streaming distribution.

StreamingDistributionSummaryListSerializer

StreamingDistributionSummarySerializer

StreamingLoggingConfig
A complex type that controls whether access logs are written for this streaming distribution.

StreamingLoggingConfigSerializer

StringSerializer

Tag
A complex type that contains Tag key and Tag value.

TagKeyListSerializer

TagKeySerializer

TagKeys
A complex type that contains zero or more Tag elements.

TagKeysSerializer

TagListSerializer

TagResourceRequest
The request to add tags to a CloudFront resource.

TagSerializer

TagValueSerializer

Tags
A complex type that contains zero or more Tag elements.

TagsSerializer

TimestampSerializer

TrustedSigners
A complex type that specifies the AWS accounts, if any, that you want to allow to create signed URLs for private content. If you want to require signed URLs in requests for objects in the target origin that match the PathPattern for this cache behavior, specify true for Enabled, and specify the applicable values for Quantity and Items. For more information, go to Using a Signed URL to Serve Private Content in the Amazon CloudFront Developer Guide. If you don't want to require signed URLs in requests for objects that match PathPattern, specify false for Enabled and 0 for Quantity. Omit Items. To add, change, or remove one or more trusted signers, change Enabled to true (if it's currently false), change Quantity as applicable, and specify all of the trusted signers that you want to include in the updated distribution.

TrustedSignersSerializer

UntagResourceRequest
The request to remove tags from a CloudFront resource.

UpdateCloudFrontOriginAccessIdentityRequest
The request to update an origin access identity.

UpdateCloudFrontOriginAccessIdentityResult
The returned result of the corresponding request.

UpdateCloudFrontOriginAccessIdentityResultSerializer

UpdateDistributionRequest
The request to update a distribution.

UpdateDistributionResult
The returned result of the corresponding request.

UpdateDistributionResultSerializer

UpdateStreamingDistributionRequest
The request to update a streaming distribution.

UpdateStreamingDistributionResult
The returned result of the corresponding request.

UpdateStreamingDistributionResultSerializer

ViewerCertificate
A complex type that contains information about viewer certificates for this distribution.

ViewerCertificateSerializer

ViewerProtocolPolicySerializer

Enums

CreateCloudFrontOriginAccessIdentityError
Errors returned by CreateCloudFrontOriginAccessIdentity

CreateDistributionError
Errors returned by CreateDistribution

CreateDistributionWithTagsError
Errors returned by CreateDistributionWithTags

CreateInvalidationError
Errors returned by CreateInvalidation

CreateStreamingDistributionError
Errors returned by CreateStreamingDistribution

CreateStreamingDistributionWithTagsError
Errors returned by CreateStreamingDistributionWithTags

DeleteCloudFrontOriginAccessIdentityError
Errors returned by DeleteCloudFrontOriginAccessIdentity

DeleteDistributionError
Errors returned by DeleteDistribution

DeleteStreamingDistributionError
Errors returned by DeleteStreamingDistribution

GetCloudFrontOriginAccessIdentityConfigError
Errors returned by GetCloudFrontOriginAccessIdentityConfig

GetCloudFrontOriginAccessIdentityError
Errors returned by GetCloudFrontOriginAccessIdentity

GetDistributionConfigError
Errors returned by GetDistributionConfig

GetDistributionError
Errors returned by GetDistribution

GetInvalidationError
Errors returned by GetInvalidation

GetStreamingDistributionConfigError
Errors returned by GetStreamingDistributionConfig

GetStreamingDistributionError
Errors returned by GetStreamingDistribution

ListCloudFrontOriginAccessIdentitiesError
Errors returned by ListCloudFrontOriginAccessIdentities

ListDistributionsByWebACLIdError
Errors returned by ListDistributionsByWebACLId

ListDistributionsError
Errors returned by ListDistributions

ListInvalidationsError
Errors returned by ListInvalidations

ListStreamingDistributionsError
Errors returned by ListStreamingDistributions

ListTagsForResourceError
Errors returned by ListTagsForResource

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateCloudFrontOriginAccessIdentityError
Errors returned by UpdateCloudFrontOriginAccessIdentity

UpdateDistributionError
Errors returned by UpdateDistribution

UpdateStreamingDistributionError
Errors returned by UpdateStreamingDistribution

Type Definitions

AliasList

AwsAccountNumberList

Boolean

CacheBehaviorList

CertificateSource

CloudFrontOriginAccessIdentitySummaryList

CookieNameList

CustomErrorResponseList

DistributionSummaryList

GeoRestrictionType

HeaderList

HttpVersion

Integer

InvalidationSummaryList

ItemSelection

KeyPairIdList

LocationList

Long

Method

MethodsList

MinimumProtocolVersion

OriginCustomHeadersList

OriginList

OriginProtocolPolicy

PathList

PriceClass

QueryStringCacheKeysList

ResourceARN

SSLSupportMethod

SignerList

SslProtocol

SslProtocolsList

StreamingDistributionSummaryList

TagKey
A string that contains Tag key. The string length should be between 1 and 128 characters. Valid characters include a-z, A-Z, 0-9, space, and the special characters _ - . : / = + @.

TagKeyList

TagList

TagValue

Timestamp

ViewerProtocolPolicy