Struct rusoto::cloudfront::CloudFrontClient

pub struct CloudFrontClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest { /* fields omitted */ }

A client for the CloudFront API.

Methods

impl<P> CloudFrontClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> CloudFrontClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D,
credentials_provider: P,
region: Region)
-> Self

fn create_cloud_front_origin_access_identity(&self,
input: &CreateCloudFrontOriginAccessIdentityRequest)
-> Result<CreateCloudFrontOriginAccessIdentityResult, CreateCloudFrontOriginAccessIdentityError>

Create a new origin access identity.

fn create_distribution(&self,
input: &CreateDistributionRequest)
-> Result<CreateDistributionResult, CreateDistributionError>

Create a new distribution.

fn create_distribution_with_tags(&self,
input: &CreateDistributionWithTagsRequest)
-> Result<CreateDistributionWithTagsResult, CreateDistributionWithTagsError>

Create a new distribution with tags.

fn create_invalidation(&self,
input: &CreateInvalidationRequest)
-> Result<CreateInvalidationResult, CreateInvalidationError>

Create a new invalidation.

fn create_streaming_distribution(&self,
input: &CreateStreamingDistributionRequest)
-> Result<CreateStreamingDistributionResult, CreateStreamingDistributionError>

Create a new streaming distribution.

fn create_streaming_distribution_with_tags(&self,
input: &CreateStreamingDistributionWithTagsRequest)
-> Result<CreateStreamingDistributionWithTagsResult, CreateStreamingDistributionWithTagsError>

Create a new streaming distribution with tags.

fn delete_cloud_front_origin_access_identity(&self,
input: &DeleteCloudFrontOriginAccessIdentityRequest)
-> Result<(), DeleteCloudFrontOriginAccessIdentityError>

Delete an origin access identity.

fn delete_distribution(&self,
input: &DeleteDistributionRequest)
-> Result<(), DeleteDistributionError>

Delete a distribution.

fn delete_streaming_distribution(&self,
input: &DeleteStreamingDistributionRequest)
-> Result<(), DeleteStreamingDistributionError>

Delete a streaming distribution.

fn get_cloud_front_origin_access_identity(&self,
input: &GetCloudFrontOriginAccessIdentityRequest)
-> Result<GetCloudFrontOriginAccessIdentityResult, GetCloudFrontOriginAccessIdentityError>

Get the information about an origin access identity.

fn get_cloud_front_origin_access_identity_config(&self,
input: &GetCloudFrontOriginAccessIdentityConfigRequest)
-> Result<GetCloudFrontOriginAccessIdentityConfigResult, GetCloudFrontOriginAccessIdentityConfigError>

Get the configuration information about an origin access identity.

fn get_distribution(&self,
input: &GetDistributionRequest)
-> Result<GetDistributionResult, GetDistributionError>

Get the information about a distribution.

fn get_distribution_config(&self,
input: &GetDistributionConfigRequest)
-> Result<GetDistributionConfigResult, GetDistributionConfigError>

Get the configuration information about a distribution.

fn get_invalidation(&self,
input: &GetInvalidationRequest)
-> Result<GetInvalidationResult, GetInvalidationError>

Get the information about an invalidation.

fn get_streaming_distribution(&self,
input: &GetStreamingDistributionRequest)
-> Result<GetStreamingDistributionResult, GetStreamingDistributionError>

Get the information about a streaming distribution.

fn get_streaming_distribution_config(&self,
input: &GetStreamingDistributionConfigRequest)
-> Result<GetStreamingDistributionConfigResult, GetStreamingDistributionConfigError>

Get the configuration information about a streaming distribution.

fn list_cloud_front_origin_access_identities(&self,
input: &ListCloudFrontOriginAccessIdentitiesRequest)
-> Result<ListCloudFrontOriginAccessIdentitiesResult, ListCloudFrontOriginAccessIdentitiesError>

List origin access identities.

fn list_distributions(&self,
input: &ListDistributionsRequest)
-> Result<ListDistributionsResult, ListDistributionsError>

List distributions.

fn list_distributions_by_web_acl_id(&self,
input: &ListDistributionsByWebACLIdRequest)
-> Result<ListDistributionsByWebACLIdResult, ListDistributionsByWebACLIdError>

List the distributions that are associated with a specified AWS WAF web ACL.

fn list_invalidations(&self,
input: &ListInvalidationsRequest)
-> Result<ListInvalidationsResult, ListInvalidationsError>

List invalidation batches.

fn list_streaming_distributions(&self,
input: &ListStreamingDistributionsRequest)
-> Result<ListStreamingDistributionsResult, ListStreamingDistributionsError>

List streaming distributions.

fn list_tags_for_resource(&self,
input: &ListTagsForResourceRequest)
-> Result<ListTagsForResourceResult, ListTagsForResourceError>

List tags for a CloudFront resource.

fn tag_resource(&self,
input: &TagResourceRequest)
-> Result<(), TagResourceError>

Add tags to a CloudFront resource.

fn untag_resource(&self,
input: &UntagResourceRequest)
-> Result<(), UntagResourceError>

Remove tags from a CloudFront resource.

fn update_cloud_front_origin_access_identity(&self,
input: &UpdateCloudFrontOriginAccessIdentityRequest)
-> Result<UpdateCloudFrontOriginAccessIdentityResult, UpdateCloudFrontOriginAccessIdentityError>

Update an origin access identity.

fn update_distribution(&self,
input: &UpdateDistributionRequest)
-> Result<UpdateDistributionResult, UpdateDistributionError>

Update a distribution.

fn update_streaming_distribution(&self,
input: &UpdateStreamingDistributionRequest)
-> Result<UpdateStreamingDistributionResult, UpdateStreamingDistributionError>

Update a streaming distribution.