Struct rusoto::cloudfront::DistributionConfig

pub struct DistributionConfig {
    pub aliases: Option<Aliases>,
    pub cache_behaviors: Option<CacheBehaviors>,
    pub caller_reference: String,
    pub comment: String,
    pub custom_error_responses: Option<CustomErrorResponses>,
    pub default_cache_behavior: DefaultCacheBehavior,
    pub default_root_object: Option<String>,
    pub enabled: Boolean,
    pub http_version: Option<HttpVersion>,
    pub logging: Option<LoggingConfig>,
    pub origins: Origins,
    pub price_class: Option<PriceClass>,
    pub restrictions: Option<Restrictions>,
    pub viewer_certificate: Option<ViewerCertificate>,
    pub web_acl_id: Option<String>,
}

A distribution Configuration.

Fields
aliases: Option<Aliases>

A complex type that contains information about CNAMEs (alternate domain names), if any, for this distribution.

cache_behaviors: Option<CacheBehaviors>

A complex type that contains zero or more CacheBehavior elements.

caller_reference: String

A unique number that ensures the request can't be replayed. If the CallerReference is new (no matter the content of the DistributionConfig object), a new distribution is created. If the CallerReference is a value you already sent in a previous request to create a distribution, and the content of the DistributionConfig is identical to the original request (ignoring white space), the response includes the same information returned to the original request. If the CallerReference is a value you already sent in a previous request to create a distribution but the content of the DistributionConfig is different from the original request, CloudFront returns a DistributionAlreadyExists error.

comment: String

Any comments you want to include about the distribution.

custom_error_responses: Option<CustomErrorResponses>

A complex type that contains zero or more CustomErrorResponse elements.

default_cache_behavior: DefaultCacheBehavior

A complex type that describes the default cache behavior if you do not specify a CacheBehavior element or if files don't match any of the values of PathPattern in CacheBehavior elements.You must create exactly one default cache behavior.

default_root_object: Option<String>

The object that you want CloudFront to return (for example, index.html) when an end user requests the root URL for your distribution (http://www.example.com) instead of an object in your distribution (http://www.example.com/index.html). Specifying a default root object avoids exposing the contents of your distribution. If you don't want to specify a default root object when you create a distribution, include an empty DefaultRootObject element. To delete the default root object from an existing distribution, update the distribution configuration and include an empty DefaultRootObject element. To replace the default root object, update the distribution configuration and specify the new object.

enabled: Boolean

Whether the distribution is enabled to accept end user requests for content.

http_version: Option<HttpVersion>

(Optional) Specify the maximum HTTP version that you want viewers to use to communicate with CloudFront. The default value for new web distributions is http2. Viewers that don't support HTTP/2 will automatically use an earlier version.

logging: Option<LoggingConfig>

A complex type that controls whether access logs are written for the distribution.

origins: Origins

A complex type that contains information about origins for this distribution.

price_class: Option<PriceClass>

A complex type that contains information about price class for this distribution.

restrictions: Option<Restrictions> viewer_certificate: Option<ViewerCertificate> web_acl_id: Option<String>

(Optional) If you're using AWS WAF to filter CloudFront requests, the Id of the AWS WAF web ACL that is associated with the distribution.