Struct rusoto::cloudhsm::CloudHsmClient

pub struct CloudHsmClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the CloudHSM API.

Methods

impl<P> CloudHsmClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> CloudHsmClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn add_tags_to_resource(&self, input: &AddTagsToResourceRequest) -> Result<AddTagsToResourceResponse, AddTagsToResourceError>

Adds or overwrites one or more tags for the specified AWS CloudHSM resource.

Each tag consists of a key and a value. Tag keys must be unique to each resource.

fn create_hapg(&self, input: &CreateHapgRequest) -> Result<CreateHapgResponse, CreateHapgError>

Creates a high-availability partition group. A high-availability partition group is a group of partitions that spans multiple physical HSMs.

fn create_hsm(&self, input: &CreateHsmRequest) -> Result<CreateHsmResponse, CreateHsmError>

Creates an uninitialized HSM instance.

There is an upfront fee charged for each HSM instance that you create with the CreateHsm operation. If you accidentally provision an HSM and want to request a refund, delete the instance using the DeleteHsm operation, go to the AWS Support Center, create a new case, and select Account and Billing Support.

It can take up to 20 minutes to create and provision an HSM. You can monitor the status of the HSM with the DescribeHsm operation. The HSM is ready to be initialized when the status changes to RUNNING.

fn create_luna_client(&self, input: &CreateLunaClientRequest) -> Result<CreateLunaClientResponse, CreateLunaClientError>

Creates an HSM client.

fn delete_hapg(&self, input: &DeleteHapgRequest) -> Result<DeleteHapgResponse, DeleteHapgError>

Deletes a high-availability partition group.

fn delete_hsm(&self, input: &DeleteHsmRequest) -> Result<DeleteHsmResponse, DeleteHsmError>

Deletes an HSM. After completion, this operation cannot be undone and your key material cannot be recovered.

fn delete_luna_client(&self, input: &DeleteLunaClientRequest) -> Result<DeleteLunaClientResponse, DeleteLunaClientError>

Deletes a client.

fn describe_hapg(&self, input: &DescribeHapgRequest) -> Result<DescribeHapgResponse, DescribeHapgError>

Retrieves information about a high-availability partition group.

fn describe_hsm(&self, input: &DescribeHsmRequest) -> Result<DescribeHsmResponse, DescribeHsmError>

Retrieves information about an HSM. You can identify the HSM by its ARN or its serial number.

fn describe_luna_client(&self, input: &DescribeLunaClientRequest) -> Result<DescribeLunaClientResponse, DescribeLunaClientError>

Retrieves information about an HSM client.

fn get_config(&self, input: &GetConfigRequest) -> Result<GetConfigResponse, GetConfigError>

Gets the configuration files necessary to connect to all high availability partition groups the client is associated with.

fn list_available_zones(&self, input: &ListAvailableZonesRequest) -> Result<ListAvailableZonesResponse, ListAvailableZonesError>

Lists the Availability Zones that have available AWS CloudHSM capacity.

fn list_hapgs(&self, input: &ListHapgsRequest) -> Result<ListHapgsResponse, ListHapgsError>

Lists the high-availability partition groups for the account.

This operation supports pagination with the use of the NextToken member. If more results are available, the NextToken member of the response contains a token that you pass in the next call to ListHapgs to retrieve the next set of items.

fn list_hsms(&self, input: &ListHsmsRequest) -> Result<ListHsmsResponse, ListHsmsError>

Retrieves the identifiers of all of the HSMs provisioned for the current customer.

This operation supports pagination with the use of the NextToken member. If more results are available, the NextToken member of the response contains a token that you pass in the next call to ListHsms to retrieve the next set of items.

fn list_luna_clients(&self, input: &ListLunaClientsRequest) -> Result<ListLunaClientsResponse, ListLunaClientsError>

Lists all of the clients.

This operation supports pagination with the use of the NextToken member. If more results are available, the NextToken member of the response contains a token that you pass in the next call to ListLunaClients to retrieve the next set of items.

fn list_tags_for_resource(&self, input: &ListTagsForResourceRequest) -> Result<ListTagsForResourceResponse, ListTagsForResourceError>

Returns a list of all tags for the specified AWS CloudHSM resource.

fn modify_hapg(&self, input: &ModifyHapgRequest) -> Result<ModifyHapgResponse, ModifyHapgError>

Modifies an existing high-availability partition group.

fn modify_hsm(&self, input: &ModifyHsmRequest) -> Result<ModifyHsmResponse, ModifyHsmError>

Modifies an HSM.

This operation can result in the HSM being offline for up to 15 minutes while the AWS CloudHSM service is reconfigured. If you are modifying a production HSM, you should ensure that your AWS CloudHSM service is configured for high availability, and consider executing this operation during a maintenance window.

fn modify_luna_client(&self, input: &ModifyLunaClientRequest) -> Result<ModifyLunaClientResponse, ModifyLunaClientError>

Modifies the certificate used by the client.

This action can potentially start a workflow to install the new certificate on the client's HSMs.

fn remove_tags_from_resource(&self, input: &RemoveTagsFromResourceRequest) -> Result<RemoveTagsFromResourceResponse, RemoveTagsFromResourceError>

Removes one or more tags from the specified AWS CloudHSM resource.

To remove a tag, specify only the tag key to remove (not the value). To overwrite the value for an existing tag, use AddTagsToResource.