Struct rusoto::cloudhsm::DescribeHsmResponse

pub struct DescribeHsmResponse {
    pub availability_zone: Option<AZ>,
    pub eni_id: Option<EniId>,
    pub eni_ip: Option<IpAddress>,
    pub hsm_arn: Option<HsmArn>,
    pub hsm_type: Option<String>,
    pub iam_role_arn: Option<IamRoleArn>,
    pub partitions: Option<PartitionList>,
    pub serial_number: Option<HsmSerialNumber>,
    pub server_cert_last_updated: Option<Timestamp>,
    pub server_cert_uri: Option<String>,
    pub software_version: Option<String>,
    pub ssh_key_last_updated: Option<Timestamp>,
    pub ssh_public_key: Option<SshKey>,
    pub status: Option<HsmStatus>,
    pub status_details: Option<String>,
    pub subnet_id: Option<SubnetId>,
    pub subscription_end_date: Option<Timestamp>,
    pub subscription_start_date: Option<Timestamp>,
    pub subscription_type: Option<SubscriptionType>,
    pub vendor_name: Option<String>,
    pub vpc_id: Option<VpcId>,
}

Contains the output of the DescribeHsm operation.

Fields
availability_zone: Option<AZ>

The Availability Zone that the HSM is in.

eni_id: Option<EniId>

The identifier of the elastic network interface (ENI) attached to the HSM.

eni_ip: Option<IpAddress>

The IP address assigned to the HSM's ENI.

hsm_arn: Option<HsmArn>

The ARN of the HSM.

hsm_type: Option<String>

The HSM model type.

iam_role_arn: Option<IamRoleArn>

The ARN of the IAM role assigned to the HSM.

partitions: Option<PartitionList>

The list of partitions on the HSM.

serial_number: Option<HsmSerialNumber>

The serial number of the HSM.

server_cert_last_updated: Option<Timestamp>

The date and time that the server certificate was last updated.

server_cert_uri: Option<String>

The URI of the certificate server.

software_version: Option<String>

The HSM software version.

ssh_key_last_updated: Option<Timestamp>

The date and time that the SSH key was last updated.

ssh_public_key: Option<SshKey>

The public SSH key.

status: Option<HsmStatus>

The status of the HSM.

status_details: Option<String>

Contains additional information about the status of the HSM.

subnet_id: Option<SubnetId>

The identifier of the subnet that the HSM is in.

subscription_end_date: Option<Timestamp>

The subscription end date.

subscription_start_date: Option<Timestamp>

The subscription start date.

subscription_type: Option<SubscriptionType> vendor_name: Option<String>

The name of the HSM vendor.

vpc_id: Option<VpcId>

The identifier of the VPC that the HSM is in.