Struct rusoto::cloudhsm::ModifyHsmRequest

pub struct ModifyHsmRequest {
    pub eni_ip: Option<IpAddress>,
    pub external_id: Option<ExternalId>,
    pub hsm_arn: HsmArn,
    pub iam_role_arn: Option<IamRoleArn>,
    pub subnet_id: Option<SubnetId>,
    pub syslog_ip: Option<IpAddress>,
}

Contains the inputs for the ModifyHsm operation.

Fields
eni_ip: Option<IpAddress>

The new IP address for the elastic network interface (ENI) attached to the HSM.

If the HSM is moved to a different subnet, and an IP address is not specified, an IP address will be randomly chosen from the CIDR range of the new subnet.

external_id: Option<ExternalId>

The new external ID.

hsm_arn: HsmArn

The ARN of the HSM to modify.

iam_role_arn: Option<IamRoleArn>

The new IAM role ARN.

subnet_id: Option<SubnetId>

The new identifier of the subnet that the HSM is in. The new subnet must be in the same Availability Zone as the current subnet.

syslog_ip: Option<IpAddress>

The new IP address for the syslog monitoring server. The AWS CloudHSM service only supports one syslog monitoring server.