Module rusoto::cloudtrail

AWS CloudTrail

Structs

AddTagsRequest
Specifies the tags to add to a trail.

AddTagsResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

CloudTrailClient
A client for the CloudTrail API.

CreateTrailRequest
Specifies the settings for each trail.

CreateTrailResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

DeleteTrailRequest
The request that specifies the name of a trail to delete.

DeleteTrailResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

DescribeTrailsRequest
Returns information about the trail.

DescribeTrailsResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

Event
Contains information about an event that was returned by a lookup request. The result includes a representation of a CloudTrail event.

GetTrailStatusRequest
The name of a trail about which you want the current status.

GetTrailStatusResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

ListPublicKeysRequest
Requests the public keys for a specified time range.

ListPublicKeysResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

ListTagsRequest
Specifies a list of trail tags to return.

ListTagsResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

LookupAttribute
Specifies an attribute and value that filter the events returned.

LookupEventsRequest
Contains a request for LookupEvents.

LookupEventsResponse
Contains a response to a LookupEvents action.

PublicKey
Contains information about a returned public key.

RemoveTagsRequest
Specifies the tags to remove from a trail.

RemoveTagsResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

Resource
Specifies the type and name of a resource referenced by an event.

ResourceTag
A resource tag.

StartLoggingRequest
The request to CloudTrail to start logging AWS API calls for an account.

StartLoggingResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

StopLoggingRequest
Passes the request to CloudTrail to stop logging AWS API calls for the specified account.

StopLoggingResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

Tag
A custom key-value pair associated with a resource such as a CloudTrail trail.

Trail
The settings for a trail.

UpdateTrailRequest
Specifies settings to update for the trail.

UpdateTrailResponse
Returns the objects or data listed below if successful. Otherwise, returns an error.

Enums

AddTagsError

CreateTrailError

DeleteTrailError

DescribeTrailsError

GetTrailStatusError

ListPublicKeysError

ListTagsError

LookupEventsError

RemoveTagsError

StartLoggingError

StopLoggingError

UpdateTrailError

Type Definitions

Boolean

ByteBuffer

Date

EventsList

LookupAttributeKey

LookupAttributesList

MaxResults

NextToken

PublicKeyList

ResourceIdList

ResourceList
A list of resources referenced by the event returned.

ResourceTagList
A list of resource tags.

TagsList
A list of tags.

TrailList

TrailNameList