Enum rusoto::cloudtrail::DeleteTrailError

pub enum DeleteTrailError {
    TrailNotFound(String),
    InvalidTrailName(String),
    InvalidHomeRegion(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Variants
TrailNotFound(String)

This exception is thrown when the trail with the given name is not found.

InvalidTrailName(String)

This exception is thrown when the provided trail name is not valid. Trail names must meet the following requirements:

Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-)

Start with a letter or number, and end with a letter or number

Be between 3 and 128 characters

Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are invalid.

Not be in IP address format (for example, 192.168.5.4)

InvalidHomeRegion(String)

This exception is thrown when an operation is called on a trail from a region other than the region in which the trail was created.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl DeleteTrailError

fn from_body(body: &str) -> DeleteTrailError