Struct rusoto::cloudtrail::CreateTrailResponse

pub struct CreateTrailResponse {
    pub cloud_watch_logs_log_group_arn: Option<String>,
    pub cloud_watch_logs_role_arn: Option<String>,
    pub include_global_service_events: Option<Boolean>,
    pub is_multi_region_trail: Option<Boolean>,
    pub kms_key_id: Option<String>,
    pub log_file_validation_enabled: Option<Boolean>,
    pub name: Option<String>,
    pub s_3_bucket_name: Option<String>,
    pub s_3_key_prefix: Option<String>,
    pub sns_topic_a_r_n: Option<String>,
    pub sns_topic_name: Option<String>,
    pub trail_a_r_n: Option<String>,
}

Returns the objects or data listed below if successful. Otherwise, returns an error.

Fields
cloud_watch_logs_log_group_arn: Option<String>

Specifies the Amazon Resource Name (ARN) of the log group to which CloudTrail logs will be delivered.

cloud_watch_logs_role_arn: Option<String>

Specifies the role for the CloudWatch Logs endpoint to assume to write to a user's log group.

include_global_service_events: Option<Boolean>

Specifies whether the trail is publishing events from global services such as IAM to the log files.

is_multi_region_trail: Option<Boolean>

Specifies whether the trail exists in one region or in all regions.

kms_key_id: Option<String>

Specifies the KMS key ID that encrypts the logs delivered by CloudTrail. The value is a fully specified ARN to a KMS key in the format:

arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

log_file_validation_enabled: Option<Boolean>

Specifies whether log file integrity validation is enabled.

name: Option<String>

Specifies the name of the trail.

s_3_bucket_name: Option<String>

Specifies the name of the Amazon S3 bucket designated for publishing log files.

s_3_key_prefix: Option<String>

Specifies the Amazon S3 key prefix that comes after the name of the bucket you have designated for log file delivery. For more information, see Finding Your CloudTrail Log Files.

sns_topic_a_r_n: Option<String>

Specifies the ARN of the Amazon SNS topic that CloudTrail uses to send notifications when log files are delivered. The format of a topic ARN is:

arn:aws:sns:us-east-1:123456789012:MyTopic

sns_topic_name: Option<String>

Specifies the name of the Amazon SNS topic that CloudTrail uses to send notifications when log files are delivered.

trail_a_r_n: Option<String>

Specifies the ARN of the trail that was created. The format of a trail ARN is:

arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail