Struct rusoto::cloudwatch::GetMetricStatisticsInput

pub struct GetMetricStatisticsInput {
    pub dimensions: Option<Dimensions>,
    pub end_time: Timestamp,
    pub metric_name: MetricName,
    pub namespace: Namespace,
    pub period: Period,
    pub start_time: Timestamp,
    pub statistics: Statistics,
    pub unit: Option<StandardUnit>,
}

Describes the inputs for GetMetricStatistics.

Fields
dimensions: Option<Dimensions>

A list of dimensions describing qualities of the metric.

end_time: Timestamp

The time stamp to use for determining the last datapoint to return. The value specified is exclusive; results will include datapoints up to the time stamp specified. The time stamp must be in ISO 8601 UTC format (e.g., 2014-09-03T23:00:00Z).

metric_name: MetricName

The name of the metric, with or without spaces.

namespace: Namespace

The namespace of the metric, with or without spaces.

period: Period

The granularity, in seconds, of the returned datapoints. A Period can be as short as one minute (60 seconds) or as long as one day (86,400 seconds), and must be a multiple of 60. The default value is 60.

start_time: Timestamp

The time stamp to use for determining the first datapoint to return. The value specified is inclusive; results include datapoints with the time stamp specified. The time stamp must be in ISO 8601 UTC format (e.g., 2014-09-03T23:00:00Z).

The specified start time is rounded down to the nearest value. Datapoints are returned for start times up to two weeks in the past. Specified start times that are more than two weeks in the past will not return datapoints for metrics that are older than two weeks.

Data that is timestamped 24 hours or more in the past may take in excess of 48 hours to become available from submission time using GetMetricStatistics.

statistics: Statistics

The metric statistics to return. For information about specific statistics returned by GetMetricStatistics, see Statistics in the Amazon CloudWatch Developer Guide.

unit: Option<StandardUnit>

The specific unit for a given metric. Metrics may be reported in multiple units. Not supplying a unit results in all units being returned. If the metric only ever reports one unit, specifying a unit will have no effect.