Struct rusoto::cloudwatch::MetricAlarm

pub struct MetricAlarm {
    pub actions_enabled: Option<ActionsEnabled>,
    pub alarm_actions: Option<ResourceList>,
    pub alarm_arn: Option<AlarmArn>,
    pub alarm_configuration_updated_timestamp: Option<Timestamp>,
    pub alarm_description: Option<AlarmDescription>,
    pub alarm_name: Option<AlarmName>,
    pub comparison_operator: Option<ComparisonOperator>,
    pub dimensions: Option<Dimensions>,
    pub evaluation_periods: Option<EvaluationPeriods>,
    pub extended_statistic: Option<ExtendedStatistic>,
    pub insufficient_data_actions: Option<ResourceList>,
    pub metric_name: Option<MetricName>,
    pub namespace: Option<Namespace>,
    pub ok_actions: Option<ResourceList>,
    pub period: Option<Period>,
    pub state_reason: Option<StateReason>,
    pub state_reason_data: Option<StateReasonData>,
    pub state_updated_timestamp: Option<Timestamp>,
    pub state_value: Option<StateValue>,
    pub statistic: Option<Statistic>,
    pub threshold: Option<Threshold>,
    pub unit: Option<StandardUnit>,
}

Represents an alarm.

Fields
actions_enabled: Option<ActionsEnabled>

Indicates whether actions should be executed during any changes to the alarm state.

alarm_actions: Option<ResourceList>

The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN).

alarm_arn: Option<AlarmArn>

The Amazon Resource Name (ARN) of the alarm.

alarm_configuration_updated_timestamp: Option<Timestamp>

The time stamp of the last update to the alarm configuration.

alarm_description: Option<AlarmDescription>

The description of the alarm.

alarm_name: Option<AlarmName>

The name of the alarm.

comparison_operator: Option<ComparisonOperator>

The arithmetic operation to use when comparing the specified statistic and threshold. The specified statistic value is used as the first operand.

dimensions: Option<Dimensions>

The dimensions for the metric associated with the alarm.

evaluation_periods: Option<EvaluationPeriods>

The number of periods over which data is compared to the specified threshold.

extended_statistic: Option<ExtendedStatistic>

The percentile statistic for the metric associated with the alarm. Specify a value between p0.0 and p100.

insufficient_data_actions: Option<ResourceList>

The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN).

metric_name: Option<MetricName>

The name of the metric associated with the alarm.

namespace: Option<Namespace>

The namespace of the metric associated with the alarm.

ok_actions: Option<ResourceList>

The actions to execute when this alarm transitions to the OK state from any other state. Each action is specified as an Amazon Resource Name (ARN).

period: Option<Period>

The period, in seconds, over which the statistic is applied.

state_reason: Option<StateReason>

An explanation for the alarm state, in text format.

state_reason_data: Option<StateReasonData>

An explanation for the alarm state, in JSON format.

state_updated_timestamp: Option<Timestamp>

The time stamp of the last update to the alarm state.

state_value: Option<StateValue>

The state value for the alarm.

statistic: Option<Statistic>

The statistic for the metric associated with the alarm, other than percentile. For percentile statistics, use ExtendedStatistic.

threshold: Option<Threshold>

The value to compare with the specified statistic.

unit: Option<StandardUnit>

The unit of the metric associated with the alarm.