Struct rusoto::cloudwatch::PutMetricAlarmInput

pub struct PutMetricAlarmInput {
    pub actions_enabled: Option<ActionsEnabled>,
    pub alarm_actions: Option<ResourceList>,
    pub alarm_description: Option<AlarmDescription>,
    pub alarm_name: AlarmName,
    pub comparison_operator: ComparisonOperator,
    pub dimensions: Option<Dimensions>,
    pub evaluation_periods: EvaluationPeriods,
    pub extended_statistic: Option<ExtendedStatistic>,
    pub insufficient_data_actions: Option<ResourceList>,
    pub metric_name: MetricName,
    pub namespace: Namespace,
    pub ok_actions: Option<ResourceList>,
    pub period: Period,
    pub statistic: Option<Statistic>,
    pub threshold: Threshold,
    pub unit: Option<StandardUnit>,
}

Fields
actions_enabled: Option<ActionsEnabled>

Indicates whether actions should be executed during any changes to the alarm state.

alarm_actions: Option<ResourceList>

The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN).

Valid Values: arn:aws:automate:region:ec2:stop | arn:aws:automate:region:ec2:terminate | arn:aws:automate:region:ec2:recover

Valid Values (for use with IAM roles): arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Stop/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Terminate/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Reboot/1.0

alarm_description: Option<AlarmDescription>

The description for the alarm.

alarm_name: AlarmName

The name for the alarm. This name must be unique within the AWS account.

comparison_operator: ComparisonOperator

The arithmetic operation to use when comparing the specified statistic and threshold. The specified statistic value is used as the first operand.

dimensions: Option<Dimensions>

The dimensions for the metric associated with the alarm.

evaluation_periods: EvaluationPeriods

The number of periods over which data is compared to the specified threshold.

extended_statistic: Option<ExtendedStatistic>

The percentile statistic for the metric associated with the alarm. Specify a value between p0.0 and p100.

insufficient_data_actions: Option<ResourceList>

The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN).

Valid Values: arn:aws:automate:region:ec2:stop | arn:aws:automate:region:ec2:terminate | arn:aws:automate:region:ec2:recover

Valid Values (for use with IAM roles): arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Stop/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Terminate/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Reboot/1.0

metric_name: MetricName

The name for the metric associated with the alarm.

namespace: Namespace

The namespace for the metric associated with the alarm.

ok_actions: Option<ResourceList>

The actions to execute when this alarm transitions to an OK state from any other state. Each action is specified as an Amazon Resource Name (ARN).

Valid Values: arn:aws:automate:region:ec2:stop | arn:aws:automate:region:ec2:terminate | arn:aws:automate:region:ec2:recover

Valid Values (for use with IAM roles): arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Stop/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Terminate/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Reboot/1.0

period: Period

The period, in seconds, over which the specified statistic is applied.

statistic: Option<Statistic>

The statistic for the metric associated with the alarm, other than percentile. For percentile statistics, use ExtendedStatistic.

threshold: Threshold

The value against which the specified statistic is compared.

unit: Option<StandardUnit>

The unit of measure for the statistic. For example, the units for the Amazon EC2 NetworkIn metric are Bytes because NetworkIn tracks the number of bytes that an instance receives on all network interfaces. You can also specify a unit when you create a custom metric. Units help provide conceptual meaning to your data. Metric data points that specify a unit of measure, such as Percent, are aggregated separately.

If you specify a unit, you must use a unit that is appropriate for the metric. Otherwise, the Amazon CloudWatch alarm can get stuck in the INSUFFICIENT DATA state.