Module rusoto::codecommit

AWS CodeCommit

If you're using the service, you're probably looking for CodeCommitClient.

Structs

BatchGetRepositoriesInput
Represents the input of a batch get repositories operation.

BatchGetRepositoriesOutput
Represents the output of a batch get repositories operation.

BranchInfo
Returns information about a branch.

CodeCommitClient
A client for the CodeCommit API.

Commit
Returns information about a specific commit.

CreateBranchInput
Represents the input of a create branch operation.

CreateRepositoryInput
Represents the input of a create repository operation.

CreateRepositoryOutput
Represents the output of a create repository operation.

DeleteRepositoryInput
Represents the input of a delete repository operation.

DeleteRepositoryOutput
Represents the output of a delete repository operation.

GetBranchInput
Represents the input of a get branch operation.

GetBranchOutput
Represents the output of a get branch operation.

GetCommitInput
Represents the input of a get commit operation.

GetCommitOutput
Represents the output of a get commit operation.

GetRepositoryInput
Represents the input of a get repository operation.

GetRepositoryOutput
Represents the output of a get repository operation.

GetRepositoryTriggersInput
Represents the input of a get repository triggers operation.

GetRepositoryTriggersOutput
Represents the output of a get repository triggers operation.

ListBranchesInput
Represents the input of a list branches operation.

ListBranchesOutput
Represents the output of a list branches operation.

ListRepositoriesInput
Represents the input of a list repositories operation.

ListRepositoriesOutput
Represents the output of a list repositories operation.

PutRepositoryTriggersInput
Represents the input ofa put repository triggers operation.

PutRepositoryTriggersOutput
Represents the output of a put repository triggers operation.

RepositoryMetadata
Information about a repository.

RepositoryNameIdPair
Information about a repository name and ID.

RepositoryTrigger
Information about a trigger for a repository.

RepositoryTriggerExecutionFailure
A trigger failed to run.

TestRepositoryTriggersInput
Represents the input of a test repository triggers operation.

TestRepositoryTriggersOutput
Represents the output of a test repository triggers operation.

UpdateDefaultBranchInput
Represents the input of an update default branch operation.

UpdateRepositoryDescriptionInput
Represents the input of an update repository description operation.

UpdateRepositoryNameInput
Represents the input of an update repository description operation.

UserInfo
Information about the user who made a specified commit.

Enums

BatchGetRepositoriesError
Errors returned by BatchGetRepositories

CreateBranchError
Errors returned by CreateBranch

CreateRepositoryError
Errors returned by CreateRepository

DeleteRepositoryError
Errors returned by DeleteRepository

GetBranchError
Errors returned by GetBranch

GetCommitError
Errors returned by GetCommit

GetRepositoryError
Errors returned by GetRepository

GetRepositoryTriggersError
Errors returned by GetRepositoryTriggers

ListBranchesError
Errors returned by ListBranches

ListRepositoriesError
Errors returned by ListRepositories

PutRepositoryTriggersError
Errors returned by PutRepositoryTriggers

TestRepositoryTriggersError
Errors returned by TestRepositoryTriggers

UpdateDefaultBranchError
Errors returned by UpdateDefaultBranch

UpdateRepositoryDescriptionError
Errors returned by UpdateRepositoryDescription

UpdateRepositoryNameError
Errors returned by UpdateRepositoryName

Type Definitions

AccountId

AdditionalData

Arn

BranchName

BranchNameList

CloneUrlHttp

CloneUrlSsh

CommitId

CreationDate

Date

Email

LastModifiedDate

Message

Name

NextToken

ObjectId

OrderEnum

ParentList

RepositoryDescription

RepositoryId

RepositoryMetadataList

RepositoryName

RepositoryNameIdPairList

RepositoryNameList

RepositoryNotFoundList

RepositoryTriggerCustomData

RepositoryTriggerEventEnum

RepositoryTriggerEventList

RepositoryTriggerExecutionFailureList

RepositoryTriggerExecutionFailureMessage

RepositoryTriggerName

RepositoryTriggerNameList

RepositoryTriggersConfigurationId

RepositoryTriggersList

SortByEnum