Struct rusoto::codecommit::CodeCommitClient

pub struct CodeCommitClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the CodeCommit API.

Methods

impl<P> CodeCommitClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> CodeCommitClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn batch_get_repositories(&self, input: &BatchGetRepositoriesInput) -> Result<BatchGetRepositoriesOutput, BatchGetRepositoriesError>

Returns information about one or more repositories.

The description field for a repository accepts all HTML characters and all valid Unicode characters. Applications that do not HTML-encode the description and display it in a web page could expose users to potentially malicious code. Make sure that you HTML-encode the description field in any application that uses this API to display the repository description on a web page.

fn create_branch(&self, input: &CreateBranchInput) -> Result<(), CreateBranchError>

Creates a new branch in a repository and points the branch to a commit.

Calling the create branch operation does not set a repository's default branch. To do this, call the update default branch operation.

fn create_repository(&self, input: &CreateRepositoryInput) -> Result<CreateRepositoryOutput, CreateRepositoryError>

Creates a new, empty repository.

fn delete_repository(&self, input: &DeleteRepositoryInput) -> Result<DeleteRepositoryOutput, DeleteRepositoryError>

Deletes a repository. If a specified repository was already deleted, a null repository ID will be returned.
Deleting a repository also deletes all associated objects and metadata. After a repository is deleted, all future push calls to the deleted repository will fail.

fn get_branch(&self, input: &GetBranchInput) -> Result<GetBranchOutput, GetBranchError>

Returns information about a repository branch, including its name and the last commit ID.

fn get_commit(&self, input: &GetCommitInput) -> Result<GetCommitOutput, GetCommitError>

Returns information about a commit, including commit message and committer information.

fn get_repository(&self, input: &GetRepositoryInput) -> Result<GetRepositoryOutput, GetRepositoryError>

Returns information about a repository.

The description field for a repository accepts all HTML characters and all valid Unicode characters. Applications that do not HTML-encode the description and display it in a web page could expose users to potentially malicious code. Make sure that you HTML-encode the description field in any application that uses this API to display the repository description on a web page.

fn get_repository_triggers(&self, input: &GetRepositoryTriggersInput) -> Result<GetRepositoryTriggersOutput, GetRepositoryTriggersError>

Gets information about triggers configured for a repository.

fn list_branches(&self, input: &ListBranchesInput) -> Result<ListBranchesOutput, ListBranchesError>

Gets information about one or more branches in a repository.

fn list_repositories(&self, input: &ListRepositoriesInput) -> Result<ListRepositoriesOutput, ListRepositoriesError>

Gets information about one or more repositories.

fn put_repository_triggers(&self, input: &PutRepositoryTriggersInput) -> Result<PutRepositoryTriggersOutput, PutRepositoryTriggersError>

Replaces all triggers for a repository. This can be used to create or delete triggers.

fn test_repository_triggers(&self, input: &TestRepositoryTriggersInput) -> Result<TestRepositoryTriggersOutput, TestRepositoryTriggersError>

Tests the functionality of repository triggers by sending information to the trigger target. If real data is available in the repository, the test will send data from the last commit. If no data is available, sample data will be generated.

fn update_default_branch(&self, input: &UpdateDefaultBranchInput) -> Result<(), UpdateDefaultBranchError>

Sets or changes the default branch name for the specified repository.

If you use this operation to change the default branch name to the current default branch name, a success message is returned even though the default branch did not change.

fn update_repository_description(&self, input: &UpdateRepositoryDescriptionInput) -> Result<(), UpdateRepositoryDescriptionError>

Sets or changes the comment or description for a repository.

The description field for a repository accepts all HTML characters and all valid Unicode characters. Applications that do not HTML-encode the description and display it in a web page could expose users to potentially malicious code. Make sure that you HTML-encode the description field in any application that uses this API to display the repository description on a web page.

fn update_repository_name(&self, input: &UpdateRepositoryNameInput) -> Result<(), UpdateRepositoryNameError>

Renames a repository. The repository name must be unique across the calling AWS account. In addition, repository names are limited to 100 alphanumeric, dash, and underscore characters, and cannot include certain characters. The suffix ".git" is prohibited. For a full description of the limits on repository names, see Limits in the AWS CodeCommit User Guide.