Module rusoto::codedeploy

AWS CodeDeploy

Structs

AddTagsToOnPremisesInstancesInput
Represents the input of, and adds tags to, an on-premises instance operation.

ApplicationInfo
Information about an application.

AutoScalingGroup
Information about an Auto Scaling group.

BatchGetApplicationRevisionsInput
Represents the input of a batch get application revisions operation.

BatchGetApplicationRevisionsOutput
Represents the output of a batch get application revisions operation.

BatchGetApplicationsInput
Represents the input of a batch get applications operation.

BatchGetApplicationsOutput
Represents the output of a batch get applications operation.

BatchGetDeploymentGroupsInput
Represents the input of a batch get deployment groups operation.

BatchGetDeploymentGroupsOutput
Represents the output of a batch get deployment groups operation.

BatchGetDeploymentInstancesInput
Represents the input of a batch get deployment instances operation.

BatchGetDeploymentInstancesOutput
Represents the output of a batch get deployment instance operation.

BatchGetDeploymentsInput
Represents the input of a batch get deployments operation.

BatchGetDeploymentsOutput
Represents the output of a batch get deployments operation.

BatchGetOnPremisesInstancesInput
Represents the input of a batch get on-premises instances operation.

BatchGetOnPremisesInstancesOutput
Represents the output of a batch get on-premises instances operation.

CodeDeployClient
A client for the CodeDeploy API.

CreateApplicationInput
Represents the input of a create application operation.

CreateApplicationOutput
Represents the output of a create application operation.

CreateDeploymentConfigInput
Represents the input of a create deployment configuration operation.

CreateDeploymentConfigOutput
Represents the output of a create deployment configuration operation.

CreateDeploymentGroupInput
Represents the input of a create deployment group operation.

CreateDeploymentGroupOutput
Represents the output of a create deployment group operation.

CreateDeploymentInput
Represents the input of a create deployment operation.

CreateDeploymentOutput
Represents the output of a create deployment operation.

DeleteApplicationInput
Represents the input of a delete application operation.

DeleteDeploymentConfigInput
Represents the input of a delete deployment configuration operation.

DeleteDeploymentGroupInput
Represents the input of a delete deployment group operation.

DeleteDeploymentGroupOutput
Represents the output of a delete deployment group operation.

DeploymentConfigInfo
Information about a deployment configuration.

DeploymentGroupInfo
Information about a deployment group.

DeploymentInfo
Information about a deployment.

DeploymentOverview
Information about the deployment status of the instances in the deployment.

DeregisterOnPremisesInstanceInput
Represents the input of a deregister on-premises instance operation.

Diagnostics
Diagnostic information about executable scripts that are part of a deployment.

EC2TagFilter
Information about a tag filter.

ErrorInformation
Information about a deployment error.

GenericRevisionInfo
Information about an application revision.

GetApplicationInput
Represents the input of a get application operation.

GetApplicationOutput
Represents the output of a get application operation.

GetApplicationRevisionInput
Represents the input of a get application revision operation.

GetApplicationRevisionOutput
Represents the output of a get application revision operation.

GetDeploymentConfigInput
Represents the input of a get deployment configuration operation.

GetDeploymentConfigOutput
Represents the output of a get deployment configuration operation.

GetDeploymentGroupInput
Represents the input of a get deployment group operation.

GetDeploymentGroupOutput
Represents the output of a get deployment group operation.

GetDeploymentInput
Represents the input of a get deployment operation.

GetDeploymentInstanceInput
Represents the input of a get deployment instance operation.

GetDeploymentInstanceOutput
Represents the output of a get deployment instance operation.

GetDeploymentOutput
Represents the output of a get deployment operation.

GetOnPremisesInstanceInput
Represents the input of a get on-premises instance operation.

GetOnPremisesInstanceOutput
Represents the output of a get on-premises instance operation.

GitHubLocation
Information about the location of application artifacts stored in GitHub.

InstanceInfo
Information about an on-premises instance.

InstanceSummary
Information about an instance in a deployment.

LifecycleEvent
Information about a deployment lifecycle event.

ListApplicationRevisionsInput
Represents the input of a list application revisions operation.

ListApplicationRevisionsOutput
Represents the output of a list application revisions operation.

ListApplicationsInput
Represents the input of a list applications operation.

ListApplicationsOutput
Represents the output of a list applications operation.

ListDeploymentConfigsInput
Represents the input of a list deployment configurations operation.

ListDeploymentConfigsOutput
Represents the output of a list deployment configurations operation.

ListDeploymentGroupsInput
Represents the input of a list deployment groups operation.

ListDeploymentGroupsOutput
Represents the output of a list deployment groups operation.

ListDeploymentInstancesInput
Represents the input of a list deployment instances operation.

ListDeploymentInstancesOutput
Represents the output of a list deployment instances operation.

ListDeploymentsInput
Represents the input of a list deployments operation.

ListDeploymentsOutput
Represents the output of a list deployments operation.

ListOnPremisesInstancesInput

Represents the input of a list on-premises instances operation.
.

ListOnPremisesInstancesOutput
Represents the output of list on-premises instances operation.

MinimumHealthyHosts
Information about minimum healthy instance.

RegisterApplicationRevisionInput
Represents the input of a register application revision operation.

RegisterOnPremisesInstanceInput
Represents the input of the register on-premises instance operation.

RemoveTagsFromOnPremisesInstancesInput
Represents the input of a remove tags from on-premises instances operation.

RevisionInfo
Information about an application revision.

RevisionLocation
Information about the location of an application revision.

S3Location
Information about the location of application artifacts stored in Amazon S3.

StopDeploymentInput
Represents the input of a stop deployment operation.

StopDeploymentOutput
Represents the output of a stop deployment operation.

Tag
Information about a tag.

TagFilter
Information about an on-premises instance tag filter.

TimeRange
Information about a time range.

TriggerConfig
Information about notification triggers for the deployment group.

UpdateApplicationInput
Represents the input of an update application operation.

UpdateDeploymentGroupInput
Represents the input of an update deployment group operation.

UpdateDeploymentGroupOutput
Represents the output of an update deployment group operation.

Enums

AddTagsToOnPremisesInstancesError

BatchGetApplicationRevisionsError

BatchGetApplicationsError

BatchGetDeploymentGroupsError

BatchGetDeploymentInstancesError

BatchGetDeploymentsError

BatchGetOnPremisesInstancesError

CreateApplicationError

CreateDeploymentConfigError

CreateDeploymentError

CreateDeploymentGroupError

DeleteApplicationError

DeleteDeploymentConfigError

DeleteDeploymentGroupError

DeregisterOnPremisesInstanceError

GetApplicationError

GetApplicationRevisionError

GetDeploymentConfigError

GetDeploymentError

GetDeploymentGroupError

GetDeploymentInstanceError

GetOnPremisesInstanceError

ListApplicationRevisionsError

ListApplicationsError

ListDeploymentConfigsError

ListDeploymentGroupsError

ListDeploymentInstancesError

ListDeploymentsError

ListOnPremisesInstancesError

RegisterApplicationRevisionError

RegisterOnPremisesInstanceError

RemoveTagsFromOnPremisesInstancesError

StopDeploymentError

UpdateApplicationError

UpdateDeploymentGroupError

Type Definitions

ApplicationId

ApplicationName

ApplicationRevisionSortBy

ApplicationsInfoList

ApplicationsList

AutoScalingGroupHook

AutoScalingGroupList

AutoScalingGroupName

AutoScalingGroupNameList

Boolean

BundleType

CommitId

DeploymentConfigId

DeploymentConfigName

DeploymentConfigsList

DeploymentCreator

DeploymentGroupId

DeploymentGroupInfoList

DeploymentGroupName

DeploymentGroupsList

DeploymentId

DeploymentStatus

DeploymentStatusList

DeploymentsInfoList

DeploymentsList

Description

EC2TagFilterList

EC2TagFilterType

ETag

ErrorCode

ErrorMessage

IamUserArn

InstanceArn

InstanceCount

InstanceId

InstanceInfoList

InstanceName

InstanceNameList

InstanceStatus

InstanceStatusList

InstanceSummaryList

InstancesList

Key

LifecycleErrorCode

LifecycleEventList

LifecycleEventName

LifecycleEventStatus

LifecycleMessage

ListStateFilterAction

LogTail

Message

MinimumHealthyHostsType

MinimumHealthyHostsValue

NextToken

RegistrationStatus

Repository

RevisionInfoList

RevisionLocationList

RevisionLocationType

Role

S3Bucket

S3Key

ScriptName

SortOrder

StopStatus

TagFilterList

TagFilterType

TagList

Timestamp

TriggerConfigList

TriggerEventType

TriggerEventTypeList

TriggerName

TriggerTargetArn

Value

VersionId