Enum rusoto::codedeploy::CreateDeploymentGroupError

pub enum CreateDeploymentGroupError {
    ApplicationNameRequired(String),
    InvalidApplicationName(String),
    ApplicationDoesNotExist(String),
    DeploymentGroupNameRequired(String),
    InvalidDeploymentGroupName(String),
    DeploymentGroupAlreadyExists(String),
    InvalidEC2Tag(String),
    InvalidTag(String),
    InvalidAutoScalingGroup(String),
    InvalidDeploymentConfigName(String),
    DeploymentConfigDoesNotExist(String),
    RoleRequired(String),
    InvalidRole(String),
    DeploymentGroupLimitExceeded(String),
    LifecycleHookLimitExceeded(String),
    InvalidTriggerConfig(String),
    TriggerTargetsLimitExceeded(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Variants
ApplicationNameRequired(String)

The minimum number of required application names was not specified.

InvalidApplicationName(String)

The application name was specified in an invalid format.

ApplicationDoesNotExist(String)

The application does not exist with the applicable IAM user or AWS account.

DeploymentGroupNameRequired(String)

The deployment group name was not specified.

InvalidDeploymentGroupName(String)

The deployment group name was specified in an invalid format.

DeploymentGroupAlreadyExists(String)

A deployment group with the specified name already exists with the applicable IAM user or AWS account.

InvalidEC2Tag(String)

The tag was specified in an invalid format.

InvalidTag(String)

The specified tag was specified in an invalid format.

InvalidAutoScalingGroup(String)

The Auto Scaling group was specified in an invalid format or does not exist.

InvalidDeploymentConfigName(String)

The deployment configuration name was specified in an invalid format.

DeploymentConfigDoesNotExist(String)

The deployment configuration does not exist with the applicable IAM user or AWS account.

RoleRequired(String)

The role ID was not specified.

InvalidRole(String)

The service role ARN was specified in an invalid format. Or, if an Auto Scaling group was specified, the specified service role does not grant the appropriate permissions to Auto Scaling.

DeploymentGroupLimitExceeded(String)

The deployment groups limit was exceeded.

LifecycleHookLimitExceeded(String)

The limit for lifecycle hooks was exceeded.

InvalidTriggerConfig(String)

The trigger was specified in an invalid format.

TriggerTargetsLimitExceeded(String)

The maximum allowed number of triggers was exceeded.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl CreateDeploymentGroupError

fn from_body(body: &str) -> CreateDeploymentGroupError