Struct rusoto::codedeploy::CodeDeployClient

pub struct CodeDeployClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the CodeDeploy API.

Methods

impl<P> CodeDeployClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> CodeDeployClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn add_tags_to_on_premises_instances(&self, input: &AddTagsToOnPremisesInstancesInput) -> Result<(), AddTagsToOnPremisesInstancesError>

Adds tags to on-premises instances.

fn batch_get_application_revisions(&self, input: &BatchGetApplicationRevisionsInput) -> Result<BatchGetApplicationRevisionsOutput, BatchGetApplicationRevisionsError>

Gets information about one or more application revisions.

fn batch_get_applications(&self, input: &BatchGetApplicationsInput) -> Result<BatchGetApplicationsOutput, BatchGetApplicationsError>

Gets information about one or more applications.

fn batch_get_deployment_groups(&self, input: &BatchGetDeploymentGroupsInput) -> Result<BatchGetDeploymentGroupsOutput, BatchGetDeploymentGroupsError>

Get information about one or more deployment groups.

fn batch_get_deployment_instances(&self, input: &BatchGetDeploymentInstancesInput) -> Result<BatchGetDeploymentInstancesOutput, BatchGetDeploymentInstancesError>

Gets information about one or more instance that are part of a deployment group.

fn batch_get_deployments(&self, input: &BatchGetDeploymentsInput) -> Result<BatchGetDeploymentsOutput, BatchGetDeploymentsError>

Gets information about one or more deployments.

fn batch_get_on_premises_instances(&self, input: &BatchGetOnPremisesInstancesInput) -> Result<BatchGetOnPremisesInstancesOutput, BatchGetOnPremisesInstancesError>

Gets information about one or more on-premises instances.

fn create_application(&self, input: &CreateApplicationInput) -> Result<CreateApplicationOutput, CreateApplicationError>

Creates an application.

fn create_deployment(&self, input: &CreateDeploymentInput) -> Result<CreateDeploymentOutput, CreateDeploymentError>

Deploys an application revision through the specified deployment group.

fn create_deployment_config(&self, input: &CreateDeploymentConfigInput) -> Result<CreateDeploymentConfigOutput, CreateDeploymentConfigError>

Creates a deployment configuration.

fn create_deployment_group(&self, input: &CreateDeploymentGroupInput) -> Result<CreateDeploymentGroupOutput, CreateDeploymentGroupError>

Creates a deployment group to which application revisions will be deployed.

fn delete_application(&self, input: &DeleteApplicationInput) -> Result<(), DeleteApplicationError>

Deletes an application.

fn delete_deployment_config(&self, input: &DeleteDeploymentConfigInput) -> Result<(), DeleteDeploymentConfigError>

Deletes a deployment configuration.
A deployment configuration cannot be deleted if it is currently in use. Predefined configurations cannot be deleted.

fn delete_deployment_group(&self, input: &DeleteDeploymentGroupInput) -> Result<DeleteDeploymentGroupOutput, DeleteDeploymentGroupError>

Deletes a deployment group.

fn deregister_on_premises_instance(&self, input: &DeregisterOnPremisesInstanceInput) -> Result<(), DeregisterOnPremisesInstanceError>

Deregisters an on-premises instance.

fn get_application(&self, input: &GetApplicationInput) -> Result<GetApplicationOutput, GetApplicationError>

Gets information about an application.

fn get_application_revision(&self, input: &GetApplicationRevisionInput) -> Result<GetApplicationRevisionOutput, GetApplicationRevisionError>

Gets information about an application revision.

fn get_deployment(&self, input: &GetDeploymentInput) -> Result<GetDeploymentOutput, GetDeploymentError>

Gets information about a deployment.

fn get_deployment_config(&self, input: &GetDeploymentConfigInput) -> Result<GetDeploymentConfigOutput, GetDeploymentConfigError>

Gets information about a deployment configuration.

fn get_deployment_group(&self, input: &GetDeploymentGroupInput) -> Result<GetDeploymentGroupOutput, GetDeploymentGroupError>

Gets information about a deployment group.

fn get_deployment_instance(&self, input: &GetDeploymentInstanceInput) -> Result<GetDeploymentInstanceOutput, GetDeploymentInstanceError>

Gets information about an instance as part of a deployment.

fn get_on_premises_instance(&self, input: &GetOnPremisesInstanceInput) -> Result<GetOnPremisesInstanceOutput, GetOnPremisesInstanceError>

Gets information about an on-premises instance.

fn list_application_revisions(&self, input: &ListApplicationRevisionsInput) -> Result<ListApplicationRevisionsOutput, ListApplicationRevisionsError>

Lists information about revisions for an application.

fn list_applications(&self, input: &ListApplicationsInput) -> Result<ListApplicationsOutput, ListApplicationsError>

Lists the applications registered with the applicable IAM user or AWS account.

fn list_deployment_configs(&self, input: &ListDeploymentConfigsInput) -> Result<ListDeploymentConfigsOutput, ListDeploymentConfigsError>

Lists the deployment configurations with the applicable IAM user or AWS account.

fn list_deployment_groups(&self, input: &ListDeploymentGroupsInput) -> Result<ListDeploymentGroupsOutput, ListDeploymentGroupsError>

Lists the deployment groups for an application registered with the applicable IAM user or AWS account.

fn list_deployment_instances(&self, input: &ListDeploymentInstancesInput) -> Result<ListDeploymentInstancesOutput, ListDeploymentInstancesError>

Lists the instance for a deployment associated with the applicable IAM user or AWS account.

fn list_deployments(&self, input: &ListDeploymentsInput) -> Result<ListDeploymentsOutput, ListDeploymentsError>

Lists the deployments in a deployment group for an application registered with the applicable IAM user or AWS account.

fn list_on_premises_instances(&self, input: &ListOnPremisesInstancesInput) -> Result<ListOnPremisesInstancesOutput, ListOnPremisesInstancesError>

Gets a list of names for one or more on-premises instances.

Unless otherwise specified, both registered and deregistered on-premises instance names will be listed. To list only registered or deregistered on-premises instance names, use the registration status parameter.

fn register_application_revision(&self, input: &RegisterApplicationRevisionInput) -> Result<(), RegisterApplicationRevisionError>

Registers with AWS CodeDeploy a revision for the specified application.

fn register_on_premises_instance(&self, input: &RegisterOnPremisesInstanceInput) -> Result<(), RegisterOnPremisesInstanceError>

Registers an on-premises instance.

fn remove_tags_from_on_premises_instances(&self, input: &RemoveTagsFromOnPremisesInstancesInput) -> Result<(), RemoveTagsFromOnPremisesInstancesError>

Removes one or more tags from one or more on-premises instances.

fn stop_deployment(&self, input: &StopDeploymentInput) -> Result<StopDeploymentOutput, StopDeploymentError>

Attempts to stop an ongoing deployment.

fn update_application(&self, input: &UpdateApplicationInput) -> Result<(), UpdateApplicationError>

Changes the name of an application.

fn update_deployment_group(&self, input: &UpdateDeploymentGroupInput) -> Result<UpdateDeploymentGroupOutput, UpdateDeploymentGroupError>

Changes information about a deployment group.