Struct rusoto::codedeploy::CreateDeploymentGroupInput

pub struct CreateDeploymentGroupInput {
    pub application_name: ApplicationName,
    pub auto_scaling_groups: Option<AutoScalingGroupNameList>,
    pub deployment_config_name: Option<DeploymentConfigName>,
    pub deployment_group_name: DeploymentGroupName,
    pub ec_2_tag_filters: Option<EC2TagFilterList>,
    pub on_premises_instance_tag_filters: Option<TagFilterList>,
    pub service_role_arn: Role,
    pub trigger_configurations: Option<TriggerConfigList>,
}

Represents the input of a create deployment group operation.

Fields
application_name: ApplicationName

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

auto_scaling_groups: Option<AutoScalingGroupNameList>

A list of associated Auto Scaling groups.

deployment_config_name: Option<DeploymentConfigName>

If specified, the deployment configuration name can be either one of the predefined configurations provided with AWS CodeDeploy or a custom deployment configuration that you create by calling the create deployment configuration operation.

CodeDeployDefault.OneAtATime is the default deployment configuration. It is used if a configuration isn't specified for the deployment or the deployment group.

The predefined deployment configurations include the following:

CodeDeployDefault.AllAtOnce attempts to deploy an application revision to as many instance as possible at once. The status of the overall deployment will be displayed as Succeeded if the application revision is deployed to one or more of the instances. The status of the overall deployment will be displayed as Failed if the application revision is not deployed to any of the instances. Using an example of nine instance, CodeDeployDefault.AllAtOnce will attempt to deploy to all nine instance at once. The overall deployment will succeed if deployment to even a single instance is successful; it will fail only if deployments to all nine instance fail.

CodeDeployDefault.HalfAtATime deploys to up to half of the instances at a time (with fractions rounded down). The overall deployment succeeds if the application revision is deployed to at least half of the instances (with fractions rounded up); otherwise, the deployment fails. In the example of nine instances, it will deploy to up to four instance at a time. The overall deployment succeeds if deployment to five or more instances succeed; otherwise, the deployment fails. The deployment may be successfully deployed to some instances even if the overall deployment fails.

CodeDeployDefault.OneAtATime deploys the application revision to only one instance at a time.

For deployment groups that contain more than one instance:

The overall deployment succeeds if the application revision is deployed to all of the instances. The exception to this rule is if deployment to the last instance fails, the overall deployment still succeeds. This is because AWS CodeDeploy allows only one instance at a time to be taken offline with the CodeDeployDefault.OneAtATime configuration.

The overall deployment fails as soon as the application revision fails to be deployed to any but the last instance. The deployment may be successfully deployed to some instances even if the overall deployment fails.

In an example using nine instance, it will deploy to one instance at a time. The overall deployment succeeds if deployment to the first eight instance is successful; the overall deployment fails if deployment to any of the first eight instance fails.

For deployment groups that contain only one instance, the overall deployment is successful only if deployment to the single instance is successful

deployment_group_name: DeploymentGroupName

The name of a new deployment group for the specified application.

ec_2_tag_filters: Option<EC2TagFilterList>

The Amazon EC2 tags on which to filter.

on_premises_instance_tag_filters: Option<TagFilterList>

The on-premises instance tags on which to filter.

service_role_arn: Role

A service role ARN that allows AWS CodeDeploy to act on the user's behalf when interacting with AWS services.

trigger_configurations: Option<TriggerConfigList>

Information about triggers to create when the deployment group is created.