Struct rusoto::codedeploy::DeploymentInfo

pub struct DeploymentInfo {
    pub application_name: Option<ApplicationName>,
    pub complete_time: Option<Timestamp>,
    pub create_time: Option<Timestamp>,
    pub creator: Option<DeploymentCreator>,
    pub deployment_config_name: Option<DeploymentConfigName>,
    pub deployment_group_name: Option<DeploymentGroupName>,
    pub deployment_id: Option<DeploymentId>,
    pub deployment_overview: Option<DeploymentOverview>,
    pub description: Option<Description>,
    pub error_information: Option<ErrorInformation>,
    pub ignore_application_stop_failures: Option<Boolean>,
    pub revision: Option<RevisionLocation>,
    pub start_time: Option<Timestamp>,
    pub status: Option<DeploymentStatus>,
}

Information about a deployment.

Fields
application_name: Option<ApplicationName>

The application name.

complete_time: Option<Timestamp>

A timestamp indicating when the deployment was complete.

create_time: Option<Timestamp>

A timestamp indicating when the deployment was created.

creator: Option<DeploymentCreator>

The means by which the deployment was created:

user: A user created the deployment.

autoscaling: Auto Scaling created the deployment.

deployment_config_name: Option<DeploymentConfigName>

The deployment configuration name.

deployment_group_name: Option<DeploymentGroupName>

The deployment group name.

deployment_id: Option<DeploymentId>

The deployment ID.

deployment_overview: Option<DeploymentOverview>

A summary of the deployment status of the instances in the deployment.

description: Option<Description>

A comment about the deployment.

error_information: Option<ErrorInformation>

Information about any error associated with this deployment.

ignore_application_stop_failures: Option<Boolean>

If true, then if the deployment causes the ApplicationStop deployment lifecycle event to an instance to fail, the deployment to that instance will not be considered to have failed at that point and will continue on to the BeforeInstall deployment lifecycle event.

If false or not specified, then if the deployment causes the ApplicationStop deployment lifecycle event to an instance to fail, the deployment to that instance will stop, and the deployment to that instance will be considered to have failed.

revision: Option<RevisionLocation>

Information about the location of stored application artifacts and the service from which to retrieve them.

start_time: Option<Timestamp>

A timestamp indicating when the deployment was deployed to the deployment group.

In some cases, the reported value of the start time may be later than the complete time. This is due to differences in the clock settings of back-end servers that participate in the deployment process.

status: Option<DeploymentStatus>

The current state of the deployment as a whole.