Struct rusoto::codedeploy::ErrorInformation

pub struct ErrorInformation {
    pub code: Option<ErrorCode>,
    pub message: Option<ErrorMessage>,
}

Information about a deployment error.

Fields
code: Option<ErrorCode>

The error code:

APPLICATION_MISSING: The application was missing. This error code will most likely be raised if the application is deleted after the deployment is created but before it is started.

DEPLOYMENT_GROUP_MISSING: The deployment group was missing. This error code will most likely be raised if the deployment group is deleted after the deployment is created but before it is started.

HEALTH_CONSTRAINTS: The deployment failed on too many instances to be successfully deployed within the instance health constraints specified.

HEALTH_CONSTRAINTS_INVALID: The revision cannot be successfully deployed within the instance health constraints specified.

IAM_ROLE_MISSING: The service role cannot be accessed.

IAM_ROLE_PERMISSIONS: The service role does not have the correct permissions.

INTERNAL_ERROR: There was an internal error.

NO_EC2_SUBSCRIPTION: The calling account is not subscribed to the Amazon EC2 service.

NO_INSTANCES: No instance were specified, or no instance can be found.

OVER_MAX_INSTANCES: The maximum number of instance was exceeded.

THROTTLED: The operation was throttled because the calling account exceeded the throttling limits of one or more AWS services.

TIMEOUT: The deployment has timed out.

REVISION_MISSING: The revision ID was missing. This error code will most likely be raised if the revision is deleted after the deployment is created but before it is started.

message: Option<ErrorMessage>

An accompanying error message.