Struct rusoto::codedeploy::ListApplicationRevisionsInput

pub struct ListApplicationRevisionsInput {
    pub application_name: ApplicationName,
    pub deployed: Option<ListStateFilterAction>,
    pub next_token: Option<NextToken>,
    pub s_3_bucket: Option<S3Bucket>,
    pub s_3_key_prefix: Option<S3Key>,
    pub sort_by: Option<ApplicationRevisionSortBy>,
    pub sort_order: Option<SortOrder>,
}

Represents the input of a list application revisions operation.

Fields
application_name: ApplicationName

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

deployed: Option<ListStateFilterAction>

Whether to list revisions based on whether the revision is the target revision of an deployment group:

include: List revisions that are target revisions of a deployment group.

exclude: Do not list revisions that are target revisions of a deployment group.

ignore: List all revisions.

next_token: Option<NextToken>

An identifier returned from the previous list application revisions call. It can be used to return the next set of applications in the list.

s_3_bucket: Option<S3Bucket>

An Amazon S3 bucket name to limit the search for revisions.

If set to null, all of the user's buckets will be searched.

s_3_key_prefix: Option<S3Key>

A key prefix for the set of Amazon S3 objects to limit the search for revisions.

sort_by: Option<ApplicationRevisionSortBy>

The column name to use to sort the list results:

registerTime: Sort by the time the revisions were registered with AWS CodeDeploy.

firstUsedTime: Sort by the time the revisions were first used in a deployment.

lastUsedTime: Sort by the time the revisions were last used in a deployment.

If not specified or set to null, the results will be returned in an arbitrary order.

sort_order: Option<SortOrder>

The order in which to sort the list results:

ascending: ascending order.

descending: descending order.

If not specified, the results will be sorted in ascending order.

If set to null, the results will be sorted in an arbitrary order.