Struct rusoto::codepipeline::ActionDeclaration

pub struct ActionDeclaration {
    pub action_type_id: ActionTypeId,
    pub configuration: Option<ActionConfigurationMap>,
    pub input_artifacts: Option<InputArtifactList>,
    pub name: ActionName,
    pub output_artifacts: Option<OutputArtifactList>,
    pub role_arn: Option<RoleArn>,
    pub run_order: Option<ActionRunOrder>,
}

Represents information about an action declaration.

Fields
action_type_id: ActionTypeId

The configuration information for the action type.

configuration: Option<ActionConfigurationMap>

The action declaration's configuration.

input_artifacts: Option<InputArtifactList>

The name or ID of the artifact consumed by the action, such as a test or build artifact.

name: ActionName

The action declaration's name.

output_artifacts: Option<OutputArtifactList>

The name or ID of the result of the action declaration, such as a test or build artifact.

role_arn: Option<RoleArn>

The ARN of the IAM service role that will perform the declared action. This is assumed through the roleArn for the pipeline.

run_order: Option<ActionRunOrder>

The order in which actions are run.