Struct rusoto::codepipeline::ActionExecution

pub struct ActionExecution {
    pub error_details: Option<ErrorDetails>,
    pub external_execution_id: Option<ExecutionId>,
    pub external_execution_url: Option<Url>,
    pub last_status_change: Option<Timestamp>,
    pub percent_complete: Option<Percentage>,
    pub status: Option<ActionExecutionStatus>,
    pub summary: Option<ExecutionSummary>,
}

Represents information about how an action runs.

Fields
error_details: Option<ErrorDetails>

The details of an error returned by a URL external to AWS.

external_execution_id: Option<ExecutionId>

The external ID of the run of the action.

external_execution_url: Option<Url>

The URL of a resource external to AWS that will be used when running the action, for example an external repository URL.

last_status_change: Option<Timestamp>

The last status change of the action.

percent_complete: Option<Percentage>

A percentage of completeness of the action as it runs.

status: Option<ActionExecutionStatus>

The status of the action, or for a completed action, the last status of the action.

summary: Option<ExecutionSummary>

A summary of the run of the action.