Struct rusoto::codepipeline::CodePipelineClient

pub struct CodePipelineClient<P, D> where P: ProvideAwsCredentials,
        D: DispatchSignedRequest { /* fields omitted */ }

A client for the CodePipeline API.

Methods

impl<P, D> CodePipelineClient<P, D> where P: ProvideAwsCredentials,
D: DispatchSignedRequest

fn new(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn acknowledge_job(&self,
input: &AcknowledgeJobInput)
-> Result<AcknowledgeJobOutput, AcknowledgeJobError>

Returns information about a specified job and whether that job has been received by the job worker. Only used for custom actions.

fn acknowledge_third_party_job(&self,
input: &AcknowledgeThirdPartyJobInput)
-> Result<AcknowledgeThirdPartyJobOutput, AcknowledgeThirdPartyJobError>

Confirms a job worker has received the specified job. Only used for partner actions.

fn create_custom_action_type(&self,
input: &CreateCustomActionTypeInput)
-> Result<CreateCustomActionTypeOutput, CreateCustomActionTypeError>

Creates a new custom action that can be used in all pipelines associated with the AWS account. Only used for custom actions.

fn create_pipeline(&self,
input: &CreatePipelineInput)
-> Result<CreatePipelineOutput, CreatePipelineError>

Creates a pipeline.

fn delete_custom_action_type(&self,
input: &DeleteCustomActionTypeInput)
-> Result<(), DeleteCustomActionTypeError>

Marks a custom action as deleted. PollForJobs for the custom action will fail after the action is marked for deletion. Only used for custom actions.

You cannot recreate a custom action after it has been deleted unless you increase the version number of the action.

fn delete_pipeline(&self,
input: &DeletePipelineInput)
-> Result<(), DeletePipelineError>

Deletes the specified pipeline.

fn disable_stage_transition(&self,
input: &DisableStageTransitionInput)
-> Result<(), DisableStageTransitionError>

Prevents artifacts in a pipeline from transitioning to the next stage in the pipeline.

fn enable_stage_transition(&self,
input: &EnableStageTransitionInput)
-> Result<(), EnableStageTransitionError>

Enables artifacts in a pipeline to transition to a stage in a pipeline.

fn get_job_details(&self,
input: &GetJobDetailsInput)
-> Result<GetJobDetailsOutput, GetJobDetailsError>

Returns information about a job. Only used for custom actions.

When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts. Additionally, this API returns any secret values defined for the action.

fn get_pipeline(&self,
input: &GetPipelineInput)
-> Result<GetPipelineOutput, GetPipelineError>

Returns the metadata, structure, stages, and actions of a pipeline. Can be used to return the entire structure of a pipeline in JSON format, which can then be modified and used to update the pipeline structure with UpdatePipeline.

fn get_pipeline_execution(&self,
input: &GetPipelineExecutionInput)
-> Result<GetPipelineExecutionOutput, GetPipelineExecutionError>

Returns information about an execution of a pipeline, including details about artifacts, the pipeline execution ID, and the name, version, and status of the pipeline.

fn get_pipeline_state(&self,
input: &GetPipelineStateInput)
-> Result<GetPipelineStateOutput, GetPipelineStateError>

Returns information about the state of a pipeline, including the stages and actions.

fn get_third_party_job_details(&self,
input: &GetThirdPartyJobDetailsInput)
-> Result<GetThirdPartyJobDetailsOutput, GetThirdPartyJobDetailsError>

Requests the details of a job for a third party action. Only used for partner actions.

When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts. Additionally, this API returns any secret values defined for the action.

fn list_action_types(&self,
input: &ListActionTypesInput)
-> Result<ListActionTypesOutput, ListActionTypesError>

Gets a summary of all AWS CodePipeline action types associated with your account.

fn list_pipelines(&self,
input: &ListPipelinesInput)
-> Result<ListPipelinesOutput, ListPipelinesError>

Gets a summary of all of the pipelines associated with your account.

fn poll_for_jobs(&self,
input: &PollForJobsInput)
-> Result<PollForJobsOutput, PollForJobsError>

Returns information about any jobs for AWS CodePipeline to act upon.

When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts. Additionally, this API returns any secret values defined for the action.

fn poll_for_third_party_jobs(&self,
input: &PollForThirdPartyJobsInput)
-> Result<PollForThirdPartyJobsOutput, PollForThirdPartyJobsError>

Determines whether there are any third party jobs for a job worker to act on. Only used for partner actions.

When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts.

fn put_action_revision(&self,
input: &PutActionRevisionInput)
-> Result<PutActionRevisionOutput, PutActionRevisionError>

Provides information to AWS CodePipeline about new revisions to a source.

fn put_approval_result(&self,
input: &PutApprovalResultInput)
-> Result<PutApprovalResultOutput, PutApprovalResultError>

Provides the response to a manual approval request to AWS CodePipeline. Valid responses include Approved and Rejected.

fn put_job_failure_result(&self,
input: &PutJobFailureResultInput)
-> Result<(), PutJobFailureResultError>

Represents the failure of a job as returned to the pipeline by a job worker. Only used for custom actions.

fn put_job_success_result(&self,
input: &PutJobSuccessResultInput)
-> Result<(), PutJobSuccessResultError>

Represents the success of a job as returned to the pipeline by a job worker. Only used for custom actions.

fn put_third_party_job_failure_result(&self,
input: &PutThirdPartyJobFailureResultInput)
-> Result<(), PutThirdPartyJobFailureResultError>

Represents the failure of a third party job as returned to the pipeline by a job worker. Only used for partner actions.

fn put_third_party_job_success_result(&self,
input: &PutThirdPartyJobSuccessResultInput)
-> Result<(), PutThirdPartyJobSuccessResultError>

Represents the success of a third party job as returned to the pipeline by a job worker. Only used for partner actions.

fn retry_stage_execution(&self,
input: &RetryStageExecutionInput)
-> Result<RetryStageExecutionOutput, RetryStageExecutionError>

Resumes the pipeline execution by retrying the last failed actions in a stage.

fn start_pipeline_execution(&self,
input: &StartPipelineExecutionInput)
-> Result<StartPipelineExecutionOutput, StartPipelineExecutionError>

Starts the specified pipeline. Specifically, it begins processing the latest commit to the source location specified as part of the pipeline.

fn update_pipeline(&self,
input: &UpdatePipelineInput)
-> Result<UpdatePipelineOutput, UpdatePipelineError>

Updates a specified pipeline with edits or changes to its structure. Use a JSON file with the pipeline structure in conjunction with UpdatePipeline to provide the full structure of the pipeline. Updating the pipeline increases the version number of the pipeline by 1.