Enum rusoto::cognitoidentity::UnlinkIdentityError

pub enum UnlinkIdentityError {
    InvalidParameter(String),
    ResourceNotFound(String),
    NotAuthorized(String),
    ResourceConflict(String),
    TooManyRequests(String),
    InternalError(String),
    ExternalService(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Variants
InvalidParameter(String)

Thrown for missing or bad input parameter(s).

ResourceNotFound(String)

Thrown when the requested resource (for example, a dataset or record) does not exist.

NotAuthorized(String)

Thrown when a user is not authorized to access the requested resource.

ResourceConflict(String)

Thrown when a user tries to use a login which is already linked to another account.

TooManyRequests(String)

Thrown when a request is throttled.

InternalError(String)

Thrown when the service encounters an error during processing the request.

ExternalService(String)

An exception thrown when a dependent service such as Facebook or Twitter is not responding

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl UnlinkIdentityError

fn from_body(body: &str) -> UnlinkIdentityError