Struct rusoto::cognitoidentity::CognitoIdentityClient

pub struct CognitoIdentityClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the Amazon Cognito Identity API.

Methods

impl<P> CognitoIdentityClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> CognitoIdentityClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn create_identity_pool(&self, input: &CreateIdentityPoolInput) -> Result<IdentityPool, CreateIdentityPoolError>

Creates a new identity pool. The identity pool is a store of user identity information that is specific to your AWS account. The limit on identity pools is 60 per account. The keys for SupportedLoginProviders are as follows:

Facebook: graph.facebook.com

Google: accounts.google.com

Amazon: www.amazon.com

Twitter: api.twitter.com

Digits: www.digits.com

You must use AWS Developer credentials to call this API.

fn delete_identities(&self, input: &DeleteIdentitiesInput) -> Result<DeleteIdentitiesResponse, DeleteIdentitiesError>

Deletes identities from an identity pool. You can specify a list of 1-60 identities that you want to delete.

You must use AWS Developer credentials to call this API.

fn delete_identity_pool(&self, input: &DeleteIdentityPoolInput) -> Result<(), DeleteIdentityPoolError>

Deletes a user pool. Once a pool is deleted, users will not be able to authenticate with the pool.

You must use AWS Developer credentials to call this API.

fn describe_identity(&self, input: &DescribeIdentityInput) -> Result<IdentityDescription, DescribeIdentityError>

Returns metadata related to the given identity, including when the identity was created and any associated linked logins.

You must use AWS Developer credentials to call this API.

fn describe_identity_pool(&self, input: &DescribeIdentityPoolInput) -> Result<IdentityPool, DescribeIdentityPoolError>

Gets details about a particular identity pool, including the pool name, ID description, creation date, and current number of users.

You must use AWS Developer credentials to call this API.

fn get_credentials_for_identity(&self, input: &GetCredentialsForIdentityInput) -> Result<GetCredentialsForIdentityResponse, GetCredentialsForIdentityError>

Returns credentials for the provided identity ID. Any provided logins will be validated against supported login providers. If the token is for cognito-identity.amazonaws.com, it will be passed through to AWS Security Token Service with the appropriate role for the token.

This is a public API. You do not need any credentials to call this API.

fn get_id(&self, input: &GetIdInput) -> Result<GetIdResponse, GetIdError>

Generates (or retrieves) a Cognito ID. Supplying multiple logins will create an implicit linked account.

This is a public API. You do not need any credentials to call this API.

fn get_identity_pool_roles(&self, input: &GetIdentityPoolRolesInput) -> Result<GetIdentityPoolRolesResponse, GetIdentityPoolRolesError>

Gets the roles for an identity pool.

You must use AWS Developer credentials to call this API.

fn get_open_id_token(&self, input: &GetOpenIdTokenInput) -> Result<GetOpenIdTokenResponse, GetOpenIdTokenError>

Gets an OpenID token, using a known Cognito ID. This known Cognito ID is returned by GetId. You can optionally add additional logins for the identity. Supplying multiple logins creates an implicit link.

The OpenId token is valid for 15 minutes.

This is a public API. You do not need any credentials to call this API.

fn get_open_id_token_for_developer_identity(&self, input: &GetOpenIdTokenForDeveloperIdentityInput) -> Result<GetOpenIdTokenForDeveloperIdentityResponse, GetOpenIdTokenForDeveloperIdentityError>

Registers (or retrieves) a Cognito IdentityId and an OpenID Connect token for a user authenticated by your backend authentication process. Supplying multiple logins will create an implicit linked account. You can only specify one developer provider as part of the Logins map, which is linked to the identity pool. The developer provider is the "domain" by which Cognito will refer to your users.

You can use GetOpenIdTokenForDeveloperIdentity to create a new identity and to link new logins (that is, user credentials issued by a public provider or developer provider) to an existing identity. When you want to create a new identity, the IdentityId should be null. When you want to associate a new login with an existing authenticated/unauthenticated identity, you can do so by providing the existing IdentityId. This API will create the identity in the specified IdentityPoolId.

You must use AWS Developer credentials to call this API.

fn list_identities(&self, input: &ListIdentitiesInput) -> Result<ListIdentitiesResponse, ListIdentitiesError>

Lists the identities in a pool.

You must use AWS Developer credentials to call this API.

fn list_identity_pools(&self, input: &ListIdentityPoolsInput) -> Result<ListIdentityPoolsResponse, ListIdentityPoolsError>

Lists all of the Cognito identity pools registered for your account.

You must use AWS Developer credentials to call this API.

fn lookup_developer_identity(&self, input: &LookupDeveloperIdentityInput) -> Result<LookupDeveloperIdentityResponse, LookupDeveloperIdentityError>

Retrieves the IdentityID associated with a DeveloperUserIdentifier or the list of DeveloperUserIdentifiers associated with an IdentityId for an existing identity. Either IdentityID or DeveloperUserIdentifier must not be null. If you supply only one of these values, the other value will be searched in the database and returned as a part of the response. If you supply both, DeveloperUserIdentifier will be matched against IdentityID. If the values are verified against the database, the response returns both values and is the same as the request. Otherwise a ResourceConflictException is thrown.

You must use AWS Developer credentials to call this API.

fn merge_developer_identities(&self, input: &MergeDeveloperIdentitiesInput) -> Result<MergeDeveloperIdentitiesResponse, MergeDeveloperIdentitiesError>

Merges two users having different IdentityIds, existing in the same identity pool, and identified by the same developer provider. You can use this action to request that discrete users be merged and identified as a single user in the Cognito environment. Cognito associates the given source user (SourceUserIdentifier) with the IdentityId of the DestinationUserIdentifier. Only developer-authenticated users can be merged. If the users to be merged are associated with the same public provider, but as two different users, an exception will be thrown.

You must use AWS Developer credentials to call this API.

fn set_identity_pool_roles(&self, input: &SetIdentityPoolRolesInput) -> Result<(), SetIdentityPoolRolesError>

Sets the roles for an identity pool. These roles are used when making calls to GetCredentialsForIdentity action.

You must use AWS Developer credentials to call this API.

fn unlink_developer_identity(&self, input: &UnlinkDeveloperIdentityInput) -> Result<(), UnlinkDeveloperIdentityError>

Unlinks a DeveloperUserIdentifier from an existing identity. Unlinked developer users will be considered new identities next time they are seen. If, for a given Cognito identity, you remove all federated identities as well as the developer user identifier, the Cognito identity becomes inaccessible.

You must use AWS Developer credentials to call this API.

fn unlink_identity(&self, input: &UnlinkIdentityInput) -> Result<(), UnlinkIdentityError>

Unlinks a federated identity from an existing account. Unlinked logins will be considered new identities next time they are seen. Removing the last linked login will make this identity inaccessible.

This is a public API. You do not need any credentials to call this API.

fn update_identity_pool(&self, input: &IdentityPool) -> Result<IdentityPool, UpdateIdentityPoolError>

Updates a user pool.

You must use AWS Developer credentials to call this API.