Struct rusoto::cognitoidentity::MergeDeveloperIdentitiesInput

pub struct MergeDeveloperIdentitiesInput {
    pub destination_user_identifier: DeveloperUserIdentifier,
    pub developer_provider_name: DeveloperProviderName,
    pub identity_pool_id: IdentityPoolId,
    pub source_user_identifier: DeveloperUserIdentifier,
}

Input to the MergeDeveloperIdentities action.

Fields
destination_user_identifier: DeveloperUserIdentifier

User identifier for the destination user. The value should be a DeveloperUserIdentifier.

developer_provider_name: DeveloperProviderName

The "domain" by which Cognito will refer to your users. This is a (pseudo) domain name that you provide while creating an identity pool. This name acts as a placeholder that allows your backend and the Cognito service to communicate about the developer provider. For the DeveloperProviderName, you can use letters as well as period (.), underscore (_), and dash (-).

identity_pool_id: IdentityPoolId

An identity pool ID in the format REGION:GUID.

source_user_identifier: DeveloperUserIdentifier

User identifier for the source user. The value should be a DeveloperUserIdentifier.