Enum rusoto::config::GetResourceConfigHistoryError

pub enum GetResourceConfigHistoryError {
    InvalidTimeRange(String),
    InvalidLimit(String),
    InvalidNextToken(String),
    NoAvailableConfigurationRecorder(String),
    ResourceNotDiscovered(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Variants
InvalidTimeRange(String)

The specified time range is not valid. The earlier time is not chronologically before the later time.

InvalidLimit(String)

The specified limit is outside the allowable range.

InvalidNextToken(String)

The specified next token is invalid. Specify the nextToken string that was returned in the previous response to get the next page of results.

NoAvailableConfigurationRecorder(String)

There are no configuration recorders available to provide the role needed to describe your resources. Create a configuration recorder.

ResourceNotDiscovered(String)

You have specified a resource that is either unknown or has not been discovered.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl GetResourceConfigHistoryError

fn from_body(body: &str) -> GetResourceConfigHistoryError