Enum rusoto::config::PutConfigRuleError

pub enum PutConfigRuleError {
    InvalidParameterValue(String),
    MaxNumberOfConfigRulesExceeded(String),
    ResourceInUse(String),
    InsufficientPermissions(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Variants
InvalidParameterValue(String)

One or more of the specified parameters are invalid. Verify that your parameters are valid and try again.

MaxNumberOfConfigRulesExceeded(String)

Failed to add the AWS Config rule because the account already contains the maximum number of 25 rules. Consider deleting any deactivated rules before adding new rules.

ResourceInUse(String)

The rule is currently being deleted. Wait for a while and try again.

InsufficientPermissions(String)

Indicates one of the following errors:

The rule cannot be created because the IAM role assigned to AWS Config lacks permissions to perform the config:Put* action.

The AWS Lambda function cannot be invoked. Check the function ARN, and check the function's permissions.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl PutConfigRuleError

fn from_body(body: &str) -> PutConfigRuleError