Struct rusoto::config::ConfigRule

pub struct ConfigRule {
    pub config_rule_arn: Option<String>,
    pub config_rule_id: Option<String>,
    pub config_rule_name: Option<StringWithCharLimit64>,
    pub config_rule_state: Option<ConfigRuleState>,
    pub description: Option<EmptiableStringWithCharLimit256>,
    pub input_parameters: Option<StringWithCharLimit256>,
    pub maximum_execution_frequency: Option<MaximumExecutionFrequency>,
    pub scope: Option<Scope>,
    pub source: Source,
}

An AWS Lambda function that evaluates configuration items to assess whether your AWS resources comply with your desired configurations. This function can run when AWS Config detects a configuration change to an AWS resource, or when it delivers a configuration snapshot of the resources in the account.

For more information about developing and using AWS Config rules, see Evaluating AWS Resource Configurations with AWS Config in the AWS Config Developer Guide.

Fields
config_rule_arn: Option<String>

The Amazon Resource Name (ARN) of the AWS Config rule.

config_rule_id: Option<String>

The ID of the AWS Config rule.

config_rule_name: Option<StringWithCharLimit64>

The name that you assign to the AWS Config rule. The name is required if you are adding a new rule.

config_rule_state: Option<ConfigRuleState>

Indicates whether the AWS Config rule is active or currently being deleted by AWS Config.

AWS Config sets the state of a rule to DELETING temporarily after you use the DeleteConfigRule request to delete the rule. After AWS Config finishes deleting a rule, the rule and all of its evaluations are erased and no longer available.

You cannot add a rule to AWS Config that has the state set to DELETING. If you want to delete a rule, you must use the DeleteConfigRule request.

description: Option<EmptiableStringWithCharLimit256>

The description that you provide for the AWS Config rule.

input_parameters: Option<StringWithCharLimit256>

A string in JSON format that is passed to the AWS Config rule Lambda function.

maximum_execution_frequency: Option<MaximumExecutionFrequency>

The maximum frequency at which the AWS Config rule runs evaluations.

If your rule is periodic, meaning it runs an evaluation when AWS Config delivers a configuration snapshot, then it cannot run evaluations more frequently than AWS Config delivers the snapshots. For periodic rules, set the value of the MaximumExecutionFrequency key to be equal to or greater than the value of the deliveryFrequency key, which is part of ConfigSnapshotDeliveryProperties. To update the frequency with which AWS Config delivers your snapshots, use the PutDeliveryChannel action.

scope: Option<Scope>

Defines which resources can trigger an evaluation for the rule. The scope can include one or more resource types, a combination of one resource type and one resource ID, or a combination of a tag key and value. Specify a scope to constrain the resources that can trigger an evaluation for the rule. If you do not specify a scope, evaluations are triggered when any resource in the recording group changes.

source: Source

Provides the rule owner (AWS or customer), the rule identifier, and the events that cause the function to evaluate your AWS resources.