Struct rusoto::config::ConfigRuleEvaluationStatus

pub struct ConfigRuleEvaluationStatus {
    pub config_rule_arn: Option<String>,
    pub config_rule_id: Option<String>,
    pub config_rule_name: Option<StringWithCharLimit64>,
    pub first_activated_time: Option<Date>,
    pub first_evaluation_started: Option<Boolean>,
    pub last_error_code: Option<String>,
    pub last_error_message: Option<String>,
    pub last_failed_evaluation_time: Option<Date>,
    pub last_failed_invocation_time: Option<Date>,
    pub last_successful_evaluation_time: Option<Date>,
    pub last_successful_invocation_time: Option<Date>,
}

Status information for your AWS managed Config rules. The status includes information such as the last time the rule ran, the last time it failed, and the related error for the last failure.

This action does not return status information about customer managed Config rules.

Fields
config_rule_arn: Option<String>

The Amazon Resource Name (ARN) of the AWS Config rule.

config_rule_id: Option<String>

The ID of the AWS Config rule.

config_rule_name: Option<StringWithCharLimit64>

The name of the AWS Config rule.

first_activated_time: Option<Date>

The time that you first activated the AWS Config rule.

first_evaluation_started: Option<Boolean>

Indicates whether AWS Config has evaluated your resources against the rule at least once.

true - AWS Config has evaluated your AWS resources against the rule at least once.

false - AWS Config has not once finished evaluating your AWS resources against the rule.

last_error_code: Option<String>

The error code that AWS Config returned when the rule last failed.

last_error_message: Option<String>

The error message that AWS Config returned when the rule last failed.

last_failed_evaluation_time: Option<Date>

The time that AWS Config last failed to evaluate your AWS resources against the rule.

last_failed_invocation_time: Option<Date>

The time that AWS Config last failed to invoke the AWS Config rule to evaluate your AWS resources.

last_successful_evaluation_time: Option<Date>

The time that AWS Config last successfully evaluated your AWS resources against the rule.

last_successful_invocation_time: Option<Date>

The time that AWS Config last successfully invoked the AWS Config rule to evaluate your AWS resources.