Struct rusoto::config::Source

pub struct Source {
    pub owner: Option<Owner>,
    pub source_details: Option<SourceDetails>,
    pub source_identifier: Option<StringWithCharLimit256>,
}

Provides the AWS Config rule owner (AWS or customer), the rule identifier, and the events that trigger the evaluation of your AWS resources.

Fields
owner: Option<Owner>

Indicates whether AWS or the customer owns and manages the AWS Config rule.

source_details: Option<SourceDetails>

Provides the source and type of the event that causes AWS Config to evaluate your AWS resources.

source_identifier: Option<StringWithCharLimit256>

For AWS managed Config rules, a pre-defined identifier from a list. To reference the list, see Using AWS Managed Config Rules.

For customer managed Config rules, the identifier is the Amazon Resource Name (ARN) of the rule's AWS Lambda function.