Module rusoto::datapipeline

AWS Data Pipeline

Structs

ActivatePipelineInput
Contains the parameters for ActivatePipeline.

ActivatePipelineOutput
Contains the output of ActivatePipeline.

AddTagsInput
Contains the parameters for AddTags.

AddTagsOutput
Contains the output of AddTags.

CreatePipelineInput
Contains the parameters for CreatePipeline.

CreatePipelineOutput
Contains the output of CreatePipeline.

DataPipelineClient
A client for the AWS Data Pipeline API.

DeactivatePipelineInput
Contains the parameters for DeactivatePipeline.

DeactivatePipelineOutput
Contains the output of DeactivatePipeline.

DeletePipelineInput
Contains the parameters for DeletePipeline.

DescribeObjectsInput
Contains the parameters for DescribeObjects.

DescribeObjectsOutput
Contains the output of DescribeObjects.

DescribePipelinesInput
Contains the parameters for DescribePipelines.

DescribePipelinesOutput
Contains the output of DescribePipelines.

EvaluateExpressionInput
Contains the parameters for EvaluateExpression.

EvaluateExpressionOutput
Contains the output of EvaluateExpression.

Field
A key-value pair that describes a property of a pipeline object. The value is specified as either a string value (StringValue) or a reference to another object (RefValue) but not as both.

GetPipelineDefinitionInput
Contains the parameters for GetPipelineDefinition.

GetPipelineDefinitionOutput
Contains the output of GetPipelineDefinition.

InstanceIdentity

Identity information for the EC2 instance that is hosting the task runner. You can get this value by calling a metadata URI from the EC2 instance. For more information, see Instance Metadata in the Amazon Elastic Compute Cloud User Guide. Passing in this value proves that your task runner is running on an EC2 instance, and ensures the proper AWS Data Pipeline service charges are applied to your pipeline.

ListPipelinesInput
Contains the parameters for ListPipelines.

ListPipelinesOutput
Contains the output of ListPipelines.

Operator
Contains a logical operation for comparing the value of a field with a specified value.

ParameterAttribute
The attributes allowed or specified with a parameter object.

ParameterObject
Contains information about a parameter object.

ParameterValue
A value or list of parameter values.

PipelineDescription
Contains pipeline metadata.

PipelineIdName
Contains the name and identifier of a pipeline.

PipelineObject
Contains information about a pipeline object. This can be a logical, physical, or physical attempt pipeline object. The complete set of components of a pipeline defines the pipeline.

PollForTaskInput
Contains the parameters for PollForTask.

PollForTaskOutput
Contains the output of PollForTask.

PutPipelineDefinitionInput
Contains the parameters for PutPipelineDefinition.

PutPipelineDefinitionOutput
Contains the output of PutPipelineDefinition.

Query
Defines the query to run against an object.

QueryObjectsInput
Contains the parameters for QueryObjects.

QueryObjectsOutput
Contains the output of QueryObjects.

RemoveTagsInput
Contains the parameters for RemoveTags.

RemoveTagsOutput
Contains the output of RemoveTags.

ReportTaskProgressInput
Contains the parameters for ReportTaskProgress.

ReportTaskProgressOutput
Contains the output of ReportTaskProgress.

ReportTaskRunnerHeartbeatInput
Contains the parameters for ReportTaskRunnerHeartbeat.

ReportTaskRunnerHeartbeatOutput
Contains the output of ReportTaskRunnerHeartbeat.

Selector
A comparision that is used to determine whether a query should return this object.

SetStatusInput
Contains the parameters for SetStatus.

SetTaskStatusInput
Contains the parameters for SetTaskStatus.

SetTaskStatusOutput
Contains the output of SetTaskStatus.

Tag
Tags are key/value pairs defined by a user and associated with a pipeline to control access. AWS Data Pipeline allows you to associate ten tags per pipeline. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

TaskObject
Contains information about a pipeline task that is assigned to a task runner.

ValidatePipelineDefinitionInput
Contains the parameters for ValidatePipelineDefinition.

ValidatePipelineDefinitionOutput
Contains the output of ValidatePipelineDefinition.

ValidationError
Defines a validation error. Validation errors prevent pipeline activation. The set of validation errors that can be returned are defined by AWS Data Pipeline.

ValidationWarning
Defines a validation warning. Validation warnings do not prevent pipeline activation. The set of validation warnings that can be returned are defined by AWS Data Pipeline.

Enums

ActivatePipelineError

AddTagsError

CreatePipelineError

DeactivatePipelineError

DeletePipelineError

DescribeObjectsError

DescribePipelinesError

EvaluateExpressionError

GetPipelineDefinitionError

ListPipelinesError

PollForTaskError

PutPipelineDefinitionError

QueryObjectsError

RemoveTagsError

ReportTaskProgressError

ReportTaskRunnerHeartbeatError

SetStatusError

SetTaskStatusError

ValidatePipelineDefinitionError

Type Definitions

AttributeNameString

AttributeValueString

Boolean

CancelActive

ErrorMessage

FieldList

FieldNameString

FieldStringValue

Id

IdList

Int

LongString

OperatorType

ParameterAttributeList

ParameterObjectList

ParameterValueList

PipelineDescriptionList

PipelineList

PipelineObjectList

PipelineObjectMap

SelectorList
The list of Selectors that define queries on individual fields.

StringList

TagKey

TagList

TagValue

TaskId

TaskStatus

Timestamp

ValidationErrors

ValidationMessage

ValidationMessages

ValidationWarnings