Struct rusoto::datapipeline::PutPipelineDefinitionOutput

pub struct PutPipelineDefinitionOutput {
    pub errored: Boolean,
    pub validation_errors: Option<ValidationErrors>,
    pub validation_warnings: Option<ValidationWarnings>,
}

Contains the output of PutPipelineDefinition.

Fields
errored: Boolean

Indicates whether there were validation errors, and the pipeline definition is stored but cannot be activated until you correct the pipeline and call PutPipelineDefinition to commit the corrected pipeline.

validation_errors: Option<ValidationErrors>

The validation errors that are associated with the objects defined in pipelineObjects.

validation_warnings: Option<ValidationWarnings>

The validation warnings that are associated with the objects defined in pipelineObjects.