Struct rusoto::devicefarm::DeviceFarmClient

pub struct DeviceFarmClient<P, D> where P: ProvideAwsCredentials,
        D: DispatchSignedRequest { /* fields omitted */ }

A client for the AWS Device Farm API.

Methods

impl<P, D> DeviceFarmClient<P, D> where P: ProvideAwsCredentials,
D: DispatchSignedRequest

fn new(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn create_device_pool(&self,
input: &CreateDevicePoolRequest)
-> Result<CreateDevicePoolResult, CreateDevicePoolError>

Creates a device pool.

fn create_project(&self,
input: &CreateProjectRequest)
-> Result<CreateProjectResult, CreateProjectError>

Creates a new project.

fn create_remote_access_session(&self,
input: &CreateRemoteAccessSessionRequest)
-> Result<CreateRemoteAccessSessionResult, CreateRemoteAccessSessionError>

Specifies and starts a remote access session.

fn create_upload(&self,
input: &CreateUploadRequest)
-> Result<CreateUploadResult, CreateUploadError>

Uploads an app or test scripts.

fn delete_device_pool(&self,
input: &DeleteDevicePoolRequest)
-> Result<DeleteDevicePoolResult, DeleteDevicePoolError>

Deletes a device pool given the pool ARN. Does not allow deletion of curated pools owned by the system.

fn delete_project(&self,
input: &DeleteProjectRequest)
-> Result<DeleteProjectResult, DeleteProjectError>

Deletes an AWS Device Farm project, given the project ARN.

Note Deleting this resource does not stop an in-progress run.

fn delete_remote_access_session(&self,
input: &DeleteRemoteAccessSessionRequest)
-> Result<DeleteRemoteAccessSessionResult, DeleteRemoteAccessSessionError>

Deletes a completed remote access session and its results.

fn delete_run(&self,
input: &DeleteRunRequest)
-> Result<DeleteRunResult, DeleteRunError>

Deletes the run, given the run ARN.

Note Deleting this resource does not stop an in-progress run.

fn delete_upload(&self,
input: &DeleteUploadRequest)
-> Result<DeleteUploadResult, DeleteUploadError>

Deletes an upload given the upload ARN.

fn get_account_settings(&self,
input: &GetAccountSettingsRequest)
-> Result<GetAccountSettingsResult, GetAccountSettingsError>

Returns the number of unmetered iOS and/or unmetered Android devices that have been purchased by the account.

fn get_device(&self,
input: &GetDeviceRequest)
-> Result<GetDeviceResult, GetDeviceError>

Gets information about a unique device type.

fn get_device_pool(&self,
input: &GetDevicePoolRequest)
-> Result<GetDevicePoolResult, GetDevicePoolError>

Gets information about a device pool.

fn get_device_pool_compatibility(&self,
input: &GetDevicePoolCompatibilityRequest)
-> Result<GetDevicePoolCompatibilityResult, GetDevicePoolCompatibilityError>

Gets information about compatibility with a device pool.

fn get_job(&self, input: &GetJobRequest) -> Result<GetJobResult, GetJobError>

Gets information about a job.

fn get_offering_status(&self,
input: &GetOfferingStatusRequest)
-> Result<GetOfferingStatusResult, GetOfferingStatusError>

Gets the current status and future status of all offerings purchased by an AWS account. The response indicates how many offerings are currently available and the offerings that will be available in the next period. The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

fn get_project(&self,
input: &GetProjectRequest)
-> Result<GetProjectResult, GetProjectError>

Gets information about a project.

fn get_remote_access_session(&self,
input: &GetRemoteAccessSessionRequest)
-> Result<GetRemoteAccessSessionResult, GetRemoteAccessSessionError>

Returns a link to a currently running remote access session.

fn get_run(&self, input: &GetRunRequest) -> Result<GetRunResult, GetRunError>

Gets information about a run.

fn get_suite(&self,
input: &GetSuiteRequest)
-> Result<GetSuiteResult, GetSuiteError>

Gets information about a suite.

fn get_test(&self,
input: &GetTestRequest)
-> Result<GetTestResult, GetTestError>

Gets information about a test.

fn get_upload(&self,
input: &GetUploadRequest)
-> Result<GetUploadResult, GetUploadError>

Gets information about an upload.

fn install_to_remote_access_session(&self,
input: &InstallToRemoteAccessSessionRequest)
-> Result<InstallToRemoteAccessSessionResult, InstallToRemoteAccessSessionError>

Installs an application to the device in a remote access session. For Android applications, the file must be in .apk format. For iOS applications, the file must be in .ipa format.

fn list_artifacts(&self,
input: &ListArtifactsRequest)
-> Result<ListArtifactsResult, ListArtifactsError>

Gets information about artifacts.

fn list_device_pools(&self,
input: &ListDevicePoolsRequest)
-> Result<ListDevicePoolsResult, ListDevicePoolsError>

Gets information about device pools.

fn list_devices(&self,
input: &ListDevicesRequest)
-> Result<ListDevicesResult, ListDevicesError>

Gets information about unique device types.

fn list_jobs(&self,
input: &ListJobsRequest)
-> Result<ListJobsResult, ListJobsError>

Gets information about jobs.

fn list_offering_transactions(&self,
input: &ListOfferingTransactionsRequest)
-> Result<ListOfferingTransactionsResult, ListOfferingTransactionsError>

Returns a list of all historical purchases, renewals, and system renewal transactions for an AWS account. The list is paginated and ordered by a descending timestamp (most recent transactions are first). The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

fn list_offerings(&self,
input: &ListOfferingsRequest)
-> Result<ListOfferingsResult, ListOfferingsError>

Returns a list of products or offerings that the user can manage through the API. Each offering record indicates the recurring price per unit and the frequency for that offering. The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

fn list_projects(&self,
input: &ListProjectsRequest)
-> Result<ListProjectsResult, ListProjectsError>

Gets information about projects.

fn list_remote_access_sessions(&self,
input: &ListRemoteAccessSessionsRequest)
-> Result<ListRemoteAccessSessionsResult, ListRemoteAccessSessionsError>

Returns a list of all currently running remote access sessions.

fn list_runs(&self,
input: &ListRunsRequest)
-> Result<ListRunsResult, ListRunsError>

Gets information about runs, given an AWS Device Farm project ARN.

fn list_samples(&self,
input: &ListSamplesRequest)
-> Result<ListSamplesResult, ListSamplesError>

Gets information about samples, given an AWS Device Farm project ARN

fn list_suites(&self,
input: &ListSuitesRequest)
-> Result<ListSuitesResult, ListSuitesError>

Gets information about suites.

fn list_tests(&self,
input: &ListTestsRequest)
-> Result<ListTestsResult, ListTestsError>

Gets information about tests.

fn list_unique_problems(&self,
input: &ListUniqueProblemsRequest)
-> Result<ListUniqueProblemsResult, ListUniqueProblemsError>

Gets information about unique problems.

fn list_uploads(&self,
input: &ListUploadsRequest)
-> Result<ListUploadsResult, ListUploadsError>

Gets information about uploads, given an AWS Device Farm project ARN.

fn purchase_offering(&self,
input: &PurchaseOfferingRequest)
-> Result<PurchaseOfferingResult, PurchaseOfferingError>

Immediately purchases offerings for an AWS account. Offerings renew with the latest total purchased quantity for an offering, unless the renewal was overridden. The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

fn renew_offering(&self,
input: &RenewOfferingRequest)
-> Result<RenewOfferingResult, RenewOfferingError>

Explicitly sets the quantity of devices to renew for an offering, starting from the effectiveDate of the next period. The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

fn schedule_run(&self,
input: &ScheduleRunRequest)
-> Result<ScheduleRunResult, ScheduleRunError>

Schedules a run.

fn stop_remote_access_session(&self,
input: &StopRemoteAccessSessionRequest)
-> Result<StopRemoteAccessSessionResult, StopRemoteAccessSessionError>

Ends a specified remote access session.

fn stop_run(&self,
input: &StopRunRequest)
-> Result<StopRunResult, StopRunError>

Initiates a stop request for the current test run. AWS Device Farm will immediately stop the run on devices where tests have not started executing, and you will not be billed for these devices. On devices where tests have started executing, Setup Suite and Teardown Suite tests will run to completion before stopping execution on those devices. You will be billed for Setup, Teardown, and any tests that were in progress or already completed.

fn update_device_pool(&self,
input: &UpdateDevicePoolRequest)
-> Result<UpdateDevicePoolResult, UpdateDevicePoolError>

Modifies the name, description, and rules in a device pool given the attributes and the pool ARN. Rule updates are all-or-nothing, meaning they can only be updated as a whole (or not at all).

fn update_project(&self,
input: &UpdateProjectRequest)
-> Result<UpdateProjectResult, UpdateProjectError>

Modifies the specified project name, given the project ARN and a new name.