Struct rusoto::devicefarm::GetDevicePoolCompatibilityRequest

pub struct GetDevicePoolCompatibilityRequest {
    pub app_arn: Option<AmazonResourceName>,
    pub device_pool_arn: AmazonResourceName,
    pub test_type: Option<TestType>,
}

Represents a request to the get device pool compatibility operation.

Fields
app_arn: Option<AmazonResourceName>

The ARN of the app that is associated with the specified device pool.

device_pool_arn: AmazonResourceName

The device pool's ARN.

test_type: Option<TestType>

The test type for the specified device pool.

Allowed values include the following:

BUILTIN_FUZZ: The built-in fuzz type.

BUILTIN_EXPLORER: For Android, an app explorer that will traverse an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT: The Appium Java JUnit type.

APPIUM_JAVA_TESTNG: The Appium Java TestNG type.

APPIUM_PYTHON: The Appium Python type.

APPIUM_WEB_JAVA_JUNIT: The Appium Java JUnit type for Web apps.

APPIUM_WEB_JAVA_TESTNG: The Appium Java TestNG type for Web apps.

APPIUM_WEB_PYTHON: The Appium Python type for Web apps.

CALABASH: The Calabash type.

INSTRUMENTATION: The Instrumentation type.

UIAUTOMATION: The uiautomation type.

UIAUTOMATOR: The uiautomator type.

XCTEST: The XCode test type.

XCTEST_UI: The XCode UI test type.