Struct rusoto::devicefarm::Run

pub struct Run {
    pub arn: Option<AmazonResourceName>,
    pub billing_method: Option<BillingMethod>,
    pub completed_jobs: Option<Integer>,
    pub counters: Option<Counters>,
    pub created: Option<DateTime>,
    pub device_minutes: Option<DeviceMinutes>,
    pub message: Option<Message>,
    pub name: Option<Name>,
    pub platform: Option<DevicePlatform>,
    pub result: Option<ExecutionResult>,
    pub started: Option<DateTime>,
    pub status: Option<ExecutionStatus>,
    pub stopped: Option<DateTime>,
    pub total_jobs: Option<Integer>,
    pub type_: Option<TestType>,
}

Represents an app on a set of devices with a specific test and configuration.

Fields
arn: Option<AmazonResourceName>

The run's ARN.

billing_method: Option<BillingMethod>

Specifies the billing method for a test run: metered or unmetered. If the parameter is not specified, the default value is unmetered.

completed_jobs: Option<Integer>

The total number of completed jobs.

counters: Option<Counters>

The run's result counters.

created: Option<DateTime>

When the run was created.

device_minutes: Option<DeviceMinutes>

Represents the total (metered or unmetered) minutes used by the test run.

message: Option<Message>

A message about the run's result.

name: Option<Name>

The run's name.

platform: Option<DevicePlatform>

The run's platform.

Allowed values include:

ANDROID: The Android platform.

IOS: The iOS platform.

result: Option<ExecutionResult>

The run's result.

Allowed values include:

PENDING: A pending condition.

PASSED: A passing condition.

WARNED: A warning condition.

FAILED: A failed condition.

SKIPPED: A skipped condition.

ERRORED: An error condition.

STOPPED: A stopped condition.

started: Option<DateTime>

The run's start time.

status: Option<ExecutionStatus>

The run's status.

Allowed values include:

PENDING: A pending status.

PENDING_CONCURRENCY: A pending concurrency status.

PENDING_DEVICE: A pending device status.

PROCESSING: A processing status.

SCHEDULING: A scheduling status.

PREPARING: A preparing status.

RUNNING: A running status.

COMPLETED: A completed status.

STOPPING: A stopping status.

stopped: Option<DateTime>

The run's stop time.

total_jobs: Option<Integer>

The total number of jobs for the run.

type_: Option<TestType>

The run's type.

Must be one of the following values:

BUILTIN_FUZZ: The built-in fuzz type.

BUILTIN_EXPLORER: For Android, an app explorer that will traverse an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT: The Appium Java JUnit type.

APPIUM_JAVA_TESTNG: The Appium Java TestNG type.

APPIUM_PYTHON: The Appium Python type.

APPIUM_WEB_JAVA_JUNIT: The Appium Java JUnit type for Web apps.

APPIUM_WEB_JAVA_TESTNG: The Appium Java TestNG type for Web apps.

APPIUM_WEB_PYTHON: The Appium Python type for Web apps.

CALABASH: The Calabash type.

INSTRUMENTATION: The Instrumentation type.

UIAUTOMATION: The uiautomation type.

UIAUTOMATOR: The uiautomator type.

XCTEST: The XCode test type.

XCTEST_UI: The XCode UI test type.