Struct rusoto::devicefarm::ScheduleRunRequest

pub struct ScheduleRunRequest {
    pub app_arn: Option<AmazonResourceName>,
    pub configuration: Option<ScheduleRunConfiguration>,
    pub device_pool_arn: AmazonResourceName,
    pub name: Option<Name>,
    pub project_arn: AmazonResourceName,
    pub test: ScheduleRunTest,
}

Represents a request to the schedule run operation.

Fields
app_arn: Option<AmazonResourceName>

The ARN of the app to schedule a run.

configuration: Option<ScheduleRunConfiguration>

Information about the settings for the run to be scheduled.

device_pool_arn: AmazonResourceName

The ARN of the device pool for the run to be scheduled.

name: Option<Name>

The name for the run to be scheduled.

project_arn: AmazonResourceName

The ARN of the project for the run to be scheduled.

test: ScheduleRunTest

Information about the test for the run to be scheduled.