Struct rusoto::devicefarm::Test

pub struct Test {
    pub arn: Option<AmazonResourceName>,
    pub counters: Option<Counters>,
    pub created: Option<DateTime>,
    pub device_minutes: Option<DeviceMinutes>,
    pub message: Option<Message>,
    pub name: Option<Name>,
    pub result: Option<ExecutionResult>,
    pub started: Option<DateTime>,
    pub status: Option<ExecutionStatus>,
    pub stopped: Option<DateTime>,
    pub type_: Option<TestType>,
}

Represents a condition that is evaluated.

Fields
arn: Option<AmazonResourceName>

The test's ARN.

counters: Option<Counters>

The test's result counters.

created: Option<DateTime>

When the test was created.

device_minutes: Option<DeviceMinutes>

Represents the total (metered or unmetered) minutes used by the test.

message: Option<Message>

A message about the test's result.

name: Option<Name>

The test's name.

result: Option<ExecutionResult>

The test's result.

Allowed values include:

PENDING: A pending condition.

PASSED: A passing condition.

WARNED: A warning condition.

FAILED: A failed condition.

SKIPPED: A skipped condition.

ERRORED: An error condition.

STOPPED: A stopped condition.

started: Option<DateTime>

The test's start time.

status: Option<ExecutionStatus>

The test's status.

Allowed values include:

PENDING: A pending status.

PENDING_CONCURRENCY: A pending concurrency status.

PENDING_DEVICE: A pending device status.

PROCESSING: A processing status.

SCHEDULING: A scheduling status.

PREPARING: A preparing status.

RUNNING: A running status.

COMPLETED: A completed status.

STOPPING: A stopping status.

stopped: Option<DateTime>

The test's stop time.

type_: Option<TestType>

The test's type.

Must be one of the following values:

BUILTIN_FUZZ: The built-in fuzz type.

BUILTIN_EXPLORER: For Android, an app explorer that will traverse an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT: The Appium Java JUnit type.

APPIUM_JAVA_TESTNG: The Appium Java TestNG type.

APPIUM_PYTHON: The Appium Python type.

APPIUM_WEB_JAVA_JUNIT: The Appium Java JUnit type for Web apps.

APPIUM_WEB_JAVA_TESTNG: The Appium Java TestNG type for Web apps.

APPIUM_WEB_PYTHON: The Appium Python type for Web apps.

CALABASH: The Calabash type.

INSTRUMENTATION: The Instrumentation type.

UIAUTOMATION: The uiautomation type.

UIAUTOMATOR: The uiautomator type.

XCTEST: The XCode test type.

XCTEST_UI: The XCode UI test type.