Struct rusoto::devicefarm::Upload

pub struct Upload {
    pub arn: Option<AmazonResourceName>,
    pub content_type: Option<ContentType>,
    pub created: Option<DateTime>,
    pub message: Option<Message>,
    pub metadata: Option<Metadata>,
    pub name: Option<Name>,
    pub status: Option<UploadStatus>,
    pub type_: Option<UploadType>,
    pub url: Option<URL>,
}

An app or a set of one or more tests to upload or that have been uploaded.

Fields
arn: Option<AmazonResourceName>

The upload's ARN.

content_type: Option<ContentType>

The upload's content type (for example, "application/octet-stream").

created: Option<DateTime>

When the upload was created.

message: Option<Message>

A message about the upload's result.

metadata: Option<Metadata>

The upload's metadata. For example, for Android, this contains information that is parsed from the manifest and is displayed in the AWS Device Farm console after the associated app is uploaded.

name: Option<Name>

The upload's file name.

status: Option<UploadStatus>

The upload's status.

Must be one of the following values:

FAILED: A failed status.

INITIALIZED: An initialized status.

PROCESSING: A processing status.

SUCCEEDED: A succeeded status.

type_: Option<UploadType>

The upload's type.

Must be one of the following values:

ANDROID_APP: An Android upload.

IOS_APP: An iOS upload.

WEB_APP: A web appliction upload.

EXTERNAL_DATA: An external data upload.

APPIUM_JAVA_JUNIT_TEST_PACKAGE: An Appium Java JUnit test package upload.

APPIUM_JAVA_TESTNG_TEST_PACKAGE: An Appium Java TestNG test package upload.

APPIUM_PYTHON_TEST_PACKAGE: An Appium Python test package upload.

APPIUM_WEB_JAVA_JUNIT_TEST_PACKAGE: An Appium Java JUnit test package upload.

APPIUM_WEB_JAVA_TESTNG_TEST_PACKAGE: An Appium Java TestNG test package upload.

APPIUM_WEB_PYTHON_TEST_PACKAGE: An Appium Python test package upload.

CALABASH_TEST_PACKAGE: A Calabash test package upload.

INSTRUMENTATION_TEST_PACKAGE: An instrumentation upload.

UIAUTOMATION_TEST_PACKAGE: A uiautomation test package upload.

UIAUTOMATOR_TEST_PACKAGE: A uiautomator test package upload.

XCTEST_TEST_PACKAGE: An XCode test package upload.

XCTEST_UI_TEST_PACKAGE: An XCode UI test package upload.

url: Option<URL>

The pre-signed Amazon S3 URL that was used to store a file through a corresponding PUT request.