Struct rusoto::directconnect::AllocatePublicVirtualInterfaceRequest

pub struct AllocatePublicVirtualInterfaceRequest {
    pub connection_id: ConnectionId,
    pub new_public_virtual_interface_allocation: NewPublicVirtualInterfaceAllocation,
    pub owner_account: OwnerAccount,
}

Container for the parameters to the AllocatePublicVirtualInterface operation.

Fields
connection_id: ConnectionId

The connection ID on which the public virtual interface is provisioned.

Default: None

new_public_virtual_interface_allocation: NewPublicVirtualInterfaceAllocation

Detailed information for the public virtual interface to be provisioned.

Default: None

owner_account: OwnerAccount

The AWS account that will own the new public virtual interface.

Default: None