Struct rusoto::directconnect::Interconnect

pub struct Interconnect {
    pub bandwidth: Option<Bandwidth>,
    pub interconnect_id: Option<InterconnectId>,
    pub interconnect_name: Option<InterconnectName>,
    pub interconnect_state: Option<InterconnectState>,
    pub location: Option<LocationCode>,
    pub region: Option<Region>,
}

An interconnect is a connection that can host other connections.

Like a standard AWS Direct Connect connection, an interconnect represents the physical connection between an AWS Direct Connect partner's network and a specific Direct Connect location. An AWS Direct Connect partner who owns an interconnect can provision hosted connections on the interconnect for their end customers, thereby providing the end customers with connectivity to AWS services.

The resources of the interconnect, including bandwidth and VLAN numbers, are shared by all of the hosted connections on the interconnect, and the owner of the interconnect determines how these resources are assigned.

Fields
bandwidth: Option<Bandwidth> interconnect_id: Option<InterconnectId> interconnect_name: Option<InterconnectName> interconnect_state: Option<InterconnectState> location: Option<LocationCode> region: Option<Region>