Module rusoto::ds

AWS Directory Service

If you're using the service, you're probably looking for DirectoryServiceClient.

Structs

AddIpRoutesRequest

AddIpRoutesResult

AddTagsToResourceRequest

AddTagsToResourceResult

Attribute
Represents a named directory attribute.

CancelSchemaExtensionRequest

CancelSchemaExtensionResult

Computer
Contains information about a computer account in a directory.

ConditionalForwarder
Points to a remote domain with which you are setting up a trust relationship. Conditional forwarders are required in order to set up a trust relationship with another domain.

ConnectDirectoryRequest
Contains the inputs for the ConnectDirectory operation.

ConnectDirectoryResult
Contains the results of the ConnectDirectory operation.

CreateAliasRequest
Contains the inputs for the CreateAlias operation.

CreateAliasResult
Contains the results of the CreateAlias operation.

CreateComputerRequest
Contains the inputs for the CreateComputer operation.

CreateComputerResult
Contains the results for the CreateComputer operation.

CreateConditionalForwarderRequest
Initiates the creation of a conditional forwarder for your AWS Directory Service for Microsoft Active Directory. Conditional forwarders are required in order to set up a trust relationship with another domain.

CreateConditionalForwarderResult
The result of a CreateConditinalForwarder request.

CreateDirectoryRequest
Contains the inputs for the CreateDirectory operation.

CreateDirectoryResult
Contains the results of the CreateDirectory operation.

CreateMicrosoftADRequest
Creates a Microsoft AD in the AWS cloud.

CreateMicrosoftADResult
Result of a CreateMicrosoftAD request.

CreateSnapshotRequest
Contains the inputs for the CreateSnapshot operation.

CreateSnapshotResult
Contains the results of the CreateSnapshot operation.

CreateTrustRequest
AWS Directory Service for Microsoft Active Directory allows you to configure trust relationships. For example, you can establish a trust between your Microsoft AD in the AWS cloud, and your existing on-premises Microsoft Active Directory. This would allow you to provide users and groups access to resources in either domain, with a single set of credentials.

This action initiates the creation of the AWS side of a trust relationship between a Microsoft AD in the AWS cloud and an external domain.

CreateTrustResult
The result of a CreateTrust request.

DeleteConditionalForwarderRequest
Deletes a conditional forwarder.

DeleteConditionalForwarderResult
The result of a DeleteConditionalForwarder request.

DeleteDirectoryRequest
Contains the inputs for the DeleteDirectory operation.

DeleteDirectoryResult
Contains the results of the DeleteDirectory operation.

DeleteSnapshotRequest
Contains the inputs for the DeleteSnapshot operation.

DeleteSnapshotResult
Contains the results of the DeleteSnapshot operation.

DeleteTrustRequest
Deletes the local side of an existing trust relationship between the Microsoft AD in the AWS cloud and the external domain.

DeleteTrustResult
The result of a DeleteTrust request.

DeregisterEventTopicRequest
Removes the specified directory as a publisher to the specified SNS topic.

DeregisterEventTopicResult
The result of a DeregisterEventTopic request.

DescribeConditionalForwardersRequest
Describes a conditional forwarder.

DescribeConditionalForwardersResult
The result of a DescribeConditionalForwarder request.

DescribeDirectoriesRequest
Contains the inputs for the DescribeDirectories operation.

DescribeDirectoriesResult
Contains the results of the DescribeDirectories operation.

DescribeEventTopicsRequest
Describes event topics.

DescribeEventTopicsResult
The result of a DescribeEventTopic request.

DescribeSnapshotsRequest
Contains the inputs for the DescribeSnapshots operation.

DescribeSnapshotsResult
Contains the results of the DescribeSnapshots operation.

DescribeTrustsRequest
Describes the trust relationships for a particular Microsoft AD in the AWS cloud. If no input parameters are are provided, such as directory ID or trust ID, this request describes all the trust relationships.

DescribeTrustsResult
The result of a DescribeTrust request.

DirectoryConnectSettings
Contains information for the ConnectDirectory operation when an AD Connector directory is being created.

DirectoryConnectSettingsDescription
Contains information about an AD Connector directory.

DirectoryDescription
Contains information about an AWS Directory Service directory.

DirectoryLimits
Contains directory limit information for a region.

DirectoryServiceClient
A client for the Directory Service API.

DirectoryVpcSettings
Contains VPC information for the CreateDirectory or CreateMicrosoftAD operation.

DirectoryVpcSettingsDescription
Contains information about the directory.

DisableRadiusRequest
Contains the inputs for the DisableRadius operation.

DisableRadiusResult
Contains the results of the DisableRadius operation.

DisableSsoRequest
Contains the inputs for the DisableSso operation.

DisableSsoResult
Contains the results of the DisableSso operation.

EnableRadiusRequest
Contains the inputs for the EnableRadius operation.

EnableRadiusResult
Contains the results of the EnableRadius operation.

EnableSsoRequest
Contains the inputs for the EnableSso operation.

EnableSsoResult
Contains the results of the EnableSso operation.

EventTopic
Information about SNS topic and AWS Directory Service directory associations.

GetDirectoryLimitsRequest
Contains the inputs for the GetDirectoryLimits operation.

GetDirectoryLimitsResult
Contains the results of the GetDirectoryLimits operation.

GetSnapshotLimitsRequest
Contains the inputs for the GetSnapshotLimits operation.

GetSnapshotLimitsResult
Contains the results of the GetSnapshotLimits operation.

IpRoute
IP address block. This is often the address block of the DNS server used for your on-premises domain.

IpRouteInfo
Information about one or more IP address blocks.

ListIpRoutesRequest

ListIpRoutesResult

ListSchemaExtensionsRequest

ListSchemaExtensionsResult

ListTagsForResourceRequest

ListTagsForResourceResult

RadiusSettings
Contains information about a Remote Authentication Dial In User Service (RADIUS) server.

RegisterEventTopicRequest
Registers a new event topic.

RegisterEventTopicResult
The result of a RegisterEventTopic request.

RemoveIpRoutesRequest

RemoveIpRoutesResult

RemoveTagsFromResourceRequest

RemoveTagsFromResourceResult

RestoreFromSnapshotRequest
An object representing the inputs for the RestoreFromSnapshot operation.

RestoreFromSnapshotResult
Contains the results of the RestoreFromSnapshot operation.

SchemaExtensionInfo
Information about a schema extension.

Snapshot
Describes a directory snapshot.

SnapshotLimits
Contains manual snapshot limit information for a directory.

StartSchemaExtensionRequest

StartSchemaExtensionResult

Tag
Metadata assigned to a directory consisting of a key-value pair.

Trust
Describes a trust relationship between an Microsoft AD in the AWS cloud and an external domain.

UpdateConditionalForwarderRequest
Updates a conditional forwarder.

UpdateConditionalForwarderResult
The result of an UpdateConditionalForwarder request.

UpdateRadiusRequest
Contains the inputs for the UpdateRadius operation.

UpdateRadiusResult
Contains the results of the UpdateRadius operation.

VerifyTrustRequest
Initiates the verification of an existing trust relationship between a Microsoft AD in the AWS cloud and an external domain.

VerifyTrustResult
Result of a VerifyTrust request.

Enums

AddIpRoutesError
Errors returned by AddIpRoutes

AddTagsToResourceError
Errors returned by AddTagsToResource

CancelSchemaExtensionError
Errors returned by CancelSchemaExtension

ConnectDirectoryError
Errors returned by ConnectDirectory

CreateAliasError
Errors returned by CreateAlias

CreateComputerError
Errors returned by CreateComputer

CreateConditionalForwarderError
Errors returned by CreateConditionalForwarder

CreateDirectoryError
Errors returned by CreateDirectory

CreateMicrosoftADError
Errors returned by CreateMicrosoftAD

CreateSnapshotError
Errors returned by CreateSnapshot

CreateTrustError
Errors returned by CreateTrust

DeleteConditionalForwarderError
Errors returned by DeleteConditionalForwarder

DeleteDirectoryError
Errors returned by DeleteDirectory

DeleteSnapshotError
Errors returned by DeleteSnapshot

DeleteTrustError
Errors returned by DeleteTrust

DeregisterEventTopicError
Errors returned by DeregisterEventTopic

DescribeConditionalForwardersError
Errors returned by DescribeConditionalForwarders

DescribeDirectoriesError
Errors returned by DescribeDirectories

DescribeEventTopicsError
Errors returned by DescribeEventTopics

DescribeSnapshotsError
Errors returned by DescribeSnapshots

DescribeTrustsError
Errors returned by DescribeTrusts

DisableRadiusError
Errors returned by DisableRadius

DisableSsoError
Errors returned by DisableSso

EnableRadiusError
Errors returned by EnableRadius

EnableSsoError
Errors returned by EnableSso

GetDirectoryLimitsError
Errors returned by GetDirectoryLimits

GetSnapshotLimitsError
Errors returned by GetSnapshotLimits

ListIpRoutesError
Errors returned by ListIpRoutes

ListSchemaExtensionsError
Errors returned by ListSchemaExtensions

ListTagsForResourceError
Errors returned by ListTagsForResource

RegisterEventTopicError
Errors returned by RegisterEventTopic

RemoveIpRoutesError
Errors returned by RemoveIpRoutes

RemoveTagsFromResourceError
Errors returned by RemoveTagsFromResource

RestoreFromSnapshotError
Errors returned by RestoreFromSnapshot

StartSchemaExtensionError
Errors returned by StartSchemaExtension

UpdateConditionalForwarderError
Errors returned by UpdateConditionalForwarder

UpdateRadiusError
Errors returned by UpdateRadius

VerifyTrustError
Errors returned by VerifyTrust

Type Definitions

AccessUrl

AddedDateTime

AliasName

AttributeName

AttributeValue

Attributes

AvailabilityZone

AvailabilityZones

CidrIp

CidrIps

CloudOnlyDirectoriesLimitReached

ComputerName

ComputerPassword

ConditionalForwarders

ConnectPassword

ConnectedDirectoriesLimitReached

CreateSnapshotBeforeSchemaExtension

CreatedDateTime

DeleteAssociatedConditionalForwarder

Description

DirectoryDescriptions
A list of directory descriptions.

DirectoryId

DirectoryIds
A list of directory identifiers.

DirectoryName

DirectoryShortName

DirectorySize

DirectoryStage

DirectoryType

DnsIpAddrs

EndDateTime

EventTopics

ExceptionMessage
The descriptive message for the exception.

IpAddr

IpAddrs

IpRouteStatusMsg

IpRouteStatusReason

IpRoutes

IpRoutesInfo

LastUpdatedDateTime

LaunchTime

LdifContent

Limit

ManualSnapshotsLimitReached

NextToken

OrganizationalUnitDN

Password

PortNumber

RadiusAuthenticationProtocol

RadiusDisplayLabel

RadiusRetries

RadiusSharedSecret

RadiusStatus

RadiusTimeout

RemoteDomainName

RemoteDomainNames

ReplicationScope

RequestId
The AWS request identifier.

ResourceId

SID

SchemaExtensionId

SchemaExtensionStatus

SchemaExtensionStatusReason

SchemaExtensionsInfo

SecurityGroupId

Server

Servers

SnapshotId

SnapshotIds
A list of directory snapshot identifiers.

SnapshotName

SnapshotStatus

SnapshotType

Snapshots
A list of descriptions of directory snapshots.

SsoEnabled

StageReason

StartDateTime

StartTime

StateLastUpdatedDateTime

SubnetId

SubnetIds

TagKey

TagKeys

TagValue

Tags

TopicArn

TopicName

TopicNames

TopicStatus

TrustDirection

TrustId

TrustIds

TrustPassword

TrustState

TrustStateReason

TrustType

Trusts

UpdateSecurityGroupForDirectoryControllers

UseSameUsername

UserName

VpcId