Struct rusoto::ds::CreateDirectoryRequest

pub struct CreateDirectoryRequest {
    pub description: Option<Description>,
    pub name: DirectoryName,
    pub password: Password,
    pub short_name: Option<DirectoryShortName>,
    pub size: DirectorySize,
    pub vpc_settings: Option<DirectoryVpcSettings>,
}

Contains the inputs for the CreateDirectory operation.

Fields
description: Option<Description>

A textual description for the directory.

name: DirectoryName

The fully qualified name for the directory, such as corp.example.com.

password: Password

The password for the directory administrator. The directory creation process creates a directory administrator account with the username Administrator and this password.

short_name: Option<DirectoryShortName>

The short name of the directory, such as CORP.

size: DirectorySize

The size of the directory.

vpc_settings: Option<DirectoryVpcSettings>

A DirectoryVpcSettings object that contains additional information for the operation.