Struct rusoto::ds::CreateMicrosoftADRequest

pub struct CreateMicrosoftADRequest {
    pub description: Option<Description>,
    pub name: DirectoryName,
    pub password: Password,
    pub short_name: Option<DirectoryShortName>,
    pub vpc_settings: DirectoryVpcSettings,
}

Creates a Microsoft AD in the AWS cloud.

Fields
description: Option<Description>

A textual description for the directory. This label will appear on the AWS console Directory Details page after the directory is created.

name: DirectoryName

The fully qualified domain name for the directory, such as corp.example.com. This name will resolve inside your VPC only. It does not need to be publicly resolvable.

password: Password

The password for the default administrative user named Admin.

short_name: Option<DirectoryShortName>

The NetBIOS name for your domain. A short identifier for your domain, such as CORP. If you don't specify a NetBIOS name, it will default to the first part of your directory DNS. For example, CORP for the directory DNS corp.example.com.

vpc_settings: DirectoryVpcSettings