Struct rusoto::ds::CreateTrustRequest

pub struct CreateTrustRequest {
    pub conditional_forwarder_ip_addrs: Option<DnsIpAddrs>,
    pub directory_id: DirectoryId,
    pub remote_domain_name: RemoteDomainName,
    pub trust_direction: TrustDirection,
    pub trust_password: TrustPassword,
    pub trust_type: Option<TrustType>,
}

AWS Directory Service for Microsoft Active Directory allows you to configure trust relationships. For example, you can establish a trust between your Microsoft AD in the AWS cloud, and your existing on-premises Microsoft Active Directory. This would allow you to provide users and groups access to resources in either domain, with a single set of credentials.

This action initiates the creation of the AWS side of a trust relationship between a Microsoft AD in the AWS cloud and an external domain.

Fields
conditional_forwarder_ip_addrs: Option<DnsIpAddrs> directory_id: DirectoryId

The Directory ID of the Microsoft AD in the AWS cloud for which to establish the trust relationship.

remote_domain_name: RemoteDomainName

The Fully Qualified Domain Name (FQDN) of the external domain for which to create the trust relationship.

trust_direction: TrustDirection

The direction of the trust relationship.

trust_password: TrustPassword

The trust password. The must be the same password that was used when creating the trust relationship on the external domain.

trust_type: Option<TrustType>

The trust relationship type.