Struct rusoto::ds::DescribeDirectoriesRequest

pub struct DescribeDirectoriesRequest {
    pub directory_ids: Option<DirectoryIds>,
    pub limit: Option<Limit>,
    pub next_token: Option<NextToken>,
}

Contains the inputs for the DescribeDirectories operation.

Fields
directory_ids: Option<DirectoryIds>

A list of identifiers of the directories for which to obtain the information. If this member is null, all directories that belong to the current account are returned.

An empty list results in an InvalidParameterException being thrown.

limit: Option<Limit>

The maximum number of items to return. If this value is zero, the maximum number of items is specified by the limitations of the operation.

next_token: Option<NextToken>

The DescribeDirectoriesResult.NextToken value from a previous call to DescribeDirectories. Pass null if this is the first call.