Struct rusoto::ds::DescribeEventTopicsRequest

pub struct DescribeEventTopicsRequest {
    pub directory_id: Option<DirectoryId>,
    pub topic_names: Option<TopicNames>,
}

Fields
directory_id: Option<DirectoryId>

The Directory ID for which to get the list of associated SNS topics. If this member is null, associations for all Directory IDs are returned.

topic_names: Option<TopicNames>

A list of SNS topic names for which to obtain the information. If this member is null, all associations for the specified Directory ID are returned.

An empty list results in an InvalidParameterException being thrown.