Struct rusoto::ds::DescribeSnapshotsRequest

pub struct DescribeSnapshotsRequest {
    pub directory_id: Option<DirectoryId>,
    pub limit: Option<Limit>,
    pub next_token: Option<NextToken>,
    pub snapshot_ids: Option<SnapshotIds>,
}

Contains the inputs for the DescribeSnapshots operation.

Fields
directory_id: Option<DirectoryId>

The identifier of the directory for which to retrieve snapshot information.

limit: Option<Limit>

The maximum number of objects to return.

next_token: Option<NextToken>

The DescribeSnapshotsResult.NextToken value from a previous call to DescribeSnapshots. Pass null if this is the first call.

snapshot_ids: Option<SnapshotIds>

A list of identifiers of the snapshots to obtain the information for. If this member is null or empty, all snapshots are returned using the Limit and NextToken members.