Struct rusoto::ds::DescribeTrustsRequest

pub struct DescribeTrustsRequest {
    pub directory_id: Option<DirectoryId>,
    pub limit: Option<Limit>,
    pub next_token: Option<NextToken>,
    pub trust_ids: Option<TrustIds>,
}

Describes the trust relationships for a particular Microsoft AD in the AWS cloud. If no input parameters are are provided, such as directory ID or trust ID, this request describes all the trust relationships.

Fields
directory_id: Option<DirectoryId>

The Directory ID of the AWS directory that is a part of the requested trust relationship.

limit: Option<Limit>

The maximum number of objects to return.

next_token: Option<NextToken>

The DescribeTrustsResult.NextToken value from a previous call to DescribeTrusts. Pass null if this is the first call.

trust_ids: Option<TrustIds>

A list of identifiers of the trust relationships for which to obtain the information. If this member is null, all trust relationships that belong to the current account are returned.

An empty list results in an InvalidParameterException being thrown.