Struct rusoto::ds::DirectoryDescription

pub struct DirectoryDescription {
    pub access_url: Option<AccessUrl>,
    pub alias: Option<AliasName>,
    pub connect_settings: Option<DirectoryConnectSettingsDescription>,
    pub description: Option<Description>,
    pub directory_id: Option<DirectoryId>,
    pub dns_ip_addrs: Option<DnsIpAddrs>,
    pub launch_time: Option<LaunchTime>,
    pub name: Option<DirectoryName>,
    pub radius_settings: Option<RadiusSettings>,
    pub radius_status: Option<RadiusStatus>,
    pub short_name: Option<DirectoryShortName>,
    pub size: Option<DirectorySize>,
    pub sso_enabled: Option<SsoEnabled>,
    pub stage: Option<DirectoryStage>,
    pub stage_last_updated_date_time: Option<LastUpdatedDateTime>,
    pub stage_reason: Option<StageReason>,
    pub type_: Option<DirectoryType>,
    pub vpc_settings: Option<DirectoryVpcSettingsDescription>,
}

Contains information about an AWS Directory Service directory.

Fields

access_url

The access URL for the directory, such as http://alias.awsapps.com. If no alias has been created for the directory, alias is the directory identifier, such as d-XXXXXXXXXX.

alias

The alias for the directory. If no alias has been created for the directory, the alias is the directory identifier, such as d-XXXXXXXXXX.

connect_settings

A DirectoryConnectSettingsDescription object that contains additional information about an AD Connector directory. This member is only present if the directory is an AD Connector directory.

description

The textual description for the directory.

directory_id

The directory identifier.

dns_ip_addrs

The IP addresses of the DNS servers for the directory. For a Simple AD or Microsoft AD directory, these are the IP addresses of the Simple AD or Microsoft AD directory servers. For an AD Connector directory, these are the IP addresses of the DNS servers or domain controllers in the on-premises directory to which the AD Connector is connected.

launch_time

Specifies when the directory was created.

name

The fully-qualified name of the directory.

radius_settings

A RadiusSettings object that contains information about the RADIUS server configured for this directory.

radius_status

The status of the RADIUS MFA server connection.

short_name

The short name of the directory.

size

The directory size.

sso_enabled

Indicates if single-sign on is enabled for the directory. For more information, see EnableSso and DisableSso.

stage

The current stage of the directory.

stage_last_updated_date_time

The date and time that the stage was last updated.

stage_reason

Additional information about the directory stage.

type_

The directory size.

vpc_settings

A DirectoryVpcSettingsDescription object that contains additional information about a directory. This member is only present if the directory is a Simple AD or Managed AD directory.