Struct rusoto::dynamodb::AttributeValue

pub struct AttributeValue {
    pub b: BinaryAttributeValue,
    pub null: NullAttributeValue,
    pub ss: StringSetAttributeValue,
    pub m: MapAttributeValue,
    pub l: ListAttributeValue,
    pub n: NumberAttributeValue,
    pub s: StringAttributeValue,
    pub bool: BooleanAttributeValue,
    pub bs: BinarySetAttributeValue,
    pub ns: NumberSetAttributeValue,
}

Represents the data for an attribute. You can set one, and only one, of the elements. Each attribute in an item is a name-value pair. An attribute can be single- valued or multi-valued set. For example, a book item can have title and authors attributes. Each book has one title but can have many authors. The multi-valued attribute is a set; duplicate values are not allowed.

Fields

b

A Binary data type.

null

A Null data type.

ss

A String Set data type.

m

A Map of attribute values.

l

A List of attribute values.

n

A Number data type.

s

A String data type.

bool

A Boolean data type.

bs

A Binary Set data type.

ns

A Number Set data type.