Struct rusoto::dynamodb::ConsumedCapacity

pub struct ConsumedCapacity {
    pub capacity_units: Option<ConsumedCapacityUnits>,
    pub global_secondary_indexes: Option<SecondaryIndexesCapacityMap>,
    pub local_secondary_indexes: Option<SecondaryIndexesCapacityMap>,
    pub table: Option<Capacity>,
    pub table_name: Option<TableName>,
}

The capacity units consumed by an operation. The data returned includes the total provisioned throughput consumed, along with statistics for the table and any indexes involved in the operation. ConsumedCapacity is only returned if the request asked for it. For more information, see Provisioned Throughput in the Amazon DynamoDB Developer Guide.

Fields
capacity_units: Option<ConsumedCapacityUnits>

The total number of capacity units consumed by the operation.

global_secondary_indexes: Option<SecondaryIndexesCapacityMap>

The amount of throughput consumed on each global index affected by the operation.

local_secondary_indexes: Option<SecondaryIndexesCapacityMap>

The amount of throughput consumed on each local index affected by the operation.

table: Option<Capacity>

The amount of throughput consumed on the table affected by the operation.

table_name: Option<TableName>

The name of the table that was affected by the operation.