Struct rusoto::dynamodb::GlobalSecondaryIndex

pub struct GlobalSecondaryIndex {
    pub key_schema: KeySchema,
    pub index_name: IndexName,
    pub projection: Projection,
    pub provisioned_throughput: ProvisionedThroughput,
}

Represents the properties of a global secondary index.

Fields

key_schema

The complete key schema for a global secondary index, which consists of one or more pairs of attribute names and key types (HASH or RANGE).

index_name

The name of the global secondary index. The name must be unique among all other indexes on this table.

projection

provisioned_throughput