Struct rusoto::dynamodb::GlobalSecondaryIndexDescription

pub struct GlobalSecondaryIndexDescription {
    pub index_size_bytes: Long,
    pub index_name: IndexName,
    pub projection: Projection,
    pub provisioned_throughput: ProvisionedThroughputDescription,
    pub index_status: IndexStatus,
    pub backfilling: Backfilling,
    pub key_schema: KeySchema,
    pub index_arn: String,
    pub item_count: Long,
}

Represents the properties of a global secondary index.

Fields

index_size_bytes

The total size of the specified index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

index_name

The name of the global secondary index.

projection

provisioned_throughput

index_status

The current state of the global secondary index: * CREATING - The index is being created. * UPDATING - The index is being updated. * DELETING - The index is being deleted. * ACTIVE - The index is ready for use.

backfilling

Indicates whether the index is currently backfilling. Backfilling is the process of reading items from the table and determining whether they can be added to the index. (Not all items will qualify: For example, a hash key attribute cannot have any duplicates.) If an item can be added to the index, DynamoDB will do so. After all items have been processed, the backfilling operation is complete and Backfilling is false. For indexes that were created during a CreateTable operation, the Backfilling attribute does not appear in the DescribeTable output.

key_schema

The complete key schema for the global secondary index, consisting of one or more pairs of attribute names and key types (HASH or RANGE).

index_arn

The Amazon Resource Name (ARN) that uniquely identifies the index.

item_count

The number of items in the specified index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.