Struct rusoto::dynamodb::GlobalSecondaryIndexUpdate

pub struct GlobalSecondaryIndexUpdate {
    pub Create: Option<CreateGlobalSecondaryIndexAction>,
    pub Update: Option<UpdateGlobalSecondaryIndexAction>,
    pub Delete: Option<DeleteGlobalSecondaryIndexAction>,
}

Represents one of the following: * A new global secondary index to be added to an existing table. * New provisioned throughput parameters for an existing global secondary index. * An existing global secondary index to be removed from an existing table.

Fields

Create

The parameters required for creating a global secondary index on an existing table: * IndexName * KeySchema * AttributeDefinitions * Projection * ProvisionedThroughput

Update

The name of an existing global secondary index, along with new provisioned throughput settings to be applied to that index.

Delete

The name of an existing global secondary index to be removed.