Struct rusoto::dynamodb::ItemCollectionMetrics

pub struct ItemCollectionMetrics {
    pub item_collection_key: ItemCollectionKeyAttributeMap,
    pub size_estimate_range_gb: ItemCollectionSizeEstimateRange,
}

Information about item collections, if any, that were affected by the operation. ItemCollectionMetrics is only returned if the request asked for it. If the table does not have any local secondary indexes, this information is not returned in the response.

Fields

item_collection_key

The hash key value of the item collection. This value is the same as the hash key of the item.

size_estimate_range_gb

An estimate of item collection size, in gigabytes. This value is a two-element array containing a lower bound and an upper bound for the estimate. The estimate includes the size of all the items in the table, plus the size of all attributes projected into all of the local secondary indexes on that table. Use this estimate to measure whether a local secondary index is approaching its size limit. The estimate is subject to change over time; therefore, do not rely on the precision or accuracy of the estimate.