Struct rusoto::dynamodb::LimitExceededException

pub struct LimitExceededException {
    pub message: ErrorMessage,
}

The number of concurrent table requests (cumulative number of tables in the CREATING, DELETING or UPDATING state) exceeds the maximum allowed of 10. Also, for tables with secondary indexes, only one of those tables can be in the CREATING state at any point in time. Do not attempt to create more than one such table simultaneously. The total limit of tables in the ACTIVE state is 250.

Fields

message

Too many operations for a given subscriber.