Struct rusoto::dynamodb::LocalSecondaryIndex

pub struct LocalSecondaryIndex {
    pub KeySchema: KeySchema,
    pub IndexName: IndexName,
    pub Projection: Projection,
}

Represents the properties of a local secondary index.

Fields

KeySchema

The complete key schema for the local secondary index, consisting of one or more pairs of attribute names and key types (HASH or RANGE).

IndexName

The name of the local secondary index. The name must be unique among all other indexes on this table.

Projection