Struct rusoto::dynamodb::Projection

pub struct Projection {
    pub non_key_attributes: Option<NonKeyAttributeNameList>,
    pub projection_type: Option<ProjectionType>,
}

Represents attributes that are copied (projected) from the table into an index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

Fields
non_key_attributes: Option<NonKeyAttributeNameList>

Represents the non-key attribute names which will be projected into the index.

For local secondary indexes, the total count of NonKeyAttributes summed across all of the local secondary indexes, must not exceed 20. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

projection_type: Option<ProjectionType>

The set of attributes that are projected into the index:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes are in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.