Struct rusoto::dynamodb::ProvisionedThroughput

pub struct ProvisionedThroughput {
    pub write_capacity_units: PositiveLongObject,
    pub read_capacity_units: PositiveLongObject,
}

Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation. For current minimum and maximum provisioned throughput values, see Limits in the Amazon DynamoDB Developer Guide.

Fields

write_capacity_units

The maximum number of writes consumed per second before DynamoDB returns a ThrottlingException. For more information, see Specifying Read and Write Re quirements in the Amazon DynamoDB Developer Guide.

read_capacity_units

The maximum number of strongly consistent reads consumed per second before DynamoDB returns a ThrottlingException. For more information, see Specifying Read and Write Requirements in the Amazon DynamoDB Developer Guide.