Struct rusoto::dynamodb::ProvisionedThroughputDescription

pub struct ProvisionedThroughputDescription {
    pub NumberOfDecreasesToday: Option<PositiveLongObject>,
    pub WriteCapacityUnits: Option<PositiveLongObject>,
    pub LastIncreaseDateTime: Option<Date>,
    pub ReadCapacityUnits: Option<PositiveLongObject>,
    pub LastDecreaseDateTime: Option<Date>,
}

Represents the provisioned throughput settings for the table, consisting of read and write capacity units, along with data about increases and decreases.

Fields

NumberOfDecreasesToday

The number of provisioned throughput decreases for this table during this UTC calendar day. For current maximums on provisioned throughput decreases, see L imits in the Amazon DynamoDB Developer Guide.

WriteCapacityUnits

The maximum number of writes consumed per second before DynamoDB returns a ThrottlingException.

LastIncreaseDateTime

The date and time of the last provisioned throughput increase for this table.

ReadCapacityUnits

The maximum number of strongly consistent reads consumed per second before DynamoDB returns a ThrottlingException. Eventually consistent reads require less effort than strongly consistent reads, so a setting of 50 ReadCapacityUnits per second provides 100 eventually consistent ReadCapacityUnits per second.

LastDecreaseDateTime

The date and time of the last provisioned throughput decrease for this table.