Struct rusoto::dynamodb::ProvisionedThroughputExceededException

pub struct ProvisionedThroughputExceededException {
    pub message: ErrorMessage,
}

Your request rate is too high. The AWS SDKs for DynamoDB automatically retry requests that receive this exception. Your request is eventually successful, unless your retry queue is too large to finish. Reduce the frequency of requests and use exponential backoff. For more information, go to Error Retries and Exponential Backoff in the Amazon DynamoDB Developer Guide.

Fields

message

You exceeded your maximum allowed provisioned throughput.