Struct rusoto::dynamodb::PutItemOutput

pub struct PutItemOutput {
    pub Attributes: Option<AttributeMap>,
    pub ItemCollectionMetrics: Option<ItemCollectionMetrics>,
    pub ConsumedCapacity: Option<ConsumedCapacity>,
}

Represents the output of a PutItem operation.

Fields

Attributes

The attribute values as they appeared before the PutItem operation, but only if ReturnValues is specified as ALL_OLD in the request. Each element consists of an attribute name and an attribute value.

ItemCollectionMetrics

Information about item collections, if any, that were affected by the operation. ItemCollectionMetrics is only returned if the request asked for it. If the table does not have any local secondary indexes, this information is not returned in the response. Each ItemCollectionMetrics element consists of: * ItemCollectionKey - The hash key value of the item collection. This is the same as the hash key of the item. * SizeEstimateRange - An estimate of item collection size, in gigabytes. This value is a two-element array containing a lower bound and an upper bound for the estimate. The estimate includes the size of all the items in the table, plus the size of all attributes projected into all of the local secondary indexes on that table. Use this estimate to measure whether a local secondary index is approaching its size limit. The estimate is subject to change over time; therefore, do not rely on the precision or accuracy of the estimate.

ConsumedCapacity