Struct rusoto::dynamodb::ScanOutput

pub struct ScanOutput {
    pub consumed_capacity: Option<ConsumedCapacity>,
    pub count: Option<Integer>,
    pub items: Option<ItemList>,
    pub last_evaluated_key: Option<Key>,
    pub scanned_count: Option<Integer>,
}

Represents the output of a Scan operation.

Fields
consumed_capacity: Option<ConsumedCapacity> count: Option<Integer>

The number of items in the response.

If you set ScanFilter in the request, then Count is the number of items returned after the filter was applied, and ScannedCount is the number of matching items before the filter was applied.

If you did not use a filter in the request, then Count is the same as ScannedCount.

items: Option<ItemList>

An array of item attributes that match the scan criteria. Each element in this array consists of an attribute name and the value for that attribute.

last_evaluated_key: Option<Key>

The primary key of the item where the operation stopped, inclusive of the previous result set. Use this value to start a new operation, excluding this value in the new request.

If LastEvaluatedKey is empty, then the "last page" of results has been processed and there is no more data to be retrieved.

If LastEvaluatedKey is not empty, it does not necessarily mean that there is more data in the result set. The only way to know when you have reached the end of the result set is when LastEvaluatedKey is empty.

scanned_count: Option<Integer>

The number of items evaluated, before any ScanFilter is applied. A high ScannedCount value with few, or no, Count results indicates an inefficient Scan operation. For more information, see Count and ScannedCount in the Amazon DynamoDB Developer Guide.

If you did not use a filter in the request, then ScannedCount is the same as Count.