Struct rusoto::dynamodb::StreamSpecification

pub struct StreamSpecification {
    pub stream_view_type: StreamViewType,
    pub stream_enabled: StreamEnabled,
}

Represents the DynamoDB Streams configuration for a table in DynamoDB.

Fields

stream_view_type

The DynamoDB Streams settings for the table. These settings consist of: * StreamEnabled - Indicates whether DynamoDB Streams is enabled (true) or disabled (false) on the table. * StreamViewType - When an item in the table is modified, StreamViewType determines what information is written to the stream for this table. Valid values for StreamViewType are: * KEYS_ONLY - Only the key attributes of the modified item are written to the stream. * NEW_IMAGE - The entire item, as it appears after it was modified, is written to the stream. * OLD_IMAGE - The entire item, as it appeared before it was modified, is written to the stream. * NEW_AND_OLD_IMAGES - Both the new and the old item images of the item are written to the stream.

stream_enabled

Indicates whether DynamoDB Streams is enabled (true) or disabled (false) on the table.