Struct rusoto::dynamodb::TableDescription

pub struct TableDescription {
    pub TableArn: Option<String>,
    pub LocalSecondaryIndexes: Option<LocalSecondaryIndexDescriptionList>,
    pub AttributeDefinitions: Option<AttributeDefinitions>,
    pub GlobalSecondaryIndexes: Option<GlobalSecondaryIndexDescriptionList>,
    pub ProvisionedThroughput: Option<ProvisionedThroughputDescription>,
    pub TableSizeBytes: Option<Long>,
    pub TableName: Option<TableName>,
    pub TableStatus: Option<TableStatus>,
    pub StreamSpecification: Option<StreamSpecification>,
    pub LatestStreamLabel: Option<String>,
    pub KeySchema: Option<KeySchema>,
    pub ItemCount: Option<Long>,
    pub CreationDateTime: Option<Date>,
    pub LatestStreamArn: Option<StreamArn>,
}

Represents the properties of a table.

Fields

TableArn

The Amazon Resource Name (ARN) that uniquely identifies the table.

LocalSecondaryIndexes

Represents one or more local secondary indexes on the table. Each index is scoped to a given hash key value. Tables with one or more local secondary indexes are subject to an item collection size limit, where the amount of data within a given item collection cannot exceed 10 GB. Each element is composed of: * IndexName - The name of the local secondary index. * KeySchema - Specifies the complete index key schema. The attribute names in the key schema must be between 1 and 255 characters (inclusive). The key schema must begin with the same hash key attribute as the table. * Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of: * ProjectionType - One of the following: * KEYS_ONLY - Only the index and primary keys are projected into the index. * INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes are in NonKeyAttributes. * ALL - All of the table attributes are projected into the index. * NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 20. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total. * IndexSizeBytes - Represents the total size of the index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value. * ItemCount - Represents the number of items in the index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value. If the table is in the DELETING state, no information about indexes will be returned.

AttributeDefinitions

An array of AttributeDefinition objects. Each of these objects describes one attribute in the table and index key schema. Each AttributeDefinition object in this array is composed of: * AttributeName - The name of the attribute. * AttributeType - The data type for the attribute.

GlobalSecondaryIndexes

The global secondary indexes, if any, on the table. Each index is scoped to a given hash key value. Each element is composed of: * Backfilling - If true, then the index is currently in the backfilling phase. Backfilling occurs only when a new global secondary index is added to the table; it is the process by which DynamoDB populates the new index with data from the table. (This attribute does not appear for indexes that were created during a CreateTable operation.) * IndexName - The name of the global secondary index. * IndexSizeBytes - The total size of the global secondary index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value. * IndexStatus - The current status of the global secondary index: * CREATING - The index is being created. * UPDATING - The index is being updated. * DELETING - The index is being deleted. * ACTIVE - The index is ready for use. * ItemCount - The number of items in the global secondary index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value. * KeySchema - Specifies the complete index key schema. The attribute names in the key schema must be between 1 and 255 characters (inclusive). The key schema must begin with the same hash key attribute as the table. * Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of: * ProjectionType - One of the following: * KEYS_ONLY - Only the index and primary keys are projected into the index. * INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes are in NonKeyAttributes. * ALL - All of the table attributes are projected into the index. * NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 20. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total. * ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units, along with data about increases and decreases. If the table is in the DELETING state, no information about indexes will be returned.

ProvisionedThroughput

The provisioned throughput settings for the table, consisting of read and write capacity units, along with data about increases and decreases.

TableSizeBytes

The total size of the specified table, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

TableName

The name of the table.

TableStatus

The current state of the table: * CREATING - The table is being created. * UPDATING - The table is being updated. * DELETING - The table is being deleted. * ACTIVE - The table is ready for use.

StreamSpecification

The current DynamoDB Streams configuration for the table.

LatestStreamLabel

A timestamp, in ISO 8601 format, for this stream. Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique: * the AWS customer ID. * the table name. * the StreamLabel.

KeySchema

The primary key structure for the table. Each KeySchemaElement consists of: * AttributeName - The name of the attribute. * KeyType - The key type for the attribute. Can be either HASH or RANGE. For more information about primary keys, see Primary Key in the Amazon DynamoDB Developer Guide.

ItemCount

The number of items in the specified table. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

CreationDateTime

The date and time when the table was created, in UNIX epoch time format.

LatestStreamArn

The Amazon Resource Name (ARN) that uniquely identifies the latest stream for this table.