Struct rusoto::dynamodb::UpdateTableInput

pub struct UpdateTableInput {
    pub GlobalSecondaryIndexUpdates: Option<GlobalSecondaryIndexUpdateList>,
    pub ProvisionedThroughput: Option<ProvisionedThroughput>,
    pub StreamSpecification: Option<StreamSpecification>,
    pub TableName: TableName,
    pub AttributeDefinitions: Option<AttributeDefinitions>,
}

Represents the input of an UpdateTable operation.

Fields

GlobalSecondaryIndexUpdates

An array of one or more global secondary indexes for the table. For each index in the array, you can request one action: * Create - add a new global secondary index to the table. * Update - modify the provisioned throughput settings of an existing global secondary index. * Delete - remove a global secondary index from the table. For more information, see Managing Global Secondary Indexes in the Amazon DynamoDB Developer Guide.

ProvisionedThroughput

StreamSpecification

Represents the DynamoDB Streams configuration for the table. You will receive a ResourceInUseException if you attempt to enable a stream on a table that already has a stream, or if you attempt to disable a stream on a table which does not have a stream.

TableName

The name of the table to be updated.

AttributeDefinitions

An array of attributes that describe the key schema for the table and indexes. If you are adding a new global secondary index to the table, AttributeDefinitions must include the key element(s) of the new index.