Struct rusoto::dynamodb::WriteRequest

pub struct WriteRequest {
    pub put_request: PutRequest,
    pub delete_request: DeleteRequest,
}

Represents an operation to perform - either DeleteItem or PutItem. You can only request one of these operations, not both, in a single WriteRequest. If you do need to perform both of these operations, you will need to provide two separate WriteRequest objects.

Fields

put_request

A request to perform a PutItem operation.

delete_request

A request to perform a DeleteItem operation.