Module rusoto::dynamodbstreams

AWS DynamoDB Streams

If you're using the service, you're probably looking for DynamoDbStreamsClient.

Structs

AttributeValue
Represents the data for an attribute. You can set one, and only one, of the elements.

Each attribute in an item is a name-value pair. An attribute can be single-valued or multi-valued set. For example, a book item can have title and authors attributes. Each book has one title but can have many authors. The multi-valued attribute is a set; duplicate values are not allowed.

DescribeStreamInput
Represents the input of a DescribeStream operation.

DescribeStreamOutput
Represents the output of a DescribeStream operation.

DynamoDbStreamsClient
A client for the Amazon DynamoDB Streams API.

GetRecordsInput
Represents the input of a GetRecords operation.

GetRecordsOutput
Represents the output of a GetRecords operation.

GetShardIteratorInput
Represents the input of a GetShardIterator operation.

GetShardIteratorOutput
Represents the output of a GetShardIterator operation.

KeySchemaElement

Represents a single element of a key schema. A key schema specifies the attributes that make up the primary key of a table, or the key attributes of an index.

A KeySchemaElement represents exactly one attribute of the primary key. For example, a simple primary key (partition key) would be represented by one KeySchemaElement. A composite primary key (partition key and sort key) would require one KeySchemaElement for the partition key, and another KeySchemaElement for the sort key.

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

ListStreamsInput
Represents the input of a ListStreams operation.

ListStreamsOutput
Represents the output of a ListStreams operation.

Record
A description of a unique event within a stream.

SequenceNumberRange
The beginning and ending sequence numbers for the stream records contained within a shard.

Shard
A uniquely identified group of stream records within a stream.

Stream
Represents all of the data describing a particular stream.

StreamDescription
Represents all of the data describing a particular stream.

StreamRecord
A description of a single data modification that was performed on an item in a DynamoDB table.

Enums

DescribeStreamError
Errors returned by DescribeStream

GetRecordsError
Errors returned by GetRecords

GetShardIteratorError
Errors returned by GetShardIterator

ListStreamsError
Errors returned by ListStreams

Type Definitions

AttributeMap

AttributeName

BinaryAttributeValue

BinarySetAttributeValue

BooleanAttributeValue

Date

ErrorMessage

KeySchema

KeySchemaAttributeName

KeyType

ListAttributeValue

MapAttributeValue

NullAttributeValue

NumberAttributeValue

NumberSetAttributeValue

OperationType

PositiveIntegerObject

PositiveLongObject

RecordList

SequenceNumber

ShardDescriptionList

ShardId

ShardIterator

ShardIteratorType

StreamArn

StreamList

StreamStatus

StreamViewType

StringAttributeValue

StringSetAttributeValue

TableName