Struct rusoto::dynamodbstreams::DynamoDbStreamsClient

pub struct DynamoDbStreamsClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the Amazon DynamoDB Streams API.

Methods

impl<P> DynamoDbStreamsClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> DynamoDbStreamsClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn describe_stream(&self, input: &DescribeStreamInput) -> Result<DescribeStreamOutput, DescribeStreamError>

Returns information about a stream, including the current status of the stream, its Amazon Resource Name (ARN), the composition of its shards, and its corresponding DynamoDB table.

You can call DescribeStream at a maximum rate of 10 times per second.

Each shard in the stream has a SequenceNumberRange associated with it. If the SequenceNumberRange has a StartingSequenceNumber but no EndingSequenceNumber, then the shard is still open (able to receive more stream records). If both StartingSequenceNumber and EndingSequenceNumber are present, the that shared is closed and can no longer receive more data.

fn get_records(&self, input: &GetRecordsInput) -> Result<GetRecordsOutput, GetRecordsError>

Retrieves the stream records from a given shard.

Specify a shard iterator using the ShardIterator parameter. The shard iterator specifies the position in the shard from which you want to start reading stream records sequentially. If there are no stream records available in the portion of the shard that the iterator points to, GetRecords returns an empty list. Note that it might take multiple calls to get to a portion of the shard that contains stream records.

GetRecords can retrieve a maximum of 1 MB of data or 2000 stream records, whichever comes first.

fn get_shard_iterator(&self, input: &GetShardIteratorInput) -> Result<GetShardIteratorOutput, GetShardIteratorError>

Returns a shard iterator. A shard iterator provides information about how to retrieve the stream records from within a shard. Use the shard iterator in a subsequent GetRecords request to read the stream records from the shard.

A shard iterator expires 15 minutes after it is returned to the requester.

fn list_streams(&self, input: &ListStreamsInput) -> Result<ListStreamsOutput, ListStreamsError>

Returns an array of stream ARNs associated with the current account and endpoint. If the TableName parameter is present, then ListStreams will return only the streams ARNs for that table.

You can call ListStreams at a maximum rate of 5 times per second.