Struct rusoto::dynamodbstreams::GetRecordsOutput

pub struct GetRecordsOutput {
    pub next_shard_iterator: Option<ShardIterator>,
    pub records: Option<RecordList>,
}

Represents the output of a GetRecords operation.

Fields
next_shard_iterator: Option<ShardIterator>

The next position in the shard from which to start sequentially reading stream records. If set to null, the shard has been closed and the requested iterator will not return any more data.

records: Option<RecordList>

The stream records from the shard, which were retrieved using the shard iterator.