Struct rusoto::dynamodbstreams::StreamRecord

pub struct StreamRecord {
    pub keys: Option<AttributeMap>,
    pub new_image: Option<AttributeMap>,
    pub old_image: Option<AttributeMap>,
    pub sequence_number: Option<SequenceNumber>,
    pub size_bytes: Option<PositiveLongObject>,
    pub stream_view_type: Option<StreamViewType>,
}

A description of a single data modification that was performed on an item in a DynamoDB table.

Fields
keys: Option<AttributeMap>

The primary key attribute(s) for the DynamoDB item that was modified.

new_image: Option<AttributeMap>

The item in the DynamoDB table as it appeared after it was modified.

old_image: Option<AttributeMap>

The item in the DynamoDB table as it appeared before it was modified.

sequence_number: Option<SequenceNumber>

The sequence number of the stream record.

size_bytes: Option<PositiveLongObject>

The size of the stream record, in bytes.

stream_view_type: Option<StreamViewType>

The type of data from the modified DynamoDB item that was captured in this stream record:

KEYS_ONLY - only the key attributes of the modified item.

NEW_IMAGE - the entire item, as it appears after it was modified.

OLD_IMAGE - the entire item, as it appeared before it was modified.

NEW_AND_OLD_IMAGES — both the new and the old item images of the item.