Module rusoto::ec2

Amazon EC2 Container Service

Structs

AcceptVpcPeeringConnectionRequest
Contains the parameters for AcceptVpcPeeringConnection.

AcceptVpcPeeringConnectionResult
Contains the output of AcceptVpcPeeringConnection.

AccountAttribute
Describes an account attribute.

AccountAttributeValue
Describes a value of an account attribute.

ActiveInstance
Describes a running instance in a Spot fleet.

Address
Describes an Elastic IP address.

AllocateAddressRequest
Contains the parameters for AllocateAddress.

AllocateAddressResult
Contains the output of AllocateAddress.

AllocateHostsRequest
Contains the parameters for AllocateHosts.

AllocateHostsResult
Contains the output of AllocateHosts.

AssignPrivateIpAddressesRequest
Contains the parameters for AssignPrivateIpAddresses.

AssociateAddressRequest
Contains the parameters for AssociateAddress.

AssociateAddressResult
Contains the output of AssociateAddress.

AssociateDhcpOptionsRequest
Contains the parameters for AssociateDhcpOptions.

AssociateRouteTableRequest
Contains the parameters for AssociateRouteTable.

AssociateRouteTableResult
Contains the output of AssociateRouteTable.

AttachClassicLinkVpcRequest
Contains the parameters for AttachClassicLinkVpc.

AttachClassicLinkVpcResult
Contains the output of AttachClassicLinkVpc.

AttachInternetGatewayRequest
Contains the parameters for AttachInternetGateway.

AttachNetworkInterfaceRequest
Contains the parameters for AttachNetworkInterface.

AttachNetworkInterfaceResult
Contains the output of AttachNetworkInterface.

AttachVolumeRequest
Contains the parameters for AttachVolume.

AttachVpnGatewayRequest
Contains the parameters for AttachVpnGateway.

AttachVpnGatewayResult
Contains the output of AttachVpnGateway.

AttributeBooleanValue
The value to use when a resource attribute accepts a Boolean value.

AttributeValue
The value to use for a resource attribute.

AuthorizeSecurityGroupEgressRequest
Contains the parameters for AuthorizeSecurityGroupEgress.

AuthorizeSecurityGroupIngressRequest
Contains the parameters for AuthorizeSecurityGroupIngress.

AvailabilityZone
Describes an Availability Zone.

AvailabilityZoneMessage
Describes a message about an Availability Zone.

AvailableCapacity
The capacity information for instances launched onto the Dedicated host.

BlobAttributeValue

BlockDeviceMapping
Describes a block device mapping.

BundleInstanceRequest
Contains the parameters for BundleInstance.

BundleInstanceResult
Contains the output of BundleInstance.

BundleTask
Describes a bundle task.

BundleTaskError
Describes an error for BundleInstance.

CancelBundleTaskRequest
Contains the parameters for CancelBundleTask.

CancelBundleTaskResult
Contains the output of CancelBundleTask.

CancelConversionRequest
Contains the parameters for CancelConversionTask.

CancelExportTaskRequest
Contains the parameters for CancelExportTask.

CancelImportTaskRequest
Contains the parameters for CancelImportTask.

CancelImportTaskResult
Contains the output for CancelImportTask.

CancelReservedInstancesListingRequest
Contains the parameters for CancelReservedInstancesListing.

CancelReservedInstancesListingResult
Contains the output of CancelReservedInstancesListing.

CancelSpotFleetRequestsError
Describes a Spot fleet error.

CancelSpotFleetRequestsErrorItem
Describes a Spot fleet request that was not successfully canceled.

CancelSpotFleetRequestsRequest
Contains the parameters for CancelSpotFleetRequests.

CancelSpotFleetRequestsResponse
Contains the output of CancelSpotFleetRequests.

CancelSpotFleetRequestsSuccessItem
Describes a Spot fleet request that was successfully canceled.

CancelSpotInstanceRequestsRequest
Contains the parameters for CancelSpotInstanceRequests.

CancelSpotInstanceRequestsResult
Contains the output of CancelSpotInstanceRequests.

CancelledSpotInstanceRequest
Describes a request to cancel a Spot instance.

ClassicLinkDnsSupport
Describes the ClassicLink DNS support status of a VPC.

ClassicLinkInstance
Describes a linked EC2-Classic instance.

ClientData
Describes the client-specific data.

ConfirmProductInstanceRequest
Contains the parameters for ConfirmProductInstance.

ConfirmProductInstanceResult
Contains the output of ConfirmProductInstance.

ConversionTask
Describes a conversion task.

CopyImageRequest
Contains the parameters for CopyImage.

CopyImageResult
Contains the output of CopyImage.

CopySnapshotRequest
Contains the parameters for CopySnapshot.

CopySnapshotResult
Contains the output of CopySnapshot.

CreateCustomerGatewayRequest
Contains the parameters for CreateCustomerGateway.

CreateCustomerGatewayResult
Contains the output of CreateCustomerGateway.

CreateDhcpOptionsRequest
Contains the parameters for CreateDhcpOptions.

CreateDhcpOptionsResult
Contains the output of CreateDhcpOptions.

CreateFlowLogsRequest
Contains the parameters for CreateFlowLogs.

CreateFlowLogsResult
Contains the output of CreateFlowLogs.

CreateImageRequest
Contains the parameters for CreateImage.

CreateImageResult
Contains the output of CreateImage.

CreateInstanceExportTaskRequest
Contains the parameters for CreateInstanceExportTask.

CreateInstanceExportTaskResult
Contains the output for CreateInstanceExportTask.

CreateInternetGatewayRequest
Contains the parameters for CreateInternetGateway.

CreateInternetGatewayResult
Contains the output of CreateInternetGateway.

CreateKeyPairRequest
Contains the parameters for CreateKeyPair.

CreateNatGatewayRequest
Contains the parameters for CreateNatGateway.

CreateNatGatewayResult
Contains the output of CreateNatGateway.

CreateNetworkAclEntryRequest
Contains the parameters for CreateNetworkAclEntry.

CreateNetworkAclRequest
Contains the parameters for CreateNetworkAcl.

CreateNetworkAclResult
Contains the output of CreateNetworkAcl.

CreateNetworkInterfaceRequest
Contains the parameters for CreateNetworkInterface.

CreateNetworkInterfaceResult
Contains the output of CreateNetworkInterface.

CreatePlacementGroupRequest
Contains the parameters for CreatePlacementGroup.

CreateReservedInstancesListingRequest
Contains the parameters for CreateReservedInstancesListing.

CreateReservedInstancesListingResult
Contains the output of CreateReservedInstancesListing.

CreateRouteRequest
Contains the parameters for CreateRoute.

CreateRouteResult
Contains the output of CreateRoute.

CreateRouteTableRequest
Contains the parameters for CreateRouteTable.

CreateRouteTableResult
Contains the output of CreateRouteTable.

CreateSecurityGroupRequest
Contains the parameters for CreateSecurityGroup.

CreateSecurityGroupResult
Contains the output of CreateSecurityGroup.

CreateSnapshotRequest
Contains the parameters for CreateSnapshot.

CreateSpotDatafeedSubscriptionRequest
Contains the parameters for CreateSpotDatafeedSubscription.

CreateSpotDatafeedSubscriptionResult
Contains the output of CreateSpotDatafeedSubscription.

CreateSubnetRequest
Contains the parameters for CreateSubnet.

CreateSubnetResult
Contains the output of CreateSubnet.

CreateTagsRequest
Contains the parameters for CreateTags.

CreateVolumePermission
Describes the user or group to be added or removed from the permissions for a volume.

CreateVolumePermissionModifications
Describes modifications to the permissions for a volume.

CreateVolumeRequest
Contains the parameters for CreateVolume.

CreateVpcEndpointRequest
Contains the parameters for CreateVpcEndpoint.

CreateVpcEndpointResult
Contains the output of CreateVpcEndpoint.

CreateVpcPeeringConnectionRequest
Contains the parameters for CreateVpcPeeringConnection.

CreateVpcPeeringConnectionResult
Contains the output of CreateVpcPeeringConnection.

CreateVpcRequest
Contains the parameters for CreateVpc.

CreateVpcResult
Contains the output of CreateVpc.

CreateVpnConnectionRequest
Contains the parameters for CreateVpnConnection.

CreateVpnConnectionResult
Contains the output of CreateVpnConnection.

CreateVpnConnectionRouteRequest
Contains the parameters for CreateVpnConnectionRoute.

CreateVpnGatewayRequest
Contains the parameters for CreateVpnGateway.

CreateVpnGatewayResult
Contains the output of CreateVpnGateway.

CustomerGateway
Describes a customer gateway.

DeleteCustomerGatewayRequest
Contains the parameters for DeleteCustomerGateway.

DeleteDhcpOptionsRequest
Contains the parameters for DeleteDhcpOptions.

DeleteFlowLogsRequest
Contains the parameters for DeleteFlowLogs.

DeleteFlowLogsResult
Contains the output of DeleteFlowLogs.

DeleteInternetGatewayRequest
Contains the parameters for DeleteInternetGateway.

DeleteKeyPairRequest
Contains the parameters for DeleteKeyPair.

DeleteNatGatewayRequest
Contains the parameters for DeleteNatGateway.

DeleteNatGatewayResult
Contains the output of DeleteNatGateway.

DeleteNetworkAclEntryRequest
Contains the parameters for DeleteNetworkAclEntry.

DeleteNetworkAclRequest
Contains the parameters for DeleteNetworkAcl.

DeleteNetworkInterfaceRequest
Contains the parameters for DeleteNetworkInterface.

DeletePlacementGroupRequest
Contains the parameters for DeletePlacementGroup.

DeleteRouteRequest
Contains the parameters for DeleteRoute.

DeleteRouteTableRequest
Contains the parameters for DeleteRouteTable.

DeleteSecurityGroupRequest
Contains the parameters for DeleteSecurityGroup.

DeleteSnapshotRequest
Contains the parameters for DeleteSnapshot.

DeleteSpotDatafeedSubscriptionRequest
Contains the parameters for DeleteSpotDatafeedSubscription.

DeleteSubnetRequest
Contains the parameters for DeleteSubnet.

DeleteTagsRequest
Contains the parameters for DeleteTags.

DeleteVolumeRequest
Contains the parameters for DeleteVolume.

DeleteVpcEndpointsRequest
Contains the parameters for DeleteVpcEndpoints.

DeleteVpcEndpointsResult
Contains the output of DeleteVpcEndpoints.

DeleteVpcPeeringConnectionRequest
Contains the parameters for DeleteVpcPeeringConnection.

DeleteVpcPeeringConnectionResult
Contains the output of DeleteVpcPeeringConnection.

DeleteVpcRequest
Contains the parameters for DeleteVpc.

DeleteVpnConnectionRequest
Contains the parameters for DeleteVpnConnection.

DeleteVpnConnectionRouteRequest
Contains the parameters for DeleteVpnConnectionRoute.

DeleteVpnGatewayRequest
Contains the parameters for DeleteVpnGateway.

DeregisterImageRequest
Contains the parameters for DeregisterImage.

DescribeAccountAttributesRequest
Contains the parameters for DescribeAccountAttributes.

DescribeAccountAttributesResult
Contains the output of DescribeAccountAttributes.

DescribeAddressesRequest
Contains the parameters for DescribeAddresses.

DescribeAddressesResult
Contains the output of DescribeAddresses.

DescribeAvailabilityZonesRequest
Contains the parameters for DescribeAvailabilityZones.

DescribeAvailabilityZonesResult
Contains the output of DescribeAvailabiltyZones.

DescribeBundleTasksRequest
Contains the parameters for DescribeBundleTasks.

DescribeBundleTasksResult
Contains the output of DescribeBundleTasks.

DescribeClassicLinkInstancesRequest
Contains the parameters for DescribeClassicLinkInstances.

DescribeClassicLinkInstancesResult
Contains the output of DescribeClassicLinkInstances.

DescribeConversionTasksRequest
Contains the parameters for DescribeConversionTasks.

DescribeConversionTasksResult
Contains the output for DescribeConversionTasks.

DescribeCustomerGatewaysRequest
Contains the parameters for DescribeCustomerGateways.

DescribeCustomerGatewaysResult
Contains the output of DescribeCustomerGateways.

DescribeDhcpOptionsRequest
Contains the parameters for DescribeDhcpOptions.

DescribeDhcpOptionsResult
Contains the output of DescribeDhcpOptions.

DescribeExportTasksRequest
Contains the parameters for DescribeExportTasks.

DescribeExportTasksResult
Contains the output for DescribeExportTasks.

DescribeFlowLogsRequest
Contains the parameters for DescribeFlowLogs.

DescribeFlowLogsResult
Contains the output of DescribeFlowLogs.

DescribeHostsRequest
Contains the parameters for DescribeHosts.

DescribeHostsResult
Contains the output of DescribeHosts.

DescribeIdFormatRequest
Contains the parameters for DescribeIdFormat.

DescribeIdFormatResult
Contains the output of DescribeIdFormat.

DescribeImageAttributeRequest
Contains the parameters for DescribeImageAttribute.

DescribeImagesRequest
Contains the parameters for DescribeImages.

DescribeImagesResult
Contains the output of DescribeImages.

DescribeImportImageTasksRequest
Contains the parameters for DescribeImportImageTasks.

DescribeImportImageTasksResult
Contains the output for DescribeImportImageTasks.

DescribeImportSnapshotTasksRequest
Contains the parameters for DescribeImportSnapshotTasks.

DescribeImportSnapshotTasksResult
Contains the output for DescribeImportSnapshotTasks.

DescribeInstanceAttributeRequest
Contains the parameters for DescribeInstanceAttribute.

DescribeInstanceStatusRequest
Contains the parameters for DescribeInstanceStatus.

DescribeInstanceStatusResult
Contains the output of DescribeInstanceStatus.

DescribeInstancesRequest
Contains the parameters for DescribeInstances.

DescribeInstancesResult
Contains the output of DescribeInstances.

DescribeInternetGatewaysRequest
Contains the parameters for DescribeInternetGateways.

DescribeInternetGatewaysResult
Contains the output of DescribeInternetGateways.

DescribeKeyPairsRequest
Contains the parameters for DescribeKeyPairs.

DescribeKeyPairsResult
Contains the output of DescribeKeyPairs.

DescribeMovingAddressesRequest
Contains the parameters for DescribeMovingAddresses.

DescribeMovingAddressesResult
Contains the output of DescribeMovingAddresses.

DescribeNatGatewaysRequest
Contains the parameters for DescribeNatGateways.

DescribeNatGatewaysResult
Contains the output of DescribeNatGateways.

DescribeNetworkAclsRequest
Contains the parameters for DescribeNetworkAcls.

DescribeNetworkAclsResult
Contains the output of DescribeNetworkAcls.

DescribeNetworkInterfaceAttributeRequest
Contains the parameters for DescribeNetworkInterfaceAttribute.

DescribeNetworkInterfaceAttributeResult
Contains the output of DescribeNetworkInterfaceAttribute.

DescribeNetworkInterfacesRequest
Contains the parameters for DescribeNetworkInterfaces.

DescribeNetworkInterfacesResult
Contains the output of DescribeNetworkInterfaces.

DescribePlacementGroupsRequest
Contains the parameters for DescribePlacementGroups.

DescribePlacementGroupsResult
Contains the output of DescribePlacementGroups.

DescribePrefixListsRequest
Contains the parameters for DescribePrefixLists.

DescribePrefixListsResult
Contains the output of DescribePrefixLists.

DescribeRegionsRequest
Contains the parameters for DescribeRegions.

DescribeRegionsResult
Contains the output of DescribeRegions.

DescribeReservedInstancesListingsRequest
Contains the parameters for DescribeReservedInstancesListings.

DescribeReservedInstancesListingsResult
Contains the output of DescribeReservedInstancesListings.

DescribeReservedInstancesModificationsRequest
Contains the parameters for DescribeReservedInstancesModifications.

DescribeReservedInstancesModificationsResult
Contains the output of DescribeReservedInstancesModifications.

DescribeReservedInstancesOfferingsRequest
Contains the parameters for DescribeReservedInstancesOfferings.

DescribeReservedInstancesOfferingsResult
Contains the output of DescribeReservedInstancesOfferings.

DescribeReservedInstancesRequest
Contains the parameters for DescribeReservedInstances.

DescribeReservedInstancesResult
Contains the output for DescribeReservedInstances.

DescribeRouteTablesRequest
Contains the parameters for DescribeRouteTables.

DescribeRouteTablesResult
Contains the output of DescribeRouteTables.

DescribeScheduledInstanceAvailabilityRequest
Contains the parameters for DescribeScheduledInstanceAvailability.

DescribeScheduledInstanceAvailabilityResult
Contains the output of DescribeScheduledInstanceAvailability.

DescribeScheduledInstancesRequest
Contains the parameters for DescribeScheduledInstances.

DescribeScheduledInstancesResult
Contains the output of DescribeScheduledInstances.

DescribeSecurityGroupReferencesRequest

DescribeSecurityGroupReferencesResult

DescribeSecurityGroupsRequest
Contains the parameters for DescribeSecurityGroups.

DescribeSecurityGroupsResult
Contains the output of DescribeSecurityGroups.

DescribeSnapshotAttributeRequest
Contains the parameters for DescribeSnapshotAttribute.

DescribeSnapshotAttributeResult
Contains the output of DescribeSnapshotAttribute.

DescribeSnapshotsRequest
Contains the parameters for DescribeSnapshots.

DescribeSnapshotsResult
Contains the output of DescribeSnapshots.

DescribeSpotDatafeedSubscriptionRequest
Contains the parameters for DescribeSpotDatafeedSubscription.

DescribeSpotDatafeedSubscriptionResult
Contains the output of DescribeSpotDatafeedSubscription.

DescribeSpotFleetInstancesRequest
Contains the parameters for DescribeSpotFleetInstances.

DescribeSpotFleetInstancesResponse
Contains the output of DescribeSpotFleetInstances.

DescribeSpotFleetRequestHistoryRequest
Contains the parameters for DescribeSpotFleetRequestHistory.

DescribeSpotFleetRequestHistoryResponse
Contains the output of DescribeSpotFleetRequestHistory.

DescribeSpotFleetRequestsRequest
Contains the parameters for DescribeSpotFleetRequests.

DescribeSpotFleetRequestsResponse
Contains the output of DescribeSpotFleetRequests.

DescribeSpotInstanceRequestsRequest
Contains the parameters for DescribeSpotInstanceRequests.

DescribeSpotInstanceRequestsResult
Contains the output of DescribeSpotInstanceRequests.

DescribeSpotPriceHistoryRequest
Contains the parameters for DescribeSpotPriceHistory.

DescribeSpotPriceHistoryResult
Contains the output of DescribeSpotPriceHistory.

DescribeStaleSecurityGroupsRequest

DescribeStaleSecurityGroupsResult

DescribeSubnetsRequest
Contains the parameters for DescribeSubnets.

DescribeSubnetsResult
Contains the output of DescribeSubnets.

DescribeTagsRequest
Contains the parameters for DescribeTags.

DescribeTagsResult
Contains the output of DescribeTags.

DescribeVolumeAttributeRequest
Contains the parameters for DescribeVolumeAttribute.

DescribeVolumeAttributeResult
Contains the output of DescribeVolumeAttribute.

DescribeVolumeStatusRequest
Contains the parameters for DescribeVolumeStatus.

DescribeVolumeStatusResult
Contains the output of DescribeVolumeStatus.

DescribeVolumesRequest
Contains the parameters for DescribeVolumes.

DescribeVolumesResult
Contains the output of DescribeVolumes.

DescribeVpcAttributeRequest
Contains the parameters for DescribeVpcAttribute.

DescribeVpcAttributeResult
Contains the output of DescribeVpcAttribute.

DescribeVpcClassicLinkDnsSupportRequest
Contains the parameters for DescribeVpcClassicLinkDnsSupport.

DescribeVpcClassicLinkDnsSupportResult
Contains the output of DescribeVpcClassicLinkDnsSupport.

DescribeVpcClassicLinkRequest
Contains the parameters for DescribeVpcClassicLink.

DescribeVpcClassicLinkResult
Contains the output of DescribeVpcClassicLink.

DescribeVpcEndpointServicesRequest
Contains the parameters for DescribeVpcEndpointServices.

DescribeVpcEndpointServicesResult
Contains the output of DescribeVpcEndpointServices.

DescribeVpcEndpointsRequest
Contains the parameters for DescribeVpcEndpoints.

DescribeVpcEndpointsResult
Contains the output of DescribeVpcEndpoints.

DescribeVpcPeeringConnectionsRequest
Contains the parameters for DescribeVpcPeeringConnections.

DescribeVpcPeeringConnectionsResult
Contains the output of DescribeVpcPeeringConnections.

DescribeVpcsRequest
Contains the parameters for DescribeVpcs.

DescribeVpcsResult
Contains the output of DescribeVpcs.

DescribeVpnConnectionsRequest
Contains the parameters for DescribeVpnConnections.

DescribeVpnConnectionsResult
Contains the output of DescribeVpnConnections.

DescribeVpnGatewaysRequest
Contains the parameters for DescribeVpnGateways.

DescribeVpnGatewaysResult
Contains the output of DescribeVpnGateways.

DetachClassicLinkVpcRequest
Contains the parameters for DetachClassicLinkVpc.

DetachClassicLinkVpcResult
Contains the output of DetachClassicLinkVpc.

DetachInternetGatewayRequest
Contains the parameters for DetachInternetGateway.

DetachNetworkInterfaceRequest
Contains the parameters for DetachNetworkInterface.

DetachVolumeRequest
Contains the parameters for DetachVolume.

DetachVpnGatewayRequest
Contains the parameters for DetachVpnGateway.

DhcpConfiguration
Describes a DHCP configuration option.

DhcpOptions
Describes a set of DHCP options.

DisableVgwRoutePropagationRequest
Contains the parameters for DisableVgwRoutePropagation.

DisableVpcClassicLinkDnsSupportRequest
Contains the parameters for DisableVpcClassicLinkDnsSupport.

DisableVpcClassicLinkDnsSupportResult
Contains the output of DisableVpcClassicLinkDnsSupport.

DisableVpcClassicLinkRequest
Contains the parameters for DisableVpcClassicLink.

DisableVpcClassicLinkResult
Contains the output of DisableVpcClassicLink.

DisassociateAddressRequest
Contains the parameters for DisassociateAddress.

DisassociateRouteTableRequest
Contains the parameters for DisassociateRouteTable.

DiskImage
Describes a disk image.

DiskImageDescription
Describes a disk image.

DiskImageDetail
Describes a disk image.

DiskImageVolumeDescription
Describes a disk image volume.

EbsBlockDevice
Describes a block device for an EBS volume.

EbsInstanceBlockDevice
Describes a parameter used to set up an EBS volume in a block device mapping.

EbsInstanceBlockDeviceSpecification
Describes information used to set up an EBS volume specified in a block device mapping.

Ec2Client
A client for the Amazon EC2 API.

EnableVgwRoutePropagationRequest
Contains the parameters for EnableVgwRoutePropagation.

EnableVolumeIORequest
Contains the parameters for EnableVolumeIO.

EnableVpcClassicLinkDnsSupportRequest
Contains the parameters for EnableVpcClassicLinkDnsSupport.

EnableVpcClassicLinkDnsSupportResult
Contains the output of EnableVpcClassicLinkDnsSupport.

EnableVpcClassicLinkRequest
Contains the parameters for EnableVpcClassicLink.

EnableVpcClassicLinkResult
Contains the output of EnableVpcClassicLink.

EventInformation
Describes a Spot fleet event.

ExportTask
Describes an instance export task.

ExportToS3Task
Describes the format and location for an instance export task.

ExportToS3TaskSpecification
Describes an instance export task.

Filter
A filter name and value pair that is used to return a more specific list of results. Filters can be used to match a set of resources by various criteria, such as tags, attributes, or IDs.

FlowLog
Describes a flow log.

GetConsoleOutputRequest
Contains the parameters for GetConsoleOutput.

GetConsoleOutputResult
Contains the output of GetConsoleOutput.

GetPasswordDataRequest
Contains the parameters for GetPasswordData.

GetPasswordDataResult
Contains the output of GetPasswordData.

GroupIdentifier
Describes a security group.

HistoryRecord
Describes an event in the history of the Spot fleet request.

Host
Describes the properties of the Dedicated host.

HostInstance
Describes an instance running on a Dedicated host.

HostProperties
Describes properties of a Dedicated host.

IamInstanceProfile
Describes an IAM instance profile.

IamInstanceProfileSpecification
Describes an IAM instance profile.

IcmpTypeCode
Describes the ICMP type and code.

IdFormat
Describes the ID format for a resource.

Image
Describes an image.

ImageAttribute
Describes an image attribute.

ImageDiskContainer
Describes the disk container object for an import image task.

ImportImageRequest
Contains the parameters for ImportImage.

ImportImageResult
Contains the output for ImportImage.

ImportImageTask
Describes an import image task.

ImportInstanceLaunchSpecification
Describes the launch specification for VM import.

ImportInstanceRequest
Contains the parameters for ImportInstance.

ImportInstanceResult
Contains the output for ImportInstance.

ImportInstanceTaskDetails
Describes an import instance task.

ImportInstanceVolumeDetailItem
Describes an import volume task.

ImportKeyPairRequest
Contains the parameters for ImportKeyPair.

ImportKeyPairResult
Contains the output of ImportKeyPair.

ImportSnapshotRequest
Contains the parameters for ImportSnapshot.

ImportSnapshotResult
Contains the output for ImportSnapshot.

ImportSnapshotTask
Describes an import snapshot task.

ImportVolumeRequest
Contains the parameters for ImportVolume.

ImportVolumeResult
Contains the output for ImportVolume.

ImportVolumeTaskDetails
Describes an import volume task.

Instance
Describes an instance.

InstanceAttribute
Describes an instance attribute.

InstanceBlockDeviceMapping
Describes a block device mapping.

InstanceBlockDeviceMappingSpecification
Describes a block device mapping entry.

InstanceCapacity
Information about the instance type that the Dedicated host supports.

InstanceCount
Describes a Reserved Instance listing state.

InstanceExportDetails
Describes an instance to export.

InstanceMonitoring
Describes the monitoring information of the instance.

InstanceNetworkInterface
Describes a network interface.

InstanceNetworkInterfaceAssociation
Describes association information for an Elastic IP address.

InstanceNetworkInterfaceAttachment
Describes a network interface attachment.

InstanceNetworkInterfaceSpecification
Describes a network interface.

InstancePrivateIpAddress
Describes a private IP address.

InstanceState
Describes the current state of the instance.

InstanceStateChange
Describes an instance state change.

InstanceStatus
Describes the status of an instance.

InstanceStatusDetails
Describes the instance status.

InstanceStatusEvent
Describes a scheduled event for an instance.

InstanceStatusSummary
Describes the status of an instance.

InternetGateway
Describes an Internet gateway.

InternetGatewayAttachment
Describes the attachment of a VPC to an Internet gateway.

IpPermission
Describes a security group rule.

IpRange
Describes an IP range.

KeyPair
Describes a key pair.

KeyPairInfo
Describes a key pair.

LaunchPermission
Describes a launch permission.

LaunchPermissionModifications
Describes a launch permission modification.

LaunchSpecification
Describes the launch specification for an instance.

ModifyHostsRequest
Contains the parameters for ModifyHosts.

ModifyHostsResult
Contains the output of ModifyHosts.

ModifyIdFormatRequest
Contains the parameters of ModifyIdFormat.

ModifyImageAttributeRequest
Contains the parameters for ModifyImageAttribute.

ModifyInstanceAttributeRequest
Contains the parameters for ModifyInstanceAttribute.

ModifyInstancePlacementRequest
Contains the parameters for ModifyInstancePlacement.

ModifyInstancePlacementResult
Contains the output of ModifyInstancePlacement.

ModifyNetworkInterfaceAttributeRequest
Contains the parameters for ModifyNetworkInterfaceAttribute.

ModifyReservedInstancesRequest
Contains the parameters for ModifyReservedInstances.

ModifyReservedInstancesResult
Contains the output of ModifyReservedInstances.

ModifySnapshotAttributeRequest
Contains the parameters for ModifySnapshotAttribute.

ModifySpotFleetRequestRequest
Contains the parameters for ModifySpotFleetRequest.

ModifySpotFleetRequestResponse
Contains the output of ModifySpotFleetRequest.

ModifySubnetAttributeRequest
Contains the parameters for ModifySubnetAttribute.

ModifyVolumeAttributeRequest
Contains the parameters for ModifyVolumeAttribute.

ModifyVpcAttributeRequest
Contains the parameters for ModifyVpcAttribute.

ModifyVpcEndpointRequest
Contains the parameters for ModifyVpcEndpoint.

ModifyVpcEndpointResult
Contains the output of ModifyVpcEndpoint.

ModifyVpcPeeringConnectionOptionsRequest

ModifyVpcPeeringConnectionOptionsResult

MonitorInstancesRequest
Contains the parameters for MonitorInstances.

MonitorInstancesResult
Contains the output of MonitorInstances.

Monitoring
Describes the monitoring for the instance.

MoveAddressToVpcRequest
Contains the parameters for MoveAddressToVpc.

MoveAddressToVpcResult
Contains the output of MoveAddressToVpc.

MovingAddressStatus
Describes the status of a moving Elastic IP address.

NatGateway
Describes a NAT gateway.

NatGatewayAddress
Describes the IP addresses and network interface associated with a NAT gateway.

NetworkAcl
Describes a network ACL.

NetworkAclAssociation
Describes an association between a network ACL and a subnet.

NetworkAclEntry
Describes an entry in a network ACL.

NetworkInterface
Describes a network interface.

NetworkInterfaceAssociation
Describes association information for an Elastic IP address.

NetworkInterfaceAttachment
Describes a network interface attachment.

NetworkInterfaceAttachmentChanges
Describes an attachment change.

NetworkInterfacePrivateIpAddress
Describes the private IP address of a network interface.

NewDhcpConfiguration

PeeringConnectionOptions
Describes the VPC peering connection options.

PeeringConnectionOptionsRequest
The VPC peering connection options.

Placement
Describes the placement for the instance.

PlacementGroup
Describes a placement group.

PortRange
Describes a range of ports.

PrefixList
Describes prefixes for AWS services.

PrefixListId
The ID of the prefix.

PriceSchedule
Describes the price for a Reserved Instance.

PriceScheduleSpecification
Describes the price for a Reserved Instance.

PricingDetail
Describes a Reserved Instance offering.

PrivateIpAddressSpecification
Describes a secondary private IP address for a network interface.

ProductCode
Describes a product code.

PropagatingVgw
Describes a virtual private gateway propagating route.

PurchaseRequest
Describes a request to purchase Scheduled Instances.

PurchaseReservedInstancesOfferingRequest
Contains the parameters for PurchaseReservedInstancesOffering.

PurchaseReservedInstancesOfferingResult
Contains the output of PurchaseReservedInstancesOffering.

PurchaseScheduledInstancesRequest
Contains the parameters for PurchaseScheduledInstances.

PurchaseScheduledInstancesResult
Contains the output of PurchaseScheduledInstances.

RebootInstancesRequest
Contains the parameters for RebootInstances.

RecurringCharge
Describes a recurring charge.

Region
Describes a region.

RegisterImageRequest
Contains the parameters for RegisterImage.

RegisterImageResult
Contains the output of RegisterImage.

RejectVpcPeeringConnectionRequest
Contains the parameters for RejectVpcPeeringConnection.

RejectVpcPeeringConnectionResult
Contains the output of RejectVpcPeeringConnection.

ReleaseAddressRequest
Contains the parameters for ReleaseAddress.

ReleaseHostsRequest
Contains the parameters for ReleaseHosts.

ReleaseHostsResult
Contains the output of ReleaseHosts.

ReplaceNetworkAclAssociationRequest
Contains the parameters for ReplaceNetworkAclAssociation.

ReplaceNetworkAclAssociationResult
Contains the output of ReplaceNetworkAclAssociation.

ReplaceNetworkAclEntryRequest
Contains the parameters for ReplaceNetworkAclEntry.

ReplaceRouteRequest
Contains the parameters for ReplaceRoute.

ReplaceRouteTableAssociationRequest
Contains the parameters for ReplaceRouteTableAssociation.

ReplaceRouteTableAssociationResult
Contains the output of ReplaceRouteTableAssociation.

ReportInstanceStatusRequest
Contains the parameters for ReportInstanceStatus.

RequestSpotFleetRequest
Contains the parameters for RequestSpotFleet.

RequestSpotFleetResponse
Contains the output of RequestSpotFleet.

RequestSpotInstancesRequest
Contains the parameters for RequestSpotInstances.

RequestSpotInstancesResult
Contains the output of RequestSpotInstances.

RequestSpotLaunchSpecification
Describes the launch specification for an instance.

Reservation
Describes a reservation.

ReservedInstanceLimitPrice
Describes the limit price of a Reserved Instance offering.

ReservedInstances
Describes a Reserved Instance.

ReservedInstancesConfiguration
Describes the configuration settings for the modified Reserved Instances.

ReservedInstancesId
Describes the ID of a Reserved Instance.

ReservedInstancesListing
Describes a Reserved Instance listing.

ReservedInstancesModification
Describes a Reserved Instance modification.

ReservedInstancesModificationResult
Describes the modification request/s.

ReservedInstancesOffering
Describes a Reserved Instance offering.

ResetImageAttributeRequest
Contains the parameters for ResetImageAttribute.

ResetInstanceAttributeRequest
Contains the parameters for ResetInstanceAttribute.

ResetNetworkInterfaceAttributeRequest
Contains the parameters for ResetNetworkInterfaceAttribute.

ResetSnapshotAttributeRequest
Contains the parameters for ResetSnapshotAttribute.

RestoreAddressToClassicRequest
Contains the parameters for RestoreAddressToClassic.

RestoreAddressToClassicResult
Contains the output of RestoreAddressToClassic.

RevokeSecurityGroupEgressRequest
Contains the parameters for RevokeSecurityGroupEgress.

RevokeSecurityGroupIngressRequest
Contains the parameters for RevokeSecurityGroupIngress.

Route
Describes a route in a route table.

RouteTable
Describes a route table.

RouteTableAssociation
Describes an association between a route table and a subnet.

RunInstancesMonitoringEnabled
Describes the monitoring for the instance.

RunInstancesRequest
Contains the parameters for RunInstances.

RunScheduledInstancesRequest
Contains the parameters for RunScheduledInstances.

RunScheduledInstancesResult
Contains the output of RunScheduledInstances.

S3Storage
Describes the storage parameters for S3 and S3 buckets for an instance store-backed AMI.

ScheduledInstance
Describes a Scheduled Instance.

ScheduledInstanceAvailability
Describes a schedule that is available for your Scheduled Instances.

ScheduledInstanceRecurrence
Describes the recurring schedule for a Scheduled Instance.

ScheduledInstanceRecurrenceRequest
Describes the recurring schedule for a Scheduled Instance.

ScheduledInstancesBlockDeviceMapping
Describes a block device mapping for a Scheduled Instance.

ScheduledInstancesEbs
Describes an EBS volume for a Scheduled Instance.

ScheduledInstancesIamInstanceProfile
Describes an IAM instance profile for a Scheduled Instance.

ScheduledInstancesLaunchSpecification
Describes the launch specification for a Scheduled Instance.

If you are launching the Scheduled Instance in EC2-VPC, you must specify the ID of the subnet. You can specify the subnet using either SubnetId or NetworkInterface.

ScheduledInstancesMonitoring
Describes whether monitoring is enabled for a Scheduled Instance.

ScheduledInstancesNetworkInterface
Describes a network interface for a Scheduled Instance.

ScheduledInstancesPlacement
Describes the placement for a Scheduled Instance.

ScheduledInstancesPrivateIpAddressConfig
Describes a private IP address for a Scheduled Instance.

SecurityGroup
Describes a security group

SecurityGroupReference
Describes a VPC with a security group that references your security group.

SlotDateTimeRangeRequest
Describes the time period for a Scheduled Instance to start its first schedule. The time period must span less than one day.

SlotStartTimeRangeRequest
Describes the time period for a Scheduled Instance to start its first schedule.

Snapshot
Describes a snapshot.

SnapshotDetail
Describes the snapshot created from the imported disk.

SnapshotDiskContainer
The disk container object for the import snapshot request.

SnapshotTaskDetail
Details about the import snapshot task.

SpotDatafeedSubscription
Describes the data feed for a Spot instance.

SpotFleetLaunchSpecification
Describes the launch specification for one or more Spot instances.

SpotFleetMonitoring
Describes whether monitoring is enabled.

SpotFleetRequestConfig
Describes a Spot fleet request.

SpotFleetRequestConfigData
Describes the configuration of a Spot fleet request.

SpotInstanceRequest
Describes a Spot instance request.

SpotInstanceStateFault
Describes a Spot instance state change.

SpotInstanceStatus
Describes the status of a Spot instance request.

SpotPlacement
Describes Spot instance placement.

SpotPrice
Describes the maximum hourly price (bid) for any Spot instance launched to fulfill the request.

StaleIpPermission
Describes a stale rule in a security group.

StaleSecurityGroup
Describes a stale security group (a security group that contains stale rules).

StartInstancesRequest
Contains the parameters for StartInstances.

StartInstancesResult
Contains the output of StartInstances.

StateReason
Describes a state change.

StopInstancesRequest
Contains the parameters for StopInstances.

StopInstancesResult
Contains the output of StopInstances.

Storage
Describes the storage location for an instance store-backed AMI.

Subnet
Describes a subnet.

Tag
Describes a tag.

TagDescription
Describes a tag.

TerminateInstancesRequest
Contains the parameters for TerminateInstances.

TerminateInstancesResult
Contains the output of TerminateInstances.

UnassignPrivateIpAddressesRequest
Contains the parameters for UnassignPrivateIpAddresses.

UnmonitorInstancesRequest
Contains the parameters for UnmonitorInstances.

UnmonitorInstancesResult
Contains the output of UnmonitorInstances.

UnsuccessfulItem
Information about items that were not successfully processed in a batch call.

UnsuccessfulItemError
Information about the error that occurred. For more information about errors, see Error Codes.

UserBucket
Describes the S3 bucket for the disk image.

UserBucketDetails
Describes the S3 bucket for the disk image.

UserData
Describes the user data to be made available to an instance.

UserIdGroupPair
Describes a security group and AWS account ID pair.

VgwTelemetry
Describes telemetry for a VPN tunnel.

Volume
Describes a volume.

VolumeAttachment
Describes volume attachment details.

VolumeDetail
Describes an EBS volume.

VolumeStatusAction
Describes a volume status operation code.

VolumeStatusDetails
Describes a volume status.

VolumeStatusEvent
Describes a volume status event.

VolumeStatusInfo
Describes the status of a volume.

VolumeStatusItem
Describes the volume status.

Vpc
Describes a VPC.

VpcAttachment
Describes an attachment between a virtual private gateway and a VPC.

VpcClassicLink
Describes whether a VPC is enabled for ClassicLink.

VpcEndpoint
Describes a VPC endpoint.

VpcPeeringConnection
Describes a VPC peering connection.

VpcPeeringConnectionOptionsDescription
Describes the VPC peering connection options.

VpcPeeringConnectionStateReason
Describes the status of a VPC peering connection.

VpcPeeringConnectionVpcInfo
Describes a VPC in a VPC peering connection.

VpnConnection
Describes a VPN connection.

VpnConnectionOptions
Describes VPN connection options.

VpnConnectionOptionsSpecification
Describes VPN connection options.

VpnGateway
Describes a virtual private gateway.

VpnStaticRoute
Describes a static route for a VPN connection.

Enums

AcceptVpcPeeringConnectionError

AllocateAddressError

AllocateHostsError

AssignPrivateIpAddressesError

AssociateAddressError

AssociateDhcpOptionsError

AssociateRouteTableError

AttachClassicLinkVpcError

AttachInternetGatewayError

AttachNetworkInterfaceError

AttachVolumeError

AttachVpnGatewayError

AuthorizeSecurityGroupEgressError

AuthorizeSecurityGroupIngressError

BundleInstanceError

CancelBundleTaskError

CancelConversionTaskError

CancelExportTaskError

CancelImportTaskError

CancelReservedInstancesListingError

CancelSpotFleetRequestsErrorType

CancelSpotInstanceRequestsError

ConfirmProductInstanceError

CopyImageError

CopySnapshotError

CreateCustomerGatewayError

CreateDhcpOptionsError

CreateFlowLogsError

CreateImageError

CreateInstanceExportTaskError

CreateInternetGatewayError

CreateKeyPairError

CreateNatGatewayError

CreateNetworkAclEntryError

CreateNetworkAclError

CreateNetworkInterfaceError

CreatePlacementGroupError

CreateReservedInstancesListingError

CreateRouteError

CreateRouteTableError

CreateSecurityGroupError

CreateSnapshotError

CreateSpotDatafeedSubscriptionError

CreateSubnetError

CreateTagsError

CreateVolumeError

CreateVpcEndpointError

CreateVpcError

CreateVpcPeeringConnectionError

CreateVpnConnectionError

CreateVpnConnectionRouteError

CreateVpnGatewayError

DeleteCustomerGatewayError

DeleteDhcpOptionsError

DeleteFlowLogsError

DeleteInternetGatewayError

DeleteKeyPairError

DeleteNatGatewayError

DeleteNetworkAclEntryError

DeleteNetworkAclError

DeleteNetworkInterfaceError

DeletePlacementGroupError

DeleteRouteError

DeleteRouteTableError

DeleteSecurityGroupError

DeleteSnapshotError

DeleteSpotDatafeedSubscriptionError

DeleteSubnetError

DeleteTagsError

DeleteVolumeError

DeleteVpcEndpointsError

DeleteVpcError

DeleteVpcPeeringConnectionError

DeleteVpnConnectionError

DeleteVpnConnectionRouteError

DeleteVpnGatewayError

DeregisterImageError

DescribeAccountAttributesError

DescribeAddressesError

DescribeAvailabilityZonesError

DescribeBundleTasksError

DescribeClassicLinkInstancesError

DescribeConversionTasksError

DescribeCustomerGatewaysError

DescribeDhcpOptionsError

DescribeExportTasksError

DescribeFlowLogsError

DescribeHostsError

DescribeIdFormatError

DescribeImageAttributeError

DescribeImagesError

DescribeImportImageTasksError

DescribeImportSnapshotTasksError

DescribeInstanceAttributeError

DescribeInstanceStatusError

DescribeInstancesError

DescribeInternetGatewaysError

DescribeKeyPairsError

DescribeMovingAddressesError

DescribeNatGatewaysError

DescribeNetworkAclsError

DescribeNetworkInterfaceAttributeError

DescribeNetworkInterfacesError

DescribePlacementGroupsError

DescribePrefixListsError

DescribeRegionsError

DescribeReservedInstancesError

DescribeReservedInstancesListingsError

DescribeReservedInstancesModificationsError

DescribeReservedInstancesOfferingsError

DescribeRouteTablesError

DescribeScheduledInstanceAvailabilityError

DescribeScheduledInstancesError

DescribeSecurityGroupReferencesError

DescribeSecurityGroupsError

DescribeSnapshotAttributeError

DescribeSnapshotsError

DescribeSpotDatafeedSubscriptionError

DescribeSpotFleetInstancesError

DescribeSpotFleetRequestHistoryError

DescribeSpotFleetRequestsError

DescribeSpotInstanceRequestsError

DescribeSpotPriceHistoryError

DescribeStaleSecurityGroupsError

DescribeSubnetsError

DescribeTagsError

DescribeVolumeAttributeError

DescribeVolumeStatusError

DescribeVolumesError

DescribeVpcAttributeError

DescribeVpcClassicLinkDnsSupportError

DescribeVpcClassicLinkError

DescribeVpcEndpointServicesError

DescribeVpcEndpointsError

DescribeVpcPeeringConnectionsError

DescribeVpcsError

DescribeVpnConnectionsError

DescribeVpnGatewaysError

DetachClassicLinkVpcError

DetachInternetGatewayError

DetachNetworkInterfaceError

DetachVolumeError

DetachVpnGatewayError

DisableVgwRoutePropagationError

DisableVpcClassicLinkDnsSupportError

DisableVpcClassicLinkError

DisassociateAddressError

DisassociateRouteTableError

EnableVgwRoutePropagationError

EnableVolumeIOError

EnableVpcClassicLinkDnsSupportError

EnableVpcClassicLinkError

GetConsoleOutputError

GetPasswordDataError

ImportImageError

ImportInstanceError

ImportKeyPairError

ImportSnapshotError

ImportVolumeError

ModifyHostsError

ModifyIdFormatError

ModifyImageAttributeError

ModifyInstanceAttributeError

ModifyInstancePlacementError

ModifyNetworkInterfaceAttributeError

ModifyReservedInstancesError

ModifySnapshotAttributeError

ModifySpotFleetRequestError

ModifySubnetAttributeError

ModifyVolumeAttributeError

ModifyVpcAttributeError

ModifyVpcEndpointError

ModifyVpcPeeringConnectionOptionsError

MonitorInstancesError

MoveAddressToVpcError

PurchaseReservedInstancesOfferingError

PurchaseScheduledInstancesError

RebootInstancesError

RegisterImageError

RejectVpcPeeringConnectionError

ReleaseAddressError

ReleaseHostsError

ReplaceNetworkAclAssociationError

ReplaceNetworkAclEntryError

ReplaceRouteError

ReplaceRouteTableAssociationError

ReportInstanceStatusError

RequestSpotFleetError

RequestSpotInstancesError

ResetImageAttributeError

ResetInstanceAttributeError

ResetNetworkInterfaceAttributeError

ResetSnapshotAttributeError

RestoreAddressToClassicError

RevokeSecurityGroupEgressError

RevokeSecurityGroupIngressError

RunInstancesError

RunScheduledInstancesError

StartInstancesError

StopInstancesError

TerminateInstancesError

UnassignPrivateIpAddressesError

UnmonitorInstancesError

Type Definitions

AccountAttributeList

AccountAttributeName

AccountAttributeNameStringList

AccountAttributeValueList

ActiveInstanceSet

AddressList

Affinity

AllocationIdList

AllocationState

AllocationStrategy

ArchitectureValues

AttachmentStatus

AutoPlacement

AvailabilityZoneList

AvailabilityZoneMessageList

AvailabilityZoneState

AvailableInstanceCapacityList

BatchState

Blob

BlockDeviceMappingList

BlockDeviceMappingRequestList

Boolean

BundleIdStringList

BundleTaskList

BundleTaskState

CancelBatchErrorCode

CancelSpotFleetRequestsErrorSet

CancelSpotFleetRequestsSuccessSet

CancelSpotInstanceRequestState

CancelledSpotInstanceRequestList

ClassicLinkDnsSupportList

ClassicLinkInstanceList

ContainerFormat

ConversionIdStringList

ConversionTaskState

CreateVolumePermissionList

CurrencyCodeValues

CustomerGatewayIdStringList

CustomerGatewayList

DatafeedSubscriptionState

DateTime

DescribeConversionTaskList

DeviceType

DhcpConfigurationList

DhcpConfigurationValueList

DhcpOptionsIdStringList

DhcpOptionsList

DiskImageFormat

DiskImageList

DomainType

Double

EventCode

EventType

ExcessCapacityTerminationPolicy

ExecutableByStringList

ExportEnvironment

ExportTaskIdStringList

ExportTaskList

ExportTaskState

FilterList

Float

FlowLogSet

FlowLogsResourceType

GatewayType

GroupIdStringList

GroupIdentifierList

GroupIds

GroupNameStringList

HistoryRecords

HostInstanceList

HostList

HostTenancy

HypervisorType

IdFormatList

ImageAttributeName

ImageDiskContainerList

ImageIdStringList

ImageList

ImageState

ImageTypeValues

ImportImageTaskList

ImportInstanceVolumeDetailSet

ImportSnapshotTaskList

ImportTaskIdList

InstanceAttributeName

InstanceBlockDeviceMappingList

InstanceBlockDeviceMappingSpecificationList

InstanceCountList

InstanceIdSet

InstanceIdStringList

InstanceLifecycleType

InstanceList

InstanceMonitoringList

InstanceNetworkInterfaceList

InstanceNetworkInterfaceSpecificationList

InstancePrivateIpAddressList

InstanceStateChangeList

InstanceStateName

InstanceStatusDetailsList

InstanceStatusEventList

InstanceStatusList

InstanceType

InstanceTypeList

Integer

InternetGatewayAttachmentList

InternetGatewayList

IpPermissionList

IpRangeList

IpRanges

KeyNameStringList

KeyPairList

LaunchPermissionList

LaunchSpecsList

ListingState

ListingStatus

Long

MaxResults

MonitoringState

MoveStatus

MovingAddressStatusSet

NatGatewayAddressList

NatGatewayList

NatGatewayState

NetworkAclAssociationList

NetworkAclEntryList

NetworkAclList

NetworkInterfaceAttribute

NetworkInterfaceIdList

NetworkInterfaceList

NetworkInterfacePrivateIpAddressList

NetworkInterfaceStatus

NetworkInterfaceType

NewDhcpConfigurationList

NextToken

OccurrenceDayRequestSet

OccurrenceDaySet

OfferingTypeValues

OperationType

OwnerStringList

PermissionGroup

PlacementGroupList

PlacementGroupState

PlacementGroupStringList

PlacementStrategy

PlatformValues

PrefixListIdList

PrefixListIdSet

PrefixListSet

PriceScheduleList

PriceScheduleSpecificationList

PricingDetailsList

PrivateIpAddressConfigSet

PrivateIpAddressSpecificationList

PrivateIpAddressStringList

ProductCodeList

ProductCodeStringList

ProductCodeValues

ProductDescriptionList

PropagatingVgwList

PublicIpStringList

PurchaseRequestSet

PurchasedScheduledInstanceSet

RIProductDescription

ReasonCodesList

RecurringChargeFrequency

RecurringChargesList

RegionList

RegionNameStringList

ReportInstanceReasonCodes

ReportStatusType

RequestHostIdList

ReservationList

ReservedInstanceState

ReservedInstancesConfigurationList

ReservedInstancesIdStringList

ReservedInstancesList

ReservedInstancesListingList

ReservedInstancesModificationIdStringList

ReservedInstancesModificationList

ReservedInstancesModificationResultList

ReservedInstancesOfferingIdStringList

ReservedInstancesOfferingList

ReservedIntancesIds

ResetImageAttributeName

ResourceIdList

ResourceType

ResponseHostIdList

RestorableByStringList

RouteList

RouteOrigin

RouteState

RouteTableAssociationList

RouteTableList

RuleAction

ScheduledInstanceAvailabilitySet

ScheduledInstanceIdRequestSet

ScheduledInstanceSet

ScheduledInstancesBlockDeviceMappingSet

ScheduledInstancesNetworkInterfaceSet

ScheduledInstancesSecurityGroupIdSet

SecurityGroupIdStringList

SecurityGroupList

SecurityGroupReferences

SecurityGroupStringList

ShutdownBehavior

SnapshotAttributeName

SnapshotDetailList

SnapshotIdStringList

SnapshotList

SnapshotState

SpotFleetRequestConfigSet

SpotInstanceRequestIdList

SpotInstanceRequestList

SpotInstanceState

SpotInstanceType

SpotPriceHistoryList

StaleIpPermissionSet

StaleSecurityGroupSet

State

Status

StatusName

StatusType

SubnetIdStringList

SubnetList

SubnetState

SummaryStatus

TagDescriptionList

TagList

TelemetryStatus

Tenancy

TrafficType

UnsuccessfulItemList

UnsuccessfulItemSet

UserGroupStringList

UserIdGroupPairList

UserIdGroupPairSet

UserIdStringList

ValueStringList

VgwTelemetryList

VirtualizationType

VolumeAttachmentList

VolumeAttachmentState

VolumeAttributeName

VolumeIdStringList

VolumeList

VolumeState

VolumeStatusActionsList

VolumeStatusDetailsList

VolumeStatusEventsList

VolumeStatusInfoStatus

VolumeStatusList

VolumeStatusName

VolumeType

VpcAttachmentList

VpcAttributeName

VpcClassicLinkIdList

VpcClassicLinkList

VpcEndpointSet

VpcIdStringList

VpcList

VpcPeeringConnectionList

VpcPeeringConnectionStateReasonCode

VpcState

VpnConnectionIdStringList

VpnConnectionList

VpnGatewayIdStringList

VpnGatewayList

VpnState

VpnStaticRouteList

VpnStaticRouteSource

ZoneNameStringList