Struct rusoto::ec2::CreateFlowLogsRequest

pub struct CreateFlowLogsRequest {
    pub client_token: Option<String>,
    pub deliver_logs_permission_arn: String,
    pub log_group_name: String,
    pub resource_ids: ValueStringList,
    pub resource_type: FlowLogsResourceType,
    pub traffic_type: TrafficType,
}

Contains the parameters for CreateFlowLogs.

Fields
client_token: Option<String>

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

deliver_logs_permission_arn: String

The ARN for the IAM role that's used to post flow logs to a CloudWatch Logs log group.

log_group_name: String

The name of the CloudWatch log group.

resource_ids: ValueStringList

One or more subnet, network interface, or VPC IDs.

Constraints: Maximum of 1000 resources

resource_type: FlowLogsResourceType

The type of resource on which to create the flow log.

traffic_type: TrafficType

The type of traffic to log.