Struct rusoto::ec2::CreateNetworkAclEntryRequest

pub struct CreateNetworkAclEntryRequest {
    pub cidr_block: String,
    pub dry_run: Option<Boolean>,
    pub egress: Boolean,
    pub icmp_type_code: Option<IcmpTypeCode>,
    pub network_acl_id: String,
    pub port_range: Option<PortRange>,
    pub protocol: String,
    pub rule_action: RuleAction,
    pub rule_number: Integer,
}

Contains the parameters for CreateNetworkAclEntry.

Fields
cidr_block: String

The network range to allow or deny, in CIDR notation (for example 172.16.0.0/24).

dry_run: Option<Boolean>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

egress: Boolean

Indicates whether this is an egress rule (rule is applied to traffic leaving the subnet).

icmp_type_code: Option<IcmpTypeCode>

ICMP protocol: The ICMP type and code. Required if specifying ICMP for the protocol.

network_acl_id: String

The ID of the network ACL.

port_range: Option<PortRange>

TCP or UDP protocols: The range of ports the rule applies to.

protocol: String

The protocol. A value of -1 means all protocols.

rule_action: RuleAction

Indicates whether to allow or deny the traffic that matches the rule.

rule_number: Integer

The rule number for the entry (for example, 100). ACL entries are processed in ascending order by rule number.

Constraints: Positive integer from 1 to 32766