Struct rusoto::ec2::DescribeDhcpOptionsRequest

pub struct DescribeDhcpOptionsRequest {
    pub dhcp_options_ids: Option<DhcpOptionsIdStringList>,
    pub dry_run: Option<Boolean>,
    pub filters: Option<FilterList>,
}

Contains the parameters for DescribeDhcpOptions.

Fields
dhcp_options_ids: Option<DhcpOptionsIdStringList>

The IDs of one or more DHCP options sets.

Default: Describes all your DHCP options sets.

dry_run: Option<Boolean>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters: Option<FilterList>

One or more filters.

dhcp-options-id - The ID of a set of DHCP options.

key - The key for one of the options (for example, domain-name).

value - The value for one of the options.

tag:key=value - The key/value combination of a tag assigned to the resource.

tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter.

tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter.