Struct rusoto::ec2::DescribeInstanceStatusRequest

pub struct DescribeInstanceStatusRequest {
    pub dry_run: Option<Boolean>,
    pub filters: Option<FilterList>,
    pub include_all_instances: Option<Boolean>,
    pub instance_ids: Option<InstanceIdStringList>,
    pub max_results: Option<Integer>,
    pub next_token: Option<String>,
}

Contains the parameters for DescribeInstanceStatus.

Fields
dry_run: Option<Boolean>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters: Option<FilterList>

One or more filters.

availability-zone - The Availability Zone of the instance.

event.code - The code for the scheduled event (instance-reboot | system-reboot | system-maintenance | instance-retirement | instance-stop).

event.description - A description of the event.

event.not-after - The latest end time for the scheduled event (for example, 2014-09-15T17:15:20.000Z).

event.not-before - The earliest start time for the scheduled event (for example, 2014-09-15T17:15:20.000Z).

instance-state-code - The code for the instance state, as a 16-bit unsigned integer. The high byte is an opaque internal value and should be ignored. The low byte is set based on the state represented. The valid values are 0 (pending), 16 (running), 32 (shutting-down), 48 (terminated), 64 (stopping), and 80 (stopped).

instance-state-name - The state of the instance (pending | running | shutting-down | terminated | stopping | stopped).

instance-status.reachability - Filters on instance status where the name is reachability (passed | failed | initializing | insufficient-data).

instance-status.status - The status of the instance (ok | impaired | initializing | insufficient-data | not-applicable).

system-status.reachability - Filters on system status where the name is reachability (passed | failed | initializing | insufficient-data).

system-status.status - The system status of the instance (ok | impaired | initializing | insufficient-data | not-applicable).

include_all_instances: Option<Boolean>

When true, includes the health status for all instances. When false, includes the health status for running instances only.

Default: false

instance_ids: Option<InstanceIdStringList>

One or more instance IDs.

Default: Describes all your instances.

Constraints: Maximum 100 explicitly specified instance IDs.

max_results: Option<Integer>

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 5 and 1000. You cannot specify this parameter and the instance IDs parameter in the same call.

next_token: Option<String>

The token to retrieve the next page of results.