Struct rusoto::ec2::Host

pub struct Host {
    pub auto_placement: Option<AutoPlacement>,
    pub availability_zone: Option<String>,
    pub available_capacity: Option<AvailableCapacity>,
    pub client_token: Option<String>,
    pub host_id: Option<String>,
    pub host_properties: Option<HostProperties>,
    pub host_reservation_id: Option<String>,
    pub instances: Option<HostInstanceList>,
    pub state: Option<AllocationState>,
}

Describes the properties of the Dedicated host.

Fields
auto_placement: Option<AutoPlacement>

Whether auto-placement is on or off.

availability_zone: Option<String>

The Availability Zone of the Dedicated host.

available_capacity: Option<AvailableCapacity>

The number of new instances that can be launched onto the Dedicated host.

client_token: Option<String>

Unique, case-sensitive identifier you provide to ensure idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide.

host_id: Option<String>

The ID of the Dedicated host.

host_properties: Option<HostProperties>

The hardware specifications of the Dedicated host.

host_reservation_id: Option<String>

The reservation ID of the Dedicated host. This returns a null response if the Dedicated host doesn't have an associated reservation.

instances: Option<HostInstanceList>

The IDs and instance type that are currently running on the Dedicated host.

state: Option<AllocationState>

The Dedicated host's state.