Struct rusoto::ec2::InstanceAttribute

pub struct InstanceAttribute {
    pub block_device_mappings: Option<InstanceBlockDeviceMappingList>,
    pub disable_api_termination: Option<AttributeBooleanValue>,
    pub ebs_optimized: Option<AttributeBooleanValue>,
    pub groups: Option<GroupIdentifierList>,
    pub instance_id: Option<String>,
    pub instance_initiated_shutdown_behavior: Option<AttributeValue>,
    pub instance_type: Option<AttributeValue>,
    pub kernel_id: Option<AttributeValue>,
    pub product_codes: Option<ProductCodeList>,
    pub ramdisk_id: Option<AttributeValue>,
    pub root_device_name: Option<AttributeValue>,
    pub source_dest_check: Option<AttributeBooleanValue>,
    pub sriov_net_support: Option<AttributeValue>,
    pub user_data: Option<AttributeValue>,
}

Describes an instance attribute.

Fields
block_device_mappings: Option<InstanceBlockDeviceMappingList>

The block device mapping of the instance.

disable_api_termination: Option<AttributeBooleanValue>

If the value is true, you can't terminate the instance through the Amazon EC2 console, CLI, or API; otherwise, you can.

ebs_optimized: Option<AttributeBooleanValue>

Indicates whether the instance is optimized for EBS I/O.

groups: Option<GroupIdentifierList>

The security groups associated with the instance.

instance_id: Option<String>

The ID of the instance.

instance_initiated_shutdown_behavior: Option<AttributeValue>

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

instance_type: Option<AttributeValue>

The instance type.

kernel_id: Option<AttributeValue>

The kernel ID.

product_codes: Option<ProductCodeList>

A list of product codes.

ramdisk_id: Option<AttributeValue>

The RAM disk ID.

root_device_name: Option<AttributeValue>

The name of the root device (for example, /dev/sda1 or /dev/xvda).

source_dest_check: Option<AttributeBooleanValue>

Indicates whether source/destination checking is enabled. A value of true means checking is enabled, and false means checking is disabled. This value must be false for a NAT instance to perform NAT.

sriov_net_support: Option<AttributeValue> user_data: Option<AttributeValue>

The Base64-encoded MIME user data.