Struct rusoto::ec2::InstanceNetworkInterfaceSpecification

pub struct InstanceNetworkInterfaceSpecification {
    pub associate_public_ip_address: Option<Boolean>,
    pub delete_on_termination: Option<Boolean>,
    pub description: Option<String>,
    pub device_index: Option<Integer>,
    pub groups: Option<SecurityGroupIdStringList>,
    pub network_interface_id: Option<String>,
    pub private_ip_address: Option<String>,
    pub private_ip_addresses: Option<PrivateIpAddressSpecificationList>,
    pub secondary_private_ip_address_count: Option<Integer>,
    pub subnet_id: Option<String>,
}

Describes a network interface.

Fields
associate_public_ip_address: Option<Boolean>

Indicates whether to assign a public IP address to an instance you launch in a VPC. The public IP address can only be assigned to a network interface for eth0, and can only be assigned to a new network interface, not an existing one. You cannot specify more than one network interface in the request. If launching into a default subnet, the default value is true.

delete_on_termination: Option<Boolean>

If set to true, the interface is deleted when the instance is terminated. You can specify true only if creating a new network interface when launching an instance.

description: Option<String>

The description of the network interface. Applies only if creating a network interface when launching an instance.

device_index: Option<Integer>

The index of the device on the instance for the network interface attachment. If you are specifying a network interface in a RunInstances request, you must provide the device index.

groups: Option<SecurityGroupIdStringList>

The IDs of the security groups for the network interface. Applies only if creating a network interface when launching an instance.

network_interface_id: Option<String>

The ID of the network interface.

private_ip_address: Option<String>

The private IP address of the network interface. Applies only if creating a network interface when launching an instance.

private_ip_addresses: Option<PrivateIpAddressSpecificationList>

One or more private IP addresses to assign to the network interface. Only one private IP address can be designated as primary.

secondary_private_ip_address_count: Option<Integer>

The number of secondary private IP addresses. You can't specify this option and specify more than one private IP address using the private IP addresses option.

subnet_id: Option<String>

The ID of the subnet associated with the network string. Applies only if creating a network interface when launching an instance.