Struct rusoto::ec2::ModifyImageAttributeRequest

pub struct ModifyImageAttributeRequest {
    pub attribute: Option<String>,
    pub description: Option<AttributeValue>,
    pub dry_run: Option<Boolean>,
    pub image_id: String,
    pub launch_permission: Option<LaunchPermissionModifications>,
    pub operation_type: Option<OperationType>,
    pub product_codes: Option<ProductCodeStringList>,
    pub user_groups: Option<UserGroupStringList>,
    pub user_ids: Option<UserIdStringList>,
    pub value: Option<String>,
}

Contains the parameters for ModifyImageAttribute.

Fields
attribute: Option<String>

The name of the attribute to modify.

description: Option<AttributeValue>

A description for the AMI.

dry_run: Option<Boolean>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

image_id: String

The ID of the AMI.

launch_permission: Option<LaunchPermissionModifications>

A launch permission modification.

operation_type: Option<OperationType>

The operation type.

product_codes: Option<ProductCodeStringList>

One or more product codes. After you add a product code to an AMI, it can't be removed. This is only valid when modifying the productCodes attribute.

user_groups: Option<UserGroupStringList>

One or more user groups. This is only valid when modifying the launchPermission attribute.

user_ids: Option<UserIdStringList>

One or more AWS account IDs. This is only valid when modifying the launchPermission attribute.

value: Option<String>

The value of the attribute being modified. This is only valid when modifying the description attribute.