Struct rusoto::ec2::NetworkInterface

pub struct NetworkInterface {
    pub association: Option<NetworkInterfaceAssociation>,
    pub attachment: Option<NetworkInterfaceAttachment>,
    pub availability_zone: Option<String>,
    pub description: Option<String>,
    pub groups: Option<GroupIdentifierList>,
    pub interface_type: Option<NetworkInterfaceType>,
    pub mac_address: Option<String>,
    pub network_interface_id: Option<String>,
    pub owner_id: Option<String>,
    pub private_dns_name: Option<String>,
    pub private_ip_address: Option<String>,
    pub private_ip_addresses: Option<NetworkInterfacePrivateIpAddressList>,
    pub requester_id: Option<String>,
    pub requester_managed: Option<Boolean>,
    pub source_dest_check: Option<Boolean>,
    pub status: Option<NetworkInterfaceStatus>,
    pub subnet_id: Option<String>,
    pub tag_set: Option<TagList>,
    pub vpc_id: Option<String>,
}

Describes a network interface.

Fields

association

The association information for an Elastic IP associated with the network interface.

attachment

The network interface attachment.

availability_zone

The Availability Zone.

description

A description.

groups

Any security groups for the network interface.

interface_type

The type of interface.

mac_address

The MAC address.

network_interface_id

The ID of the network interface.

owner_id

The AWS account ID of the owner of the network interface.

private_dns_name

The private DNS name.

private_ip_address

The IP address of the network interface within the subnet.

private_ip_addresses

The private IP addresses associated with the network interface.

requester_id

The ID of the entity that launched the instance on your behalf (for example, AWS Management Console or Auto Scaling).

requester_managed

Indicates whether the network interface is being managed by AWS.

source_dest_check

Indicates whether traffic to or from the instance is validated.

status

The status of the network interface.

subnet_id

The ID of the subnet.

tag_set

Any tags assigned to the network interface.

vpc_id

The ID of the VPC.