Struct rusoto::ec2::RegisterImageRequest

pub struct RegisterImageRequest {
    pub architecture: Option<ArchitectureValues>,
    pub block_device_mappings: Option<BlockDeviceMappingRequestList>,
    pub description: Option<String>,
    pub dry_run: Option<Boolean>,
    pub image_location: Option<String>,
    pub kernel_id: Option<String>,
    pub name: String,
    pub ramdisk_id: Option<String>,
    pub root_device_name: Option<String>,
    pub sriov_net_support: Option<String>,
    pub virtualization_type: Option<String>,
}

Contains the parameters for RegisterImage.

Fields
architecture: Option<ArchitectureValues>

The architecture of the AMI.

Default: For Amazon EBS-backed AMIs, i386. For instance store-backed AMIs, the architecture specified in the manifest file.

block_device_mappings: Option<BlockDeviceMappingRequestList>

One or more block device mapping entries.

description: Option<String>

A description for your AMI.

dry_run: Option<Boolean>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

image_location: Option<String>

The full path to your AMI manifest in Amazon S3 storage.

kernel_id: Option<String>

The ID of the kernel.

name: String

A name for your AMI.

Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

ramdisk_id: Option<String>

The ID of the RAM disk.

root_device_name: Option<String>

The name of the root device (for example, /dev/sda1, or /dev/xvda).

sriov_net_support: Option<String>

Set to simple to enable enhanced networking for the AMI and any instances that you launch from the AMI.

There is no way to disable enhanced networking at this time.

This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

virtualization_type: Option<String>

The type of virtualization.

Default: paravirtual