Struct rusoto::ec2::RunInstancesRequest

pub struct RunInstancesRequest {
    pub additional_info: Option<String>,
    pub block_device_mappings: Option<BlockDeviceMappingRequestList>,
    pub client_token: Option<String>,
    pub disable_api_termination: Option<Boolean>,
    pub dry_run: Option<Boolean>,
    pub ebs_optimized: Option<Boolean>,
    pub iam_instance_profile: Option<IamInstanceProfileSpecification>,
    pub image_id: String,
    pub instance_initiated_shutdown_behavior: Option<ShutdownBehavior>,
    pub instance_type: Option<InstanceType>,
    pub kernel_id: Option<String>,
    pub key_name: Option<String>,
    pub max_count: Integer,
    pub min_count: Integer,
    pub monitoring: Option<RunInstancesMonitoringEnabled>,
    pub network_interfaces: Option<InstanceNetworkInterfaceSpecificationList>,
    pub placement: Option<Placement>,
    pub private_ip_address: Option<String>,
    pub ramdisk_id: Option<String>,
    pub security_group_ids: Option<SecurityGroupIdStringList>,
    pub security_groups: Option<SecurityGroupStringList>,
    pub subnet_id: Option<String>,
    pub user_data: Option<String>,
}

Contains the parameters for RunInstances.

Fields

additional_info

Reserved.

block_device_mappings

The block device mapping.

client_token

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

Constraints: Maximum 64 ASCII characters

disable_api_termination

If you set this parameter to true, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. If you set this parameter to true and then later want to be able to terminate the instance, you must first change the value of the disableApiTermination attribute to false using ModifyInstanceAttribute. Alternatively, if you set InstanceInitiatedShutdownBehavior to terminate, you can terminate the instance by running the shutdown command from the instance.

Default: false

dry_run

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

ebs_optimized

Indicates whether the instance is optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance.

Default: false

iam_instance_profile

The IAM instance profile.

image_id

The ID of the AMI, which you can get by calling DescribeImages.

instance_initiated_shutdown_behavior

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

Default: stop

instance_type

The instance type. For more information, see Instance Types in the Amazon Elastic Compute Cloud User Guide.

Default: m1.small

kernel_id

The ID of the kernel.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB in the Amazon Elastic Compute Cloud User Guide.

key_name

The name of the key pair. You can create a key pair using CreateKeyPair or ImportKeyPair.

If you do not specify a key pair, you can't connect to the instance unless you choose an AMI that is configured to allow users another way to log in.

max_count

The maximum number of instances to launch. If you specify more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches the largest possible number of instances above MinCount.

Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 FAQ.

min_count

The minimum number of instances to launch. If you specify a minimum that is more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches no instances.

Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 General FAQ.

monitoring

The monitoring for the instance.

network_interfaces

One or more network interfaces.

placement

The placement for the instance.

private_ip_address

[EC2-VPC] The primary IP address. You must specify a value from the IP address range of the subnet.

Only one private IP address can be designated as primary. Therefore, you can't specify this parameter if PrivateIpAddresses.n.Primary is set to true and PrivateIpAddresses.n.PrivateIpAddress is set to an IP address.

Default: We select an IP address from the IP address range of the subnet.

ramdisk_id

The ID of the RAM disk.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB in the Amazon Elastic Compute Cloud User Guide.

security_group_ids

One or more security group IDs. You can create a security group using CreateSecurityGroup.

Default: Amazon EC2 uses the default security group.

security_groups

[EC2-Classic, default VPC] One or more security group names. For a nondefault VPC, you must use security group IDs instead.

Default: Amazon EC2 uses the default security group.

subnet_id

[EC2-VPC] The ID of the subnet to launch the instance into.

user_data

Data to configure the instance, or a script to run during instance launch. For more information, see Running Commands on Your Linux Instance at Launch (Linux) and Adding User Data (Windows). For API calls, the text must be base64-encoded. For command line tools, the encoding is performed for you, and you can load the text from a file.