Struct rusoto::ec2::SpotFleetRequestConfigData

pub struct SpotFleetRequestConfigData {
    pub allocation_strategy: Option<AllocationStrategy>,
    pub client_token: Option<String>,
    pub excess_capacity_termination_policy: Option<ExcessCapacityTerminationPolicy>,
    pub iam_fleet_role: String,
    pub launch_specifications: LaunchSpecsList,
    pub spot_price: String,
    pub target_capacity: Integer,
    pub terminate_instances_with_expiration: Option<Boolean>,
    pub valid_from: Option<DateTime>,
    pub valid_until: Option<DateTime>,
}

Describes the configuration of a Spot fleet request.

Fields
allocation_strategy: Option<AllocationStrategy>

Indicates how to allocate the target capacity across the Spot pools specified by the Spot fleet request. The default is lowestPrice.

client_token: Option<String>

A unique, case-sensitive identifier you provide to ensure idempotency of your listings. This helps avoid duplicate listings. For more information, see Ensuring Idempotency.

excess_capacity_termination_policy: Option<ExcessCapacityTerminationPolicy>

Indicates whether running Spot instances should be terminated if the target capacity of the Spot fleet request is decreased below the current size of the Spot fleet.

iam_fleet_role: String

Grants the Spot fleet permission to terminate Spot instances on your behalf when you cancel its Spot fleet request using CancelSpotFleetRequests or when the Spot fleet request expires, if you set terminateInstancesWithExpiration.

launch_specifications: LaunchSpecsList

Information about the launch specifications for the Spot fleet request.

spot_price: String

The bid price per unit hour.

target_capacity: Integer

The number of units to request. You can choose to set the target capacity in terms of instances or a performance characteristic that is important to your application workload, such as vCPUs, memory, or I/O.

terminate_instances_with_expiration: Option<Boolean>

Indicates whether running Spot instances should be terminated when the Spot fleet request expires.

valid_from: Option<DateTime>

The start date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The default is to start fulfilling the request immediately.

valid_until: Option<DateTime>

The end date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). At this point, no new Spot instance requests are placed or enabled to fulfill the request.