Struct rusoto::ec2::Subnet

pub struct Subnet {
    pub availability_zone: Option<String>,
    pub available_ip_address_count: Option<Integer>,
    pub cidr_block: Option<String>,
    pub default_for_az: Option<Boolean>,
    pub map_public_ip_on_launch: Option<Boolean>,
    pub state: Option<SubnetState>,
    pub subnet_id: Option<String>,
    pub tags: Option<TagList>,
    pub vpc_id: Option<String>,
}

Describes a subnet.

Fields
availability_zone: Option<String>

The Availability Zone of the subnet.

available_ip_address_count: Option<Integer>

The number of unused IP addresses in the subnet. Note that the IP addresses for any stopped instances are considered unavailable.

cidr_block: Option<String>

The CIDR block assigned to the subnet.

default_for_az: Option<Boolean>

Indicates whether this is the default subnet for the Availability Zone.

map_public_ip_on_launch: Option<Boolean>

Indicates whether instances launched in this subnet receive a public IP address.

state: Option<SubnetState>

The current state of the subnet.

subnet_id: Option<String>

The ID of the subnet.

tags: Option<TagList>

Any tags assigned to the subnet.

vpc_id: Option<String>

The ID of the VPC the subnet is in.