Struct rusoto::ecr::DescribeRepositoriesRequest

pub struct DescribeRepositoriesRequest {
    pub max_results: Option<MaxResults>,
    pub next_token: Option<NextToken>,
    pub registry_id: Option<RegistryId>,
    pub repository_names: Option<RepositoryNameList>,
}

Fields
max_results: Option<MaxResults>

The maximum number of repository results returned by DescribeRepositories in paginated output. When this parameter is used, DescribeRepositories only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeRepositories request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then DescribeRepositories returns up to 100 results and a nextToken value, if applicable.

next_token: Option<NextToken>

The nextToken value returned from a previous paginated DescribeRepositories request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

registry_id: Option<RegistryId>

The AWS account ID associated with the registry that contains the repositories to be described. If you do not specify a registry, the default registry is assumed.

repository_names: Option<RepositoryNameList>

A list of repositories to describe. If this parameter is omitted, then all repositories in a registry are described.