Struct rusoto::ecr::PutImageRequest

pub struct PutImageRequest {
    pub image_manifest: ImageManifest,
    pub registry_id: Option<RegistryId>,
    pub repository_name: RepositoryName,
}

Fields
image_manifest: ImageManifest

The image manifest corresponding to the image to be uploaded.

registry_id: Option<RegistryId>

The AWS account ID associated with the registry that contains the repository in which to put the image. If you do not specify a registry, the default registry is assumed.

repository_name: RepositoryName

The name of the repository in which to put the image.