Enum rusoto::ecs::ListContainerInstancesError

pub enum ListContainerInstancesError {
    Server(String),
    Client(String),
    InvalidParameter(String),
    ClusterNotFound(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Variants
Server(String)

These errors are usually caused by a server issue.

Client(String)

These errors are usually caused by a client action, such as using an action or resource on behalf of a user that doesn't have permission to use the action or resource, or specifying an identifier that is not valid.

InvalidParameter(String)

The specified parameter is invalid. Review the available parameters for the API request.

ClusterNotFound(String)

The specified cluster could not be found. You can view your available clusters with ListClusters. Amazon ECS clusters are region-specific.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl ListContainerInstancesError

fn from_body(body: &str) -> ListContainerInstancesError