Struct rusoto::ecs::Cluster

pub struct Cluster {
    pub status: Option<String>,
    pub clusterName: Option<String>,
    pub registeredContainerInstancesCount: Option<Integer>,
    pub pendingTasksCount: Option<Integer>,
    pub runningTasksCount: Option<Integer>,
    pub activeServicesCount: Option<Integer>,
    pub clusterArn: Option<String>,
}

A regional grouping of one or more container instances on which you can run task requests. Each account receives a default cluster the first time you use the Amazon ECS service, but you may also create other clusters. Clusters may contain more than one instance type simultaneously.

Fields

status

The status of the cluster. The valid values are ACTIVE or INACTIVE. ACTIVE indicates that you can register container instances with the cluster and the associated instances can accept tasks.

clusterName

A user-generated string that you use to identify your cluster.

registeredContainerInstancesCount

The number of container instances registered into the cluster.

pendingTasksCount

The number of tasks in the cluster that are in the PENDING state.

runningTasksCount

The number of tasks in the cluster that are in the RUNNING state.

activeServicesCount

The number of services that are running on the cluster in an ACTIVE state. You can view these services with ListServices.

clusterArn

The Amazon Resource Name (ARN) that identifies the cluster. The ARN contains the arn:aws:ecs namespace, followed by the region of the cluster, the AWS account ID of the cluster owner, the cluster namespace, and then the cluster name. For example, arn:aws:ecs:_region_:_012345678910_:cluster/_test_.