Struct rusoto::ecs::ContainerInstance

pub struct ContainerInstance {
    pub agent_connected: Option<Boolean>,
    pub agent_update_status: Option<AgentUpdateStatus>,
    pub attributes: Option<Attributes>,
    pub container_instance_arn: Option<String>,
    pub ec_2_instance_id: Option<String>,
    pub pending_tasks_count: Option<Integer>,
    pub registered_resources: Option<Resources>,
    pub remaining_resources: Option<Resources>,
    pub running_tasks_count: Option<Integer>,
    pub status: Option<String>,
    pub version_info: Option<VersionInfo>,
}

An EC2 instance that is running the Amazon ECS agent and has been registered with a cluster.

Fields
agent_connected: Option<Boolean>

This parameter returns true if the agent is actually connected to Amazon ECS. Registered instances with an agent that may be unhealthy or stopped return false, and instances without a connected agent cannot accept placement requests.

agent_update_status: Option<AgentUpdateStatus>

The status of the most recent agent update. If an update has never been requested, this value is NULL.

attributes: Option<Attributes>

The attributes set for the container instance by the Amazon ECS container agent at instance registration.

container_instance_arn: Option<String>

The Amazon Resource Name (ARN) of the container instance. The ARN contains the arn:aws:ecs namespace, followed by the region of the container instance, the AWS account ID of the container instance owner, the container-instance namespace, and then the container instance ID. For example, arn:aws:ecs:region:aws_account_id:container-instance/container_instance_ID .

ec_2_instance_id: Option<String>

The EC2 instance ID of the container instance.

pending_tasks_count: Option<Integer>

The number of tasks on the container instance that are in the PENDING status.

registered_resources: Option<Resources>

The registered resources on the container instance that are in use by current tasks.

remaining_resources: Option<Resources>

The remaining resources of the container instance that are available for new tasks.

running_tasks_count: Option<Integer>

The number of tasks on the container instance that are in the RUNNING status.

status: Option<String>

The status of the container instance. The valid values are ACTIVE or INACTIVE. ACTIVE indicates that the container instance can accept tasks.

version_info: Option<VersionInfo>

The version information for the Amazon ECS container agent and Docker daemon running on the container instance.