Struct rusoto::ecs::ContainerOverride

pub struct ContainerOverride {
    pub environment: Option<EnvironmentVariables>,
    pub command: Option<StringList>,
    pub name: Option<String>,
}

The overrides that should be sent to a container.

Fields

environment

The environment variables to send to the container. You can add new environment variables, which are added to the container at launch, or you can override the existing environment variables from the Docker image or the task definition.

command

The command to send to the container that overrides the default command from the Docker image or the task definition.

name

The name of the container that receives the override.