Struct rusoto::ecs::CreateServiceRequest

pub struct CreateServiceRequest {
    pub loadBalancers: Option<LoadBalancers>,
    pub desiredCount: BoxedInteger,
    pub cluster: Option<String>,
    pub serviceName: String,
    pub role: Option<String>,
    pub clientToken: Option<String>,
    pub taskDefinition: String,
}

Fields

loadBalancers

A list of load balancer objects, containing the load balancer name, the container name (as it appears in a container definition), and the container port to access from the load balancer.

desiredCount

The number of instantiations of the specified task definition to place and keep running on your cluster.

cluster

The short name or full Amazon Resource Name (ARN) of the cluster on which to run your service. If you do not specify a cluster, the default cluster is assumed.

serviceName

The name of your service. Up to 255 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed. Service names must be unique within a cluster, but you can have similarly named services in multiple clusters within a region or across multiple regions.

role

The name or full Amazon Resource Name (ARN) of the IAM role that allows your Amazon ECS container agent to make calls to your load balancer on your behalf. This parameter is only required if you are using a load balancer with your service.

clientToken

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. Up to 32 ASCII characters are allowed.

taskDefinition

The family and revision (family:revision) or full Amazon Resource Name (ARN) of the task definition to run in your service. If a revision is not specified, the latest ACTIVE revision is used.