Struct rusoto::ecs::Deployment

pub struct Deployment {
    pub status: Option<String>,
    pub pendingCount: Option<Integer>,
    pub createdAt: Option<Timestamp>,
    pub desiredCount: Option<Integer>,
    pub taskDefinition: Option<String>,
    pub updatedAt: Option<Timestamp>,
    pub id: Option<String>,
    pub runningCount: Option<Integer>,
}

The details of an Amazon ECS service deployment.

Fields

status

The status of the deployment. Valid values are PRIMARY (for the most recent deployment), ACTIVE (for previous deployments that still have tasks running, but are being replaced with the PRIMARY deployment), and INACTIVE (for deployments that have been completely replaced).

pendingCount

The number of tasks in the deployment that are in the PENDING status.

createdAt

The Unix time in seconds and milliseconds when the service was created.

desiredCount

The most recent desired count of tasks that was specified for the service to deploy or maintain.

taskDefinition

The most recent task definition that was specified for the service to use.

updatedAt

The Unix time in seconds and milliseconds when the service was last updated.

id

The ID of the deployment.

runningCount

The number of tasks in the deployment that are in the RUNNING status.