Struct rusoto::ecs::ECSClient

pub struct ECSClient<'a> {
    // some fields omitted
}

Methods

impl<'a> ECSClient<'a>

fn new<P: AWSCredentialsProvider + 'a>(creds: P, region: &'a Region) -> ECSClient<'a>

fn create_service(&mut self, input: &CreateServiceRequest) -> Result<CreateServiceResponse>

Runs and maintains a desired number of tasks from a specified task definition. If the number of tasks running in a service drops below desiredCount, Amazon ECS spawns another instantiation of the task in the specified cluster.

fn submit_container_state_change(&mut self, input: &SubmitContainerStateChangeRequest) -> Result<SubmitContainerStateChangeResponse>

This action is only used by the Amazon EC2 Container Service agent, and it is not intended for use outside of the agent. Sent to acknowledge that a container changed states.

fn create_cluster(&mut self, input: &CreateClusterRequest) -> Result<CreateClusterResponse>

Creates a new Amazon ECS cluster. By default, your account receives a default cluster when you launch your first container instance. However, you can create your own cluster with a unique name with the CreateCluster action.

fn submit_task_state_change(&mut self, input: &SubmitTaskStateChangeRequest) -> Result<SubmitTaskStateChangeResponse>

This action is only used by the Amazon EC2 Container Service agent, and it is not intended for use outside of the agent. Sent to acknowledge that a task changed states.

fn list_task_definition_families(&mut self, input: &ListTaskDefinitionFamiliesRequest) -> Result<ListTaskDefinitionFamiliesResponse>

Returns a list of task definition families that are registered to your account (which may include task definition families that no longer have any ACTIVE task definitions). You can filter the results with the familyPrefix parameter.

fn delete_service(&mut self, input: &DeleteServiceRequest) -> Result<DeleteServiceResponse>

Deletes a specified service within a cluster.

fn list_container_instances(&mut self, input: &ListContainerInstancesRequest) -> Result<ListContainerInstancesResponse>

Returns a list of container instances in a specified cluster.

fn describe_services(&mut self, input: &DescribeServicesRequest) -> Result<DescribeServicesResponse>

Describes the specified services running in your cluster.

fn delete_cluster(&mut self, input: &DeleteClusterRequest) -> Result<DeleteClusterResponse>

Deletes the specified cluster. You must deregister all container instances from this cluster before you may delete it. You can list the container instances in a cluster with ListContainerInstances and deregister them with DeregisterContainerInstance.

fn update_service(&mut self, input: &UpdateServiceRequest) -> Result<UpdateServiceResponse>

Modify the desired count or task definition used in a service. You can add to or subtract from the number of instantiations of a task definition in a service by specifying the cluster that the service is running in and a new desiredCount parameter. You can use UpdateService to modify your task definition and deploy a new version of your service, one task at a time. If you modify the task definition with UpdateService, Amazon ECS spawns a task with the new version of the task definition and then stops an old task after the new version is running. Because UpdateService starts a new version of the task before stopping an old version, your cluster must have capacity to support one more instantiation of the task when UpdateService is run. If your cluster cannot support another instantiation of the task used in your service, you can reduce the desired count of your service by one before modifying the task definition. When UpdateService replaces a task during an update, the equivalent of docker stop is issued to the containers running in the task. This results in a SIGTERM and a 30-second timeout, after which SIGKILL is sent and the containers are forcibly stopped. If the container handles the SIGTERM gracefully and exits within 30 seconds from receiving it, no SIGKILL is sent.

fn list_clusters(&mut self, input: &ListClustersRequest) -> Result<ListClustersResponse>

Returns a list of existing clusters.

fn deregister_task_definition(&mut self, input: &DeregisterTaskDefinitionRequest) -> Result<DeregisterTaskDefinitionResponse>

Deregisters the specified task definition by family and revision. Upon deregistration, the task definition is marked as INACTIVE. Existing tasks and services that reference an INACTIVE task definition continue to run without disruption. Existing services that reference an INACTIVE task definition can still scale up or down by modifying the service's desired count. You cannot use an INACTIVE task definition to run new tasks or create new services, and you cannot update an existing service to reference an INACTIVE task definition (although there may be up to a 10 minute window following deregistration where these restrictions have not yet taken effect).

fn register_task_definition(&mut self, input: &RegisterTaskDefinitionRequest) -> Result<RegisterTaskDefinitionResponse>

Registers a new task definition from the supplied family and containerDefinitions. Optionally, you can add data volumes to your containers with the volumes parameter. For more information about task definition parameters and defaults, see Amazon ECS Task Definitions in the Amazon EC2 Container Service Developer Guide.

fn stop_task(&mut self, input: &StopTaskRequest) -> Result<StopTaskResponse>

Stops a running task. When StopTask is called on a task, the equivalent of docker stop is issued to the containers running in the task. This results in a SIGTERM and a 30-second timeout, after which SIGKILL is sent and the containers are forcibly stopped. If the container handles the SIGTERM gracefully and exits within 30 seconds from receiving it, no SIGKILL is sent.

fn list_task_definitions(&mut self, input: &ListTaskDefinitionsRequest) -> Result<ListTaskDefinitionsResponse>

Returns a list of task definitions that are registered to your account. You can filter the results by family name with the familyPrefix parameter or by status with the status parameter.

fn describe_clusters(&mut self, input: &DescribeClustersRequest) -> Result<DescribeClustersResponse>

Describes one or more of your clusters.

fn update_container_agent(&mut self, input: &UpdateContainerAgentRequest) -> Result<UpdateContainerAgentResponse>

Updates the Amazon ECS container agent on a specified container instance. Updating the Amazon ECS container agent does not interrupt running tasks or services on the container instance. The process for updating the agent differs depending on whether your container instance was launched with the Amazon ECS- optimized AMI or another operating system. UpdateContainerAgent requires the Amazon ECS-optimized AMI or Amazon Linux with the ecs-init service installed and running. For help updating the Amazon ECS container agent on other operating systems, see Manually Updating the Amazon ECS Container Agent in the Amazon EC2 Container Service Developer Guide.

fn list_services(&mut self, input: &ListServicesRequest) -> Result<ListServicesResponse>

Lists the services that are running in a specified cluster.

fn start_task(&mut self, input: &StartTaskRequest) -> Result<StartTaskResponse>

Starts a new task from the specified task definition on the specified container instance or instances. To use the default Amazon ECS scheduler to place your task, use RunTask instead. The list of container instances to start tasks on is limited to 10.

fn register_container_instance(&mut self, input: &RegisterContainerInstanceRequest) -> Result<RegisterContainerInstanceResponse>

This action is only used by the Amazon EC2 Container Service agent, and it is not intended for use outside of the agent. Registers an EC2 instance into the specified cluster. This instance becomes available to place containers on.

fn describe_task_definition(&mut self, input: &DescribeTaskDefinitionRequest) -> Result<DescribeTaskDefinitionResponse>

Describes a task definition. You can specify a family and revision to find information about a specific task definition, or you can simply specify the family to find the latest ACTIVE revision in that family. You can only describe INACTIVE task definitions while an active task or service references them.

fn describe_tasks(&mut self, input: &DescribeTasksRequest) -> Result<DescribeTasksResponse>

Describes a specified task or tasks.

fn deregister_container_instance(&mut self, input: &DeregisterContainerInstanceRequest) -> Result<DeregisterContainerInstanceResponse>

Deregisters an Amazon ECS container instance from the specified cluster. This instance is no longer available to run tasks. If you intend to use the container instance for some other purpose after deregistration, you should stop all of the tasks running on the container instance before deregistration to avoid any orphaned tasks from consuming resources. Deregistering a container instance removes the instance from a cluster, but it does not terminate the EC2 instance; if you are finished using the instance, be sure to terminate it in the Amazon EC2 console to stop billing. When you terminate a container instance, it is automatically deregistered from your cluster.

fn list_tasks(&mut self, input: &ListTasksRequest) -> Result<ListTasksResponse>

Returns a list of tasks for a specified cluster. You can filter the results by family name, by a particular container instance, or by the desired status of the task with the family, containerInstance, and desiredStatus parameters.

fn describe_container_instances(&mut self, input: &DescribeContainerInstancesRequest) -> Result<DescribeContainerInstancesResponse>

Describes Amazon EC2 Container Service container instances. Returns metadata about registered and remaining resources on each container instance requested.

fn run_task(&mut self, input: &RunTaskRequest) -> Result<RunTaskResponse>

Start a task using random placement and the default Amazon ECS scheduler. To use your own scheduler or place a task on a specific container instance, use StartTask instead. The count parameter is limited to 10 tasks per call.

fn discover_poll_endpoint(&mut self, input: &DiscoverPollEndpointRequest) -> Result<DiscoverPollEndpointResponse>

This action is only used by the Amazon EC2 Container Service agent, and it is not intended for use outside of the agent. Returns an endpoint for the Amazon EC2 Container Service agent to poll for updates.