Struct rusoto::ecs::EcsClient

pub struct EcsClient<P, D> where P: ProvideAwsCredentials,
        D: DispatchSignedRequest { /* fields omitted */ }

A client for the Amazon ECS API.

Methods

impl<P, D> EcsClient<P, D> where P: ProvideAwsCredentials,
D: DispatchSignedRequest

fn new(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn create_cluster(&self,
input: &CreateClusterRequest)
-> Result<CreateClusterResponse, CreateClusterError>

Creates a new Amazon ECS cluster. By default, your account receives a default cluster when you launch your first container instance. However, you can create your own cluster with a unique name with the CreateCluster action.

fn create_service(&self,
input: &CreateServiceRequest)
-> Result<CreateServiceResponse, CreateServiceError>

Runs and maintains a desired number of tasks from a specified task definition. If the number of tasks running in a service drops below desiredCount, Amazon ECS spawns another copy of the task in the specified cluster. To update an existing service, see UpdateService.

In addition to maintaining the desired count of tasks in your service, you can optionally run your service behind a load balancer. The load balancer distributes traffic across the tasks that are associated with the service. For more information, see Service Load Balancing in the Amazon EC2 Container Service Developer Guide.

You can optionally specify a deployment configuration for your service. During a deployment (which is triggered by changing the task definition or the desired count of a service with an UpdateService operation), the service scheduler uses the minimumHealthyPercent and maximumPercent parameters to determine the deployment strategy.

The minimumHealthyPercent represents a lower limit on the number of your service's tasks that must remain in the RUNNING state during a deployment, as a percentage of the desiredCount (rounded up to the nearest integer). This parameter enables you to deploy without using additional cluster capacity. For example, if your service has a desiredCount of four tasks and a minimumHealthyPercent of 50%, the scheduler may stop two existing tasks to free up cluster capacity before starting two new tasks. Tasks for services that do not use a load balancer are considered healthy if they are in the RUNNING state; tasks for services that do use a load balancer are considered healthy if they are in the RUNNING state and the container instance it is hosted on is reported as healthy by the load balancer. The default value for minimumHealthyPercent is 50% in the console and 100% for the AWS CLI, the AWS SDKs, and the APIs.

The maximumPercent parameter represents an upper limit on the number of your service's tasks that are allowed in the RUNNING or PENDING state during a deployment, as a percentage of the desiredCount (rounded down to the nearest integer). This parameter enables you to define the deployment batch size. For example, if your service has a desiredCount of four tasks and a maximumPercent value of 200%, the scheduler may start four new tasks before stopping the four older tasks (provided that the cluster resources required to do this are available). The default value for maximumPercent is 200%.

When the service scheduler launches new tasks, it determines task placement in your cluster with the following logic:

Determine which of the container instances in your cluster can support your service's task definition (for example, they have the required CPU, memory, ports, and container instance attributes).

By default, the service scheduler attempts to balance tasks across Availability Zones in this manner (although you can choose a different placement strategy with the placementStrategy parameter):

Sort the valid container instances by the fewest number of running tasks for this service in the same Availability Zone as the instance. For example, if zone A has one running service task and zones B and C each have zero, valid container instances in either zone B or C are considered optimal for placement.

Place the new service task on a valid container instance in an optimal Availability Zone (based on the previous steps), favoring container instances with the fewest number of running tasks for this service.

fn delete_attributes(&self,
input: &DeleteAttributesRequest)
-> Result<DeleteAttributesResponse, DeleteAttributesError>

Deletes one or more attributes from an Amazon ECS resource.

fn delete_cluster(&self,
input: &DeleteClusterRequest)
-> Result<DeleteClusterResponse, DeleteClusterError>

Deletes the specified cluster. You must deregister all container instances from this cluster before you may delete it. You can list the container instances in a cluster with ListContainerInstances and deregister them with DeregisterContainerInstance.

fn delete_service(&self,
input: &DeleteServiceRequest)
-> Result<DeleteServiceResponse, DeleteServiceError>

Deletes a specified service within a cluster. You can delete a service if you have no running tasks in it and the desired task count is zero. If the service is actively maintaining tasks, you cannot delete it, and you must update the service to a desired task count of zero. For more information, see UpdateService.

When you delete a service, if there are still running tasks that require cleanup, the service status moves from ACTIVE to DRAINING, and the service is no longer visible in the console or in ListServices API operations. After the tasks have stopped, then the service status moves from DRAINING to INACTIVE. Services in the DRAINING or INACTIVE status can still be viewed with DescribeServices API operations; however, in the future, INACTIVE services may be cleaned up and purged from Amazon ECS record keeping, and DescribeServices API operations on those services will return a ServiceNotFoundException error.

fn deregister_container_instance(&self,
input: &DeregisterContainerInstanceRequest)
-> Result<DeregisterContainerInstanceResponse, DeregisterContainerInstanceError>

Deregisters an Amazon ECS container instance from the specified cluster. This instance is no longer available to run tasks.

If you intend to use the container instance for some other purpose after deregistration, you should stop all of the tasks running on the container instance before deregistration to avoid any orphaned tasks from consuming resources.

Deregistering a container instance removes the instance from a cluster, but it does not terminate the EC2 instance; if you are finished using the instance, be sure to terminate it in the Amazon EC2 console to stop billing.

If you terminate a running container instance, Amazon ECS automatically deregisters the instance from your cluster (stopped container instances or instances with disconnected agents are not automatically deregistered when terminated).

fn deregister_task_definition(&self,
input: &DeregisterTaskDefinitionRequest)
-> Result<DeregisterTaskDefinitionResponse, DeregisterTaskDefinitionError>

Deregisters the specified task definition by family and revision. Upon deregistration, the task definition is marked as INACTIVE. Existing tasks and services that reference an INACTIVE task definition continue to run without disruption. Existing services that reference an INACTIVE task definition can still scale up or down by modifying the service's desired count.

You cannot use an INACTIVE task definition to run new tasks or create new services, and you cannot update an existing service to reference an INACTIVE task definition (although there may be up to a 10 minute window following deregistration where these restrictions have not yet taken effect).

fn describe_clusters(&self,
input: &DescribeClustersRequest)
-> Result<DescribeClustersResponse, DescribeClustersError>

Describes one or more of your clusters.

fn describe_container_instances(&self,
input: &DescribeContainerInstancesRequest)
-> Result<DescribeContainerInstancesResponse, DescribeContainerInstancesError>

Describes Amazon EC2 Container Service container instances. Returns metadata about registered and remaining resources on each container instance requested.

fn describe_services(&self,
input: &DescribeServicesRequest)
-> Result<DescribeServicesResponse, DescribeServicesError>

Describes the specified services running in your cluster.

fn describe_task_definition(&self,
input: &DescribeTaskDefinitionRequest)
-> Result<DescribeTaskDefinitionResponse, DescribeTaskDefinitionError>

Describes a task definition. You can specify a family and revision to find information about a specific task definition, or you can simply specify the family to find the latest ACTIVE revision in that family.

You can only describe INACTIVE task definitions while an active task or service references them.

fn describe_tasks(&self,
input: &DescribeTasksRequest)
-> Result<DescribeTasksResponse, DescribeTasksError>

Describes a specified task or tasks.

fn discover_poll_endpoint(&self,
input: &DiscoverPollEndpointRequest)
-> Result<DiscoverPollEndpointResponse, DiscoverPollEndpointError>

This action is only used by the Amazon EC2 Container Service agent, and it is not intended for use outside of the agent.

Returns an endpoint for the Amazon EC2 Container Service agent to poll for updates.

fn list_attributes(&self,
input: &ListAttributesRequest)
-> Result<ListAttributesResponse, ListAttributesError>

Lists the attributes for Amazon ECS resources within a specified target type and cluster. When you specify a target type and cluster, LisAttributes returns a list of attribute objects, one for each attribute on each resource. You can filter the list of results to a single attribute name to only return results that have that name. You can also filter the results by attribute name and value, for example, to see which container instances in a cluster are running a Linux AMI (ecs.os-type=linux).

fn list_clusters(&self,
input: &ListClustersRequest)
-> Result<ListClustersResponse, ListClustersError>

Returns a list of existing clusters.

fn list_container_instances(&self,
input: &ListContainerInstancesRequest)
-> Result<ListContainerInstancesResponse, ListContainerInstancesError>

Returns a list of container instances in a specified cluster. You can filter the results of a ListContainerInstances operation with cluster query language statements inside the filter parameter. For more information, see Cluster Query Language in the Amazon EC2 Container Service Developer Guide.

fn list_services(&self,
input: &ListServicesRequest)
-> Result<ListServicesResponse, ListServicesError>

Lists the services that are running in a specified cluster.

fn list_task_definition_families(&self,
input: &ListTaskDefinitionFamiliesRequest)
-> Result<ListTaskDefinitionFamiliesResponse, ListTaskDefinitionFamiliesError>

Returns a list of task definition families that are registered to your account (which may include task definition families that no longer have any ACTIVE task definition revisions).

You can filter out task definition families that do not contain any ACTIVE task definition revisions by setting the status parameter to ACTIVE. You can also filter the results with the familyPrefix parameter.

fn list_task_definitions(&self,
input: &ListTaskDefinitionsRequest)
-> Result<ListTaskDefinitionsResponse, ListTaskDefinitionsError>

Returns a list of task definitions that are registered to your account. You can filter the results by family name with the familyPrefix parameter or by status with the status parameter.

fn list_tasks(&self,
input: &ListTasksRequest)
-> Result<ListTasksResponse, ListTasksError>

Returns a list of tasks for a specified cluster. You can filter the results by family name, by a particular container instance, or by the desired status of the task with the family, containerInstance, and desiredStatus parameters.

Recently-stopped tasks might appear in the returned results. Currently, stopped tasks appear in the returned results for at least one hour.

fn put_attributes(&self,
input: &PutAttributesRequest)
-> Result<PutAttributesResponse, PutAttributesError>

Create or update an attribute on an Amazon ECS resource. If the attribute does not already exist on the given target, it is created; if it does exist, it is replaced with the new value.

fn register_container_instance(&self,
input: &RegisterContainerInstanceRequest)
-> Result<RegisterContainerInstanceResponse, RegisterContainerInstanceError>

This action is only used by the Amazon EC2 Container Service agent, and it is not intended for use outside of the agent.

Registers an EC2 instance into the specified cluster. This instance becomes available to place containers on.

fn register_task_definition(&self,
input: &RegisterTaskDefinitionRequest)
-> Result<RegisterTaskDefinitionResponse, RegisterTaskDefinitionError>

Registers a new task definition from the supplied family and containerDefinitions. Optionally, you can add data volumes to your containers with the volumes parameter. For more information about task definition parameters and defaults, see Amazon ECS Task Definitions in the Amazon EC2 Container Service Developer Guide.

You can specify an IAM role for your task with the taskRoleArn parameter. When you specify an IAM role for a task, its containers can then use the latest versions of the AWS CLI or SDKs to make API requests to the AWS services that are specified in the IAM policy associated with the role. For more information, see IAM Roles for Tasks in the Amazon EC2 Container Service Developer Guide.

You can specify a Docker networking mode for the containers in your task definition with the networkMode parameter. The available network modes correspond to those described in Network settings in the Docker run reference.

fn run_task(&self,
input: &RunTaskRequest)
-> Result<RunTaskResponse, RunTaskError>

Starts a new task using the specified task definition.

You can allow Amazon ECS to place tasks for you, or you can customize how Amazon ECS places tasks using placement constraints and placement strategies. For more information, see Scheduling Tasks in the Amazon EC2 Container Service Developer Guide.

Alternatively, you can use StartTask to use your own scheduler or place tasks manually on specific container instances.

fn start_task(&self,
input: &StartTaskRequest)
-> Result<StartTaskResponse, StartTaskError>

Starts a new task from the specified task definition on the specified container instance or instances.

Alternatively, you can use RunTask to place tasks for you. For more information, see Scheduling Tasks in the Amazon EC2 Container Service Developer Guide.

fn stop_task(&self,
input: &StopTaskRequest)
-> Result<StopTaskResponse, StopTaskError>

Stops a running task.

When StopTask is called on a task, the equivalent of docker stop is issued to the containers running in the task. This results in a SIGTERM and a 30-second timeout, after which SIGKILL is sent and the containers are forcibly stopped. If the container handles the SIGTERM gracefully and exits within 30 seconds from receiving it, no SIGKILL is sent.

fn submit_container_state_change(&self,
input: &SubmitContainerStateChangeRequest)
-> Result<SubmitContainerStateChangeResponse, SubmitContainerStateChangeError>

This action is only used by the Amazon EC2 Container Service agent, and it is not intended for use outside of the agent.

Sent to acknowledge that a container changed states.

fn submit_task_state_change(&self,
input: &SubmitTaskStateChangeRequest)
-> Result<SubmitTaskStateChangeResponse, SubmitTaskStateChangeError>

This action is only used by the Amazon EC2 Container Service agent, and it is not intended for use outside of the agent.

Sent to acknowledge that a task changed states.

fn update_container_agent(&self,
input: &UpdateContainerAgentRequest)
-> Result<UpdateContainerAgentResponse, UpdateContainerAgentError>

Updates the Amazon ECS container agent on a specified container instance. Updating the Amazon ECS container agent does not interrupt running tasks or services on the container instance. The process for updating the agent differs depending on whether your container instance was launched with the Amazon ECS-optimized AMI or another operating system.

UpdateContainerAgent requires the Amazon ECS-optimized AMI or Amazon Linux with the ecs-init service installed and running. For help updating the Amazon ECS container agent on other operating systems, see Manually Updating the Amazon ECS Container Agent in the Amazon EC2 Container Service Developer Guide.

fn update_service(&self,
input: &UpdateServiceRequest)
-> Result<UpdateServiceResponse, UpdateServiceError>

Modifies the desired count, deployment configuration, or task definition used in a service.

You can add to or subtract from the number of instantiations of a task definition in a service by specifying the cluster that the service is running in and a new desiredCount parameter.

You can use UpdateService to modify your task definition and deploy a new version of your service.

You can also update the deployment configuration of a service. When a deployment is triggered by updating the task definition of a service, the service scheduler uses the deployment configuration parameters, minimumHealthyPercent and maximumPercent, to determine the deployment strategy.

If the minimumHealthyPercent is below 100%, the scheduler can ignore the desiredCount temporarily during a deployment. For example, if your service has a desiredCount of four tasks, a minimumHealthyPercent of 50% allows the scheduler to stop two existing tasks before starting two new tasks. Tasks for services that do not use a load balancer are considered healthy if they are in the RUNNING state; tasks for services that do use a load balancer are considered healthy if they are in the RUNNING state and the container instance it is hosted on is reported as healthy by the load balancer.

The maximumPercent parameter represents an upper limit on the number of running tasks during a deployment, which enables you to define the deployment batch size. For example, if your service has a desiredCount of four tasks, a maximumPercent value of 200% starts four new tasks before stopping the four older tasks (provided that the cluster resources required to do this are available).

When UpdateService stops a task during a deployment, the equivalent of docker stop is issued to the containers running in the task. This results in a SIGTERM and a 30-second timeout, after which SIGKILL is sent and the containers are forcibly stopped. If the container handles the SIGTERM gracefully and exits within 30 seconds from receiving it, no SIGKILL is sent.

When the service scheduler launches new tasks, it determines task placement in your cluster with the following logic:

Determine which of the container instances in your cluster can support your service's task definition (for example, they have the required CPU, memory, ports, and container instance attributes).

By default, the service scheduler attempts to balance tasks across Availability Zones in this manner (although you can choose a different placement strategy with the placementStrategy parameter):

Sort the valid container instances by the fewest number of running tasks for this service in the same Availability Zone as the instance. For example, if zone A has one running service task and zones B and C each have zero, valid container instances in either zone B or C are considered optimal for placement.

Place the new service task on a valid container instance in an optimal Availability Zone (based on the previous steps), favoring container instances with the fewest number of running tasks for this service.

When the service scheduler stops running tasks, it attempts to maintain balance across the Availability Zones in your cluster with the following logic:

Sort the container instances by the largest number of running tasks for this service in the same Availability Zone as the instance. For example, if zone A has one running service task and zones B and C each have two, container instances in either zone B or C are considered optimal for termination.

Stop the task on a container instance in an optimal Availability Zone (based on the previous steps), favoring container instances with the largest number of running tasks for this service.