Struct rusoto::ecs::ListTasksRequest

pub struct ListTasksRequest {
    pub cluster: Option<String>,
    pub container_instance: Option<String>,
    pub desired_status: Option<DesiredStatus>,
    pub family: Option<String>,
    pub max_results: Option<BoxedInteger>,
    pub next_token: Option<String>,
    pub service_name: Option<String>,
    pub started_by: Option<String>,
}

Fields
cluster: Option<String>

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the tasks to list. If you do not specify a cluster, the default cluster is assumed..

container_instance: Option<String>

The container instance ID or full Amazon Resource Name (ARN) of the container instance with which to filter the ListTasks results. Specifying a containerInstance limits the results to tasks that belong to that container instance.

desired_status: Option<DesiredStatus>

The task status with which to filter the ListTasks results. Specifying a desiredStatus of STOPPED limits the results to tasks that are in the STOPPED status, which can be useful for debugging tasks that are not starting properly or have died or finished. The default status filter is RUNNING.

family: Option<String>

The name of the family with which to filter the ListTasks results. Specifying a family limits the results to tasks that belong to that family.

max_results: Option<BoxedInteger>

The maximum number of task results returned by ListTasks in paginated output. When this parameter is used, ListTasks only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListTasks request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListTasks returns up to 100 results and a nextToken value if applicable.

next_token: Option<String>

The nextToken value returned from a previous paginated ListTasks request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

service_name: Option<String>

The name of the service with which to filter the ListTasks results. Specifying a serviceName limits the results to tasks that belong to that service.

started_by: Option<String>

The startedBy value with which to filter the task results. Specifying a startedBy value limits the results to tasks that were started with that value.