Struct rusoto::ecs::RegisterTaskDefinitionRequest

pub struct RegisterTaskDefinitionRequest {
    pub container_definitions: ContainerDefinitions,
    pub family: String,
    pub volumes: Option<VolumeList>,
}

Fields
container_definitions: ContainerDefinitions

A list of container definitions in JSON format that describe the different containers that make up your task.

family: String

You must specify a family for a task definition, which allows you to track multiple versions of the same task definition. The family is used as a name for your task definition. Up to 255 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

volumes: Option<VolumeList>

A list of volume definitions in JSON format that containers in your task may use.