Struct rusoto::ecs::Service

pub struct Service {
    pub status: Option<String>,
    pub taskDefinition: Option<String>,
    pub pendingCount: Option<Integer>,
    pub loadBalancers: Option<LoadBalancers>,
    pub roleArn: Option<String>,
    pub desiredCount: Option<Integer>,
    pub serviceName: Option<String>,
    pub clusterArn: Option<String>,
    pub serviceArn: Option<String>,
    pub deployments: Option<Deployments>,
    pub events: Option<ServiceEvents>,
    pub runningCount: Option<Integer>,
}

Details on a service within a cluster

Fields

status

The status of the service. The valid values are ACTIVE, DRAINING, or INACTIVE.

taskDefinition

The task definition to use for tasks in the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

pendingCount

The number of tasks in the cluster that are in the PENDING state.

loadBalancers

A list of load balancer objects, containing the load balancer name, the container name (as it appears in a container definition), and the container port to access from the load balancer.

roleArn

The Amazon Resource Name (ARN) of the IAM role associated with the service that allows the Amazon ECS container agent to register container instances with a load balancer.

desiredCount

The desired number of instantiations of the task definition to keep running on the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

serviceName

A user-generated string that you can use to identify your service.

clusterArn

The Amazon Resource Name (ARN) of the of the cluster that hosts the service.

serviceArn

The Amazon Resource Name (ARN) that identifies the service. The ARN contains the arn:aws:ecs namespace, followed by the region of the service, the AWS account ID of the service owner, the service namespace, and then the service name. For example, arn:aws:ecs:_region_:_012345678910_:service/_my-service_.

deployments

The current state of deployments for the service.

events

The event stream for your service. A maximum of 100 of the latest events are displayed.

runningCount

The number of tasks in the cluster that are in the RUNNING state.