Struct rusoto::ecs::StartTaskRequest

pub struct StartTaskRequest {
    pub cluster: Option<String>,
    pub overrides: Option<TaskOverride>,
    pub containerInstances: StringList,
    pub taskDefinition: String,
    pub startedBy: Option<String>,
}

Fields

cluster

The short name or full Amazon Resource Name (ARN) of the cluster on which to start your task. If you do not specify a cluster, the default cluster is assumed..

overrides

A list of container overrides in JSON format that specify the name of a container in the specified task definition and the overrides it should receive. You can override the default command for a container (that is specified in the task definition or Docker image) with a command override. You can also override existing environment variables (that are specified in the task definition or Docker image) on a container or add new environment variables to it with an environment override. A total of 8192 characters are allowed for overrides. This limit includes the JSON formatting characters of the override structure.

containerInstances

The container instance IDs or full Amazon Resource Name (ARN) entries for the container instances on which you would like to place your task. The list of container instances to start tasks on is limited to 10.

taskDefinition

The family and revision (family:revision) or full Amazon Resource Name (ARN) of the task definition to start. If a revision is not specified, the latest ACTIVE revision is used.

startedBy

An optional tag specified when a task is started. For example if you automatically trigger a task to run a batch process job, you could apply a unique identifier for that job to your task with the startedBy parameter. You can then identify which tasks belong to that job by filtering the results of a ListTasks call with the startedBy value. If a task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.