Struct rusoto::ecs::TaskDefinition

pub struct TaskDefinition {
    pub container_definitions: Option<ContainerDefinitions>,
    pub family: Option<String>,
    pub requires_attributes: Option<RequiresAttributes>,
    pub revision: Option<Integer>,
    pub status: Option<TaskDefinitionStatus>,
    pub task_definition_arn: Option<String>,
    pub volumes: Option<VolumeList>,
}

Details of a task definition.

Fields
container_definitions: Option<ContainerDefinitions>

A list of container definitions in JSON format that describe the different containers that make up your task. For more information about container definition parameters and defaults, see Amazon ECS Task Definitions in the Amazon EC2 Container Service Developer Guide.

family: Option<String>

The family of your task definition, used as the definition name.

requires_attributes: Option<RequiresAttributes>

The container instance attributes required by your task.

revision: Option<Integer>

The revision of the task in a particular family. The revision is a version number of a task definition in a family. When you register a task definition for the first time, the revision is 1; each time you register a new revision of a task definition in the same family, the revision value always increases by one (even if you have deregistered previous revisions in this family).

status: Option<TaskDefinitionStatus>

The status of the task definition.

task_definition_arn: Option<String>

The full Amazon Resource Name (ARN) of the task definition.

volumes: Option<VolumeList>

The list of volumes in a task. For more information about volume definition parameters and defaults, see Amazon ECS Task Definitions in the Amazon EC2 Container Service Developer Guide.