Struct rusoto::ecs::UpdateServiceRequest

pub struct UpdateServiceRequest {
    pub cluster: Option<String>,
    pub deployment_configuration: Option<DeploymentConfiguration>,
    pub desired_count: Option<BoxedInteger>,
    pub service: String,
    pub task_definition: Option<String>,
}

Fields
cluster: Option<String>

The short name or full Amazon Resource Name (ARN) of the cluster that your service is running on. If you do not specify a cluster, the default cluster is assumed.

deployment_configuration: Option<DeploymentConfiguration>

Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

desired_count: Option<BoxedInteger>

The number of instantiations of the task to place and keep running in your service.

service: String

The name of the service to update.

task_definition: Option<String>

The family and revision (family:revision) or full Amazon Resource Name (ARN) of the task definition to run in your service. If a revision is not specified, the latest ACTIVE revision is used. If you modify the task definition with UpdateService, Amazon ECS spawns a task with the new version of the task definition and then stops an old task after the new version is running.