Struct rusoto::elasticbeanstalk::ConfigurationOptionDescription

pub struct ConfigurationOptionDescription {
    pub change_severity: Option<ConfigurationOptionSeverity>,
    pub default_value: Option<ConfigurationOptionDefaultValue>,
    pub max_length: Option<OptionRestrictionMaxLength>,
    pub max_value: Option<OptionRestrictionMaxValue>,
    pub min_value: Option<OptionRestrictionMinValue>,
    pub name: Option<ConfigurationOptionName>,
    pub namespace: Option<OptionNamespace>,
    pub regex: Option<OptionRestrictionRegex>,
    pub user_defined: Option<UserDefinedOption>,
    pub value_options: Option<ConfigurationOptionPossibleValues>,
    pub value_type: Option<ConfigurationOptionValueType>,
}

Describes the possible values for a configuration option.

Fields
change_severity: Option<ConfigurationOptionSeverity>

An indication of which action is required if the value for this configuration option changes:

NoInterruption : There is no interruption to the environment or application availability.

RestartEnvironment : The environment is entirely restarted, all AWS resources are deleted and recreated, and the environment is unavailable during the process.

RestartApplicationServer : The environment is available the entire time. However, a short application outage occurs when the application servers on the running Amazon EC2 instances are restarted.

default_value: Option<ConfigurationOptionDefaultValue>

The default value for this configuration option.

max_length: Option<OptionRestrictionMaxLength>

If specified, the configuration option must be a string value no longer than this value.

max_value: Option<OptionRestrictionMaxValue>

If specified, the configuration option must be a numeric value less than this value.

min_value: Option<OptionRestrictionMinValue>

If specified, the configuration option must be a numeric value greater than this value.

name: Option<ConfigurationOptionName>

The name of the configuration option.

namespace: Option<OptionNamespace>

A unique namespace identifying the option's associated AWS resource.

regex: Option<OptionRestrictionRegex>

If specified, the configuration option must be a string value that satisfies this regular expression.

user_defined: Option<UserDefinedOption>

An indication of whether the user defined this configuration option:

true : This configuration option was defined by the user. It is a valid choice for specifying if this as an Option to Remove when updating configuration settings.

false : This configuration was not defined by the user.

Constraint: You can remove only UserDefined options from a configuration.

Valid Values: true | false

value_options: Option<ConfigurationOptionPossibleValues>

If specified, values for the configuration option are selected from this list.

value_type: Option<ConfigurationOptionValueType>

An indication of which type of values this option has and whether it is allowable to select one or more than one of the possible values:

Scalar : Values for this option are a single selection from the possible values, or an unformatted string, or numeric value governed by the MIN/MAX/Regex constraints.

List : Values for this option are multiple selections from the possible values.

Boolean : Values for this option are either true or false .

Json : Values for this option are a JSON representation of a ConfigDocument.