Struct rusoto::elasticbeanstalk::ElasticBeanstalkClient

pub struct ElasticBeanstalkClient<P, D> where P: ProvideAwsCredentials,
        D: DispatchSignedRequest { /* fields omitted */ }

A client for the Elastic Beanstalk API.

Methods

impl<P, D> ElasticBeanstalkClient<P, D> where P: ProvideAwsCredentials,
D: DispatchSignedRequest

fn new(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn abort_environment_update(&self,
input: &AbortEnvironmentUpdateMessage)
-> Result<(), AbortEnvironmentUpdateError>

Cancels in-progress environment configuration update or application version deployment.

fn apply_environment_managed_action(&self,
input: &ApplyEnvironmentManagedActionRequest)
-> Result<ApplyEnvironmentManagedActionResult, ApplyEnvironmentManagedActionError>

Applies a scheduled managed action immediately. A managed action can be applied only if its status is Scheduled. Get the status and action ID of a managed action with DescribeEnvironmentManagedActions.

fn check_dns_availability(&self,
input: &CheckDNSAvailabilityMessage)
-> Result<CheckDNSAvailabilityResultMessage, CheckDNSAvailabilityError>

Checks if the specified CNAME is available.

fn compose_environments(&self,
input: &ComposeEnvironmentsMessage)
-> Result<EnvironmentDescriptionsMessage, ComposeEnvironmentsError>

Create or update a group of environments that each run a separate component of a single application. Takes a list of version labels that specify application source bundles for each of the environments to create or update. The name of each environment and other required information must be included in the source bundles in an environment manifest named env.yaml. See Compose Environments for details.

fn create_application(&self,
input: &CreateApplicationMessage)
-> Result<ApplicationDescriptionMessage, CreateApplicationError>

Creates an application that has one configuration template named default and no application versions.

fn create_application_version(&self,
input: &CreateApplicationVersionMessage)
-> Result<ApplicationVersionDescriptionMessage, CreateApplicationVersionError>

Creates an application version for the specified application. You can create an application version from a source bundle in Amazon S3, a commit in AWS CodeCommit, or the output of an AWS CodeBuild build as follows:

Specify a commit in an AWS CodeCommit repository with SourceBuildInformation.

Specify a build in an AWS CodeBuild with SourceBuildInformation and BuildConfiguration.

Specify a source bundle in S3 with SourceBundle

Omit both SourceBuildInformation and SourceBundle to use the default sample application.

Once you create an application version with a specified Amazon S3 bucket and key location, you cannot change that Amazon S3 location. If you change the Amazon S3 location, you receive an exception when you attempt to launch an environment from the application version.

fn create_configuration_template(&self,
input: &CreateConfigurationTemplateMessage)
-> Result<ConfigurationSettingsDescription, CreateConfigurationTemplateError>

Creates a configuration template. Templates are associated with a specific application and are used to deploy different versions of the application with the same configuration settings.

Related Topics

DescribeConfigurationOptions

DescribeConfigurationSettings

ListAvailableSolutionStacks

fn create_environment(&self,
input: &CreateEnvironmentMessage)
-> Result<EnvironmentDescription, CreateEnvironmentError>

Launches an environment for the specified application using the specified configuration.

fn create_storage_location(&self)
-> Result<CreateStorageLocationResultMessage, CreateStorageLocationError>

Creates the Amazon S3 storage location for the account.

This location is used to store user log files.

fn delete_application(&self,
input: &DeleteApplicationMessage)
-> Result<(), DeleteApplicationError>

Deletes the specified application along with all associated versions and configurations. The application versions will not be deleted from your Amazon S3 bucket.

You cannot delete an application that has a running environment.

fn delete_application_version(&self,
input: &DeleteApplicationVersionMessage)
-> Result<(), DeleteApplicationVersionError>

Deletes the specified version from the specified application.

You cannot delete an application version that is associated with a running environment.

fn delete_configuration_template(&self,
input: &DeleteConfigurationTemplateMessage)
-> Result<(), DeleteConfigurationTemplateError>

Deletes the specified configuration template.

When you launch an environment using a configuration template, the environment gets a copy of the template. You can delete or modify the environment's copy of the template without affecting the running environment.

fn delete_environment_configuration(&self,
input: &DeleteEnvironmentConfigurationMessage)
-> Result<(), DeleteEnvironmentConfigurationError>

Deletes the draft configuration associated with the running environment.

Updating a running environment with any configuration changes creates a draft configuration set. You can get the draft configuration using DescribeConfigurationSettings while the update is in progress or if the update fails. The DeploymentStatus for the draft configuration indicates whether the deployment is in process or has failed. The draft configuration remains in existence until it is deleted with this action.

fn describe_application_versions(&self,
input: &DescribeApplicationVersionsMessage)
-> Result<ApplicationVersionDescriptionsMessage, DescribeApplicationVersionsError>

Retrieve a list of application versions.

fn describe_applications(&self,
input: &DescribeApplicationsMessage)
-> Result<ApplicationDescriptionsMessage, DescribeApplicationsError>

Returns the descriptions of existing applications.

fn describe_configuration_options(&self,
input: &DescribeConfigurationOptionsMessage)
-> Result<ConfigurationOptionsDescription, DescribeConfigurationOptionsError>

Describes the configuration options that are used in a particular configuration template or environment, or that a specified solution stack defines. The description includes the values the options, their default values, and an indication of the required action on a running environment if an option value is changed.

fn describe_configuration_settings(&self,
input: &DescribeConfigurationSettingsMessage)
-> Result<ConfigurationSettingsDescriptions, DescribeConfigurationSettingsError>

Returns a description of the settings for the specified configuration set, that is, either a configuration template or the configuration set associated with a running environment.

When describing the settings for the configuration set associated with a running environment, it is possible to receive two sets of setting descriptions. One is the deployed configuration set, and the other is a draft configuration of an environment that is either in the process of deployment or that failed to deploy.

Related Topics

DeleteEnvironmentConfiguration

fn describe_environment_health(&self,
input: &DescribeEnvironmentHealthRequest)
-> Result<DescribeEnvironmentHealthResult, DescribeEnvironmentHealthError>

Returns information about the overall health of the specified environment. The DescribeEnvironmentHealth operation is only available with AWS Elastic Beanstalk Enhanced Health.

fn describe_environment_managed_action_history(&self,
input: &DescribeEnvironmentManagedActionHistoryRequest)
-> Result<DescribeEnvironmentManagedActionHistoryResult, DescribeEnvironmentManagedActionHistoryError>

Lists an environment's completed and failed managed actions.

fn describe_environment_managed_actions(&self,
input: &DescribeEnvironmentManagedActionsRequest)
-> Result<DescribeEnvironmentManagedActionsResult, DescribeEnvironmentManagedActionsError>

Lists an environment's upcoming and in-progress managed actions.

fn describe_environment_resources(&self,
input: &DescribeEnvironmentResourcesMessage)
-> Result<EnvironmentResourceDescriptionsMessage, DescribeEnvironmentResourcesError>

Returns AWS resources for this environment.

fn describe_environments(&self,
input: &DescribeEnvironmentsMessage)
-> Result<EnvironmentDescriptionsMessage, DescribeEnvironmentsError>

Returns descriptions for existing environments.

fn describe_events(&self,
input: &DescribeEventsMessage)
-> Result<EventDescriptionsMessage, DescribeEventsError>

Returns list of event descriptions matching criteria up to the last 6 weeks.

This action returns the most recent 1,000 events from the specified NextToken.

fn describe_instances_health(&self,
input: &DescribeInstancesHealthRequest)
-> Result<DescribeInstancesHealthResult, DescribeInstancesHealthError>

Retrives detailed information about the health of instances in your AWS Elastic Beanstalk. This operation requires enhanced health reporting.

fn list_available_solution_stacks(&self)
-> Result<ListAvailableSolutionStacksResultMessage, ListAvailableSolutionStacksError>

Returns a list of the available solution stack names.

fn rebuild_environment(&self,
input: &RebuildEnvironmentMessage)
-> Result<(), RebuildEnvironmentError>

Deletes and recreates all of the AWS resources (for example: the Auto Scaling group, load balancer, etc.) for a specified environment and forces a restart.

fn request_environment_info(&self,
input: &RequestEnvironmentInfoMessage)
-> Result<(), RequestEnvironmentInfoError>

Initiates a request to compile the specified type of information of the deployed environment.

Setting the InfoType to tail compiles the last lines from the application server log files of every Amazon EC2 instance in your environment.

Setting the InfoType to bundle compresses the application server log files for every Amazon EC2 instance into a .zip file. Legacy and .NET containers do not support bundle logs.

Use RetrieveEnvironmentInfo to obtain the set of logs.

Related Topics

RetrieveEnvironmentInfo

fn restart_app_server(&self,
input: &RestartAppServerMessage)
-> Result<(), RestartAppServerError>

Causes the environment to restart the application container server running on each Amazon EC2 instance.

fn retrieve_environment_info(&self,
input: &RetrieveEnvironmentInfoMessage)
-> Result<RetrieveEnvironmentInfoResultMessage, RetrieveEnvironmentInfoError>

Retrieves the compiled information from a RequestEnvironmentInfo request.

Related Topics

RequestEnvironmentInfo

fn swap_environment_cnam_es(&self,
input: &SwapEnvironmentCNAMEsMessage)
-> Result<(), SwapEnvironmentCNAMEsError>

Swaps the CNAMEs of two environments.

fn terminate_environment(&self,
input: &TerminateEnvironmentMessage)
-> Result<EnvironmentDescription, TerminateEnvironmentError>

Terminates the specified environment.

fn update_application(&self,
input: &UpdateApplicationMessage)
-> Result<ApplicationDescriptionMessage, UpdateApplicationError>

Updates the specified application to have the specified properties.

If a property (for example, description) is not provided, the value remains unchanged. To clear these properties, specify an empty string.

fn update_application_resource_lifecycle(&self,
input: &UpdateApplicationResourceLifecycleMessage)
-> Result<ApplicationResourceLifecycleDescriptionMessage, UpdateApplicationResourceLifecycleError>

Modifies lifecycle settings for an application.

fn update_application_version(&self,
input: &UpdateApplicationVersionMessage)
-> Result<ApplicationVersionDescriptionMessage, UpdateApplicationVersionError>

Updates the specified application version to have the specified properties.

If a property (for example, description) is not provided, the value remains unchanged. To clear properties, specify an empty string.

fn update_configuration_template(&self,
input: &UpdateConfigurationTemplateMessage)
-> Result<ConfigurationSettingsDescription, UpdateConfigurationTemplateError>

Updates the specified configuration template to have the specified properties or configuration option values.

If a property (for example, ApplicationName) is not provided, its value remains unchanged. To clear such properties, specify an empty string.

Related Topics

DescribeConfigurationOptions

fn update_environment(&self,
input: &UpdateEnvironmentMessage)
-> Result<EnvironmentDescription, UpdateEnvironmentError>

Updates the environment description, deploys a new application version, updates the configuration settings to an entirely new configuration template, or updates select configuration option values in the running environment.

Attempting to update both the release and configuration is not allowed and AWS Elastic Beanstalk returns an InvalidParameterCombination error.

When updating the configuration settings to a new template or individual settings, a draft configuration is created and DescribeConfigurationSettings for this environment returns two setting descriptions with different DeploymentStatus values.

fn validate_configuration_settings(&self,
input: &ValidateConfigurationSettingsMessage)
-> Result<ConfigurationSettingsValidationMessages, ValidateConfigurationSettingsError>

Takes a set of configuration settings and either a configuration template or environment, and determines whether those values are valid.

This action returns a list of messages indicating any errors or warnings associated with the selection of option values.