Struct rusoto::elastictranscoder::CreatePresetRequest

pub struct CreatePresetRequest {
    pub audio: Option<AudioParameters>,
    pub container: PresetContainer,
    pub description: Option<Description>,
    pub name: Name,
    pub thumbnails: Option<Thumbnails>,
    pub video: Option<VideoParameters>,
}

The CreatePresetRequest structure.

Fields
audio: Option<AudioParameters>

A section of the request body that specifies the audio parameters.

container: PresetContainer

The container type for the output file. Valid values include flac, flv, fmp4, gif, mp3, mp4, mpg, mxf, oga, ogg, ts, and webm.

description: Option<Description>

A description of the preset.

name: Name

The name of the preset. We recommend that the name be unique within the AWS account, but uniqueness is not enforced.

thumbnails: Option<Thumbnails>

A section of the request body that specifies the thumbnail parameters, if any.

video: Option<VideoParameters>

A section of the request body that specifies the video parameters.