Struct rusoto::elastictranscoder::JobInput

pub struct JobInput {
    pub aspect_ratio: Option<AspectRatio>,
    pub container: Option<JobContainer>,
    pub detected_properties: Option<DetectedProperties>,
    pub encryption: Option<Encryption>,
    pub frame_rate: Option<FrameRate>,
    pub interlaced: Option<Interlaced>,
    pub key: Option<Key>,
    pub resolution: Option<Resolution>,
}

Information about the file that you're transcoding.

Fields
aspect_ratio: Option<AspectRatio>

The aspect ratio of the input file. If you want Elastic Transcoder to automatically detect the aspect ratio of the input file, specify auto. If you want to specify the aspect ratio for the output file, enter one of the following values:

1:1, 4:3, 3:2, 16:9

If you specify a value other than auto, Elastic Transcoder disables automatic detection of the aspect ratio.

container: Option<JobContainer>

The container type for the input file. If you want Elastic Transcoder to automatically detect the container type of the input file, specify auto. If you want to specify the container type for the input file, enter one of the following values:

3gp, aac, asf, avi, divx, flv, m4a, mkv, mov, mp3, mp4, mpeg, mpeg-ps, mpeg-ts, mxf, ogg, vob, wav, webm

detected_properties: Option<DetectedProperties>

The detected properties of the input file.

encryption: Option<Encryption>

The encryption settings, if any, that are used for decrypting your input files. If your input file is encrypted, you must specify the mode that Elastic Transcoder will use to decrypt your file.

frame_rate: Option<FrameRate>

The frame rate of the input file. If you want Elastic Transcoder to automatically detect the frame rate of the input file, specify auto. If you want to specify the frame rate for the input file, enter one of the following values:

10, 15, 23.97, 24, 25, 29.97, 30, 60

If you specify a value other than auto, Elastic Transcoder disables automatic detection of the frame rate.

interlaced: Option<Interlaced>

Whether the input file is interlaced. If you want Elastic Transcoder to automatically detect whether the input file is interlaced, specify auto. If you want to specify whether the input file is interlaced, enter one of the following values:

true, false

If you specify a value other than auto, Elastic Transcoder disables automatic detection of interlacing.

key: Option<Key>

The name of the file to transcode. Elsewhere in the body of the JSON block is the the ID of the pipeline to use for processing the job. The InputBucket object in that pipeline tells Elastic Transcoder which Amazon S3 bucket to get the file from.

If the file name includes a prefix, such as cooking/lasagna.mpg, include the prefix in the key. If the file isn't in the specified bucket, Elastic Transcoder returns an error.

resolution: Option<Resolution>

This value must be auto, which causes Elastic Transcoder to automatically detect the resolution of the input file.