Struct rusoto::elastictranscoder::ListJobsByStatusRequest

pub struct ListJobsByStatusRequest {
    pub ascending: Option<Ascending>,
    pub page_token: Option<Id>,
    pub status: JobStatus,
}

The ListJobsByStatusRequest structure.

Fields
ascending: Option<Ascending>

To list jobs in chronological order by the date and time that they were submitted, enter true. To list jobs in reverse chronological order, enter false.

page_token: Option<Id>

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

status: JobStatus

To get information about all of the jobs associated with the current AWS account that have a given status, specify the following status: Submitted, Progressing, Complete, Canceled, or Error.