Struct rusoto::elastictranscoder::ListPipelinesRequest

pub struct ListPipelinesRequest {
    pub ascending: Option<Ascending>,
    pub page_token: Option<Id>,
}

The ListPipelineRequest structure.

Fields
ascending: Option<Ascending>

To list pipelines in chronological order by the date and time that they were created, enter true. To list pipelines in reverse chronological order, enter false.

page_token: Option<Id>

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.