Struct rusoto::elastictranscoder::Pipeline

pub struct Pipeline {
    pub arn: Option<String>,
    pub aws_kms_key_arn: Option<KeyArn>,
    pub content_config: Option<PipelineOutputConfig>,
    pub id: Option<Id>,
    pub input_bucket: Option<BucketName>,
    pub name: Option<Name>,
    pub notifications: Option<Notifications>,
    pub output_bucket: Option<BucketName>,
    pub role: Option<Role>,
    pub status: Option<PipelineStatus>,
    pub thumbnail_config: Option<PipelineOutputConfig>,
}

The pipeline (queue) that is used to manage jobs.

Fields
arn: Option<String>

The Amazon Resource Name (ARN) for the pipeline.

aws_kms_key_arn: Option<KeyArn>

The AWS Key Management Service (AWS KMS) key that you want to use with this pipeline.

If you use either S3 or S3-AWS-KMS as your Encryption:Mode, you don't need to provide a key with your job because a default key, known as an AWS-KMS key, is created for you automatically. You need to provide an AWS-KMS key only if you want to use a non-default AWS-KMS key, or if you are using an Encryption:Mode of AES-PKCS7, AES-CTR, or AES-GCM.

content_config: Option<PipelineOutputConfig>

Information about the Amazon S3 bucket in which you want Elastic Transcoder to save transcoded files and playlists. Either you specify both ContentConfig and ThumbnailConfig, or you specify OutputBucket.

Bucket: The Amazon S3 bucket in which you want Elastic Transcoder to save transcoded files and playlists.

Permissions: A list of the users and/or predefined Amazon S3 groups you want to have access to transcoded files and playlists, and the type of access that you want them to have.

GranteeType: The type of value that appears in the Grantee object:

Canonical: Either the canonical user ID for an AWS account or an origin access identity for an Amazon CloudFront distribution.

Email: The registered email address of an AWS account.

Group: One of the following predefined Amazon S3 groups: AllUsers, AuthenticatedUsers, or LogDelivery.

Grantee: The AWS user or group that you want to have access to transcoded files and playlists.

Access: The permission that you want to give to the AWS user that is listed in Grantee. Valid values include:

READ: The grantee can read the objects and metadata for objects that Elastic Transcoder adds to the Amazon S3 bucket.

READ_ACP: The grantee can read the object ACL for objects that Elastic Transcoder adds to the Amazon S3 bucket.

WRITE_ACP: The grantee can write the ACL for the objects that Elastic Transcoder adds to the Amazon S3 bucket.

FULL_CONTROL: The grantee has READ, READ_ACP, and WRITE_ACP permissions for the objects that Elastic Transcoder adds to the Amazon S3 bucket.

StorageClass: The Amazon S3 storage class, Standard or ReducedRedundancy, that you want Elastic Transcoder to assign to the video files and playlists that it stores in your Amazon S3 bucket.

id: Option<Id>

The identifier for the pipeline. You use this value to identify the pipeline in which you want to perform a variety of operations, such as creating a job or a preset.

input_bucket: Option<BucketName>

The Amazon S3 bucket from which Elastic Transcoder gets media files for transcoding and the graphics files, if any, that you want to use for watermarks.

name: Option<Name>

The name of the pipeline. We recommend that the name be unique within the AWS account, but uniqueness is not enforced.

Constraints: Maximum 40 characters

notifications: Option<Notifications>

The Amazon Simple Notification Service (Amazon SNS) topic that you want to notify to report job status.
To receive notifications, you must also subscribe to the new topic in the Amazon SNS console.

Progressing (optional): The Amazon Simple Notification Service (Amazon SNS) topic that you want to notify when Elastic Transcoder has started to process the job.

Completed (optional): The Amazon SNS topic that you want to notify when Elastic Transcoder has finished processing the job.

Warning (optional): The Amazon SNS topic that you want to notify when Elastic Transcoder encounters a warning condition.

Error (optional): The Amazon SNS topic that you want to notify when Elastic Transcoder encounters an error condition.

output_bucket: Option<BucketName>

The Amazon S3 bucket in which you want Elastic Transcoder to save transcoded files, thumbnails, and playlists. Either you specify this value, or you specify both ContentConfig and ThumbnailConfig.

role: Option<Role>

The IAM Amazon Resource Name (ARN) for the role that Elastic Transcoder uses to transcode jobs for this pipeline.

status: Option<PipelineStatus>

The current status of the pipeline:

Active: The pipeline is processing jobs.

Paused: The pipeline is not currently processing jobs.

thumbnail_config: Option<PipelineOutputConfig>

Information about the Amazon S3 bucket in which you want Elastic Transcoder to save thumbnail files. Either you specify both ContentConfig and ThumbnailConfig, or you specify OutputBucket.

Bucket: The Amazon S3 bucket in which you want Elastic Transcoder to save thumbnail files.

Permissions: A list of the users and/or predefined Amazon S3 groups you want to have access to thumbnail files, and the type of access that you want them to have.

GranteeType: The type of value that appears in the Grantee object:

Canonical: Either the canonical user ID for an AWS account or an origin access identity for an Amazon CloudFront distribution. A canonical user ID is not the same as an AWS account number.

Email: The registered email address of an AWS account.

Group: One of the following predefined Amazon S3 groups: AllUsers, AuthenticatedUsers, or LogDelivery.

Grantee: The AWS user or group that you want to have access to thumbnail files.

Access: The permission that you want to give to the AWS user that is listed in Grantee. Valid values include:

READ: The grantee can read the thumbnails and metadata for thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

READ_ACP: The grantee can read the object ACL for thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

WRITE_ACP: The grantee can write the ACL for the thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

FULL_CONTROL: The grantee has READ, READ_ACP, and WRITE_ACP permissions for the thumbnails that Elastic Transcoder adds to the Amazon S3 bucket.

StorageClass: The Amazon S3 storage class, Standard or ReducedRedundancy, that you want Elastic Transcoder to assign to the thumbnails that it stores in your Amazon S3 bucket.