Struct rusoto::elastictranscoder::Preset

pub struct Preset {
    pub arn: Option<String>,
    pub audio: Option<AudioParameters>,
    pub container: Option<PresetContainer>,
    pub description: Option<Description>,
    pub id: Option<Id>,
    pub name: Option<Name>,
    pub thumbnails: Option<Thumbnails>,
    pub type_: Option<PresetType>,
    pub video: Option<VideoParameters>,
}

Presets are templates that contain most of the settings for transcoding media files from one format to another. Elastic Transcoder includes some default presets for common formats, for example, several iPod and iPhone versions. You can also create your own presets for formats that aren't included among the default presets. You specify which preset you want to use when you create a job.

Fields
arn: Option<String>

The Amazon Resource Name (ARN) for the preset.

audio: Option<AudioParameters>

A section of the response body that provides information about the audio preset values.

container: Option<PresetContainer>

The container type for the output file. Valid values include flac, flv, fmp4, gif, mp3, mp4, mpg, mxf, oga, ogg, ts, and webm.

description: Option<Description>

A description of the preset.

id: Option<Id>

Identifier for the new preset. You use this value to get settings for the preset or to delete it.

name: Option<Name>

The name of the preset.

thumbnails: Option<Thumbnails>

A section of the response body that provides information about the thumbnail preset values, if any.

type_: Option<PresetType>

Whether the preset is a default preset provided by Elastic Transcoder (System) or a preset that you have defined (Custom).

video: Option<VideoParameters>

A section of the response body that provides information about the video preset values.