Struct rusoto::elastictranscoder::TestRoleRequest

pub struct TestRoleRequest {
    pub input_bucket: BucketName,
    pub output_bucket: BucketName,
    pub role: Role,
    pub topics: SnsTopics,
}

The TestRoleRequest structure.

Fields
input_bucket: BucketName

The Amazon S3 bucket that contains media files to be transcoded. The action attempts to read from this bucket.

output_bucket: BucketName

The Amazon S3 bucket that Elastic Transcoder will write transcoded media files to. The action attempts to read from this bucket.

role: Role

The IAM Amazon Resource Name (ARN) for the role that you want Elastic Transcoder to test.

topics: SnsTopics

The ARNs of one or more Amazon Simple Notification Service (Amazon SNS) topics that you want the action to send a test notification to.