Module rusoto::elb

AWS Elastic Load Balancing

If you're using the service, you're probably looking for ElbClient.

Structs

AccessLog
Information about the AccessLog attribute.

AddAvailabilityZonesInput
Contains the parameters for EnableAvailabilityZonesForLoadBalancer.

AddAvailabilityZonesOutput
Contains the output of EnableAvailabilityZonesForLoadBalancer.

AddTagsInput
Contains the parameters for AddTags.

AddTagsOutput
Contains the output of AddTags.

AdditionalAttribute
This data type is reserved.

AppCookieStickinessPolicy
Information about a policy for application-controlled session stickiness.

ApplySecurityGroupsToLoadBalancerInput
Contains the parameters for ApplySecurityGroupsToLoadBalancer.

ApplySecurityGroupsToLoadBalancerOutput
Contains the output of ApplySecurityGroupsToLoadBalancer.

AttachLoadBalancerToSubnetsInput
Contains the parameters for AttachLoaBalancerToSubnets.

AttachLoadBalancerToSubnetsOutput
Contains the output of AttachLoadBalancerToSubnets.

BackendServerDescription
Information about the configuration of an EC2 instance.

ConfigureHealthCheckInput
Contains the parameters for ConfigureHealthCheck.

ConfigureHealthCheckOutput
Contains the output of ConfigureHealthCheck.

ConnectionDraining
Information about the ConnectionDraining attribute.

ConnectionSettings
Information about the ConnectionSettings attribute.

CreateAccessPointInput
Contains the parameters for CreateLoadBalancer.

CreateAccessPointOutput
Contains the output for CreateLoadBalancer.

CreateAppCookieStickinessPolicyInput
Contains the parameters for CreateAppCookieStickinessPolicy.

CreateAppCookieStickinessPolicyOutput
Contains the output for CreateAppCookieStickinessPolicy.

CreateLBCookieStickinessPolicyInput
Contains the parameters for CreateLBCookieStickinessPolicy.

CreateLBCookieStickinessPolicyOutput
Contains the output for CreateLBCookieStickinessPolicy.

CreateLoadBalancerListenerInput
Contains the parameters for CreateLoadBalancerListeners.

CreateLoadBalancerListenerOutput
Contains the parameters for CreateLoadBalancerListener.

CreateLoadBalancerPolicyInput
Contains the parameters for CreateLoadBalancerPolicy.

CreateLoadBalancerPolicyOutput
Contains the output of CreateLoadBalancerPolicy.

CrossZoneLoadBalancing
Information about the CrossZoneLoadBalancing attribute.

DeleteAccessPointInput
Contains the parameters for DeleteLoadBalancer.

DeleteAccessPointOutput
Contains the output of DeleteLoadBalancer.

DeleteLoadBalancerListenerInput
Contains the parameters for DeleteLoadBalancerListeners.

DeleteLoadBalancerListenerOutput
Contains the output of DeleteLoadBalancerListeners.

DeleteLoadBalancerPolicyInput
Contains the parameters for DeleteLoadBalancerPolicy.

DeleteLoadBalancerPolicyOutput
Contains the output of DeleteLoadBalancerPolicy.

DeregisterEndPointsInput
Contains the parameters for DeregisterInstancesFromLoadBalancer.

DeregisterEndPointsOutput
Contains the output of DeregisterInstancesFromLoadBalancer.

DescribeAccessPointsInput
Contains the parameters for DescribeLoadBalancers.

DescribeAccessPointsOutput
Contains the parameters for DescribeLoadBalancers.

DescribeEndPointStateInput
Contains the parameters for DescribeInstanceHealth.

DescribeEndPointStateOutput
Contains the output for DescribeInstanceHealth.

DescribeLoadBalancerAttributesInput
Contains the parameters for DescribeLoadBalancerAttributes.

DescribeLoadBalancerAttributesOutput
Contains the output of DescribeLoadBalancerAttributes.

DescribeLoadBalancerPoliciesInput
Contains the parameters for DescribeLoadBalancerPolicies.

DescribeLoadBalancerPoliciesOutput
Contains the output of DescribeLoadBalancerPolicies.

DescribeLoadBalancerPolicyTypesInput
Contains the parameters for DescribeLoadBalancerPolicyTypes.

DescribeLoadBalancerPolicyTypesOutput
Contains the output of DescribeLoadBalancerPolicyTypes.

DescribeTagsInput
Contains the parameters for DescribeTags.

DescribeTagsOutput
Contains the output for DescribeTags.

DetachLoadBalancerFromSubnetsInput
Contains the parameters for DetachLoadBalancerFromSubnets.

DetachLoadBalancerFromSubnetsOutput
Contains the output of DetachLoadBalancerFromSubnets.

ElbClient
A client for the Elastic Load Balancing API.

HealthCheck
Information about a health check.

Instance
The ID of an EC2 instance.

InstanceState
Information about the state of an EC2 instance.

LBCookieStickinessPolicy
Information about a policy for duration-based session stickiness.

Listener
Information about a listener.

For information about the protocols and the ports supported by Elastic Load Balancing, see Listeners for Your Classic Load Balancer in the Classic Load Balancers Guide.

ListenerDescription
The policies enabled for a listener.

LoadBalancerAttributes
The attributes for a load balancer.

LoadBalancerDescription
Information about a load balancer.

ModifyLoadBalancerAttributesInput
Contains the parameters for ModifyLoadBalancerAttributes.

ModifyLoadBalancerAttributesOutput
Contains the output of ModifyLoadBalancerAttributes.

Policies
The policies for a load balancer.

PolicyAttribute
Information about a policy attribute.

PolicyAttributeDescription
Information about a policy attribute.

PolicyAttributeTypeDescription
Information about a policy attribute type.

PolicyDescription
Information about a policy.

PolicyTypeDescription
Information about a policy type.

RegisterEndPointsInput
Contains the parameters for RegisterInstancesWithLoadBalancer.

RegisterEndPointsOutput
Contains the output of RegisterInstancesWithLoadBalancer.

RemoveAvailabilityZonesInput
Contains the parameters for DisableAvailabilityZonesForLoadBalancer.

RemoveAvailabilityZonesOutput
Contains the output for DisableAvailabilityZonesForLoadBalancer.

RemoveTagsInput
Contains the parameters for RemoveTags.

RemoveTagsOutput
Contains the output of RemoveTags.

SetLoadBalancerListenerSSLCertificateInput
Contains the parameters for SetLoadBalancerListenerSSLCertificate.

SetLoadBalancerListenerSSLCertificateOutput
Contains the output of SetLoadBalancerListenerSSLCertificate.

SetLoadBalancerPoliciesForBackendServerInput
Contains the parameters for SetLoadBalancerPoliciesForBackendServer.

SetLoadBalancerPoliciesForBackendServerOutput
Contains the output of SetLoadBalancerPoliciesForBackendServer.

SetLoadBalancerPoliciesOfListenerInput
Contains the parameters for SetLoadBalancePoliciesOfListener.

SetLoadBalancerPoliciesOfListenerOutput
Contains the output of SetLoadBalancePoliciesOfListener.

SourceSecurityGroup
Information about a source security group.

Tag
Information about a tag.

TagDescription
The tags associated with a load balancer.

TagKeyOnly
The key of a tag.

Enums

AddTagsError
Errors returned by AddTags

ApplySecurityGroupsToLoadBalancerError
Errors returned by ApplySecurityGroupsToLoadBalancer

AttachLoadBalancerToSubnetsError
Errors returned by AttachLoadBalancerToSubnets

ConfigureHealthCheckError
Errors returned by ConfigureHealthCheck

CreateAppCookieStickinessPolicyError
Errors returned by CreateAppCookieStickinessPolicy

CreateLBCookieStickinessPolicyError
Errors returned by CreateLBCookieStickinessPolicy

CreateLoadBalancerError
Errors returned by CreateLoadBalancer

CreateLoadBalancerListenersError
Errors returned by CreateLoadBalancerListeners

CreateLoadBalancerPolicyError
Errors returned by CreateLoadBalancerPolicy

DeleteLoadBalancerError
Errors returned by DeleteLoadBalancer

DeleteLoadBalancerListenersError
Errors returned by DeleteLoadBalancerListeners

DeleteLoadBalancerPolicyError
Errors returned by DeleteLoadBalancerPolicy

DeregisterInstancesFromLoadBalancerError
Errors returned by DeregisterInstancesFromLoadBalancer

DescribeInstanceHealthError
Errors returned by DescribeInstanceHealth

DescribeLoadBalancerAttributesError
Errors returned by DescribeLoadBalancerAttributes

DescribeLoadBalancerPoliciesError
Errors returned by DescribeLoadBalancerPolicies

DescribeLoadBalancerPolicyTypesError
Errors returned by DescribeLoadBalancerPolicyTypes

DescribeLoadBalancersError
Errors returned by DescribeLoadBalancers

DescribeTagsError
Errors returned by DescribeTags

DetachLoadBalancerFromSubnetsError
Errors returned by DetachLoadBalancerFromSubnets

DisableAvailabilityZonesForLoadBalancerError
Errors returned by DisableAvailabilityZonesForLoadBalancer

EnableAvailabilityZonesForLoadBalancerError
Errors returned by EnableAvailabilityZonesForLoadBalancer

ModifyLoadBalancerAttributesError
Errors returned by ModifyLoadBalancerAttributes

RegisterInstancesWithLoadBalancerError
Errors returned by RegisterInstancesWithLoadBalancer

RemoveTagsError
Errors returned by RemoveTags

SetLoadBalancerListenerSSLCertificateError
Errors returned by SetLoadBalancerListenerSSLCertificate

SetLoadBalancerPoliciesForBackendServerError
Errors returned by SetLoadBalancerPoliciesForBackendServer

SetLoadBalancerPoliciesOfListenerError
Errors returned by SetLoadBalancerPoliciesOfListener

Type Definitions

AccessLogEnabled

AccessLogInterval

AccessLogPrefix

AccessPointName

AccessPointPort

AdditionalAttributeKey

AdditionalAttributeValue

AdditionalAttributes

AppCookieStickinessPolicies

AttributeName

AttributeType

AttributeValue

AvailabilityZone

AvailabilityZones

BackendServerDescriptions

Cardinality

ConnectionDrainingEnabled

ConnectionDrainingTimeout

CookieExpirationPeriod

CookieName

CreatedTime

CrossZoneLoadBalancingEnabled

DNSName

DefaultValue

Description

EndPointPort

HealthCheckInterval

HealthCheckTarget

HealthCheckTimeout

HealthyThreshold

IdleTimeout

InstanceId

InstancePort

InstanceStates

Instances

LBCookieStickinessPolicies

ListenerDescriptions

Listeners

LoadBalancerDescriptions

LoadBalancerNames

LoadBalancerNamesMax20

LoadBalancerScheme

Marker

PageSize

PolicyAttributeDescriptions

PolicyAttributeTypeDescriptions

PolicyAttributes

PolicyDescriptions

PolicyName

PolicyNames

PolicyTypeDescriptions

PolicyTypeName

PolicyTypeNames

Ports

Protocol

ReasonCode

S3BucketName

SSLCertificateId

SecurityGroupId

SecurityGroupName

SecurityGroupOwnerAlias

SecurityGroups

State

SubnetId

Subnets

TagDescriptions

TagKey

TagKeyList

TagList

TagValue

UnhealthyThreshold

VPCId