Module rusoto::elbv2

AWS Application Load Balancing

If you're using the service, you're probably looking for ElbClient.

Structs

Action
Information about an action.

AddTagsInput
Contains the parameters for AddTags.

AddTagsOutput
Contains the output of AddTags.

AvailabilityZone
Information about an Availability Zone.

Certificate
Information about an SSL server certificate deployed on a load balancer.

Cipher
Information about a cipher used in a policy.

CreateListenerInput
Contains the parameters for CreateListener.

CreateListenerOutput
Contains the output of CreateListener.

CreateLoadBalancerInput
Contains the parameters for CreateLoadBalancer.

CreateLoadBalancerOutput
Contains the output of CreateLoadBalancer.

CreateRuleInput
Contains the parameters for CreateRule.

CreateRuleOutput
Contains the output of CreateRule.

CreateTargetGroupInput
Contains the parameters for CreateTargetGroup.

CreateTargetGroupOutput
Contains the output of CreateTargetGroup.

DeleteListenerInput
Contains the parameters for DeleteListener.

DeleteListenerOutput
Contains the output of DeleteListener.

DeleteLoadBalancerInput
Contains the parameters for DeleteLoadBalancer.

DeleteLoadBalancerOutput
Contains the output of DeleteLoadBalancer.

DeleteRuleInput
Contains the parameters for DeleteRule.

DeleteRuleOutput
Contains the output of DeleteRule.

DeleteTargetGroupInput
Contains the parameters for DeleteTargetGroup.

DeleteTargetGroupOutput
Contains the output of DeleteTargetGroup.

DeregisterTargetsInput
Contains the parameters for DeregisterTargets.

DeregisterTargetsOutput
Contains the output of DeregisterTargets.

DescribeListenersInput
Contains the parameters for DescribeListeners.

DescribeListenersOutput
Contains the output of DescribeListeners.

DescribeLoadBalancerAttributesInput
Contains the parameters for DescribeLoadBalancerAttributes.

DescribeLoadBalancerAttributesOutput
Contains the output of DescribeLoadBalancerAttributes.

DescribeLoadBalancersInput
Contains the parameters for DescribeLoadBalancers.

DescribeLoadBalancersOutput
Contains the output of DescribeLoadBalancers.

DescribeRulesInput
Contains the parameters for DescribeRules.

DescribeRulesOutput
Contains the output of DescribeRules.

DescribeSSLPoliciesInput
Contains the parameters for DescribeSSLPolicies.

DescribeSSLPoliciesOutput
Contains the output of DescribeSSLPolicies.

DescribeTagsInput
Contains the parameters for DescribeTags.

DescribeTagsOutput
Contains the output of DescribeTags.

DescribeTargetGroupAttributesInput
Contains the parameters for DescribeTargetGroupAttributes.

DescribeTargetGroupAttributesOutput
Contains the output of DescribeTargetGroupAttributes.

DescribeTargetGroupsInput
Contains the parameters for DescribeTargetGroups.

DescribeTargetGroupsOutput
Contains the output of DescribeTargetGroups.

DescribeTargetHealthInput
Contains the parameters for DescribeTargetHealth.

DescribeTargetHealthOutput
Contains the output of DescribeTargetHealth.

ElbClient
A client for the Elastic Load Balancing v2 API.

Listener
Information about a listener.

LoadBalancer
Information about a load balancer.

LoadBalancerAttribute
Information about a load balancer attribute.

LoadBalancerState
Information about the state of the load balancer.

Matcher
Information to use when checking for a successful response from a target.

ModifyListenerInput
Contains the parameters for ModifyListener.

ModifyListenerOutput
Contains the output of ModifyListener.

ModifyLoadBalancerAttributesInput
Contains the parameters for ModifyLoadBalancerAttributes.

ModifyLoadBalancerAttributesOutput
Contains the output of ModifyLoadBalancerAttributes.

ModifyRuleInput
Contains the parameters for ModifyRules.

ModifyRuleOutput
Contains the output of ModifyRules.

ModifyTargetGroupAttributesInput
Contains the parameters for ModifyTargetGroupAttributes.

ModifyTargetGroupAttributesOutput
Contains the output of ModifyTargetGroupAttributes.

ModifyTargetGroupInput
Contains the parameters for ModifyTargetGroup.

ModifyTargetGroupOutput
Contains the output of ModifyTargetGroup.

RegisterTargetsInput
Contains the parameters for RegisterTargets.

RegisterTargetsOutput
Contains the output of RegisterTargets.

RemoveTagsInput
Contains the parameters for RemoveTags.

RemoveTagsOutput
Contains the output of RemoveTags.

Rule
Information about a rule.

RuleCondition
Information about a condition for a rule.

RulePriorityPair
Information about the priorities for the rules for a listener.

SetRulePrioritiesInput
Contains the parameters for SetRulePriorities.

SetRulePrioritiesOutput
Contains the output of SetRulePriorities.

SetSecurityGroupsInput
Contains the parameters for SetSecurityGroups.

SetSecurityGroupsOutput
Contains the output of SetSecurityGroups.

SetSubnetsInput
Contains the parameters for SetSubnets.

SetSubnetsOutput
Contains the output of SetSubnets.

SslPolicy
Information about a policy used for SSL negotiation.

Tag
Information about a tag.

TagDescription
The tags associated with a resource.

TargetDescription
Information about a target.

TargetGroup
Information about a target group.

TargetGroupAttribute
Information about a target group attribute.

TargetHealth
Information about the current health of a target.

TargetHealthDescription
Information about the health of a target.

Enums

AddTagsError
Errors returned by AddTags

CreateListenerError
Errors returned by CreateListener

CreateLoadBalancerError
Errors returned by CreateLoadBalancer

CreateRuleError
Errors returned by CreateRule

CreateTargetGroupError
Errors returned by CreateTargetGroup

DeleteListenerError
Errors returned by DeleteListener

DeleteLoadBalancerError
Errors returned by DeleteLoadBalancer

DeleteRuleError
Errors returned by DeleteRule

DeleteTargetGroupError
Errors returned by DeleteTargetGroup

DeregisterTargetsError
Errors returned by DeregisterTargets

DescribeListenersError
Errors returned by DescribeListeners

DescribeLoadBalancerAttributesError
Errors returned by DescribeLoadBalancerAttributes

DescribeLoadBalancersError
Errors returned by DescribeLoadBalancers

DescribeRulesError
Errors returned by DescribeRules

DescribeSSLPoliciesError
Errors returned by DescribeSSLPolicies

DescribeTagsError
Errors returned by DescribeTags

DescribeTargetGroupAttributesError
Errors returned by DescribeTargetGroupAttributes

DescribeTargetGroupsError
Errors returned by DescribeTargetGroups

DescribeTargetHealthError
Errors returned by DescribeTargetHealth

ModifyListenerError
Errors returned by ModifyListener

ModifyLoadBalancerAttributesError
Errors returned by ModifyLoadBalancerAttributes

ModifyRuleError
Errors returned by ModifyRule

ModifyTargetGroupAttributesError
Errors returned by ModifyTargetGroupAttributes

ModifyTargetGroupError
Errors returned by ModifyTargetGroup

RegisterTargetsError
Errors returned by RegisterTargets

RemoveTagsError
Errors returned by RemoveTags

SetRulePrioritiesError
Errors returned by SetRulePriorities

SetSecurityGroupsError
Errors returned by SetSecurityGroups

SetSubnetsError
Errors returned by SetSubnets

Type Definitions

ActionTypeEnum

Actions

AvailabilityZones

CanonicalHostedZoneId

CertificateArn

CertificateList

CipherName

CipherPriority

Ciphers

ConditionFieldName

CreatedTime

DNSName

Description

HealthCheckIntervalSeconds

HealthCheckPort

HealthCheckThresholdCount

HealthCheckTimeoutSeconds

HttpCode

IsDefault

ListOfString

ListenerArn

ListenerArns

Listeners

LoadBalancerArn

LoadBalancerArns

LoadBalancerAttributeKey

LoadBalancerAttributeValue

LoadBalancerAttributes

LoadBalancerName

LoadBalancerNames

LoadBalancerSchemeEnum

LoadBalancerStateEnum

LoadBalancerTypeEnum

LoadBalancers

Marker

PageSize

Path

Port

ProtocolEnum

ResourceArn

ResourceArns

RuleArn

RuleArns

RuleConditionList

RulePriority

RulePriorityList

Rules

SecurityGroupId

SecurityGroups

SslPolicies

SslPolicyName

SslPolicyNames

SslProtocol

SslProtocols

StateReason

StringValue

SubnetId

Subnets

TagDescriptions

TagKey

TagKeys

TagList

TagValue

TargetDescriptions

TargetGroupArn

TargetGroupArns

TargetGroupAttributeKey

TargetGroupAttributeValue

TargetGroupAttributes

TargetGroupName

TargetGroupNames

TargetGroups

TargetHealthDescriptions

TargetHealthReasonEnum

TargetHealthStateEnum

TargetId

VpcId

ZoneName