Struct rusoto::elbv2::ElbClient

pub struct ElbClient<P, D> where P: ProvideAwsCredentials,
        D: DispatchSignedRequest { /* fields omitted */ }

A client for the Elastic Load Balancing v2 API.

Methods

impl<P, D> ElbClient<P, D> where P: ProvideAwsCredentials,
D: DispatchSignedRequest

fn new(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn add_tags(&self, input: &AddTagsInput) -> Result<AddTagsOutput, AddTagsError>

Adds the specified tags to the specified resource. You can tag your Application Load Balancers and your target groups.

Each tag consists of a key and an optional value. If a resource already has a tag with the same key, AddTags updates its value.

To list the current tags for your resources, use DescribeTags. To remove tags from your resources, use RemoveTags.

fn create_listener(&self,
input: &CreateListenerInput)
-> Result<CreateListenerOutput, CreateListenerError>

Creates a listener for the specified Application Load Balancer.

You can create up to 10 listeners per load balancer.

To update a listener, use ModifyListener. When you are finished with a listener, you can delete it using DeleteListener. If you are finished with both the listener and the load balancer, you can delete them both using DeleteLoadBalancer.

For more information, see Listeners for Your Application Load Balancers in the Application Load Balancers Guide.

fn create_load_balancer(&self,
input: &CreateLoadBalancerInput)
-> Result<CreateLoadBalancerOutput, CreateLoadBalancerError>

Creates an Application Load Balancer.

To create listeners for your load balancer, use CreateListener. You can add security groups, subnets, and tags when you create your load balancer, or you can add them later using SetSecurityGroups, SetSubnets, and AddTags.

To describe your current load balancers, see DescribeLoadBalancers. When you are finished with a load balancer, you can delete it using DeleteLoadBalancer.

You can create up to 20 load balancers per region per account. You can request an increase for the number of load balancers for your account. For more information, see Limits for Your Application Load Balancer in the Application Load Balancers Guide.

For more information, see Application Load Balancers in the Application Load Balancers Guide.

fn create_rule(&self,
input: &CreateRuleInput)
-> Result<CreateRuleOutput, CreateRuleError>

Creates a rule for the specified listener.

Each rule can have one action and one condition. Rules are evaluated in priority order, from the lowest value to the highest value. When the condition for a rule is met, the specified action is taken. If no conditions are met, the default action for the default rule is taken. For more information, see Listener Rules in the Application Load Balancers Guide.

To view your current rules, use DescribeRules. To update a rule, use ModifyRule. To set the priorities of your rules, use SetRulePriorities. To delete a rule, use DeleteRule.

fn create_target_group(&self,
input: &CreateTargetGroupInput)
-> Result<CreateTargetGroupOutput, CreateTargetGroupError>

Creates a target group.

To register targets with the target group, use RegisterTargets. To update the health check settings for the target group, use ModifyTargetGroup. To monitor the health of targets in the target group, use DescribeTargetHealth.

To route traffic to the targets in a target group, specify the target group in an action using CreateListener or CreateRule.

To delete a target group, use DeleteTargetGroup.

For more information, see Target Groups for Your Application Load Balancers in the Application Load Balancers Guide.

fn delete_listener(&self,
input: &DeleteListenerInput)
-> Result<DeleteListenerOutput, DeleteListenerError>

Deletes the specified listener.

Alternatively, your listener is deleted when you delete the load balancer it is attached to using DeleteLoadBalancer.

fn delete_load_balancer(&self,
input: &DeleteLoadBalancerInput)
-> Result<DeleteLoadBalancerOutput, DeleteLoadBalancerError>

Deletes the specified Application Load Balancer and its attached listeners.

You can't delete a load balancer if deletion protection is enabled. If the load balancer does not exist or has already been deleted, the call succeeds.

Deleting a load balancer does not affect its registered targets. For example, your EC2 instances continue to run and are still registered to their target groups. If you no longer need these EC2 instances, you can stop or terminate them.

fn delete_rule(&self,
input: &DeleteRuleInput)
-> Result<DeleteRuleOutput, DeleteRuleError>

Deletes the specified rule.

fn delete_target_group(&self,
input: &DeleteTargetGroupInput)
-> Result<DeleteTargetGroupOutput, DeleteTargetGroupError>

Deletes the specified target group.

You can delete a target group if it is not referenced by any actions. Deleting a target group also deletes any associated health checks.

fn deregister_targets(&self,
input: &DeregisterTargetsInput)
-> Result<DeregisterTargetsOutput, DeregisterTargetsError>

Deregisters the specified targets from the specified target group. After the targets are deregistered, they no longer receive traffic from the load balancer.

fn describe_listeners(&self,
input: &DescribeListenersInput)
-> Result<DescribeListenersOutput, DescribeListenersError>

Describes the specified listeners or the listeners for the specified Application Load Balancer. You must specify either a load balancer or one or more listeners.

fn describe_load_balancer_attributes(&self,
input: &DescribeLoadBalancerAttributesInput)
-> Result<DescribeLoadBalancerAttributesOutput, DescribeLoadBalancerAttributesError>

Describes the attributes for the specified Application Load Balancer.

fn describe_load_balancers(&self,
input: &DescribeLoadBalancersInput)
-> Result<DescribeLoadBalancersOutput, DescribeLoadBalancersError>

Describes the specified Application Load Balancers or all of your Application Load Balancers.

To describe the listeners for a load balancer, use DescribeListeners. To describe the attributes for a load balancer, use DescribeLoadBalancerAttributes.

fn describe_rules(&self,
input: &DescribeRulesInput)
-> Result<DescribeRulesOutput, DescribeRulesError>

Describes the specified rules or the rules for the specified listener. You must specify either a listener or one or more rules.

fn describe_ssl_policies(&self,
input: &DescribeSSLPoliciesInput)
-> Result<DescribeSSLPoliciesOutput, DescribeSSLPoliciesError>

Describes the specified policies or all policies used for SSL negotiation.

Note that the only supported policy at this time is ELBSecurityPolicy-2015-05.

fn describe_tags(&self,
input: &DescribeTagsInput)
-> Result<DescribeTagsOutput, DescribeTagsError>

Describes the tags for the specified resources.

fn describe_target_group_attributes(&self,
input: &DescribeTargetGroupAttributesInput)
-> Result<DescribeTargetGroupAttributesOutput, DescribeTargetGroupAttributesError>

Describes the attributes for the specified target group.

fn describe_target_groups(&self,
input: &DescribeTargetGroupsInput)
-> Result<DescribeTargetGroupsOutput, DescribeTargetGroupsError>

Describes the specified target groups or all of your target groups. By default, all target groups are described. Alternatively, you can specify one of the following to filter the results: the ARN of the load balancer, the names of one or more target groups, or the ARNs of one or more target groups.

To describe the targets for a target group, use DescribeTargetHealth. To describe the attributes of a target group, use DescribeTargetGroupAttributes.

fn describe_target_health(&self,
input: &DescribeTargetHealthInput)
-> Result<DescribeTargetHealthOutput, DescribeTargetHealthError>

Describes the health of the specified targets or all of your targets.

fn modify_listener(&self,
input: &ModifyListenerInput)
-> Result<ModifyListenerOutput, ModifyListenerError>

Modifies the specified properties of the specified listener.

Any properties that you do not specify retain their current values. However, changing the protocol from HTTPS to HTTP removes the security policy and SSL certificate properties. If you change the protocol from HTTP to HTTPS, you must add the security policy.

fn modify_load_balancer_attributes(&self,
input: &ModifyLoadBalancerAttributesInput)
-> Result<ModifyLoadBalancerAttributesOutput, ModifyLoadBalancerAttributesError>

Modifies the specified attributes of the specified Application Load Balancer.

If any of the specified attributes can't be modified as requested, the call fails. Any existing attributes that you do not modify retain their current values.

fn modify_rule(&self,
input: &ModifyRuleInput)
-> Result<ModifyRuleOutput, ModifyRuleError>

Modifies the specified rule.

Any existing properties that you do not modify retain their current values.

To modify the default action, use ModifyListener.

fn modify_target_group(&self,
input: &ModifyTargetGroupInput)
-> Result<ModifyTargetGroupOutput, ModifyTargetGroupError>

Modifies the health checks used when evaluating the health state of the targets in the specified target group.

To monitor the health of the targets, use DescribeTargetHealth.

fn modify_target_group_attributes(&self,
input: &ModifyTargetGroupAttributesInput)
-> Result<ModifyTargetGroupAttributesOutput, ModifyTargetGroupAttributesError>

Modifies the specified attributes of the specified target group.

fn register_targets(&self,
input: &RegisterTargetsInput)
-> Result<RegisterTargetsOutput, RegisterTargetsError>

Registers the specified targets with the specified target group.

By default, the load balancer routes requests to registered targets using the protocol and port number for the target group. Alternatively, you can override the port for a target when you register it.

The target must be in the virtual private cloud (VPC) that you specified for the target group. If the target is an EC2 instance, it can't be in the stopped or running state when you register it.

To remove a target from a target group, use DeregisterTargets.

fn remove_tags(&self,
input: &RemoveTagsInput)
-> Result<RemoveTagsOutput, RemoveTagsError>

Removes the specified tags from the specified resource.

To list the current tags for your resources, use DescribeTags.

fn set_rule_priorities(&self,
input: &SetRulePrioritiesInput)
-> Result<SetRulePrioritiesOutput, SetRulePrioritiesError>

Sets the priorities of the specified rules.

You can reorder the rules as long as there are no priority conflicts in the new order. Any existing rules that you do not specify retain their current priority.

fn set_security_groups(&self,
input: &SetSecurityGroupsInput)
-> Result<SetSecurityGroupsOutput, SetSecurityGroupsError>

Associates the specified security groups with the specified load balancer. The specified security groups override the previously associated security groups.

fn set_subnets(&self,
input: &SetSubnetsInput)
-> Result<SetSubnetsOutput, SetSubnetsError>

Enables the Availability Zone for the specified subnets for the specified load balancer. The specified subnets replace the previously enabled subnets.