Module rusoto::emr

Amazon Elastic MapReduce

Structs

AddInstanceGroupsInput
Input to an AddInstanceGroups call.

AddInstanceGroupsOutput
Output from an AddInstanceGroups call.

AddJobFlowStepsInput
The input argument to the AddJobFlowSteps operation.

AddJobFlowStepsOutput
The output for the AddJobFlowSteps operation.

AddTagsInput
This input identifies a cluster and a list of tags to attach.

AddTagsOutput
This output indicates the result of adding tags to a resource.

Application

An application is any Amazon or third-party software that you can add to the cluster. This structure contains a list of strings that indicates the software to use with the cluster and accepts a user argument list. Amazon EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action argument. For more information, see Launch a Job Flow on the MapR Distribution for Hadoop. Currently supported values are:

"mapr-m3" - launch the job flow using MapR M3 Edition.

"mapr-m5" - launch the job flow using MapR M5 Edition.

"mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition, respectively.

In Amazon EMR releases 4.0 and greater, the only accepted parameter is the application name. To pass arguments to applications, you supply a configuration for each application.

BootstrapActionConfig
Configuration of a bootstrap action.

BootstrapActionDetail
Reports the configuration of a bootstrap action in a job flow.

Cluster
The detailed description of the cluster.

ClusterStateChangeReason
The reason that the cluster changed to its current state.

ClusterStatus
The detailed status of the cluster.

ClusterSummary
The summary description of the cluster.

ClusterTimeline
Represents the timeline of the cluster's lifecycle.

Command
An entity describing an executable that runs on a cluster.

Configuration

Amazon EMR releases 4.x or later.

Specifies a hardware and software configuration of the EMR cluster. This includes configurations for applications and software bundled with Amazon EMR. The Configuration object is a JSON object which is defined by a classification and a set of properties. Configurations can be nested, so a configuration may have its own Configuration objects listed.

DescribeClusterInput
This input determines which cluster to describe.

DescribeClusterOutput
This output contains the description of the cluster.

DescribeJobFlowsInput
The input for the DescribeJobFlows operation.

DescribeJobFlowsOutput
The output for the DescribeJobFlows operation.

DescribeStepInput
This input determines which step to describe.

DescribeStepOutput
This output contains the description of the cluster step.

EbsBlockDevice
Configuration of requested EBS block device associated with the instance group.

EbsBlockDeviceConfig
Configuration of requested EBS block device associated with the instance group with count of volumes that will be associated to every instance.

EbsConfiguration

EbsVolume
EBS block device that's attached to an EC2 instance.

Ec2InstanceAttributes
Provides information about the EC2 instances in a cluster grouped by category. For example, key name, subnet ID, IAM instance profile, and so on.

EmrClient
A client for the Amazon EMR API.

HadoopJarStepConfig
A job flow step consisting of a JAR file whose main function will be executed. The main function submits a job for Hadoop to execute and waits for the job to finish or fail.

HadoopStepConfig
A cluster step consisting of a JAR file whose main function will be executed. The main function submits a job for Hadoop to execute and waits for the job to finish or fail.

Instance
Represents an EC2 instance provisioned as part of cluster.

InstanceGroup
This entity represents an instance group, which is a group of instances that have common purpose. For example, CORE instance group is used for HDFS.

InstanceGroupConfig
Configuration defining a new instance group.

InstanceGroupDetail
Detailed information about an instance group.

InstanceGroupModifyConfig
Modify an instance group size.

InstanceGroupStateChangeReason
The status change reason details for the instance group.

InstanceGroupStatus
The details of the instance group status.

InstanceGroupTimeline
The timeline of the instance group lifecycle.

InstanceResizePolicy
Custom policy for requesting termination protection or termination of specific instances when shrinking an instance group.

InstanceStateChangeReason
The details of the status change reason for the instance.

InstanceStatus
The instance status details.

InstanceTimeline
The timeline of the instance lifecycle.

JobFlowDetail
A description of a job flow.

JobFlowExecutionStatusDetail
Describes the status of the job flow.

JobFlowInstancesConfig
A description of the Amazon EC2 instance running the job flow. A valid JobFlowInstancesConfig must contain at least InstanceGroups, which is the recommended configuration. However, a valid alternative is to have MasterInstanceType, SlaveInstanceType, and InstanceCount (all three must be present).

JobFlowInstancesDetail
Specify the type of Amazon EC2 instances to run the job flow on.

KeyValue
A key value pair.

ListBootstrapActionsInput
This input determines which bootstrap actions to retrieve.

ListBootstrapActionsOutput
This output contains the boostrap actions detail .

ListClustersInput
This input determines how the ListClusters action filters the list of clusters that it returns.

ListClustersOutput
This contains a ClusterSummaryList with the cluster details; for example, the cluster IDs, names, and status.

ListInstanceGroupsInput
This input determines which instance groups to retrieve.

ListInstanceGroupsOutput
This input determines which instance groups to retrieve.

ListInstancesInput
This input determines which instances to list.

ListInstancesOutput
This output contains the list of instances.

ListStepsInput
This input determines which steps to list.

ListStepsOutput
This output contains the list of steps returned in reverse order. This means that the last step is the first element in the list.

ModifyInstanceGroupsInput
Change the size of some instance groups.

PlacementType
The Amazon EC2 location for the job flow.

RemoveTagsInput
This input identifies a cluster and a list of tags to remove.

RemoveTagsOutput
This output indicates the result of removing tags from a resource.

RunJobFlowInput
Input to the RunJobFlow operation.

RunJobFlowOutput
The result of the RunJobFlow operation.

ScriptBootstrapActionConfig
Configuration of the script to run during a bootstrap action.

SetTerminationProtectionInput
The input argument to the TerminationProtection operation.

SetVisibleToAllUsersInput
The input to the SetVisibleToAllUsers action.

ShrinkPolicy
Policy for customizing shrink operations. Allows configuration of decommissioning timeout and targeted instance shrinking.

Step
This represents a step in a cluster.

StepConfig
Specification of a job flow step.

StepDetail
Combines the execution state and configuration of a step.

StepExecutionStatusDetail
The execution state of a step.

StepStateChangeReason
The details of the step state change reason.

StepStatus
The execution status details of the cluster step.

StepSummary
The summary of the cluster step.

StepTimeline
The timeline of the cluster step lifecycle.

SupportedProductConfig
The list of supported product configurations which allow user-supplied arguments. EMR accepts these arguments and forwards them to the corresponding installation script as bootstrap action arguments.

Tag
A key/value pair containing user-defined metadata that you can associate with an Amazon EMR resource. Tags make it easier to associate clusters in various ways, such as grouping clu\ sters to track your Amazon EMR resource allocation costs. For more information, see Tagging Amazon EMR Resources.

TerminateJobFlowsInput
Input to the TerminateJobFlows operation.

VolumeSpecification
EBS volume specifications such as volume type, IOPS, and size(GiB) that will be requested for the EBS volume attached to an EC2 instance in the cluster.

Enums

AddInstanceGroupsError

AddJobFlowStepsError

AddTagsError

DescribeClusterError

DescribeJobFlowsError

DescribeStepError

ListBootstrapActionsError

ListClustersError

ListInstanceGroupsError

ListInstancesError

ListStepsError

ModifyInstanceGroupsError

RemoveTagsError

RunJobFlowError

SetTerminationProtectionError

SetVisibleToAllUsersError

TerminateJobFlowsError

Type Definitions

ActionOnFailure

ApplicationList

Boolean

BooleanObject

BootstrapActionConfigList

BootstrapActionDetailList

ClusterId

ClusterState

ClusterStateChangeReasonCode

ClusterStateList

ClusterSummaryList

CommandList

ConfigurationList

Date

EC2InstanceIdsList

EC2InstanceIdsToTerminateList

EbsBlockDeviceConfigList

EbsBlockDeviceList

EbsVolumeList

ErrorCode

ErrorMessage

InstanceGroupConfigList

InstanceGroupDetailList

InstanceGroupId

InstanceGroupIdsList

InstanceGroupList

InstanceGroupModifyConfigList

InstanceGroupState

InstanceGroupStateChangeReasonCode

InstanceGroupType

InstanceGroupTypeList

InstanceId

InstanceList

InstanceRoleType

InstanceState

InstanceStateChangeReasonCode

InstanceStateList

InstanceType

Integer

JobFlowDetailList

JobFlowExecutionState

The type of instance.

A small instance

A large instance

JobFlowExecutionStateList

KeyValueList

Marker

MarketType

NewSupportedProductsList

ResourceId

SecurityGroupsList

StepConfigList

StepDetailList

StepExecutionState

StepId

StepIdsList

StepState

StepStateChangeReasonCode

StepStateList

StepSummaryList

StringList

StringMap

SupportedProductsList

TagList

XmlString

XmlStringList

XmlStringMaxLen256