Struct rusoto::emr::Application

pub struct Application {
    pub additional_info: Option<StringMap>,
    pub args: Option<StringList>,
    pub name: Option<String>,
    pub version: Option<String>,
}

An application is any Amazon or third-party software that you can add to the cluster. This structure contains a list of strings that indicates the software to use with the cluster and accepts a user argument list. Amazon EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action argument. For more information, see Launch a Job Flow on the MapR Distribution for Hadoop. Currently supported values are:

"mapr-m3" - launch the job flow using MapR M3 Edition.

"mapr-m5" - launch the job flow using MapR M5 Edition.

"mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition, respectively.

In Amazon EMR releases 4.0 and greater, the only accepted parameter is the application name. To pass arguments to applications, you supply a configuration for each application.

Fields
additional_info: Option<StringMap>

This option is for advanced users only. This is meta information about third-party applications that third-party vendors use for testing purposes.

args: Option<StringList>

Arguments for Amazon EMR to pass to the application.

name: Option<String>

The name of the application.

version: Option<String>

The version of the application.