Struct rusoto::emr::Cluster

pub struct Cluster {
    pub applications: Option<ApplicationList>,
    pub auto_terminate: Option<Boolean>,
    pub configurations: Option<ConfigurationList>,
    pub ec_2_instance_attributes: Option<Ec2InstanceAttributes>,
    pub id: Option<ClusterId>,
    pub log_uri: Option<String>,
    pub master_public_dns_name: Option<String>,
    pub name: Option<String>,
    pub normalized_instance_hours: Option<Integer>,
    pub release_label: Option<String>,
    pub requested_ami_version: Option<String>,
    pub running_ami_version: Option<String>,
    pub service_role: Option<String>,
    pub status: Option<ClusterStatus>,
    pub tags: Option<TagList>,
    pub termination_protected: Option<Boolean>,
    pub visible_to_all_users: Option<Boolean>,
}

The detailed description of the cluster.

Fields
applications: Option<ApplicationList>

The applications installed on this cluster.

auto_terminate: Option<Boolean>

Specifies whether the cluster should terminate after completing all steps.

configurations: Option<ConfigurationList>

Amazon EMR releases 4.x or later.

The list of Configurations supplied to the EMR cluster.

ec_2_instance_attributes: Option<Ec2InstanceAttributes> id: Option<ClusterId>

The unique identifier for the cluster.

log_uri: Option<String>

The path to the Amazon S3 location where logs for this cluster are stored.

master_public_dns_name: Option<String>

The public DNS name of the master EC2 instance.

name: Option<String>

The name of the cluster.

normalized_instance_hours: Option<Integer>

An approximation of the cost of the job flow, represented in m1.small/hours. This value is incremented one time for every hour an m1.small instance runs. Larger instances are weighted more, so an EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.

release_label: Option<String>

The release label for the Amazon EMR release. For Amazon EMR 3.x and 2.x AMIs, use amiVersion instead instead of ReleaseLabel.

requested_ami_version: Option<String>

The AMI version requested for this cluster.

running_ami_version: Option<String>

The AMI version running on this cluster.

service_role: Option<String>

The IAM role that will be assumed by the Amazon EMR service to access AWS resources on your behalf.

status: Option<ClusterStatus>

The current status details about the cluster.

tags: Option<TagList>

A list of tags associated with a cluster.

termination_protected: Option<Boolean>

Indicates whether Amazon EMR will lock the cluster to prevent the EC2 instances from being terminated by an API call or user intervention, or in the event of a cluster error.

visible_to_all_users: Option<Boolean>

Indicates whether the job flow is visible to all IAM users of the AWS account associated with the job flow. If this value is set to true, all IAM users of that AWS account can view and manage the job flow if they have the proper policy permissions set. If this value is false, only the IAM user that created the cluster can view and manage it. This value can be changed using the SetVisibleToAllUsers action.