Struct rusoto::emr::Ec2InstanceAttributes

pub struct Ec2InstanceAttributes {
    pub additional_master_security_groups: Option<StringList>,
    pub additional_slave_security_groups: Option<StringList>,
    pub ec_2_availability_zone: Option<String>,
    pub ec_2_key_name: Option<String>,
    pub ec_2_subnet_id: Option<String>,
    pub emr_managed_master_security_group: Option<String>,
    pub emr_managed_slave_security_group: Option<String>,
    pub iam_instance_profile: Option<String>,
    pub service_access_security_group: Option<String>,
}

Provides information about the EC2 instances in a cluster grouped by category. For example, key name, subnet ID, IAM instance profile, and so on.

Fields
additional_master_security_groups: Option<StringList>

A list of additional Amazon EC2 security group IDs for the master node.

additional_slave_security_groups: Option<StringList>

A list of additional Amazon EC2 security group IDs for the slave nodes.

ec_2_availability_zone: Option<String>

The Availability Zone in which the cluster will run.

ec_2_key_name: Option<String>

The name of the Amazon EC2 key pair to use when connecting with SSH into the master node as a user named "hadoop".

ec_2_subnet_id: Option<String>

To launch the job flow in Amazon VPC, set this parameter to the identifier of the Amazon VPC subnet where you want the job flow to launch. If you do not specify this value, the job flow is launched in the normal AWS cloud, outside of a VPC.

Amazon VPC currently does not support cluster compute quadruple extra large (cc1.4xlarge) instances. Thus, you cannot specify the cc1.4xlarge instance type for nodes of a job flow launched in a VPC.

emr_managed_master_security_group: Option<String>

The identifier of the Amazon EC2 security group for the master node.

emr_managed_slave_security_group: Option<String>

The identifier of the Amazon EC2 security group for the slave nodes.

iam_instance_profile: Option<String>

The IAM role that was specified when the job flow was launched. The EC2 instances of the job flow assume this role.

service_access_security_group: Option<String>

The identifier of the Amazon EC2 security group for the Amazon EMR service to access clusters in VPC private subnets.