Struct rusoto::emr::InstanceGroup

pub struct InstanceGroup {
    pub bid_price: Option<String>,
    pub configurations: Option<ConfigurationList>,
    pub ebs_block_devices: Option<EbsBlockDeviceList>,
    pub ebs_optimized: Option<BooleanObject>,
    pub id: Option<InstanceGroupId>,
    pub instance_group_type: Option<InstanceGroupType>,
    pub instance_type: Option<InstanceType>,
    pub market: Option<MarketType>,
    pub name: Option<String>,
    pub requested_instance_count: Option<Integer>,
    pub running_instance_count: Option<Integer>,
    pub shrink_policy: Option<ShrinkPolicy>,
    pub status: Option<InstanceGroupStatus>,
}

This entity represents an instance group, which is a group of instances that have common purpose. For example, CORE instance group is used for HDFS.

Fields
bid_price: Option<String>

The bid price for each EC2 instance in the instance group when launching nodes as Spot Instances, expressed in USD.

configurations: Option<ConfigurationList>

Amazon EMR releases 4.x or later.

The list of configurations supplied for an EMR cluster instance group. You can specify a separate configuration for each instance group (master, core, and task).

ebs_block_devices: Option<EbsBlockDeviceList>

The EBS block devices that are mapped to this instance group.

ebs_optimized: Option<BooleanObject>

If the instance group is EBS-optimized. An Amazon EBS–optimized instance uses an optimized configuration stack and provides additional, dedicated capacity for Amazon EBS I/O.

id: Option<InstanceGroupId>

The identifier of the instance group.

instance_group_type: Option<InstanceGroupType>

The type of the instance group. Valid values are MASTER, CORE or TASK.

instance_type: Option<InstanceType>

The EC2 instance type for all instances in the instance group.

market: Option<MarketType>

The marketplace to provision instances for this group. Valid values are ON_DEMAND or SPOT.

name: Option<String>

The name of the instance group.

requested_instance_count: Option<Integer>

The target number of instances for the instance group.

running_instance_count: Option<Integer>

The number of instances currently running in this instance group.

shrink_policy: Option<ShrinkPolicy>

Policy for customizing shrink operations.

status: Option<InstanceGroupStatus>

The current status of the instance group.