Struct rusoto::emr::InstanceGroupDetail

pub struct InstanceGroupDetail {
    pub bid_price: Option<XmlStringMaxLen256>,
    pub creation_date_time: Date,
    pub end_date_time: Option<Date>,
    pub instance_group_id: Option<XmlStringMaxLen256>,
    pub instance_request_count: Integer,
    pub instance_role: InstanceRoleType,
    pub instance_running_count: Integer,
    pub instance_type: InstanceType,
    pub last_state_change_reason: Option<XmlString>,
    pub market: MarketType,
    pub name: Option<XmlStringMaxLen256>,
    pub ready_date_time: Option<Date>,
    pub start_date_time: Option<Date>,
    pub state: InstanceGroupState,
}

Detailed information about an instance group.

Fields
bid_price: Option<XmlStringMaxLen256>

Bid price for EC2 Instances when launching nodes as Spot Instances, expressed in USD.

creation_date_time: Date

The date/time the instance group was created.

end_date_time: Option<Date>

The date/time the instance group was terminated.

instance_group_id: Option<XmlStringMaxLen256>

Unique identifier for the instance group.

instance_request_count: Integer

Target number of instances to run in the instance group.

instance_role: InstanceRoleType

Instance group role in the cluster

instance_running_count: Integer

Actual count of running instances.

instance_type: InstanceType

Amazon EC2 Instance type.

last_state_change_reason: Option<XmlString>

Details regarding the state of the instance group.

market: MarketType

Market type of the Amazon EC2 instances used to create a cluster node.

name: Option<XmlStringMaxLen256>

Friendly name for the instance group.

ready_date_time: Option<Date>

The date/time the instance group was available to the cluster.

start_date_time: Option<Date>

The date/time the instance group was started.

state: InstanceGroupState

State of instance group. The following values are deprecated: STARTING, TERMINATED, and FAILED.