Struct rusoto::emr::JobFlowInstancesConfig

pub struct JobFlowInstancesConfig {
    pub additional_master_security_groups: Option<SecurityGroupsList>,
    pub additional_slave_security_groups: Option<SecurityGroupsList>,
    pub ec_2_key_name: Option<XmlStringMaxLen256>,
    pub ec_2_subnet_id: Option<XmlStringMaxLen256>,
    pub emr_managed_master_security_group: Option<XmlStringMaxLen256>,
    pub emr_managed_slave_security_group: Option<XmlStringMaxLen256>,
    pub hadoop_version: Option<XmlStringMaxLen256>,
    pub instance_count: Option<Integer>,
    pub instance_groups: Option<InstanceGroupConfigList>,
    pub keep_job_flow_alive_when_no_steps: Option<Boolean>,
    pub master_instance_type: Option<InstanceType>,
    pub placement: Option<PlacementType>,
    pub service_access_security_group: Option<XmlStringMaxLen256>,
    pub slave_instance_type: Option<InstanceType>,
    pub termination_protected: Option<Boolean>,
}

A description of the Amazon EC2 instance running the job flow. A valid JobFlowInstancesConfig must contain at least InstanceGroups, which is the recommended configuration. However, a valid alternative is to have MasterInstanceType, SlaveInstanceType, and InstanceCount (all three must be present).

Fields
additional_master_security_groups: Option<SecurityGroupsList>

A list of additional Amazon EC2 security group IDs for the master node.

additional_slave_security_groups: Option<SecurityGroupsList>

A list of additional Amazon EC2 security group IDs for the slave nodes.

ec_2_key_name: Option<XmlStringMaxLen256>

The name of the Amazon EC2 key pair that can be used to ssh to the master node as the user called "hadoop."

ec_2_subnet_id: Option<XmlStringMaxLen256>

To launch the job flow in Amazon Virtual Private Cloud (Amazon VPC), set this parameter to the identifier of the Amazon VPC subnet where you want the job flow to launch. If you do not specify this value, the job flow is launched in the normal Amazon Web Services cloud, outside of an Amazon VPC.

Amazon VPC currently does not support cluster compute quadruple extra large (cc1.4xlarge) instances. Thus you cannot specify the cc1.4xlarge instance type for nodes of a job flow launched in a Amazon VPC.

emr_managed_master_security_group: Option<XmlStringMaxLen256>

The identifier of the Amazon EC2 security group for the master node.

emr_managed_slave_security_group: Option<XmlStringMaxLen256>

The identifier of the Amazon EC2 security group for the slave nodes.

hadoop_version: Option<XmlStringMaxLen256>

The Hadoop version for the job flow. Valid inputs are "0.18" (deprecated), "0.20" (deprecated), "0.20.205" (deprecated), "1.0.3", "2.2.0", or "2.4.0". If you do not set this value, the default of 0.18 is used, unless the AmiVersion parameter is set in the RunJobFlow call, in which case the default version of Hadoop for that AMI version is used.

instance_count: Option<Integer>

The number of Amazon EC2 instances used to execute the job flow.

instance_groups: Option<InstanceGroupConfigList>

Configuration for the job flow's instance groups.

keep_job_flow_alive_when_no_steps: Option<Boolean>

Specifies whether the job flow should be kept alive after completing all steps.

master_instance_type: Option<InstanceType>

The EC2 instance type of the master node.

placement: Option<PlacementType>

The Availability Zone the job flow will run in.

service_access_security_group: Option<XmlStringMaxLen256>

The identifier of the Amazon EC2 security group for the Amazon EMR service to access clusters in VPC private subnets.

slave_instance_type: Option<InstanceType>

The EC2 instance type of the slave nodes.

termination_protected: Option<Boolean>

Specifies whether to lock the job flow to prevent the Amazon EC2 instances from being terminated by API call, user intervention, or in the event of a job flow error.