Struct rusoto::emr::JobFlowInstancesDetail

pub struct JobFlowInstancesDetail {
    pub ec_2_key_name: Option<XmlStringMaxLen256>,
    pub ec_2_subnet_id: Option<XmlStringMaxLen256>,
    pub hadoop_version: Option<XmlStringMaxLen256>,
    pub instance_count: Integer,
    pub instance_groups: Option<InstanceGroupDetailList>,
    pub keep_job_flow_alive_when_no_steps: Option<Boolean>,
    pub master_instance_id: Option<XmlString>,
    pub master_instance_type: InstanceType,
    pub master_public_dns_name: Option<XmlString>,
    pub normalized_instance_hours: Option<Integer>,
    pub placement: Option<PlacementType>,
    pub slave_instance_type: InstanceType,
    pub termination_protected: Option<Boolean>,
}

Specify the type of Amazon EC2 instances to run the job flow on.

Fields
ec_2_key_name: Option<XmlStringMaxLen256>

The name of an Amazon EC2 key pair that can be used to ssh to the master node of job flow.

ec_2_subnet_id: Option<XmlStringMaxLen256>

For job flows launched within Amazon Virtual Private Cloud, this value specifies the identifier of the subnet where the job flow was launched.

hadoop_version: Option<XmlStringMaxLen256>

The Hadoop version for the job flow.

instance_count: Integer

The number of Amazon EC2 instances in the cluster. If the value is 1, the same instance serves as both the master and slave node. If the value is greater than 1, one instance is the master node and all others are slave nodes.

instance_groups: Option<InstanceGroupDetailList>

Details about the job flow's instance groups.

keep_job_flow_alive_when_no_steps: Option<Boolean>

Specifies whether the job flow should terminate after completing all steps.

master_instance_id: Option<XmlString>

The Amazon EC2 instance identifier of the master node.

master_instance_type: InstanceType

The Amazon EC2 master node instance type.

master_public_dns_name: Option<XmlString>

The DNS name of the master node.

normalized_instance_hours: Option<Integer>

An approximation of the cost of the job flow, represented in m1.small/hours. This value is incremented once for every hour an m1.small runs. Larger instances are weighted more, so an Amazon EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.

placement: Option<PlacementType>

The Amazon EC2 Availability Zone for the job flow.

slave_instance_type: InstanceType

The Amazon EC2 slave node instance type.

termination_protected: Option<Boolean>

Specifies whether the Amazon EC2 instances in the cluster are protected from termination by API calls, user intervention, or in the event of a job flow error.