Struct rusoto::emr::ListClustersInput

pub struct ListClustersInput {
    pub cluster_states: Option<ClusterStateList>,
    pub created_after: Option<Date>,
    pub created_before: Option<Date>,
    pub marker: Option<Marker>,
}

This input determines how the ListClusters action filters the list of clusters that it returns.

Fields
cluster_states: Option<ClusterStateList>

The cluster state filters to apply when listing clusters.

created_after: Option<Date>

The creation date and time beginning value filter for listing clusters .

created_before: Option<Date>

The creation date and time end value filter for listing clusters .

marker: Option<Marker>

The pagination token that indicates the next set of results to retrieve.