Struct rusoto::emr::RunJobFlowInput

pub struct RunJobFlowInput {
    pub additional_info: Option<XmlString>,
    pub ami_version: Option<XmlStringMaxLen256>,
    pub applications: Option<ApplicationList>,
    pub bootstrap_actions: Option<BootstrapActionConfigList>,
    pub configurations: Option<ConfigurationList>,
    pub instances: JobFlowInstancesConfig,
    pub job_flow_role: Option<XmlString>,
    pub log_uri: Option<XmlString>,
    pub name: XmlStringMaxLen256,
    pub new_supported_products: Option<NewSupportedProductsList>,
    pub release_label: Option<XmlStringMaxLen256>,
    pub service_role: Option<XmlString>,
    pub steps: Option<StepConfigList>,
    pub supported_products: Option<SupportedProductsList>,
    pub tags: Option<TagList>,
    pub visible_to_all_users: Option<Boolean>,
}

Input to the RunJobFlow operation.

Fields
additional_info: Option<XmlString>

A JSON string for selecting additional features.

ami_version: Option<XmlStringMaxLen256>

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and greater, use ReleaseLabel.

The version of the Amazon Machine Image (AMI) to use when launching Amazon EC2 instances in the job flow. The following values are valid:

The version number of the AMI to use, for example, "2.0."

If the AMI supports multiple versions of Hadoop (for example, AMI 1.0 supports both Hadoop 0.18 and 0.20) you can use the JobFlowInstancesConfig HadoopVersion parameter to modify the version of Hadoop from the defaults shown above.

For details about the AMI versions currently supported by Amazon Elastic MapReduce, go to AMI Versions Supported in Elastic MapReduce in the Amazon Elastic MapReduce Developer's Guide.

applications: Option<ApplicationList>

Amazon EMR releases 4.x or later.

A list of applications for the cluster. Valid values are: "Hadoop", "Hive", "Mahout", "Pig", and "Spark." They are case insensitive.

bootstrap_actions: Option<BootstrapActionConfigList>

A list of bootstrap actions that will be run before Hadoop is started on the cluster nodes.

configurations: Option<ConfigurationList>

Amazon EMR releases 4.x or later.

The list of configurations supplied for the EMR cluster you are creating.

instances: JobFlowInstancesConfig

A specification of the number and type of Amazon EC2 instances on which to run the job flow.

job_flow_role: Option<XmlString>

Also called instance profile and EC2 role. An IAM role for an EMR cluster. The EC2 instances of the cluster assume this role. The default role is EMR_EC2_DefaultRole. In order to use the default role, you must have already created it using the CLI or console.

log_uri: Option<XmlString>

The location in Amazon S3 to write the log files of the job flow. If a value is not provided, logs are not created.

name: XmlStringMaxLen256

The name of the job flow.

new_supported_products: Option<NewSupportedProductsList>

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and greater, use Applications.

A list of strings that indicates third-party software to use with the job flow that accepts a user argument list. EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action arguments. For more information, see Launch a Job Flow on the MapR Distribution for Hadoop. Currently supported values are:

"mapr-m3" - launch the cluster using MapR M3 Edition.

"mapr-m5" - launch the cluster using MapR M5 Edition.

"mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition respectively.

"mapr-m7" - launch the cluster using MapR M7 Edition.

"hunk" - launch the cluster with the Hunk Big Data Analtics Platform.

"hue"- launch the cluster with Hue installed.

"spark" - launch the cluster with Apache Spark installed.

"ganglia" - launch the cluster with the Ganglia Monitoring System installed.

release_label: Option<XmlStringMaxLen256>

Amazon EMR releases 4.x or later.

The release label for the Amazon EMR release. For Amazon EMR 3.x and 2.x AMIs, use amiVersion instead instead of ReleaseLabel.

service_role: Option<XmlString>

The IAM role that will be assumed by the Amazon EMR service to access AWS resources on your behalf.

steps: Option<StepConfigList>

A list of steps to be executed by the job flow.

supported_products: Option<SupportedProductsList>

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and greater, use Applications.

A list of strings that indicates third-party software to use with the job flow. For more information, go to Use Third Party Applications with Amazon EMR. Currently supported values are:

"mapr-m3" - launch the job flow using MapR M3 Edition.

"mapr-m5" - launch the job flow using MapR M5 Edition.

tags: Option<TagList>

A list of tags to associate with a cluster and propagate to Amazon EC2 instances.

visible_to_all_users: Option<Boolean>

Whether the job flow is visible to all IAM users of the AWS account associated with the job flow. If this value is set to true, all IAM users of that AWS account can view and (if they have the proper policy permissions set) manage the job flow. If it is set to false, only the IAM user that created the job flow can view and manage it.