Module rusoto::ets

Amazon Elastic Transcoder

Structs

AccessDeniedException
General authentication failure. The request was not signed correctly.

Artwork
The file to be used as album art. There can be multiple artworks associated with an audio file, to a maximum of 20.

To remove artwork or leave the artwork empty, you can either set Artwork to null, or set the Merge Policy to "Replace" and use an empty Artwork array.

To pass through existing artwork unchanged, set the Merge Policy to "Prepend", "Append", or "Fallback", and use an empty Artwork array.

AudioCodecOptions
Options associated with your audio codec.

AudioParameters
Parameters required for transcoding audio.

CancelJobRequest
The CancelJobRequest structure.

CancelJobResponse
The response body contains a JSON object. If the job is successfully canceled, the value of Success is true.

CaptionFormat
The file format of the output captions. If you leave this value blank, Elastic Transcoder returns an error.

CaptionSource
A source file for the input sidecar captions used during the transcoding process.

Captions
The captions to be created, if any.

Clip
Settings for one clip in a composition. All jobs in a playlist must have the same clip settings.

CreateJobOutput
The CreateJobOutput structure.

CreateJobPlaylist
Information about the master playlist.

CreateJobRequest
The CreateJobRequest structure.

CreateJobResponse
The CreateJobResponse structure.

CreatePipelineRequest
The CreatePipelineRequest structure.

CreatePipelineResponse
When you create a pipeline, Elastic Transcoder returns the values that you specified in the request.

CreatePresetRequest
The CreatePresetRequest structure.

CreatePresetResponse
The CreatePresetResponse structure.

DeletePipelineRequest
The DeletePipelineRequest structure.

DeletePipelineResponse
The DeletePipelineResponse structure.

DeletePresetRequest
The DeletePresetRequest structure.

DeletePresetResponse
The DeletePresetResponse structure.

DetectedProperties
The detected properties of the input file. Elastic Transcoder identifies these values from the input file.

Encryption
The encryption settings, if any, that are used for decrypting your input files or encrypting your output files. If your input file is encrypted, you must specify the mode that Elastic Transcoder will use to decrypt your file, otherwise you must specify the mode you want Elastic Transcoder to use to encrypt your output files.

EtsClient
A client for the Amazon Elastic Transcoder API.

HlsContentProtection
The HLS content protection settings, if any, that you want Elastic Transcoder to apply to your output files.

IncompatibleVersionException

InternalServiceException
Elastic Transcoder encountered an unexpected exception while trying to fulfill the request.

Job
A section of the response body that provides information about the job that is created.

JobAlbumArt
The .jpg or .png file associated with an audio file.

JobInput
Information about the file that you're transcoding.

JobOutput
Outputs recommended instead.If you specified one output for a job, information about that output. If you specified multiple outputs for a job, the Output object lists information about the first output. This duplicates the information that is listed for the first output in the Outputs object.

JobWatermark
Watermarks can be in .png or .jpg format. If you want to display a watermark that is not rectangular, use the .png format, which supports transparency.

LimitExceededException
Too many operations for a given AWS account. For example, the number of pipelines exceeds the maximum allowed.

ListJobsByPipelineRequest
The ListJobsByPipelineRequest structure.

ListJobsByPipelineResponse
The ListJobsByPipelineResponse structure.

ListJobsByStatusRequest
The ListJobsByStatusRequest structure.

ListJobsByStatusResponse
The ListJobsByStatusResponse structure.

ListPipelinesRequest
The ListPipelineRequest structure.

ListPipelinesResponse
A list of the pipelines associated with the current AWS account.

ListPresetsRequest
The ListPresetsRequest structure.

ListPresetsResponse
The ListPresetsResponse structure.

Notifications

The Amazon Simple Notification Service (Amazon SNS) topic or topics to notify in order to report job status.
To receive notifications, you must also subscribe to the new topic in the Amazon SNS console.

Permission
The Permission structure.

Pipeline
The pipeline (queue) that is used to manage jobs.

PipelineOutputConfig
The PipelineOutputConfig structure.

PlayReadyDrm
The PlayReady DRM settings, if any, that you want Elastic Transcoder to apply to the output files associated with this playlist.

PlayReady DRM encrypts your media files using AES-CTR encryption.

If you use DRM for an HLSv3 playlist, your outputs must have a master playlist.

Playlist
Use Only for Fragmented MP4 or MPEG-TS Outputs. If you specify a preset for which the value of Container is fmp4 (Fragmented MP4) or ts (MPEG-TS), Playlists contains information about the master playlists that you want Elastic Transcoder to create. We recommend that you create only one master playlist per output format. The maximum number of master playlists in a job is 30.

Preset
Presets are templates that contain most of the settings for transcoding media files from one format to another. Elastic Transcoder includes some default presets for common formats, for example, several iPod and iPhone versions. You can also create your own presets for formats that aren't included among the default presets. You specify which preset you want to use when you create a job.

PresetWatermark
Settings for the size, location, and opacity of graphics that you want Elastic Transcoder to overlay over videos that are transcoded using this preset. You can specify settings for up to four watermarks. Watermarks appear in the specified size and location, and with the specified opacity for the duration of the transcoded video.

Watermarks can be in .png or .jpg format. If you want to display a watermark that is not rectangular, use the .png format, which supports transparency.

When you create a job that uses this preset, you specify the .png or .jpg graphics that you want Elastic Transcoder to include in the transcoded videos. You can specify fewer graphics in the job than you specify watermark settings in the preset, which allows you to use the same preset for up to four watermarks that have different dimensions.

ReadJobRequest
The ReadJobRequest structure.

ReadJobResponse
The ReadJobResponse structure.

ReadPipelineRequest
The ReadPipelineRequest structure.

ReadPipelineResponse
The ReadPipelineResponse structure.

ReadPresetRequest
The ReadPresetRequest structure.

ReadPresetResponse
The ReadPresetResponse structure.

ResourceInUseException
The resource you are attempting to change is in use. For example, you are attempting to delete a pipeline that is currently in use.

ResourceNotFoundException
The requested resource does not exist or is not available. For example, the pipeline to which you're trying to add a job doesn't exist or is still being created.

TestRoleRequest
The TestRoleRequest structure.

TestRoleResponse
The TestRoleResponse structure.

Thumbnails
Thumbnails for videos.

TimeSpan
Settings that determine when a clip begins and how long it lasts.

Timing
Details about the timing of a job.

UpdatePipelineNotificationsRequest
The UpdatePipelineNotificationsRequest structure.

UpdatePipelineNotificationsResponse
The UpdatePipelineNotificationsResponse structure.

UpdatePipelineRequest
The UpdatePipelineRequest structure.

UpdatePipelineResponse
When you update a pipeline, Elastic Transcoder returns the values that you specified in the request.

UpdatePipelineStatusRequest
The UpdatePipelineStatusRequest structure.

UpdatePipelineStatusResponse
When you update status for a pipeline, Elastic Transcoder returns the values that you specified in the request.

ValidationException
One or more required parameter values were not provided in the request.

VideoParameters
The VideoParameters structure.

Warning
Elastic Transcoder returns a warning if the resources used by your pipeline are not in the same region as the pipeline.

Using resources in the same region, such as your Amazon S3 buckets, Amazon SNS notification topics, and AWS KMS key, reduces processing time and prevents cross-regional charges.

Type Definitions

AccessControl

AccessControls

Artworks

Ascending

AspectRatio

AudioBitDepth

AudioBitOrder

AudioBitRate

AudioChannels

AudioCodec

AudioCodecProfile

AudioPackingMode

AudioSampleRate

AudioSigned

Base64EncodedString

BucketName

CaptionFormatFormat

CaptionFormatPattern

CaptionFormats

CaptionMergePolicy

CaptionSources

CodecOption

CodecOptions

Composition

CreateJobOutputs

CreateJobPlaylists

Description

Digits

DigitsOrAuto

EncryptionMode

ExceptionMessages

Filename

FixedGOP

FloatString

FrameRate

Grantee

GranteeType

HlsContentProtectionMethod

HorizontalAlign

Id

Interlaced

JobContainer

JobOutputs

JobStatus

JobWatermarks

Jobs

JpgOrPng

Key

KeyArn

KeyIdGuid

KeyStoragePolicy

KeyframesMaxDist

MaxFrameRate

MergePolicy

Name

NonEmptyBase64EncodedString

NullableInteger

NullableLong

OneTo512String

Opacity

OutputKeys

PaddingPolicy

Permissions

PipelineStatus

Pipelines

PixelsOrPercent

PlayReadyDrmFormatString

PlaylistFormat

Playlists

PresetContainer

PresetType

PresetWatermarkId

PresetWatermarks

Presets

Resolution

Role

Rotate

SizingPolicy

SnsTopic

SnsTopics

StorageClass

Success

Target

ThumbnailPattern

ThumbnailResolution

Time

TimeOffset

UserMetadata

VerticalAlign

VideoBitRate

VideoCodec

Warnings

WatermarkKey

WatermarkSizingPolicy

ZeroTo255String

ZeroTo512String