Struct rusoto::events::PutTargetsRequest

pub struct PutTargetsRequest {
    pub rule: RuleName,
    pub targets: TargetList,
}

Container for the parameters to the PutTargets operation.

Fields
rule: RuleName

The name of the rule you want to add targets to.

targets: TargetList

List of targets you want to update or add to the rule.