Struct rusoto::events::Target

pub struct Target {
    pub arn: TargetArn,
    pub id: TargetId,
    pub input: Option<TargetInput>,
    pub input_path: Option<TargetInputPath>,
}

Targets are the resources that can be invoked when a rule is triggered. For example, AWS Lambda functions, Amazon Kinesis streams, and built-in targets.

Input and InputPath are mutually-exclusive and optional parameters of a target. When a rule is triggered due to a matched event, if for a target:

Neither Input nor InputPath is specified, then the entire event is passed to the target in JSON form.

InputPath is specified in the form of JSONPath (e.g. $.detail), then only the part of the event specified in the path is passed to the target (e.g. only the detail part of the event is passed).

Input is specified in the form of a valid JSON, then the matched event is overridden with this constant.

Fields
arn: TargetArn

The Amazon Resource Name (ARN) associated of the target.

id: TargetId

The unique target assignment ID.

input: Option<TargetInput>

Valid JSON text passed to the target. For more information about JSON text, see The JavaScript Object Notation (JSON) Data Interchange Format.

input_path: Option<TargetInputPath>

The value of the JSONPath that is used for extracting part of the matched event when passing it to the target. For more information about JSON paths, see JSONPath.