Struct rusoto::firehose::ElasticsearchDestinationConfiguration

pub struct ElasticsearchDestinationConfiguration {
    pub buffering_hints: Option<ElasticsearchBufferingHints>,
    pub cloud_watch_logging_options: Option<CloudWatchLoggingOptions>,
    pub domain_a_r_n: ElasticsearchDomainARN,
    pub index_name: ElasticsearchIndexName,
    pub index_rotation_period: Option<ElasticsearchIndexRotationPeriod>,
    pub retry_options: Option<ElasticsearchRetryOptions>,
    pub role_a_r_n: RoleARN,
    pub s_3_backup_mode: Option<ElasticsearchS3BackupMode>,
    pub s_3_configuration: S3DestinationConfiguration,
    pub type_name: ElasticsearchTypeName,
}

Describes the configuration of a destination in Amazon ES.

Fields
buffering_hints: Option<ElasticsearchBufferingHints>

Buffering options. If no value is specified, ElasticsearchBufferingHints object default values are used.

cloud_watch_logging_options: Option<CloudWatchLoggingOptions>

Describes CloudWatch logging options for your delivery stream.

domain_a_r_n: ElasticsearchDomainARN

The ARN of the Amazon ES domain. The IAM role must have permission for DescribeElasticsearchDomain, DescribeElasticsearchDomains , and DescribeElasticsearchDomainConfig after assuming RoleARN.

index_name: ElasticsearchIndexName

The Elasticsearch index name.

index_rotation_period: Option<ElasticsearchIndexRotationPeriod>

The Elasticsearch index rotation period. Index rotation appends a timestamp to the IndexName to facilitate expiration of old data. For more information, see Index Rotation for Amazon Elasticsearch Service Destination. Default value is OneDay.

retry_options: Option<ElasticsearchRetryOptions>

Configures retry behavior in the event that Firehose is unable to deliver documents to Amazon ES. Default value is 300 (5 minutes).

role_a_r_n: RoleARN

The ARN of the IAM role to be assumed by Firehose for calling the Amazon ES Configuration API and for indexing documents. For more information, see Amazon S3 Bucket Access.

s_3_backup_mode: Option<ElasticsearchS3BackupMode>

Defines how documents should be delivered to Amazon S3. When set to FailedDocumentsOnly, Firehose writes any documents that could not be indexed to the configured Amazon S3 destination, with elasticsearch-failed/ appended to the key prefix. When set to AllDocuments, Firehose delivers all incoming records to Amazon S3, and also writes failed documents with elasticsearch-failed/ appended to the prefix. For more information, see Amazon S3 Backup for Amazon Elasticsearch Service Destination. Default value is FailedDocumentsOnly.

s_3_configuration: S3DestinationConfiguration type_name: ElasticsearchTypeName

The Elasticsearch type name.