Struct rusoto::firehose::RedshiftDestinationConfiguration

pub struct RedshiftDestinationConfiguration {
    pub cluster_j_d_b_c_u_r_l: ClusterJDBCURL,
    pub copy_command: CopyCommand,
    pub password: Password,
    pub role_a_r_n: RoleARN,
    pub s_3_configuration: S3DestinationConfiguration,
    pub username: Username,
}

Describes the configuration of a destination in Amazon Redshift.

Fields
cluster_j_d_b_c_u_r_l: ClusterJDBCURL

The database connection string.

copy_command: CopyCommand

The COPY command.

password: Password

The user password.

role_a_r_n: RoleARN

The ARN of the AWS credentials.

s_3_configuration: S3DestinationConfiguration

The S3 configuration for the intermediate location from which Amazon Redshift obtains data. Restrictions are described in the topic for CreateDeliveryStream.

The compression formats SNAPPY or ZIP cannot be specified in RedshiftDestinationConfiguration.S3Configuration because the Amazon Redshift COPY operation that reads from the S3 bucket doesn't support these compression formats.

username: Username

The name of the user.