Module rusoto::iam

Amazon Identity and Access Management

If you're using the service, you're probably looking for IamClient.

Structs

AccessKey

Contains information about an AWS access key.

This data type is used as a response element in the CreateAccessKey and ListAccessKeys actions.

The SecretAccessKey value is returned only in response to CreateAccessKey. You can get a secret access key only when you first create an access key; you cannot recover the secret access key later. If you lose a secret access key, you must create a new access key.

AccessKeyLastUsed
Contains information about the last time an AWS access key was used.

This data type is used as a response element in the GetAccessKeyLastUsed action.

AccessKeyMetadata
Contains information about an AWS access key, without its secret key.

This data type is used as a response element in the ListAccessKeys action.

AddClientIDToOpenIDConnectProviderRequest

AddRoleToInstanceProfileRequest

AddUserToGroupRequest

AttachGroupPolicyRequest

AttachRolePolicyRequest

AttachUserPolicyRequest

AttachedPolicy
Contains information about an attached policy.

An attached policy is a managed policy that has been attached to a user, group, or role. This data type is used as a response element in the ListAttachedGroupPolicies, ListAttachedRolePolicies, ListAttachedUserPolicies, and GetAccountAuthorizationDetails actions.

For more information about managed policies, refer to Managed Policies and Inline Policies in the Using IAM guide.

ChangePasswordRequest

ContextEntry
Contains information about a condition context key. It includes the name of the key and specifies the value (or values, if the context key supports multiple values) to use in the simulation. This information is used when evaluating the Condition elements of the input policies.

This data type is used as an input parameter to SimulateCustomPolicy and SimulateCustomPolicy .

CreateAccessKeyRequest

CreateAccessKeyResponse
Contains the response to a successful CreateAccessKey request.

CreateAccountAliasRequest

CreateGroupRequest

CreateGroupResponse
Contains the response to a successful CreateGroup request.

CreateInstanceProfileRequest

CreateInstanceProfileResponse
Contains the response to a successful CreateInstanceProfile request.

CreateLoginProfileRequest

CreateLoginProfileResponse
Contains the response to a successful CreateLoginProfile request.

CreateOpenIDConnectProviderRequest

CreateOpenIDConnectProviderResponse
Contains the response to a successful CreateOpenIDConnectProvider request.

CreatePolicyRequest

CreatePolicyResponse
Contains the response to a successful CreatePolicy request.

CreatePolicyVersionRequest

CreatePolicyVersionResponse
Contains the response to a successful CreatePolicyVersion request.

CreateRoleRequest

CreateRoleResponse
Contains the response to a successful CreateRole request.

CreateSAMLProviderRequest

CreateSAMLProviderResponse
Contains the response to a successful CreateSAMLProvider request.

CreateServiceSpecificCredentialRequest

CreateServiceSpecificCredentialResponse

CreateUserRequest

CreateUserResponse
Contains the response to a successful CreateUser request.

CreateVirtualMFADeviceRequest

CreateVirtualMFADeviceResponse
Contains the response to a successful CreateVirtualMFADevice request.

DeactivateMFADeviceRequest

DeleteAccessKeyRequest

DeleteAccountAliasRequest

DeleteGroupPolicyRequest

DeleteGroupRequest

DeleteInstanceProfileRequest

DeleteLoginProfileRequest

DeleteOpenIDConnectProviderRequest

DeletePolicyRequest

DeletePolicyVersionRequest

DeleteRolePolicyRequest

DeleteRoleRequest

DeleteSAMLProviderRequest

DeleteSSHPublicKeyRequest

DeleteServerCertificateRequest

DeleteServiceSpecificCredentialRequest

DeleteSigningCertificateRequest

DeleteUserPolicyRequest

DeleteUserRequest

DeleteVirtualMFADeviceRequest

DetachGroupPolicyRequest

DetachRolePolicyRequest

DetachUserPolicyRequest

EnableMFADeviceRequest

EvaluationResult
Contains the results of a simulation.

This data type is used by the return parameter of SimulateCustomPolicy and SimulatePrincipalPolicy .

GenerateCredentialReportResponse
Contains the response to a successful GenerateCredentialReport request.

GetAccessKeyLastUsedRequest

GetAccessKeyLastUsedResponse
Contains the response to a successful GetAccessKeyLastUsed request. It is also returned as a member of the AccessKeyMetaData structure returned by the ListAccessKeys action.

GetAccountAuthorizationDetailsRequest

GetAccountAuthorizationDetailsResponse
Contains the response to a successful GetAccountAuthorizationDetails request.

GetAccountPasswordPolicyResponse
Contains the response to a successful GetAccountPasswordPolicy request.

GetAccountSummaryResponse
Contains the response to a successful GetAccountSummary request.

GetContextKeysForCustomPolicyRequest

GetContextKeysForPolicyResponse
Contains the response to a successful GetContextKeysForPrincipalPolicy or GetContextKeysForCustomPolicy request.

GetContextKeysForPrincipalPolicyRequest

GetCredentialReportResponse
Contains the response to a successful GetCredentialReport request.

GetGroupPolicyRequest

GetGroupPolicyResponse
Contains the response to a successful GetGroupPolicy request.

GetGroupRequest

GetGroupResponse
Contains the response to a successful GetGroup request.

GetInstanceProfileRequest

GetInstanceProfileResponse
Contains the response to a successful GetInstanceProfile request.

GetLoginProfileRequest

GetLoginProfileResponse
Contains the response to a successful GetLoginProfile request.

GetOpenIDConnectProviderRequest

GetOpenIDConnectProviderResponse
Contains the response to a successful GetOpenIDConnectProvider request.

GetPolicyRequest

GetPolicyResponse
Contains the response to a successful GetPolicy request.

GetPolicyVersionRequest

GetPolicyVersionResponse
Contains the response to a successful GetPolicyVersion request.

GetRolePolicyRequest

GetRolePolicyResponse
Contains the response to a successful GetRolePolicy request.

GetRoleRequest

GetRoleResponse
Contains the response to a successful GetRole request.

GetSAMLProviderRequest

GetSAMLProviderResponse
Contains the response to a successful GetSAMLProvider request.

GetSSHPublicKeyRequest

GetSSHPublicKeyResponse
Contains the response to a successful GetSSHPublicKey request.

GetServerCertificateRequest

GetServerCertificateResponse
Contains the response to a successful GetServerCertificate request.

GetUserPolicyRequest

GetUserPolicyResponse
Contains the response to a successful GetUserPolicy request.

GetUserRequest

GetUserResponse
Contains the response to a successful GetUser request.

Group

Contains information about an IAM group entity.

This data type is used as a response element in the following actions:

CreateGroup

GetGroup

ListGroups

GroupDetail
Contains information about an IAM group, including all of the group's policies.

This data type is used as a response element in the GetAccountAuthorizationDetails action.

IamClient
A client for the IAM API.

InstanceProfile

Contains information about an instance profile.

This data type is used as a response element in the following actions:

CreateInstanceProfile

GetInstanceProfile

ListInstanceProfiles

ListInstanceProfilesForRole

ListAccessKeysRequest

ListAccessKeysResponse
Contains the response to a successful ListAccessKeys request.

ListAccountAliasesRequest

ListAccountAliasesResponse
Contains the response to a successful ListAccountAliases request.

ListAttachedGroupPoliciesRequest

ListAttachedGroupPoliciesResponse
Contains the response to a successful ListAttachedGroupPolicies request.

ListAttachedRolePoliciesRequest

ListAttachedRolePoliciesResponse
Contains the response to a successful ListAttachedRolePolicies request.

ListAttachedUserPoliciesRequest

ListAttachedUserPoliciesResponse
Contains the response to a successful ListAttachedUserPolicies request.

ListEntitiesForPolicyRequest

ListEntitiesForPolicyResponse
Contains the response to a successful ListEntitiesForPolicy request.

ListGroupPoliciesRequest

ListGroupPoliciesResponse
Contains the response to a successful ListGroupPolicies request.

ListGroupsForUserRequest

ListGroupsForUserResponse
Contains the response to a successful ListGroupsForUser request.

ListGroupsRequest

ListGroupsResponse
Contains the response to a successful ListGroups request.

ListInstanceProfilesForRoleRequest

ListInstanceProfilesForRoleResponse
Contains the response to a successful ListInstanceProfilesForRole request.

ListInstanceProfilesRequest

ListInstanceProfilesResponse
Contains the response to a successful ListInstanceProfiles request.

ListMFADevicesRequest

ListMFADevicesResponse
Contains the response to a successful ListMFADevices request.

ListOpenIDConnectProvidersRequest

ListOpenIDConnectProvidersResponse
Contains the response to a successful ListOpenIDConnectProviders request.

ListPoliciesRequest

ListPoliciesResponse
Contains the response to a successful ListPolicies request.

ListPolicyVersionsRequest

ListPolicyVersionsResponse
Contains the response to a successful ListPolicyVersions request.

ListRolePoliciesRequest

ListRolePoliciesResponse
Contains the response to a successful ListRolePolicies request.

ListRolesRequest

ListRolesResponse
Contains the response to a successful ListRoles request.

ListSAMLProvidersRequest

ListSAMLProvidersResponse
Contains the response to a successful ListSAMLProviders request.

ListSSHPublicKeysRequest

ListSSHPublicKeysResponse
Contains the response to a successful ListSSHPublicKeys request.

ListServerCertificatesRequest

ListServerCertificatesResponse
Contains the response to a successful ListServerCertificates request.

ListServiceSpecificCredentialsRequest

ListServiceSpecificCredentialsResponse

ListSigningCertificatesRequest

ListSigningCertificatesResponse
Contains the response to a successful ListSigningCertificates request.

ListUserPoliciesRequest

ListUserPoliciesResponse
Contains the response to a successful ListUserPolicies request.

ListUsersRequest

ListUsersResponse
Contains the response to a successful ListUsers request.

ListVirtualMFADevicesRequest

ListVirtualMFADevicesResponse
Contains the response to a successful ListVirtualMFADevices request.

LoginProfile
Contains the user name and password create date for a user.

This data type is used as a response element in the CreateLoginProfile and GetLoginProfile actions.

MFADevice
Contains information about an MFA device.

This data type is used as a response element in the ListMFADevices action.

ManagedPolicyDetail
Contains information about a managed policy, including the policy's ARN, versions, and the number of principal entities (users, groups, and roles) that the policy is attached to.

This data type is used as a response element in the GetAccountAuthorizationDetails action.

For more information about managed policies, see Managed Policies and Inline Policies in the Using IAM guide.

OpenIDConnectProviderListEntry
Contains the Amazon Resource Name (ARN) for an IAM OpenID Connect provider.

PasswordPolicy
Contains information about the account password policy.

This data type is used as a response element in the GetAccountPasswordPolicy action.

Policy
Contains information about a managed policy.

This data type is used as a response element in the CreatePolicy, GetPolicy, and ListPolicies actions.

For more information about managed policies, refer to Managed Policies and Inline Policies in the Using IAM guide.

PolicyDetail
Contains information about an IAM policy, including the policy document.

This data type is used as a response element in the GetAccountAuthorizationDetails action.

PolicyGroup
Contains information about a group that a managed policy is attached to.

This data type is used as a response element in the ListEntitiesForPolicy action.

For more information about managed policies, refer to Managed Policies and Inline Policies in the Using IAM guide.

PolicyRole
Contains information about a role that a managed policy is attached to.

This data type is used as a response element in the ListEntitiesForPolicy action.

For more information about managed policies, refer to Managed Policies and Inline Policies in the Using IAM guide.

PolicyUser
Contains information about a user that a managed policy is attached to.

This data type is used as a response element in the ListEntitiesForPolicy action.

For more information about managed policies, refer to Managed Policies and Inline Policies in the Using IAM guide.

PolicyVersion
Contains information about a version of a managed policy.

This data type is used as a response element in the CreatePolicyVersion, GetPolicyVersion, ListPolicyVersions, and GetAccountAuthorizationDetails actions.

For more information about managed policies, refer to Managed Policies and Inline Policies in the Using IAM guide.

Position
Contains the row and column of a location of a Statement element in a policy document.

This data type is used as a member of the Statement type.

PutGroupPolicyRequest

PutRolePolicyRequest

PutUserPolicyRequest

RemoveClientIDFromOpenIDConnectProviderRequest

RemoveRoleFromInstanceProfileRequest

RemoveUserFromGroupRequest

ResetServiceSpecificCredentialRequest

ResetServiceSpecificCredentialResponse

ResourceSpecificResult
Contains the result of the simulation of a single API action call on a single resource.

This data type is used by a member of the EvaluationResult data type.

ResyncMFADeviceRequest

Role

Contains information about an IAM role.

This data type is used as a response element in the following actions:

CreateRole

GetRole

ListRoles

RoleDetail
Contains information about an IAM role, including all of the role's policies.

This data type is used as a response element in the GetAccountAuthorizationDetails action.

SAMLProviderListEntry
Contains the list of SAML providers for this account.

SSHPublicKey
Contains information about an SSH public key.

This data type is used as a response element in the GetSSHPublicKey and UploadSSHPublicKey actions.

SSHPublicKeyMetadata
Contains information about an SSH public key, without the key's body or fingerprint.

This data type is used as a response element in the ListSSHPublicKeys action.

ServerCertificate
Contains information about a server certificate.

This data type is used as a response element in the GetServerCertificate action.

ServerCertificateMetadata
Contains information about a server certificate without its certificate body, certificate chain, and private key.

This data type is used as a response element in the UploadServerCertificate and ListServerCertificates actions.

ServiceSpecificCredential
Contains the details of a service specific credential.

ServiceSpecificCredentialMetadata
Contains additional details about a service-specific credential.

SetDefaultPolicyVersionRequest

SigningCertificate
Contains information about an X.509 signing certificate.

This data type is used as a response element in the UploadSigningCertificate and ListSigningCertificates actions.

SimulateCustomPolicyRequest

SimulatePolicyResponse
Contains the response to a successful SimulatePrincipalPolicy or SimulateCustomPolicy request.

SimulatePrincipalPolicyRequest

Statement
Contains a reference to a Statement element in a policy document that determines the result of the simulation.

This data type is used by the MatchedStatements member of the EvaluationResult type.

UpdateAccessKeyRequest

UpdateAccountPasswordPolicyRequest

UpdateAssumeRolePolicyRequest

UpdateGroupRequest

UpdateLoginProfileRequest

UpdateOpenIDConnectProviderThumbprintRequest

UpdateSAMLProviderRequest

UpdateSAMLProviderResponse
Contains the response to a successful UpdateSAMLProvider request.

UpdateSSHPublicKeyRequest

UpdateServerCertificateRequest

UpdateServiceSpecificCredentialRequest

UpdateSigningCertificateRequest

UpdateUserRequest

UploadSSHPublicKeyRequest

UploadSSHPublicKeyResponse
Contains the response to a successful UploadSSHPublicKey request.

UploadServerCertificateRequest

UploadServerCertificateResponse
Contains the response to a successful UploadServerCertificate request.

UploadSigningCertificateRequest

UploadSigningCertificateResponse
Contains the response to a successful UploadSigningCertificate request.

User

Contains information about an IAM user entity.

This data type is used as a response element in the following actions:

CreateUser

GetUser

ListUsers

UserDetail
Contains information about an IAM user, including all the user's policies and all the IAM groups the user is in.

This data type is used as a response element in the GetAccountAuthorizationDetails action.

VirtualMFADevice
Contains information about a virtual MFA device.

Enums

AddClientIDToOpenIDConnectProviderError
Errors returned by AddClientIDToOpenIDConnectProvider

AddRoleToInstanceProfileError
Errors returned by AddRoleToInstanceProfile

AddUserToGroupError
Errors returned by AddUserToGroup

AttachGroupPolicyError
Errors returned by AttachGroupPolicy

AttachRolePolicyError
Errors returned by AttachRolePolicy

AttachUserPolicyError
Errors returned by AttachUserPolicy

ChangePasswordError
Errors returned by ChangePassword

CreateAccessKeyError
Errors returned by CreateAccessKey

CreateAccountAliasError
Errors returned by CreateAccountAlias

CreateGroupError
Errors returned by CreateGroup

CreateInstanceProfileError
Errors returned by CreateInstanceProfile

CreateLoginProfileError
Errors returned by CreateLoginProfile

CreateOpenIDConnectProviderError
Errors returned by CreateOpenIDConnectProvider

CreatePolicyError
Errors returned by CreatePolicy

CreatePolicyVersionError
Errors returned by CreatePolicyVersion

CreateRoleError
Errors returned by CreateRole

CreateSAMLProviderError
Errors returned by CreateSAMLProvider

CreateServiceSpecificCredentialError
Errors returned by CreateServiceSpecificCredential

CreateUserError
Errors returned by CreateUser

CreateVirtualMFADeviceError
Errors returned by CreateVirtualMFADevice

DeactivateMFADeviceError
Errors returned by DeactivateMFADevice

DeleteAccessKeyError
Errors returned by DeleteAccessKey

DeleteAccountAliasError
Errors returned by DeleteAccountAlias

DeleteAccountPasswordPolicyError
Errors returned by DeleteAccountPasswordPolicy

DeleteGroupError
Errors returned by DeleteGroup

DeleteGroupPolicyError
Errors returned by DeleteGroupPolicy

DeleteInstanceProfileError
Errors returned by DeleteInstanceProfile

DeleteLoginProfileError
Errors returned by DeleteLoginProfile

DeleteOpenIDConnectProviderError
Errors returned by DeleteOpenIDConnectProvider

DeletePolicyError
Errors returned by DeletePolicy

DeletePolicyVersionError
Errors returned by DeletePolicyVersion

DeleteRoleError
Errors returned by DeleteRole

DeleteRolePolicyError
Errors returned by DeleteRolePolicy

DeleteSAMLProviderError
Errors returned by DeleteSAMLProvider

DeleteSSHPublicKeyError
Errors returned by DeleteSSHPublicKey

DeleteServerCertificateError
Errors returned by DeleteServerCertificate

DeleteServiceSpecificCredentialError
Errors returned by DeleteServiceSpecificCredential

DeleteSigningCertificateError
Errors returned by DeleteSigningCertificate

DeleteUserError
Errors returned by DeleteUser

DeleteUserPolicyError
Errors returned by DeleteUserPolicy

DeleteVirtualMFADeviceError
Errors returned by DeleteVirtualMFADevice

DetachGroupPolicyError
Errors returned by DetachGroupPolicy

DetachRolePolicyError
Errors returned by DetachRolePolicy

DetachUserPolicyError
Errors returned by DetachUserPolicy

EnableMFADeviceError
Errors returned by EnableMFADevice

GenerateCredentialReportError
Errors returned by GenerateCredentialReport

GetAccessKeyLastUsedError
Errors returned by GetAccessKeyLastUsed

GetAccountAuthorizationDetailsError
Errors returned by GetAccountAuthorizationDetails

GetAccountPasswordPolicyError
Errors returned by GetAccountPasswordPolicy

GetAccountSummaryError
Errors returned by GetAccountSummary

GetContextKeysForCustomPolicyError
Errors returned by GetContextKeysForCustomPolicy

GetContextKeysForPrincipalPolicyError
Errors returned by GetContextKeysForPrincipalPolicy

GetCredentialReportError
Errors returned by GetCredentialReport

GetGroupError
Errors returned by GetGroup

GetGroupPolicyError
Errors returned by GetGroupPolicy

GetInstanceProfileError
Errors returned by GetInstanceProfile

GetLoginProfileError
Errors returned by GetLoginProfile

GetOpenIDConnectProviderError
Errors returned by GetOpenIDConnectProvider

GetPolicyError
Errors returned by GetPolicy

GetPolicyVersionError
Errors returned by GetPolicyVersion

GetRoleError
Errors returned by GetRole

GetRolePolicyError
Errors returned by GetRolePolicy

GetSAMLProviderError
Errors returned by GetSAMLProvider

GetSSHPublicKeyError
Errors returned by GetSSHPublicKey

GetServerCertificateError
Errors returned by GetServerCertificate

GetUserError
Errors returned by GetUser

GetUserPolicyError
Errors returned by GetUserPolicy

ListAccessKeysError
Errors returned by ListAccessKeys

ListAccountAliasesError
Errors returned by ListAccountAliases

ListAttachedGroupPoliciesError
Errors returned by ListAttachedGroupPolicies

ListAttachedRolePoliciesError
Errors returned by ListAttachedRolePolicies

ListAttachedUserPoliciesError
Errors returned by ListAttachedUserPolicies

ListEntitiesForPolicyError
Errors returned by ListEntitiesForPolicy

ListGroupPoliciesError
Errors returned by ListGroupPolicies

ListGroupsError
Errors returned by ListGroups

ListGroupsForUserError
Errors returned by ListGroupsForUser

ListInstanceProfilesError
Errors returned by ListInstanceProfiles

ListInstanceProfilesForRoleError
Errors returned by ListInstanceProfilesForRole

ListMFADevicesError
Errors returned by ListMFADevices

ListOpenIDConnectProvidersError
Errors returned by ListOpenIDConnectProviders

ListPoliciesError
Errors returned by ListPolicies

ListPolicyVersionsError
Errors returned by ListPolicyVersions

ListRolePoliciesError
Errors returned by ListRolePolicies

ListRolesError
Errors returned by ListRoles

ListSAMLProvidersError
Errors returned by ListSAMLProviders

ListSSHPublicKeysError
Errors returned by ListSSHPublicKeys

ListServerCertificatesError
Errors returned by ListServerCertificates

ListServiceSpecificCredentialsError
Errors returned by ListServiceSpecificCredentials

ListSigningCertificatesError
Errors returned by ListSigningCertificates

ListUserPoliciesError
Errors returned by ListUserPolicies

ListUsersError
Errors returned by ListUsers

ListVirtualMFADevicesError
Errors returned by ListVirtualMFADevices

PutGroupPolicyError
Errors returned by PutGroupPolicy

PutRolePolicyError
Errors returned by PutRolePolicy

PutUserPolicyError
Errors returned by PutUserPolicy

RemoveClientIDFromOpenIDConnectProviderError
Errors returned by RemoveClientIDFromOpenIDConnectProvider

RemoveRoleFromInstanceProfileError
Errors returned by RemoveRoleFromInstanceProfile

RemoveUserFromGroupError
Errors returned by RemoveUserFromGroup

ResetServiceSpecificCredentialError
Errors returned by ResetServiceSpecificCredential

ResyncMFADeviceError
Errors returned by ResyncMFADevice

SetDefaultPolicyVersionError
Errors returned by SetDefaultPolicyVersion

SimulateCustomPolicyError
Errors returned by SimulateCustomPolicy

SimulatePrincipalPolicyError
Errors returned by SimulatePrincipalPolicy

UpdateAccessKeyError
Errors returned by UpdateAccessKey

UpdateAccountPasswordPolicyError
Errors returned by UpdateAccountPasswordPolicy

UpdateAssumeRolePolicyError
Errors returned by UpdateAssumeRolePolicy

UpdateGroupError
Errors returned by UpdateGroup

UpdateLoginProfileError
Errors returned by UpdateLoginProfile

UpdateOpenIDConnectProviderThumbprintError
Errors returned by UpdateOpenIDConnectProviderThumbprint

UpdateSAMLProviderError
Errors returned by UpdateSAMLProvider

UpdateSSHPublicKeyError
Errors returned by UpdateSSHPublicKey

UpdateServerCertificateError
Errors returned by UpdateServerCertificate

UpdateServiceSpecificCredentialError
Errors returned by UpdateServiceSpecificCredential

UpdateSigningCertificateError
Errors returned by UpdateSigningCertificate

UpdateUserError
Errors returned by UpdateUser

UploadSSHPublicKeyError
Errors returned by UploadSSHPublicKey

UploadServerCertificateError
Errors returned by UploadServerCertificate

UploadSigningCertificateError
Errors returned by UploadSigningCertificate

Type Definitions

AccessKeyIdType

AccessKeyMetadataListType
Contains a list of access key metadata.

This data type is used as a response element in the ListAccessKeys action.

AccessKeySecretType

AccountAliasListType

AccountAliasType

ActionNameListType

ActionNameType

ArnType
The Amazon Resource Name (ARN). ARNs are unique identifiers for AWS resources.

For more information about ARNs, go to Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

AssignmentStatusType

AttachedPoliciesListType

AttachmentCountType

AuthenticationCodeType

BooleanObjectType

BooleanType

BootstrapDatum

CertificateBodyType

CertificateChainType

CertificateIdType

CertificateListType
Contains a list of signing certificates.

This data type is used as a response element in the ListSigningCertificates action.

ClientIDListType

ClientIDType

ColumnNumber

ContextEntryListType

ContextKeyNameType

ContextKeyNamesResultListType

ContextKeyTypeEnum

ContextKeyValueListType

ContextKeyValueType

CredentialReportExpiredExceptionMessage

CredentialReportNotPresentExceptionMessage

CredentialReportNotReadyExceptionMessage

DateType

DeleteConflictMessage

DuplicateCertificateMessage

DuplicateSSHPublicKeyMessage

EncodingType

EntityAlreadyExistsMessage

EntityListType

EntityTemporarilyUnmodifiableMessage

EntityType

EvalDecisionDetailsType

EvalDecisionSourceType

EvaluationResultsListType

ExistingUserNameType

GroupDetailListType

GroupListType
Contains a list of IAM groups.

This data type is used as a response element in the ListGroups action.

GroupNameListType

GroupNameType

IdType

InstanceProfileListType
Contains a list of instance profiles.

InstanceProfileNameType

InvalidAuthenticationCodeMessage

InvalidCertificateMessage

InvalidInputMessage

InvalidPublicKeyMessage

InvalidUserTypeMessage

KeyPairMismatchMessage

LimitExceededMessage

LineNumber

MalformedCertificateMessage

MalformedPolicyDocumentMessage

ManagedPolicyDetailListType

MarkerType

MaxItemsType

MaxPasswordAgeType

MfaDeviceListType
Contains a list of MFA devices.

This data type is used as a response element in the ListMFADevices and ListVirtualMFADevices actions.

MinimumPasswordLengthType

NoSuchEntityMessage

OpenIDConnectProviderListType
Contains a list of IAM OpenID Connect providers.

OpenIDConnectProviderUrlType
Contains a URL that specifies the endpoint for an OpenID Connect provider.

PasswordPolicyViolationMessage

PasswordReusePreventionType

PasswordType

PathPrefixType

PathType

PolicyDescriptionType

PolicyDetailListType

PolicyDocumentType

PolicyDocumentVersionListType

PolicyEvaluationDecisionType

PolicyEvaluationErrorMessage

PolicyGroupListType

PolicyIdentifierType

PolicyListType

PolicyNameListType
Contains a list of policy names.

This data type is used as a response element in the ListPolicies action.

PolicyNameType

PolicyPathType

PolicyRoleListType

PolicyScopeType

PolicySourceType

PolicyUserListType

PolicyVersionIdType

PrivateKeyType

PublicKeyFingerprintType

PublicKeyIdType

PublicKeyMaterialType

ReportContentType

ReportFormatType

ReportStateDescriptionType

ReportStateType

ResourceHandlingOptionType

ResourceNameListType

ResourceNameType

ResourceSpecificResultListType

RoleDetailListType

RoleListType
Contains a list of IAM roles.

This data type is used as a response element in the ListRoles action.

RoleNameType

SAMLMetadataDocumentType

SAMLProviderListType

SAMLProviderNameType

SSHPublicKeyListType

SerialNumberType

ServerCertificateMetadataListType

ServerCertificateNameType

ServiceFailureExceptionMessage

ServiceName

ServiceNotSupportedMessage

ServicePassword

ServiceSpecificCredentialId

ServiceSpecificCredentialsListType

ServiceUserName

SimulationPolicyListType

StatementListType

StatusType

StringType

SummaryKeyType

SummaryMapType

SummaryValueType

ThumbprintListType
Contains a list of thumbprints of identity provider server certificates.

ThumbprintType
Contains a thumbprint for an identity provider's server certificate.

The identity provider's server certificate thumbprint is the hex-encoded SHA-1 hash value of the self-signed X.509 certificate used by the domain where the OpenID Connect provider makes its keys available. It is always a 40-character string.

UnrecognizedPublicKeyEncodingMessage

UserDetailListType

UserListType
Contains a list of users.

This data type is used as a response element in the GetGroup and ListUsers actions.

UserNameType

VirtualMFADeviceListType

VirtualMFADeviceName