Struct rusoto::iam::UpdateServerCertificateRequest

pub struct UpdateServerCertificateRequest {
    pub new_path: Option<PathType>,
    pub new_server_certificate_name: Option<ServerCertificateNameType>,
    pub server_certificate_name: ServerCertificateNameType,
}

Fields
new_path: Option<PathType>

The new path for the server certificate. Include this only if you are updating the server certificate's path.

The regex pattern for this parameter is a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (\u0021) thru the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

new_server_certificate_name: Option<ServerCertificateNameType>

The new name for the server certificate. Include this only if you are updating the server certificate's name. The name of the certificate cannot contain any spaces.

The regex pattern for this parameter is a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

server_certificate_name: ServerCertificateNameType

The name of the server certificate that you want to update.

The regex pattern for this parameter is a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-