Struct rusoto::importexport::ImportExportClient

pub struct ImportExportClient<P, D> where P: ProvideAwsCredentials,
        D: DispatchSignedRequest { /* fields omitted */ }

A client for the AWS Import/Export API.

Methods

impl<P, D> ImportExportClient<P, D> where P: ProvideAwsCredentials,
D: DispatchSignedRequest

fn new(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn cancel_job(&self,
input: &CancelJobInput)
-> Result<CancelJobOutput, CancelJobError>

This operation cancels a specified job. Only the job owner can cancel it. The operation fails if the job has already started or is complete.

fn create_job(&self,
input: &CreateJobInput)
-> Result<CreateJobOutput, CreateJobError>

This operation initiates the process of scheduling an upload or download of your data. You include in the request a manifest that describes the data transfer specifics. The response to the request includes a job ID, which you can use in other operations, a signature that you use to identify your storage device, and the address where you should ship your storage device.

fn get_shipping_label(&self,
input: &GetShippingLabelInput)
-> Result<GetShippingLabelOutput, GetShippingLabelError>

This operation generates a pre-paid UPS shipping label that you will use to ship your device to AWS for processing.

fn get_status(&self,
input: &GetStatusInput)
-> Result<GetStatusOutput, GetStatusError>

This operation returns information about a job, including where the job is in the processing pipeline, the status of the results, and the signature value associated with the job. You can only return information about jobs you own.

fn list_jobs(&self,
input: &ListJobsInput)
-> Result<ListJobsOutput, ListJobsError>

This operation returns the jobs associated with the requester. AWS Import/Export lists the jobs in reverse chronological order based on the date of creation. For example if Job Test1 was created 2009Dec30 and Test2 was created 2010Feb05, the ListJobs operation would return Test2 followed by Test1.

fn update_job(&self,
input: &UpdateJobInput)
-> Result<UpdateJobOutput, UpdateJobError>

You use this operation to change the parameters specified in the original manifest file by supplying a new manifest file. The manifest file attached to this request replaces the original manifest file. You can only use the operation after a CreateJob request but before the data transfer starts and you can only use it on jobs you own.