Module rusoto::inspector

Amazon Inspector

Structs

AddAttributesToFindingsRequest

AddAttributesToFindingsResponse

AgentAlreadyRunningAssessment
Used in the exception error that is thrown if you start an assessment run for an assessment target that includes an EC2 instance that is already participating in another started assessment run.

AgentFilter
Contains information about an Amazon Inspector agent. This data type is used as a request parameter in the ListAssessmentRunAgents action.

AgentPreview
Used as a response element in the PreviewAgents action.

AssessmentRun
A snapshot of an Amazon Inspector assessment run that contains the findings of the assessment run .

Used as the response element in the DescribeAssessmentRuns action.

AssessmentRunAgent
Contains information about an Amazon Inspector agent. This data type is used as a response element in the ListAssessmentRunAgents action.

AssessmentRunFilter
Used as the request parameter in the ListAssessmentRuns action.

AssessmentRunNotification
Used as one of the elements of the AssessmentRun data type.

AssessmentRunStateChange
Used as one of the elements of the AssessmentRun data type.

AssessmentTarget
Contains information about an Amazon Inspector application. This data type is used as the response element in the DescribeAssessmentTargets action.

AssessmentTargetFilter
Used as the request parameter in the ListAssessmentTargets action.

AssessmentTemplate
Contains information about an Amazon Inspector assessment template. This data type is used as the response element in the DescribeAssessmentTemplates action.

AssessmentTemplateFilter
Used as the request parameter in the ListAssessmentTemplates action.

AssetAttributes
A collection of attributes of the host from which the finding is generated.

Attribute
This data type is used as a request parameter in the AddAttributesToFindings and CreateAssessmentTemplate actions.

CreateAssessmentTargetRequest

CreateAssessmentTargetResponse

CreateAssessmentTemplateRequest

CreateAssessmentTemplateResponse

CreateResourceGroupRequest

CreateResourceGroupResponse

DeleteAssessmentRunRequest

DeleteAssessmentTargetRequest

DeleteAssessmentTemplateRequest

DescribeAssessmentRunsRequest

DescribeAssessmentRunsResponse

DescribeAssessmentTargetsRequest

DescribeAssessmentTargetsResponse

DescribeAssessmentTemplatesRequest

DescribeAssessmentTemplatesResponse

DescribeCrossAccountAccessRoleResponse

DescribeFindingsRequest

DescribeFindingsResponse

DescribeResourceGroupsRequest

DescribeResourceGroupsResponse

DescribeRulesPackagesRequest

DescribeRulesPackagesResponse

DurationRange
This data type is used in the AssessmentTemplateFilter data type.

EventSubscription
This data type is used in the Subscription data type.

FailedItemDetails
Includes details about the failed items.

Finding
Contains information about an Amazon Inspector finding. This data type is used as the response element in the DescribeFindings action.

FindingFilter
This data type is used as a request parameter in the ListFindings action.

GetTelemetryMetadataRequest

GetTelemetryMetadataResponse

InspectorClient
A client for the Amazon Inspector API.

InspectorServiceAttributes
This data type is used in the Finding data type.

ListAssessmentRunAgentsRequest

ListAssessmentRunAgentsResponse

ListAssessmentRunsRequest

ListAssessmentRunsResponse

ListAssessmentTargetsRequest

ListAssessmentTargetsResponse

ListAssessmentTemplatesRequest

ListAssessmentTemplatesResponse

ListEventSubscriptionsRequest

ListEventSubscriptionsResponse

ListFindingsRequest

ListFindingsResponse

ListRulesPackagesRequest

ListRulesPackagesResponse

ListTagsForResourceRequest

ListTagsForResourceResponse

PreviewAgentsRequest

PreviewAgentsResponse

RegisterCrossAccountAccessRoleRequest

RemoveAttributesFromFindingsRequest

RemoveAttributesFromFindingsResponse

ResourceGroup
Contains information about a resource group. The resource group defines a set of tags that, when queried, identify the AWS resources that make up the assessment target. This data type is used as the response element in the DescribeResourceGroups action.

ResourceGroupTag
This data type is used as one of the elements of the ResourceGroup data type.

RulesPackage
Contains information about an Amazon Inspector rules package. This data type is used as the response element in the DescribeRulesPackages action.

SetTagsForResourceRequest

StartAssessmentRunRequest

StartAssessmentRunResponse

StopAssessmentRunRequest

SubscribeToEventRequest

Subscription
This data type is used as a response element in the ListEventSubscriptions action.

Tag
A key and value pair. This data type is used as a request parameter in the SetTagsForResource action and a response element in the ListTagsForResource action.

TelemetryMetadata
The metadata about the Amazon Inspector application data metrics collected by the agent. This data type is used as the response element in the GetTelemetryMetadata action.

TimestampRange
This data type is used in the AssessmentRunFilter data type.

UnsubscribeFromEventRequest

UpdateAssessmentTargetRequest

Enums

AddAttributesToFindingsError

CreateAssessmentTargetError

CreateAssessmentTemplateError

CreateResourceGroupError

DeleteAssessmentRunError

DeleteAssessmentTargetError

DeleteAssessmentTemplateError

DescribeAssessmentRunsError

DescribeAssessmentTargetsError

DescribeAssessmentTemplatesError

DescribeCrossAccountAccessRoleError

DescribeFindingsError

DescribeResourceGroupsError

DescribeRulesPackagesError

GetTelemetryMetadataError

ListAssessmentRunAgentsError

ListAssessmentRunsError

ListAssessmentTargetsError

ListAssessmentTemplatesError

ListEventSubscriptionsError

ListFindingsError

ListRulesPackagesError

ListTagsForResourceError

PreviewAgentsError

RegisterCrossAccountAccessRoleError

RemoveAttributesFromFindingsError

SetTagsForResourceError

StartAssessmentRunError

StopAssessmentRunError

SubscribeToEventError

UnsubscribeFromEventError

UpdateAssessmentTargetError

Type Definitions

AccessDeniedErrorCode

AddRemoveAttributesFindingArnList

AgentAlreadyRunningAssessmentList

AgentHealth

AgentHealthCode

AgentHealthCodeList

AgentHealthList

AgentId

AgentIdList

AgentPreviewList

AmiId

Arn

AssessmentRulesPackageArnList

AssessmentRunAgentList

AssessmentRunDuration

AssessmentRunInProgressArnList

AssessmentRunList

AssessmentRunName

AssessmentRunNotificationList

AssessmentRunNotificationSnsStatusCode

AssessmentRunState

AssessmentRunStateChangeList

AssessmentRunStateList

AssessmentTargetList

AssessmentTargetName

AssessmentTemplateList

AssessmentTemplateName

AssessmentTemplateRulesPackageArnList

AssetType

AttributeKey

AttributeList

AttributeValue

AutoScalingGroup

AutoScalingGroupList

BatchDescribeArnList

Bool

ErrorMessage

EventSubscriptionList

FailedItemErrorCode

FailedItems

FilterRulesPackageArnList

FindingId

FindingList

Hostname

InspectorEvent

InvalidCrossAccountRoleErrorCode

InvalidInputErrorCode

IocConfidence

Ipv4Address

Ipv4AddressList

LimitExceededErrorCode

ListEventSubscriptionsMaxResults

ListMaxResults

ListParentArnList

ListReturnedArnList

Locale

Long

Message

MessageType

NamePattern

NoSuchEntityErrorCode

NumericSeverity

NumericVersion

PaginationToken

PreviewAgentsMaxResults

ProviderName

ResourceGroupList

ResourceGroupTags

RuleName

RuleNameList

RulesPackageList

RulesPackageName

ServiceName

Severity

SeverityList

SubscriptionList

TagKey

TagList

TagValue

TelemetryMetadataList

Text

Timestamp

UserAttributeKeyList

UserAttributeList

Version