Struct rusoto::inspector::AssessmentRunAgent

pub struct AssessmentRunAgent {
    pub agent_health: AgentHealth,
    pub agent_health_code: AgentHealthCode,
    pub agent_health_details: Option<Message>,
    pub agent_id: AgentId,
    pub assessment_run_arn: Arn,
    pub auto_scaling_group: Option<AutoScalingGroup>,
    pub telemetry_metadata: TelemetryMetadataList,
}

Contains information about an Amazon Inspector agent. This data type is used as a response element in the ListAssessmentRunAgents action.

Fields
agent_health: AgentHealth

The current health state of the agent.

agent_health_code: AgentHealthCode

The detailed health state of the agent.

agent_health_details: Option<Message>

The description for the agent health code.

agent_id: AgentId

The AWS account of the EC2 instance where the agent is installed.

assessment_run_arn: Arn

The ARN of the assessment run that is associated with the agent.

auto_scaling_group: Option<AutoScalingGroup>

The Auto Scaling group of the EC2 instance that is specified by the agent ID.

telemetry_metadata: TelemetryMetadataList

The Amazon Inspector application data metrics that are collected by the agent.