Struct rusoto::inspector::AssetAttributes

pub struct AssetAttributes {
    pub agent_id: Option<AgentId>,
    pub ami_id: Option<AmiId>,
    pub auto_scaling_group: Option<AutoScalingGroup>,
    pub hostname: Option<Hostname>,
    pub ipv_4_addresses: Option<Ipv4AddressList>,
    pub schema_version: NumericVersion,
}

A collection of attributes of the host from which the finding is generated.

Fields
agent_id: Option<AgentId>

The ID of the agent that is installed on the EC2 instance where the finding is generated.

ami_id: Option<AmiId>

The ID of the Amazon Machine Image (AMI) that is installed on the EC2 instance where the finding is generated.

auto_scaling_group: Option<AutoScalingGroup>

The Auto Scaling group of the EC2 instance where the finding is generated.

hostname: Option<Hostname>

The hostname of the EC2 instance where the finding is generated.

ipv_4_addresses: Option<Ipv4AddressList>

The list of IP v4 addresses of the EC2 instance where the finding is generated.

schema_version: NumericVersion

The schema version of this data type.