Struct rusoto::inspector::CreateAssessmentTemplateRequest

pub struct CreateAssessmentTemplateRequest {
    pub assessment_target_arn: Arn,
    pub assessment_template_name: AssessmentTemplateName,
    pub duration_in_seconds: AssessmentRunDuration,
    pub rules_package_arns: AssessmentTemplateRulesPackageArnList,
    pub user_attributes_for_findings: Option<UserAttributeList>,
}

Fields
assessment_target_arn: Arn

The ARN that specifies the assessment target for which you want to create the assessment template.

assessment_template_name: AssessmentTemplateName

The user-defined name that identifies the assessment template that you want to create. You can create several assessment templates for an assessment target. The names of the assessment templates that correspond to a particular assessment target must be unique.

duration_in_seconds: AssessmentRunDuration

The duration of the assessment run in seconds. The default value is 3600 seconds (one hour).

rules_package_arns: AssessmentTemplateRulesPackageArnList

The ARNs that specify the rules packages that you want to attach to the assessment template.

user_attributes_for_findings: Option<UserAttributeList>

The user-defined attributes that are assigned to every finding that is generated by the assessment run that uses this assessment template.