Struct rusoto::inspector::Finding

pub struct Finding {
    pub arn: Arn,
    pub asset_attributes: Option<AssetAttributes>,
    pub asset_type: Option<AssetType>,
    pub attributes: AttributeList,
    pub confidence: Option<IocConfidence>,
    pub created_at: Timestamp,
    pub description: Option<Text>,
    pub id: Option<FindingId>,
    pub indicator_of_compromise: Option<Bool>,
    pub numeric_severity: Option<NumericSeverity>,
    pub recommendation: Option<Text>,
    pub schema_version: Option<NumericVersion>,
    pub service: Option<ServiceName>,
    pub service_attributes: Option<InspectorServiceAttributes>,
    pub severity: Option<Severity>,
    pub title: Option<Text>,
    pub updated_at: Timestamp,
    pub user_attributes: UserAttributeList,
}

Contains information about an Amazon Inspector finding. This data type is used as the response element in the DescribeFindings action.

Fields
arn: Arn

The ARN that specifies the finding.

asset_attributes: Option<AssetAttributes>

A collection of attributes of the host from which the finding is generated.

asset_type: Option<AssetType>

The type of the host from which the finding is generated.

attributes: AttributeList

The system-defined attributes for the finding.

confidence: Option<IocConfidence>

This data element is currently not used.

created_at: Timestamp

The time when the finding was generated.

description: Option<Text>

The description of the finding.

id: Option<FindingId>

The ID of the finding.

indicator_of_compromise: Option<Bool>

This data element is currently not used.

numeric_severity: Option<NumericSeverity>

The numeric value of the finding severity.

recommendation: Option<Text>

The recommendation for the finding.

schema_version: Option<NumericVersion>

The schema version of this data type.

service: Option<ServiceName>

The data element is set to "Inspector".

service_attributes: Option<InspectorServiceAttributes> severity: Option<Severity>

The finding severity. Values can be set to High, Medium, Low, and Informational.

title: Option<Text>

The name of the finding.

updated_at: Timestamp

The time when AddAttributesToFindings is called.

user_attributes: UserAttributeList

The user-defined attributes that are assigned to the finding.