Struct rusoto::inspector::InspectorClient

pub struct InspectorClient<P, D> where P: ProvideAwsCredentials,
        D: DispatchSignedRequest { /* fields omitted */ }

A client for the Amazon Inspector API.

Methods

impl<P, D> InspectorClient<P, D> where P: ProvideAwsCredentials,
D: DispatchSignedRequest

fn new(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn add_attributes_to_findings(&self,
input: &AddAttributesToFindingsRequest)
-> Result<AddAttributesToFindingsResponse, AddAttributesToFindingsError>

Assigns attributes (key and value pairs) to the findings that are specified by the ARNs of the findings.

fn create_assessment_target(&self,
input: &CreateAssessmentTargetRequest)
-> Result<CreateAssessmentTargetResponse, CreateAssessmentTargetError>

Creates a new assessment target using the ARN of the resource group that is generated by CreateResourceGroup. You can create up to 50 assessment targets per AWS account. You can run up to 500 concurrent agents per AWS account. For more information, see Amazon Inspector Assessment Targets.

fn create_assessment_template(&self,
input: &CreateAssessmentTemplateRequest)
-> Result<CreateAssessmentTemplateResponse, CreateAssessmentTemplateError>

Creates an assessment template for the assessment target that is specified by the ARN of the assessment target.

fn create_resource_group(&self,
input: &CreateResourceGroupRequest)
-> Result<CreateResourceGroupResponse, CreateResourceGroupError>

Creates a resource group using the specified set of tags (key and value pairs) that are used to select the EC2 instances to be included in an Amazon Inspector assessment target. The created resource group is then used to create an Amazon Inspector assessment target. For more information, see CreateAssessmentTarget.

fn delete_assessment_run(&self,
input: &DeleteAssessmentRunRequest)
-> Result<(), DeleteAssessmentRunError>

Deletes the assessment run that is specified by the ARN of the assessment run.

fn delete_assessment_target(&self,
input: &DeleteAssessmentTargetRequest)
-> Result<(), DeleteAssessmentTargetError>

Deletes the assessment target that is specified by the ARN of the assessment target.

fn delete_assessment_template(&self,
input: &DeleteAssessmentTemplateRequest)
-> Result<(), DeleteAssessmentTemplateError>

Deletes the assessment template that is specified by the ARN of the assessment template.

fn describe_assessment_runs(&self,
input: &DescribeAssessmentRunsRequest)
-> Result<DescribeAssessmentRunsResponse, DescribeAssessmentRunsError>

Describes the assessment runs that are specified by the ARNs of the assessment runs.

fn describe_assessment_targets(&self,
input: &DescribeAssessmentTargetsRequest)
-> Result<DescribeAssessmentTargetsResponse, DescribeAssessmentTargetsError>

Describes the assessment targets that are specified by the ARNs of the assessment targets.

fn describe_assessment_templates(&self,
input: &DescribeAssessmentTemplatesRequest)
-> Result<DescribeAssessmentTemplatesResponse, DescribeAssessmentTemplatesError>

Describes the assessment templates that are specified by the ARNs of the assessment templates.

fn describe_cross_account_access_role(&self)
-> Result<DescribeCrossAccountAccessRoleResponse, DescribeCrossAccountAccessRoleError>

Describes the IAM role that enables Amazon Inspector to access your AWS account.

fn describe_findings(&self,
input: &DescribeFindingsRequest)
-> Result<DescribeFindingsResponse, DescribeFindingsError>

Describes the findings that are specified by the ARNs of the findings.

fn describe_resource_groups(&self,
input: &DescribeResourceGroupsRequest)
-> Result<DescribeResourceGroupsResponse, DescribeResourceGroupsError>

Describes the resource groups that are specified by the ARNs of the resource groups.

fn describe_rules_packages(&self,
input: &DescribeRulesPackagesRequest)
-> Result<DescribeRulesPackagesResponse, DescribeRulesPackagesError>

Describes the rules packages that are specified by the ARNs of the rules packages.

fn get_telemetry_metadata(&self,
input: &GetTelemetryMetadataRequest)
-> Result<GetTelemetryMetadataResponse, GetTelemetryMetadataError>

Information about the data that is collected for the specified assessment run.

fn list_assessment_run_agents(&self,
input: &ListAssessmentRunAgentsRequest)
-> Result<ListAssessmentRunAgentsResponse, ListAssessmentRunAgentsError>

Lists the agents of the assessment runs that are specified by the ARNs of the assessment runs.

fn list_assessment_runs(&self,
input: &ListAssessmentRunsRequest)
-> Result<ListAssessmentRunsResponse, ListAssessmentRunsError>

Lists the assessment runs that correspond to the assessment templates that are specified by the ARNs of the assessment templates.

fn list_assessment_targets(&self,
input: &ListAssessmentTargetsRequest)
-> Result<ListAssessmentTargetsResponse, ListAssessmentTargetsError>

Lists the ARNs of the assessment targets within this AWS account. For more information about assessment targets, see Amazon Inspector Assessment Targets.

fn list_assessment_templates(&self,
input: &ListAssessmentTemplatesRequest)
-> Result<ListAssessmentTemplatesResponse, ListAssessmentTemplatesError>

Lists the assessment templates that correspond to the assessment targets that are specified by the ARNs of the assessment targets.

fn list_event_subscriptions(&self,
input: &ListEventSubscriptionsRequest)
-> Result<ListEventSubscriptionsResponse, ListEventSubscriptionsError>

Lists all the event subscriptions for the assessment template that is specified by the ARN of the assessment template. For more information, see SubscribeToEvent and UnsubscribeFromEvent.

fn list_findings(&self,
input: &ListFindingsRequest)
-> Result<ListFindingsResponse, ListFindingsError>

Lists findings that are generated by the assessment runs that are specified by the ARNs of the assessment runs.

fn list_rules_packages(&self,
input: &ListRulesPackagesRequest)
-> Result<ListRulesPackagesResponse, ListRulesPackagesError>

Lists all available Amazon Inspector rules packages.

fn list_tags_for_resource(&self,
input: &ListTagsForResourceRequest)
-> Result<ListTagsForResourceResponse, ListTagsForResourceError>

Lists all tags associated with an assessment template.

fn preview_agents(&self,
input: &PreviewAgentsRequest)
-> Result<PreviewAgentsResponse, PreviewAgentsError>

Previews the agents installed on the EC2 instances that are part of the specified assessment target.

fn register_cross_account_access_role(&self,
input: &RegisterCrossAccountAccessRoleRequest)
-> Result<(), RegisterCrossAccountAccessRoleError>

Registers the IAM role that Amazon Inspector uses to list your EC2 instances at the start of the assessment run or when you call the PreviewAgents action.

fn remove_attributes_from_findings(&self,
input: &RemoveAttributesFromFindingsRequest)
-> Result<RemoveAttributesFromFindingsResponse, RemoveAttributesFromFindingsError>

Removes entire attributes (key and value pairs) from the findings that are specified by the ARNs of the findings where an attribute with the specified key exists.

fn set_tags_for_resource(&self,
input: &SetTagsForResourceRequest)
-> Result<(), SetTagsForResourceError>

Sets tags (key and value pairs) to the assessment template that is specified by the ARN of the assessment template.

fn start_assessment_run(&self,
input: &StartAssessmentRunRequest)
-> Result<StartAssessmentRunResponse, StartAssessmentRunError>

Starts the assessment run specified by the ARN of the assessment template. For this API to function properly, you must not exceed the limit of running up to 500 concurrent agents per AWS account.

fn stop_assessment_run(&self,
input: &StopAssessmentRunRequest)
-> Result<(), StopAssessmentRunError>

Stops the assessment run that is specified by the ARN of the assessment run.

fn subscribe_to_event(&self,
input: &SubscribeToEventRequest)
-> Result<(), SubscribeToEventError>

Enables the process of sending Amazon Simple Notification Service (SNS) notifications about a specified event to a specified SNS topic.

fn unsubscribe_from_event(&self,
input: &UnsubscribeFromEventRequest)
-> Result<(), UnsubscribeFromEventError>

Disables the process of sending Amazon Simple Notification Service (SNS) notifications about a specified event to a specified SNS topic.

fn update_assessment_target(&self,
input: &UpdateAssessmentTargetRequest)
-> Result<(), UpdateAssessmentTargetError>

Updates the assessment target that is specified by the ARN of the assessment target.