Module rusoto::iot

Amazon IoT

Structs

AcceptCertificateTransferRequest
The input for the AcceptCertificateTransfer operation.

Action
Describes the actions associated with a rule.

AttachPrincipalPolicyRequest
The input for the AttachPrincipalPolicy operation.

AttachThingPrincipalRequest
The input for the AttachThingPrincipal operation.

AttachThingPrincipalResponse
The output from the AttachThingPrincipal operation.

AttributePayload
The attribute payload, a JSON string containing up to three key-value pairs (for example, {\"attributes\":{\"string1\":\"string2\"}}).

CACertificate
A CA certificate.

CACertificateDescription
Describes a CA certificate.

CancelCertificateTransferRequest
The input for the CancelCertificateTransfer operation.

Certificate
Information about a certificate.

CertificateDescription
Describes a certificate.

CloudwatchAlarmAction
Describes an action that updates a CloudWatch alarm.

CloudwatchMetricAction
Describes an action that captures a CloudWatch metric.

CreateCertificateFromCsrRequest
The input for the CreateCertificateFromCsr operation.

CreateCertificateFromCsrResponse
The output from the CreateCertificateFromCsr operation.

CreateKeysAndCertificateRequest
The input for the CreateKeysAndCertificate operation.

CreateKeysAndCertificateResponse
The output of the CreateKeysAndCertificate operation.

CreatePolicyRequest
The input for the CreatePolicy operation.

CreatePolicyResponse
The output from the CreatePolicy operation.

CreatePolicyVersionRequest
The input for the CreatePolicyVersion operation.

CreatePolicyVersionResponse
The output of the CreatePolicyVersion operation.

CreateThingRequest
The input for the CreateThing operation.

CreateThingResponse
The output of the CreateThing operation.

CreateTopicRuleRequest
The input for the CreateTopicRule operation.

DeleteCACertificateRequest
Input for the DeleteCACertificate operation.

DeleteCACertificateResponse
The output for the DeleteCACertificate operation.

DeleteCertificateRequest
The input for the DeleteCertificate operation.

DeletePolicyRequest
The input for the DeletePolicy operation.

DeletePolicyVersionRequest
The input for the DeletePolicyVersion operation.

DeleteRegistrationCodeRequest
The input for the DeleteRegistrationCode operation.

DeleteRegistrationCodeResponse
The output for the DeleteRegistrationCode operation.

DeleteThingRequest
The input for the DeleteThing operation.

DeleteThingResponse
The output of the DeleteThing operation.

DeleteTopicRuleRequest
The input for the DeleteTopicRule operation.

DescribeCACertificateRequest
The input for the DescribeCACertificate operation.

DescribeCACertificateResponse
The output from the DescribeCACertificate operation.

DescribeCertificateRequest
The input for the DescribeCertificate operation.

DescribeCertificateResponse
The output of the DescribeCertificate operation.

DescribeEndpointRequest
The input for the DescribeEndpoint operation.

DescribeEndpointResponse
The output from the DescribeEndpoint operation.

DescribeThingRequest
The input for the DescribeThing operation.

DescribeThingResponse
The output from the DescribeThing operation.

DetachPrincipalPolicyRequest
The input for the DetachPrincipalPolicy operation.

DetachThingPrincipalRequest
The input for the DetachThingPrincipal operation.

DetachThingPrincipalResponse
The output from the DetachThingPrincipal operation.

DisableTopicRuleRequest
The input for the DisableTopicRuleRequest operation.

DynamoDBAction
Describes an action to write to a DynamoDB table.

The tableName, hashKeyField, and rangeKeyField values must match the values used when you created the table.

The hashKeyValue and rangeKeyvalue fields use a substitution template syntax. These templates provide data at runtime. The syntax is as follows: ${sql-expression}.

You can specify any valid expression in a WHERE or SELECT clause, including JSON properties, comparisons, calculations, and functions. For example, the following field uses the third level of the topic:

"hashKeyValue": "${topic(3)}"

The following field uses the timestamp:

"rangeKeyValue": "${timestamp()}"

ElasticsearchAction
Describes an action that writes data to an Amazon Elasticsearch Service; domain.

EnableTopicRuleRequest
The input for the EnableTopicRuleRequest operation.

FirehoseAction
Describes an action that writes data to an Amazon Kinesis Firehose stream.

GetLoggingOptionsRequest
The input for the GetLoggingOptions operation.

GetLoggingOptionsResponse
The output from the GetLoggingOptions operation.

GetPolicyRequest
The input for the GetPolicy operation.

GetPolicyResponse
The output from the GetPolicy operation.

GetPolicyVersionRequest
The input for the GetPolicyVersion operation.

GetPolicyVersionResponse
The output from the GetPolicyVersion operation.

GetRegistrationCodeRequest
The input to the GetRegistrationCode operation.

GetRegistrationCodeResponse
The output from the GetRegistrationCode operation.

GetTopicRuleRequest
The input for the GetTopicRule operation.

GetTopicRuleResponse
The output from the GetTopicRule operation.

IotClient
A client for the AWS IoT API.

KeyPair
Describes a key pair.

KinesisAction
Describes an action to write data to an Amazon Kinesis stream.

LambdaAction
Describes an action to invoke a Lambda function.

ListCACertificatesRequest
Input for the ListCACertificates operation.

ListCACertificatesResponse
The output from the ListCACertificates operation.

ListCertificatesByCARequest
The input to the ListCertificatesByCA operation.

ListCertificatesByCAResponse
The output of the ListCertificatesByCA operation.

ListCertificatesRequest
The input for the ListCertificates operation.

ListCertificatesResponse
The output of the ListCertificates operation.

ListPoliciesRequest
The input for the ListPolicies operation.

ListPoliciesResponse
The output from the ListPolicies operation.

ListPolicyVersionsRequest
The input for the ListPolicyVersions operation.

ListPolicyVersionsResponse
The output from the ListPolicyVersions operation.

ListPrincipalPoliciesRequest
The input for the ListPrincipalPolicies operation.

ListPrincipalPoliciesResponse
The output from the ListPrincipalPolicies operation.

ListPrincipalThingsRequest
The input for the ListPrincipalThings operation.

ListPrincipalThingsResponse
The output from the ListPrincipalThings operation.

ListThingPrincipalsRequest
The input for the ListThingPrincipal operation.

ListThingPrincipalsResponse
The output from the ListThingPrincipals operation.

ListThingsRequest
The input for the ListThings operation.

ListThingsResponse
The output from the ListThings operation.

ListTopicRulesRequest
The input for the ListTopicRules operation.

ListTopicRulesResponse
The output from the ListTopicRules operation.

LoggingOptionsPayload
Describes the logging options payload.

Policy
Describes an AWS IoT policy.

PolicyVersion
Describes a policy version.

RegisterCACertificateRequest
The input to the RegisterCACertificate operation.

RegisterCACertificateResponse
The output from the RegisterCACertificateResponse operation.

RegisterCertificateRequest
The input to the RegisterCertificate operation.

RegisterCertificateResponse
The output from the RegisterCertificate operation.

RejectCertificateTransferRequest
The input for the RejectCertificateTransfer operation.

ReplaceTopicRuleRequest
The input for the ReplaceTopicRule operation.

RepublishAction
Describes an action to republish to another topic.

S3Action
Describes an action to write data to an Amazon S3 bucket.

SetDefaultPolicyVersionRequest
The input for the SetDefaultPolicyVersion operation.

SetLoggingOptionsRequest
The input for the SetLoggingOptions operation.

SnsAction
Describes an action to publish to an Amazon SNS topic.

SqsAction
Describes an action to publish data to an Amazon SQS queue.

ThingAttribute
Describes a thing attribute.

TopicRule
Describes a rule.

TopicRuleListItem
Describes a rule.

TopicRulePayload
Describes a rule.

TransferCertificateRequest
The input for the TransferCertificate operation.

TransferCertificateResponse
The output from the TransferCertificate operation.

TransferData
Data used to transfer a certificate to an AWS account.

UpdateCACertificateRequest
The input to the UpdateCACertificate operation.

UpdateCertificateRequest
The input for the UpdateCertificate operation.

UpdateThingRequest
The input for the UpdateThing operation.

UpdateThingResponse
The output from the UpdateThing operation.

Enums

AcceptCertificateTransferError

AttachPrincipalPolicyError

AttachThingPrincipalError

CancelCertificateTransferError

CreateCertificateFromCsrError

CreateKeysAndCertificateError

CreatePolicyError

CreatePolicyVersionError

CreateThingError

CreateTopicRuleError

DeleteCACertificateError

DeleteCertificateError

DeletePolicyError

DeletePolicyVersionError

DeleteRegistrationCodeError

DeleteThingError

DeleteTopicRuleError

DescribeCACertificateError

DescribeCertificateError

DescribeEndpointError

DescribeThingError

DetachPrincipalPolicyError

DetachThingPrincipalError

DisableTopicRuleError

EnableTopicRuleError

GetLoggingOptionsError

GetPolicyError

GetPolicyVersionError

GetRegistrationCodeError

GetTopicRuleError

ListCACertificatesError

ListCertificatesByCAError

ListCertificatesError

ListPoliciesError

ListPolicyVersionsError

ListPrincipalPoliciesError

ListPrincipalThingsError

ListThingPrincipalsError

ListThingsError

ListTopicRulesError

RegisterCACertificateError

RegisterCertificateError

RejectCertificateTransferError

ReplaceTopicRuleError

SetDefaultPolicyVersionError

SetLoggingOptionsError

TransferCertificateError

UpdateCACertificateError

UpdateCertificateError

UpdateThingError

Type Definitions

ActionList

AlarmName

AscendingOrder

AttributeName

AttributeValue

Attributes

AwsAccountId

AwsArn

AwsIotSqlVersion

BucketName

CACertificateStatus

CACertificates

CertificateArn

CertificateId

CertificatePem

CertificateSigningRequest

CertificateStatus

Certificates

ClientId

CreatedAtDate

DateType

DeliveryStreamName

Description

ElasticsearchEndpoint

ElasticsearchId

ElasticsearchIndex

ElasticsearchType

EndpointAddress

ErrorMessage

FunctionArn

HashKeyField

HashKeyValue

IsDefaultVersion

IsDisabled

Key

LogLevel

Marker

MaxResults

Message

MessageFormat

MetricName

MetricNamespace

MetricTimestamp

MetricUnit

MetricValue

NextToken

PageSize

PartitionKey

PayloadField

Policies

PolicyArn

PolicyDocument

PolicyName

PolicyVersionId

PolicyVersions

Principal

PrincipalArn

Principals

PrivateKey

PublicKey

QueueUrl

RangeKeyField

RangeKeyValue

RegistrationCode

RuleArn

RuleName

SQL

SetAsActive

SetAsDefault

StateReason

StateValue

StreamName

TableName

ThingArn

ThingAttributeList

ThingName

ThingNameList

Topic

TopicPattern

TopicRuleList

UseBase64