Struct rusoto::iot::DynamoDBAction

pub struct DynamoDBAction {
    pub hash_key_field: HashKeyField,
    pub hash_key_value: HashKeyValue,
    pub payload_field: Option<PayloadField>,
    pub range_key_field: RangeKeyField,
    pub range_key_value: RangeKeyValue,
    pub role_arn: AwsArn,
    pub table_name: TableName,
}

Describes an action to write to a DynamoDB table.

The tableName, hashKeyField, and rangeKeyField values must match the values used when you created the table.

The hashKeyValue and rangeKeyvalue fields use a substitution template syntax. These templates provide data at runtime. The syntax is as follows: ${sql-expression}.

You can specify any valid expression in a WHERE or SELECT clause, including JSON properties, comparisons, calculations, and functions. For example, the following field uses the third level of the topic:

"hashKeyValue": "${topic(3)}"

The following field uses the timestamp:

"rangeKeyValue": "${timestamp()}"

Fields
hash_key_field: HashKeyField

The hash key name.

hash_key_value: HashKeyValue

The hash key value.

payload_field: Option<PayloadField>

The action payload. This name can be customized.

range_key_field: RangeKeyField

The range key name.

range_key_value: RangeKeyValue

The range key value.

role_arn: AwsArn

The ARN of the IAM role that grants access to the DynamoDB table.

table_name: TableName

The name of the DynamoDB table.