Struct rusoto::kinesis::StreamDescription

pub struct StreamDescription {
    pub enhanced_monitoring: EnhancedMonitoringList,
    pub has_more_shards: BooleanObject,
    pub retention_period_hours: RetentionPeriodHours,
    pub shards: ShardList,
    pub stream_a_r_n: StreamARN,
    pub stream_name: StreamName,
    pub stream_status: StreamStatus,
}

Represents the output for DescribeStream.

Fields
enhanced_monitoring: EnhancedMonitoringList

Represents the current enhanced monitoring settings of the stream.

has_more_shards: BooleanObject

If set to true, more shards in the stream are available to describe.

retention_period_hours: RetentionPeriodHours

The current retention period, in hours.

shards: ShardList

The shards that comprise the stream.

stream_a_r_n: StreamARN

The Amazon Resource Name (ARN) for the stream being described.

stream_name: StreamName

The name of the stream being described.

stream_status: StreamStatus

The current status of the stream being described. The stream status is one of the following states:

CREATING - The stream is being created. Amazon Kinesis immediately returns and sets StreamStatus to CREATING.

DELETING - The stream is being deleted. The specified stream is in the DELETING state until Amazon Kinesis completes the deletion.

ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream.

UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.