Module rusoto::kms

KMS bindings for Rust

Check Helper for convenience functions.

Structs

AliasListEntry
Contains information about an alias.

AlreadyExistsException
The request was rejected because it attempted to create a resource that already exists.

CancelKeyDeletionRequest

CancelKeyDeletionResponse

CreateAliasRequest

CreateGrantRequest

CreateGrantResponse

CreateKeyRequest

CreateKeyResponse

DecryptRequest

DecryptResponse

DeleteAliasRequest

DependencyTimeoutException
The system timed out while trying to fulfill the request. The request can be retried.

DescribeKeyRequest

DescribeKeyResponse

DisableKeyRequest

DisableKeyRotationRequest

DisabledException
The request was rejected because the specified key was marked as disabled.

EnableKeyRequest

EnableKeyRotationRequest

EncryptRequest

EncryptResponse

GenerateDataKeyRequest

GenerateDataKeyResponse

GenerateDataKeyWithoutPlaintextRequest

GenerateDataKeyWithoutPlaintextResponse

GenerateRandomRequest

GenerateRandomResponse

GetKeyPolicyRequest

GetKeyPolicyResponse

GetKeyRotationStatusRequest

GetKeyRotationStatusResponse

GrantConstraints
A structure for specifying the conditions under which the operations permitted by the grant are allowed. You can use this structure to allow the operations permitted by the grant only when a specified encryption context is present. For more information about encryption context, see Encryption Context in the AWS Key Management Service Developer Guide.

GrantListEntry
Contains information about an entry in a list of grants.

InvalidAliasNameException
The request was rejected because the specified alias name is not valid.

InvalidArnException
The request was rejected because a specified ARN was not valid.

InvalidCiphertextException
The request was rejected because the specified ciphertext has been corrupted or is otherwise invalid.

InvalidGrantIdException
The request was rejected because the specified GrantId is not valid.

InvalidGrantTokenException
The request was rejected because a grant token provided as part of the request is invalid.

InvalidKeyUsageException
The request was rejected because the specified KeySpec parameter is not valid. The currently supported value is ENCRYPT/DECRYPT.

InvalidMarkerException
The request was rejected because the marker that specifies where pagination should next begin is not valid.

KMSClient

KMSError

KMSHelper

KMSInternalException
The request was rejected because an internal exception occurred. The request can be retried.

KMSInvalidStateException
The request was rejected because the state of the specified resource is not valid for this request. For more information about how key state affects the use of a customer master key (CMK), go to How Key State Affects the Use of a Customer Master Key in the AWS Key Management Service Developer Guide.

KeyListEntry
Contains information about each entry in the key list.

KeyMetadata
Contains metadata about a customer master key (CMK). This data type is used as a response element for the CreateKey and DescribeKey operations.

KeyUnavailableException
The request was rejected because the key was not available. The request can be retried.

LimitExceededException
The request was rejected because a limit was exceeded. For more information, see Limits in the AWS Key Management Service Developer Guide.

ListAliasesRequest

ListAliasesResponse

ListGrantsRequest

ListGrantsResponse

ListKeyPoliciesRequest

ListKeyPoliciesResponse

ListKeysRequest

ListKeysResponse

ListRetirableGrantsRequest

MalformedPolicyDocumentException
The request was rejected because the specified policy is not syntactically or semantically correct.

NotFoundException
The request was rejected because the specified entity or resource could not be found.

PutKeyPolicyRequest

ReEncryptRequest

ReEncryptResponse

RetireGrantRequest

RevokeGrantRequest

ScheduleKeyDeletionRequest

ScheduleKeyDeletionResponse

UnsupportedOperationException
The request was rejected because a specified parameter is not supported.

UpdateAliasRequest

UpdateKeyDescriptionRequest

Type Definitions

AWSAccountIdType

AliasList

AliasNameType

ArnType

BooleanType

CiphertextType

DataKeySpec

DateType

DescriptionType

EncryptionContextKey

EncryptionContextType

EncryptionContextValue

ErrorMessageType

GrantIdType

GrantList

GrantNameType

GrantOperation

GrantOperationList

GrantTokenList

GrantTokenType

KeyIdType

KeyList

KeyState

KeyUsageType

LimitType

MarkerType

NumberOfBytesType

PendingWindowInDaysType

PlaintextType

PolicyNameList

PolicyNameType

PolicyType

PrincipalIdType

Result