Enum rusoto::kms::ListAliasesError

pub enum ListAliasesError {
    DependencyTimeout(String),
    InvalidMarker(String),
    KMSInternal(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Variants
DependencyTimeout(String)

The system timed out while trying to fulfill the request. The request can be retried.

InvalidMarker(String)

The request was rejected because the marker that specifies where pagination should next begin is not valid.

KMSInternal(String)

The request was rejected because an internal exception occurred. The request can be retried.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl ListAliasesError

fn from_body(body: &str) -> ListAliasesError