Struct rusoto::kms::CreateKeyRequest

pub struct CreateKeyRequest {
    pub bypass_policy_lockout_safety_check: Option<BooleanType>,
    pub description: Option<DescriptionType>,
    pub key_usage: Option<KeyUsageType>,
    pub policy: Option<PolicyType>,
}

Fields
bypass_policy_lockout_safety_check: Option<BooleanType>

A flag to indicate whether to bypass the key policy lockout safety check.

Setting this value to true increases the likelihood that the CMK becomes unmanageable. Do not set this value to true indiscriminately.

For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide.

Use this parameter only when you include a policy in the request and you intend to prevent the principal making the request from making a subsequent PutKeyPolicy request on the CMK.

The default value is false.

description: Option<DescriptionType>

A description of the CMK.

Use a description that helps you decide whether the CMK is appropriate for a task.

key_usage: Option<KeyUsageType>

The intended use of the CMK.

You can use CMKs only for symmetric encryption and decryption.

policy: Option<PolicyType>

The key policy to attach to the CMK.

If you specify a key policy, it must meet the following criteria:

It must allow the principal making the CreateKey request to make a subsequent PutKeyPolicy request on the CMK. This reduces the likelihood that the CMK becomes unmanageable. For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide.

The principal(s) specified in the key policy must exist and be visible to AWS KMS. When you create a new AWS principal (for example, an IAM user or role), you might need to enforce a delay before specifying the new principal in a key policy because the new principal might not immediately be visible to AWS KMS. For more information, see Changes that I make are not always immediately visible in the IAM User Guide.

If you do not specify a policy, AWS KMS attaches a default key policy to the CMK. For more information, see Default Key Policy in the AWS Key Management Service Developer Guide.

The policy size limit is 32 KiB (32768 bytes).