Struct rusoto::kms::GenerateDataKeyResponse

pub struct GenerateDataKeyResponse {
    pub ciphertext_blob: Option<CiphertextType>,
    pub key_id: Option<KeyIdType>,
    pub plaintext: Option<PlaintextType>,
}

Fields
ciphertext_blob: Option<CiphertextType>

Ciphertext that contains the encrypted data key. You must store the blob and enough information to reconstruct the encryption context so that the data encrypted by using the key can later be decrypted. You must provide both the ciphertext blob and the encryption context to the Decrypt API to recover the plaintext data key and decrypt the object.

If you are using the CLI, the value is Base64 encoded. Otherwise, it is not encoded.

key_id: Option<KeyIdType>

System generated unique identifier of the key to be used to decrypt the encrypted copy of the data key.

plaintext: Option<PlaintextType>

Plaintext that contains the data key. Use this for encryption and decryption and then remove it from memory as soon as possible.