Struct rusoto::kms::GrantListEntry

pub struct GrantListEntry {
    pub Operations: Option<GrantOperationList>,
    pub KeyId: Option<KeyIdType>,
    pub Name: Option<GrantNameType>,
    pub RetiringPrincipal: Option<PrincipalIdType>,
    pub GranteePrincipal: Option<PrincipalIdType>,
    pub GrantId: Option<GrantIdType>,
    pub IssuingAccount: Option<PrincipalIdType>,
    pub CreationDate: Option<DateType>,
    pub Constraints: Option<GrantConstraints>,
}

Contains information about an entry in a list of grants.

Fields

Operations

The list of operations permitted by the grant.

KeyId

The unique identifier for the customer master key (CMK) to which the grant applies.

Name

The friendly name that identifies the grant. If a name was provided in the CreateGrant request, that name is returned. Otherwise this value is null.

RetiringPrincipal

The principal that can retire the grant.

GranteePrincipal

The principal that receives the grant's permissions.

GrantId

The unique identifier for the grant.

IssuingAccount

The AWS account under which the grant was issued.

CreationDate

The date and time when the grant was created.

Constraints

The conditions under which the grant's operations are allowed.