Struct rusoto::kms::KeyMetadata

pub struct KeyMetadata {
    pub KeyId: KeyIdType,
    pub Description: Option<DescriptionType>,
    pub DeletionDate: Option<DateType>,
    pub Enabled: Option<BooleanType>,
    pub KeyUsage: Option<KeyUsageType>,
    pub KeyState: Option<KeyState>,
    pub CreationDate: Option<DateType>,
    pub Arn: Option<ArnType>,
    pub AWSAccountId: Option<AWSAccountIdType>,
}

Contains metadata about a customer master key (CMK). This data type is used as a response element for the CreateKey and DescribeKey operations.

Fields

KeyId

The globally unique identifier for the key.

Description

The friendly description of the key.

DeletionDate

The date and time after which AWS KMS deletes the customer master key (CMK). This value is present only when KeyState is PendingDeletion, otherwise this value is null.

Enabled

Specifies whether the key is enabled. When KeyState is Enabled this value is true, otherwise it is false.

KeyUsage

The cryptographic operations for which you can use the key. Currently the only allowed value is ENCRYPT_DECRYPT, which means you can use the key for the Encrypt and Decrypt operations.

KeyState

The state of the customer master key (CMK). For more information about how key state affects the use of a CMK, go to How Key State Affects the Use of a Customer Master Key in the AWS Key Management Service Developer Guide.

CreationDate

The date and time when the key was created.

Arn

The Amazon Resource Name (ARN) of the key. For examples, see AWS Key Management Service (AWS KMS) in the Example ARNs section of the AWS General Reference.

AWSAccountId

The twelve-digit account ID of the AWS account that owns the key.