Struct rusoto::kms::KmsClient

pub struct KmsClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the KMS API.

Methods

impl<P> KmsClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> KmsClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn cancel_key_deletion(&self, input: &CancelKeyDeletionRequest) -> Result<CancelKeyDeletionResponse, CancelKeyDeletionError>

Cancels the deletion of a customer master key (CMK). When this operation is successful, the CMK is set to the Disabled state. To enable a CMK, use EnableKey.

For more information about scheduling and canceling deletion of a CMK, see Deleting Customer Master Keys in the AWS Key Management Service Developer Guide.

fn create_alias(&self, input: &CreateAliasRequest) -> Result<(), CreateAliasError>

Creates a display name for a customer master key. An alias can be used to identify a key and should be unique. The console enforces a one-to-one mapping between the alias and a key. An alias name can contain only alphanumeric characters, forward slashes (/), underscores (_), and dashes (-). An alias must start with the word "alias" followed by a forward slash (alias/). An alias that begins with "aws" after the forward slash (alias/aws...) is reserved by Amazon Web Services (AWS).

The alias and the key it is mapped to must be in the same AWS account and the same region.

To map an alias to a different key, call UpdateAlias.

fn create_grant(&self, input: &CreateGrantRequest) -> Result<CreateGrantResponse, CreateGrantError>

Adds a grant to a key to specify who can use the key and under what conditions. Grants are alternate permission mechanisms to key policies.

For more information about grants, see Grants in the AWS Key Management Service Developer Guide.

fn create_key(&self, input: &CreateKeyRequest) -> Result<CreateKeyResponse, CreateKeyError>

Creates a customer master key (CMK).

You can use a CMK to encrypt small amounts of data (4 KiB or less) directly, but CMKs are more commonly used to encrypt data encryption keys (DEKs), which are used to encrypt raw data. For more information about DEKs and the difference between CMKs and DEKs, see the following:

The GenerateDataKey operation

AWS Key Management Service Concepts in the AWS Key Management Service Developer Guide

fn decrypt(&self, input: &DecryptRequest) -> Result<DecryptResponse, DecryptError>

Decrypts ciphertext. Ciphertext is plaintext that has been previously encrypted by using any of the following functions:

GenerateDataKey

GenerateDataKeyWithoutPlaintext

Encrypt

Note that if a caller has been granted access permissions to all keys (through, for example, IAM user policies that grant Decrypt permission on all resources), then ciphertext encrypted by using keys in other accounts where the key grants access to the caller can be decrypted. To remedy this, we recommend that you do not grant Decrypt access in an IAM user policy. Instead grant Decrypt access only in key policies. If you must grant Decrypt access in an IAM user policy, you should scope the resource to specific keys or to specific trusted accounts.

fn delete_alias(&self, input: &DeleteAliasRequest) -> Result<(), DeleteAliasError>

Deletes the specified alias. To map an alias to a different key, call UpdateAlias.

fn describe_key(&self, input: &DescribeKeyRequest) -> Result<DescribeKeyResponse, DescribeKeyError>

Provides detailed information about the specified customer master key.

fn disable_key(&self, input: &DisableKeyRequest) -> Result<(), DisableKeyError>

Sets the state of a customer master key (CMK) to disabled, thereby preventing its use for cryptographic operations. For more information about how key state affects the use of a CMK, see How Key State Affects the Use of a Customer Master Key in the AWS Key Management Service Developer Guide.

fn disable_key_rotation(&self, input: &DisableKeyRotationRequest) -> Result<(), DisableKeyRotationError>

Disables rotation of the specified key.

fn enable_key(&self, input: &EnableKeyRequest) -> Result<(), EnableKeyError>

Marks a key as enabled, thereby permitting its use.

fn enable_key_rotation(&self, input: &EnableKeyRotationRequest) -> Result<(), EnableKeyRotationError>

Enables rotation of the specified customer master key.

fn encrypt(&self, input: &EncryptRequest) -> Result<EncryptResponse, EncryptError>

Encrypts plaintext into ciphertext by using a customer master key. The Encrypt function has two primary use cases:

You can encrypt up to 4 KB of arbitrary data such as an RSA key, a database password, or other sensitive customer information.

If you are moving encrypted data from one region to another, you can use this API to encrypt in the new region the plaintext data key that was used to encrypt the data in the original region. This provides you with an encrypted copy of the data key that can be decrypted in the new region and used there to decrypt the encrypted data.

Unless you are moving encrypted data from one region to another, you don't use this function to encrypt a generated data key within a region. You retrieve data keys already encrypted by calling the GenerateDataKey or GenerateDataKeyWithoutPlaintext function. Data keys don't need to be encrypted again by calling Encrypt.

If you want to encrypt data locally in your application, you can use the GenerateDataKey function to return a plaintext data encryption key and a copy of the key encrypted under the customer master key (CMK) of your choosing.

fn generate_data_key(&self, input: &GenerateDataKeyRequest) -> Result<GenerateDataKeyResponse, GenerateDataKeyError>

Generates a data key that you can use in your application to locally encrypt data. This call returns a plaintext version of the key in the Plaintext field of the response object and an encrypted copy of the key in the CiphertextBlob field. The key is encrypted by using the master key specified by the KeyId field. To decrypt the encrypted key, pass it to the Decrypt API.

We recommend that you use the following pattern to locally encrypt data: call the GenerateDataKey API, use the key returned in the Plaintext response field to locally encrypt data, and then erase the plaintext data key from memory. Store the encrypted data key (contained in the CiphertextBlob field) alongside of the locally encrypted data.

You should not call the Encrypt function to re-encrypt your data keys within a region. GenerateDataKey always returns the data key encrypted and tied to the customer master key that will be used to decrypt it. There is no need to decrypt it twice.

If you decide to use the optional EncryptionContext parameter, you must also store the context in full or at least store enough information along with the encrypted data to be able to reconstruct the context when submitting the ciphertext to the Decrypt API. It is a good practice to choose a context that you can reconstruct on the fly to better secure the ciphertext. For more information about how this parameter is used, see Encryption Context.

To decrypt data, pass the encrypted data key to the Decrypt API. Decrypt uses the associated master key to decrypt the encrypted data key and returns it as plaintext. Use the plaintext data key to locally decrypt your data and then erase the key from memory. You must specify the encryption context, if any, that you specified when you generated the key. The encryption context is logged by CloudTrail, and you can use this log to help track the use of particular data.

fn generate_data_key_without_plaintext(&self, input: &GenerateDataKeyWithoutPlaintextRequest) -> Result<GenerateDataKeyWithoutPlaintextResponse, GenerateDataKeyWithoutPlaintextError>

Returns a data key encrypted by a customer master key without the plaintext copy of that key. Otherwise, this API functions exactly like GenerateDataKey. You can use this API to, for example, satisfy an audit requirement that an encrypted key be made available without exposing the plaintext copy of that key.

fn generate_random(&self, input: &GenerateRandomRequest) -> Result<GenerateRandomResponse, GenerateRandomError>

Generates an unpredictable byte string.

fn get_key_policy(&self, input: &GetKeyPolicyRequest) -> Result<GetKeyPolicyResponse, GetKeyPolicyError>

Retrieves a policy attached to the specified key.

fn get_key_rotation_status(&self, input: &GetKeyRotationStatusRequest) -> Result<GetKeyRotationStatusResponse, GetKeyRotationStatusError>

Retrieves a Boolean value that indicates whether key rotation is enabled for the specified key.

fn list_aliases(&self, input: &ListAliasesRequest) -> Result<ListAliasesResponse, ListAliasesError>

Lists all of the key aliases in the account.

fn list_grants(&self, input: &ListGrantsRequest) -> Result<ListGrantsResponse, ListGrantsError>

List the grants for a specified key.

fn list_key_policies(&self, input: &ListKeyPoliciesRequest) -> Result<ListKeyPoliciesResponse, ListKeyPoliciesError>

Retrieves a list of policies attached to a key.

fn list_keys(&self, input: &ListKeysRequest) -> Result<ListKeysResponse, ListKeysError>

Lists the customer master keys.

fn list_retirable_grants(&self, input: &ListRetirableGrantsRequest) -> Result<ListGrantsResponse, ListRetirableGrantsError>

Returns a list of all grants for which the grant's RetiringPrincipal matches the one specified.

A typical use is to list all grants that you are able to retire. To retire a grant, use RetireGrant.

fn put_key_policy(&self, input: &PutKeyPolicyRequest) -> Result<(), PutKeyPolicyError>

Attaches a key policy to the specified customer master key (CMK).

For more information about key policies, see Key Policies in the AWS Key Management Service Developer Guide.

fn re_encrypt(&self, input: &ReEncryptRequest) -> Result<ReEncryptResponse, ReEncryptError>

Encrypts data on the server side with a new customer master key without exposing the plaintext of the data on the client side. The data is first decrypted and then encrypted. This operation can also be used to change the encryption context of a ciphertext.

Unlike other actions, ReEncrypt is authorized twice - once as ReEncryptFrom on the source key and once as ReEncryptTo on the destination key. We therefore recommend that you include the "action":"kms:ReEncrypt*" statement in your key policies to permit re-encryption from or to the key. The statement is included automatically when you authorize use of the key through the console but must be included manually when you set a policy by using the PutKeyPolicy function.

fn retire_grant(&self, input: &RetireGrantRequest) -> Result<(), RetireGrantError>

Retires a grant. You can retire a grant when you're done using it to clean up. You should revoke a grant when you intend to actively deny operations that depend on it. The following are permitted to call this API:

The account that created the grant

The RetiringPrincipal, if present

The GranteePrincipal, if RetireGrant is a grantee operation

The grant to retire must be identified by its grant token or by a combination of the key ARN and the grant ID. A grant token is a unique variable-length base64-encoded string. A grant ID is a 64 character unique identifier of a grant. Both are returned by the CreateGrant function.

fn revoke_grant(&self, input: &RevokeGrantRequest) -> Result<(), RevokeGrantError>

Revokes a grant. You can revoke a grant to actively deny operations that depend on it.

fn schedule_key_deletion(&self, input: &ScheduleKeyDeletionRequest) -> Result<ScheduleKeyDeletionResponse, ScheduleKeyDeletionError>

Schedules the deletion of a customer master key (CMK). You may provide a waiting period, specified in days, before deletion occurs. If you do not provide a waiting period, the default period of 30 days is used. When this operation is successful, the state of the CMK changes to PendingDeletion. Before the waiting period ends, you can use CancelKeyDeletion to cancel the deletion of the CMK. After the waiting period ends, AWS KMS deletes the CMK and all AWS KMS data associated with it, including all aliases that point to it.

Deleting a CMK is a destructive and potentially dangerous operation. When a CMK is deleted, all data that was encrypted under the CMK is rendered unrecoverable. To restrict the use of a CMK without deleting it, use DisableKey.

For more information about scheduling a CMK for deletion, see Deleting Customer Master Keys in the AWS Key Management Service Developer Guide.

fn update_alias(&self, input: &UpdateAliasRequest) -> Result<(), UpdateAliasError>

Updates an alias to map it to a different key.

An alias is not a property of a key. Therefore, an alias can be mapped to and unmapped from an existing key without changing the properties of the key.

An alias name can contain only alphanumeric characters, forward slashes (/), underscores (_), and dashes (-). An alias must start with the word "alias" followed by a forward slash (alias/). An alias that begins with "aws" after the forward slash (alias/aws...) is reserved by Amazon Web Services (AWS).

The alias and the key it is mapped to must be in the same AWS account and the same region.

fn update_key_description(&self, input: &UpdateKeyDescriptionRequest) -> Result<(), UpdateKeyDescriptionError>

Updates the description of a key.