Struct rusoto::kms::ListGrantsRequest

pub struct ListGrantsRequest {
    pub key_id: KeyIdType,
    pub limit: Option<LimitType>,
    pub marker: Option<MarkerType>,
}

Fields
key_id: KeyIdType

A unique identifier for the customer master key. This value can be a globally unique identifier or the fully specified ARN to a key.

Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

limit: Option<LimitType>

When paginating results, specify the maximum number of items to return in the response. If additional items exist beyond the number you specify, the Truncated element in the response is set to true.

This value is optional. If you include a value, it must be between 1 and 100, inclusive. If you do not include a value, it defaults to 50.

marker: Option<MarkerType>

Use this parameter only when paginating results and only in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the response you just received.