Struct rusoto::kms::ReEncryptRequest

pub struct ReEncryptRequest {
    pub ciphertext_blob: CiphertextType,
    pub destination_encryption_context: Option<EncryptionContextType>,
    pub destination_key_id: KeyIdType,
    pub grant_tokens: Option<GrantTokenList>,
    pub source_encryption_context: Option<EncryptionContextType>,
}

Fields
ciphertext_blob: CiphertextType

Ciphertext of the data to re-encrypt.

destination_encryption_context: Option<EncryptionContextType>

Encryption context to be used when the data is re-encrypted.

destination_key_id: KeyIdType

A unique identifier for the customer master key used to re-encrypt the data. This value can be a globally unique identifier, a fully specified ARN to either an alias or a key, or an alias name prefixed by "alias/".

Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Alias ARN Example - arn:aws:kms:us-east-1:123456789012:alias/MyAliasName

Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

Alias Name Example - alias/MyAliasName

grant_tokens: Option<GrantTokenList>

A list of grant tokens.

For more information, see Grant Tokens in the AWS Key Management Service Developer Guide.

source_encryption_context: Option<EncryptionContextType>

Encryption context used to encrypt and decrypt the data specified in the CiphertextBlob parameter.