Struct rusoto::lambda::FunctionConfiguration

pub struct FunctionConfiguration {
    pub code_sha_256: Option<String>,
    pub code_size: Option<Long>,
    pub dead_letter_config: Option<DeadLetterConfig>,
    pub description: Option<Description>,
    pub environment: Option<EnvironmentResponse>,
    pub function_arn: Option<FunctionArn>,
    pub function_name: Option<FunctionName>,
    pub handler: Option<Handler>,
    pub kms_key_arn: Option<KMSKeyArn>,
    pub last_modified: Option<Timestamp>,
    pub memory_size: Option<MemorySize>,
    pub role: Option<RoleArn>,
    pub runtime: Option<Runtime>,
    pub timeout: Option<Timeout>,
    pub version: Option<Version>,
    pub vpc_config: Option<VpcConfigResponse>,
}

A complex type that describes function metadata.

Fields
code_sha_256: Option<String>

It is the SHA256 hash of your function deployment package.

code_size: Option<Long>

The size, in bytes, of the function .zip file you uploaded.

dead_letter_config: Option<DeadLetterConfig>

The parent object that contains the target Amazon Resource Name (ARN) of an Amazon SQS queue or Amazon SNS topic.

description: Option<Description>

The user-provided description.

environment: Option<EnvironmentResponse>

The parent object that contains your environment's configuration settings.

function_arn: Option<FunctionArn>

The Amazon Resource Name (ARN) assigned to the function.

function_name: Option<FunctionName>

The name of the function.

handler: Option<Handler>

The function Lambda calls to begin executing your function.

kms_key_arn: Option<KMSKeyArn>

The Amazon Resource Name (ARN) of the KMS key used to encrypt your function's environment variables. If empty, it means you are using the AWS Lambda default service key.

last_modified: Option<Timestamp>

The time stamp of the last time you updated the function.

memory_size: Option<MemorySize>

The memory size, in MB, you configured for the function. Must be a multiple of 64 MB.

role: Option<RoleArn>

The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources.

runtime: Option<Runtime>

The runtime environment for the Lambda function.

To use the Node.js runtime v4.3, set the value to "nodejs4.3". To use earlier runtime (v0.10.42), set the value to "nodejs".

timeout: Option<Timeout>

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

version: Option<Version>

The version of the Lambda function.

vpc_config: Option<VpcConfigResponse>

VPC configuration associated with your Lambda function.