Module rusoto::logs

Amazon CloudWatch Logs

Structs

CancelExportTaskRequest

CloudWatchLogsClient
A client for the Amazon CloudWatch Logs API.

CreateExportTaskRequest

CreateExportTaskResponse

CreateLogGroupRequest

CreateLogStreamRequest

DeleteDestinationRequest

DeleteLogGroupRequest

DeleteLogStreamRequest

DeleteMetricFilterRequest

DeleteRetentionPolicyRequest

DeleteSubscriptionFilterRequest

DescribeDestinationsRequest

DescribeDestinationsResponse

DescribeExportTasksRequest

DescribeExportTasksResponse

DescribeLogGroupsRequest

DescribeLogGroupsResponse

DescribeLogStreamsRequest

DescribeLogStreamsResponse

DescribeMetricFiltersRequest

DescribeMetricFiltersResponse

DescribeSubscriptionFiltersRequest

DescribeSubscriptionFiltersResponse

Destination
A cross account destination that is the recipient of subscription log events.

ExportTask
Represents an export task.

ExportTaskExecutionInfo
Represents the status of an export task.

ExportTaskStatus
Represents the status of an export task.

FilterLogEventsRequest

FilterLogEventsResponse

FilteredLogEvent
Represents a matched event from a FilterLogEvents request.

GetLogEventsRequest

GetLogEventsResponse

InputLogEvent
A log event is a record of some activity that was recorded by the application or resource being monitored. The log event record that CloudWatch Logs understands contains two properties: the timestamp of when the event occurred, and the raw event message.

LogGroup

LogStream
A log stream is sequence of log events from a single emitter of logs.

MetricFilter
Metric filters can be used to express how CloudWatch Logs would extract metric observations from ingested log events and transform them to metric data in a CloudWatch metric.

MetricFilterMatchRecord

MetricTransformation

OutputLogEvent

PutDestinationPolicyRequest

PutDestinationRequest

PutDestinationResponse

PutLogEventsRequest

PutLogEventsResponse

PutMetricFilterRequest

PutRetentionPolicyRequest

PutSubscriptionFilterRequest

RejectedLogEventsInfo

SearchedLogStream
An object indicating the search status of a log stream in a FilterLogEvents request.

SubscriptionFilter

TestMetricFilterRequest

TestMetricFilterResponse

Enums

CancelExportTaskError

CreateExportTaskError

CreateLogGroupError

CreateLogStreamError

DeleteDestinationError

DeleteLogGroupError

DeleteLogStreamError

DeleteMetricFilterError

DeleteRetentionPolicyError

DeleteSubscriptionFilterError

DescribeDestinationsError

DescribeExportTasksError

DescribeLogGroupsError

DescribeLogStreamsError

DescribeMetricFiltersError

DescribeSubscriptionFiltersError

FilterLogEventsError

GetLogEventsError

PutDestinationError

PutDestinationPolicyError

PutLogEventsError

PutMetricFilterError

PutRetentionPolicyError

PutSubscriptionFilterError

TestMetricFilterError

Type Definitions

AccessPolicy

Arn

Days
Specifies the number of days you want to retain log events in the specified log group. Possible values are: 1, 3, 5, 7, 14, 30, 60, 90, 120, 150, 180, 365, 400, 545, 731, 1827, 3653.

Descending

DescribeLimit
The maximum number of results to return.

DestinationArn

DestinationName

Destinations

EventId

EventMessage

EventNumber

EventsLimit
The maximum number of events to return.

ExportDestinationBucket

ExportDestinationPrefix

ExportTaskId

ExportTaskName

ExportTaskStatusCode

ExportTaskStatusMessage

ExportTasks
A list of export tasks.

ExtractedValues

FilterCount
The number of metric filters associated with the log group.

FilterName
A name for a metric or subscription filter.

FilterPattern
A symbolic description of how CloudWatch Logs should interpret the data in each log event. For example, a log event may contain timestamps, IP addresses, strings, and so on. You use the filter pattern to specify what to look for in the log event message.

FilteredLogEvents
A list of matched FilteredLogEvent objects returned from a FilterLogEvents request.

InputLogEvents
A list of log events belonging to a log stream.

InputLogStreamNames
A list of log stream names.

Interleaved

LogEventIndex

LogGroupName

LogGroups
A list of log groups.

LogStreamName

LogStreamSearchedCompletely

LogStreams
A list of log streams.

MetricFilterMatches

MetricFilters

MetricName
The name of the CloudWatch metric to which the monitored log information should be published. For example, you may publish to a metric called ErrorCount.

MetricNamespace
The destination namespace of the new CloudWatch metric.

MetricTransformations

MetricValue
What to publish to the metric. For example, if you're counting the occurrences of a particular term like "Error", the value will be "1" for each occurrence. If you're counting the bytes transferred the published value will be the value in the log event.

NextToken
A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous request. The token expires after 24 hours.

OrderBy

OutputLogEvents

RoleArn

SearchedLogStreams
A list of SearchedLogStream objects indicating the search status for log streams in a FilterLogEvents request.

SequenceToken
A string token used for making PutLogEvents requests. A sequenceToken can only be used once, and PutLogEvents requests must include the sequenceToken obtained from the response of the previous request.

StartFromHead

StoredBytes

SubscriptionFilters

TargetArn

TestEventMessages

Timestamp
A point in time expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

Token

Value