Struct rusoto::logs::CloudWatchLogsClient

pub struct CloudWatchLogsClient<P, D> where P: ProvideAwsCredentials,
        D: DispatchSignedRequest { /* fields omitted */ }

A client for the Amazon CloudWatch Logs API.

Methods

impl<P, D> CloudWatchLogsClient<P, D> where P: ProvideAwsCredentials,
D: DispatchSignedRequest

fn new(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn cancel_export_task(&self,
input: &CancelExportTaskRequest)
-> Result<(), CancelExportTaskError>

Cancels the specified export task.

The task must be in the PENDING or RUNNING state.

fn create_export_task(&self,
input: &CreateExportTaskRequest)
-> Result<CreateExportTaskResponse, CreateExportTaskError>

Creates an export task, which allows you to efficiently export data from a log group to an Amazon S3 bucket.

This is an asynchronous call. If all the required information is provided, this operation initiates an export task and responds with the ID of the task. After the task has started, you can use DescribeExportTasks to get the status of the export task. Each account can only have one active (RUNNING or PENDING) export task at a time. To cancel an export task, use CancelExportTask.

You can export logs from multiple log groups or multiple time ranges to the same S3 bucket. To separate out log data for each export task, you can specify a prefix that will be used as the Amazon S3 key prefix for all exported objects.

fn create_log_group(&self,
input: &CreateLogGroupRequest)
-> Result<(), CreateLogGroupError>

Creates a log group with the specified name.

You can create up to 5000 log groups per account.

You must use the following guidelines when naming a log group:

Log group names must be unique within a region for an AWS account.

Log group names can be between 1 and 512 characters long.

Log group names consist of the following characters: a-z, A-Z, 0-9, '_' (underscore), '-' (hyphen), '/' (forward slash), and '.' (period).

fn create_log_stream(&self,
input: &CreateLogStreamRequest)
-> Result<(), CreateLogStreamError>

Creates a log stream for the specified log group.

There is no limit on the number of log streams that you can create for a log group.

You must use the following guidelines when naming a log stream:

Log stream names must be unique within the log group.

Log stream names can be between 1 and 512 characters long.

The ':' (colon) and '*' (asterisk) characters are not allowed.

fn delete_destination(&self,
input: &DeleteDestinationRequest)
-> Result<(), DeleteDestinationError>

Deletes the specified destination, and eventually disables all the subscription filters that publish to it. This operation does not delete the physical resource encapsulated by the destination.

fn delete_log_group(&self,
input: &DeleteLogGroupRequest)
-> Result<(), DeleteLogGroupError>

Deletes the specified log group and permanently deletes all the archived log events associated with the log group.

fn delete_log_stream(&self,
input: &DeleteLogStreamRequest)
-> Result<(), DeleteLogStreamError>

Deletes the specified log stream and permanently deletes all the archived log events associated with the log stream.

fn delete_metric_filter(&self,
input: &DeleteMetricFilterRequest)
-> Result<(), DeleteMetricFilterError>

Deletes the specified metric filter.

fn delete_retention_policy(&self,
input: &DeleteRetentionPolicyRequest)
-> Result<(), DeleteRetentionPolicyError>

Deletes the specified retention policy.

Log events do not expire if they belong to log groups without a retention policy.

fn delete_subscription_filter(&self,
input: &DeleteSubscriptionFilterRequest)
-> Result<(), DeleteSubscriptionFilterError>

Deletes the specified subscription filter.

fn describe_destinations(&self,
input: &DescribeDestinationsRequest)
-> Result<DescribeDestinationsResponse, DescribeDestinationsError>

Lists all your destinations. The results are ASCII-sorted by destination name.

fn describe_export_tasks(&self,
input: &DescribeExportTasksRequest)
-> Result<DescribeExportTasksResponse, DescribeExportTasksError>

Lists the specified export tasks. You can list all your export tasks or filter the results based on task ID or task status.

fn describe_log_groups(&self,
input: &DescribeLogGroupsRequest)
-> Result<DescribeLogGroupsResponse, DescribeLogGroupsError>

Lists the specified log groups. You can list all your log groups or filter the results by prefix. The results are ASCII-sorted by log group name.

fn describe_log_streams(&self,
input: &DescribeLogStreamsRequest)
-> Result<DescribeLogStreamsResponse, DescribeLogStreamsError>

Lists the log streams for the specified log group. You can list all the log streams or filter the results by prefix. You can also control how the results are ordered.

This operation has a limit of five transactions per second, after which transactions are throttled.

fn describe_metric_filters(&self,
input: &DescribeMetricFiltersRequest)
-> Result<DescribeMetricFiltersResponse, DescribeMetricFiltersError>

Lists the specified metric filters. You can list all the metric filters or filter the results by log name, prefix, metric name, and metric namespace. The results are ASCII-sorted by filter name.

fn describe_subscription_filters(&self,
input: &DescribeSubscriptionFiltersRequest)
-> Result<DescribeSubscriptionFiltersResponse, DescribeSubscriptionFiltersError>

Lists the subscription filters for the specified log group. You can list all the subscription filters or filter the results by prefix. The results are ASCII-sorted by filter name.

fn filter_log_events(&self,
input: &FilterLogEventsRequest)
-> Result<FilterLogEventsResponse, FilterLogEventsError>

Lists log events from the specified log group. You can list all the log events or filter the results using a filter pattern, a time range, and the name of the log stream.

By default, this operation returns as many log events as can fit in 1MB (up to 10,000 log events), or all the events found within the time range that you specify. If the results include a token, then there are more log events available, and you can get additional results by specifying the token in a subsequent call.

fn get_log_events(&self,
input: &GetLogEventsRequest)
-> Result<GetLogEventsResponse, GetLogEventsError>

Lists log events from the specified log stream. You can list all the log events or filter using a time range.

By default, this operation returns as many log events as can fit in a response size of 1MB (up to 10,000 log events). If the results include tokens, there are more log events available. You can get additional log events by specifying one of the tokens in a subsequent call.

fn list_tags_log_group(&self,
input: &ListTagsLogGroupRequest)
-> Result<ListTagsLogGroupResponse, ListTagsLogGroupError>

Lists the tags for the specified log group.

To add tags, use TagLogGroup. To remove tags, use UntagLogGroup.

fn put_destination(&self,
input: &PutDestinationRequest)
-> Result<PutDestinationResponse, PutDestinationError>

Creates or updates a destination. A destination encapsulates a physical resource (such as a Kinesis stream) and enables you to subscribe to a real-time stream of log events of a different account, ingested using PutLogEvents. Currently, the only supported physical resource is a Amazon Kinesis stream belonging to the same account as the destination.

A destination controls what is written to its Amazon Kinesis stream through an access policy. By default, PutDestination does not set any access policy with the destination, which means a cross-account user cannot call PutSubscriptionFilter against this destination. To enable this, the destination owner must call PutDestinationPolicy after PutDestination.

fn put_destination_policy(&self,
input: &PutDestinationPolicyRequest)
-> Result<(), PutDestinationPolicyError>

Creates or updates an access policy associated with an existing destination. An access policy is an IAM policy document that is used to authorize claims to register a subscription filter against a given destination.

fn put_log_events(&self,
input: &PutLogEventsRequest)
-> Result<PutLogEventsResponse, PutLogEventsError>

Uploads a batch of log events to the specified log stream.

You must include the sequence token obtained from the response of the previous call. An upload in a newly created log stream does not require a sequence token. You can also get the sequence token using DescribeLogStreams.

The batch of events must satisfy the following constraints:

The maximum batch size is 1,048,576 bytes, and this size is calculated as the sum of all event messages in UTF-8, plus 26 bytes for each log event.

None of the log events in the batch can be more than 2 hours in the future.

None of the log events in the batch can be older than 14 days or the retention period of the log group.

The log events in the batch must be in chronological ordered by their timestamp (the time the event occurred, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC).

The maximum number of log events in a batch is 10,000.

A batch of log events in a single request cannot span more than 24 hours. Otherwise, the operation fails.

fn put_metric_filter(&self,
input: &PutMetricFilterRequest)
-> Result<(), PutMetricFilterError>

Creates or updates a metric filter and associates it with the specified log group. Metric filters allow you to configure rules to extract metric data from log events ingested through PutLogEvents.

The maximum number of metric filters that can be associated with a log group is 100.

fn put_retention_policy(&self,
input: &PutRetentionPolicyRequest)
-> Result<(), PutRetentionPolicyError>

Sets the retention of the specified log group. A retention policy allows you to configure the number of days you want to retain log events in the specified log group.

fn put_subscription_filter(&self,
input: &PutSubscriptionFilterRequest)
-> Result<(), PutSubscriptionFilterError>

Creates or updates a subscription filter and associates it with the specified log group. Subscription filters allow you to subscribe to a real-time stream of log events ingested through PutLogEvents and have them delivered to a specific destination. Currently, the supported destinations are:

An Amazon Kinesis stream belonging to the same account as the subscription filter, for same-account delivery.

A logical destination that belongs to a different account, for cross-account delivery.

An Amazon Kinesis Firehose stream that belongs to the same account as the subscription filter, for same-account delivery.

An AWS Lambda function that belongs to the same account as the subscription filter, for same-account delivery.

There can only be one subscription filter associated with a log group.

fn tag_log_group(&self,
input: &TagLogGroupRequest)
-> Result<(), TagLogGroupError>

Adds or updates the specified tags for the specified log group.

To list the tags for a log group, use ListTagsLogGroup. To remove tags, use UntagLogGroup.

For more information about tags, see Tag Log Groups in Amazon CloudWatch Logs in the Amazon CloudWatch Logs User Guide.

fn test_metric_filter(&self,
input: &TestMetricFilterRequest)
-> Result<TestMetricFilterResponse, TestMetricFilterError>

Tests the filter pattern of a metric filter against a sample of log event messages. You can use this operation to validate the correctness of a metric filter pattern.

fn untag_log_group(&self,
input: &UntagLogGroupRequest)
-> Result<(), UntagLogGroupError>

Removes the specified tags from the specified log group.

To list the tags for a log group, use ListTagsLogGroup. To add tags, use UntagLogGroup.