Struct rusoto::logs::CreateExportTaskRequest

pub struct CreateExportTaskRequest {
    pub destination: ExportDestinationBucket,
    pub destination_prefix: Option<ExportDestinationPrefix>,
    pub from: Timestamp,
    pub log_group_name: LogGroupName,
    pub log_stream_name_prefix: Option<LogStreamName>,
    pub task_name: Option<ExportTaskName>,
    pub to: Timestamp,
}

Fields
destination: ExportDestinationBucket

Name of Amazon S3 bucket to which the log data will be exported.

Note: Only buckets in the same AWS region are supported.

destination_prefix: Option<ExportDestinationPrefix>

Prefix that will be used as the start of Amazon S3 key for every object exported. If not specified, this defaults to 'exportedlogs'.

from: Timestamp

A point in time expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. It indicates the start time of the range for the request. Events with a timestamp prior to this time will not be exported.

log_group_name: LogGroupName

The name of the log group to export.

log_stream_name_prefix: Option<LogStreamName>

Will only export log streams that match the provided logStreamNamePrefix. If you don't specify a value, no prefix filter is applied.

task_name: Option<ExportTaskName>

The name of the export task.

to: Timestamp

A point in time expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. It indicates the end time of the range for the request. Events with a timestamp later than this time will not be exported.