Struct rusoto::logs::DescribeSubscriptionFiltersRequest

pub struct DescribeSubscriptionFiltersRequest {
    pub filter_name_prefix: Option<FilterName>,
    pub limit: Option<DescribeLimit>,
    pub log_group_name: LogGroupName,
    pub next_token: Option<NextToken>,
}

Fields
filter_name_prefix: Option<FilterName>

Will only return subscription filters that match the provided filterNamePrefix. If you don't specify a value, no prefix filter is applied.

limit: Option<DescribeLimit> log_group_name: LogGroupName

The log group name for which subscription filters are to be listed.

next_token: Option<NextToken>