Struct rusoto::logs::GetLogEventsRequest

pub struct GetLogEventsRequest {
    pub end_time: Option<Timestamp>,
    pub limit: Option<EventsLimit>,
    pub log_group_name: LogGroupName,
    pub log_stream_name: LogStreamName,
    pub next_token: Option<NextToken>,
    pub start_from_head: Option<StartFromHead>,
    pub start_time: Option<Timestamp>,
}

Fields
end_time: Option<Timestamp> limit: Option<EventsLimit>

The maximum number of log events returned in the response. If you don't specify a value, the request would return as many log events as can fit in a response size of 1MB, up to 10,000 log events.

log_group_name: LogGroupName

The name of the log group to query.

log_stream_name: LogStreamName

The name of the log stream to query.

next_token: Option<NextToken>

A string token used for pagination that points to the next page of results. It must be a value obtained from the nextForwardToken or nextBackwardToken fields in the response of the previous GetLogEvents request.

start_from_head: Option<StartFromHead>

If set to true, the earliest log events would be returned first. The default is false (the latest log events are returned first).

start_time: Option<Timestamp>