Struct rusoto::logs::PutSubscriptionFilterRequest

pub struct PutSubscriptionFilterRequest {
    pub destination_arn: DestinationArn,
    pub filter_name: FilterName,
    pub filter_pattern: FilterPattern,
    pub log_group_name: LogGroupName,
    pub role_arn: Option<RoleArn>,
}

Fields
destination_arn: DestinationArn

The ARN of the destination to deliver matching log events to. Currently, the supported destinations are:

An Amazon Kinesis stream belonging to the same account as the subscription filter, for same-account delivery.

A logical destination (used via an ARN of Destination) belonging to a different account, for cross-account delivery.

An Amazon Kinesis Firehose stream belonging to the same account as the subscription filter, for same-account delivery.

An AWS Lambda function belonging to the same account as the subscription filter, for same-account delivery.

filter_name: FilterName

A name for the subscription filter.

filter_pattern: FilterPattern

A valid CloudWatch Logs filter pattern for subscribing to a filtered stream of log events.

log_group_name: LogGroupName

The name of the log group to associate the subscription filter with.

role_arn: Option<RoleArn>

The ARN of an IAM role that grants CloudWatch Logs permissions to deliver ingested log events to the destination stream. You don't need to provide the ARN when you are working with a logical destination (used via an ARN of Destination) for cross-account delivery.