Module rusoto::machinelearning

Amazon Machine Learning

Structs

BatchPrediction
Represents the output of GetBatchPrediction operation.

The content consists of the detailed metadata, the status, and the data file information of a Batch Prediction.

CreateBatchPredictionInput

CreateBatchPredictionOutput
Represents the output of a CreateBatchPrediction operation, and is an acknowledgement that Amazon ML received the request.

The CreateBatchPrediction operation is asynchronous. You can poll for status updates by using the GetBatchPrediction operation and checking the Status parameter of the result.

CreateDataSourceFromRDSInput

CreateDataSourceFromRDSOutput
Represents the output of a CreateDataSourceFromRDS operation, and is an acknowledgement that Amazon ML received the request.

The CreateDataSourceFromRDS operation is asynchronous. You can poll for updates by using the GetBatchPrediction operation and checking the Status parameter. You can inspect the Message when Status shows up as FAILED. You can also check the progress of the copy operation by going to the DataPipeline console and looking up the pipeline using the pipelineId from the describe call.

CreateDataSourceFromRedshiftInput

CreateDataSourceFromRedshiftOutput
Represents the output of a CreateDataSourceFromRedshift operation, and is an acknowledgement that Amazon ML received the request.

The CreateDataSourceFromRedshift operation is asynchronous. You can poll for updates by using the GetBatchPrediction operation and checking the Status parameter.

CreateDataSourceFromS3Input

CreateDataSourceFromS3Output
Represents the output of a CreateDataSourceFromS3 operation, and is an acknowledgement that Amazon ML received the request.

The CreateDataSourceFromS3 operation is asynchronous. You can poll for updates by using the GetBatchPrediction operation and checking the Status parameter.

CreateEvaluationInput

CreateEvaluationOutput
Represents the output of a CreateEvaluation operation, and is an acknowledgement that Amazon ML received the request.

CreateEvaluation operation is asynchronous. You can poll for status updates by using the GetEvaluation operation and checking the Status parameter.

CreateMLModelInput

CreateMLModelOutput
Represents the output of a CreateMLModel operation, and is an acknowledgement that Amazon ML received the request.

The CreateMLModel operation is asynchronous. You can poll for status updates by using the GetMLModel operation and checking the Status parameter.

CreateRealtimeEndpointInput

CreateRealtimeEndpointOutput

Represents the output of an CreateRealtimeEndpoint operation.

The result contains the MLModelId and the endpoint information for the MLModel.

The endpoint information includes the URI of the MLModel; that is, the location to send online prediction requests for the specified MLModel.

DataSource
Represents the output of the GetDataSource operation.

The content consists of the detailed metadata and data file information and the current status of the DataSource.

DeleteBatchPredictionInput

DeleteBatchPredictionOutput
Represents the output of a DeleteBatchPrediction operation.

You can use the GetBatchPrediction operation and check the value of the Status parameter to see whether a BatchPrediction is marked as DELETED.

DeleteDataSourceInput

DeleteDataSourceOutput
Represents the output of a DeleteDataSource operation.

DeleteEvaluationInput

DeleteEvaluationOutput
Represents the output of a DeleteEvaluation operation. The output indicates that Amazon Machine Learning (Amazon ML) received the request.

You can use the GetEvaluation operation and check the value of the Status parameter to see whether an Evaluation is marked as DELETED.

DeleteMLModelInput

DeleteMLModelOutput
Represents the output of a DeleteMLModel operation.

You can use the GetMLModel operation and check the value of the Status parameter to see whether an MLModel is marked as DELETED.

DeleteRealtimeEndpointInput

DeleteRealtimeEndpointOutput
Represents the output of an DeleteRealtimeEndpoint operation.

The result contains the MLModelId and the endpoint information for the MLModel.

DescribeBatchPredictionsInput

DescribeBatchPredictionsOutput
Represents the output of a DescribeBatchPredictions operation. The content is essentially a list of BatchPredictions.

DescribeDataSourcesInput

DescribeDataSourcesOutput
Represents the query results from a DescribeDataSources operation. The content is essentially a list of DataSource.

DescribeEvaluationsInput

DescribeEvaluationsOutput
Represents the query results from a DescribeEvaluations operation. The content is essentially a list of Evaluation.

DescribeMLModelsInput

DescribeMLModelsOutput
Represents the output of a DescribeMLModels operation. The content is essentially a list of MLModel.

Evaluation
Represents the output of GetEvaluation operation.

The content consists of the detailed metadata and data file information and the current status of the Evaluation.

GetBatchPredictionInput

GetBatchPredictionOutput
Represents the output of a GetBatchPrediction operation and describes a BatchPrediction.

GetDataSourceInput

GetDataSourceOutput
Represents the output of a GetDataSource operation and describes a DataSource.

GetEvaluationInput

GetEvaluationOutput
Represents the output of a GetEvaluation operation and describes an Evaluation.

GetMLModelInput

GetMLModelOutput
Represents the output of a GetMLModel operation, and provides detailed information about a MLModel.

MLModel
Represents the output of a GetMLModel operation.

The content consists of the detailed metadata and the current status of the MLModel.

MachineLearningClient
A client for the Amazon Machine Learning API.

PerformanceMetrics
Measurements of how well the MLModel performed on known observations. One of the following metrics is returned, based on the type of the MLModel:

BinaryAUC: The binary MLModel uses the Area Under the Curve (AUC) technique to measure performance.

RegressionRMSE: The regression MLModel uses the Root Mean Square Error (RMSE) technique to measure performance. RMSE measures the difference between predicted and actual values for a single variable.

MulticlassAvgFScore: The multiclass MLModel uses the F1 score technique to measure performance.

For more information about performance metrics, please see the Amazon Machine Learning Developer Guide.

PredictInput

PredictOutput

Prediction

The output from a Predict operation:

Details - Contains the following attributes: DetailsAttributes.PREDICTIVE_MODEL_TYPE - REGRESSION | BINARY | MULTICLASS DetailsAttributes.ALGORITHM - SGD

PredictedLabel - Present for either a BINARY or MULTICLASS MLModel request.

PredictedScores - Contains the raw classification score corresponding to each label.

PredictedValue - Present for a REGRESSION MLModel request.

RDSDataSpec
The data specification of an Amazon Relational Database Service (Amazon RDS) DataSource.

RDSDatabase
The database details of an Amazon RDS database.

RDSDatabaseCredentials
The database credentials to connect to a database on an RDS DB instance.

RDSMetadata
The datasource details that are specific to Amazon RDS.

RealtimeEndpointInfo
Describes the real-time endpoint information for an MLModel.

RedshiftDataSpec
Describes the data specification of an Amazon Redshift DataSource.

RedshiftDatabase
Describes the database details required to connect to an Amazon Redshift database.

RedshiftDatabaseCredentials
Describes the database credentials for connecting to a database on an Amazon Redshift cluster.

RedshiftMetadata
Describes the DataSource details specific to Amazon Redshift.

S3DataSpec
Describes the data specification of a DataSource.

UpdateBatchPredictionInput

UpdateBatchPredictionOutput
Represents the output of an UpdateBatchPrediction operation.

You can see the updated content by using the GetBatchPrediction operation.

UpdateDataSourceInput

UpdateDataSourceOutput
Represents the output of an UpdateDataSource operation.

You can see the updated content by using the GetBatchPrediction operation.

UpdateEvaluationInput

UpdateEvaluationOutput
Represents the output of an UpdateEvaluation operation.

You can see the updated content by using the GetEvaluation operation.

UpdateMLModelInput

UpdateMLModelOutput
Represents the output of an UpdateMLModel operation.

You can see the updated content by using the GetMLModel operation.

Enums

CreateBatchPredictionError

CreateDataSourceFromRDSError

CreateDataSourceFromRedshiftError

CreateDataSourceFromS3Error

CreateEvaluationError

CreateMLModelError

CreateRealtimeEndpointError

DeleteBatchPredictionError

DeleteDataSourceError

DeleteEvaluationError

DeleteMLModelError

DeleteRealtimeEndpointError

DescribeBatchPredictionsError

DescribeDataSourcesError

DescribeEvaluationsError

DescribeMLModelsError

GetBatchPredictionError

GetDataSourceError

GetEvaluationError

GetMLModelError

PredictError

UpdateBatchPredictionError

UpdateDataSourceError

UpdateEvaluationError

UpdateMLModelError

Type Definitions

Algorithm

The function used to train a MLModel. Training choices supported by Amazon ML include the following:

SGD - Stochastic Gradient Descent.

RandomForest - Random forest of decision trees.

AwsUserArn
An Amazon Web Service (AWS) user account identifier. The account identifier can be an AWS root account or an AWS Identity and Access Management (IAM) user.

BatchPredictionFilterVariable

A list of the variables to use in searching or filtering BatchPrediction.

CreatedAt - Sets the search criteria to BatchPrediction creation date.

Status - Sets the search criteria to BatchPrediction status.

Name - Sets the search criteria to the contents of BatchPrediction Name.

IAMUser - Sets the search criteria to the user account that invoked the BatchPrediction creation.

MLModelId - Sets the search criteria to the MLModel used in the BatchPrediction.

DataSourceId - Sets the search criteria to the DataSource used in the BatchPrediction.

DataURI - Sets the search criteria to the data file(s) used in the BatchPrediction. The URL can identify either a file or an Amazon Simple Storage Service (Amazon S3) bucket or directory.

BatchPredictions

ComparatorValue
The value specified in a filtering condition. The ComparatorValue becomes the reference value when matching or evaluating data values in filtering and searching functions.

ComputeStatistics

DataRearrangement

DataSchema
The schema of a DataSource. The DataSchema defines the structure of the observation data in the data file(s) referenced in the DataSource. The DataSource schema is expressed in JSON format.

DataSchema is not required if you specify a DataSchemaUri

{ "version": "1.0", "recordAnnotationFieldName": "F1", "recordWeightFieldName": "F2", "targetFieldName": "F3", "dataFormat": "CSV", "dataFileContainsHeader": true, "variables": [ { "fieldName": "F1", "fieldType": "TEXT" }, { "fieldName": "F2", "fieldType": "NUMERIC" }, { "fieldName": "F3", "fieldType": "CATEGORICAL" }, { "fieldName": "F4", "fieldType": "NUMERIC" }, { "fieldName": "F5", "fieldType": "CATEGORICAL" }, { "fieldName": "F6", "fieldType": "TEXT" }, { "fieldName": "F7", "fieldType": "WEIGHTED_INT_SEQUENCE" }, { "fieldName": "F8", "fieldType": "WEIGHTED_STRING_SEQUENCE" } ], "excludedVariableNames": [ "F6" ] }

DataSourceFilterVariable

A list of the variables to use in searching or filtering DataSource.

CreatedAt - Sets the search criteria to DataSource creation date.

Status - Sets the search criteria to DataSource status.

Name - Sets the search criteria to the contents of DataSource Name.

DataUri - Sets the search criteria to the URI of data files used to create the DataSource. The URI can identify either a file or an Amazon Simple Storage Service (Amazon S3) bucket or directory.

IAMUser - Sets the search criteria to the user account that invoked the DataSource creation.

Note
The variable names should match the variable names in the DataSource.

DataSources

DetailsAttributes
Contains the key values of DetailsMap: PredictiveModelType - Indicates the type of the MLModel. Algorithm - Indicates the algorithm was used for the MLModel.

DetailsMap
Provides any additional details regarding the prediction.

DetailsValue

EDPPipelineId

EDPResourceRole

EDPSecurityGroupId

EDPSecurityGroupIds

EDPServiceRole

EDPSubnetId

EntityId

EntityName
A user-supplied name or description of the Amazon ML resource.

EntityStatus

Entity status with the following possible values:

PENDING

INPROGRESS

FAILED

COMPLETED

DELETED

EpochTime
A timestamp represented in epoch time.

ErrorCode

ErrorMessage

EvaluationFilterVariable

A list of the variables to use in searching or filtering Evaluation.

CreatedAt - Sets the search criteria to Evaluation creation date.

Status - Sets the search criteria to Evaluation status.

Name - Sets the search criteria to the contents of Evaluation Name.

IAMUser - Sets the search criteria to the user account that invoked an evaluation.

MLModelId - Sets the search criteria to the Predictor that was evaluated.

DataSourceId - Sets the search criteria to the DataSource used in evaluation.

DataUri - Sets the search criteria to the data file(s) used in evaluation. The URL can identify either a file or an Amazon Simple Storage Service (Amazon S3) bucket or directory.

Evaluations

FloatLabel

IntegerType
Integer type that is a 32-bit signed number.

Label

LongType
Long integer type that is a 64-bit signed number.

MLModelFilterVariable

MLModelName

MLModelType

MLModels

Message
Description of the most recent details about an entity.

PageLimit

PerformanceMetricsProperties

PerformanceMetricsPropertyKey

PerformanceMetricsPropertyValue

PresignedS3Url

RDSDatabaseName
The name of a database hosted on an RDS DB instance.

RDSDatabasePassword
The password to be used by Amazon ML to connect to a database on an RDS DB instance. The password should have sufficient permissions to execute the RDSSelectQuery query.

RDSDatabaseUsername
The username to be used by Amazon ML to connect to database on an Amazon RDS instance. The username should have sufficient permissions to execute an RDSSelectSqlQuery query.

RDSInstanceIdentifier
Identifier of RDS DB Instances.

RDSSelectSqlQuery
The SQL query to be executed against the Amazon RDS database. The SQL query should be valid for the Amazon RDS type being used.

RealtimeEndpointStatus

Recipe

Record
A map of variable name-value pairs that represent an observation.

RedshiftClusterIdentifier
The ID of an Amazon Redshift cluster.

RedshiftDatabaseName
The name of a database hosted on an Amazon Redshift cluster.

RedshiftDatabasePassword
A password to be used by Amazon ML to connect to a database on an Amazon Redshift cluster. The password should have sufficient permissions to execute a RedshiftSelectSqlQuery query. The password should be valid for an Amazon Redshift USER.

RedshiftDatabaseUsername
A username to be used by Amazon Machine Learning (Amazon ML)to connect to a database on an Amazon Redshift cluster. The username should have sufficient permissions to execute the RedshiftSelectSqlQuery query. The username should be valid for an Amazon Redshift USER.

RedshiftSelectSqlQuery
Describes the SQL query to execute on the Amazon Redshift database. The SQL query should be valid for an Amazon Redshift SELECT.

RoleARN
The Amazon Resource Name (ARN) of an AWS IAM Role such as the following: arn:aws:iam::account:role/rolename.

S3Url
A reference to a file or bucket on Amazon Simple Storage Service (Amazon S3).

ScoreThreshold

ScoreValue

ScoreValuePerLabelMap
Provides the raw classification score corresponding to each label.

SortOrder

The sort order specified in a listing condition. Possible values include the following:

asc - Present the information in ascending order (from A-Z).

dsc - Present the information in descending order (from Z-A).

StringType
String type.

TrainingParameters

VariableName
The name of a variable. Currently it's used to specify the name of the target value, label, weight, and tags.

VariableValue
The value of a variable. Currently it's used to specify values of the target value, weights, and tag variables and for filtering variable values.

Verbose
Specifies whether a describe operation should return exhaustive or abbreviated information.

VipURL