Struct rusoto::machinelearning::GetDataSourceOutput

pub struct GetDataSourceOutput {
    pub compute_statistics: Option<ComputeStatistics>,
    pub created_at: Option<EpochTime>,
    pub created_by_iam_user: Option<AwsUserArn>,
    pub data_location_s_3: Option<S3Url>,
    pub data_rearrangement: Option<DataRearrangement>,
    pub data_size_in_bytes: Option<LongType>,
    pub data_source_id: Option<EntityId>,
    pub data_source_schema: Option<DataSchema>,
    pub last_updated_at: Option<EpochTime>,
    pub log_uri: Option<PresignedS3Url>,
    pub message: Option<Message>,
    pub name: Option<EntityName>,
    pub number_of_files: Option<LongType>,
    pub r_d_s_metadata: Option<RDSMetadata>,
    pub redshift_metadata: Option<RedshiftMetadata>,
    pub role_a_r_n: Option<RoleARN>,
    pub status: Option<EntityStatus>,
}

Represents the output of a GetDataSource operation and describes a DataSource.

Fields
compute_statistics: Option<ComputeStatistics>

The parameter is true if statistics need to be generated from the observation data.

created_at: Option<EpochTime>

The time that the DataSource was created. The time is expressed in epoch time.

created_by_iam_user: Option<AwsUserArn>

The AWS user account from which the DataSource was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

data_location_s_3: Option<S3Url>

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

data_rearrangement: Option<DataRearrangement>

A JSON string that captures the splitting rearrangement requirement of the DataSource.

data_size_in_bytes: Option<LongType>

The total size of observations in the data files.

data_source_id: Option<EntityId>

The ID assigned to the DataSource at creation. This value should be identical to the value of the DataSourceId in the request.

data_source_schema: Option<DataSchema>

The schema used by all of the data files of this DataSource.
Note
This parameter is provided as part of the verbose format.

last_updated_at: Option<EpochTime>

The time of the most recent edit to the DataSource. The time is expressed in epoch time.

log_uri: Option<PresignedS3Url>

A link to the file containining logs of either create DataSource operation.

message: Option<Message>

The description of the most recent details about creating the DataSource.

name: Option<EntityName>

A user-supplied name or description of the DataSource.

number_of_files: Option<LongType>

The number of data files referenced by the DataSource.

r_d_s_metadata: Option<RDSMetadata> redshift_metadata: Option<RedshiftMetadata> role_a_r_n: Option<RoleARN> status: Option<EntityStatus>

The current status of the DataSource. This element can have one of the following values:

PENDING - Amazon Machine Language (Amazon ML) submitted a request to create a DataSource.

INPROGRESS - The creation process is underway.

FAILED - The request to create a DataSource did not run to completion. It is not usable.

COMPLETED - The creation process completed successfully.

DELETED - The DataSource is marked as deleted. It is not usable.