Struct rusoto::machinelearning::GetEvaluationOutput

pub struct GetEvaluationOutput {
    pub created_at: Option<EpochTime>,
    pub created_by_iam_user: Option<AwsUserArn>,
    pub evaluation_data_source_id: Option<EntityId>,
    pub evaluation_id: Option<EntityId>,
    pub input_data_location_s_3: Option<S3Url>,
    pub last_updated_at: Option<EpochTime>,
    pub log_uri: Option<PresignedS3Url>,
    pub m_l_model_id: Option<EntityId>,
    pub message: Option<Message>,
    pub name: Option<EntityName>,
    pub performance_metrics: Option<PerformanceMetrics>,
    pub status: Option<EntityStatus>,
}

Represents the output of a GetEvaluation operation and describes an Evaluation.

Fields
created_at: Option<EpochTime>

The time that the Evaluation was created. The time is expressed in epoch time.

created_by_iam_user: Option<AwsUserArn>

The AWS user account that invoked the evaluation. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

evaluation_data_source_id: Option<EntityId>

The DataSource used for this evaluation.

evaluation_id: Option<EntityId>

The evaluation ID which is same as the EvaluationId in the request.

input_data_location_s_3: Option<S3Url>

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

last_updated_at: Option<EpochTime>

The time of the most recent edit to the BatchPrediction. The time is expressed in epoch time.

log_uri: Option<PresignedS3Url>

A link to the file that contains logs of the CreateEvaluation operation.

m_l_model_id: Option<EntityId>

The ID of the MLModel that was the focus of the evaluation.

message: Option<Message>

A description of the most recent details about evaluating the MLModel.

name: Option<EntityName>

A user-supplied name or description of the Evaluation.

performance_metrics: Option<PerformanceMetrics>

Measurements of how well the MLModel performed using observations referenced by the DataSource. One of the following metric is returned based on the type of the MLModel:

BinaryAUC: A binary MLModel uses the Area Under the Curve (AUC) technique to measure performance.

RegressionRMSE: A regression MLModel uses the Root Mean Square Error (RMSE) technique to measure performance. RMSE measures the difference between predicted and actual values for a single variable.

MulticlassAvgFScore: A multiclass MLModel uses the F1 score technique to measure performance.

For more information about performance metrics, please see the Amazon Machine Learning Developer Guide.

status: Option<EntityStatus>

The status of the evaluation. This element can have one of the following values:

PENDING - Amazon Machine Language (Amazon ML) submitted a request to evaluate an MLModel.

INPROGRESS - The evaluation is underway.

FAILED - The request to evaluate an MLModel did not run to completion. It is not usable.

COMPLETED - The evaluation process completed successfully.

DELETED - The Evaluation is marked as deleted. It is not usable.