Struct rusoto::machinelearning::GetMLModelOutput

pub struct GetMLModelOutput {
    pub created_at: Option<EpochTime>,
    pub created_by_iam_user: Option<AwsUserArn>,
    pub endpoint_info: Option<RealtimeEndpointInfo>,
    pub input_data_location_s_3: Option<S3Url>,
    pub last_updated_at: Option<EpochTime>,
    pub log_uri: Option<PresignedS3Url>,
    pub m_l_model_id: Option<EntityId>,
    pub m_l_model_type: Option<MLModelType>,
    pub message: Option<Message>,
    pub name: Option<MLModelName>,
    pub recipe: Option<Recipe>,
    pub schema: Option<DataSchema>,
    pub score_threshold: Option<ScoreThreshold>,
    pub score_threshold_last_updated_at: Option<EpochTime>,
    pub size_in_bytes: Option<LongType>,
    pub status: Option<EntityStatus>,
    pub training_data_source_id: Option<EntityId>,
    pub training_parameters: Option<TrainingParameters>,
}

Represents the output of a GetMLModel operation, and provides detailed information about a MLModel.

Fields
created_at: Option<EpochTime>

The time that the MLModel was created. The time is expressed in epoch time.

created_by_iam_user: Option<AwsUserArn>

The AWS user account from which the MLModel was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

endpoint_info: Option<RealtimeEndpointInfo>

The current endpoint of the MLModel

input_data_location_s_3: Option<S3Url>

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

last_updated_at: Option<EpochTime>

The time of the most recent edit to the MLModel. The time is expressed in epoch time.

log_uri: Option<PresignedS3Url>

A link to the file that contains logs of the CreateMLModel operation.

m_l_model_id: Option<EntityId>

The MLModel ID which is same as the MLModelId in the request.

m_l_model_type: Option<MLModelType>

Identifies the MLModel category. The following are the available types:

REGRESSION -- Produces a numeric result. For example, "What listing price should a house have?"

BINARY -- Produces one of two possible results. For example, "Is this an e-commerce website?"

MULTICLASS -- Produces more than two possible results. For example, "Is this a HIGH, LOW or MEDIUM risk trade?"

message: Option<Message>

Description of the most recent details about accessing the MLModel.

name: Option<MLModelName>

A user-supplied name or description of the MLModel.

recipe: Option<Recipe>

The recipe to use when training the MLModel. The Recipe provides detailed information about the observation data to use during training, as well as manipulations to perform on the observation data during training.
Note
This parameter is provided as part of the verbose format.

schema: Option<DataSchema>

The schema used by all of the data files referenced by the DataSource.
Note
This parameter is provided as part of the verbose format.

score_threshold: Option<ScoreThreshold>

The scoring threshold is used in binary classification MLModels, and marks the boundary between a positive prediction and a negative prediction.

Output values greater than or equal to the threshold receive a positive result from the MLModel, such as true. Output values less than the threshold receive a negative response from the MLModel, such as false.

score_threshold_last_updated_at: Option<EpochTime>

The time of the most recent edit to the ScoreThreshold. The time is expressed in epoch time.

size_in_bytes: Option<LongType> status: Option<EntityStatus>

The current status of the MLModel. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to describe a MLModel.

INPROGRESS - The request is processing.

FAILED - The request did not run to completion. It is not usable.

COMPLETED - The request completed successfully.

DELETED - The MLModel is marked as deleted. It is not usable.

training_data_source_id: Option<EntityId>

The ID of the training DataSource.

training_parameters: Option<TrainingParameters>

A list of the training parameters in the MLModel. The list is implemented as a map of key/value pairs.

The following is the current set of training parameters:

sgd.l1RegularizationAmount - Coefficient regularization L1 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to zero, resulting in a sparse feature set. If you use this parameter, specify a small value, such as 1.0E-04 or 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is not to use L1 normalization. The parameter cannot be used when L2 is specified. Use this parameter sparingly.

sgd.l2RegularizationAmount - Coefficient regularization L2 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to small, nonzero values. If you use this parameter, specify a small value, such as 1.0E-04 or 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is not to use L2 normalization. This parameter cannot be used when L1 is specified. Use this parameter sparingly.

sgd.maxPasses - The number of times that the training process traverses the observations to build the MLModel. The value is an integer that ranges from 1 to 10000. The default value is 10.

sgd.maxMLModelSizeInBytes - The maximum allowed size of the model. Depending on the input data, the model size might affect performance.

The value is an integer that ranges from 100000 to 2147483648. The default value is 33554432.