Struct rusoto::machinelearning::MLModel

pub struct MLModel {
    pub algorithm: Option<Algorithm>,
    pub created_at: Option<EpochTime>,
    pub created_by_iam_user: Option<AwsUserArn>,
    pub endpoint_info: Option<RealtimeEndpointInfo>,
    pub input_data_location_s_3: Option<S3Url>,
    pub last_updated_at: Option<EpochTime>,
    pub m_l_model_id: Option<EntityId>,
    pub m_l_model_type: Option<MLModelType>,
    pub message: Option<Message>,
    pub name: Option<MLModelName>,
    pub score_threshold: Option<ScoreThreshold>,
    pub score_threshold_last_updated_at: Option<EpochTime>,
    pub size_in_bytes: Option<LongType>,
    pub status: Option<EntityStatus>,
    pub training_data_source_id: Option<EntityId>,
    pub training_parameters: Option<TrainingParameters>,
}

Represents the output of a GetMLModel operation.

The content consists of the detailed metadata and the current status of the MLModel.

Fields
algorithm: Option<Algorithm>

The algorithm used to train the MLModel. The following algorithm is supported:

SGD -- Stochastic gradient descent. The goal of SGD is to minimize the gradient of the loss function.

created_at: Option<EpochTime>

The time that the MLModel was created. The time is expressed in epoch time.

created_by_iam_user: Option<AwsUserArn>

The AWS user account from which the MLModel was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

endpoint_info: Option<RealtimeEndpointInfo>

The current endpoint of the MLModel.

input_data_location_s_3: Option<S3Url>

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

last_updated_at: Option<EpochTime>

The time of the most recent edit to the MLModel. The time is expressed in epoch time.

m_l_model_id: Option<EntityId>

The ID assigned to the MLModel at creation.

m_l_model_type: Option<MLModelType>

Identifies the MLModel category. The following are the available types:

REGRESSION - Produces a numeric result. For example, "What listing price should a house have?".

BINARY - Produces one of two possible results. For example, "Is this a child-friendly web site?".

MULTICLASS - Produces more than two possible results. For example, "Is this a HIGH, LOW or MEDIUM risk trade?".

message: Option<Message>

A description of the most recent details about accessing the MLModel.

name: Option<MLModelName>

A user-supplied name or description of the MLModel.

score_threshold: Option<ScoreThreshold> score_threshold_last_updated_at: Option<EpochTime>

The time of the most recent edit to the ScoreThreshold. The time is expressed in epoch time.

size_in_bytes: Option<LongType> status: Option<EntityStatus>

The current status of an MLModel. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to create an MLModel.

INPROGRESS - The creation process is underway.

FAILED - The request to create an MLModel did not run to completion. It is not usable.

COMPLETED - The creation process completed successfully.

DELETED - The MLModel is marked as deleted. It is not usable.

training_data_source_id: Option<EntityId>

The ID of the training DataSource. The CreateMLModel operation uses the TrainingDataSourceId.

training_parameters: Option<TrainingParameters>

A list of the training parameters in the MLModel. The list is implemented as a map of key/value pairs.

The following is the current set of training parameters:

sgd.l1RegularizationAmount - Coefficient regularization L1 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to zero, resulting in a sparse feature set. If you use this parameter, specify a small value, such as 1.0E-04 or 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is not to use L1 normalization. The parameter cannot be used when L2 is specified. Use this parameter sparingly.

sgd.l2RegularizationAmount - Coefficient regularization L2 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to small, nonzero values. If you use this parameter, specify a small value, such as 1.0E-04 or 1.0E-08.

The valus is a double that ranges from 0 to MAX_DOUBLE. The default is not to use L2 normalization. This cannot be used when L1 is specified. Use this parameter sparingly.

sgd.maxPasses - Number of times that the training process traverses the observations to build the MLModel. The value is an integer that ranges from 1 to 10000. The default value is 10.

sgd.maxMLModelSizeInBytes - Maximum allowed size of the model. Depending on the input data, the model size might affect performance.

The value is an integer that ranges from 100000 to 2147483648. The default value is 33554432.