Struct rusoto::machinelearning::RDSDataSpec

pub struct RDSDataSpec {
    pub data_rearrangement: Option<DataRearrangement>,
    pub data_schema: Option<DataSchema>,
    pub data_schema_uri: Option<S3Url>,
    pub database_credentials: RDSDatabaseCredentials,
    pub database_information: RDSDatabase,
    pub resource_role: EDPResourceRole,
    pub s_3_staging_location: S3Url,
    pub security_group_ids: EDPSecurityGroupIds,
    pub select_sql_query: RDSSelectSqlQuery,
    pub service_role: EDPServiceRole,
    pub subnet_id: EDPSubnetId,
}

The data specification of an Amazon Relational Database Service (Amazon RDS) DataSource.

Fields
data_rearrangement: Option<DataRearrangement>

DataRearrangement - A JSON string that represents the splitting requirement of a DataSource.

Sample - "{\"splitting\":{\"percentBegin\":10,\"percentEnd\":60}}"

data_schema: Option<DataSchema>

A JSON string that represents the schema for an Amazon RDS DataSource. The DataSchema defines the structure of the observation data in the data file(s) referenced in the DataSource.

A DataSchema is not required if you specify a DataSchemaUri

Define your DataSchema as a series of key-value pairs. attributes and excludedVariableNames have an array of key-value pairs for their value. Use the following format to define your DataSchema.

{ "version": "1.0",

"recordAnnotationFieldName": "F1",

"recordWeightFieldName": "F2",

"targetFieldName": "F3",

"dataFormat": "CSV",

"dataFileContainsHeader": true,

"attributes": [

{ "fieldName": "F1", "fieldType": "TEXT" }, { "fieldName": "F2", "fieldType": "NUMERIC" }, { "fieldName": "F3", "fieldType": "CATEGORICAL" }, { "fieldName": "F4", "fieldType": "NUMERIC" }, { "fieldName": "F5", "fieldType": "CATEGORICAL" }, { "fieldName": "F6", "fieldType": "TEXT" }, { "fieldName": "F7", "fieldType": "WEIGHTED_INT_SEQUENCE" }, { "fieldName": "F8", "fieldType": "WEIGHTED_STRING_SEQUENCE" } ],

"excludedVariableNames": [ "F6" ] }
<?oxy_insert_end>

data_schema_uri: Option<S3Url>

The Amazon S3 location of the DataSchema.

database_credentials: RDSDatabaseCredentials

The AWS Identity and Access Management (IAM) credentials that are used connect to the Amazon RDS database.

database_information: RDSDatabase

Describes the DatabaseName and InstanceIdentifier of an an Amazon RDS database.

resource_role: EDPResourceRole

The role (DataPipelineDefaultResourceRole) assumed by an Amazon Elastic Compute Cloud (Amazon EC2) instance to carry out the copy operation from Amazon RDS to an Amazon S3 task. For more information, see Role templates for data pipelines.

s_3_staging_location: S3Url

The Amazon S3 location for staging Amazon RDS data. The data retrieved from Amazon RDS using SelectSqlQuery is stored in this location.

security_group_ids: EDPSecurityGroupIds

The security group IDs to be used to access a VPC-based RDS DB instance. Ensure that there are appropriate ingress rules set up to allow access to the RDS DB instance. This attribute is used by Data Pipeline to carry out the copy operation from Amazon RDS to an Amazon S3 task.

select_sql_query: RDSSelectSqlQuery

The query that is used to retrieve the observation data for the DataSource.

service_role: EDPServiceRole

The role (DataPipelineDefaultRole) assumed by AWS Data Pipeline service to monitor the progress of the copy task from Amazon RDS to Amazon S3. For more information, see Role templates for data pipelines.

subnet_id: EDPSubnetId

The subnet ID to be used to access a VPC-based RDS DB instance. This attribute is used by Data Pipeline to carry out the copy task from Amazon RDS to Amazon S3.