Struct rusoto::machinelearning::RedshiftDataSpec

pub struct RedshiftDataSpec {
    pub data_rearrangement: Option<DataRearrangement>,
    pub data_schema: Option<DataSchema>,
    pub data_schema_uri: Option<S3Url>,
    pub database_credentials: RedshiftDatabaseCredentials,
    pub database_information: RedshiftDatabase,
    pub s_3_staging_location: S3Url,
    pub select_sql_query: RedshiftSelectSqlQuery,
}

Describes the data specification of an Amazon Redshift DataSource.

Fields
data_rearrangement: Option<DataRearrangement>

Describes the splitting specifications for a DataSource.

data_schema: Option<DataSchema>

A JSON string that represents the schema for an Amazon Redshift DataSource. The DataSchema defines the structure of the observation data in the data file(s) referenced in the DataSource.

A DataSchema is not required if you specify a DataSchemaUri.

Define your DataSchema as a series of key-value pairs. attributes and excludedVariableNames have an array of key-value pairs for their value. Use the following format to define your DataSchema.

{ "version": "1.0",

"recordAnnotationFieldName": "F1",

"recordWeightFieldName": "F2",

"targetFieldName": "F3",

"dataFormat": "CSV",

"dataFileContainsHeader": true,

"attributes": [

{ "fieldName": "F1", "fieldType": "TEXT" }, { "fieldName": "F2", "fieldType": "NUMERIC" }, { "fieldName": "F3", "fieldType": "CATEGORICAL" }, { "fieldName": "F4", "fieldType": "NUMERIC" }, { "fieldName": "F5", "fieldType": "CATEGORICAL" }, { "fieldName": "F6", "fieldType": "TEXT" }, { "fieldName": "F7", "fieldType": "WEIGHTED_INT_SEQUENCE" }, { "fieldName": "F8", "fieldType": "WEIGHTED_STRING_SEQUENCE" } ],

"excludedVariableNames": [ "F6" ] }

data_schema_uri: Option<S3Url>

Describes the schema location for an Amazon Redshift DataSource.

database_credentials: RedshiftDatabaseCredentials

Describes AWS Identity and Access Management (IAM) credentials that are used connect to the Amazon Redshift database.

database_information: RedshiftDatabase

Describes the DatabaseName and ClusterIdentifier for an Amazon Redshift DataSource.

s_3_staging_location: S3Url

Describes an Amazon S3 location to store the result set of the SelectSqlQuery query.

select_sql_query: RedshiftSelectSqlQuery

Describes the SQL Query to execute on an Amazon Redshift database for an Amazon Redshift DataSource.