Struct rusoto::opsworks::App

pub struct App {
    pub app_id: Option<String>,
    pub app_source: Option<Source>,
    pub attributes: Option<AppAttributes>,
    pub created_at: Option<String>,
    pub data_sources: Option<DataSources>,
    pub description: Option<String>,
    pub domains: Option<Strings>,
    pub enable_ssl: Option<Boolean>,
    pub environment: Option<EnvironmentVariables>,
    pub name: Option<String>,
    pub shortname: Option<String>,
    pub ssl_configuration: Option<SslConfiguration>,
    pub stack_id: Option<String>,
    pub type_: Option<AppType>,
}

A description of the app.

Fields
app_id: Option<String>

The app ID.

app_source: Option<Source>

A Source object that describes the app repository.

attributes: Option<AppAttributes>

The stack attributes.

created_at: Option<String>

When the app was created.

data_sources: Option<DataSources>

The app's data sources.

description: Option<String>

A description of the app.

domains: Option<Strings>

The app vhost settings with multiple domains separated by commas. For example: 'www.example.com, example.com'

enable_ssl: Option<Boolean>

Whether to enable SSL for the app.

environment: Option<EnvironmentVariables>

An array of EnvironmentVariable objects that specify environment variables to be associated with the app. After you deploy the app, these variables are defined on the associated app server instances. For more information, see Environment Variables.
There is no specific limit on the number of environment variables. However, the size of the associated data structure - which includes the variables' names, values, and protected flag values - cannot exceed 10 KB (10240 Bytes). This limit should accommodate most if not all use cases, but if you do exceed it, you will cause an exception (API) with an "Environment: is too large (maximum is 10KB)" message.

name: Option<String>

The app name.

shortname: Option<String>

The app's short name.

ssl_configuration: Option<SslConfiguration>

An SslConfiguration object with the SSL configuration.

stack_id: Option<String>

The app stack ID.

type_: Option<AppType>

The app type.