Struct rusoto::opsworks::CreateInstanceRequest

pub struct CreateInstanceRequest {
    pub agent_version: Option<String>,
    pub ami_id: Option<String>,
    pub architecture: Option<Architecture>,
    pub auto_scaling_type: Option<AutoScalingType>,
    pub availability_zone: Option<String>,
    pub block_device_mappings: Option<BlockDeviceMappings>,
    pub ebs_optimized: Option<Boolean>,
    pub hostname: Option<String>,
    pub install_updates_on_boot: Option<Boolean>,
    pub instance_type: String,
    pub layer_ids: Strings,
    pub os: Option<String>,
    pub root_device_type: Option<RootDeviceType>,
    pub ssh_key_name: Option<String>,
    pub stack_id: String,
    pub subnet_id: Option<String>,
    pub tenancy: Option<String>,
    pub virtualization_type: Option<String>,
}

Fields
agent_version: Option<String>

The default AWS OpsWorks agent version. You have the following options:

INHERIT - Use the stack's default agent version setting.

version_number - Use the specified agent version. This value overrides the stack's default setting. To update the agent version, edit the instance configuration and specify a new version. AWS OpsWorks then automatically installs that version on the instance.

The default setting is INHERIT. To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions.

ami_id: Option<String>

A custom AMI ID to be used to create the instance. The AMI should be based on one of the supported operating systems. For more information, see Using Custom AMIs.
If you specify a custom AMI, you must set Os to Custom.

architecture: Option<Architecture>

The instance architecture. The default option is x86_64. Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families and Types.

auto_scaling_type: Option<AutoScalingType>

For load-based or time-based instances, the type. Windows stacks can use only time-based instances.

availability_zone: Option<String>

The instance Availability Zone. For more information, see Regions and Endpoints.

block_device_mappings: Option<BlockDeviceMappings>

An array of BlockDeviceMapping objects that specify the instance's block devices. For more information, see Block Device Mapping. Note that block device mappings are not supported for custom AMIs.

ebs_optimized: Option<Boolean>

Whether to create an Amazon EBS-optimized instance.

hostname: Option<String>

The instance host name.

install_updates_on_boot: Option<Boolean>

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true to ensure that your instances have the latest security updates.

instance_type: String

The instance type, such as t2.micro. For a list of supported instance types, open the stack in the console, choose Instances, and choose + Instance. The Size list contains the currently supported types. For more information, see Instance Families and Types. The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

layer_ids: Strings

An array that contains the instance's layer IDs.

os: Option<String>

The instance's operating system, which must be set to one of the following.

A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2015.03, Red Hat Enterprise Linux 7, Ubuntu 12.04 LTS, or Ubuntu 14.04 LTS.

Microsoft Windows Server 2012 R2 Base.

A custom AMI: Custom.

For more information on the supported operating systems, see AWS OpsWorks Operating Systems.

The default option is the current Amazon Linux version. If you set this parameter to Custom, you must use the CreateInstance action's AmiId parameter to specify the custom AMI that you want to use. Block device mappings are not supported if the value is Custom. For more information on the supported operating systems, see Operating SystemsFor more information on how to use custom AMIs with AWS OpsWorks, see Using Custom AMIs.

root_device_type: Option<RootDeviceType>

The instance root device type. For more information, see Storage for the Root Device.

ssh_key_name: Option<String>

The instance's Amazon EC2 key-pair name.

stack_id: String

The stack ID.

subnet_id: Option<String>

The ID of the instance's subnet. If the stack is running in a VPC, you can use this parameter to override the stack's default subnet ID value and direct AWS OpsWorks to launch the instance in a different subnet.

tenancy: Option<String>

The instance's tenancy option. The default option is no tenancy, or if the instance is running in a VPC, inherit tenancy settings from the VPC. The following are valid values for this parameter: dedicated, default, or host. Because there are costs associated with changes in tenancy options, we recommend that you research tenancy options before choosing them for your instances. For more information about dedicated hosts, see Dedicated Hosts Overview and Amazon EC2 Dedicated Hosts. For more information about dedicated instances, see Dedicated Instances and Amazon EC2 Dedicated Instances.

virtualization_type: Option<String>

The instance's virtualization type, paravirtual or hvm.