Struct rusoto::opsworks::ElasticLoadBalancer

pub struct ElasticLoadBalancer {
    pub availability_zones: Option<Strings>,
    pub dns_name: Option<String>,
    pub ec_2_instance_ids: Option<Strings>,
    pub elastic_load_balancer_name: Option<String>,
    pub layer_id: Option<String>,
    pub region: Option<String>,
    pub stack_id: Option<String>,
    pub subnet_ids: Option<Strings>,
    pub vpc_id: Option<String>,
}

Describes an Elastic Load Balancing instance.

Fields
availability_zones: Option<Strings>

A list of Availability Zones.

dns_name: Option<String>

The instance's public DNS name.

ec_2_instance_ids: Option<Strings>

A list of the EC2 instances that the Elastic Load Balancing instance is managing traffic for.

elastic_load_balancer_name: Option<String>

The Elastic Load Balancing instance's name.

layer_id: Option<String>

The ID of the layer that the instance is attached to.

region: Option<String>

The instance's AWS region.

stack_id: Option<String>

The ID of the stack that the instance is associated with.

subnet_ids: Option<Strings>

A list of subnet IDs, if the stack is running in a VPC.

vpc_id: Option<String>

The VPC ID.