Struct rusoto::opsworks::RegisterInstanceRequest

pub struct RegisterInstanceRequest {
    pub hostname: Option<String>,
    pub instance_identity: Option<InstanceIdentity>,
    pub private_ip: Option<String>,
    pub public_ip: Option<String>,
    pub rsa_public_key: Option<String>,
    pub rsa_public_key_fingerprint: Option<String>,
    pub stack_id: String,
}

Fields
hostname: Option<String>

The instance's hostname.

instance_identity: Option<InstanceIdentity>

An InstanceIdentity object that contains the instance's identity.

private_ip: Option<String>

The instance's private IP address.

public_ip: Option<String>

The instance's public IP address.

rsa_public_key: Option<String>

The instances public RSA key. This key is used to encrypt communication between the instance and the service.

rsa_public_key_fingerprint: Option<String>

The instances public RSA key fingerprint.

stack_id: String

The ID of the stack that the instance is to be registered with.