Struct rusoto::opsworks::SetLoadBasedAutoScalingRequest

pub struct SetLoadBasedAutoScalingRequest {
    pub down_scaling: Option<AutoScalingThresholds>,
    pub enable: Option<Boolean>,
    pub layer_id: String,
    pub up_scaling: Option<AutoScalingThresholds>,
}

Fields
down_scaling: Option<AutoScalingThresholds>

An AutoScalingThresholds object with the downscaling threshold configuration. If the load falls below these thresholds for a specified amount of time, AWS OpsWorks stops a specified number of instances.

enable: Option<Boolean>

Enables load-based auto scaling for the layer.

layer_id: String

The layer ID.

up_scaling: Option<AutoScalingThresholds>

An AutoScalingThresholds object with the upscaling threshold configuration. If the load exceeds these thresholds for a specified amount of time, AWS OpsWorks starts a specified number of instances.