Struct rusoto::opsworks::SetPermissionRequest

pub struct SetPermissionRequest {
    pub allow_ssh: Option<Boolean>,
    pub allow_sudo: Option<Boolean>,
    pub iam_user_arn: String,
    pub level: Option<String>,
    pub stack_id: String,
}

Fields
allow_ssh: Option<Boolean>

The user is allowed to use SSH to communicate with the instance.

allow_sudo: Option<Boolean>

The user is allowed to use sudo to elevate privileges.

iam_user_arn: String

The user's IAM ARN.

level: Option<String>

The user's permission level, which must be set to one of the following strings. You cannot set your own permissions level.

deny

show

deploy

manage

iam_only

For more information on the permissions associated with these levels, see Managing User Permissions.

stack_id: String

The stack ID.