Struct rusoto::opsworks::Source

pub struct Source {
    pub password: Option<String>,
    pub revision: Option<String>,
    pub ssh_key: Option<String>,
    pub type_: Option<SourceType>,
    pub url: Option<String>,
    pub username: Option<String>,
}

Contains the information required to retrieve an app or cookbook from a repository. For more information, see Creating Apps or Custom Recipes and Cookbooks.

Fields
password: Option<String>

When included in a request, the parameter depends on the repository type.

For Amazon S3 bundles, set Password to the appropriate IAM secret access key.

For HTTP bundles and Subversion repositories, set Password to the password.

For more information on how to safely handle IAM credentials, see .

In responses, AWS OpsWorks returns *****FILTERED***** instead of the actual value.

revision: Option<String>

The application's version. AWS OpsWorks enables you to easily deploy new versions of an application. One of the simplest approaches is to have branches or revisions in your repository that represent different versions that can potentially be deployed.

ssh_key: Option<String>

In requests, the repository's SSH key.

In responses, AWS OpsWorks returns *****FILTERED***** instead of the actual value.

type_: Option<SourceType>

The repository type.

url: Option<String>

The source URL.

username: Option<String>

This parameter depends on the repository type.

For Amazon S3 bundles, set Username to the appropriate IAM access key ID.

For HTTP bundles, Git repositories, and Subversion repositories, set Username to the user name.