Struct rusoto::opsworks::UpdateInstanceRequest

pub struct UpdateInstanceRequest {
    pub agent_version: Option<String>,
    pub ami_id: Option<String>,
    pub architecture: Option<Architecture>,
    pub auto_scaling_type: Option<AutoScalingType>,
    pub ebs_optimized: Option<Boolean>,
    pub hostname: Option<String>,
    pub install_updates_on_boot: Option<Boolean>,
    pub instance_id: String,
    pub instance_type: Option<String>,
    pub layer_ids: Option<Strings>,
    pub os: Option<String>,
    pub ssh_key_name: Option<String>,
}

Fields
agent_version: Option<String>

The default AWS OpsWorks agent version. You have the following options:

INHERIT - Use the stack's default agent version setting.

version_number - Use the specified agent version. This value overrides the stack's default setting. To update the agent version, you must edit the instance configuration and specify a new version. AWS OpsWorks then automatically installs that version on the instance.

The default setting is INHERIT. To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions.

ami_id: Option<String>

A custom AMI ID to be used to create the instance. The AMI must be based on one of the supported operating systems. For more information, see Instances
If you specify a custom AMI, you must set Os to Custom.

architecture: Option<Architecture>

The instance architecture. Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families and Types.

auto_scaling_type: Option<AutoScalingType>

For load-based or time-based instances, the type. Windows stacks can use only time-based instances.

ebs_optimized: Option<Boolean>

This property cannot be updated.

hostname: Option<String>

The instance host name.

install_updates_on_boot: Option<Boolean>

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

instance_id: String

The instance ID.

instance_type: Option<String>

The instance type, such as t2.micro. For a list of supported instance types, open the stack in the console, choose Instances, and choose + Instance. The Size list contains the currently supported types. For more information, see Instance Families and Types. The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

layer_ids: Option<Strings>

The instance's layer IDs.

os: Option<String>

The instance's operating system, which must be set to one of the following.

A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2015.03, Red Hat Enterprise Linux 7, Ubuntu 12.04 LTS, or Ubuntu 14.04 LTS.

Microsoft Windows Server 2012 R2 Base.

A custom AMI: Custom.

For more information on the supported operating systems, see AWS OpsWorks Operating Systems.

The default option is the current Amazon Linux version. If you set this parameter to Custom, you must use the AmiId parameter to specify the custom AMI that you want to use. For more information on the supported operating systems, see Operating Systems. For more information on how to use custom AMIs with OpsWorks, see Using Custom AMIs.
You can specify a different Linux operating system for the updated stack, but you cannot change from Linux to Windows or Windows to Linux.

ssh_key_name: Option<String>

The instance's Amazon EC2 key name.