Module rusoto::rds

AWS Relational Database Service

If you're using the service, you're probably looking for RdsClient.

Structs

AccountAttributesMessage
Data returned by the DescribeAccountAttributes action.

AccountQuota
Describes a quota for an AWS account, for example, the number of DB instances allowed.

AddRoleToDBClusterMessage

AddSourceIdentifierToSubscriptionMessage

AddSourceIdentifierToSubscriptionResult

AddTagsToResourceMessage

ApplyPendingMaintenanceActionMessage

ApplyPendingMaintenanceActionResult

AuthorizeDBSecurityGroupIngressMessage

AuthorizeDBSecurityGroupIngressResult

AvailabilityZone

Contains Availability Zone information.

This data type is used as an element in the following data type:

OrderableDBInstanceOption

Certificate
A CA certificate for an AWS account.

CertificateMessage
Data returned by the DescribeCertificates action.

CharacterSet
This data type is used as a response element in the action DescribeDBEngineVersions.

CopyDBClusterParameterGroupMessage

CopyDBClusterParameterGroupResult

CopyDBClusterSnapshotMessage

CopyDBClusterSnapshotResult

CopyDBParameterGroupMessage

CopyDBParameterGroupResult

CopyDBSnapshotMessage

CopyDBSnapshotResult

CopyOptionGroupMessage

CopyOptionGroupResult

CreateDBClusterMessage

CreateDBClusterParameterGroupMessage

CreateDBClusterParameterGroupResult

CreateDBClusterResult

CreateDBClusterSnapshotMessage

CreateDBClusterSnapshotResult

CreateDBInstanceMessage

CreateDBInstanceReadReplicaMessage

CreateDBInstanceReadReplicaResult

CreateDBInstanceResult

CreateDBParameterGroupMessage

CreateDBParameterGroupResult

CreateDBSecurityGroupMessage

CreateDBSecurityGroupResult

CreateDBSnapshotMessage

CreateDBSnapshotResult

CreateDBSubnetGroupMessage

CreateDBSubnetGroupResult

CreateEventSubscriptionMessage

CreateEventSubscriptionResult

CreateOptionGroupMessage

CreateOptionGroupResult

DBCluster
Contains the result of a successful invocation of the following actions:

CreateDBCluster

DeleteDBCluster

FailoverDBCluster

ModifyDBCluster

RestoreDBClusterFromSnapshot

RestoreDBClusterToPointInTime

This data type is used as a response element in the DescribeDBClusters action.

DBClusterMember
Contains information about an instance that is part of a DB cluster.

DBClusterMessage
Contains the result of a successful invocation of the DescribeDBClusters action.

DBClusterOptionGroupStatus
Contains status information for a DB cluster option group.

DBClusterParameterGroup
Contains the result of a successful invocation of the CreateDBClusterParameterGroup or CopyDBClusterParameterGroup action.

This data type is used as a request parameter in the DeleteDBClusterParameterGroup action, and as a response element in the DescribeDBClusterParameterGroups action.

DBClusterParameterGroupDetails
Provides details about a DB cluster parameter group including the parameters in the DB cluster parameter group.

DBClusterParameterGroupNameMessage

DBClusterParameterGroupsMessage

DBClusterRole
Describes an AWS Identity and Access Management (IAM) role that is associated with a DB cluster.

DBClusterSnapshot
Contains the result of a successful invocation of the following actions:

CreateDBClusterSnapshot

DeleteDBClusterSnapshot

This data type is used as a response element in the DescribeDBClusterSnapshots action.

DBClusterSnapshotAttribute
Contains the name and values of a manual DB cluster snapshot attribute.

Manual DB cluster snapshot attributes are used to authorize other AWS accounts to restore a manual DB cluster snapshot. For more information, see the ModifyDBClusterSnapshotAttribute API action.

DBClusterSnapshotAttributesResult
Contains the results of a successful call to the DescribeDBClusterSnapshotAttributes API action.

Manual DB cluster snapshot attributes are used to authorize other AWS accounts to copy or restore a manual DB cluster snapshot. For more information, see the ModifyDBClusterSnapshotAttribute API action.

DBClusterSnapshotMessage
Provides a list of DB cluster snapshots for the user as the result of a call to the DescribeDBClusterSnapshots action.

DBEngineVersion
This data type is used as a response element in the action DescribeDBEngineVersions.

DBEngineVersionMessage
Contains the result of a successful invocation of the DescribeDBEngineVersions action.

DBInstance
Contains the result of a successful invocation of the following actions:

CreateDBInstance

DeleteDBInstance

ModifyDBInstance

This data type is used as a response element in the DescribeDBInstances action.

DBInstanceMessage
Contains the result of a successful invocation of the DescribeDBInstances action.

DBInstanceStatusInfo
Provides a list of status information for a DB instance.

DBParameterGroup
Contains the result of a successful invocation of the CreateDBParameterGroup action.

This data type is used as a request parameter in the DeleteDBParameterGroup action, and as a response element in the DescribeDBParameterGroups action.

DBParameterGroupDetails
Contains the result of a successful invocation of the DescribeDBParameters action.

DBParameterGroupNameMessage
Contains the result of a successful invocation of the ModifyDBParameterGroup or ResetDBParameterGroup action.

DBParameterGroupStatus

The status of the DB parameter group.

This data type is used as a response element in the following actions:

CreateDBInstance

CreateDBInstanceReadReplica

DeleteDBInstance

ModifyDBInstance

RebootDBInstance

RestoreDBInstanceFromDBSnapshot

DBParameterGroupsMessage
Contains the result of a successful invocation of the DescribeDBParameterGroups action.

DBSecurityGroup
Contains the result of a successful invocation of the following actions:

DescribeDBSecurityGroups

AuthorizeDBSecurityGroupIngress

CreateDBSecurityGroup

RevokeDBSecurityGroupIngress

This data type is used as a response element in the DescribeDBSecurityGroups action.

DBSecurityGroupMembership

This data type is used as a response element in the following actions:

ModifyDBInstance

RebootDBInstance

RestoreDBInstanceFromDBSnapshot

RestoreDBInstanceToPointInTime

DBSecurityGroupMessage
Contains the result of a successful invocation of the DescribeDBSecurityGroups action.

DBSnapshot
Contains the result of a successful invocation of the following actions:

CreateDBSnapshot

DeleteDBSnapshot

This data type is used as a response element in the DescribeDBSnapshots action.

DBSnapshotAttribute
Contains the name and values of a manual DB snapshot attribute

Manual DB snapshot attributes are used to authorize other AWS accounts to restore a manual DB snapshot. For more information, see the ModifyDBSnapshotAttribute API.

DBSnapshotAttributesResult
Contains the results of a successful call to the DescribeDBSnapshotAttributes API action.

Manual DB snapshot attributes are used to authorize other AWS accounts to copy or restore a manual DB snapshot. For more information, see the ModifyDBSnapshotAttribute API action.

DBSnapshotMessage
Contains the result of a successful invocation of the DescribeDBSnapshots action.

DBSubnetGroup
Contains the result of a successful invocation of the following actions:

CreateDBSubnetGroup

ModifyDBSubnetGroup

DescribeDBSubnetGroups

DeleteDBSubnetGroup

This data type is used as a response element in the DescribeDBSubnetGroups action.

DBSubnetGroupMessage
Contains the result of a successful invocation of the DescribeDBSubnetGroups action.

DeleteDBClusterMessage

DeleteDBClusterParameterGroupMessage

DeleteDBClusterResult

DeleteDBClusterSnapshotMessage

DeleteDBClusterSnapshotResult

DeleteDBInstanceMessage

DeleteDBInstanceResult

DeleteDBParameterGroupMessage

DeleteDBSecurityGroupMessage

DeleteDBSnapshotMessage

DeleteDBSnapshotResult

DeleteDBSubnetGroupMessage

DeleteEventSubscriptionMessage

DeleteEventSubscriptionResult

DeleteOptionGroupMessage

DescribeAccountAttributesMessage

DescribeCertificatesMessage

DescribeDBClusterParameterGroupsMessage

DescribeDBClusterParametersMessage

DescribeDBClusterSnapshotAttributesMessage

DescribeDBClusterSnapshotAttributesResult

DescribeDBClusterSnapshotsMessage

DescribeDBClustersMessage

DescribeDBEngineVersionsMessage

DescribeDBInstancesMessage

DescribeDBLogFilesDetails
This data type is used as a response element to DescribeDBLogFiles.

DescribeDBLogFilesMessage

DescribeDBLogFilesResponse
The response from a call to DescribeDBLogFiles.

DescribeDBParameterGroupsMessage

DescribeDBParametersMessage

DescribeDBSecurityGroupsMessage

DescribeDBSnapshotAttributesMessage

DescribeDBSnapshotAttributesResult

DescribeDBSnapshotsMessage

DescribeDBSubnetGroupsMessage

DescribeEngineDefaultClusterParametersMessage

DescribeEngineDefaultClusterParametersResult

DescribeEngineDefaultParametersMessage

DescribeEngineDefaultParametersResult

DescribeEventCategoriesMessage

DescribeEventSubscriptionsMessage

DescribeEventsMessage

DescribeOptionGroupOptionsMessage

DescribeOptionGroupsMessage

DescribeOrderableDBInstanceOptionsMessage

DescribePendingMaintenanceActionsMessage

DescribeReservedDBInstancesMessage

DescribeReservedDBInstancesOfferingsMessage

DescribeSourceRegionsMessage

DomainMembership
An Active Directory Domain membership record associated with the DB instance.

DownloadDBLogFilePortionDetails
This data type is used as a response element to DownloadDBLogFilePortion.

DownloadDBLogFilePortionMessage

EC2SecurityGroup

This data type is used as a response element in the following actions:

AuthorizeDBSecurityGroupIngress

DescribeDBSecurityGroups

RevokeDBSecurityGroupIngress

Endpoint

This data type is used as a response element in the following actions:

CreateDBInstance

DescribeDBInstances

DeleteDBInstance

EngineDefaults
Contains the result of a successful invocation of the DescribeEngineDefaultParameters action.

Event
This data type is used as a response element in the DescribeEvents action.

EventCategoriesMap
Contains the results of a successful invocation of the DescribeEventCategories action.

EventCategoriesMessage
Data returned from the DescribeEventCategories action.

EventSubscription
Contains the results of a successful invocation of the DescribeEventSubscriptions action.

EventSubscriptionsMessage
Data returned by the DescribeEventSubscriptions action.

EventsMessage
Contains the result of a successful invocation of the DescribeEvents action.

FailoverDBClusterMessage

FailoverDBClusterResult

Filter
This type is not currently supported.

IPRange
This data type is used as a response element in the DescribeDBSecurityGroups action.

ListTagsForResourceMessage

ModifyDBClusterMessage

ModifyDBClusterParameterGroupMessage

ModifyDBClusterResult

ModifyDBClusterSnapshotAttributeMessage

ModifyDBClusterSnapshotAttributeResult

ModifyDBInstanceMessage

ModifyDBInstanceResult

ModifyDBParameterGroupMessage

ModifyDBSnapshotAttributeMessage

ModifyDBSnapshotAttributeResult

ModifyDBSubnetGroupMessage

ModifyDBSubnetGroupResult

ModifyEventSubscriptionMessage

ModifyEventSubscriptionResult

ModifyOptionGroupMessage

ModifyOptionGroupResult

OptionConfiguration
A list of all available options

OptionGroup

OptionGroupMembership
Provides information on the option groups the DB instance is a member of.

OptionGroupOption
Available option.

OptionGroupOptionSetting
Option group option settings are used to display settings available for each option with their default values and other information. These values are used with the DescribeOptionGroupOptions action.

OptionGroupOptionsMessage

OptionGroups
List of option groups.

OptionSetting
Option settings are the actual settings being applied or configured for that option. It is used when you modify an option group or describe option groups. For example, the NATIVE_NETWORK_ENCRYPTION option has a setting called SQLNET.ENCRYPTION_SERVER that can have several different values.

OptionVersion
The version for an option. Option group option versions are returned by the DescribeOptionGroupOptions action.

OrderableDBInstanceOption
Contains a list of available options for a DB instance

This data type is used as a response element in the DescribeOrderableDBInstanceOptions action.

OrderableDBInstanceOptionsMessage
Contains the result of a successful invocation of the DescribeOrderableDBInstanceOptions action.

Parameter
This data type is used as a request parameter in the ModifyDBParameterGroup and ResetDBParameterGroup actions.

This data type is used as a response element in the DescribeEngineDefaultParameters and DescribeDBParameters actions.

PendingMaintenanceAction
Provides information about a pending maintenance action for a resource.

PendingMaintenanceActionsMessage
Data returned from the DescribePendingMaintenanceActions action.

PendingModifiedValues
This data type is used as a response element in the ModifyDBInstance action.

PromoteReadReplicaDBClusterMessage

PromoteReadReplicaDBClusterResult

PromoteReadReplicaMessage

PromoteReadReplicaResult

PurchaseReservedDBInstancesOfferingMessage

PurchaseReservedDBInstancesOfferingResult

RDSOption
Option details.

RdsClient
A client for the Amazon RDS API.

RebootDBInstanceMessage

RebootDBInstanceResult

RecurringCharge
This data type is used as a response element in the DescribeReservedDBInstances and DescribeReservedDBInstancesOfferings actions.

RemoveRoleFromDBClusterMessage

RemoveSourceIdentifierFromSubscriptionMessage

RemoveSourceIdentifierFromSubscriptionResult

RemoveTagsFromResourceMessage

ReservedDBInstance
This data type is used as a response element in the DescribeReservedDBInstances and PurchaseReservedDBInstancesOffering actions.

ReservedDBInstanceMessage
Contains the result of a successful invocation of the DescribeReservedDBInstances action.

ReservedDBInstancesOffering
This data type is used as a response element in the DescribeReservedDBInstancesOfferings action.

ReservedDBInstancesOfferingMessage
Contains the result of a successful invocation of the DescribeReservedDBInstancesOfferings action.

ResetDBClusterParameterGroupMessage

ResetDBParameterGroupMessage

ResourcePendingMaintenanceActions
Describes the pending maintenance actions for a resource.

RestoreDBClusterFromS3Message

RestoreDBClusterFromS3Result

RestoreDBClusterFromSnapshotMessage

RestoreDBClusterFromSnapshotResult

RestoreDBClusterToPointInTimeMessage

RestoreDBClusterToPointInTimeResult

RestoreDBInstanceFromDBSnapshotMessage

RestoreDBInstanceFromDBSnapshotResult

RestoreDBInstanceToPointInTimeMessage

RestoreDBInstanceToPointInTimeResult

RevokeDBSecurityGroupIngressMessage

RevokeDBSecurityGroupIngressResult

SourceRegion
Contains an AWS Region name as the result of a successful call to the DescribeSourceRegions action.

SourceRegionMessage
Contains the result of a successful invocation of the DescribeSourceRegions action.

Subnet
This data type is used as a response element in the DescribeDBSubnetGroups action.

Tag
Metadata assigned to an Amazon RDS resource consisting of a key-value pair.

TagListMessage

Timezone
A time zone associated with a DBInstance or a DBSnapshot. This data type is an element in the response to the DescribeDBInstances, the DescribeDBSnapshots, and the DescribeDBEngineVersions actions.

UpgradeTarget
The version of the database engine that a DB instance can be upgraded to.

VpcSecurityGroupMembership
This data type is used as a response element for queries on VPC security group membership.

Enums

AddRoleToDBClusterError
Errors returned by AddRoleToDBCluster

AddSourceIdentifierToSubscriptionError
Errors returned by AddSourceIdentifierToSubscription

AddTagsToResourceError
Errors returned by AddTagsToResource

ApplyPendingMaintenanceActionError
Errors returned by ApplyPendingMaintenanceAction

AuthorizeDBSecurityGroupIngressError
Errors returned by AuthorizeDBSecurityGroupIngress

CopyDBClusterParameterGroupError
Errors returned by CopyDBClusterParameterGroup

CopyDBClusterSnapshotError
Errors returned by CopyDBClusterSnapshot

CopyDBParameterGroupError
Errors returned by CopyDBParameterGroup

CopyDBSnapshotError
Errors returned by CopyDBSnapshot

CopyOptionGroupError
Errors returned by CopyOptionGroup

CreateDBClusterError
Errors returned by CreateDBCluster

CreateDBClusterParameterGroupError
Errors returned by CreateDBClusterParameterGroup

CreateDBClusterSnapshotError
Errors returned by CreateDBClusterSnapshot

CreateDBInstanceError
Errors returned by CreateDBInstance

CreateDBInstanceReadReplicaError
Errors returned by CreateDBInstanceReadReplica

CreateDBParameterGroupError
Errors returned by CreateDBParameterGroup

CreateDBSecurityGroupError
Errors returned by CreateDBSecurityGroup

CreateDBSnapshotError
Errors returned by CreateDBSnapshot

CreateDBSubnetGroupError
Errors returned by CreateDBSubnetGroup

CreateEventSubscriptionError
Errors returned by CreateEventSubscription

CreateOptionGroupError
Errors returned by CreateOptionGroup

DeleteDBClusterError
Errors returned by DeleteDBCluster

DeleteDBClusterParameterGroupError
Errors returned by DeleteDBClusterParameterGroup

DeleteDBClusterSnapshotError
Errors returned by DeleteDBClusterSnapshot

DeleteDBInstanceError
Errors returned by DeleteDBInstance

DeleteDBParameterGroupError
Errors returned by DeleteDBParameterGroup

DeleteDBSecurityGroupError
Errors returned by DeleteDBSecurityGroup

DeleteDBSnapshotError
Errors returned by DeleteDBSnapshot

DeleteDBSubnetGroupError
Errors returned by DeleteDBSubnetGroup

DeleteEventSubscriptionError
Errors returned by DeleteEventSubscription

DeleteOptionGroupError
Errors returned by DeleteOptionGroup

DescribeAccountAttributesError
Errors returned by DescribeAccountAttributes

DescribeCertificatesError
Errors returned by DescribeCertificates

DescribeDBClusterParameterGroupsError
Errors returned by DescribeDBClusterParameterGroups

DescribeDBClusterParametersError
Errors returned by DescribeDBClusterParameters

DescribeDBClusterSnapshotAttributesError
Errors returned by DescribeDBClusterSnapshotAttributes

DescribeDBClusterSnapshotsError
Errors returned by DescribeDBClusterSnapshots

DescribeDBClustersError
Errors returned by DescribeDBClusters

DescribeDBEngineVersionsError
Errors returned by DescribeDBEngineVersions

DescribeDBInstancesError
Errors returned by DescribeDBInstances

DescribeDBLogFilesError
Errors returned by DescribeDBLogFiles

DescribeDBParameterGroupsError
Errors returned by DescribeDBParameterGroups

DescribeDBParametersError
Errors returned by DescribeDBParameters

DescribeDBSecurityGroupsError
Errors returned by DescribeDBSecurityGroups

DescribeDBSnapshotAttributesError
Errors returned by DescribeDBSnapshotAttributes

DescribeDBSnapshotsError
Errors returned by DescribeDBSnapshots

DescribeDBSubnetGroupsError
Errors returned by DescribeDBSubnetGroups

DescribeEngineDefaultClusterParametersError
Errors returned by DescribeEngineDefaultClusterParameters

DescribeEngineDefaultParametersError
Errors returned by DescribeEngineDefaultParameters

DescribeEventCategoriesError
Errors returned by DescribeEventCategories

DescribeEventSubscriptionsError
Errors returned by DescribeEventSubscriptions

DescribeEventsError
Errors returned by DescribeEvents

DescribeOptionGroupOptionsError
Errors returned by DescribeOptionGroupOptions

DescribeOptionGroupsError
Errors returned by DescribeOptionGroups

DescribeOrderableDBInstanceOptionsError
Errors returned by DescribeOrderableDBInstanceOptions

DescribePendingMaintenanceActionsError
Errors returned by DescribePendingMaintenanceActions

DescribeReservedDBInstancesError
Errors returned by DescribeReservedDBInstances

DescribeReservedDBInstancesOfferingsError
Errors returned by DescribeReservedDBInstancesOfferings

DescribeSourceRegionsError
Errors returned by DescribeSourceRegions

DownloadDBLogFilePortionError
Errors returned by DownloadDBLogFilePortion

FailoverDBClusterError
Errors returned by FailoverDBCluster

ListTagsForResourceError
Errors returned by ListTagsForResource

ModifyDBClusterError
Errors returned by ModifyDBCluster

ModifyDBClusterParameterGroupError
Errors returned by ModifyDBClusterParameterGroup

ModifyDBClusterSnapshotAttributeError
Errors returned by ModifyDBClusterSnapshotAttribute

ModifyDBInstanceError
Errors returned by ModifyDBInstance

ModifyDBParameterGroupError
Errors returned by ModifyDBParameterGroup

ModifyDBSnapshotAttributeError
Errors returned by ModifyDBSnapshotAttribute

ModifyDBSubnetGroupError
Errors returned by ModifyDBSubnetGroup

ModifyEventSubscriptionError
Errors returned by ModifyEventSubscription

ModifyOptionGroupError
Errors returned by ModifyOptionGroup

PromoteReadReplicaDBClusterError
Errors returned by PromoteReadReplicaDBCluster

PromoteReadReplicaError
Errors returned by PromoteReadReplica

PurchaseReservedDBInstancesOfferingError
Errors returned by PurchaseReservedDBInstancesOffering

RebootDBInstanceError
Errors returned by RebootDBInstance

RemoveRoleFromDBClusterError
Errors returned by RemoveRoleFromDBCluster

RemoveSourceIdentifierFromSubscriptionError
Errors returned by RemoveSourceIdentifierFromSubscription

RemoveTagsFromResourceError
Errors returned by RemoveTagsFromResource

ResetDBClusterParameterGroupError
Errors returned by ResetDBClusterParameterGroup

ResetDBParameterGroupError
Errors returned by ResetDBParameterGroup

RestoreDBClusterFromS3Error
Errors returned by RestoreDBClusterFromS3

RestoreDBClusterFromSnapshotError
Errors returned by RestoreDBClusterFromSnapshot

RestoreDBClusterToPointInTimeError
Errors returned by RestoreDBClusterToPointInTime

RestoreDBInstanceFromDBSnapshotError
Errors returned by RestoreDBInstanceFromDBSnapshot

RestoreDBInstanceToPointInTimeError
Errors returned by RestoreDBInstanceToPointInTime

RevokeDBSecurityGroupIngressError
Errors returned by RevokeDBSecurityGroupIngress

Type Definitions

AccountQuotaList

ApplyMethod

AttributeValueList

AvailabilityZoneList

AvailabilityZones

Boolean

BooleanOptional

CertificateList

DBClusterList

DBClusterMemberList

DBClusterOptionGroupMemberships

DBClusterParameterGroupList

DBClusterRoles

DBClusterSnapshotAttributeList

DBClusterSnapshotList

DBEngineVersionList

DBInstanceList

DBInstanceStatusInfoList

DBParameterGroupList

DBParameterGroupStatusList

DBSecurityGroupMembershipList

DBSecurityGroupNameList

DBSecurityGroups

DBSnapshotAttributeList

DBSnapshotList

DBSubnetGroups

DescribeDBLogFilesList

DomainMembershipList
List of Active Directory Domain membership records associated with a DB instance.

Double

EC2SecurityGroupList

EventCategoriesList

EventCategoriesMapList

EventList

EventSubscriptionsList

FilterList

FilterValueList

IPRangeList

Integer

IntegerOptional

KeyList

Long

OptionConfigurationList

OptionGroupMembershipList

OptionGroupOptionSettingsList

OptionGroupOptionVersionsList

OptionGroupOptionsList
List of available option group options.

OptionGroupsList

OptionNamesList

OptionSettingConfigurationList

OptionSettingsList

OptionsConflictsWith

OptionsDependedOn

OptionsList

OrderableDBInstanceOptionsList

ParametersList

PendingMaintenanceActionDetails

PendingMaintenanceActions

ReadReplicaDBInstanceIdentifierList

ReadReplicaIdentifierList

RecurringChargeList

ReservedDBInstanceList

ReservedDBInstancesOfferingList

SourceIdsList

SourceRegionList

SourceType

SubnetIdentifierList

SubnetList

SupportedCharacterSetsList

SupportedTimezonesList

TStamp

TagList
A list of tags.

ValidUpgradeTargetList

VpcSecurityGroupIdList

VpcSecurityGroupMembershipList