Struct rusoto::rds::RdsClient

pub struct RdsClient<P, D> where P: ProvideAwsCredentials,
        D: DispatchSignedRequest { /* fields omitted */ }

A client for the Amazon RDS API.

Methods

impl<P, D> RdsClient<P, D> where P: ProvideAwsCredentials,
D: DispatchSignedRequest

fn new(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn add_role_to_db_cluster(&self,
input: &AddRoleToDBClusterMessage)
-> Result<(), AddRoleToDBClusterError>

Associates an Identity and Access Management (IAM) role from an Aurora DB cluster. For more information, see Authorizing Amazon Aurora to Access Other AWS Services On Your Behalf.

fn add_source_identifier_to_subscription(&self,
input: &AddSourceIdentifierToSubscriptionMessage)
-> Result<AddSourceIdentifierToSubscriptionResult, AddSourceIdentifierToSubscriptionError>

Adds a source identifier to an existing RDS event notification subscription.

fn add_tags_to_resource(&self,
input: &AddTagsToResourceMessage)
-> Result<(), AddTagsToResourceError>

Adds metadata tags to an Amazon RDS resource. These tags can also be used with cost allocation reporting to track cost associated with Amazon RDS resources, or used in a Condition statement in an IAM policy for Amazon RDS.

For an overview on tagging Amazon RDS resources, see Tagging Amazon RDS Resources.

fn apply_pending_maintenance_action(&self,
input: &ApplyPendingMaintenanceActionMessage)
-> Result<ApplyPendingMaintenanceActionResult, ApplyPendingMaintenanceActionError>

Applies a pending maintenance action to a resource (for example, to a DB instance).

fn authorize_db_security_group_ingress(&self,
input: &AuthorizeDBSecurityGroupIngressMessage)
-> Result<AuthorizeDBSecurityGroupIngressResult, AuthorizeDBSecurityGroupIngressError>

Enables ingress to a DBSecurityGroup using one of two forms of authorization. First, EC2 or VPC security groups can be added to the DBSecurityGroup if the application using the database is running on EC2 or VPC instances. Second, IP ranges are available if the application accessing your database is running on the Internet. Required parameters for this API are one of CIDR range, EC2SecurityGroupId for VPC, or (EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId for non-VPC).

You cannot authorize ingress from an EC2 security group in one region to an Amazon RDS DB instance in another. You cannot authorize ingress from a VPC security group in one VPC to an Amazon RDS DB instance in another.

For an overview of CIDR ranges, go to the Wikipedia Tutorial.

fn copy_db_cluster_parameter_group(&self,
input: &CopyDBClusterParameterGroupMessage)
-> Result<CopyDBClusterParameterGroupResult, CopyDBClusterParameterGroupError>

Copies the specified DB cluster parameter group.

fn copy_db_cluster_snapshot(&self,
input: &CopyDBClusterSnapshotMessage)
-> Result<CopyDBClusterSnapshotResult, CopyDBClusterSnapshotError>

Creates a snapshot of a DB cluster. For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn copy_db_parameter_group(&self,
input: &CopyDBParameterGroupMessage)
-> Result<CopyDBParameterGroupResult, CopyDBParameterGroupError>

Copies the specified DB parameter group.

fn copy_db_snapshot(&self,
input: &CopyDBSnapshotMessage)
-> Result<CopyDBSnapshotResult, CopyDBSnapshotError>

Copies the specified DB snapshot. The source DB snapshot must be in the "available" state.

To copy a DB snapshot from a shared manual DB snapshot, SourceDBSnapshotIdentifier must be the Amazon Resource Name (ARN) of the shared DB snapshot.

You can copy an encrypted DB snapshot from another AWS Region. In that case, the region where you call the CopyDBSnapshot action is the destination region for the encrypted DB snapshot to be copied to. To copy an encrypted DB snapshot from another region, you must provide the following values:

KmsKeyId - The AWS Key Management System (KMS) key identifier for the key to use to encrypt the copy of the DB snapshot in the destination region.

PreSignedUrl - A URL that contains a Signature Version 4 signed request for the CopyDBSnapshot action to be called in the source region where the DB snapshot will be copied from. The presigned URL must be a valid request for the CopyDBSnapshot API action that can be executed in the source region that contains the encrypted DB snapshot to be copied.

The presigned URL request must contain the following parameter values:

DestinationRegion - The AWS Region that the encrypted DB snapshot will be copied to. This region is the same one where the CopyDBSnapshot action is called that contains this presigned URL.

For example, if you copy an encrypted DB snapshot from the us-west-2 region to the us-east-1 region, then you will call the CopyDBSnapshot action in the us-east-1 region and provide a presigned URL that contains a call to the CopyDBSnapshot action in the us-west-2 region. For this example, the DestinationRegion in the presigned URL must be set to the us-east-1 region.

KmsKeyId - The KMS key identifier for the key to use to encrypt the copy of the DB snapshot in the destination region. This identifier is the same for both the CopyDBSnapshot action that is called in the destination region, and the action contained in the presigned URL.

SourceDBSnapshotIdentifier - The DB snapshot identifier for the encrypted snapshot to be copied. This identifier must be in the Amazon Resource Name (ARN) format for the source region. For example, if you copy an encrypted DB snapshot from the us-west-2 region, then your SourceDBSnapshotIdentifier looks like this example: arn:aws:rds:us-west-2:123456789012:snapshot:mysql-instance1-snapshot-20161115.

To learn how to generate a Signature Version 4 signed request, see Authenticating Requests: Using Query Parameters (AWS Signature Version 4) and Signature Version 4 Signing Process.

TargetDBSnapshotIdentifier - The identifier for the new copy of the DB snapshot in the destination region.

SourceDBSnapshotIdentifier - The DB snapshot identifier for the encrypted snapshot to be copied. This identifier must be in the ARN format for the source region and is the same value as the SourceDBSnapshotIdentifier in the presigned URL.

For more information on copying encrypted snapshots from one region to another, see Copying an Encrypted DB Snapshot to Another Region in the Amazon RDS User Guide.

fn copy_option_group(&self,
input: &CopyOptionGroupMessage)
-> Result<CopyOptionGroupResult, CopyOptionGroupError>

Copies the specified option group.

fn create_db_cluster(&self,
input: &CreateDBClusterMessage)
-> Result<CreateDBClusterResult, CreateDBClusterError>

Creates a new Amazon Aurora DB cluster.

You can use the ReplicationSourceIdentifier parameter to create the DB cluster as a Read Replica of another DB cluster.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn create_db_cluster_parameter_group(&self,
input: &CreateDBClusterParameterGroupMessage)
-> Result<CreateDBClusterParameterGroupResult, CreateDBClusterParameterGroupError>

Creates a new DB cluster parameter group.

Parameters in a DB cluster parameter group apply to all of the instances in a DB cluster.

A DB cluster parameter group is initially created with the default parameters for the database engine used by instances in the DB cluster. To provide custom values for any of the parameters, you must modify the group after creating it using ModifyDBClusterParameterGroup. Once you've created a DB cluster parameter group, you need to associate it with your DB cluster using ModifyDBCluster. When you associate a new DB cluster parameter group with a running DB cluster, you need to reboot the DB instances in the DB cluster without failover for the new DB cluster parameter group and associated settings to take effect.

After you create a DB cluster parameter group, you should wait at least 5 minutes before creating your first DB cluster that uses that DB cluster parameter group as the default parameter group. This allows Amazon RDS to fully complete the create action before the DB cluster parameter group is used as the default for a new DB cluster. This is especially important for parameters that are critical when creating the default database for a DB cluster, such as the character set for the default database defined by the character_set_database parameter. You can use the Parameter Groups option of the Amazon RDS console or the DescribeDBClusterParameters command to verify that your DB cluster parameter group has been created or modified.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn create_db_cluster_snapshot(&self,
input: &CreateDBClusterSnapshotMessage)
-> Result<CreateDBClusterSnapshotResult, CreateDBClusterSnapshotError>

Creates a snapshot of a DB cluster. For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn create_db_instance(&self,
input: &CreateDBInstanceMessage)
-> Result<CreateDBInstanceResult, CreateDBInstanceError>

Creates a new DB instance.

fn create_db_instance_read_replica(&self,
input: &CreateDBInstanceReadReplicaMessage)
-> Result<CreateDBInstanceReadReplicaResult, CreateDBInstanceReadReplicaError>

Creates a DB instance for a DB instance running MySQL, MariaDB, or PostgreSQL that acts as a Read Replica of a source DB instance.

All Read Replica DB instances are created as Single-AZ deployments with backups disabled. All other DB instance attributes (including DB security groups and DB parameter groups) are inherited from the source DB instance, except as specified below.

The source DB instance must have backup retention enabled.

fn create_db_parameter_group(&self,
input: &CreateDBParameterGroupMessage)
-> Result<CreateDBParameterGroupResult, CreateDBParameterGroupError>

Creates a new DB parameter group.

A DB parameter group is initially created with the default parameters for the database engine used by the DB instance. To provide custom values for any of the parameters, you must modify the group after creating it using ModifyDBParameterGroup. Once you've created a DB parameter group, you need to associate it with your DB instance using ModifyDBInstance. When you associate a new DB parameter group with a running DB instance, you need to reboot the DB instance without failover for the new DB parameter group and associated settings to take effect.

After you create a DB parameter group, you should wait at least 5 minutes before creating your first DB instance that uses that DB parameter group as the default parameter group. This allows Amazon RDS to fully complete the create action before the parameter group is used as the default for a new DB instance. This is especially important for parameters that are critical when creating the default database for a DB instance, such as the character set for the default database defined by the character_set_database parameter. You can use the Parameter Groups option of the Amazon RDS console or the DescribeDBParameters command to verify that your DB parameter group has been created or modified.

fn create_db_security_group(&self,
input: &CreateDBSecurityGroupMessage)
-> Result<CreateDBSecurityGroupResult, CreateDBSecurityGroupError>

Creates a new DB security group. DB security groups control access to a DB instance.

fn create_db_snapshot(&self,
input: &CreateDBSnapshotMessage)
-> Result<CreateDBSnapshotResult, CreateDBSnapshotError>

Creates a DBSnapshot. The source DBInstance must be in "available" state.

fn create_db_subnet_group(&self,
input: &CreateDBSubnetGroupMessage)
-> Result<CreateDBSubnetGroupResult, CreateDBSubnetGroupError>

Creates a new DB subnet group. DB subnet groups must contain at least one subnet in at least two AZs in the region.

fn create_event_subscription(&self,
input: &CreateEventSubscriptionMessage)
-> Result<CreateEventSubscriptionResult, CreateEventSubscriptionError>

Creates an RDS event notification subscription. This action requires a topic ARN (Amazon Resource Name) created by either the RDS console, the SNS console, or the SNS API. To obtain an ARN with SNS, you must create a topic in Amazon SNS and subscribe to the topic. The ARN is displayed in the SNS console.

You can specify the type of source (SourceType) you want to be notified of, provide a list of RDS sources (SourceIds) that triggers the events, and provide a list of event categories (EventCategories) for events you want to be notified of. For example, you can specify SourceType = db-instance, SourceIds = mydbinstance1, mydbinstance2 and EventCategories = Availability, Backup.

If you specify both the SourceType and SourceIds, such as SourceType = db-instance and SourceIdentifier = myDBInstance1, you will be notified of all the db-instance events for the specified source. If you specify a SourceType but do not specify a SourceIdentifier, you will receive notice of the events for that source type for all your RDS sources. If you do not specify either the SourceType nor the SourceIdentifier, you will be notified of events generated from all RDS sources belonging to your customer account.

fn create_option_group(&self,
input: &CreateOptionGroupMessage)
-> Result<CreateOptionGroupResult, CreateOptionGroupError>

Creates a new option group. You can create up to 20 option groups.

fn delete_db_cluster(&self,
input: &DeleteDBClusterMessage)
-> Result<DeleteDBClusterResult, DeleteDBClusterError>

The DeleteDBCluster action deletes a previously provisioned DB cluster. When you delete a DB cluster, all automated backups for that DB cluster are deleted and cannot be recovered. Manual DB cluster snapshots of the specified DB cluster are not deleted.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn delete_db_cluster_parameter_group(&self,
input: &DeleteDBClusterParameterGroupMessage)
-> Result<(), DeleteDBClusterParameterGroupError>

Deletes a specified DB cluster parameter group. The DB cluster parameter group to be deleted cannot be associated with any DB clusters.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn delete_db_cluster_snapshot(&self,
input: &DeleteDBClusterSnapshotMessage)
-> Result<DeleteDBClusterSnapshotResult, DeleteDBClusterSnapshotError>

Deletes a DB cluster snapshot. If the snapshot is being copied, the copy operation is terminated.

The DB cluster snapshot must be in the available state to be deleted.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn delete_db_instance(&self,
input: &DeleteDBInstanceMessage)
-> Result<DeleteDBInstanceResult, DeleteDBInstanceError>

The DeleteDBInstance action deletes a previously provisioned DB instance. When you delete a DB instance, all automated backups for that instance are deleted and cannot be recovered. Manual DB snapshots of the DB instance to be deleted by DeleteDBInstance are not deleted.

If you request a final DB snapshot the status of the Amazon RDS DB instance is deleting until the DB snapshot is created. The API action DescribeDBInstance is used to monitor the status of this operation. The action cannot be canceled or reverted once submitted.

Note that when a DB instance is in a failure state and has a status of failed, incompatible-restore, or incompatible-network, you can only delete it when the SkipFinalSnapshot parameter is set to true.

If the specified DB instance is part of an Amazon Aurora DB cluster, you cannot delete the DB instance if the following are true:

The DB cluster is a Read Replica of another Amazon Aurora DB cluster.

The DB instance is the only instance in the DB cluster.

To delete a DB instance in this case, first call the PromoteReadReplicaDBCluster API action to promote the DB cluster so it's no longer a Read Replica. After the promotion completes, then call the DeleteDBInstance API action to delete the final instance in the DB cluster.

fn delete_db_parameter_group(&self,
input: &DeleteDBParameterGroupMessage)
-> Result<(), DeleteDBParameterGroupError>

Deletes a specified DBParameterGroup. The DBParameterGroup to be deleted cannot be associated with any DB instances.

fn delete_db_security_group(&self,
input: &DeleteDBSecurityGroupMessage)
-> Result<(), DeleteDBSecurityGroupError>

Deletes a DB security group.

The specified DB security group must not be associated with any DB instances.

fn delete_db_snapshot(&self,
input: &DeleteDBSnapshotMessage)
-> Result<DeleteDBSnapshotResult, DeleteDBSnapshotError>

Deletes a DBSnapshot. If the snapshot is being copied, the copy operation is terminated.

The DBSnapshot must be in the available state to be deleted.

fn delete_db_subnet_group(&self,
input: &DeleteDBSubnetGroupMessage)
-> Result<(), DeleteDBSubnetGroupError>

Deletes a DB subnet group.

The specified database subnet group must not be associated with any DB instances.

fn delete_event_subscription(&self,
input: &DeleteEventSubscriptionMessage)
-> Result<DeleteEventSubscriptionResult, DeleteEventSubscriptionError>

Deletes an RDS event notification subscription.

fn delete_option_group(&self,
input: &DeleteOptionGroupMessage)
-> Result<(), DeleteOptionGroupError>

Deletes an existing option group.

fn describe_account_attributes(&self,
input: &DescribeAccountAttributesMessage)
-> Result<AccountAttributesMessage, DescribeAccountAttributesError>

Lists all of the attributes for a customer account. The attributes include Amazon RDS quotas for the account, such as the number of DB instances allowed. The description for a quota includes the quota name, current usage toward that quota, and the quota's maximum value.

This command does not take any parameters.

fn describe_certificates(&self,
input: &DescribeCertificatesMessage)
-> Result<CertificateMessage, DescribeCertificatesError>

Lists the set of CA certificates provided by Amazon RDS for this AWS account.

fn describe_db_cluster_parameter_groups(&self,
input: &DescribeDBClusterParameterGroupsMessage)
-> Result<DBClusterParameterGroupsMessage, DescribeDBClusterParameterGroupsError>

Returns a list of DBClusterParameterGroup descriptions. If a DBClusterParameterGroupName parameter is specified, the list will contain only the description of the specified DB cluster parameter group.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn describe_db_cluster_parameters(&self,
input: &DescribeDBClusterParametersMessage)
-> Result<DBClusterParameterGroupDetails, DescribeDBClusterParametersError>

Returns the detailed parameter list for a particular DB cluster parameter group.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn describe_db_cluster_snapshot_attributes(&self,
input: &DescribeDBClusterSnapshotAttributesMessage)
-> Result<DescribeDBClusterSnapshotAttributesResult, DescribeDBClusterSnapshotAttributesError>

Returns a list of DB cluster snapshot attribute names and values for a manual DB cluster snapshot.

When sharing snapshots with other AWS accounts, DescribeDBClusterSnapshotAttributes returns the restore attribute and a list of IDs for the AWS accounts that are authorized to copy or restore the manual DB cluster snapshot. If all is included in the list of values for the restore attribute, then the manual DB cluster snapshot is public and can be copied or restored by all AWS accounts.

To add or remove access for an AWS account to copy or restore a manual DB cluster snapshot, or to make the manual DB cluster snapshot public or private, use the ModifyDBClusterSnapshotAttribute API action.

fn describe_db_cluster_snapshots(&self,
input: &DescribeDBClusterSnapshotsMessage)
-> Result<DBClusterSnapshotMessage, DescribeDBClusterSnapshotsError>

Returns information about DB cluster snapshots. This API action supports pagination.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn describe_db_clusters(&self,
input: &DescribeDBClustersMessage)
-> Result<DBClusterMessage, DescribeDBClustersError>

Returns information about provisioned Aurora DB clusters. This API supports pagination.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn describe_db_engine_versions(&self,
input: &DescribeDBEngineVersionsMessage)
-> Result<DBEngineVersionMessage, DescribeDBEngineVersionsError>

Returns a list of the available DB engines.

fn describe_db_instances(&self,
input: &DescribeDBInstancesMessage)
-> Result<DBInstanceMessage, DescribeDBInstancesError>

Returns information about provisioned RDS instances. This API supports pagination.

fn describe_db_log_files(&self,
input: &DescribeDBLogFilesMessage)
-> Result<DescribeDBLogFilesResponse, DescribeDBLogFilesError>

Returns a list of DB log files for the DB instance.

fn describe_db_parameter_groups(&self,
input: &DescribeDBParameterGroupsMessage)
-> Result<DBParameterGroupsMessage, DescribeDBParameterGroupsError>

Returns a list of DBParameterGroup descriptions. If a DBParameterGroupName is specified, the list will contain only the description of the specified DB parameter group.

fn describe_db_parameters(&self,
input: &DescribeDBParametersMessage)
-> Result<DBParameterGroupDetails, DescribeDBParametersError>

Returns the detailed parameter list for a particular DB parameter group.

fn describe_db_security_groups(&self,
input: &DescribeDBSecurityGroupsMessage)
-> Result<DBSecurityGroupMessage, DescribeDBSecurityGroupsError>

Returns a list of DBSecurityGroup descriptions. If a DBSecurityGroupName is specified, the list will contain only the descriptions of the specified DB security group.

fn describe_db_snapshot_attributes(&self,
input: &DescribeDBSnapshotAttributesMessage)
-> Result<DescribeDBSnapshotAttributesResult, DescribeDBSnapshotAttributesError>

Returns a list of DB snapshot attribute names and values for a manual DB snapshot.

When sharing snapshots with other AWS accounts, DescribeDBSnapshotAttributes returns the restore attribute and a list of IDs for the AWS accounts that are authorized to copy or restore the manual DB snapshot. If all is included in the list of values for the restore attribute, then the manual DB snapshot is public and can be copied or restored by all AWS accounts.

To add or remove access for an AWS account to copy or restore a manual DB snapshot, or to make the manual DB snapshot public or private, use the ModifyDBSnapshotAttribute API action.

fn describe_db_snapshots(&self,
input: &DescribeDBSnapshotsMessage)
-> Result<DBSnapshotMessage, DescribeDBSnapshotsError>

Returns information about DB snapshots. This API action supports pagination.

fn describe_db_subnet_groups(&self,
input: &DescribeDBSubnetGroupsMessage)
-> Result<DBSubnetGroupMessage, DescribeDBSubnetGroupsError>

Returns a list of DBSubnetGroup descriptions. If a DBSubnetGroupName is specified, the list will contain only the descriptions of the specified DBSubnetGroup.

For an overview of CIDR ranges, go to the Wikipedia Tutorial.

fn describe_engine_default_cluster_parameters(&self,
input: &DescribeEngineDefaultClusterParametersMessage)
-> Result<DescribeEngineDefaultClusterParametersResult, DescribeEngineDefaultClusterParametersError>

Returns the default engine and system parameter information for the cluster database engine.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn describe_engine_default_parameters(&self,
input: &DescribeEngineDefaultParametersMessage)
-> Result<DescribeEngineDefaultParametersResult, DescribeEngineDefaultParametersError>

Returns the default engine and system parameter information for the specified database engine.

fn describe_event_categories(&self,
input: &DescribeEventCategoriesMessage)
-> Result<EventCategoriesMessage, DescribeEventCategoriesError>

Displays a list of categories for all event source types, or, if specified, for a specified source type. You can see a list of the event categories and source types in the Events topic in the Amazon RDS User Guide.

fn describe_event_subscriptions(&self,
input: &DescribeEventSubscriptionsMessage)
-> Result<EventSubscriptionsMessage, DescribeEventSubscriptionsError>

Lists all the subscription descriptions for a customer account. The description for a subscription includes SubscriptionName, SNSTopicARN, CustomerID, SourceType, SourceID, CreationTime, and Status.

If you specify a SubscriptionName, lists the description for that subscription.

fn describe_events(&self,
input: &DescribeEventsMessage)
-> Result<EventsMessage, DescribeEventsError>

Returns events related to DB instances, DB security groups, DB snapshots, and DB parameter groups for the past 14 days. Events specific to a particular DB instance, DB security group, database snapshot, or DB parameter group can be obtained by providing the name as a parameter. By default, the past hour of events are returned.

fn describe_option_group_options(&self,
input: &DescribeOptionGroupOptionsMessage)
-> Result<OptionGroupOptionsMessage, DescribeOptionGroupOptionsError>

Describes all available options.

fn describe_option_groups(&self,
input: &DescribeOptionGroupsMessage)
-> Result<OptionGroups, DescribeOptionGroupsError>

Describes the available option groups.

fn describe_orderable_db_instance_options(&self,
input: &DescribeOrderableDBInstanceOptionsMessage)
-> Result<OrderableDBInstanceOptionsMessage, DescribeOrderableDBInstanceOptionsError>

Returns a list of orderable DB instance options for the specified engine.

fn describe_pending_maintenance_actions(&self,
input: &DescribePendingMaintenanceActionsMessage)
-> Result<PendingMaintenanceActionsMessage, DescribePendingMaintenanceActionsError>

Returns a list of resources (for example, DB instances) that have at least one pending maintenance action.

fn describe_reserved_db_instances(&self,
input: &DescribeReservedDBInstancesMessage)
-> Result<ReservedDBInstanceMessage, DescribeReservedDBInstancesError>

Returns information about reserved DB instances for this account, or about a specified reserved DB instance.

fn describe_reserved_db_instances_offerings(&self,
input: &DescribeReservedDBInstancesOfferingsMessage)
-> Result<ReservedDBInstancesOfferingMessage, DescribeReservedDBInstancesOfferingsError>

Lists available reserved DB instance offerings.

fn describe_source_regions(&self,
input: &DescribeSourceRegionsMessage)
-> Result<SourceRegionMessage, DescribeSourceRegionsError>

Returns a list of the source AWS regions where the current AWS region can create a Read Replica or copy a DB snapshot from. This API action supports pagination.

fn download_db_log_file_portion(&self,
input: &DownloadDBLogFilePortionMessage)
-> Result<DownloadDBLogFilePortionDetails, DownloadDBLogFilePortionError>

Downloads all or a portion of the specified log file, up to 1 MB in size.

fn failover_db_cluster(&self,
input: &FailoverDBClusterMessage)
-> Result<FailoverDBClusterResult, FailoverDBClusterError>

Forces a failover for a DB cluster.

A failover for a DB cluster promotes one of the Aurora Replicas (read-only instances) in the DB cluster to be the primary instance (the cluster writer).

Amazon Aurora will automatically fail over to an Aurora Replica, if one exists, when the primary instance fails. You can force a failover when you want to simulate a failure of a primary instance for testing. Because each instance in a DB cluster has its own endpoint address, you will need to clean up and re-establish any existing connections that use those endpoint addresses when the failover is complete.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn list_tags_for_resource(&self,
input: &ListTagsForResourceMessage)
-> Result<TagListMessage, ListTagsForResourceError>

Lists all tags on an Amazon RDS resource.

For an overview on tagging an Amazon RDS resource, see Tagging Amazon RDS Resources.

fn modify_db_cluster(&self,
input: &ModifyDBClusterMessage)
-> Result<ModifyDBClusterResult, ModifyDBClusterError>

Modify a setting for an Amazon Aurora DB cluster. You can change one or more database configuration parameters by specifying these parameters and the new values in the request. For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn modify_db_cluster_parameter_group(&self,
input: &ModifyDBClusterParameterGroupMessage)
-> Result<DBClusterParameterGroupNameMessage, ModifyDBClusterParameterGroupError>

Modifies the parameters of a DB cluster parameter group. To modify more than one parameter, submit a list of the following: ParameterName, ParameterValue, and ApplyMethod. A maximum of 20 parameters can be modified in a single request.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

Changes to dynamic parameters are applied immediately. Changes to static parameters require a reboot without failover to the DB cluster associated with the parameter group before the change can take effect.

After you create a DB cluster parameter group, you should wait at least 5 minutes before creating your first DB cluster that uses that DB cluster parameter group as the default parameter group. This allows Amazon RDS to fully complete the create action before the parameter group is used as the default for a new DB cluster. This is especially important for parameters that are critical when creating the default database for a DB cluster, such as the character set for the default database defined by the character_set_database parameter. You can use the Parameter Groups option of the Amazon RDS console or the DescribeDBClusterParameters command to verify that your DB cluster parameter group has been created or modified.

fn modify_db_cluster_snapshot_attribute(&self,
input: &ModifyDBClusterSnapshotAttributeMessage)
-> Result<ModifyDBClusterSnapshotAttributeResult, ModifyDBClusterSnapshotAttributeError>

Adds an attribute and values to, or removes an attribute and values from, a manual DB cluster snapshot.

To share a manual DB cluster snapshot with other AWS accounts, specify restore as the AttributeName and use the ValuesToAdd parameter to add a list of IDs of the AWS accounts that are authorized to restore the manual DB cluster snapshot. Use the value all to make the manual DB cluster snapshot public, which means that it can be copied or restored by all AWS accounts. Do not add the all value for any manual DB cluster snapshots that contain private information that you don't want available to all AWS accounts.

To view which AWS accounts have access to copy or restore a manual DB cluster snapshot, or whether a manual DB cluster snapshot public or private, use the DescribeDBClusterSnapshotAttributes API action.

If a manual DB cluster snapshot is encrypted, it cannot be shared.

fn modify_db_instance(&self,
input: &ModifyDBInstanceMessage)
-> Result<ModifyDBInstanceResult, ModifyDBInstanceError>

Modifies settings for a DB instance. You can change one or more database configuration parameters by specifying these parameters and the new values in the request.

fn modify_db_parameter_group(&self,
input: &ModifyDBParameterGroupMessage)
-> Result<DBParameterGroupNameMessage, ModifyDBParameterGroupError>

Modifies the parameters of a DB parameter group. To modify more than one parameter, submit a list of the following: ParameterName, ParameterValue, and ApplyMethod. A maximum of 20 parameters can be modified in a single request.

Changes to dynamic parameters are applied immediately. Changes to static parameters require a reboot without failover to the DB instance associated with the parameter group before the change can take effect.

After you modify a DB parameter group, you should wait at least 5 minutes before creating your first DB instance that uses that DB parameter group as the default parameter group. This allows Amazon RDS to fully complete the modify action before the parameter group is used as the default for a new DB instance. This is especially important for parameters that are critical when creating the default database for a DB instance, such as the character set for the default database defined by the character_set_database parameter. You can use the Parameter Groups option of the Amazon RDS console or the DescribeDBParameters command to verify that your DB parameter group has been created or modified.

fn modify_db_snapshot_attribute(&self,
input: &ModifyDBSnapshotAttributeMessage)
-> Result<ModifyDBSnapshotAttributeResult, ModifyDBSnapshotAttributeError>

Adds an attribute and values to, or removes an attribute and values from, a manual DB snapshot.

To share a manual DB snapshot with other AWS accounts, specify restore as the AttributeName and use the ValuesToAdd parameter to add a list of IDs of the AWS accounts that are authorized to restore the manual DB snapshot. Uses the value all to make the manual DB snapshot public, which means it can be copied or restored by all AWS accounts. Do not add the all value for any manual DB snapshots that contain private information that you don't want available to all AWS accounts.

To view which AWS accounts have access to copy or restore a manual DB snapshot, or whether a manual DB snapshot public or private, use the DescribeDBSnapshotAttributes API action.

If the manual DB snapshot is encrypted, it cannot be shared.

fn modify_db_subnet_group(&self,
input: &ModifyDBSubnetGroupMessage)
-> Result<ModifyDBSubnetGroupResult, ModifyDBSubnetGroupError>

Modifies an existing DB subnet group. DB subnet groups must contain at least one subnet in at least two AZs in the region.

fn modify_event_subscription(&self,
input: &ModifyEventSubscriptionMessage)
-> Result<ModifyEventSubscriptionResult, ModifyEventSubscriptionError>

Modifies an existing RDS event notification subscription. Note that you cannot modify the source identifiers using this call; to change source identifiers for a subscription, use the AddSourceIdentifierToSubscription and RemoveSourceIdentifierFromSubscription calls.

You can see a list of the event categories for a given SourceType in the Events topic in the Amazon RDS User Guide or by using the DescribeEventCategories action.

fn modify_option_group(&self,
input: &ModifyOptionGroupMessage)
-> Result<ModifyOptionGroupResult, ModifyOptionGroupError>

Modifies an existing option group.

fn promote_read_replica(&self,
input: &PromoteReadReplicaMessage)
-> Result<PromoteReadReplicaResult, PromoteReadReplicaError>

Promotes a Read Replica DB instance to a standalone DB instance.

We recommend that you enable automated backups on your Read Replica before promoting the Read Replica. This ensures that no backup is taken during the promotion process. Once the instance is promoted to a primary instance, backups are taken based on your backup settings.

fn promote_read_replica_db_cluster(&self,
input: &PromoteReadReplicaDBClusterMessage)
-> Result<PromoteReadReplicaDBClusterResult, PromoteReadReplicaDBClusterError>

Promotes a Read Replica DB cluster to a standalone DB cluster.

fn purchase_reserved_db_instances_offering(&self,
input: &PurchaseReservedDBInstancesOfferingMessage)
-> Result<PurchaseReservedDBInstancesOfferingResult, PurchaseReservedDBInstancesOfferingError>

Purchases a reserved DB instance offering.

fn reboot_db_instance(&self,
input: &RebootDBInstanceMessage)
-> Result<RebootDBInstanceResult, RebootDBInstanceError>

Rebooting a DB instance restarts the database engine service. A reboot also applies to the DB instance any modifications to the associated DB parameter group that were pending. Rebooting a DB instance results in a momentary outage of the instance, during which the DB instance status is set to rebooting. If the RDS instance is configured for MultiAZ, it is possible that the reboot will be conducted through a failover. An Amazon RDS event is created when the reboot is completed.

If your DB instance is deployed in multiple Availability Zones, you can force a failover from one AZ to the other during the reboot. You might force a failover to test the availability of your DB instance deployment or to restore operations to the original AZ after a failover occurs.

The time required to reboot is a function of the specific database engine's crash recovery process. To improve the reboot time, we recommend that you reduce database activities as much as possible during the reboot process to reduce rollback activity for in-transit transactions.

fn remove_role_from_db_cluster(&self,
input: &RemoveRoleFromDBClusterMessage)
-> Result<(), RemoveRoleFromDBClusterError>

Disassociates an Identity and Access Management (IAM) role from an Aurora DB cluster. For more information, see Authorizing Amazon Aurora to Access Other AWS Services On Your Behalf.

fn remove_source_identifier_from_subscription(&self,
input: &RemoveSourceIdentifierFromSubscriptionMessage)
-> Result<RemoveSourceIdentifierFromSubscriptionResult, RemoveSourceIdentifierFromSubscriptionError>

Removes a source identifier from an existing RDS event notification subscription.

fn remove_tags_from_resource(&self,
input: &RemoveTagsFromResourceMessage)
-> Result<(), RemoveTagsFromResourceError>

Removes metadata tags from an Amazon RDS resource.

For an overview on tagging an Amazon RDS resource, see Tagging Amazon RDS Resources.

fn reset_db_cluster_parameter_group(&self,
input: &ResetDBClusterParameterGroupMessage)
-> Result<DBClusterParameterGroupNameMessage, ResetDBClusterParameterGroupError>

Modifies the parameters of a DB cluster parameter group to the default value. To reset specific parameters submit a list of the following: ParameterName and ApplyMethod. To reset the entire DB cluster parameter group, specify the DBClusterParameterGroupName and ResetAllParameters parameters.

When resetting the entire group, dynamic parameters are updated immediately and static parameters are set to pending-reboot to take effect on the next DB instance restart or RebootDBInstance request. You must call RebootDBInstance for every DB instance in your DB cluster that you want the updated static parameter to apply to.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn reset_db_parameter_group(&self,
input: &ResetDBParameterGroupMessage)
-> Result<DBParameterGroupNameMessage, ResetDBParameterGroupError>

Modifies the parameters of a DB parameter group to the engine/system default value. To reset specific parameters submit a list of the following: ParameterName and ApplyMethod. To reset the entire DB parameter group, specify the DBParameterGroup name and ResetAllParameters parameters. When resetting the entire group, dynamic parameters are updated immediately and static parameters are set to pending-reboot to take effect on the next DB instance restart or RebootDBInstance request.

fn restore_db_cluster_from_s3(&self,
input: &RestoreDBClusterFromS3Message)
-> Result<RestoreDBClusterFromS3Result, RestoreDBClusterFromS3Error>

Creates an Amazon Aurora DB cluster from data stored in an Amazon S3 bucket. Amazon RDS must be authorized to access the Amazon S3 bucket and the data must be created using the Percona XtraBackup utility as described in Migrating Data from MySQL by Using an Amazon S3 Bucket.

fn restore_db_cluster_from_snapshot(&self,
input: &RestoreDBClusterFromSnapshotMessage)
-> Result<RestoreDBClusterFromSnapshotResult, RestoreDBClusterFromSnapshotError>

Creates a new DB cluster from a DB cluster snapshot. The target DB cluster is created from the source DB cluster restore point with the same configuration as the original source DB cluster, except that the new DB cluster is created with the default security group.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn restore_db_cluster_to_point_in_time(&self,
input: &RestoreDBClusterToPointInTimeMessage)
-> Result<RestoreDBClusterToPointInTimeResult, RestoreDBClusterToPointInTimeError>

Restores a DB cluster to an arbitrary point in time. Users can restore to any point in time before LatestRestorableTime for up to BackupRetentionPeriod days. The target DB cluster is created from the source DB cluster with the same configuration as the original DB cluster, except that the new DB cluster is created with the default DB security group.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

fn restore_db_instance_from_db_snapshot(&self,
input: &RestoreDBInstanceFromDBSnapshotMessage)
-> Result<RestoreDBInstanceFromDBSnapshotResult, RestoreDBInstanceFromDBSnapshotError>

Creates a new DB instance from a DB snapshot. The target database is created from the source database restore point with the most of original configuration with the default security group and the default DB parameter group. By default, the new DB instance is created as a single-AZ deployment except when the instance is a SQL Server instance that has an option group that is associated with mirroring; in this case, the instance becomes a mirrored AZ deployment and not a single-AZ deployment.

If your intent is to replace your original DB instance with the new, restored DB instance, then rename your original DB instance before you call the RestoreDBInstanceFromDBSnapshot action. RDS does not allow two DB instances with the same name. Once you have renamed your original DB instance with a different identifier, then you can pass the original name of the DB instance as the DBInstanceIdentifier in the call to the RestoreDBInstanceFromDBSnapshot action. The result is that you will replace the original DB instance with the DB instance created from the snapshot.

If you are restoring from a shared manual DB snapshot, the DBSnapshotIdentifier must be the ARN of the shared DB snapshot.

fn restore_db_instance_to_point_in_time(&self,
input: &RestoreDBInstanceToPointInTimeMessage)
-> Result<RestoreDBInstanceToPointInTimeResult, RestoreDBInstanceToPointInTimeError>

Restores a DB instance to an arbitrary point in time. You can restore to any point in time before the time identified by the LatestRestorableTime property. You can restore to a point up to the number of days specified by the BackupRetentionPeriod property.

The target database is created with most of the original configuration, but in a system-selected availability zone, with the default security group, the default subnet group, and the default DB parameter group. By default, the new DB instance is created as a single-AZ deployment except when the instance is a SQL Server instance that has an option group that is associated with mirroring; in this case, the instance becomes a mirrored deployment and not a single-AZ deployment.

fn revoke_db_security_group_ingress(&self,
input: &RevokeDBSecurityGroupIngressMessage)
-> Result<RevokeDBSecurityGroupIngressResult, RevokeDBSecurityGroupIngressError>

Revokes ingress from a DBSecurityGroup for previously authorized IP ranges or EC2 or VPC Security Groups. Required parameters for this API are one of CIDRIP, EC2SecurityGroupId for VPC, or (EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId).