Module rusoto::redshift

AWS Redshift

If you're using the service, you're probably looking for RedshiftClient.

Structs

AccountWithRestoreAccess
Describes an AWS customer account authorized to restore a snapshot.

AuthorizeClusterSecurityGroupIngressMessage

AuthorizeClusterSecurityGroupIngressResult

AuthorizeSnapshotAccessMessage

AuthorizeSnapshotAccessResult

AvailabilityZone
Describes an availability zone.

Cluster
Describes a cluster.

ClusterIamRole
An AWS Identity and Access Management (IAM) role that can be used by the associated Amazon Redshift cluster to access other AWS services.

ClusterNode
The identifier of a node in a cluster.

ClusterParameterGroup
Describes a parameter group.

ClusterParameterGroupDetails
Contains the output from the DescribeClusterParameters action.

ClusterParameterGroupNameMessage

ClusterParameterGroupStatus
Describes the status of a parameter group.

ClusterParameterGroupsMessage
Contains the output from the DescribeClusterParameterGroups action.

ClusterParameterStatus
Describes the status of a parameter group.

ClusterSecurityGroup
Describes a security group.

ClusterSecurityGroupMembership
Describes a cluster security group.

ClusterSecurityGroupMessage

ClusterSnapshotCopyStatus
Returns the destination region and retention period that are configured for cross-region snapshot copy.

ClusterSubnetGroup
Describes a subnet group.

ClusterSubnetGroupMessage
Contains the output from the DescribeClusterSubnetGroups action.

ClusterVersion
Describes a cluster version, including the parameter group family and description of the version.

ClusterVersionsMessage
Contains the output from the DescribeClusterVersions action.

ClustersMessage
Contains the output from the DescribeClusters action.

CopyClusterSnapshotMessage

CopyClusterSnapshotResult

CreateClusterMessage

CreateClusterParameterGroupMessage

CreateClusterParameterGroupResult

CreateClusterResult

CreateClusterSecurityGroupMessage

CreateClusterSecurityGroupResult

CreateClusterSnapshotMessage

CreateClusterSnapshotResult

CreateClusterSubnetGroupMessage

CreateClusterSubnetGroupResult

CreateEventSubscriptionMessage

CreateEventSubscriptionResult

CreateHsmClientCertificateMessage

CreateHsmClientCertificateResult

CreateHsmConfigurationMessage

CreateHsmConfigurationResult

CreateSnapshotCopyGrantMessage
The result of the CreateSnapshotCopyGrant action.

CreateSnapshotCopyGrantResult

CreateTagsMessage
Contains the output from the CreateTags action.

DefaultClusterParameters
Describes the default cluster parameters for a parameter group family.

DeleteClusterMessage

DeleteClusterParameterGroupMessage

DeleteClusterResult

DeleteClusterSecurityGroupMessage

DeleteClusterSnapshotMessage

DeleteClusterSnapshotResult

DeleteClusterSubnetGroupMessage

DeleteEventSubscriptionMessage

DeleteHsmClientCertificateMessage

DeleteHsmConfigurationMessage

DeleteSnapshotCopyGrantMessage
The result of the DeleteSnapshotCopyGrant action.

DeleteTagsMessage
Contains the output from the DeleteTags action.

DescribeClusterParameterGroupsMessage

DescribeClusterParametersMessage

DescribeClusterSecurityGroupsMessage

DescribeClusterSnapshotsMessage

DescribeClusterSubnetGroupsMessage

DescribeClusterVersionsMessage

DescribeClustersMessage

DescribeDefaultClusterParametersMessage

DescribeDefaultClusterParametersResult

DescribeEventCategoriesMessage

DescribeEventSubscriptionsMessage

DescribeEventsMessage

DescribeHsmClientCertificatesMessage

DescribeHsmConfigurationsMessage

DescribeLoggingStatusMessage

DescribeOrderableClusterOptionsMessage

DescribeReservedNodeOfferingsMessage

DescribeReservedNodesMessage

DescribeResizeMessage

DescribeSnapshotCopyGrantsMessage
The result of the DescribeSnapshotCopyGrants action.

DescribeTableRestoreStatusMessage

DescribeTagsMessage

DisableLoggingMessage

DisableSnapshotCopyMessage

DisableSnapshotCopyResult

EC2SecurityGroup
Describes an Amazon EC2 security group.

ElasticIpStatus
Describes the status of the elastic IP (EIP) address.

EnableLoggingMessage

EnableSnapshotCopyMessage

EnableSnapshotCopyResult

Endpoint
Describes a connection endpoint.

Event
Describes an event.

EventCategoriesMap
Describes event categories.

EventCategoriesMessage

EventInfoMap
Describes event information.

EventSubscription
Describes event subscriptions.

EventSubscriptionsMessage

EventsMessage

HsmClientCertificate
Returns information about an HSM client certificate. The certificate is stored in a secure Hardware Storage Module (HSM), and used by the Amazon Redshift cluster to encrypt data files.

HsmClientCertificateMessage

HsmConfiguration
Returns information about an HSM configuration, which is an object that describes to Amazon Redshift clusters the information they require to connect to an HSM where they can store database encryption keys.

HsmConfigurationMessage

HsmStatus
Describes the status of changes to HSM settings.

IPRange
Describes an IP range used in a security group.

LoggingStatus
Describes the status of logging for a cluster.

ModifyClusterIamRolesMessage

ModifyClusterIamRolesResult

ModifyClusterMessage

ModifyClusterParameterGroupMessage

ModifyClusterResult

ModifyClusterSubnetGroupMessage

ModifyClusterSubnetGroupResult

ModifyEventSubscriptionMessage

ModifyEventSubscriptionResult

ModifySnapshotCopyRetentionPeriodMessage

ModifySnapshotCopyRetentionPeriodResult

OrderableClusterOption
Describes an orderable cluster option.

OrderableClusterOptionsMessage
Contains the output from the DescribeOrderableClusterOptions action.

Parameter
Describes a parameter in a cluster parameter group.

PendingModifiedValues
Describes cluster attributes that are in a pending state. A change to one or more the attributes was requested and is in progress or will be applied.

PurchaseReservedNodeOfferingMessage

PurchaseReservedNodeOfferingResult

RebootClusterMessage

RebootClusterResult

RecurringCharge
Describes a recurring charge.

RedshiftClient
A client for the Amazon Redshift API.

ReservedNode
Describes a reserved node. You can call the DescribeReservedNodeOfferings API to obtain the available reserved node offerings.

ReservedNodeOffering
Describes a reserved node offering.

ReservedNodeOfferingsMessage

ReservedNodesMessage

ResetClusterParameterGroupMessage

ResizeProgressMessage
Describes the result of a cluster resize operation.

RestoreFromClusterSnapshotMessage

RestoreFromClusterSnapshotResult

RestoreStatus
Describes the status of a cluster restore action. Returns null if the cluster was not created by restoring a snapshot.

RestoreTableFromClusterSnapshotMessage

RestoreTableFromClusterSnapshotResult

RevokeClusterSecurityGroupIngressMessage

RevokeClusterSecurityGroupIngressResult

RevokeSnapshotAccessMessage

RevokeSnapshotAccessResult

RotateEncryptionKeyMessage

RotateEncryptionKeyResult

Snapshot
Describes a snapshot.

SnapshotCopyGrant
The snapshot copy grant that grants Amazon Redshift permission to encrypt copied snapshots with the specified customer master key (CMK) from AWS KMS in the destination region.

For more information about managing snapshot copy grants, go to Amazon Redshift Database Encryption in the Amazon Redshift Cluster Management Guide.

SnapshotCopyGrantMessage

SnapshotMessage
Contains the output from the DescribeClusterSnapshots action.

Subnet
Describes a subnet.

TableRestoreStatus
Describes the status of a RestoreTableFromClusterSnapshot operation.

TableRestoreStatusMessage

Tag
A tag consisting of a name/value pair for a resource.

TaggedResource
A tag and its associated resource.

TaggedResourceListMessage

VpcSecurityGroupMembership
Describes the members of a VPC security group.

Enums

AuthorizeClusterSecurityGroupIngressError
Errors returned by AuthorizeClusterSecurityGroupIngress

AuthorizeSnapshotAccessError
Errors returned by AuthorizeSnapshotAccess

CopyClusterSnapshotError
Errors returned by CopyClusterSnapshot

CreateClusterError
Errors returned by CreateCluster

CreateClusterParameterGroupError
Errors returned by CreateClusterParameterGroup

CreateClusterSecurityGroupError
Errors returned by CreateClusterSecurityGroup

CreateClusterSnapshotError
Errors returned by CreateClusterSnapshot

CreateClusterSubnetGroupError
Errors returned by CreateClusterSubnetGroup

CreateEventSubscriptionError
Errors returned by CreateEventSubscription

CreateHsmClientCertificateError
Errors returned by CreateHsmClientCertificate

CreateHsmConfigurationError
Errors returned by CreateHsmConfiguration

CreateSnapshotCopyGrantError
Errors returned by CreateSnapshotCopyGrant

CreateTagsError
Errors returned by CreateTags

DeleteClusterError
Errors returned by DeleteCluster

DeleteClusterParameterGroupError
Errors returned by DeleteClusterParameterGroup

DeleteClusterSecurityGroupError
Errors returned by DeleteClusterSecurityGroup

DeleteClusterSnapshotError
Errors returned by DeleteClusterSnapshot

DeleteClusterSubnetGroupError
Errors returned by DeleteClusterSubnetGroup

DeleteEventSubscriptionError
Errors returned by DeleteEventSubscription

DeleteHsmClientCertificateError
Errors returned by DeleteHsmClientCertificate

DeleteHsmConfigurationError
Errors returned by DeleteHsmConfiguration

DeleteSnapshotCopyGrantError
Errors returned by DeleteSnapshotCopyGrant

DeleteTagsError
Errors returned by DeleteTags

DescribeClusterParameterGroupsError
Errors returned by DescribeClusterParameterGroups

DescribeClusterParametersError
Errors returned by DescribeClusterParameters

DescribeClusterSecurityGroupsError
Errors returned by DescribeClusterSecurityGroups

DescribeClusterSnapshotsError
Errors returned by DescribeClusterSnapshots

DescribeClusterSubnetGroupsError
Errors returned by DescribeClusterSubnetGroups

DescribeClusterVersionsError
Errors returned by DescribeClusterVersions

DescribeClustersError
Errors returned by DescribeClusters

DescribeDefaultClusterParametersError
Errors returned by DescribeDefaultClusterParameters

DescribeEventCategoriesError
Errors returned by DescribeEventCategories

DescribeEventSubscriptionsError
Errors returned by DescribeEventSubscriptions

DescribeEventsError
Errors returned by DescribeEvents

DescribeHsmClientCertificatesError
Errors returned by DescribeHsmClientCertificates

DescribeHsmConfigurationsError
Errors returned by DescribeHsmConfigurations

DescribeLoggingStatusError
Errors returned by DescribeLoggingStatus

DescribeOrderableClusterOptionsError
Errors returned by DescribeOrderableClusterOptions

DescribeReservedNodeOfferingsError
Errors returned by DescribeReservedNodeOfferings

DescribeReservedNodesError
Errors returned by DescribeReservedNodes

DescribeResizeError
Errors returned by DescribeResize

DescribeSnapshotCopyGrantsError
Errors returned by DescribeSnapshotCopyGrants

DescribeTableRestoreStatusError
Errors returned by DescribeTableRestoreStatus

DescribeTagsError
Errors returned by DescribeTags

DisableLoggingError
Errors returned by DisableLogging

DisableSnapshotCopyError
Errors returned by DisableSnapshotCopy

EnableLoggingError
Errors returned by EnableLogging

EnableSnapshotCopyError
Errors returned by EnableSnapshotCopy

ModifyClusterError
Errors returned by ModifyCluster

ModifyClusterIamRolesError
Errors returned by ModifyClusterIamRoles

ModifyClusterParameterGroupError
Errors returned by ModifyClusterParameterGroup

ModifyClusterSubnetGroupError
Errors returned by ModifyClusterSubnetGroup

ModifyEventSubscriptionError
Errors returned by ModifyEventSubscription

ModifySnapshotCopyRetentionPeriodError
Errors returned by ModifySnapshotCopyRetentionPeriod

PurchaseReservedNodeOfferingError
Errors returned by PurchaseReservedNodeOffering

RebootClusterError
Errors returned by RebootCluster

ResetClusterParameterGroupError
Errors returned by ResetClusterParameterGroup

RestoreFromClusterSnapshotError
Errors returned by RestoreFromClusterSnapshot

RestoreTableFromClusterSnapshotError
Errors returned by RestoreTableFromClusterSnapshot

RevokeClusterSecurityGroupIngressError
Errors returned by RevokeClusterSecurityGroupIngress

RevokeSnapshotAccessError
Errors returned by RevokeSnapshotAccess

RotateEncryptionKeyError
Errors returned by RotateEncryptionKey

Type Definitions

AccountsWithRestoreAccessList

AvailabilityZoneList

Boolean

BooleanOptional

ClusterIamRoleList

ClusterList

ClusterNodesList

ClusterParameterGroupStatusList

ClusterParameterStatusList

ClusterSecurityGroupMembershipList

ClusterSecurityGroupNameList

ClusterSecurityGroups

ClusterSubnetGroups

ClusterVersionList

Double

DoubleOptional

EC2SecurityGroupList

EventCategoriesList

EventCategoriesMapList

EventInfoMapList

EventList

EventSubscriptionsList

HsmClientCertificateList

HsmConfigurationList

IPRangeList

IamRoleArnList

ImportTablesCompleted

ImportTablesInProgress

ImportTablesNotStarted

Integer

IntegerOptional

Long

LongOptional

OrderableClusterOptionsList

ParameterApplyType

ParameterGroupList

ParametersList

RecurringChargeList

ReservedNodeList

ReservedNodeOfferingList

RestorableNodeTypeList

SnapshotCopyGrantList

SnapshotList

SourceIdsList

SourceType

SubnetIdentifierList

SubnetList

TStamp

TableRestoreStatusList

TableRestoreStatusType

TagKeyList

TagList

TagValueList

TaggedResourceList

VpcSecurityGroupIdList

VpcSecurityGroupMembershipList