Module rusoto::route53

AWS Route 53

If you're using the service, you're probably looking for Route53Client.

Structs

AlarmIdentifier
A complex type that identifies the CloudWatch alarm that you want Amazon Route 53 health checkers to use to determine whether this health check is healthy.

AlarmIdentifierSerializer

AlarmNameSerializer

AliasHealthEnabledSerializer

AliasTarget

Alias resource record sets only: Information about the CloudFront distribution, Elastic Beanstalk environment, ELB load balancer, Amazon S3 bucket, or Amazon Route 53 resource record set that you're redirecting queries to. The Elastic Beanstalk environment must have a regionalized subdomain.

When creating resource record sets for a private hosted zone, note the following:

Resource record sets can't be created for CloudFront distributions in a private hosted zone.

Creating geolocation alias resource record sets or latency alias resource record sets in a private hosted zone is unsupported.

For information about creating failover resource record sets in a private hosted zone, see Configuring Failover in a Private Hosted Zone.

AliasTargetSerializer

AssociateVPCCommentSerializer

AssociateVPCWithHostedZoneRequest
A complex type that contains information about the request to associate a VPC with a private hosted zone.

AssociateVPCWithHostedZoneResponse
A complex type that contains the response information for the AssociateVPCWithHostedZone request.

Change
The information for each resource record set that you want to change.

ChangeActionSerializer

ChangeBatch
The information for a change request.

ChangeBatchSerializer

ChangeInfo
A complex type that describes change information about changes made to your hosted zone.

ChangeResourceRecordSetsRequest
A complex type that contains change information for the resource record set.

ChangeResourceRecordSetsResponse
A complex type containing the response for the request.

ChangeSerializer

ChangeTagsForResourceRequest
A complex type that contains information about the tags that you want to add, edit, or delete.

ChangeTagsForResourceResponse
Empty response for the request.

ChangesSerializer

ChildHealthCheckListSerializer

CloudWatchAlarmConfiguration
A complex type that contains information about the CloudWatch alarm that Amazon Route 53 is monitoring for this health check.

CloudWatchRegionSerializer

CreateHealthCheckRequest
A complex type that contains the health check request information.

CreateHealthCheckResponse
A complex type containing the response information for the new health check.

CreateHostedZoneRequest
A complex type containing the hosted zone request information.

CreateHostedZoneResponse
A complex type containing the response information for the hosted zone.

CreateReusableDelegationSetRequest

CreateReusableDelegationSetResponse

CreateTrafficPolicyInstanceRequest
A complex type that contains information about the resource record sets that you want to create based on a specified traffic policy.

CreateTrafficPolicyInstanceResponse
A complex type that contains the response information for the CreateTrafficPolicyInstance request.

CreateTrafficPolicyRequest
A complex type that contains information about the traffic policy that you want to create.

CreateTrafficPolicyResponse
A complex type that contains the response information for the CreateTrafficPolicy request.

CreateTrafficPolicyVersionRequest
A complex type that contains information about the traffic policy for which you want to create a new version.

CreateTrafficPolicyVersionResponse
A complex type that contains the response information for the CreateTrafficPolicyVersion request.

CreateVPCAssociationAuthorizationRequest
A complex type that contains information about the request to authorize associating a VPC with your private hosted zone. Authorization is only required when a private hosted zone and a VPC were created by using different accounts.

CreateVPCAssociationAuthorizationResponse
A complex type that contains the response information from a CreateVPCAssociationAuthorization request.

DNSNameSerializer

DelegationSet
A complex type that describes the name servers for this hosted zone.

DeleteHealthCheckRequest
This action deletes a health check. Send a DELETE request to the /2013-04-01/DeleteHealthCheckRequest resource.

DeleteHealthCheckResponse
An empty element.

DeleteHostedZoneRequest
A complex type that contains information about the hosted zone that you want to delete.

DeleteHostedZoneResponse
A complex type containing the response information for the request.

DeleteReusableDelegationSetRequest
A complex type containing the information for the delete request.

DeleteReusableDelegationSetResponse
An empty element.

DeleteTrafficPolicyInstanceRequest
A complex type that contains information about the traffic policy instance that you want to delete.

DeleteTrafficPolicyInstanceResponse
An empty element.

DeleteTrafficPolicyRequest
A request to delete a specified traffic policy version.

DeleteTrafficPolicyResponse
An empty element.

DeleteVPCAssociationAuthorizationRequest
A complex type that contains information about the request to remove authorization to associate a VPC that was created by one AWS account with a hosted zone that was created with a different AWS account.

DeleteVPCAssociationAuthorizationResponse
Empty response for the request.

Dimension
For the metric that the CloudWatch alarm is associated with, a complex type that contains information about one dimension.

DisassociateVPCCommentSerializer

DisassociateVPCFromHostedZoneRequest
A complex type that contains information about the VPC that you want to disassociate from a specified private hosted zone.

DisassociateVPCFromHostedZoneResponse
A complex type that contains the response information for the disassociate request.

EnableSNISerializer

FailureThresholdSerializer

FullyQualifiedDomainNameSerializer

GeoLocation
A complex type that contains information about a geo location.

GeoLocationContinentCodeSerializer

GeoLocationCountryCodeSerializer

GeoLocationDetails
A complex type that contains the codes and full continent, country, and subdivision names for the specified geolocation code.

GeoLocationSerializer

GeoLocationSubdivisionCodeSerializer

GetChangeRequest
The input for a GetChange request.

GetChangeResponse
A complex type that contains the ChangeInfo element.

GetCheckerIpRangesRequest
Empty request.

GetCheckerIpRangesResponse
A complex type that contains the CheckerIpRanges element.

GetGeoLocationRequest
A complex type that contains information about the request to get a geo location.

GetGeoLocationResponse
A complex type that contains the response information for the specified geolocation code.

GetHealthCheckCountRequest
To retrieve a count of all your health checks, send a GET request to the /2013-04-01/healthcheckcount resource.

GetHealthCheckCountResponse
A complex type that contains the response to a healthcheckcount request.

GetHealthCheckLastFailureReasonRequest
This action gets the reason that a specified health check failed most recently.

To get the reason for the last failure of a health check, send a GET request to the /2013-04-01/healthcheck/health check ID/lastfailurereason resource.

For information about viewing the last failure reason for a health check using the Amazon Route 53 console, see Viewing Health Check Status and the Reason for Health Check Failures in the Amazon Route 53 Developer Guide.

GetHealthCheckLastFailureReasonResponse
A complex type that contains the response to a GetHealthCheckLastFailureReason request.

GetHealthCheckRequest
This action gets information about a specified health check.

Send a GET request to the /Amazon Route 53 API version/gethealthcheckrequest resource.

For information about getting information about a health check using the Amazon Route 53 console, see Amazon Route 53 Health Checks and DNS Failover in the Amazon Route 53 Developer Guide.

GetHealthCheckResponse
A complex type that contains the response to a GetHealthCheck request.

GetHealthCheckStatusRequest
A complex type that contains information about the request to get health check status for a health check.

GetHealthCheckStatusResponse
A complex type that contains the response to a GetHealthCheck request.

GetHostedZoneCountRequest
To retrieve a count of all your hosted zones, send a GET request to the /2013-04-01/hostedzonecount resource.

GetHostedZoneCountResponse
A complex type that contains the response to a hostedzonecount request.

GetHostedZoneRequest
The input for a GetHostedZone request.

GetHostedZoneResponse
A complex type containing the response information for the hosted zone.

GetReusableDelegationSetRequest
The input for a GetReusableDelegationSet request.

GetReusableDelegationSetResponse
A complex type that contains the response to the GetReusableDelegationSet request.

GetTrafficPolicyInstanceCountRequest
To retrieve a count of all your traffic policy instances, send a GET request to the /2013-04-01/trafficpolicyinstancecount resource.

GetTrafficPolicyInstanceCountResponse
A complex type that contains information about the resource record sets that Amazon Route 53 created based on a specified traffic policy.

GetTrafficPolicyInstanceRequest
Gets information about a specified traffic policy instance.

To get information about a traffic policy instance, send a GET request to the /Amazon Route 53 API version/trafficpolicyinstance/Id resource.

GetTrafficPolicyInstanceResponse
A complex type that contains information about the resource record sets that Amazon Route 53 created based on a specified traffic policy.

GetTrafficPolicyRequest
Gets information about a specific traffic policy version. To get the information, send a GET request to the /2013-04-01/trafficpolicy resource, and specify the ID and the version of the traffic policy.

GetTrafficPolicyResponse
A complex type that contains the response information for the request.

HealthCheck
A complex type that contains information about one health check that is associated with the current AWS account.

HealthCheckConfig
A complex type that contains information about the health check.

HealthCheckConfigSerializer

HealthCheckIdSerializer

HealthCheckNonceSerializer

HealthCheckObservation
A complex type that contains the last failure reason as reported by one Amazon Route 53 health checker.

HealthCheckRegionListSerializer

HealthCheckRegionSerializer

HealthCheckTypeSerializer

HealthCheckVersionSerializer

HealthThresholdSerializer

HostedZone
A complex type that contains general information about the hosted zone.

HostedZoneConfig
A complex type that contains an optional comment about your hosted zone. If you don't want to specify a comment, omit both the HostedZoneConfig and Comment elements.

HostedZoneConfigSerializer

IPAddressSerializer

InsufficientDataHealthStatusSerializer

InvertedSerializer

IsPrivateZoneSerializer

ListGeoLocationsRequest
To get a list of geographic locations that Amazon Route 53 supports for geolocation, send a GET request to the /Amazon Route 53 API version/geolocations resource. The response to this request includes a GeoLocationDetails element for each location that Amazon Route 53 supports.

Countries are listed first, and continents are listed last. If Amazon Route 53 supports subdivisions for a country (for example, states or provinces), the subdivisions for that country are listed in alphabetical order immediately after the corresponding country.

ListGeoLocationsResponse
A complex type containing the response information for the request.

ListHealthChecksRequest

To retrieve a list of your health checks, send a GET request to the /2013-04-01/healthcheck resource. The response to this request includes a HealthChecks element with zero or more HealthCheck child elements. By default, the list of health checks is displayed on a single page. You can control the length of the page that is displayed by using the MaxItems parameter. You can use the Marker parameter to control the health check that the list begins with.

Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Amazon Route 53 returns only the first 100.

ListHealthChecksResponse
A complex type that contains the response to a ListHealthChecks request.

ListHostedZonesByNameRequest

To retrieve a list of your public and private hosted zones in ASCII order by domain name, send a GET request to the /Amazon Route 53 API version/hostedzonesbyname resource. The response to this request includes a HostedZone child element for each hosted zone that was created by the current AWS account. ListHostedZonesByName sorts hosted zones by name with the labels reversed, for example:

com.example.www.

Note the trailing dot, which can change the sort order in some circumstances.

If the domain name includes escape characters or Punycode, ListHostedZonesByName alphabetizes the domain name using the escaped or Punycoded value, which is the format that Amazon Route 53 saves in its database. For example, to create a hosted zone for exämple.com, you specify ex\344mple.com for the domain name. ListHostedZonesByName alphabetizes it as: com.ex\344mple. The labels are reversed, and it's alphabetized using the escaped value. For more information about valid domain name formats, including internationalized domain names, see DNS Domain Name Format in the Amazon Route 53 Developer Guide.

Amazon Route 53 returns up to 100 items in each response. If you have a lot of hosted zones, you can use the MaxItems parameter to list them in groups of up to 100. The response includes values that help you navigate from one group of MaxItems hosted zones to the next:

The DNSName and HostedZoneId elements in the response contain the values, if any, that you specified for the dnsname and hostedzoneid parameters in the request that produced the current response.

The MaxItems element in the response contains the value, if any, that you specified for the maxitems parameter in the request that produced the current response.

If the value of IsTruncated in the response is true, there are more hosted zones associated with the current Amazon Route 53 account.

If IsTruncated is false, this response includes the last hosted zone that is associated with the current account. The NextDNSName element and NextHostedZoneId elements are omitted from the response.

The NextDNSName and NextHostedZoneId elements in the response contain the domain name and the hosted zone ID of the next hosted zone that is associated with the current AWS account. If you want to list more hosted zones, make another call to ListHostedZonesByName, and specify the value of NextDNSName and NextHostedZoneId in the dnsname and hostedzoneid parameters, respectively.

ListHostedZonesByNameResponse
A complex type that contains the response information for the request.

ListHostedZonesRequest

To retrieve a list of your public and private hosted zones, send a GET request to the /2013-04-01/hostedzone resource. The response to this request includes a HostedZone child element for each hosted zone that was created by the current AWS account.

Amazon Route 53 returns a maximum of 100 items in each response. If you have a lot of hosted zones, you can use the maxitems parameter to list them in groups of up to 100. The response includes four values that help you navigate from one group of maxitems hosted zones to the next:

MaxItems is the value that you specified for the maxitems parameter in the request that produced the current response.

If the value of IsTruncated in the response is true, there are more hosted zones associated with the current AWS account.

If IsTruncated is false, this response includes the last hosted zone that is associated with the current account.

NextMarker is the hosted zone ID of the next hosted zone that is associated with the current AWS account. If you want to list more hosted zones, make another call to ListHostedZones, and specify the value of the NextMarker element in the marker parameter.

If IsTruncated is false, the NextMarker element is omitted from the response.

If you're making the second or subsequent call to ListHostedZones, the Marker element matches the value that you specified in the marker parameter in the previous request.

ListHostedZonesResponse

ListResourceRecordSetsRequest
The input for a ListResourceRecordSets request.

ListResourceRecordSetsResponse
A complex type that contains list information for the resource record set.

ListReusableDelegationSetsRequest

To retrieve a list of your reusable delegation sets, send a GET request to the /2013-04-01/delegationset resource. The response to this request includes a DelegationSets element with zero or more DelegationSet child elements. By default, the list of reusable delegation sets is displayed on a single page. You can control the length of the page that is displayed by using the MaxItems parameter. You can use the Marker parameter to control the delegation set that the list begins with.

Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Amazon Route 53 returns only the first 100.

ListReusableDelegationSetsResponse
A complex type that contains information about the reusable delegation sets that are associated with the current AWS account.

ListTagsForResourceRequest
A complex type containing information about a request for a list of the tags that are associated with an individual resource.

ListTagsForResourceResponse
A complex type that contains information about the health checks or hosted zones for which you want to list tags.

ListTagsForResourcesRequest
A complex type that contains information about the health checks or hosted zones for which you want to list tags.

ListTagsForResourcesResponse
A complex type containing tags for the specified resources.

ListTrafficPoliciesRequest
A complex type that contains the information about the request to list the traffic policies that are associated with the current AWS account.

ListTrafficPoliciesResponse
A complex type that contains the response information for the request.

ListTrafficPolicyInstancesByHostedZoneRequest
A request for the traffic policy instances that you created in a specified hosted zone.

ListTrafficPolicyInstancesByHostedZoneResponse
A complex type that contains the response information for the request.

ListTrafficPolicyInstancesByPolicyRequest
A complex type that contains the information about the request to list your traffic policy instances.

ListTrafficPolicyInstancesByPolicyResponse
A complex type that contains the response information for the request.

ListTrafficPolicyInstancesRequest
A complex type that contains the information about the request to list your traffic policy instances.

ListTrafficPolicyInstancesResponse
A complex type that contains the response information for the request.

ListTrafficPolicyVersionsRequest
A complex type that contains the information about the request to list your traffic policies.

ListTrafficPolicyVersionsResponse
A complex type that contains the response information for the request.

ListVPCAssociationAuthorizationsRequest
A complex type that contains information about that can be associated with your hosted zone.

ListVPCAssociationAuthorizationsResponse
A complex type that contains the response information for the request.

MaxResultsSerializer

MeasureLatencySerializer

NonceSerializer

PageMarkerSerializer

PageMaxItemsSerializer

PaginationTokenSerializer

PortSerializer

RDataSerializer

RRTypeSerializer

RequestIntervalSerializer

ResourceDescriptionSerializer

ResourceIdSerializer

ResourcePathSerializer

ResourceRecord

Information specific to the resource record.

If you are creating an alias resource record set, omit ResourceRecord.

ResourceRecordSerializer

ResourceRecordSet
Information about the resource record set to create or delete.

ResourceRecordSetFailoverSerializer

ResourceRecordSetIdentifierSerializer

ResourceRecordSetRegionSerializer

ResourceRecordSetSerializer

ResourceRecordSetWeightSerializer

ResourceRecordsSerializer

ResourceTagSet
A complex type containing a resource and its associated tags.

Route53Client
A client for the Route 53 API.

SearchStringSerializer

StatusReport
A complex type that contains the status that one Amazon Route 53 health checker reports and the time of the health check.

SubnetMaskSerializer

TTLSerializer

Tag
A complex type that contains information about a tag that you want to add or edit for the specified health check or hosted zone.

TagKeyListSerializer

TagKeySerializer

TagListSerializer

TagResourceIdListSerializer

TagResourceIdSerializer

TagResourceTypeSerializer

TagSerializer

TagValueSerializer

TestDNSAnswerRequest

Gets the value that Amazon Route 53 returns in response to a DNS request for a specified record name and type. You can optionally specify the IP address of a DNS resolver, an EDNS0 client subnet IP address, and a subnet mask.

Parameters

hostedzoneid

The ID of the hosted zone that you want Amazon Route 53 to simulate a query for.

recordname

The name of the resource record set that you want Amazon Route 53 to simulate a query for.

recordtype

The type of the resource record set.

resolverip (optional)

If you want to simulate a request from a specific DNS resolver, specify the IP address for that resolver. If you omit this value, TestDNSAnswer uses the IP address of a DNS resolver in the AWS US East region.

edns0clientsubnetip (optional)

If the resolver that you specified for resolverip supports EDNS0, specify the IP address of a client in the applicable location.

edns0clientsubnetmask (optional)

If you specify an IP address for edns0clientsubnetip, you can optionally specify the number of bits of the IP address that you want the checking tool to include in the DNS query. For example, if you specify 192.0.2.44 for edns0clientsubnetip and 24 for edns0clientsubnetmask, the checking tool will simulate a request from 192.0.2.0/24. The default value is 24 bits.

TestDNSAnswerResponse
A complex type that contains the response to a TestDNSAnswer request.

TrafficPolicy
A complex type that contains settings for a traffic policy.

TrafficPolicyCommentSerializer

TrafficPolicyDocumentSerializer

TrafficPolicyIdSerializer

TrafficPolicyInstance
A complex type that contains settings for the new traffic policy instance.

TrafficPolicyInstanceIdSerializer

TrafficPolicyNameSerializer

TrafficPolicySummary
A complex type that contains information about the latest version of one traffic policy that is associated with the current AWS account.

TrafficPolicyVersionMarkerSerializer

TrafficPolicyVersionSerializer

UpdateHealthCheckRequest
A complex type that contains the health check request information.

UpdateHealthCheckResponse

UpdateHostedZoneCommentRequest
A complex type that contains the hosted zone request information.

UpdateHostedZoneCommentResponse
A complex type that contains the response to the UpdateHostedZoneCommentRequest.

UpdateTrafficPolicyCommentRequest
A complex type that contains information about the traffic policy for which you want to update the comment.

UpdateTrafficPolicyCommentResponse
A complex type that contains the response information for the traffic policy.

UpdateTrafficPolicyInstanceRequest
A complex type that contains information about the resource record sets that you want to update based on a specified traffic policy instance.

UpdateTrafficPolicyInstanceResponse
A complex type that contains information about the resource record sets that Amazon Route 53 created based on a specified traffic policy.

VPC
A complex type that contains information about an Amazon VPC that is associated with a private hosted zone.

VPCIdSerializer

VPCRegionSerializer

VPCSerializer

Enums

AssociateVPCWithHostedZoneError
Errors returned by AssociateVPCWithHostedZone

ChangeResourceRecordSetsError
Errors returned by ChangeResourceRecordSets

ChangeTagsForResourceError
Errors returned by ChangeTagsForResource

CreateHealthCheckError
Errors returned by CreateHealthCheck

CreateHostedZoneError
Errors returned by CreateHostedZone

CreateReusableDelegationSetError
Errors returned by CreateReusableDelegationSet

CreateTrafficPolicyError
Errors returned by CreateTrafficPolicy

CreateTrafficPolicyInstanceError
Errors returned by CreateTrafficPolicyInstance

CreateTrafficPolicyVersionError
Errors returned by CreateTrafficPolicyVersion

CreateVPCAssociationAuthorizationError
Errors returned by CreateVPCAssociationAuthorization

DeleteHealthCheckError
Errors returned by DeleteHealthCheck

DeleteHostedZoneError
Errors returned by DeleteHostedZone

DeleteReusableDelegationSetError
Errors returned by DeleteReusableDelegationSet

DeleteTrafficPolicyError
Errors returned by DeleteTrafficPolicy

DeleteTrafficPolicyInstanceError
Errors returned by DeleteTrafficPolicyInstance

DeleteVPCAssociationAuthorizationError
Errors returned by DeleteVPCAssociationAuthorization

DisassociateVPCFromHostedZoneError
Errors returned by DisassociateVPCFromHostedZone

GetChangeError
Errors returned by GetChange

GetCheckerIpRangesError
Errors returned by GetCheckerIpRanges

GetGeoLocationError
Errors returned by GetGeoLocation

GetHealthCheckCountError
Errors returned by GetHealthCheckCount

GetHealthCheckError
Errors returned by GetHealthCheck

GetHealthCheckLastFailureReasonError
Errors returned by GetHealthCheckLastFailureReason

GetHealthCheckStatusError
Errors returned by GetHealthCheckStatus

GetHostedZoneCountError
Errors returned by GetHostedZoneCount

GetHostedZoneError
Errors returned by GetHostedZone

GetReusableDelegationSetError
Errors returned by GetReusableDelegationSet

GetTrafficPolicyError
Errors returned by GetTrafficPolicy

GetTrafficPolicyInstanceCountError
Errors returned by GetTrafficPolicyInstanceCount

GetTrafficPolicyInstanceError
Errors returned by GetTrafficPolicyInstance

ListGeoLocationsError
Errors returned by ListGeoLocations

ListHealthChecksError
Errors returned by ListHealthChecks

ListHostedZonesByNameError
Errors returned by ListHostedZonesByName

ListHostedZonesError
Errors returned by ListHostedZones

ListResourceRecordSetsError
Errors returned by ListResourceRecordSets

ListReusableDelegationSetsError
Errors returned by ListReusableDelegationSets

ListTagsForResourceError
Errors returned by ListTagsForResource

ListTagsForResourcesError
Errors returned by ListTagsForResources

ListTrafficPoliciesError
Errors returned by ListTrafficPolicies

ListTrafficPolicyInstancesByHostedZoneError
Errors returned by ListTrafficPolicyInstancesByHostedZone

ListTrafficPolicyInstancesByPolicyError
Errors returned by ListTrafficPolicyInstancesByPolicy

ListTrafficPolicyInstancesError
Errors returned by ListTrafficPolicyInstances

ListTrafficPolicyVersionsError
Errors returned by ListTrafficPolicyVersions

ListVPCAssociationAuthorizationsError
Errors returned by ListVPCAssociationAuthorizations

TestDNSAnswerError
Errors returned by TestDNSAnswer

UpdateHealthCheckError
Errors returned by UpdateHealthCheck

UpdateHostedZoneCommentError
Errors returned by UpdateHostedZoneComment

UpdateTrafficPolicyCommentError
Errors returned by UpdateTrafficPolicyComment

UpdateTrafficPolicyInstanceError
Errors returned by UpdateTrafficPolicyInstance

Type Definitions

AlarmName

AliasHealthEnabled

AssociateVPCComment

ChangeAction

ChangeStatus

Changes

CheckerIpRanges

ChildHealthCheckList

CloudWatchRegion

ComparisonOperator

DNSName

DNSRCode

DelegationSetNameServers

DelegationSets

DimensionField

DimensionList

DisassociateVPCComment

EnableSNI

ErrorMessage

ErrorMessages

EvaluationPeriods

FailureThreshold

FullyQualifiedDomainName

GeoLocationContinentCode

GeoLocationContinentName

GeoLocationCountryCode

GeoLocationCountryName

GeoLocationDetailsList

GeoLocationSubdivisionCode

GeoLocationSubdivisionName

HealthCheckCount

HealthCheckId

HealthCheckNonce

HealthCheckObservations

HealthCheckRegion
An Amazon EC2 Region that you want Amazon Route 53 to use to perform health checks.

HealthCheckRegionList

HealthCheckType

HealthCheckVersion

HealthChecks

HealthThreshold

HostedZoneCount

HostedZoneRRSetCount

HostedZones

IPAddress

IPAddressCidr

InsufficientDataHealthStatus

Inverted

IsPrivateZone

MaxResults

MeasureLatency

Message

MetricName

Nameserver

Namespace

Nonce

PageMarker

PageMaxItems

PageTruncated

PaginationToken

Period

Port

RData

RRType

RecordData

RecordDataEntry

A value that Amazon Route 53 returned for this resource record set. A RecordDataEntry element is one of the following:

For non-alias resource record sets, a RecordDataEntry element contains one value in the resource record set. If the resource record set contains multiple values, the response includes one RecordDataEntry element for each value.

For multiple resource record sets that have the same name and type, which includes weighted, latency, geolocation, and failover, a RecordDataEntry element contains the value from the appropriate resource record set based on the request.

For alias resource record sets that refer to AWS resources other than another resource record set, the RecordDataEntry element contains an IP address or a domain name for the AWS resource, depending on the type of resource.

For alias resource record sets that refer to other resource record sets, a RecordDataEntry element contains one value from the referenced resource record set. If the referenced resource record set contains multiple values, the response includes one RecordDataEntry element for each value.

RequestInterval

ResourceDescription

ResourceId

ResourcePath

ResourceRecordSetFailover

ResourceRecordSetIdentifier

ResourceRecordSetRegion

ResourceRecordSetWeight

ResourceRecordSets

ResourceRecords

ResourceTagSetList

ResourceURI

SearchString

Statistic

Status

SubnetMask

TTL

TagKey

TagKeyList

TagList

TagResourceId

TagResourceIdList

TagResourceType

TagValue

Threshold

TimeStamp

TrafficPolicies

TrafficPolicyComment

TrafficPolicyDocument

TrafficPolicyId

TrafficPolicyInstanceCount

TrafficPolicyInstanceId

TrafficPolicyInstanceState

TrafficPolicyInstances

TrafficPolicyName

TrafficPolicySummaries

TrafficPolicyVersion

TrafficPolicyVersionMarker

TransportProtocol

VPCId
The ID of an Amazon VPC.

VPCRegion

VPCs
A list of VPC elements.