Struct rusoto::route53::ListHostedZonesByNameRequest

pub struct ListHostedZonesByNameRequest {
    pub dns_name: Option<DNSName>,
    pub hosted_zone_id: Option<ResourceId>,
    pub max_items: Option<PageMaxItems>,
}

To retrieve a list of your public and private hosted zones in ASCII order by domain name, send a GET request to the /Amazon Route 53 API version/hostedzonesbyname resource. The response to this request includes a HostedZone child element for each hosted zone that was created by the current AWS account. ListHostedZonesByName sorts hosted zones by name with the labels reversed, for example:

com.example.www.

Note the trailing dot, which can change the sort order in some circumstances.

If the domain name includes escape characters or Punycode, ListHostedZonesByName alphabetizes the domain name using the escaped or Punycoded value, which is the format that Amazon Route 53 saves in its database. For example, to create a hosted zone for exämple.com, you specify ex\344mple.com for the domain name. ListHostedZonesByName alphabetizes it as: com.ex\344mple. The labels are reversed, and it's alphabetized using the escaped value. For more information about valid domain name formats, including internationalized domain names, see DNS Domain Name Format in the Amazon Route 53 Developer Guide.

Amazon Route 53 returns up to 100 items in each response. If you have a lot of hosted zones, you can use the MaxItems parameter to list them in groups of up to 100. The response includes values that help you navigate from one group of MaxItems hosted zones to the next:

The DNSName and HostedZoneId elements in the response contain the values, if any, that you specified for the dnsname and hostedzoneid parameters in the request that produced the current response.

The MaxItems element in the response contains the value, if any, that you specified for the maxitems parameter in the request that produced the current response.

If the value of IsTruncated in the response is true, there are more hosted zones associated with the current Amazon Route 53 account.

If IsTruncated is false, this response includes the last hosted zone that is associated with the current account. The NextDNSName element and NextHostedZoneId elements are omitted from the response.

The NextDNSName and NextHostedZoneId elements in the response contain the domain name and the hosted zone ID of the next hosted zone that is associated with the current AWS account. If you want to list more hosted zones, make another call to ListHostedZonesByName, and specify the value of NextDNSName and NextHostedZoneId in the dnsname and hostedzoneid parameters, respectively.

Fields
dns_name: Option<DNSName>

(Optional) For your first request to ListHostedZonesByName, include the dnsname parameter only if you want to specify the name of the first hosted zone in the response. If you don't include the dnsname parameter, Amazon Route 53 returns all of the hosted zones that were created by the current AWS account, in ASCII order. For subsequent requests, include both dnsname and hostedzoneid parameters. For dnsname, specify the value of NextDNSName from the previous response.

hosted_zone_id: Option<ResourceId>

(Optional) For your first request to ListHostedZonesByName, do not include the hostedzoneid parameter.

If you have more hosted zones than the value of maxitems, ListHostedZonesByName returns only the first maxitems hosted zones. To get the next group of maxitems hosted zones, submit another request to ListHostedZonesByName and include both dnsname and hostedzoneid parameters. For the value of hostedzoneid, specify the value of the NextHostedZoneId element from the previous response.

max_items: Option<PageMaxItems>

The maximum number of hosted zones to be included in the response body for this request. If you have more than maxitems hosted zones, then the value of the IsTruncated element in the response is true, and the values of NextDNSName and NextHostedZoneId specify the first hosted zone in the next group of maxitems hosted zones.