Struct rusoto::route53::ListHostedZonesRequest

pub struct ListHostedZonesRequest {
    pub delegation_set_id: Option<ResourceId>,
    pub marker: Option<PageMarker>,
    pub max_items: Option<PageMaxItems>,
}

To retrieve a list of your public and private hosted zones, send a GET request to the /2013-04-01/hostedzone resource. The response to this request includes a HostedZone child element for each hosted zone that was created by the current AWS account.

Amazon Route 53 returns a maximum of 100 items in each response. If you have a lot of hosted zones, you can use the maxitems parameter to list them in groups of up to 100. The response includes four values that help you navigate from one group of maxitems hosted zones to the next:

MaxItems is the value that you specified for the maxitems parameter in the request that produced the current response.

If the value of IsTruncated in the response is true, there are more hosted zones associated with the current AWS account.

If IsTruncated is false, this response includes the last hosted zone that is associated with the current account.

NextMarker is the hosted zone ID of the next hosted zone that is associated with the current AWS account. If you want to list more hosted zones, make another call to ListHostedZones, and specify the value of the NextMarker element in the marker parameter.

If IsTruncated is false, the NextMarker element is omitted from the response.

If you're making the second or subsequent call to ListHostedZones, the Marker element matches the value that you specified in the marker parameter in the previous request.

Fields
delegation_set_id: Option<ResourceId>

If you're using reusable delegation sets and you want to list all of the hosted zones that are associated with a reusable delegation set, specify the ID of that reusable delegation set.

marker: Option<PageMarker>

(Optional) If you have more hosted zones than the value of maxitems, ListHostedZones returns only the first maxitems hosted zones. To get the next group of maxitems hosted zones, submit another request to ListHostedZones. For the value of marker, specify the value of the NextMarker element that was returned in the previous response.

Hosted zones are listed in the order in which they were created.

max_items: Option<PageMaxItems>

(Optional) The maximum number of hosted zones to be included in the response body for this request. If you have more than maxitems hosted zones, the value of the IsTruncated element in the response is true, and the value of the NextMarker element is the hosted zone ID of the first hosted zone in the next group of maxitems hosted zones.