Struct rusoto::route53::Route53Client

pub struct Route53Client<P, D> where P: ProvideAwsCredentials,
        D: DispatchSignedRequest { /* fields omitted */ }

A client for the Route 53 API.

Methods

impl<P, D> Route53Client<P, D> where P: ProvideAwsCredentials,
D: DispatchSignedRequest

fn new(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn associate_vpc_with_hosted_zone(&self,
input: &AssociateVPCWithHostedZoneRequest)
-> Result<AssociateVPCWithHostedZoneResponse, AssociateVPCWithHostedZoneError>

Associates an Amazon VPC with a private hosted zone.

To perform the association, the VPC and the private hosted zone must already exist. You can't convert a public hosted zone into a private hosted zone.

Send a POST request to the /2013-04-01/hostedzone/hosted zone ID/associatevpc resource. The request body must include a document with an AssociateVPCWithHostedZoneRequest element. The response contains a ChangeInfo data type that you can use to track the progress of the request.

If you want to associate a VPC that was created by using one AWS account with a private hosted zone that was created by using a different account, the AWS account that created the private hosted zone must first submit a CreateVPCAssociationAuthorization request. Then the account that created the VPC must submit an AssociateVPCWithHostedZone request.

fn change_resource_record_sets(&self,
input: &ChangeResourceRecordSetsRequest)
-> Result<ChangeResourceRecordSetsResponse, ChangeResourceRecordSetsError>

Create, change, update, or delete authoritative DNS information on all Amazon Route 53 servers. Send a POST request to:

/2013-04-01/hostedzone/Amazon Route 53 hosted Zone ID/rrset resource.

The request body must include a document with a ChangeResourceRecordSetsRequest element. The request body contains a list of change items, known as a change batch. Change batches are considered transactional changes. When using the Amazon Route 53 API to change resource record sets, Amazon Route 53 either makes all or none of the changes in a change batch request. This ensures that Amazon Route 53 never partially implements the intended changes to the resource record sets in a hosted zone.

For example, a change batch request that deletes the CNAME record for www.example.com and creates an alias resource record set for www.example.com. Amazon Route 53 deletes the first resource record set and creates the second resource record set in a single operation. If either the DELETE or the CREATE action fails, then both changes (plus any other changes in the batch) fail, and the original CNAME record continues to exist.

Due to the nature of transactional changes, you can't delete the same resource record set more than once in a single change batch. If you attempt to delete the same change batch more than once, Amazon Route 53 returns an InvalidChangeBatch error.

To create resource record sets for complex routing configurations, use either the traffic flow visual editor in the Amazon Route 53 console or the API actions for traffic policies and traffic policy instances. Save the configuration as a traffic policy, then associate the traffic policy with one or more domain names (such as example.com) or subdomain names (such as www.example.com), in the same hosted zone or in multiple hosted zones. You can roll back the updates if the new configuration isn't performing as expected. For more information, see Using Traffic Flow to Route DNS Traffic in the Amazon Route 53 Developer Guide.

Use ChangeResourceRecordsSetsRequest to perform the following actions:

CREATE: Creates a resource record set that has the specified values.

DELETE: Deletes an existing resource record set that has the specified values.

UPSERT: If a resource record set does not already exist, AWS creates it. If a resource set does exist, Amazon Route 53 updates it with the values in the request.

The values that you need to include in the request depend on the type of resource record set that you're creating, deleting, or updating:

Basic resource record sets (excluding alias, failover, geolocation, latency, and weighted resource record sets)

Name

Type

TTL

Failover, geolocation, latency, or weighted resource record sets (excluding alias resource record sets)

Name

Type

TTL

SetIdentifier

Alias resource record sets (including failover alias, geolocation alias, latency alias, and weighted alias resource record sets)

Name

Type

AliasTarget (includes DNSName, EvaluateTargetHealth, and HostedZoneId)

SetIdentifier (for failover, geolocation, latency, and weighted resource record sets)

When you submit a ChangeResourceRecordSets request, Amazon Route 53 propagates your changes to all of the Amazon Route 53 authoritative DNS servers. While your changes are propagating, GetChange returns a status of PENDING. When propagation is complete, GetChange returns a status of INSYNC. Changes generally propagate to all Amazon Route 53 name servers in a few minutes. In rare circumstances, propagation can take up to 30 minutes. For more information, see GetChange

For information about the limits on a ChangeResourceRecordSets request, see Limits in the Amazon Route 53 Developer Guide.

fn change_tags_for_resource(&self,
input: &ChangeTagsForResourceRequest)
-> Result<ChangeTagsForResourceResponse, ChangeTagsForResourceError>

Adds, edits, or deletes tags for a health check or a hosted zone.

For information about using tags for cost allocation, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide.

fn create_health_check(&self,
input: &CreateHealthCheckRequest)
-> Result<CreateHealthCheckResponse, CreateHealthCheckError>

Creates a new health check.

To create a new health check, send a POST request to the /2013-04-01/healthcheck resource. The request body must include a document with a CreateHealthCheckRequest element. The response returns the CreateHealthCheckResponse element, containing the health check ID specified when adding health check to a resource record set. For information about adding health checks to resource record sets, see ResourceRecordSet$HealthCheckId in ChangeResourceRecordSets.

If you are registering EC2 instances with an Elastic Load Balancing (ELB) load balancer, do not create Amazon Route 53 health checks for the EC2 instances. When you register an EC2 instance with a load balancer, you configure settings for an ELB health check, which performs a similar function to an Amazon Route 53 health check.

You can associate health checks with failover resource record sets in a private hosted zone. Note the following:

Amazon Route 53 health checkers are outside the VPC. To check the health of an endpoint within a VPC by IP address, you must assign a public IP address to the instance in the VPC.

You can configure a health checker to check the health of an external resource that the instance relies on, such as a database server.

You can create a CloudWatch metric, associate an alarm with the metric, and then create a health check that is based on the state of the alarm. For example, you might create a CloudWatch metric that checks the status of the Amazon EC2 StatusCheckFailed metric, add an alarm to the metric, and then create a health check that is based on the state of the alarm. For information about creating CloudWatch metrics and alarms by using the CloudWatch console, see the Amazon CloudWatch User Guide.

fn create_hosted_zone(&self,
input: &CreateHostedZoneRequest)
-> Result<CreateHostedZoneResponse, CreateHostedZoneError>

Creates a new public hosted zone, used to specify how the Domain Name System (DNS) routes traffic on the Internet for a domain, such as example.com, and its subdomains.

Public hosted zones can't be converted to a private hosted zone or vice versa. Instead, create a new hosted zone with the same name and create new resource record sets.

Send a POST request to the /2013-04-01/hostedzone resource. The request body must include a document with a CreateHostedZoneRequest element. The response returns the CreateHostedZoneResponse element containing metadata about the hosted zone.

Fore more information about charges for hosted zones, see Amazon Route 53 Pricing.

Note the following:

You can't create a hosted zone for a top-level domain (TLD).

Amazon Route 53 automatically creates a default SOA record and four NS records for the zone. For more information about SOA and NS records, see NS and SOA Records that Amazon Route 53 Creates for a Hosted Zone in the Amazon Route 53 Developer Guide.

If your domain is registered with a registrar other than Amazon Route 53, you must update the name servers with your registrar to make Amazon Route 53 your DNS service. For more information, see Configuring Amazon Route 53 as your DNS Service in the Amazon Route 53 Developer's Guide.

After creating a zone, its initial status is PENDING. This means that it is not yet available on all DNS servers. The status of the zone changes to INSYNC when the NS and SOA records are available on all Amazon Route 53 DNS servers.

When trying to create a hosted zone using a reusable delegation set, specify an optional DelegationSetId, and Amazon Route 53 would assign those 4 NS records for the zone, instead of allotting a new one.

fn create_reusable_delegation_set(&self,
input: &CreateReusableDelegationSetRequest)
-> Result<CreateReusableDelegationSetResponse, CreateReusableDelegationSetError>

Creates a delegation set (a group of four name servers) that can be reused by multiple hosted zones. If a hosted zoned ID is specified, CreateReusableDelegationSet marks the delegation set associated with that zone as reusable

Send a POST request to the /2013-04-01/delegationset resource. The request body must include a document with a CreateReusableDelegationSetRequest element.

A reusable delegation set can't be associated with a private hosted zone/

For more information, including a procedure on how to create and configure a reusable delegation set (also known as white label name servers), see Configuring White Label Name Servers.

fn create_traffic_policy(&self,
input: &CreateTrafficPolicyRequest)
-> Result<CreateTrafficPolicyResponse, CreateTrafficPolicyError>

Creates a traffic policy, which you use to create multiple DNS resource record sets for one domain name (such as example.com) or one subdomain name (such as www.example.com).

Send a POST request to the /2013-04-01/trafficpolicy resource. The request body must include a document with a CreateTrafficPolicyRequest element. The response includes the CreateTrafficPolicyResponse element, which contains information about the new traffic policy.

fn create_traffic_policy_instance(&self,
input: &CreateTrafficPolicyInstanceRequest)
-> Result<CreateTrafficPolicyInstanceResponse, CreateTrafficPolicyInstanceError>

Creates resource record sets in a specified hosted zone based on the settings in a specified traffic policy version. In addition, CreateTrafficPolicyInstance associates the resource record sets with a specified domain name (such as example.com) or subdomain name (such as www.example.com). Amazon Route 53 responds to DNS queries for the domain or subdomain name by using the resource record sets that CreateTrafficPolicyInstance created.

Send a POST request to the /2013-04-01/trafficpolicyinstance resource. The request body must include a document with a CreateTrafficPolicyRequest element. The response returns the CreateTrafficPolicyInstanceResponse element, which contains information about the traffic policy instance.

fn create_traffic_policy_version(&self,
input: &CreateTrafficPolicyVersionRequest)
-> Result<CreateTrafficPolicyVersionResponse, CreateTrafficPolicyVersionError>

Creates a new version of an existing traffic policy. When you create a new version of a traffic policy, you specify the ID of the traffic policy that you want to update and a JSON-formatted document that describes the new version. You use traffic policies to create multiple DNS resource record sets for one domain name (such as example.com) or one subdomain name (such as www.example.com). You can create a maximum of 1000 versions of a traffic policy. If you reach the limit and need to create another version, you'll need to start a new traffic policy.

Send a POST request to the /2013-04-01/trafficpolicy/ resource. The request body includes a document with a CreateTrafficPolicyVersionRequest element. The response returns the CreateTrafficPolicyVersionResponse element, which contains information about the new version of the traffic policy.

fn create_vpc_association_authorization(&self,
input: &CreateVPCAssociationAuthorizationRequest)
-> Result<CreateVPCAssociationAuthorizationResponse, CreateVPCAssociationAuthorizationError>

Authorizes the AWS account that created a specified VPC to submit an AssociateVPCWithHostedZone request to associate the VPC with a specified hosted zone that was created by a different account. To submit a CreateVPCAssociationAuthorization request, you must use the account that created the hosted zone. After you authorize the association, use the account that created the VPC to submit an AssociateVPCWithHostedZone request.

If you want to associate multiple VPCs that you created by using one account with a hosted zone that you created by using a different account, you must submit one authorization request for each VPC.

Send a POST request to the /2013-04-01/hostedzone/hosted zone ID/authorizevpcassociation resource. The request body must include a document with a CreateVPCAssociationAuthorizationRequest element. The response contains information about the authorization.

fn delete_health_check(&self,
input: &DeleteHealthCheckRequest)
-> Result<DeleteHealthCheckResponse, DeleteHealthCheckError>

Deletes a health check. Send a DELETE request to the /2013-04-01/healthcheck/health check ID resource.

Amazon Route 53 does not prevent you from deleting a health check even if the health check is associated with one or more resource record sets. If you delete a health check and you don't update the associated resource record sets, the future status of the health check can't be predicted and may change. This will affect the routing of DNS queries for your DNS failover configuration. For more information, see Replacing and Deleting Health Checks in the Amazon Route 53 Developer Guide.

fn delete_hosted_zone(&self,
input: &DeleteHostedZoneRequest)
-> Result<DeleteHostedZoneResponse, DeleteHostedZoneError>

Deletes a hosted zone. Send a DELETE request to the /Amazon Route 53 API version/hostedzone/hosted zone ID resource.

Delete a hosted zone only if there are no resource record sets other than the default SOA record and NS resource record sets. If the hosted zone contains other resource record sets, delete them before deleting the hosted zone. If you try to delete a hosted zone that contains other resource record sets, Amazon Route 53 denies your request with a HostedZoneNotEmpty error. For information about deleting records from your hosted zone, see ChangeResourceRecordSets.

fn delete_reusable_delegation_set(&self,
input: &DeleteReusableDelegationSetRequest)
-> Result<DeleteReusableDelegationSetResponse, DeleteReusableDelegationSetError>

Deletes a reusable delegation set. Send a DELETE request to the /2013-04-01/delegationset/delegation set ID resource.

You can delete a reusable delegation set only if there are no associated hosted zones.

To verify that the reusable delegation set is not associated with any hosted zones, run the GetReusableDelegationSet action and specify the ID of the reusable delegation set that you want to delete.

fn delete_traffic_policy(&self,
input: &DeleteTrafficPolicyRequest)
-> Result<DeleteTrafficPolicyResponse, DeleteTrafficPolicyError>

Deletes a traffic policy.

Send a DELETE request to the /Amazon Route 53 API version/trafficpolicy resource.

fn delete_traffic_policy_instance(&self,
input: &DeleteTrafficPolicyInstanceRequest)
-> Result<DeleteTrafficPolicyInstanceResponse, DeleteTrafficPolicyInstanceError>

Deletes a traffic policy instance and all of the resource record sets that Amazon Route 53 created when you created the instance.

Send a DELETE request to the /Amazon Route 53 API version/trafficpolicy/traffic policy instance ID resource.

In the Amazon Route 53 console, traffic policy instances are known as policy records.

fn delete_vpc_association_authorization(&self,
input: &DeleteVPCAssociationAuthorizationRequest)
-> Result<DeleteVPCAssociationAuthorizationResponse, DeleteVPCAssociationAuthorizationError>

Removes authorization to submit an AssociateVPCWithHostedZone request to associate a specified VPC with a hosted zone that was created by a different account. You must use the account that created the hosted zone to submit a DeleteVPCAssociationAuthorization request.

Sending this request only prevents the AWS account that created the VPC from associating the VPC with the Amazon Route 53 hosted zone in the future. If the VPC is already associated with the hosted zone, DeleteVPCAssociationAuthorization won't disassociate the VPC from the hosted zone. If you want to delete an existing association, use DisassociateVPCFromHostedZone.

Send a DELETE request to the /2013-04-01/hostedzone/hosted zone ID/deauthorizevpcassociation resource. The request body must include a document with a DeleteVPCAssociationAuthorizationRequest element.

fn disassociate_vpc_from_hosted_zone(&self,
input: &DisassociateVPCFromHostedZoneRequest)
-> Result<DisassociateVPCFromHostedZoneResponse, DisassociateVPCFromHostedZoneError>

Disassociates a VPC from a Amazon Route 53 private hosted zone.

You can't disassociate the last VPC from a private hosted zone.

Send a POST request to the /2013-04-01/hostedzone/hosted zone ID/disassociatevpc resource. The request body must include a document with a DisassociateVPCFromHostedZoneRequest element. The response includes a DisassociateVPCFromHostedZoneResponse element.

You can't disassociate a VPC from a private hosted zone when only one VPC is associated with the hosted zone. You also can't convert a private hosted zone into a public hosted zone.

fn get_change(&self,
input: &GetChangeRequest)
-> Result<GetChangeResponse, GetChangeError>

Returns the current status of a change batch request. The status is one of the following values:

PENDING indicates that the changes in this request have not replicated to all Amazon Route 53 DNS servers. This is the initial status of all change batch requests.

INSYNC indicates that the changes have replicated to all Amazon Route 53 DNS servers.

fn get_checker_ip_ranges(&self,
input: &GetCheckerIpRangesRequest)
-> Result<GetCheckerIpRangesResponse, GetCheckerIpRangesError>

Retrieves a list of the IP ranges used by Amazon Route 53 health checkers to check the health of your resources. Send a GET request to the /Amazon Route 53 API version/checkeripranges resource. Use these IP addresses to configure router and firewall rules to allow health checkers to check the health of your resources.

fn get_geo_location(&self,
input: &GetGeoLocationRequest)
-> Result<GetGeoLocationResponse, GetGeoLocationError>

Retrieves a single geo location. Send a GET request to the /2013-04-01/geolocation resource with one of these options: continentcode | countrycode | countrycode and subdivisioncode.

fn get_health_check(&self,
input: &GetHealthCheckRequest)
-> Result<GetHealthCheckResponse, GetHealthCheckError>

Gets information about a specified health check. Send a GET request to the /2013-04-01/healthcheck/health check ID resource. For more information about using the console to perform this operation, see Amazon Route 53 Health Checks and DNS Failover in the Amazon Route 53 Developer Guide.

fn get_health_check_count(&self,
input: &GetHealthCheckCountRequest)
-> Result<GetHealthCheckCountResponse, GetHealthCheckCountError>

To retrieve a count of all your health checks, send a GET request to the /2013-04-01/healthcheckcount resource.

fn get_health_check_last_failure_reason(&self,
input: &GetHealthCheckLastFailureReasonRequest)
-> Result<GetHealthCheckLastFailureReasonResponse, GetHealthCheckLastFailureReasonError>

If you want to learn why a health check is currently failing or why it failed most recently (if at all), you can get the failure reason for the most recent failure. Send a GET request to the /Amazon Route 53 API version/healthcheck/health check ID/lastfailurereason resource.

fn get_health_check_status(&self,
input: &GetHealthCheckStatusRequest)
-> Result<GetHealthCheckStatusResponse, GetHealthCheckStatusError>

Gets status of a specified health check. Send a GET request to the /2013-04-01/healthcheck/health check ID/status resource. You can use this call to get a health check's current status.

fn get_hosted_zone(&self,
input: &GetHostedZoneRequest)
-> Result<GetHostedZoneResponse, GetHostedZoneError>

Retrieves the delegation set for a hosted zone, including the four name servers assigned to the hosted zone. Send a GET request to the /Amazon Route 53 API version/hostedzone/hosted zone ID resource.

fn get_hosted_zone_count(&self,
input: &GetHostedZoneCountRequest)
-> Result<GetHostedZoneCountResponse, GetHostedZoneCountError>

Retrieves a count of all your hosted zones. Send a GET request to the /2013-04-01/hostedzonecount resource.

fn get_reusable_delegation_set(&self,
input: &GetReusableDelegationSetRequest)
-> Result<GetReusableDelegationSetResponse, GetReusableDelegationSetError>

Retrieves the reusable delegation set. Send a GET request to the /2013-04-01/delegationset/delegation set ID resource.

fn get_traffic_policy(&self,
input: &GetTrafficPolicyRequest)
-> Result<GetTrafficPolicyResponse, GetTrafficPolicyError>

Gets information about a specific traffic policy version.

Send a GET request to the /Amazon Route 53 API version/trafficpolicy resource.

fn get_traffic_policy_instance(&self,
input: &GetTrafficPolicyInstanceRequest)
-> Result<GetTrafficPolicyInstanceResponse, GetTrafficPolicyInstanceError>

Gets information about a specified traffic policy instance.

Send a GET request to the /Amazon Route 53 API version/trafficpolicyinstance resource.

After you submit a CreateTrafficPolicyInstance or an UpdateTrafficPolicyInstance request, there's a brief delay while Amazon Route 53 creates the resource record sets that are specified in the traffic policy definition. For more information, see the State response element.

In the Amazon Route 53 console, traffic policy instances are known as policy records.

fn get_traffic_policy_instance_count(&self,
input: &GetTrafficPolicyInstanceCountRequest)
-> Result<GetTrafficPolicyInstanceCountResponse, GetTrafficPolicyInstanceCountError>

Gets the number of traffic policy instances that are associated with the current AWS account.

To get the number of traffic policy instances, send a GET request to the /2013-04-01/trafficpolicyinstancecount resource.

fn list_geo_locations(&self,
input: &ListGeoLocationsRequest)
-> Result<ListGeoLocationsResponse, ListGeoLocationsError>

Retrieves a list of supported geo locations. Send a GET request to the /2013-04-01/geolocations resource. The response to this request includes a GeoLocationDetailsList element for each location that Amazon Route 53 supports.

Countries are listed first, and continents are listed last. If Amazon Route 53 supports subdivisions for a country (for example, states or provinces), the subdivisions for that country are listed in alphabetical order immediately after the corresponding country.

fn list_health_checks(&self,
input: &ListHealthChecksRequest)
-> Result<ListHealthChecksResponse, ListHealthChecksError>

Retrieve a list of your health checks. Send a GET request to the /2013-04-01/healthcheck resource. The response to this request includes a HealthChecks element with zero or more HealthCheck child elements. By default, the list of health checks is displayed on a single page. You can control the length of the page that is displayed by using the MaxItems parameter. You can use the Marker parameter to control the health check that the list begins with.

For information about listing health checks using the Amazon Route 53 console, see Amazon Route 53 Health Checks and DNS Failover.

fn list_hosted_zones(&self,
input: &ListHostedZonesRequest)
-> Result<ListHostedZonesResponse, ListHostedZonesError>

To retrieve a list of your public and private hosted zones, send a GET request to the /2013-04-01/hostedzone resource. The response to this request includes a HostedZones child element for each hosted zone created by the current AWS account.

Amazon Route 53 returns a maximum of 100 items in each response. If you have a lot of hosted zones, you can use the maxitems parameter to list them in groups of up to 100. The response includes four values that help navigate from one group of maxitems hosted zones to the next:

MaxItems is the value specified for the maxitems parameter in the request that produced the current response.

If the value of IsTruncated in the response is true, there are more hosted zones associated with the current AWS account.

NextMarker is the hosted zone ID of the next hosted zone that is associated with the current AWS account. If you want to list more hosted zones, make another call to ListHostedZones, and specify the value of the NextMarker element in the marker parameter.

If IsTruncated is false, the NextMarker element is omitted from the response.

If you're making the second or subsequent call to ListHostedZones, the Marker element matches the value that you specified in the marker parameter in the previous request.

fn list_hosted_zones_by_name(&self,
input: &ListHostedZonesByNameRequest)
-> Result<ListHostedZonesByNameResponse, ListHostedZonesByNameError>

Retrieves a list of your hosted zones in lexicographic order. Send a GET request to the /2013-04-01/hostedzonesbyname resource. The response includes a HostedZones child element for each hosted zone created by the current AWS account.

ListHostedZonesByName sorts hosted zones by name with the labels reversed. For example:

com.example.www.

Note the trailing dot, which can change the sort order in some circumstances.

If the domain name includes escape characters or Punycode, ListHostedZonesByName alphabetizes the domain name using the escaped or Punycoded value, which is the format that Amazon Route 53 saves in its database. For example, to create a hosted zone for example.com, specify ex\344mple.com for the domain name. ListHostedZonesByName alphabetizes it as:

com.ex\344mple.

The labels are reversed and alphabetized using the escaped value. For more information about valid domain name formats, including internationalized domain names, see DNS Domain Name Format in the Amazon Route 53 Developer Guide.

Amazon Route 53 returns up to 100 items in each response. If you have a lot of hosted zones, use the MaxItems parameter to list them in groups of up to 100. The response includes values that help navigate from one group of MaxItems hosted zones to the next:

The DNSName and HostedZoneId elements in the response contain the values, if any, specified for the dnsname and hostedzoneid parameters in the request that produced the current response.

The MaxItems element in the response contains the value, if any, that you specified for the maxitems parameter in the request that produced the current response.

If the value of IsTruncated in the response is true, there are more hosted zones associated with the current AWS account.

If IsTruncated is false, this response includes the last hosted zone that is associated with the current account. The NextDNSName element and NextHostedZoneId elements are omitted from the response.

The NextDNSName and NextHostedZoneId elements in the response contain the domain name and the hosted zone ID of the next hosted zone that is associated with the current AWS account. If you want to list more hosted zones, make another call to ListHostedZonesByName, and specify the value of NextDNSName and NextHostedZoneId in the dnsname and hostedzoneid parameters, respectively.

fn list_resource_record_sets(&self,
input: &ListResourceRecordSetsRequest)
-> Result<ListResourceRecordSetsResponse, ListResourceRecordSetsError>

Lists the resource record sets in a specified hosted zone.

ListResourceRecordSets returns up to 100 resource record sets at a time in ASCII order, beginning at a position specified by the name and type elements. The action sorts results first by DNS name with the labels reversed, for example:

com.example.www.

Note the trailing dot, which can change the sort order in some circumstances.

When multiple records have the same DNS name, the action sorts results by the record type.

You can use the name and type elements to adjust the beginning position of the list of resource record sets returned:

If you do not specify Name or Type

The results begin with the first resource record set that the hosted zone contains.

If you specify Name but not Type

The results begin with the first resource record set in the list whose name is greater than or equal to Name.

If you specify Type but not Name

Amazon Route 53 returns the InvalidInput error.

If you specify both Name and Type

The results begin with the first resource record set in the list whose name is greater than or equal to Name, and whose type is greater than or equal to Type.

This action returns the most current version of the records. This includes records that are PENDING, and that are not yet available on all Amazon Route 53 DNS servers.

To ensure that you get an accurate listing of the resource record sets for a hosted zone at a point in time, do not submit a ChangeResourceRecordSets request while you're paging through the results of a ListResourceRecordSets request. If you do, some pages may display results without the latest changes while other pages display results with the latest changes.

fn list_reusable_delegation_sets(&self,
input: &ListReusableDelegationSetsRequest)
-> Result<ListReusableDelegationSetsResponse, ListReusableDelegationSetsError>

To retrieve a list of your reusable delegation sets, send a GET request to the /2013-04-01/delegationset resource. The response to this request includes a DelegationSets element with zero, one, or multiple DelegationSet child elements. By default, the list of delegation sets is displayed on a single page. You can control the length of the page that is displayed by using the MaxItems parameter. You can use the Marker parameter to control the delegation set that the list begins with.

Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Amazon Route 53 returns only the first 100.

fn list_tags_for_resource(&self,
input: &ListTagsForResourceRequest)
-> Result<ListTagsForResourceResponse, ListTagsForResourceError>

Lists tags for one health check or hosted zone.

For information about using tags for cost allocation, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide.

fn list_tags_for_resources(&self,
input: &ListTagsForResourcesRequest)
-> Result<ListTagsForResourcesResponse, ListTagsForResourcesError>

Lists tags for up to 10 health checks or hosted zones.

For information about using tags for cost allocation, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide.

fn list_traffic_policies(&self,
input: &ListTrafficPoliciesRequest)
-> Result<ListTrafficPoliciesResponse, ListTrafficPoliciesError>

Gets information about the latest version for every traffic policy that is associated with the current AWS account. Send a GET request to the /Amazon Route 53 API version/trafficpolicy resource.

Amazon Route 53 returns a maximum of 100 items in each response. If you have a lot of traffic policies, you can use the maxitems parameter to list them in groups of up to 100.

The response includes three values that help you navigate from one group of maxitems traffic policies to the next:

IsTruncated

If the value of IsTruncated in the response is true, there are more traffic policies associated with the current AWS account.

If IsTruncated is false, this response includes the last traffic policy that is associated with the current account.

TrafficPolicyIdMarker

If IsTruncated is true, TrafficPolicyIdMarker is the ID of the first traffic policy in the next group of MaxItems traffic policies. If you want to list more traffic policies, make another call to ListTrafficPolicies, and specify the value of the TrafficPolicyIdMarker element from the response in the TrafficPolicyIdMarker request parameter.

If IsTruncated is false, the TrafficPolicyIdMarker element is omitted from the response.

MaxItems

The value that you specified for the MaxItems parameter in the request that produced the current response.

fn list_traffic_policy_instances(&self,
input: &ListTrafficPolicyInstancesRequest)
-> Result<ListTrafficPolicyInstancesResponse, ListTrafficPolicyInstancesError>

Gets information about the traffic policy instances that you created by using the current AWS account.

After you submit an UpdateTrafficPolicyInstance request, there's a brief delay while Amazon Route 53 creates the resource record sets that are specified in the traffic policy definition. For more information, see the State response element.

Send a GET request to the /Amazon Route 53 API version/trafficpolicyinstance resource.

Amazon Route 53 returns a maximum of 100 items in each response. If you have a lot of traffic policy instances, you can use the MaxItems parameter to list them in groups of up to 100.

The response includes five values that help you navigate from one group of MaxItems traffic policy instances to the next:

IsTruncated

If the value of IsTruncated in the response is true, there are more traffic policy instances associated with the current AWS account.

If IsTruncated is false, this response includes the last traffic policy instance that is associated with the current account.

MaxItems

The value that you specified for the MaxItems parameter in the request that produced the current response.

HostedZoneIdMarker, TrafficPolicyInstanceNameMarker, and TrafficPolicyInstanceTypeMarker

If IsTruncated is true, these three values in the response represent the first traffic policy instance in the next group of MaxItems traffic policy instances. To list more traffic policy instances, make another call to ListTrafficPolicyInstances, and specify these values in the corresponding request parameters.

If IsTruncated is false, all three elements are omitted from the response.

fn list_traffic_policy_instances_by_hosted_zone(&self,
input: &ListTrafficPolicyInstancesByHostedZoneRequest)
-> Result<ListTrafficPolicyInstancesByHostedZoneResponse, ListTrafficPolicyInstancesByHostedZoneError>

Gets information about the traffic policy instances that you created in a specified hosted zone.

After you submit an UpdateTrafficPolicyInstance request, there's a brief delay while Amazon Route 53 creates the resource record sets that are specified in the traffic policy definition. For more information, see the State response element.

Send a GET request to the /Amazon Route 53 API version/trafficpolicyinstance resource and include the ID of the hosted zone.

Amazon Route 53 returns a maximum of 100 items in each response. If you have a lot of traffic policy instances, you can use the MaxItems parameter to list them in groups of up to 100.

The response includes four values that help you navigate from one group of MaxItems traffic policy instances to the next:

IsTruncated

If the value of IsTruncated in the response is true, there are more traffic policy instances associated with the current AWS account.

If IsTruncated is false, this response includes the last traffic policy instance that is associated with the current account.

MaxItems

The value that you specified for the MaxItems parameter in the request that produced the current response.

TrafficPolicyInstanceNameMarker and TrafficPolicyInstanceTypeMarker

If IsTruncated is true, these two values in the response represent the first traffic policy instance in the next group of MaxItems traffic policy instances. To list more traffic policy instances, make another call to ListTrafficPolicyInstancesByHostedZone, and specify these values in the corresponding request parameters.

If IsTruncated is false, all three elements are omitted from the response.

fn list_traffic_policy_instances_by_policy(&self,
input: &ListTrafficPolicyInstancesByPolicyRequest)
-> Result<ListTrafficPolicyInstancesByPolicyResponse, ListTrafficPolicyInstancesByPolicyError>

Gets information about the traffic policy instances that you created by using a specify traffic policy version.

After you submit a CreateTrafficPolicyInstance or an UpdateTrafficPolicyInstance request, there's a brief delay while Amazon Route 53 creates the resource record sets that are specified in the traffic policy definition. For more information, see the State response element.

Send a GET request to the /Route 53 API version/trafficpolicyinstance resource and include the ID and version of the traffic policy.

Amazon Route 53 returns a maximum of 100 items in each response. If you have a lot of traffic policy instances, you can use the MaxItems parameter to list them in groups of up to 100.

The response includes five values that help you navigate from one group of MaxItems traffic policy instances to the next:

IsTruncated

If the value of IsTruncated in the response is true, there are more traffic policy instances associated with the specified traffic policy.

If IsTruncated is false, this response includes the last traffic policy instance that is associated with the specified traffic policy.

MaxItems

The value that you specified for the MaxItems parameter in the request that produced the current response.

HostedZoneIdMarker, TrafficPolicyInstanceNameMarker, and TrafficPolicyInstanceTypeMarker

If IsTruncated is true, these values in the response represent the first traffic policy instance in the next group of MaxItems traffic policy instances. To list more traffic policy instances, make another call to ListTrafficPolicyInstancesByPolicy, and specify these values in the corresponding request parameters.

If IsTruncated is false, all three elements are omitted from the response.

fn list_traffic_policy_versions(&self,
input: &ListTrafficPolicyVersionsRequest)
-> Result<ListTrafficPolicyVersionsResponse, ListTrafficPolicyVersionsError>

Gets information about all of the versions for a specified traffic policy.

Send a GET request to the /Amazon Route 53 API version/trafficpolicy resource and specify the ID of the traffic policy for which you want to list versions.

Amazon Route 53 returns a maximum of 100 items in each response. If you have a lot of traffic policies, you can use the maxitems parameter to list them in groups of up to 100.

The response includes three values that help you navigate from one group of maxitems traffic policies to the next:

IsTruncated

If the value of IsTruncated in the response is true, there are more traffic policy versions associated with the specified traffic policy.

If IsTruncated is false, this response includes the last traffic policy version that is associated with the specified traffic policy.

TrafficPolicyVersionMarker

The ID of the next traffic policy version that is associated with the current AWS account. If you want to list more traffic policies, make another call to ListTrafficPolicyVersions, and specify the value of the TrafficPolicyVersionMarker element in the TrafficPolicyVersionMarker request parameter.

If IsTruncated is false, Amazon Route 53 omits the TrafficPolicyVersionMarker element from the response.

MaxItems

The value that you specified for the MaxItems parameter in the request that produced the current response.

fn list_vpc_association_authorizations(&self,
input: &ListVPCAssociationAuthorizationsRequest)
-> Result<ListVPCAssociationAuthorizationsResponse, ListVPCAssociationAuthorizationsError>

Gets a list of the VPCs that were created by other accounts and that can be associated with a specified hosted zone because you've submitted one or more CreateVPCAssociationAuthorization requests.

Send a GET request to the /2013-04-01/hostedzone/hosted zone ID/authorizevpcassociation resource. The response to this request includes a VPCs element with a VPC child element for each VPC that can be associated with the hosted zone.

Amazon Route 53 returns up to 50 VPCs per page. To return fewer VPCs per page, include the MaxResults parameter:

/2013-04-01/hostedzone/hosted zone ID/authorizevpcassociation?MaxItems=VPCs per page

If the response includes a NextToken element, there are more VPCs to list. To get the next page of VPCs, submit another ListVPCAssociationAuthorizations request, and include the value of the NextToken element from the response in the NextToken request parameter:

/2013-04-01/hostedzone/hosted zone ID/authorizevpcassociation?MaxItems=VPCs per page&NextToken=

fn test_dns_answer(&self,
input: &TestDNSAnswerRequest)
-> Result<TestDNSAnswerResponse, TestDNSAnswerError>

Gets the value that Amazon Route 53 returns in response to a DNS request for a specified record name and type. You can optionally specify the IP address of a DNS resolver, an EDNS0 client subnet IP address, and a subnet mask.

fn update_health_check(&self,
input: &UpdateHealthCheckRequest)
-> Result<UpdateHealthCheckResponse, UpdateHealthCheckError>

Updates an existing health check.

Send a POST request to the /2013-04-01/healthcheck/health check ID resource. The request body must include a document with an UpdateHealthCheckRequest element. For more information about updating health checks, see Creating, Updating, and Deleting Health Checks in the Amazon Route 53 Developer Guide.

fn update_hosted_zone_comment(&self,
input: &UpdateHostedZoneCommentRequest)
-> Result<UpdateHostedZoneCommentResponse, UpdateHostedZoneCommentError>

Updates the hosted zone comment. Send a POST request to the /2013-04-01/hostedzone/hosted zone ID resource.

fn update_traffic_policy_comment(&self,
input: &UpdateTrafficPolicyCommentRequest)
-> Result<UpdateTrafficPolicyCommentResponse, UpdateTrafficPolicyCommentError>

Updates the comment for a specified traffic policy version.

Send a POST request to the /2013-04-01/trafficpolicy/ resource.

The request body must include a document with an UpdateTrafficPolicyCommentRequest element.

fn update_traffic_policy_instance(&self,
input: &UpdateTrafficPolicyInstanceRequest)
-> Result<UpdateTrafficPolicyInstanceResponse, UpdateTrafficPolicyInstanceError>

Updates the resource record sets in a specified hosted zone that were created based on the settings in a specified traffic policy version.

Send a POST request to the /2013-04-01/trafficpolicyinstance/traffic policy ID resource. The request body must include a document with an UpdateTrafficPolicyInstanceRequest element.

When you update a traffic policy instance, Amazon Route 53 continues to respond to DNS queries for the root resource record set name (such as example.com) while it replaces one group of resource record sets with another. Amazon Route 53 performs the following operations:

Amazon Route 53 creates a new group of resource record sets based on the specified traffic policy. This is true regardless of how substantial the differences are between the existing resource record sets and the new resource record sets.

When all of the new resource record sets have been created, Amazon Route 53 starts to respond to DNS queries for the root resource record set name (such as example.com) by using the new resource record sets.

Amazon Route 53 deletes the old group of resource record sets that are associated with the root resource record set name.