Struct rusoto::route53domains::ContactDetail

pub struct ContactDetail {
    pub address_line_1: Option<AddressLine>,
    pub address_line_2: Option<AddressLine>,
    pub city: Option<City>,
    pub contact_type: Option<ContactType>,
    pub country_code: Option<CountryCode>,
    pub email: Option<Email>,
    pub extra_params: Option<ExtraParamList>,
    pub fax: Option<ContactNumber>,
    pub first_name: Option<ContactName>,
    pub last_name: Option<ContactName>,
    pub organization_name: Option<ContactName>,
    pub phone_number: Option<ContactNumber>,
    pub state: Option<State>,
    pub zip_code: Option<ZipCode>,
}

ContactDetail includes the following elements.

Fields
address_line_1: Option<AddressLine>

First line of the contact's address.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

address_line_2: Option<AddressLine>

Second line of contact's address, if any.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: No

city: Option<City>

The city of the contact's address.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

contact_type: Option<ContactType>

Indicates whether the contact is a person, company, association, or public organization. If you choose an option other than PERSON, you must enter an organization name, and you can't enable privacy protection for the contact.

Type: String

Default: None

Constraints: Maximum 255 characters.

Valid values: PERSON | COMPANY | ASSOCIATION | PUBLIC_BODY

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

country_code: Option<CountryCode>

Code for the country of the contact's address.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

email: Option<Email>

Email address of the contact.

Type: String

Default: None

Constraints: Maximum 254 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

extra_params: Option<ExtraParamList>

A list of name-value pairs for parameters required by certain top-level domains.

Type: Complex

Default: None

Parents: RegistrantContact, AdminContact, TechContact

Children: Name, Value

Required: No

fax: Option<ContactNumber>

Fax number of the contact.

Type: String

Default: None

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code]". For example, a US phone number might appear as "+1.1234567890".

Parents: RegistrantContact, AdminContact, TechContact

Required: No

first_name: Option<ContactName>

First name of contact.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

last_name: Option<ContactName>

Last name of contact.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

organization_name: Option<ContactName>

Name of the organization for contact types other than PERSON.

Type: String

Default: None

Constraints: Maximum 255 characters. Contact type must not be PERSON.

Parents: RegistrantContact, AdminContact, TechContact

Required: No

phone_number: Option<ContactNumber>

The phone number of the contact.

Type: String

Default: None

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code>]". For example, a US phone number might appear as "+1.1234567890".

Parents: RegistrantContact, AdminContact, TechContact

Required: Yes

state: Option<State>

The state or province of the contact's city.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: No

zip_code: Option<ZipCode>

The zip or postal code of the contact's address.

Type: String

Default: None

Constraints: Maximum 255 characters.

Parents: RegistrantContact, AdminContact, TechContact

Required: No