Struct rusoto::route53domains::GetDomainDetailResponse

pub struct GetDomainDetailResponse {
    pub abuse_contact_email: Option<Email>,
    pub abuse_contact_phone: Option<ContactNumber>,
    pub admin_contact: ContactDetail,
    pub admin_privacy: Option<Boolean>,
    pub auto_renew: Option<Boolean>,
    pub creation_date: Option<Timestamp>,
    pub dns_sec: Option<DNSSec>,
    pub domain_name: DomainName,
    pub expiration_date: Option<Timestamp>,
    pub nameservers: NameserverList,
    pub registrant_contact: ContactDetail,
    pub registrant_privacy: Option<Boolean>,
    pub registrar_name: Option<RegistrarName>,
    pub registrar_url: Option<RegistrarUrl>,
    pub registry_domain_id: Option<RegistryDomainId>,
    pub reseller: Option<Reseller>,
    pub status_list: Option<DomainStatusList>,
    pub tech_contact: ContactDetail,
    pub tech_privacy: Option<Boolean>,
    pub updated_date: Option<Timestamp>,
    pub who_is_server: Option<RegistrarWhoIsServer>,
}

The GetDomainDetail response includes the following elements.

Fields
abuse_contact_email: Option<Email>

Email address to contact to report incorrect contact information for a domain, to report that the domain is being used to send spam, to report that someone is cybersquatting on a domain name, or report some other type of abuse.

Type: String

abuse_contact_phone: Option<ContactNumber>

Phone number for reporting abuse.

Type: String

admin_contact: ContactDetail

Provides details about the domain administrative contact.

Type: Complex

Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams

admin_privacy: Option<Boolean>

Specifies whether contact information for the admin contact is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Type: Boolean

auto_renew: Option<Boolean>

Specifies whether the domain registration is set to renew automatically.

Type: Boolean

creation_date: Option<Timestamp>

The date when the domain was created as found in the response to a WHOIS query. The date format is Unix time.

dns_sec: Option<DNSSec>

Reserved for future use.

domain_name: DomainName

The name of a domain.

Type: String

expiration_date: Option<Timestamp>

The date when the registration for the domain is set to expire. The date format is Unix time.

nameservers: NameserverList

The name of the domain.

Type: String

registrant_contact: ContactDetail

Provides details about the domain registrant.

Type: Complex

Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams

registrant_privacy: Option<Boolean>

Specifies whether contact information for the registrant contact is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Type: Boolean

registrar_name: Option<RegistrarName>

Name of the registrar of the domain as identified in the registry. Amazon Route 53 domains are registered by registrar Gandi. The value is "GANDI SAS".

Type: String

registrar_url: Option<RegistrarUrl>

Web address of the registrar.

Type: String

registry_domain_id: Option<RegistryDomainId>

Reserved for future use.

reseller: Option<Reseller>

Reseller of the domain. Domains registered or transferred using Amazon Route 53 domains will have "Amazon" as the reseller.

Type: String

status_list: Option<DomainStatusList>

An array of domain name status codes, also known as Extensible Provisioning Protocol (EPP) status codes.

ICANN, the organization that maintains a central database of domain names, has developed a set of domain name status codes that tell you the status of a variety of operations on a domain name, for example, registering a domain name, transferring a domain name to another registrar, renewing the registration for a domain name, and so on. All registrars use this same set of status codes.

For a current list of domain name status codes and an explanation of what each code means, go to the ICANN website and search for epp status codes. (Search on the ICANN website; web searches sometimes return an old version of the document.)

Type: Array of String

tech_contact: ContactDetail

Provides details about the domain technical contact.

Type: Complex

Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams

tech_privacy: Option<Boolean>

Specifies whether contact information for the tech contact is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Type: Boolean

updated_date: Option<Timestamp>

The last updated date of the domain as found in the response to a WHOIS query. The date format is Unix time.

who_is_server: Option<RegistrarWhoIsServer>

The fully qualified name of the WHOIS server that can answer the WHOIS query for the domain.

Type: String