Struct rusoto::route53domains::RegisterDomainRequest

pub struct RegisterDomainRequest {
    pub admin_contact: ContactDetail,
    pub auto_renew: Option<Boolean>,
    pub domain_name: DomainName,
    pub duration_in_years: DurationInYears,
    pub idn_lang_code: Option<LangCode>,
    pub privacy_protect_admin_contact: Option<Boolean>,
    pub privacy_protect_registrant_contact: Option<Boolean>,
    pub privacy_protect_tech_contact: Option<Boolean>,
    pub registrant_contact: ContactDetail,
    pub tech_contact: ContactDetail,
}

The RegisterDomain request includes the following elements.

Fields
admin_contact: ContactDetail

Provides detailed contact information.

Type: Complex

Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams

Required: Yes

auto_renew: Option<Boolean>

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged.

Type: Boolean

Valid values: true | false

Default: true

Required: No

domain_name: DomainName

The name of a domain.

Type: String

Default: None

Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

Required: Yes

duration_in_years: DurationInYears

The number of years the domain will be registered. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain.

Type: Integer

Default: 1

Valid values: Integer from 1 to 10

Required: Yes

idn_lang_code: Option<LangCode>

Reserved for future use.

privacy_protect_admin_contact: Option<Boolean>

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Type: Boolean

Default: true

Valid values: true | false

Required: No

privacy_protect_registrant_contact: Option<Boolean>

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Type: Boolean

Default: true

Valid values: true | false

Required: No

privacy_protect_tech_contact: Option<Boolean>

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Type: Boolean

Default: true

Valid values: true | false

Required: No

registrant_contact: ContactDetail

Provides detailed contact information.

Type: Complex

Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams

Required: Yes

tech_contact: ContactDetail

Provides detailed contact information.

Type: Complex

Children: FirstName, MiddleName, LastName, ContactType, OrganizationName, AddressLine1, AddressLine2, City, State, CountryCode, ZipCode, PhoneNumber, Email, Fax, ExtraParams

Required: Yes