Struct rusoto::route53domains::Route53DomainsClient

pub struct Route53DomainsClient<P, D> where P: ProvideAwsCredentials,
        D: DispatchSignedRequest { /* fields omitted */ }

A client for the Amazon Route 53 Domains API.

Methods

impl<P, D> Route53DomainsClient<P, D> where P: ProvideAwsCredentials,
D: DispatchSignedRequest

fn new(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn check_domain_availability(&self,
input: &CheckDomainAvailabilityRequest)
-> Result<CheckDomainAvailabilityResponse, CheckDomainAvailabilityError>

This operation checks the availability of one domain name. Note that if the availability status of a domain is pending, you must submit another request to determine the availability of the domain name.

fn delete_tags_for_domain(&self,
input: &DeleteTagsForDomainRequest)
-> Result<DeleteTagsForDomainResponse, DeleteTagsForDomainError>

This operation deletes the specified tags for a domain.

All tag operations are eventually consistent; subsequent operations may not immediately represent all issued operations.

fn disable_domain_auto_renew(&self,
input: &DisableDomainAutoRenewRequest)
-> Result<DisableDomainAutoRenewResponse, DisableDomainAutoRenewError>

This operation disables automatic renewal of domain registration for the specified domain.

fn disable_domain_transfer_lock(&self,
input: &DisableDomainTransferLockRequest)
-> Result<DisableDomainTransferLockResponse, DisableDomainTransferLockError>

This operation removes the transfer lock on the domain (specifically the clientTransferProhibited status) to allow domain transfers. We recommend you refrain from performing this action unless you intend to transfer the domain to a different registrar. Successful submission returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

fn enable_domain_auto_renew(&self,
input: &EnableDomainAutoRenewRequest)
-> Result<EnableDomainAutoRenewResponse, EnableDomainAutoRenewError>

This operation configures Amazon Route 53 to automatically renew the specified domain before the domain registration expires. The cost of renewing your domain registration is billed to your AWS account.

The period during which you can renew a domain name varies by TLD. For a list of TLDs and their renewal policies, see "Renewal, restoration, and deletion times" on the website for our registrar partner, Gandi. Route 53 requires that you renew before the end of the renewal period that is listed on the Gandi website so we can complete processing before the deadline.

fn enable_domain_transfer_lock(&self,
input: &EnableDomainTransferLockRequest)
-> Result<EnableDomainTransferLockResponse, EnableDomainTransferLockError>

This operation sets the transfer lock on the domain (specifically the clientTransferProhibited status) to prevent domain transfers. Successful submission returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

fn get_contact_reachability_status(&self,
input: &GetContactReachabilityStatusRequest)
-> Result<GetContactReachabilityStatusResponse, GetContactReachabilityStatusError>

For operations that require confirmation that the email address for the registrant contact is valid, such as registering a new domain, this operation returns information about whether the registrant contact has responded.

If you want us to resend the email, use the ResendContactReachabilityEmail operation.

fn get_domain_detail(&self,
input: &GetDomainDetailRequest)
-> Result<GetDomainDetailResponse, GetDomainDetailError>

This operation returns detailed information about the domain. The domain's contact information is also returned as part of the output.

fn get_domain_suggestions(&self,
input: &GetDomainSuggestionsRequest)
-> Result<GetDomainSuggestionsResponse, GetDomainSuggestionsError>

The GetDomainSuggestions operation returns a list of suggested domain names given a string, which can either be a domain name or simply a word or phrase (without spaces).

Parameters:

DomainName (string): The basis for your domain suggestion search, a string with (or without) top-level domain specified.

SuggestionCount (int): The number of domain suggestions to be returned, maximum 50, minimum 1.

OnlyAvailable (bool): If true, availability check will be performed on suggestion results, and only available domains will be returned. If false, suggestions will be returned without checking whether the domain is actually available, and caller will have to call checkDomainAvailability for each suggestion to determine availability for registration.

fn get_operation_detail(&self,
input: &GetOperationDetailRequest)
-> Result<GetOperationDetailResponse, GetOperationDetailError>

This operation returns the current status of an operation that is not completed.

fn list_domains(&self,
input: &ListDomainsRequest)
-> Result<ListDomainsResponse, ListDomainsError>

This operation returns all the domain names registered with Amazon Route 53 for the current AWS account.

fn list_operations(&self,
input: &ListOperationsRequest)
-> Result<ListOperationsResponse, ListOperationsError>

This operation returns the operation IDs of operations that are not yet complete.

fn list_tags_for_domain(&self,
input: &ListTagsForDomainRequest)
-> Result<ListTagsForDomainResponse, ListTagsForDomainError>

This operation returns all of the tags that are associated with the specified domain.

All tag operations are eventually consistent; subsequent operations may not immediately represent all issued operations.

fn register_domain(&self,
input: &RegisterDomainRequest)
-> Result<RegisterDomainResponse, RegisterDomainError>

This operation registers a domain. Domains are registered by the AWS registrar partner, Gandi. For some top-level domains (TLDs), this operation requires extra parameters.

When you register a domain, Amazon Route 53 does the following:

Creates a Amazon Route 53 hosted zone that has the same name as the domain. Amazon Route 53 assigns four name servers to your hosted zone and automatically updates your domain registration with the names of these name servers.

Enables autorenew, so your domain registration will renew automatically each year. We'll notify you in advance of the renewal date so you can choose whether to renew the registration.

Optionally enables privacy protection, so WHOIS queries return contact information for our registrar partner, Gandi, instead of the information you entered for registrant, admin, and tech contacts.

If registration is successful, returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant is notified by email.

Charges your AWS account an amount based on the top-level domain. For more information, see Amazon Route 53 Pricing.

fn renew_domain(&self,
input: &RenewDomainRequest)
-> Result<RenewDomainResponse, RenewDomainError>

This operation renews a domain for the specified number of years. The cost of renewing your domain is billed to your AWS account.

We recommend that you renew your domain several weeks before the expiration date. Some TLD registries delete domains before the expiration date if you haven't renewed far enough in advance. For more information about renewing domain registration, see Renewing Registration for a Domain in the Amazon Route 53 documentation.

fn resend_contact_reachability_email(&self,
input: &ResendContactReachabilityEmailRequest)
-> Result<ResendContactReachabilityEmailResponse, ResendContactReachabilityEmailError>

For operations that require confirmation that the email address for the registrant contact is valid, such as registering a new domain, this operation resends the confirmation email to the current email address for the registrant contact.

fn retrieve_domain_auth_code(&self,
input: &RetrieveDomainAuthCodeRequest)
-> Result<RetrieveDomainAuthCodeResponse, RetrieveDomainAuthCodeError>

This operation returns the AuthCode for the domain. To transfer a domain to another registrar, you provide this value to the new registrar.

fn transfer_domain(&self,
input: &TransferDomainRequest)
-> Result<TransferDomainResponse, TransferDomainError>

This operation transfers a domain from another registrar to Amazon Route 53. When the transfer is complete, the domain is registered with the AWS registrar partner, Gandi.

For transfer requirements, a detailed procedure, and information about viewing the status of a domain transfer, see Transferring Registration for a Domain to Amazon Route 53 in the Amazon Route 53 Developer Guide.

If the registrar for your domain is also the DNS service provider for the domain, we highly recommend that you consider transferring your DNS service to Amazon Route 53 or to another DNS service provider before you transfer your registration. Some registrars provide free DNS service when you purchase a domain registration. When you transfer the registration, the previous registrar will not renew your domain registration and could end your DNS service at any time.
Caution! If the registrar for your domain is also the DNS service provider for the domain and you don't transfer DNS service to another provider, your website, email, and the web applications associated with the domain might become unavailable.
If the transfer is successful, this method returns an operation ID that you can use to track the progress and completion of the action. If the transfer doesn't complete successfully, the domain registrant will be notified by email.

fn update_domain_contact(&self,
input: &UpdateDomainContactRequest)
-> Result<UpdateDomainContactResponse, UpdateDomainContactError>

This operation updates the contact information for a particular domain. Information for at least one contact (registrant, administrator, or technical) must be supplied for update.

If the update is successful, this method returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

fn update_domain_contact_privacy(&self,
input: &UpdateDomainContactPrivacyRequest)
-> Result<UpdateDomainContactPrivacyResponse, UpdateDomainContactPrivacyError>

This operation updates the specified domain contact's privacy setting. When the privacy option is enabled, personal information such as postal or email address is hidden from the results of a public WHOIS query. The privacy services are provided by the AWS registrar, Gandi. For more information, see the Gandi privacy features.

This operation only affects the privacy of the specified contact type (registrant, administrator, or tech). Successful acceptance returns an operation ID that you can use with GetOperationDetail to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

fn update_domain_nameservers(&self,
input: &UpdateDomainNameserversRequest)
-> Result<UpdateDomainNameserversResponse, UpdateDomainNameserversError>

This operation replaces the current set of name servers for the domain with the specified set of name servers. If you use Amazon Route 53 as your DNS service, specify the four name servers in the delegation set for the hosted zone for the domain.

If successful, this operation returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

fn update_tags_for_domain(&self,
input: &UpdateTagsForDomainRequest)
-> Result<UpdateTagsForDomainResponse, UpdateTagsForDomainError>

This operation adds or updates tags for a specified domain.

All tag operations are eventually consistent; subsequent operations may not immediately represent all issued operations.

fn view_billing(&self,
input: &ViewBillingRequest)
-> Result<ViewBillingResponse, ViewBillingError>

This operation returns all the domain-related billing records for the current AWS account for a specified period