Struct rusoto::route53domains::UpdateTagsForDomainRequest

pub struct UpdateTagsForDomainRequest {
    pub domain_name: DomainName,
    pub tags_to_update: Option<TagList>,
}

The UpdateTagsForDomainRequest includes the following elements.

Fields
domain_name: DomainName

The domain for which you want to add or update tags.

The name of a domain.

Type: String

Default: None

Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Hyphens are allowed only when they're surrounded by letters, numbers, or other hyphens. You can't specify a hyphen at the beginning or end of a label. To specify an Internationalized Domain Name, you must convert the name to Punycode.

Required: Yes

tags_to_update: Option<TagList>

A list of the tag keys and values that you want to add or update. If you specify a key that already exists, the corresponding value will be replaced.

Type: A complex type containing a list of tags

Default: None

Required: No
'>
Each tag includes the following elements:

Key

The key (name) of a tag.

Type: String

Default: None

Valid values: Unicode characters including alphanumeric, space, and ".:/=+-@"

Constraints: Each key can be 1-128 characters long.

Required: Yes

Value

The value of a tag.

Type: String

Default: None

Valid values: Unicode characters including alphanumeric, space, and ".:/=+-@"

Constraints: Each value can be 0-256 characters long.

Required: Yes