Module rusoto::s3

The AWS S3 API.

If you're using the service, you're probably looking for S3Client.

Structs

AbortIncompleteMultipartUpload
Specifies the days since the initiation of an Incomplete Multipart Upload that Lifecycle will wait before permanently removing all parts of the upload.

AbortIncompleteMultipartUploadSerializer

AbortMultipartUploadOutput

AbortMultipartUploadRequest

AccelerateConfiguration

AccelerateConfigurationSerializer

AccessControlPolicy

AccessControlPolicySerializer

AccountIdSerializer

AllowedHeaderSerializer

AllowedHeadersSerializer

AllowedMethodSerializer

AllowedMethodsSerializer

AllowedOriginSerializer

AllowedOriginsSerializer

AnalyticsAndOperator

AnalyticsAndOperatorSerializer

AnalyticsConfiguration

AnalyticsConfigurationSerializer

AnalyticsExportDestination

AnalyticsExportDestinationSerializer

AnalyticsFilter

AnalyticsFilterSerializer

AnalyticsIdSerializer

AnalyticsS3BucketDestination

AnalyticsS3BucketDestinationSerializer

AnalyticsS3ExportFileFormatSerializer

BodySerializer

Bucket

BucketAccelerateStatusSerializer

BucketLifecycleConfiguration

BucketLifecycleConfigurationSerializer

BucketLocationConstraintSerializer

BucketLoggingStatus

BucketLoggingStatusSerializer

BucketLogsPermissionSerializer

BucketNameSerializer

BucketVersioningStatusSerializer

CORSConfiguration

CORSConfigurationSerializer

CORSRule

CORSRuleSerializer

CORSRulesSerializer

CloudFunctionConfiguration

CloudFunctionConfigurationSerializer

CloudFunctionInvocationRoleSerializer

CloudFunctionSerializer

CommonPrefix

CompleteMultipartUploadOutput

CompleteMultipartUploadRequest

CompletedMultipartUpload

CompletedMultipartUploadSerializer

CompletedPart

CompletedPartListSerializer

CompletedPartSerializer

Condition

ConditionSerializer

CopyObjectOutput

CopyObjectRequest

CopyObjectResult

CopyPartResult

CreateBucketConfiguration

CreateBucketConfigurationSerializer

CreateBucketOutput

CreateBucketRequest

CreateMultipartUploadOutput

CreateMultipartUploadRequest

DateSerializer

DaysAfterInitiationSerializer

DaysSerializer

Delete

DeleteBucketAnalyticsConfigurationRequest

DeleteBucketCorsRequest

DeleteBucketInventoryConfigurationRequest

DeleteBucketLifecycleRequest

DeleteBucketMetricsConfigurationRequest

DeleteBucketPolicyRequest

DeleteBucketReplicationRequest

DeleteBucketRequest

DeleteBucketTaggingRequest

DeleteBucketWebsiteRequest

DeleteMarkerEntry

DeleteObjectOutput

DeleteObjectRequest

DeleteObjectTaggingOutput

DeleteObjectTaggingRequest

DeleteObjectsOutput

DeleteObjectsRequest

DeleteSerializer

DeletedObject

DelimiterSerializer

Destination

DestinationSerializer

DisplayNameSerializer

ETagSerializer

EmailAddressSerializer

EncodingTypeSerializer

ErrorDocument

ErrorDocumentSerializer

EventListSerializer

EventSerializer

ExpirationStatusSerializer

ExpiredObjectDeleteMarkerSerializer

ExposeHeaderSerializer

ExposeHeadersSerializer

FetchOwnerSerializer

FilterRule
Container for key value pair that defines the criteria for the filter rule.

FilterRuleListSerializer

FilterRuleNameSerializer

FilterRuleSerializer

FilterRuleValueSerializer

GetBucketAccelerateConfigurationOutput

GetBucketAccelerateConfigurationRequest

GetBucketAclOutput

GetBucketAclRequest

GetBucketAnalyticsConfigurationOutput

GetBucketAnalyticsConfigurationRequest

GetBucketCorsOutput

GetBucketCorsRequest

GetBucketInventoryConfigurationOutput

GetBucketInventoryConfigurationRequest

GetBucketLifecycleConfigurationOutput

GetBucketLifecycleConfigurationRequest

GetBucketLifecycleOutput

GetBucketLifecycleRequest

GetBucketLocationOutput

GetBucketLocationRequest

GetBucketLoggingOutput

GetBucketLoggingRequest

GetBucketMetricsConfigurationOutput

GetBucketMetricsConfigurationRequest

GetBucketNotificationConfigurationRequest

GetBucketPolicyOutput

GetBucketPolicyRequest

GetBucketReplicationOutput

GetBucketReplicationRequest

GetBucketRequestPaymentOutput

GetBucketRequestPaymentRequest

GetBucketTaggingOutput

GetBucketTaggingRequest

GetBucketVersioningOutput

GetBucketVersioningRequest

GetBucketWebsiteOutput

GetBucketWebsiteRequest

GetObjectAclOutput

GetObjectAclRequest

GetObjectOutput

GetObjectRequest

GetObjectTaggingOutput

GetObjectTaggingRequest

GetObjectTorrentOutput

GetObjectTorrentRequest

GlacierJobParameters

GlacierJobParametersSerializer

Grant

GrantSerializer

Grantee

GranteeSerializer

GrantsSerializer

HeadBucketRequest

HeadObjectOutput

HeadObjectRequest

HostNameSerializer

HttpErrorCodeReturnedEqualsSerializer

HttpRedirectCodeSerializer

IDSerializer

IndexDocument

IndexDocumentSerializer

Initiator

InventoryConfiguration

InventoryConfigurationSerializer

InventoryDestination

InventoryDestinationSerializer

InventoryFilter

InventoryFilterSerializer

InventoryFormatSerializer

InventoryFrequencySerializer

InventoryIdSerializer

InventoryIncludedObjectVersionsSerializer

InventoryOptionalFieldSerializer

InventoryOptionalFieldsSerializer

InventoryS3BucketDestination

InventoryS3BucketDestinationSerializer

InventorySchedule

InventoryScheduleSerializer

IsEnabledSerializer

KeyMarkerSerializer

KeyPrefixEqualsSerializer

LambdaFunctionArnSerializer

LambdaFunctionConfiguration
Container for specifying the AWS Lambda notification configuration.

LambdaFunctionConfigurationListSerializer

LambdaFunctionConfigurationSerializer

LifecycleConfiguration

LifecycleConfigurationSerializer

LifecycleExpiration

LifecycleExpirationSerializer

LifecycleRule

LifecycleRuleAndOperator
This is used in a Lifecycle Rule Filter to apply a logical AND to two or more predicates. The Lifecycle Rule will apply to any object matching all of the predicates configured inside the And operator.

LifecycleRuleAndOperatorSerializer

LifecycleRuleFilter
The Filter is used to identify objects that a Lifecycle Rule applies to. A Filter must have exactly one of Prefix, Tag, or And specified.

LifecycleRuleFilterSerializer

LifecycleRuleSerializer

LifecycleRulesSerializer

ListBucketAnalyticsConfigurationsOutput

ListBucketAnalyticsConfigurationsRequest

ListBucketInventoryConfigurationsOutput

ListBucketInventoryConfigurationsRequest

ListBucketMetricsConfigurationsOutput

ListBucketMetricsConfigurationsRequest

ListBucketsOutput

ListMultipartUploadsOutput

ListMultipartUploadsRequest

ListObjectVersionsOutput

ListObjectVersionsRequest

ListObjectsOutput

ListObjectsRequest

ListObjectsV2Output

ListObjectsV2Request

ListPartsOutput

ListPartsRequest

LoggingEnabled

LoggingEnabledSerializer

MFADeleteSerializer

MarkerSerializer

MaxAgeSecondsSerializer

MaxKeysSerializer

MaxPartsSerializer

MaxUploadsSerializer

MetricsAndOperator

MetricsAndOperatorSerializer

MetricsConfiguration

MetricsConfigurationSerializer

MetricsFilter

MetricsFilterSerializer

MetricsIdSerializer

MultipartUpload

MultipartUploadIdSerializer

NoncurrentVersionExpiration
Specifies when noncurrent object versions expire. Upon expiration, Amazon S3 permanently deletes the noncurrent object versions. You set this lifecycle configuration action on a bucket that has versioning enabled (or suspended) to request that Amazon S3 delete noncurrent object versions at a specific period in the object's lifetime.

NoncurrentVersionExpirationSerializer

NoncurrentVersionTransition
Container for the transition rule that describes when noncurrent objects transition to the STANDARD_IA or GLACIER storage class. If your bucket is versioning-enabled (or versioning is suspended), you can set this action to request that Amazon S3 transition noncurrent object versions to the STANDARD_IA or GLACIER storage class at a specific period in the object's lifetime.

NoncurrentVersionTransitionListSerializer

NoncurrentVersionTransitionSerializer

NotificationConfiguration
Container for specifying the notification configuration of the bucket. If this element is empty, notifications are turned off on the bucket.

NotificationConfigurationDeprecated

NotificationConfigurationDeprecatedSerializer

NotificationConfigurationFilter
Container for object key name filtering rules. For information about key name filtering, go to Configuring Event Notifications in the Amazon Simple Storage Service Developer Guide.

NotificationConfigurationFilterSerializer

NotificationConfigurationSerializer

NotificationIdSerializer

Object

ObjectIdentifier

ObjectIdentifierListSerializer

ObjectIdentifierSerializer

ObjectKeySerializer

ObjectVersion

ObjectVersionIdSerializer

Owner

OwnerSerializer

Part

PartNumberMarkerSerializer

PartNumberSerializer

PayerSerializer

PermissionSerializer

PolicySerializer

PrefixSerializer

ProtocolSerializer

PutBucketAccelerateConfigurationRequest

PutBucketAclRequest

PutBucketAnalyticsConfigurationRequest

PutBucketCorsRequest

PutBucketInventoryConfigurationRequest

PutBucketLifecycleConfigurationRequest

PutBucketLifecycleRequest

PutBucketLoggingRequest

PutBucketMetricsConfigurationRequest

PutBucketNotificationConfigurationRequest

PutBucketNotificationRequest

PutBucketPolicyRequest

PutBucketReplicationRequest

PutBucketRequestPaymentRequest

PutBucketTaggingRequest

PutBucketVersioningRequest

PutBucketWebsiteRequest

PutObjectAclOutput

PutObjectAclRequest

PutObjectOutput

PutObjectRequest

PutObjectTaggingOutput

PutObjectTaggingRequest

QueueArnSerializer

QueueConfiguration
Container for specifying an configuration when you want Amazon S3 to publish events to an Amazon Simple Queue Service (Amazon SQS) queue.

QueueConfigurationDeprecated

QueueConfigurationDeprecatedSerializer

QueueConfigurationListSerializer

QueueConfigurationSerializer

QuietSerializer

Redirect

RedirectAllRequestsTo

RedirectAllRequestsToSerializer

RedirectSerializer

ReplaceKeyPrefixWithSerializer

ReplaceKeyWithSerializer

ReplicationConfiguration
Container for replication rules. You can add as many as 1,000 rules. Total replication configuration size can be up to 2 MB.

ReplicationConfigurationSerializer

ReplicationRule

ReplicationRuleSerializer

ReplicationRuleStatusSerializer

ReplicationRulesSerializer

RequestPaymentConfiguration

RequestPaymentConfigurationSerializer

ResponseCacheControlSerializer

ResponseContentDispositionSerializer

ResponseContentEncodingSerializer

ResponseContentLanguageSerializer

ResponseContentTypeSerializer

ResponseExpiresSerializer

RestoreObjectOutput

RestoreObjectRequest

RestoreRequest

RestoreRequestSerializer

RoleSerializer

RoutingRule

RoutingRuleSerializer

RoutingRulesSerializer

Rule

RuleSerializer

RulesSerializer

S3Client
A client for the Amazon S3 API.

S3Error

S3KeyFilter
Container for object key name prefix and suffix filtering rules.

S3KeyFilterSerializer

StartAfterSerializer

StorageClassAnalysis

StorageClassAnalysisDataExport

StorageClassAnalysisDataExportSerializer

StorageClassAnalysisSchemaVersionSerializer

StorageClassAnalysisSerializer

StorageClassSerializer

SuffixSerializer

Tag

TagSerializer

TagSetSerializer

Tagging

TaggingSerializer

TargetBucketSerializer

TargetGrant

TargetGrantSerializer

TargetGrantsSerializer

TargetPrefixSerializer

TierSerializer

TokenSerializer

TopicArnSerializer

TopicConfiguration
Container for specifying the configuration when you want Amazon S3 to publish events to an Amazon Simple Notification Service (Amazon SNS) topic.

TopicConfigurationDeprecated

TopicConfigurationDeprecatedSerializer

TopicConfigurationListSerializer

TopicConfigurationSerializer

Transition

TransitionListSerializer

TransitionSerializer

TransitionStorageClassSerializer

TypeSerializer

URISerializer

UploadIdMarkerSerializer

UploadPartCopyOutput

UploadPartCopyRequest

UploadPartOutput

UploadPartRequest

ValueSerializer

VersionIdMarkerSerializer

VersioningConfiguration

VersioningConfigurationSerializer

WebsiteConfiguration

WebsiteConfigurationSerializer

Enums

AbortMultipartUploadError
Errors returned by AbortMultipartUpload

CompleteMultipartUploadError
Errors returned by CompleteMultipartUpload

CopyObjectError
Errors returned by CopyObject

CreateBucketError
Errors returned by CreateBucket

CreateMultipartUploadError
Errors returned by CreateMultipartUpload

DeleteBucketAnalyticsConfigurationError
Errors returned by DeleteBucketAnalyticsConfiguration

DeleteBucketCorsError
Errors returned by DeleteBucketCors

DeleteBucketError
Errors returned by DeleteBucket

DeleteBucketInventoryConfigurationError
Errors returned by DeleteBucketInventoryConfiguration

DeleteBucketLifecycleError
Errors returned by DeleteBucketLifecycle

DeleteBucketMetricsConfigurationError
Errors returned by DeleteBucketMetricsConfiguration

DeleteBucketPolicyError
Errors returned by DeleteBucketPolicy

DeleteBucketReplicationError
Errors returned by DeleteBucketReplication

DeleteBucketTaggingError
Errors returned by DeleteBucketTagging

DeleteBucketWebsiteError
Errors returned by DeleteBucketWebsite

DeleteObjectError
Errors returned by DeleteObject

DeleteObjectTaggingError
Errors returned by DeleteObjectTagging

DeleteObjectsError
Errors returned by DeleteObjects

GetBucketAccelerateConfigurationError
Errors returned by GetBucketAccelerateConfiguration

GetBucketAclError
Errors returned by GetBucketAcl

GetBucketAnalyticsConfigurationError
Errors returned by GetBucketAnalyticsConfiguration

GetBucketCorsError
Errors returned by GetBucketCors

GetBucketInventoryConfigurationError
Errors returned by GetBucketInventoryConfiguration

GetBucketLifecycleConfigurationError
Errors returned by GetBucketLifecycleConfiguration

GetBucketLifecycleError
Errors returned by GetBucketLifecycle

GetBucketLocationError
Errors returned by GetBucketLocation

GetBucketLoggingError
Errors returned by GetBucketLogging

GetBucketMetricsConfigurationError
Errors returned by GetBucketMetricsConfiguration

GetBucketNotificationConfigurationError
Errors returned by GetBucketNotificationConfiguration

GetBucketNotificationError
Errors returned by GetBucketNotification

GetBucketPolicyError
Errors returned by GetBucketPolicy

GetBucketReplicationError
Errors returned by GetBucketReplication

GetBucketRequestPaymentError
Errors returned by GetBucketRequestPayment

GetBucketTaggingError
Errors returned by GetBucketTagging

GetBucketVersioningError
Errors returned by GetBucketVersioning

GetBucketWebsiteError
Errors returned by GetBucketWebsite

GetObjectAclError
Errors returned by GetObjectAcl

GetObjectError
Errors returned by GetObject

GetObjectTaggingError
Errors returned by GetObjectTagging

GetObjectTorrentError
Errors returned by GetObjectTorrent

HeadBucketError
Errors returned by HeadBucket

HeadObjectError
Errors returned by HeadObject

ListBucketAnalyticsConfigurationsError
Errors returned by ListBucketAnalyticsConfigurations

ListBucketInventoryConfigurationsError
Errors returned by ListBucketInventoryConfigurations

ListBucketMetricsConfigurationsError
Errors returned by ListBucketMetricsConfigurations

ListBucketsError
Errors returned by ListBuckets

ListMultipartUploadsError
Errors returned by ListMultipartUploads

ListObjectVersionsError
Errors returned by ListObjectVersions

ListObjectsError
Errors returned by ListObjects

ListObjectsV2Error
Errors returned by ListObjectsV2

ListPartsError
Errors returned by ListParts

PutBucketAccelerateConfigurationError
Errors returned by PutBucketAccelerateConfiguration

PutBucketAclError
Errors returned by PutBucketAcl

PutBucketAnalyticsConfigurationError
Errors returned by PutBucketAnalyticsConfiguration

PutBucketCorsError
Errors returned by PutBucketCors

PutBucketInventoryConfigurationError
Errors returned by PutBucketInventoryConfiguration

PutBucketLifecycleConfigurationError
Errors returned by PutBucketLifecycleConfiguration

PutBucketLifecycleError
Errors returned by PutBucketLifecycle

PutBucketLoggingError
Errors returned by PutBucketLogging

PutBucketMetricsConfigurationError
Errors returned by PutBucketMetricsConfiguration

PutBucketNotificationConfigurationError
Errors returned by PutBucketNotificationConfiguration

PutBucketNotificationError
Errors returned by PutBucketNotification

PutBucketPolicyError
Errors returned by PutBucketPolicy

PutBucketReplicationError
Errors returned by PutBucketReplication

PutBucketRequestPaymentError
Errors returned by PutBucketRequestPayment

PutBucketTaggingError
Errors returned by PutBucketTagging

PutBucketVersioningError
Errors returned by PutBucketVersioning

PutBucketWebsiteError
Errors returned by PutBucketWebsite

PutObjectAclError
Errors returned by PutObjectAcl

PutObjectError
Errors returned by PutObject

PutObjectTaggingError
Errors returned by PutObjectTagging

RestoreObjectError
Errors returned by RestoreObject

UploadPartCopyError
Errors returned by UploadPartCopy

UploadPartError
Errors returned by UploadPart

Type Definitions

AbortDate

AbortRuleId

AcceptRanges

AccountId

AllowedHeader

AllowedHeaders

AllowedMethod

AllowedMethods

AllowedOrigin

AllowedOrigins

AnalyticsConfigurationList

AnalyticsId

AnalyticsS3ExportFileFormat

Body

BucketAccelerateStatus

BucketCannedACL

BucketLocationConstraint

BucketLogsPermission

BucketName

BucketVersioningStatus

Buckets

CORSRules

CacheControl

CloudFunction

CloudFunctionInvocationRole

Code

CommonPrefixList

CompletedPartList

ContentDisposition

ContentEncoding

ContentLanguage

ContentLength

ContentMD5

ContentRange

ContentType

CopySource

CopySourceIfMatch

CopySourceIfModifiedSince

CopySourceIfNoneMatch

CopySourceIfUnmodifiedSince

CopySourceRange

CopySourceSSECustomerAlgorithm

CopySourceSSECustomerKey

CopySourceSSECustomerKeyMD5

CopySourceVersionId

CreationDate

Date

Days

DaysAfterInitiation

DeleteMarker

DeleteMarkerVersionId

DeleteMarkers

DeletedObjects

Delimiter

DisplayName

ETag

EmailAddress

EncodingType
Requests Amazon S3 to encode the object keys in the response and specifies the encoding method to use. An object key may contain any Unicode character; however, XML 1.0 parser cannot parse some characters, such as characters with an ASCII value from 0 to 10. For characters that are not supported in XML 1.0, you can add this parameter to request that Amazon S3 encode the keys in the response.

Errors

Event
Bucket event for which to send notifications.

EventList

Expiration

ExpirationStatus

ExpiredObjectDeleteMarker

Expires

ExposeHeader

ExposeHeaders

FetchOwner

FilterRuleList
A list of containers for key value pair that defines the criteria for the filter rule.

FilterRuleName

FilterRuleValue

GrantFullControl

GrantRead

GrantReadACP

GrantWrite

GrantWriteACP

Grants

HostName

HttpErrorCodeReturnedEquals

HttpRedirectCode

ID

IfMatch

IfModifiedSince

IfNoneMatch

IfUnmodifiedSince

Initiated

InventoryConfigurationList

InventoryFormat

InventoryFrequency

InventoryId

InventoryIncludedObjectVersions

InventoryOptionalField

InventoryOptionalFields

IsEnabled

IsLatest

IsTruncated

KeyCount

KeyMarker

KeyPrefixEquals

LambdaFunctionArn

LambdaFunctionConfigurationList

LastModified

LifecycleRules

Location

MFA

MFADelete

MFADeleteStatus

Marker

MaxAgeSeconds

MaxKeys

MaxParts

MaxUploads

Message

Metadata

MetadataDirective

MetadataKey

MetadataValue

MetricsConfigurationList

MetricsId

MissingMeta

MultipartUploadId

MultipartUploadList

NextKeyMarker

NextMarker

NextPartNumberMarker

NextToken

NextUploadIdMarker

NextVersionIdMarker

NoncurrentVersionTransitionList

NotificationId
Optional unique identifier for configurations in a notification configuration. If you don't provide one, Amazon S3 will assign an ID.

ObjectCannedACL

ObjectIdentifierList

ObjectKey

ObjectList

ObjectStorageClass

ObjectVersionId

ObjectVersionList

ObjectVersionStorageClass

PartNumber

PartNumberMarker

Parts

PartsCount

Payer

Permission

Policy

Prefix

Protocol

QueueArn

QueueConfigurationList

Quiet

Range

ReplaceKeyPrefixWith

ReplaceKeyWith

ReplicationRuleStatus

ReplicationRules

ReplicationStatus

RequestCharged
If present, indicates that the requester was successfully charged for the request.

RequestPayer
Confirms that the requester knows that she or he will be charged for the request. Bucket owners need not specify this parameter in their requests. Documentation on downloading objects from requester pays buckets can be found at http://docs.aws.amazon.com/AmazonS3/latest/dev/ObjectsinRequesterPaysBuckets.html

ResponseCacheControl

ResponseContentDisposition

ResponseContentEncoding

ResponseContentLanguage

ResponseContentType

ResponseExpires

Restore

Role

RoutingRules

Rules

SSECustomerAlgorithm

SSECustomerKey

SSECustomerKeyMD5

SSEKMSKeyId

ServerSideEncryption

Size

StartAfter

StorageClass

StorageClassAnalysisSchemaVersion

Suffix

TagCount

TagSet

TaggingDirective

TaggingHeader

TargetBucket

TargetGrants

TargetPrefix

Tier

Token

TopicArn

TopicConfigurationList

TransitionList

TransitionStorageClass

Type

URI

UploadIdMarker

Value

VersionIdMarker

WebsiteRedirectLocation