Struct rusoto::s3::CORSRule

pub struct CORSRule {
    pub allowed_headers: AllowedHeaders,
    pub expose_headers: ExposeHeaders,
    pub allowed_methods: AllowedMethods,
    pub max_age_seconds: MaxAgeSeconds,
    pub allowed_origins: AllowedOrigins,
}

Fields

allowed_headers

Specifies which headers are allowed in a pre-flight OPTIONS request.

expose_headers

One or more headers in the response that you want customers to be able to access from their applications (for example, from a JavaScript XMLHttpRequest object).

allowed_methods

Identifies HTTP methods that the domain/origin specified in the rule is allowed to execute.

max_age_seconds

The time in seconds that your browser is to cache the preflight response for the specified resource.

allowed_origins

One or more origins you want customers to be able to access the bucket from.