Struct rusoto::s3::CopyObjectRequest

pub struct CopyObjectRequest {
    pub request_payer: Option<RequestPayer>,
    pub copy_source_if_modified_since: Option<CopySourceIfModifiedSince>,
    pub copy_source_if_unmodified_since: Option<CopySourceIfUnmodifiedSince>,
    pub content_encoding: Option<ContentEncoding>,
    pub copy_source_sse_customer_key: Option<CopySourceSSECustomerKey>,
    pub storage_class: Option<StorageClass>,
    pub grant_read_acp: Option<GrantReadACP>,
    pub server_side_encryption: Option<ServerSideEncryption>,
    pub ssekms_key_id: Option<SSEKMSKeyId>,
    pub content_disposition: Option<ContentDisposition>,
    pub metadata: Option<Metadata>,
    pub sse_customer_key: Option<SSECustomerKey>,
    pub website_redirect_location: Option<WebsiteRedirectLocation>,
    pub copy_source: CopySource,
    pub expires: Option<Expires>,
    pub key: ObjectKey,
    pub cache_control: Option<CacheControl>,
    pub copy_source_sse_customer_algorithm: Option<CopySourceSSECustomerAlgorithm>,
    pub bucket: BucketName,
    pub grant_read: Option<GrantRead>,
    pub grant_write_acp: Option<GrantWriteACP>,
    pub copy_source_sse_customer_key_md5: Option<CopySourceSSECustomerKeyMD5>,
    pub acl: Option<CannedAcl>,
    pub grant_full_control: Option<GrantFullControl>,
    pub copy_source_if_match: Option<CopySourceIfMatch>,
    pub sse_customer_algorithm: Option<SSECustomerAlgorithm>,
    pub content_type: Option<ContentType>,
    pub content_language: Option<ContentLanguage>,
    pub metadata_directive: Option<MetadataDirective>,
    pub copy_source_if_none_match: Option<CopySourceIfNoneMatch>,
    pub sse_customer_key_md5: Option<SSECustomerKeyMD5>,
}

Fields

request_payer

copy_source_if_modified_since

Copies the object if it has been modified since the specified time.

copy_source_if_unmodified_since

Copies the object if it hasn't been modified since the specified time.

content_encoding

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

copy_source_sse_customer_key

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

storage_class

The type of storage to use for the object. Defaults to 'STANDARD'.

grant_read_acp

Allows grantee to read the object ACL.

server_side_encryption

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

ssekms_key_id

Specifies the AWS KMS key ID to use for object encryption. All GET and PUT requests for an object protected by AWS KMS will fail if not made via SSL or using SigV4. Documentation on configuring any of the officially supported AWS SDKs and CLI can be found at http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingAWSSDK.html#specify- signature-version

content_disposition

Specifies presentational information for the object.

metadata

A map of metadata to store with the object in S3.

sse_customer_key

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side- encryption-customer-algorithm header.

website_redirect_location

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

copy_source

The name of the source bucket and key name of the source object, separated by a slash (/). Must be URL-encoded.

expires

The date and time at which the object is no longer cacheable.

key

cache_control

Specifies caching behavior along the request/reply chain.

copy_source_sse_customer_algorithm

Specifies the algorithm to use when decrypting the source object (e.g., AES256).

bucket

grant_read

Allows grantee to read the object data and its metadata.

grant_write_acp

Allows grantee to write the ACL for the applicable object.

copy_source_sse_customer_key_md5

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

acl

The canned ACL to apply to the object.

grant_full_control

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

copy_source_if_match

Copies the object if its entity tag (ETag) matches the specified tag.

sse_customer_algorithm

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

content_type

A standard MIME type describing the format of the object data.

content_language

The language the content is in.

metadata_directive

Specifies whether the metadata is copied from the source object or replaced with metadata provided in the request.

copy_source_if_none_match

Copies the object if its entity tag (ETag) is different than the specified ETag.

sse_customer_key_md5

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.