Struct rusoto::s3::CreateBucketRequest

pub struct CreateBucketRequest {
    pub grant_full_control: Option<GrantFullControl>,
    pub create_bucket_configuration: Option<CreateBucketConfiguration>,
    pub grant_write_acp: Option<GrantWriteACP>,
    pub bucket: BucketName,
    pub acl: Option<CannedAcl>,
    pub grant_write: Option<GrantWrite>,
    pub grant_read: Option<GrantRead>,
    pub grant_read_acp: Option<GrantReadACP>,
}

Fields

grant_full_control

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

create_bucket_configuration

grant_write_acp

Allows grantee to write the ACL for the applicable bucket.

bucket

acl

The canned ACL to apply to the bucket.

grant_write

Allows grantee to create, overwrite, and delete any object in the bucket.

grant_read

Allows grantee to list the objects in the bucket.

grant_read_acp

Allows grantee to read the bucket ACL.