Struct rusoto::s3::CreateMultipartUploadOutput

pub struct CreateMultipartUploadOutput {
    pub sse_customer_algorithm: SSECustomerAlgorithm,
    pub request_charged: RequestCharged,
    pub bucket: BucketName,
    pub upload_id: MultipartUploadId,
    pub key: ObjectKey,
    pub server_side_encryption: ServerSideEncryption,
    pub sse_customer_key_md5: SSECustomerKeyMD5,
    pub ssekms_key_id: SSEKMSKeyId,
}

Fields

sse_customer_algorithm

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

request_charged

bucket

Name of the bucket to which the multipart upload was initiated.

upload_id

ID for the initiated multipart upload.

key

Object key for which the multipart upload was initiated.

server_side_encryption

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

sse_customer_key_md5

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

ssekms_key_id

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.