Struct rusoto::s3::CreateMultipartUploadRequest

pub struct CreateMultipartUploadRequest {
    pub request_payer: Option<RequestPayer>,
    pub content_encoding: Option<ContentEncoding>,
    pub storage_class: Option<StorageClass>,
    pub grant_read_acp: Option<GrantReadACP>,
    pub server_side_encryption: Option<ServerSideEncryption>,
    pub ssekms_key_id: Option<SSEKMSKeyId>,
    pub content_disposition: Option<ContentDisposition>,
    pub metadata: Option<Metadata>,
    pub sse_customer_key: Option<SSECustomerKey>,
    pub website_redirect_location: Option<WebsiteRedirectLocation>,
    pub expires: Option<Expires>,
    pub key: ObjectKey,
    pub cache_control: Option<CacheControl>,
    pub bucket: BucketName,
    pub grant_read: Option<GrantRead>,
    pub grant_write_acp: Option<GrantWriteACP>,
    pub acl: Option<ObjectCannedACL>,
    pub grant_full_control: Option<GrantFullControl>,
    pub sse_customer_algorithm: Option<SSECustomerAlgorithm>,
    pub content_type: Option<ContentType>,
    pub content_language: Option<ContentLanguage>,
    pub sse_customer_key_md5: Option<SSECustomerKeyMD5>,
}

Fields
request_payer: Option<RequestPayer> content_encoding: Option<ContentEncoding>

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

storage_class: Option<StorageClass>

The type of storage to use for the object. Defaults to 'STANDARD'.

grant_read_acp: Option<GrantReadACP>

Allows grantee to read the object ACL.

server_side_encryption: Option<ServerSideEncryption>

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

ssekms_key_id: Option<SSEKMSKeyId>

Specifies the AWS KMS key ID to use for object encryption. All GET and PUT requests for an object protected by AWS KMS will fail if not made via SSL or using SigV4. Documentation on configuring any of the officially supported AWS SDKs and CLI can be found at http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingAWSSDK.html#specify- signature-version

content_disposition: Option<ContentDisposition>

Specifies presentational information for the object.

metadata: Option<Metadata>

A map of metadata to store with the object in S3.

sse_customer_key: Option<SSECustomerKey>

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side- encryption-customer-algorithm header.

website_redirect_location: Option<WebsiteRedirectLocation>

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

expires: Option<Expires>

The date and time at which the object is no longer cacheable.

key: ObjectKey cache_control: Option<CacheControl>

Specifies caching behavior along the request/reply chain.

bucket: BucketName grant_read: Option<GrantRead>

Allows grantee to read the object data and its metadata.

grant_write_acp: Option<GrantWriteACP>

Allows grantee to write the ACL for the applicable object.

acl: Option<ObjectCannedACL>

The canned ACL to apply to the object.

grant_full_control: Option<GrantFullControl>

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

sse_customer_algorithm: Option<SSECustomerAlgorithm>

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

content_type: Option<ContentType>

A standard MIME type describing the format of the object data.

content_language: Option<ContentLanguage>

The language the content is in.

sse_customer_key_md5: Option<SSECustomerKeyMD5>

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.