Struct rusoto::s3::DeleteObjectOutput

pub struct DeleteObjectOutput {
    pub version_id: ObjectVersionId,
    pub request_charged: RequestCharged,
    pub delete_marker: DeleteMarker,
}

Fields

version_id

Returns the version ID of the delete marker created as a result of the DELETE operation.

request_charged

delete_marker

Specifies whether the versioned object that was permanently deleted was (true) or was not (false) a delete marker.