Struct rusoto::s3::ListObjectVersionsOutput

pub struct ListObjectVersionsOutput {
    pub name: BucketName,
    pub versions: ObjectVersionList,
    pub delete_markers: DeleteMarkers,
    pub next_key_marker: NextKeyMarker,
    pub delimiter: Delimiter,
    pub max_keys: MaxKeys,
    pub prefix: Prefix,
    pub key_marker: KeyMarker,
    pub next_version_id_marker: NextVersionIdMarker,
    pub encoding_type: EncodingType,
    pub is_truncated: IsTruncated,
    pub version_id_marker: VersionIdMarker,
    pub common_prefixes: CommonPrefixList,
}

Fields
name: BucketName versions: ObjectVersionList delete_markers: DeleteMarkers next_key_marker: NextKeyMarker

Use this value for the key marker request parameter in a subsequent request.

delimiter: Delimiter max_keys: MaxKeys prefix: Prefix key_marker: KeyMarker

Marks the last Key returned in a truncated response.

next_version_id_marker: NextVersionIdMarker

Use this value for the next version id marker parameter in a subsequent request.

encoding_type: EncodingType

Encoding type used by Amazon S3 to encode object keys in the response.

is_truncated: IsTruncated

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria. If your results were truncated, you can make a follow-up paginated request using the NextKeyMarker and NextVersionIdMarker response parameters as a starting place in another request to return the rest of the results.

version_id_marker: VersionIdMarker common_prefixes: CommonPrefixList