Struct rusoto::s3::ListObjectVersionsRequest

pub struct ListObjectVersionsRequest {
    pub bucket: BucketName,
    pub prefix: Option<Prefix>,
    pub max_keys: Option<MaxKeys>,
    pub delimiter: Option<Delimiter>,
    pub key_marker: Option<KeyMarker>,
    pub encoding_type: Option<EncodingType>,
    pub version_id_marker: Option<VersionIdMarker>,
}

Fields

bucket

prefix

Limits the response to keys that begin with the specified prefix.

max_keys

Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

delimiter

A delimiter is a character you use to group keys.

key_marker

Specifies the key to start with when listing objects in a bucket.

encoding_type

version_id_marker

Specifies the object version you want to start listing from.