Struct rusoto::s3::ListObjectsOutput

pub struct ListObjectsOutput {
    pub name: BucketName,
    pub next_marker: NextMarker,
    pub delimiter: Delimiter,
    pub max_keys: MaxKeys,
    pub prefix: Prefix,
    pub marker: Marker,
    pub encoding_type: EncodingType,
    pub is_truncated: IsTruncated,
    pub contents: ObjectList,
    pub common_prefixes: CommonPrefixList,
}

Fields

name

next_marker

When response is truncated (the IsTruncated element value in the response is true), you can use the key name in this field as marker in the subsequent request to get next set of objects. Amazon S3 lists objects in alphabetical order Note: This element is returned only if you have delimiter request parameter specified. If response does not include the NextMaker and it is truncated, you can use the value of the last Key in the response as the marker in the subsequent request to get the next set of object keys.

delimiter

max_keys

prefix

marker

encoding_type

Encoding type used by Amazon S3 to encode object keys in the response.

is_truncated

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria.

contents

common_prefixes