Struct rusoto::s3::ListPartsOutput

pub struct ListPartsOutput {
    pub initiator: Initiator,
    pub bucket: BucketName,
    pub next_part_number_marker: NextPartNumberMarker,
    pub parts: Parts,
    pub upload_id: MultipartUploadId,
    pub storage_class: StorageClass,
    pub key: ObjectKey,
    pub request_charged: RequestCharged,
    pub owner: Owner,
    pub max_parts: MaxParts,
    pub is_truncated: IsTruncated,
    pub part_number_marker: PartNumberMarker,
}

Fields

initiator

Identifies who initiated the multipart upload.

bucket

Name of the bucket to which the multipart upload was initiated.

next_part_number_marker

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

parts

upload_id

Upload ID identifying the multipart upload whose parts are being listed.

storage_class

The class of storage used to store the object.

key

Object key for which the multipart upload was initiated.

request_charged

owner

max_parts

Maximum number of parts that were allowed in the response.

is_truncated

Indicates whether the returned list of parts is truncated.

part_number_marker

Part number after which listing begins.