Struct rusoto::s3::LoggingEnabled

pub struct LoggingEnabled {
    pub target_prefix: TargetPrefix,
    pub target_bucket: TargetBucket,
    pub target_grants: TargetGrants,
}

Fields
target_prefix: TargetPrefix

This element lets you specify a prefix for the keys that the log files will be stored under.

target_bucket: TargetBucket

Specifies the bucket where you want Amazon S3 to store server access logs. You can have your logs delivered to any bucket that you own, including the same bucket that is being logged. You can also configure multiple buckets to deliver their logs to the same target bucket. In this case you should choose a different TargetPrefix for each source bucket so that the delivered log files can be distinguished by key.

target_grants: TargetGrants