Struct rusoto::s3::PutBucketAclRequest

pub struct PutBucketAclRequest {
    pub grant_full_control: Option<GrantFullControl>,
    pub grant_write_acp: Option<GrantWriteACP>,
    pub content_md5: Option<ContentMD5>,
    pub bucket: BucketName,
    pub acl: Option<CannedAcl>,
    pub access_control_policy: Option<AccessControlPolicy>,
    pub grant_write: Option<GrantWrite>,
    pub grant_read: Option<GrantRead>,
    pub grant_read_acp: Option<GrantReadACP>,
}

Fields
grant_full_control: Option<GrantFullControl>

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

grant_write_acp: Option<GrantWriteACP>

Allows grantee to write the ACL for the applicable bucket.

content_md5: Option<ContentMD5> bucket: BucketName acl: Option<CannedAcl>

The canned ACL to apply to the bucket.

access_control_policy: Option<AccessControlPolicy> grant_write: Option<GrantWrite>

Allows grantee to create, overwrite, and delete any object in the bucket.

grant_read: Option<GrantRead>

Allows grantee to list the objects in the bucket.

grant_read_acp: Option<GrantReadACP>

Allows grantee to read the bucket ACL.