Struct rusoto::s3::S3Client

pub struct S3Client<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

Methods

impl<P> S3Client<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> S3Client<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn list_object_versions(&self, input: &ListObjectVersionsRequest) -> Result<ListObjectVersionsOutput, S3Error>

Returns metadata about all of the versions of objects in a bucket.

fn put_bucket_policy(&self, input: &PutBucketPolicyRequest) -> Result<(), S3Error>

Replaces a policy on a bucket. If the bucket already has a policy, the one in this request completely replaces it.

fn list_objects(&self, input: &ListObjectsRequest) -> Result<ListObjectsOutput, S3Error>

Returns some or all (up to 1000) of the objects in a bucket. You can use the request parameters as selection criteria to return a subset of the objects in a bucket.

fn put_bucket_website(&self, input: &PutBucketWebsiteRequest) -> Result<(), S3Error>

Set the website configuration for a bucket.

fn put_bucket_notification(&self, input: &PutBucketNotificationRequest) -> Result<(), S3Error>

Deprecated, see the PutBucketNotificationConfiguraiton operation.

fn put_bucket_logging(&self, input: &PutBucketLoggingRequest) -> Result<(), S3Error>

Set the logging parameters for a bucket and to specify permissions for who can view and modify the logging parameters. To set the logging status of a bucket, you must be the bucket owner.

fn put_bucket_replication(&self, input: &PutBucketReplicationRequest) -> Result<(), S3Error>

Creates a new replication configuration (or replaces an existing one, if present).

fn upload_part(&self, input: &UploadPartRequest) -> Result<String, S3Error>

Uploads a part in a multipart upload. Note: After you initiate multipart upload and upload one or more parts, you must either complete or abort multipart upload in order to stop getting charged for storage of the uploaded parts. Only after you either complete or abort multipart upload, Amazon S3 frees up the parts storage and stops charging you for the parts storage.

fn put_object(&self, input: &PutObjectRequest) -> Result<PutObjectOutput, S3Error>

Adds an object to a bucket.

fn delete_bucket_cors(&self, input: &DeleteBucketCorsRequest) -> Result<(), S3Error>

Deletes the cors configuration information set for the bucket.

fn put_bucket_versioning(&self, input: &PutBucketVersioningRequest) -> Result<(), S3Error>

Sets the versioning state of an existing bucket. To set the versioning state, you must be the bucket owner.

fn get_bucket_cors(&self, input: &GetBucketCorsRequest) -> Result<GetBucketCorsOutput, S3Error>

Returns the cors configuration for the bucket.

fn put_bucket_lifecycle(&self, input: &PutBucketLifecycleRequest) -> Result<(), S3Error>

Sets lifecycle configuration for your bucket. If a lifecycle configuration exists, it replaces it.

fn get_bucket_acl(&self, input: &GetBucketAclRequest) -> Result<GetBucketAclOutput, S3Error>

Gets the access control policy for the bucket.

fn get_bucket_logging(&self, input: &GetBucketLoggingRequest) -> Result<GetBucketLoggingOutput, S3Error>

Returns the logging status of a bucket and the permissions users have to view and modify that status. To use GET, you must be the bucket owner.

fn head_bucket(&self, input: &HeadBucketRequest) -> Result<(), S3Error>

This operation is useful to determine if a bucket exists and you have permission to access it.

fn put_bucket_acl(&self, input: &PutBucketAclRequest) -> Result<(), S3Error>

Sets the permissions on a bucket using access control lists (ACL).

fn delete_bucket_website(&self, input: &DeleteBucketWebsiteRequest) -> Result<(), S3Error>

This operation removes the website configuration from the bucket.

fn delete_bucket_policy(&self, input: &DeleteBucketPolicyRequest) -> Result<(), S3Error>

Deletes the policy from the bucket.

fn get_bucket_notification_configuration(&self, input: &GetBucketNotificationConfigurationRequest) -> Result<NotificationConfiguration, S3Error>

Returns the notification configuration of a bucket.

fn delete_objects(&self, input: &DeleteObjectsRequest) -> Result<DeleteObjectsOutput, S3Error>

This operation enables you to delete multiple objects from a bucket using a single HTTP request. You may specify up to 1000 keys.

fn delete_bucket_replication(&self, input: &DeleteBucketReplicationRequest) -> Result<(), S3Error>

fn copy_object(&self, input: &CopyObjectRequest) -> Result<CopyObjectOutput, S3Error>

Creates a copy of an object that is already stored in Amazon S3.

fn list_buckets(&self) -> Result<ListBucketsOutput, S3Error>

Returns a list of all buckets owned by the authenticated sender of the request.

fn put_bucket_request_payment(&self, input: &PutBucketRequestPaymentRequest) -> Result<(), S3Error>

Sets the request payment configuration for a bucket. By default, the bucket owner pays for downloads from the bucket. This configuration parameter enables the bucket owner (only) to specify that the person requesting the download will be charged for the download. Documentation on requester pays buckets can be found at http://docs.aws.amazon.com/AmazonS3/latest/dev/RequesterPaysBuckets.html

fn put_bucket_notification_configuration(&self, input: &PutBucketNotificationConfigurationRequest) -> Result<(), S3Error>

Enables notifications of specified events for a bucket.

fn head_object(&self, input: &HeadObjectRequest) -> Result<HeadObjectOutput, S3Error>

The HEAD operation retrieves metadata from an object without returning the object itself. This operation is useful if you're only interested in an object's metadata. To use HEAD, you must have READ access to the object.

fn delete_bucket_tagging(&self, input: &DeleteBucketTaggingRequest) -> Result<(), S3Error>

Deletes the tags from the bucket.

fn get_object_torrent(&self, input: &GetObjectTorrentRequest) -> Result<GetObjectTorrentOutput, S3Error>

Return torrent files from a bucket.

fn get_bucket_lifecycle(&self, input: &GetBucketLifecycleRequest) -> Result<GetBucketLifecycleOutput, S3Error>

Returns the lifecycle configuration information set on the bucket.

fn create_bucket(&self, input: &CreateBucketRequest) -> Result<CreateBucketOutput, S3Error>

Creates a new bucket. All requests go to the us-east-1/us-standard endpoint, but can create buckets anywhere.

fn complete_multipart_upload(&self, input: &CompleteMultipartUploadRequest) -> Result<CompleteMultipartUploadOutput, S3Error>

Completes a multipart upload by assembling previously uploaded parts.

fn get_bucket_website(&self, input: &GetBucketWebsiteRequest) -> Result<GetBucketWebsiteOutput, S3Error>

Returns the website configuration for a bucket.

fn create_multipart_upload(&self, input: &CreateMultipartUploadRequest) -> Result<CreateMultipartUploadOutput, S3Error>

Initiates a multipart upload and returns an upload ID. Note: After you initiate multipart upload and upload one or more parts, you must either complete or abort multipart upload in order to stop getting charged for storage of the uploaded parts. Only after you either complete or abort multipart upload, Amazon S3 frees up the parts storage and stops charging you for the parts storage.

fn delete_bucket(&self, input: &DeleteBucketRequest, region: Region) -> Result<(), S3Error>

Deletes the bucket. All objects (including all object versions and Delete Markers) in the bucket must be deleted before the bucket itself can be deleted.

fn get_value_for_header(header_name: String, response: &HttpResponse) -> Result<String, S3Error>

fn head_object_from_response(response: &mut HttpResponse) -> Result<HeadObjectOutput, S3Error>

fn get_object_from_response(response: &mut HttpResponse) -> Result<GetObjectOutput, S3Error>

Use the Hyper resposne to populate the GetObjectOutput

fn get_object(&self, input: &GetObjectRequest) -> Result<GetObjectOutput, S3Error>

Retrieves objects from Amazon S3.

fn get_bucket_policy(&self, input: &GetBucketPolicyRequest) -> Result<GetBucketPolicyOutput, S3Error>

Returns the policy of a specified bucket.

fn get_bucket_versioning(&self, input: &GetBucketVersioningRequest) -> Result<GetBucketVersioningOutput, S3Error>

Returns the versioning state of a bucket.

fn list_multipart_uploads(&self, input: &ListMultipartUploadsRequest) -> Result<ListMultipartUploadsOutput, S3Error>

This operation lists in-progress multipart uploads.

fn get_bucket_request_payment(&self, input: &GetBucketRequestPaymentRequest) -> Result<GetBucketRequestPaymentOutput, S3Error>

Returns the request payment configuration of a bucket.

fn put_bucket_tagging(&self, input: &PutBucketTaggingRequest) -> Result<(), S3Error>

Sets the tags for a bucket.

fn get_bucket_tagging(&self, input: &GetBucketTaggingRequest) -> Result<GetBucketTaggingOutput, S3Error>

Returns the tag set associated with the bucket.

fn abort_multipart_upload(&self, input: &AbortMultipartUploadRequest) -> Result<AbortMultipartUploadOutput, S3Error>

Aborts a multipart upload. To verify that all parts have been removed, so you don't get charged for the part storage, you should call the List Parts operation and ensure the parts list is empty.

fn put_object_acl(&self, input: &PutObjectAclRequest) -> Result<PutObjectAclOutput, S3Error>

uses the acl subresource to set the access control list (ACL) permissions for an object that already exists in a bucket

fn get_bucket_location(&self, input: &GetBucketLocationRequest) -> Result<GetBucketLocationOutput, S3Error>

Returns the region the bucket resides in.

fn put_bucket_cors(&self, input: &PutBucketCorsRequest) -> Result<(), S3Error>

Sets the cors configuration for a bucket.

fn delete_bucket_lifecycle(&self, input: &DeleteBucketLifecycleRequest) -> Result<(), S3Error>

Deletes the lifecycle configuration from the bucket.

fn get_bucket_notification(&self, input: &GetBucketNotificationConfigurationRequest) -> Result<NotificationConfigurationDeprecated, S3Error>

Deprecated, see the GetBucketNotificationConfiguration operation.

fn list_parts(&self, input: &ListPartsRequest) -> Result<ListPartsOutput, S3Error>

Lists the parts that have been uploaded for a specific multipart upload.

fn get_object_acl(&self, input: &GetObjectAclRequest) -> Result<GetObjectAclOutput, S3Error>

Returns the access control list (ACL) of an object.

fn delete_object(&self, input: &DeleteObjectRequest) -> Result<DeleteObjectOutput, S3Error>

Uploads a part by copying data from an existing object as data source. Removes the null version (if there is one) of an object and inserts a delete marker, which becomes the latest version of the object. If there isn't a null version, Amazon S3 does not remove any objects.

fn restore_object(&self, input: &RestoreObjectRequest) -> Result<RestoreObjectOutput, S3Error>

Restores an archived copy of an object back into Amazon S3

fn get_bucket_replication(&self, input: &GetBucketReplicationRequest) -> Result<GetBucketReplicationOutput, S3Error>