Struct rusoto::s3::S3Helper

pub struct S3Helper<P> where P: ProvideAwsCredentials {
    // some fields omitted
}

Wraps the generated S3 client with a higher level interface

Methods

impl<P> S3Helper<P> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> S3Helper<P>

Creates a new S3 helper

fn list_buckets(&self) -> Result<ListBucketsOutput, S3Error>

Lists buckets

fn create_bucket(&self, bucket_name: &str, canned_acl: Option<CannedAcl>) -> Result<CreateBucketOutput, S3Error>

Creates bucket in default us-east-1/us-standard region.

fn create_bucket_in_region(&self, bucket_name: &str, region: Region, canned_acl: Option<CannedAcl>) -> Result<CreateBucketOutput, S3Error>

Creates bucket in specified region.

fn delete_bucket(&self, bucket_name: &str, region: Region) -> Result<(), S3Error>

Deletes specified bucket

fn get_object(&self, bucket_name: &str, object_name: &str) -> Result<GetObjectOutput, S3Error>

Download a named object from bucket

fn put_object(&self, bucket_name: &str, object_name: &str, object_as_bytes: &[u8]) -> Result<PutObjectOutput, S3Error>

Upload an object to specified bucket

fn put_object_with_reduced_redundancy(&self, bucket_name: &str, object_name: &str, object_as_bytes: &[u8]) -> Result<PutObjectOutput, S3Error>

Helper: uploads object to specified bucket using reduced redudancy storage settings

fn put_object_with_aws_encryption(&self, bucket_name: &str, object_name: &str, object_as_bytes: &[u8]) -> Result<PutObjectOutput, S3Error>

Uploads object to specified S3 bucket with server side encryption at rest.

fn put_object_with_kms_encryption(&self, bucket_name: &str, object_name: &str, object_as_bytes: &[u8], key_id: &str) -> Result<PutObjectOutput, S3Error>

Uploads object to specified S3 bucket using AWS KMS for key management of encryption at rest.

fn put_object_with_request(&self, request: &mut PutObjectRequest) -> Result<PutObjectOutput, S3Error>

Uploads object: lets sender specify options. The most generic of put_object: caller specifies the whole request.

fn put_multipart_object<T: Read>(&self, bucket_name: &str, object_name: &str, object_as_reader: &mut T) -> Result<PutObjectOutput, S3Error>

Uploads a multi-part object to specified bucket. Allows for large file uploads.

fn list_multipart_uploads_for_bucket(&self, bucket_name: &str) -> Result<ListMultipartUploadsOutput, S3Error>

Lists multipart uploads not yet completed for specified bucket

fn delete_object(&self, bucket_name: &str, object_name: &str) -> Result<DeleteObjectOutput, S3Error>

Deletes specified object from specified bucket.

fn abort_multipart_upload(&self, bucket_name: &str, object_name: &str, upload_id: &str) -> Result<AbortMultipartUploadOutput, S3Error>

Abort multipart upload.

fn multipart_upload_list_parts(&self, bucket_name: &str, object_name: &str, upload_id: &str) -> Result<ListPartsOutput, S3Error>

List parts from a multiupload request.