Struct rusoto::s3::UploadPartCopyOutput

pub struct UploadPartCopyOutput {
    pub sse_customer_algorithm: SSECustomerAlgorithm,
    pub copy_source_version_id: CopySourceVersionId,
    pub server_side_encryption: ServerSideEncryption,
    pub request_charged: RequestCharged,
    pub sse_customer_key_md5: SSECustomerKeyMD5,
    pub copy_part_result: CopyPartResult,
    pub ssekms_key_id: SSEKMSKeyId,
}

Fields
sse_customer_algorithm: SSECustomerAlgorithm

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

copy_source_version_id: CopySourceVersionId

The version of the source object that was copied, if you have enabled versioning on the source bucket.

server_side_encryption: ServerSideEncryption

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

request_charged: RequestCharged sse_customer_key_md5: SSECustomerKeyMD5

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

copy_part_result: CopyPartResult ssekms_key_id: SSEKMSKeyId

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.