Struct rusoto::s3::UploadPartCopyRequest

pub struct UploadPartCopyRequest {
    pub copy_source_if_match: Option<CopySourceIfMatch>,
    pub sse_customer_algorithm: Option<SSECustomerAlgorithm>,
    pub copy_source_sse_customer_key_md5: Option<CopySourceSSECustomerKeyMD5>,
    pub request_payer: Option<RequestPayer>,
    pub copy_source_sse_customer_key: Option<CopySourceSSECustomerKey>,
    pub copy_source_sse_customer_algorithm: Option<CopySourceSSECustomerAlgorithm>,
    pub copy_source: CopySource,
    pub copy_source_if_modified_since: Option<CopySourceIfModifiedSince>,
    pub bucket: BucketName,
    pub sse_customer_key: Option<SSECustomerKey>,
    pub copy_source_if_unmodified_since: Option<CopySourceIfUnmodifiedSince>,
    pub copy_source_if_none_match: Option<CopySourceIfNoneMatch>,
    pub upload_id: MultipartUploadId,
    pub key: ObjectKey,
    pub copy_source_range: Option<CopySourceRange>,
    pub sse_customer_key_md5: Option<SSECustomerKeyMD5>,
    pub part_number: PartNumber,
}

Fields

copy_source_if_match

Copies the object if its entity tag (ETag) matches the specified tag.

sse_customer_algorithm

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

copy_source_sse_customer_key_md5

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

request_payer

copy_source_sse_customer_key

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

copy_source_sse_customer_algorithm

Specifies the algorithm to use when decrypting the source object (e.g., AES256).

copy_source

The name of the source bucket and key name of the source object, separated by a slash (/). Must be URL-encoded.

copy_source_if_modified_since

Copies the object if it has been modified since the specified time.

bucket

sse_customer_key

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side- encryption-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

copy_source_if_unmodified_since

Copies the object if it hasn't been modified since the specified time.

copy_source_if_none_match

Copies the object if its entity tag (ETag) is different than the specified ETag.

upload_id

Upload ID identifying the multipart upload whose part is being copied.

key

copy_source_range

The range of bytes to copy from the source object. The range value must use the form bytes=first-last, where the first and last are the zero-based byte offsets to copy. For example, bytes=0-9 indicates that you want to copy the first ten bytes of the source. You can copy a range only if the source object is greater than 5 GB.

sse_customer_key_md5

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

part_number

Part number of part being copied. This is a positive integer between 1 and 10,000.