Module rusoto::sdb

Amazon SimpleDB

If you're using the service, you're probably looking for SimpleDbClient.

Structs

Attribute

BatchDeleteAttributesRequest

BatchPutAttributesRequest

CreateDomainRequest

DeletableItem

DeleteAttributesRequest

DeleteDomainRequest

DomainMetadataRequest

DomainMetadataResult

GetAttributesRequest

GetAttributesResult

Item

ListDomainsRequest

ListDomainsResult

PutAttributesRequest

ReplaceableAttribute

ReplaceableItem

SelectRequest

SelectResult

SimpleDbClient
A client for the Amazon SimpleDB API.

UpdateCondition
Specifies the conditions under which data should be updated. If an update condition is specified for a request, the data will only be updated if the condition is satisfied. For example, if an attribute with a specific name and value exists, or if a specific attribute doesn't exist.

Enums

BatchDeleteAttributesError
Errors returned by BatchDeleteAttributes

BatchPutAttributesError
Errors returned by BatchPutAttributes

CreateDomainError
Errors returned by CreateDomain

DeleteAttributesError
Errors returned by DeleteAttributes

DeleteDomainError
Errors returned by DeleteDomain

DomainMetadataError
Errors returned by DomainMetadata

GetAttributesError
Errors returned by GetAttributes

ListDomainsError
Errors returned by ListDomains

PutAttributesError
Errors returned by PutAttributes

SelectError
Errors returned by Select

Type Definitions

AttributeList

AttributeNameList

Boolean

DeletableItemList

DomainNameList

Float

Integer

ItemList

Long

ReplaceableAttributeList

ReplaceableItemList