Struct rusoto::signature::SignedRequest

pub struct SignedRequest<'a> {
    // some fields omitted
}

A data structure for all the elements of an HTTP request that are involved in the Amazon Signature Version 4 signing process

Methods

impl<'a> SignedRequest<'a>

fn new<'b>(method: &str, service: &str, region: &'a Region, path: &str) -> SignedRequest<'a>

Default constructor

fn set_content_type(&mut self, content_type: String)

fn set_hostname(&mut self, hostname: Option<String>)

fn set_payload(&mut self, payload: Option<&'a [u8]>)

fn get_method(&self) -> &str

fn get_canonical_uri(&self) -> &str

fn get_canonical_query_string(&self) -> &str

fn get_payload(&self) -> Option<&'a [u8]>

fn get_headers(&'a self) -> &'a BTreeMap<String, Vec<Vec<u8>>>

fn get_hostname(&self) -> String

fn remove_header(&mut self, key: &str)

fn add_header(&mut self, key: &str, value: &str)

Add a value to the array of headers for the specified key. Headers are kept sorted by key name for use at signing (BTreeMap)

fn add_param<S>(&mut self, key: S, value: S) where S: Into<String>

fn set_params(&mut self, params: Params)

fn sign_and_execute(&mut self, creds: &AWSCredentials) -> Response

Calculate the signature from the credentials provided and the request data Add the calculated signature to the request headers and execute it Return the hyper HTTP response