Module rusoto::sns

Amazon Simple Notification Service

If you're using the service, you're probably looking for SnsClient.

Structs

AddPermissionInput

CheckIfPhoneNumberIsOptedOutInput
The input for the CheckIfPhoneNumberIsOptedOut action.

CheckIfPhoneNumberIsOptedOutResponse
The response from the CheckIfPhoneNumberIsOptedOut action.

ConfirmSubscriptionInput
Input for ConfirmSubscription action.

ConfirmSubscriptionResponse
Response for ConfirmSubscriptions action.

CreateEndpointResponse
Response from CreateEndpoint action.

CreatePlatformApplicationInput
Input for CreatePlatformApplication action.

CreatePlatformApplicationResponse
Response from CreatePlatformApplication action.

CreatePlatformEndpointInput
Input for CreatePlatformEndpoint action.

CreateTopicInput
Input for CreateTopic action.

CreateTopicResponse
Response from CreateTopic action.

DeleteEndpointInput
Input for DeleteEndpoint action.

DeletePlatformApplicationInput
Input for DeletePlatformApplication action.

DeleteTopicInput

GetEndpointAttributesInput
Input for GetEndpointAttributes action.

GetEndpointAttributesResponse
Response from GetEndpointAttributes of the EndpointArn.

GetPlatformApplicationAttributesInput
Input for GetPlatformApplicationAttributes action.

GetPlatformApplicationAttributesResponse
Response for GetPlatformApplicationAttributes action.

GetSMSAttributesInput
The input for the GetSMSAttributes request.

GetSMSAttributesResponse
The response from the GetSMSAttributes request.

GetSubscriptionAttributesInput
Input for GetSubscriptionAttributes.

GetSubscriptionAttributesResponse
Response for GetSubscriptionAttributes action.

GetTopicAttributesInput
Input for GetTopicAttributes action.

GetTopicAttributesResponse
Response for GetTopicAttributes action.

ListEndpointsByPlatformApplicationInput
Input for ListEndpointsByPlatformApplication action.

ListEndpointsByPlatformApplicationResponse
Response for ListEndpointsByPlatformApplication action.

ListPhoneNumbersOptedOutInput
The input for the ListPhoneNumbersOptedOut action.

ListPhoneNumbersOptedOutResponse
The response from the ListPhoneNumbersOptedOut action.

ListPlatformApplicationsInput
Input for ListPlatformApplications action.

ListPlatformApplicationsResponse
Response for ListPlatformApplications action.

ListSubscriptionsByTopicInput
Input for ListSubscriptionsByTopic action.

ListSubscriptionsByTopicResponse
Response for ListSubscriptionsByTopic action.

ListSubscriptionsInput
Input for ListSubscriptions action.

ListSubscriptionsResponse
Response for ListSubscriptions action

ListTopicsInput

ListTopicsResponse
Response for ListTopics action.

MessageAttributeValue
The user-specified message attribute value. For string data types, the value attribute has the same restrictions on the content as the message body. For more information, see Publish.

Name, type, and value must not be empty or null. In addition, the message body should not be empty or null. All parts of the message attribute, including name, type, and value, are included in the message size restriction, which is currently 256 KB (262,144 bytes). For more information, see Using Amazon SNS Message Attributes.

OptInPhoneNumberInput
Input for the OptInPhoneNumber action.

OptInPhoneNumberResponse
The response for the OptInPhoneNumber action.

PlatformApplication
Platform application object.

PublishInput
Input for Publish action.

PublishResponse
Response for Publish action.

RemovePermissionInput
Input for RemovePermission action.

SetEndpointAttributesInput
Input for SetEndpointAttributes action.

SetPlatformApplicationAttributesInput
Input for SetPlatformApplicationAttributes action.

SetSMSAttributesInput
The input for the SetSMSAttributes action.

SetSMSAttributesResponse
The response for the SetSMSAttributes action.

SetSubscriptionAttributesInput
Input for SetSubscriptionAttributes action.

SetTopicAttributesInput
Input for SetTopicAttributes action.

SnsClient
A client for the Amazon SNS API.

SubscribeInput
Input for Subscribe action.

SubscribeResponse
Response for Subscribe action.

Subscription
A wrapper type for the attributes of an Amazon SNS subscription.

Topic
A wrapper type for the topic's Amazon Resource Name (ARN). To retrieve a topic's attributes, use GetTopicAttributes.

UnsubscribeInput
Input for Unsubscribe action.

Enums

AddPermissionError
Errors returned by AddPermission

CheckIfPhoneNumberIsOptedOutError
Errors returned by CheckIfPhoneNumberIsOptedOut

ConfirmSubscriptionError
Errors returned by ConfirmSubscription

CreatePlatformApplicationError
Errors returned by CreatePlatformApplication

CreatePlatformEndpointError
Errors returned by CreatePlatformEndpoint

CreateTopicError
Errors returned by CreateTopic

DeleteEndpointError
Errors returned by DeleteEndpoint

DeletePlatformApplicationError
Errors returned by DeletePlatformApplication

DeleteTopicError
Errors returned by DeleteTopic

GetEndpointAttributesError
Errors returned by GetEndpointAttributes

GetPlatformApplicationAttributesError
Errors returned by GetPlatformApplicationAttributes

GetSMSAttributesError
Errors returned by GetSMSAttributes

GetSubscriptionAttributesError
Errors returned by GetSubscriptionAttributes

GetTopicAttributesError
Errors returned by GetTopicAttributes

ListEndpointsByPlatformApplicationError
Errors returned by ListEndpointsByPlatformApplication

ListPhoneNumbersOptedOutError
Errors returned by ListPhoneNumbersOptedOut

ListPlatformApplicationsError
Errors returned by ListPlatformApplications

ListSubscriptionsByTopicError
Errors returned by ListSubscriptionsByTopic

ListSubscriptionsError
Errors returned by ListSubscriptions

ListTopicsError
Errors returned by ListTopics

OptInPhoneNumberError
Errors returned by OptInPhoneNumber

PublishError
Errors returned by Publish

RemovePermissionError
Errors returned by RemovePermission

SetEndpointAttributesError
Errors returned by SetEndpointAttributes

SetPlatformApplicationAttributesError
Errors returned by SetPlatformApplicationAttributes

SetSMSAttributesError
Errors returned by SetSMSAttributes

SetSubscriptionAttributesError
Errors returned by SetSubscriptionAttributes

SetTopicAttributesError
Errors returned by SetTopicAttributes

SubscribeError
Errors returned by Subscribe

UnsubscribeError
Errors returned by Unsubscribe

Type Definitions

Account

Action

ActionsList

AttributeName

AttributeValue

AuthenticateOnUnsubscribe

Binary

Boolean

Delegate

DelegatesList

Endpoint

Label

ListOfEndpoints

ListOfPlatformApplications

ListString

MapStringToString

Message

MessageAttributeMap

MessageId

MessageStructure

NextToken

PhoneNumber

PhoneNumberList

Protocol

Subject

SubscriptionARN

SubscriptionAttributesMap

SubscriptionsList

Token

TopicARN

TopicAttributesMap

TopicName

TopicsList