Module rusoto::sqs

SQS bindings for Rust

Check SQSHelper for convenience functions.

Structs

AddPermissionRequest

BatchEntryIdsNotDistinct
Two or more batch entries have the same Id in the request.

BatchRequestTooLong
The length of all the messages put together is more than the limit.

BatchResultErrorEntry
This is used in the responses of batch API to give a detailed description of the result of an action on each entry in the request.

ChangeMessageVisibilityBatchRequest

ChangeMessageVisibilityBatchRequestEntry
Encloses a receipt handle and an entry id for each message in ChangeMessageVisibilityBatch. All of the following parameters are list parameters that must be prefixed with ChangeMessageVisibilityBatchRequestEntry.n, where n is an integer value starting with 1. For example, a parameter list for this action might look like this: &ChangeMessageVisibilityBatchRequestEntry.1.Id=change_visibility_msg_2 &ChangeMessageVisibilityBatchRequestEntry.1.ReceiptHandle=Your_Receipt_Handle &ChangeMessageVisibilityBatchRequestEntry.1.VisibilityTimeout=45

ChangeMessageVisibilityBatchResult
For each message in the batch, the response contains a ChangeMessageVisibilityBatchResultEntry tag if the message succeeds or a BatchResultErrorEntry tag if the message fails.

ChangeMessageVisibilityBatchResultEntry
Encloses the id of an entry in ChangeMessageVisibilityBatch.

ChangeMessageVisibilityRequest

CreateQueueRequest

CreateQueueResult
Returns the QueueUrl element of the created queue.

DeleteMessageBatchRequest

DeleteMessageBatchRequestEntry
Encloses a receipt handle and an identifier for it.

DeleteMessageBatchResult
For each message in the batch, the response contains a DeleteMessageBatchResultEntry tag if the message is deleted or a BatchResultErrorEntry tag if the message cannot be deleted.

DeleteMessageBatchResultEntry
Encloses the id an entry in DeleteMessageBatch.

DeleteMessageRequest

DeleteQueueRequest

EmptyBatchRequest
Batch request does not contain an entry.

GetQueueAttributesRequest

GetQueueAttributesResult
A list of returned queue attributes.

GetQueueUrlRequest

GetQueueUrlResult
For more information, see Responses in the Amazon SQS Developer Guide.

InvalidAttributeName
The attribute referred to does not exist.

InvalidBatchEntryId
The Id of a batch entry in a batch request does not abide by the specification.

InvalidIdFormat
The receipt handle is not valid for the current version.

InvalidMessageContents
The message contains characters outside the allowed set.

ListDeadLetterSourceQueuesRequest

ListDeadLetterSourceQueuesResult
A list of your dead letter source queues.

ListQueuesRequest

ListQueuesResult
A list of your queues.

Message
An Amazon SQS message.

MessageAttributeValue
The user-specified message attribute value. For string data types, the value attribute has the same restrictions on the content as the message body. For more information, see SendMessage. Name, type, and value must not be empty or null. In addition, the message body should not be empty or null. All parts of the message attribute, including name, type, and value, are included in the message size restriction, which is currently 256 KB (262,144 bytes).

MessageNotInflight
The message referred to is not in flight.

OverLimit
The action that you requested would violate a limit. For example, ReceiveMessage returns this error if the maximum number of messages inflight has already been reached. AddPermission returns this error if the maximum number of permissions for the queue has already been reached.

PurgeQueueInProgress
Indicates that the specified queue previously received a PurgeQueue request within the last 60 seconds, the time it can take to delete the messages in the queue.

PurgeQueueRequest

QueueDeletedRecently
You must wait 60 seconds after deleting a queue before you can create another with the same name.

QueueDoesNotExist
The queue referred to does not exist.

QueueNameExists
A queue already exists with this name. Amazon SQS returns this error only if the request includes attributes whose values differ from those of the existing queue.

ReceiptHandleIsInvalid
The receipt handle provided is not valid.

ReceiveMessageRequest

ReceiveMessageResult
A list of received messages.

RemovePermissionRequest

SQSClient

SQSHelper
Easier to use SQS client: wraps SQSClient class.

SendMessageBatchRequest

SendMessageBatchRequestEntry
Contains the details of a single Amazon SQS message along with a Id.

SendMessageBatchResult
For each message in the batch, the response contains a SendMessageBatchResultEntry tag if the message succeeds or a BatchResultErrorEntry tag if the message fails.

SendMessageBatchResultEntry
Encloses a message ID for successfully enqueued message of a SendMessageBatch.

SendMessageRequest

SendMessageResult
The MD5OfMessageBody and MessageId elements.

SetQueueAttributesRequest

TooManyEntriesInBatchRequest
Batch request contains more number of entries than permissible.

UnsupportedOperation
Error code 400. Unsupported operation.

Type Definitions

AWSAccountIdList

ActionNameList

AttributeMap

AttributeNameList

BatchResultErrorEntryList

Binary

BinaryList

Boolean

ChangeMessageVisibilityBatchRequestEntryList

ChangeMessageVisibilityBatchResultEntryList

DeleteMessageBatchRequestEntryList

DeleteMessageBatchResultEntryList

Integer

MessageAttributeMap

MessageAttributeName

MessageAttributeNameList

MessageList

QueueAttributeName

QueueUrlList

SendMessageBatchRequestEntryList

SendMessageBatchResultEntryList

StringList