Struct rusoto::sqs::ChangeMessageVisibilityBatchRequestEntry

pub struct ChangeMessageVisibilityBatchRequestEntry {
    pub receipt_handle: String,
    pub visibility_timeout: Option<Integer>,
    pub id: String,
}

Encloses a receipt handle and an entry id for each message in ChangeMessageVisibilityBatch. All of the following parameters are list parameters that must be prefixed with ChangeMessageVisibilityBatchRequestEntry.n, where n is an integer value starting with 1. For example, a parameter list for this action might look like this: &ChangeMessageVisibilityBatchRequestEntry.1.Id=change_visibility_msg_2 &ChangeMessageVisibilityBatchRequestEntry.1.ReceiptHandle=Your_Receipt_Handle &ChangeMessageVisibilityBatchRequestEntry.1.VisibilityTimeout=45

Fields

receipt_handle

A receipt handle.

visibility_timeout

The new value (in seconds) for the message's visibility timeout.

id

An identifier for this particular receipt handle. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.