Struct rusoto::sqs::Message

pub struct Message {
    pub attributes: Option<AttributeMap>,
    pub body: Option<String>,
    pub m_d_5_of_body: Option<String>,
    pub m_d_5_of_message_attributes: Option<String>,
    pub message_attributes: Option<MessageAttributeMap>,
    pub message_id: Option<String>,
    pub receipt_handle: Option<String>,
}

An Amazon SQS message.

Fields
attributes: Option<AttributeMap>

SenderId, SentTimestamp, ApproximateReceiveCount, and/or ApproximateFirstReceiveTimestamp. SentTimestamp and ApproximateFirstReceiveTimestamp are each returned as an integer representing the epoch time in milliseconds.

body: Option<String>

The message's contents (not URL-encoded).

m_d_5_of_body: Option<String>

An MD5 digest of the non-URL-encoded message body string.

m_d_5_of_message_attributes: Option<String>

An MD5 digest of the non-URL-encoded message attribute string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http://www.faqs.org/rfcs/rfc1321.html.

message_attributes: Option<MessageAttributeMap>

Each message attribute consists of a Name, Type, and Value. For more information, see Message Attribute Items.

message_id: Option<String>

A unique identifier for the message. Message IDs are considered unique across all AWS accounts for an extended period of time.

receipt_handle: Option<String>

An identifier associated with the act of receiving the message. A new receipt handle is returned every time you receive a message. When deleting a message, you provide the last received receipt handle to delete the message.