Struct rusoto::sqs::MessageAttributeValue

pub struct MessageAttributeValue {
    pub binary_list_values: Option<BinaryList>,
    pub string_value: Option<String>,
    pub data_type: String,
    pub binary_value: Option<Binary>,
    pub string_list_values: Option<StringList>,
}

The user-specified message attribute value. For string data types, the value attribute has the same restrictions on the content as the message body. For more information, see SendMessage. Name, type, and value must not be empty or null. In addition, the message body should not be empty or null. All parts of the message attribute, including name, type, and value, are included in the message size restriction, which is currently 256 KB (262,144 bytes).

Fields

binary_list_values

Not implemented. Reserved for future use.

string_value

Strings are Unicode with UTF8 binary encoding. For a list of code values, see http://en.wikipedia.org/wiki/ASCII#ASCII_printable_characters.

data_type

Amazon SQS supports the following logical data types: String, Number, and Binary. In addition, you can append your own custom labels. For more information, see Message Attribute Data Types.

binary_value

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.

string_list_values

Not implemented. Reserved for future use.