Struct rusoto::sqs::ReceiveMessageRequest

pub struct ReceiveMessageRequest {
    pub queue_url: String,
    pub max_number_of_messages: Option<Integer>,
    pub wait_time_seconds: Option<Integer>,
    pub message_attribute_names: Option<MessageAttributeNameList>,
    pub visibility_timeout: Option<Integer>,
    pub attribute_names: Option<AttributeNameList>,
}

Fields

queue_url

The URL of the Amazon SQS queue to take action on.

max_number_of_messages

The maximum number of messages to return. Amazon SQS never returns more messages than this value but may return fewer. Values can be from 1 to 10. Default is 1. All of the messages are not necessarily returned.

wait_time_seconds

The duration (in seconds) for which the call will wait for a message to arrive in the queue before returning. If a message is available, the call will return sooner than WaitTimeSeconds.

message_attribute_names

The name of the message attribute, where N is the index. The message attribute name can contain the following characters: A-Z, a-z, 0-9, underscore (_), hyphen (-), and period (.). The name must not start or end with a period, and it should not have successive periods. The name is case sensitive and must be unique among all attribute names for the message. The name can be up to 256 characters long. The name cannot start with "AWS." or "Amazon." (or any variations in casing), because these prefixes are reserved for use by Amazon Web Services. When using ReceiveMessage, you can send a list of attribute names to receive, or you can return all of the attributes by specifying "All" or ".*" in your request. You can also use "foo.*" to return all message attributes starting with the "foo" prefix.

visibility_timeout

The duration (in seconds) that the received messages are hidden from subsequent retrieve requests after being retrieved by a ReceiveMessage request.

attribute_names

A list of attributes that need to be returned along with each message. The following lists the names and descriptions of the attributes that can be returned: * All - returns all values. * ApproximateFirstReceiveTimestamp - returns the time when the message was first received from the queue (epoch time in milliseconds). * ApproximateReceiveCount - returns the number of times a message has been received from the queue but not deleted. * SenderId - returns the AWS account number (or the IP address, if anonymous access is allowed) of the sender. * SentTimestamp - returns the time when the message was sent to the queue (epoch time in milliseconds).