Struct rusoto::ssm::CancelCommandRequest

pub struct CancelCommandRequest {
    pub command_id: CommandId,
    pub instance_ids: Option<InstanceIdList>,
}

Fields
command_id: CommandId

The ID of the command you want to cancel.

instance_ids: Option<InstanceIdList>

(Optional) A list of instance IDs on which you want to cancel the command. If not provided, the command is canceled on every instance on which it was requested.