Struct rusoto::ssm::Command

pub struct Command {
    pub command_id: Option<CommandId>,
    pub comment: Option<Comment>,
    pub document_name: Option<DocumentName>,
    pub expires_after: Option<DateTime>,
    pub instance_ids: Option<InstanceIdList>,
    pub notification_config: Option<NotificationConfig>,
    pub output_s3_bucket_name: Option<S3BucketName>,
    pub output_s3_key_prefix: Option<S3KeyPrefix>,
    pub parameters: Option<Parameters>,
    pub requested_date_time: Option<DateTime>,
    pub service_role: Option<ServiceRole>,
    pub status: Option<CommandStatus>,
}

Describes a command request.

Fields
command_id: Option<CommandId>

A unique identifier for this command.

comment: Option<Comment>

User-specified information about the command, such as a brief description of what the command should do.

document_name: Option<DocumentName>

The name of the SSM document requested for execution.

expires_after: Option<DateTime>

If this time is reached and the command has not already started executing, it will not execute. Calculated based on the ExpiresAfter user input provided as part of the SendCommand API.

instance_ids: Option<InstanceIdList>

The instance IDs against which this command was requested.

notification_config: Option<NotificationConfig>

Configurations for sending notifications about command status changes.

output_s3_bucket_name: Option<S3BucketName>

The S3 bucket where the responses to the command executions should be stored. This was requested when issuing the command.

output_s3_key_prefix: Option<S3KeyPrefix>

The S3 directory path inside the bucket where the responses to the command executions should be stored. This was requested when issuing the command.

parameters: Option<Parameters>

The parameter values to be inserted in the SSM document when executing the command.

requested_date_time: Option<DateTime>

The date and time the command was requested.

service_role: Option<ServiceRole>

The IAM service role that SSM uses to act on your behalf when sending notifications about command status changes.

status: Option<CommandStatus>

The status of the command.