Struct rusoto::ssm::CommandInvocation

pub struct CommandInvocation {
    pub command_id: Option<CommandId>,
    pub command_plugins: Option<CommandPluginList>,
    pub comment: Option<Comment>,
    pub document_name: Option<DocumentName>,
    pub instance_id: Option<InstanceId>,
    pub requested_date_time: Option<DateTime>,
    pub status: Option<CommandInvocationStatus>,
    pub trace_output: Option<InvocationTraceOutput>,
}

An invocation is copy of a command sent to a specific instance. A command can apply to one or more instances. A command invocation applies to one instance. For example, if a user executes SendCommand against three instances, then a command invocation is created for each requested instance ID. A command invocation returns status and detail information about a command you executed.

Fields
command_id: Option<CommandId>

The command against which this invocation was requested.

command_plugins: Option<CommandPluginList> comment: Option<Comment>

User-specified information about the command, such as a brief description of what the command should do.

document_name: Option<DocumentName>

The document name that was requested for execution.

instance_id: Option<InstanceId>

The instance ID in which this invocation was requested.

requested_date_time: Option<DateTime>

The time and date the request was sent to this instance.

status: Option<CommandInvocationStatus>

Whether or not the invocation succeeded, failed, or is pending.

trace_output: Option<InvocationTraceOutput>

Gets the trace output sent by the agent.