Struct rusoto::ssm::DocumentDescription

pub struct DocumentDescription {
    pub created_date: Option<DateTime>,
    pub description: Option<DescriptionInDocument>,
    pub hash: Option<DocumentHash>,
    pub hash_type: Option<DocumentHashType>,
    pub name: Option<DocumentARN>,
    pub owner: Option<DocumentOwner>,
    pub parameters: Option<DocumentParameterList>,
    pub platform_types: Option<PlatformTypeList>,
    pub sha_1: Option<DocumentSha1>,
    pub status: Option<DocumentStatus>,
}

Describes an SSM document.

Fields
created_date: Option<DateTime>

The date when the SSM document was created.

description: Option<DescriptionInDocument>

A description of the document.

hash: Option<DocumentHash>

The Sha256 or Sha1 hash created by the system when the document was created.

Sha1 hashes have been deprecated.

hash_type: Option<DocumentHashType>

Sha256 or Sha1.

Sha1 hashes have been deprecated.

name: Option<DocumentARN>

The name of the SSM document.

owner: Option<DocumentOwner>

The AWS user account of the person who created the document.

parameters: Option<DocumentParameterList>

A description of the parameters for a document.

platform_types: Option<PlatformTypeList>

The list of OS platforms compatible with this SSM document.

sha_1: Option<DocumentSha1>

The SHA1 hash of the document, which you can use for verification purposes.

status: Option<DocumentStatus>

The status of the SSM document.