Struct rusoto::ssm::ListCommandInvocationsRequest

pub struct ListCommandInvocationsRequest {
    pub command_id: Option<CommandId>,
    pub details: Option<Boolean>,
    pub filters: Option<CommandFilterList>,
    pub instance_id: Option<InstanceId>,
    pub max_results: Option<CommandMaxResults>,
    pub next_token: Option<NextToken>,
}

Fields
command_id: Option<CommandId>

(Optional) The invocations for a specific command ID.

details: Option<Boolean>

(Optional) If set this returns the response of the command executions and any command output. By default this is set to False.

filters: Option<CommandFilterList>

(Optional) One or more filters. Use a filter to return a more specific list of results.

instance_id: Option<InstanceId>

(Optional) The command execution details for a specific instance ID.

max_results: Option<CommandMaxResults>

(Optional) The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

next_token: Option<NextToken>

(Optional) The token for the next set of items to return. (You received this token from a previous call.)