Struct rusoto::ssm::SendCommandRequest

pub struct SendCommandRequest {
    pub comment: Option<Comment>,
    pub document_hash: Option<DocumentHash>,
    pub document_hash_type: Option<DocumentHashType>,
    pub document_name: DocumentARN,
    pub instance_ids: InstanceIdList,
    pub output_s_3_bucket_name: Option<S3BucketName>,
    pub output_s_3_key_prefix: Option<S3KeyPrefix>,
    pub parameters: Option<Parameters>,
    pub timeout_seconds: Option<TimeoutSeconds>,
}

Fields

comment

User-specified information about the command, such as a brief description of what the command should do.

document_hash

The Sha256 or Sha1 hash created by the system when the document was created.

Sha1 hashes have been deprecated.

document_hash_type

Sha256 or Sha1.

Sha1 hashes have been deprecated.

document_name

Required. The name of the SSM document to execute. This can be an SSM public document or a custom document.

instance_ids

Required. The instance IDs where the command should execute.

output_s_3_bucket_name

The name of the S3 bucket where command execution responses should be stored.

output_s_3_key_prefix

The directory structure within the S3 bucket where the responses should be stored.

parameters

The required and optional parameters specified in the SSM document being executed.

timeout_seconds

If this time is reached and the command has not already started executing, it will not execute.