Struct rusoto::ssm::SsmClient

pub struct SsmClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest { /* fields omitted */ }

A client for the Amazon SSM API.

Methods

impl<P> SsmClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> SsmClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn add_tags_to_resource(&self, input: &AddTagsToResourceRequest) -> Result<AddTagsToResourceResult, AddTagsToResourceError>

Adds or overwrites one or more tags for the specified resource. Tags are metadata that you assign to your managed instances. Tags enable you to categorize your managed instances in different ways, for example, by purpose, owner, or environment. Each tag consists of a key and an optional value, both of which you define. For example, you could define a set of tags for your account's managed instances that helps you track each instance's owner and stack level. For example: Key=Owner and Value=DbAdmin, SysAdmin, or Dev. Or Key=Stack and Value=Production, Pre-Production, or Test. Each resource can have a maximum of 10 tags.

We recommend that you devise a set of tag keys that meets your needs for each resource type. Using a consistent set of tag keys makes it easier for you to manage your resources. You can search and filter the resources based on the tags you add. Tags don't have any semantic meaning to Amazon EC2 and are interpreted strictly as a string of characters.

For more information about tags, see Tagging Your Amazon EC2 Resources in the Amazon EC2 User Guide.

fn cancel_command(&self, input: &CancelCommandRequest) -> Result<CancelCommandResult, CancelCommandError>

Attempts to cancel the command specified by the Command ID. There is no guarantee that the command will be terminated and the underlying process stopped.

fn create_activation(&self, input: &CreateActivationRequest) -> Result<CreateActivationResult, CreateActivationError>

Registers your on-premises server or virtual machine with Amazon EC2 so that you can manage these resources using Run Command. An on-premises server or virtual machine that has been registered with EC2 is called a managed instance. For more information about activations, see Setting Up Managed Instances (Linux) or Setting Up Managed Instances (Windows) in the Amazon EC2 User Guide.

fn create_association(&self, input: &CreateAssociationRequest) -> Result<CreateAssociationResult, CreateAssociationError>

Associates the specified SSM document with the specified instance.

When you associate an SSM document with an instance, the configuration agent on the instance (SSM agent for Linux and EC2Config service for Windows) processes the document and configures the instance as specified.

If you associate a document with an instance that already has an associated document, the system throws the AssociationAlreadyExists exception.

fn create_association_batch(&self, input: &CreateAssociationBatchRequest) -> Result<CreateAssociationBatchResult, CreateAssociationBatchError>

Associates the specified SSM document with the specified instances.

When you associate an SSM document with an instance, the configuration agent on the instance (SSM agent for Linux and EC2Config service for Windows) processes the document and configures the instance as specified.

If you associate a document with an instance that already has an associated document, the system throws the AssociationAlreadyExists exception.

fn create_document(&self, input: &CreateDocumentRequest) -> Result<CreateDocumentResult, CreateDocumentError>

Creates an SSM document.

After you create an SSM document, you can use CreateAssociation to associate it with one or more running instances.

fn delete_activation(&self, input: &DeleteActivationRequest) -> Result<DeleteActivationResult, DeleteActivationError>

Deletes an activation. You are not required to delete an activation. If you delete an activation, you can no longer use it to register additional managed instances. Deleting an activation does not de-register managed instances. You must manually de-register managed instances.

fn delete_association(&self, input: &DeleteAssociationRequest) -> Result<DeleteAssociationResult, DeleteAssociationError>

Disassociates the specified SSM document from the specified instance.

When you disassociate an SSM document from an instance, it does not change the configuration of the instance. To change the configuration state of an instance after you disassociate a document, you must create a new document with the desired configuration and associate it with the instance.

fn delete_document(&self, input: &DeleteDocumentRequest) -> Result<DeleteDocumentResult, DeleteDocumentError>

Deletes the SSM document and all instance associations to the document.

Before you delete the SSM document, we recommend that you use DeleteAssociation to disassociate all instances that are associated with the document.

fn deregister_managed_instance(&self, input: &DeregisterManagedInstanceRequest) -> Result<DeregisterManagedInstanceResult, DeregisterManagedInstanceError>

Removes the server or virtual machine from the list of registered servers. You can reregister the instance again at any time. If you don’t plan to use Run Command on the server, we suggest uninstalling the SSM agent first.

fn describe_activations(&self, input: &DescribeActivationsRequest) -> Result<DescribeActivationsResult, DescribeActivationsError>

Details about the activation, including: the date and time the activation was created, the expiration date, the IAM role assigned to the instances in the activation, and the number of instances activated by this registration.

fn describe_association(&self, input: &DescribeAssociationRequest) -> Result<DescribeAssociationResult, DescribeAssociationError>

Describes the associations for the specified SSM document or instance.

fn describe_document(&self, input: &DescribeDocumentRequest) -> Result<DescribeDocumentResult, DescribeDocumentError>

Describes the specified SSM document.

fn describe_document_permission(&self, input: &DescribeDocumentPermissionRequest) -> Result<DescribeDocumentPermissionResponse, DescribeDocumentPermissionError>

Describes the permissions for an SSM document. If you created the document, you are the owner. If a document is shared, it can either be shared privately (by specifying a user’s AWS account ID) or publicly (All).

fn describe_instance_information(&self, input: &DescribeInstanceInformationRequest) -> Result<DescribeInstanceInformationResult, DescribeInstanceInformationError>

Describes one or more of your instances. You can use this to get information about instances like the operating system platform, the SSM agent version (Linux), status etc. If you specify one or more instance IDs, it returns information for those instances. If you do not specify instance IDs, it returns information for all your instances. If you specify an instance ID that is not valid or an instance that you do not own, you receive an error.

fn get_document(&self, input: &GetDocumentRequest) -> Result<GetDocumentResult, GetDocumentError>

Gets the contents of the specified SSM document.

fn list_associations(&self, input: &ListAssociationsRequest) -> Result<ListAssociationsResult, ListAssociationsError>

Lists the associations for the specified SSM document or instance.

fn list_command_invocations(&self, input: &ListCommandInvocationsRequest) -> Result<ListCommandInvocationsResult, ListCommandInvocationsError>

An invocation is copy of a command sent to a specific instance. A command can apply to one or more instances. A command invocation applies to one instance. For example, if a user executes SendCommand against three instances, then a command invocation is created for each requested instance ID. ListCommandInvocations provide status about command execution.

fn list_commands(&self, input: &ListCommandsRequest) -> Result<ListCommandsResult, ListCommandsError>

Lists the commands requested by users of the AWS account.

fn list_documents(&self, input: &ListDocumentsRequest) -> Result<ListDocumentsResult, ListDocumentsError>

Describes one or more of your SSM documents.

fn list_tags_for_resource(&self, input: &ListTagsForResourceRequest) -> Result<ListTagsForResourceResult, ListTagsForResourceError>

Returns a list of the tags assigned to the specified resource.

fn modify_document_permission(&self, input: &ModifyDocumentPermissionRequest) -> Result<ModifyDocumentPermissionResponse, ModifyDocumentPermissionError>

Share a document publicly or privately. If you share a document privately, you must specify the AWS user account IDs for those people who can use the document. If you share a document publicly, you must specify All as the account ID.

fn remove_tags_from_resource(&self, input: &RemoveTagsFromResourceRequest) -> Result<RemoveTagsFromResourceResult, RemoveTagsFromResourceError>

Removes all tags from the specified resource.

fn send_command(&self, input: &SendCommandRequest) -> Result<SendCommandResult, SendCommandError>

Executes commands on one or more remote instances.

fn update_association_status(&self, input: &UpdateAssociationStatusRequest) -> Result<UpdateAssociationStatusResult, UpdateAssociationStatusError>

Updates the status of the SSM document associated with the specified instance.

fn update_managed_instance_role(&self, input: &UpdateManagedInstanceRoleRequest) -> Result<UpdateManagedInstanceRoleResult, UpdateManagedInstanceRoleError>

Assigns or changes an Amazon Identity and Access Management (IAM) role to the managed instance.