Struct rusoto::storagegateway::StorageGatewayClient

pub struct StorageGatewayClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the AWS Storage Gateway API.

Methods

impl<P> StorageGatewayClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> StorageGatewayClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn activate_gateway(&self, input: &ActivateGatewayInput) -> Result<ActivateGatewayOutput, ActivateGatewayError>

Activates the gateway you previously deployed on your host. For more information, see Activate the AWS Storage Gateway. In the activation process, you specify information such as the you want to use for storing snapshots, the time zone for scheduled snapshots the gateway snapshot schedule window, an activation key, and a name for your gateway. The activation process also associates your gateway with your account; for more information, see UpdateGatewayInformation.

You must turn on the gateway VM before you can activate your gateway.

fn add_cache(&self, input: &AddCacheInput) -> Result<AddCacheOutput, AddCacheError>

Configures one or more gateway local disks as cache for a cached-volume gateway. This operation is supported only for the gateway-cached volume architecture (see Storage Gateway Concepts).

In the request, you specify the gateway Amazon Resource Name (ARN) to which you want to add cache, and one or more disk IDs that you want to configure as cache.

fn add_tags_to_resource(&self, input: &AddTagsToResourceInput) -> Result<AddTagsToResourceOutput, AddTagsToResourceError>

Adds one or more tags to the specified resource. You use tags to add metadata to resources, which you can use to categorize these resources. For example, you can categorize resources by purpose, owner, environment, or team. Each tag consists of a key and a value, which you define. You can add tags to the following AWS Storage Gateway resources:

Storage gateways of all types

Storage Volumes

Virtual Tapes

You can create a maximum of 10 tags for each resource. Virtual tapes and storage volumes that are recovered to a new gateway maintain their tags.

fn add_upload_buffer(&self, input: &AddUploadBufferInput) -> Result<AddUploadBufferOutput, AddUploadBufferError>

Configures one or more gateway local disks as upload buffer for a specified gateway. This operation is supported for both the gateway-stored and gateway-cached volume architectures.

In the request, you specify the gateway Amazon Resource Name (ARN) to which you want to add upload buffer, and one or more disk IDs that you want to configure as upload buffer.

fn add_working_storage(&self, input: &AddWorkingStorageInput) -> Result<AddWorkingStorageOutput, AddWorkingStorageError>

Configures one or more gateway local disks as working storage for a gateway. This operation is supported only for the gateway-stored volume architecture. This operation is deprecated in cached-volumes API version 20120630. Use AddUploadBuffer instead.

Working storage is also referred to as upload buffer. You can also use the AddUploadBuffer operation to add upload buffer to a stored-volume gateway.

In the request, you specify the gateway Amazon Resource Name (ARN) to which you want to add working storage, and one or more disk IDs that you want to configure as working storage.

fn cancel_archival(&self, input: &CancelArchivalInput) -> Result<CancelArchivalOutput, CancelArchivalError>

Cancels archiving of a virtual tape to the virtual tape shelf (VTS) after the archiving process is initiated.

fn cancel_retrieval(&self, input: &CancelRetrievalInput) -> Result<CancelRetrievalOutput, CancelRetrievalError>

Cancels retrieval of a virtual tape from the virtual tape shelf (VTS) to a gateway after the retrieval process is initiated. The virtual tape is returned to the VTS.

fn create_cachedi_s_c_s_i_volume(&self, input: &CreateCachediSCSIVolumeInput) -> Result<CreateCachediSCSIVolumeOutput, CreateCachediSCSIVolumeError>

Creates a cached volume on a specified cached gateway. This operation is supported only for the gateway-cached volume architecture.

Cache storage must be allocated to the gateway before you can create a cached volume. Use the AddCache operation to add cache storage to a gateway.

In the request, you must specify the gateway, size of the volume in bytes, the iSCSI target name, an IP address on which to expose the target, and a unique client token. In response, AWS Storage Gateway creates the volume and returns information about it such as the volume Amazon Resource Name (ARN), its size, and the iSCSI target ARN that initiators can use to connect to the volume target.

fn create_snapshot(&self, input: &CreateSnapshotInput) -> Result<CreateSnapshotOutput, CreateSnapshotError>

Initiates a snapshot of a volume.

AWS Storage Gateway provides the ability to back up point-in-time snapshots of your data to Amazon Simple Storage (S3) for durable off-site recovery, as well as import the data to an Amazon Elastic Block Store (EBS) volume in Amazon Elastic Compute Cloud (EC2). You can take snapshots of your gateway volume on a scheduled or ad-hoc basis. This API enables you to take ad-hoc snapshot. For more information, see Working With Snapshots in the AWS Storage Gateway Console.

In the CreateSnapshot request you identify the volume by providing its Amazon Resource Name (ARN). You must also provide description for the snapshot. When AWS Storage Gateway takes the snapshot of specified volume, the snapshot and description appears in the AWS Storage Gateway Console. In response, AWS Storage Gateway returns you a snapshot ID. You can use this snapshot ID to check the snapshot progress or later use it when you want to create a volume from a snapshot.

To list or delete a snapshot, you must use the Amazon EC2 API. For more information, see DescribeSnapshots or DeleteSnapshot in the EC2 API reference.

Volume and snapshot IDs are changing to a longer length ID format. For more information, see the important note on the Welcome page.

fn create_snapshot_from_volume_recovery_point(&self, input: &CreateSnapshotFromVolumeRecoveryPointInput) -> Result<CreateSnapshotFromVolumeRecoveryPointOutput, CreateSnapshotFromVolumeRecoveryPointError>

Initiates a snapshot of a gateway from a volume recovery point. This operation is supported only for the gateway-cached volume architecture.

A volume recovery point is a point in time at which all data of the volume is consistent and from which you can create a snapshot. To get a list of volume recovery point for gateway-cached volumes, use ListVolumeRecoveryPoints.

In the CreateSnapshotFromVolumeRecoveryPoint request, you identify the volume by providing its Amazon Resource Name (ARN). You must also provide a description for the snapshot. When AWS Storage Gateway takes a snapshot of the specified volume, the snapshot and its description appear in the AWS Storage Gateway console. In response, AWS Storage Gateway returns you a snapshot ID. You can use this snapshot ID to check the snapshot progress or later use it when you want to create a volume from a snapshot.

To list or delete a snapshot, you must use the Amazon EC2 API. For more information, in Amazon Elastic Compute Cloud API Reference.

fn create_storedi_s_c_s_i_volume(&self, input: &CreateStorediSCSIVolumeInput) -> Result<CreateStorediSCSIVolumeOutput, CreateStorediSCSIVolumeError>

Creates a volume on a specified gateway. This operation is supported only for the gateway-stored volume architecture.

The size of the volume to create is inferred from the disk size. You can choose to preserve existing data on the disk, create volume from an existing snapshot, or create an empty volume. If you choose to create an empty gateway volume, then any existing data on the disk is erased.

In the request you must specify the gateway and the disk information on which you are creating the volume. In response, AWS Storage Gateway creates the volume and returns volume information such as the volume Amazon Resource Name (ARN), its size, and the iSCSI target ARN that initiators can use to connect to the volume target.

fn create_tape_with_barcode(&self, input: &CreateTapeWithBarcodeInput) -> Result<CreateTapeWithBarcodeOutput, CreateTapeWithBarcodeError>

Creates a virtual tape by using your own barcode. You write data to the virtual tape and then archive the tape.

Cache storage must be allocated to the gateway before you can create a virtual tape. Use the AddCache operation to add cache storage to a gateway.

fn create_tapes(&self, input: &CreateTapesInput) -> Result<CreateTapesOutput, CreateTapesError>

Creates one or more virtual tapes. You write data to the virtual tapes and then archive the tapes.

Cache storage must be allocated to the gateway before you can create virtual tapes. Use the AddCache operation to add cache storage to a gateway.

fn delete_bandwidth_rate_limit(&self, input: &DeleteBandwidthRateLimitInput) -> Result<DeleteBandwidthRateLimitOutput, DeleteBandwidthRateLimitError>

Deletes the bandwidth rate limits of a gateway. You can delete either the upload and download bandwidth rate limit, or you can delete both. If you delete only one of the limits, the other limit remains unchanged. To specify which gateway to work with, use the Amazon Resource Name (ARN) of the gateway in your request.

fn delete_chap_credentials(&self, input: &DeleteChapCredentialsInput) -> Result<DeleteChapCredentialsOutput, DeleteChapCredentialsError>

Deletes Challenge-Handshake Authentication Protocol (CHAP) credentials for a specified iSCSI target and initiator pair.

fn delete_gateway(&self, input: &DeleteGatewayInput) -> Result<DeleteGatewayOutput, DeleteGatewayError>

Deletes a gateway. To specify which gateway to delete, use the Amazon Resource Name (ARN) of the gateway in your request. The operation deletes the gateway; however, it does not delete the gateway virtual machine (VM) from your host computer.

After you delete a gateway, you cannot reactivate it. Completed snapshots of the gateway volumes are not deleted upon deleting the gateway, however, pending snapshots will not complete. After you delete a gateway, your next step is to remove it from your environment.

You no longer pay software charges after the gateway is deleted; however, your existing Amazon EBS snapshots persist and you will continue to be billed for these snapshots. You can choose to remove all remaining Amazon EBS snapshots by canceling your Amazon EC2 subscription. If you prefer not to cancel your Amazon EC2 subscription, you can delete your snapshots using the Amazon EC2 console. For more information, see the AWS Storage Gateway Detail Page.

fn delete_snapshot_schedule(&self, input: &DeleteSnapshotScheduleInput) -> Result<DeleteSnapshotScheduleOutput, DeleteSnapshotScheduleError>

Deletes a snapshot of a volume.

You can take snapshots of your gateway volumes on a scheduled or ad hoc basis. This API action enables you to delete a snapshot schedule for a volume. For more information, see Working with Snapshots. In the DeleteSnapshotSchedule request, you identify the volume by providing its Amazon Resource Name (ARN).

To list or delete a snapshot, you must use the Amazon EC2 API. in Amazon Elastic Compute Cloud API Reference.

fn delete_tape(&self, input: &DeleteTapeInput) -> Result<DeleteTapeOutput, DeleteTapeError>

Deletes the specified virtual tape.

fn delete_tape_archive(&self, input: &DeleteTapeArchiveInput) -> Result<DeleteTapeArchiveOutput, DeleteTapeArchiveError>

Deletes the specified virtual tape from the virtual tape shelf (VTS).

fn delete_volume(&self, input: &DeleteVolumeInput) -> Result<DeleteVolumeOutput, DeleteVolumeError>

Deletes the specified gateway volume that you previously created using the CreateCachediSCSIVolume or CreateStorediSCSIVolume API. For gateway-stored volumes, the local disk that was configured as the storage volume is not deleted. You can reuse the local disk to create another storage volume.

Before you delete a gateway volume, make sure there are no iSCSI connections to the volume you are deleting. You should also make sure there is no snapshot in progress. You can use the Amazon Elastic Compute Cloud (Amazon EC2) API to query snapshots on the volume you are deleting and check the snapshot status. For more information, go to DescribeSnapshots in the Amazon Elastic Compute Cloud API Reference.

In the request, you must provide the Amazon Resource Name (ARN) of the storage volume you want to delete.

fn describe_bandwidth_rate_limit(&self, input: &DescribeBandwidthRateLimitInput) -> Result<DescribeBandwidthRateLimitOutput, DescribeBandwidthRateLimitError>

Returns the bandwidth rate limits of a gateway. By default, these limits are not set, which means no bandwidth rate limiting is in effect.

This operation only returns a value for a bandwidth rate limit only if the limit is set. If no limits are set for the gateway, then this operation returns only the gateway ARN in the response body. To specify which gateway to describe, use the Amazon Resource Name (ARN) of the gateway in your request.

fn describe_cache(&self, input: &DescribeCacheInput) -> Result<DescribeCacheOutput, DescribeCacheError>

Returns information about the cache of a gateway. This operation is supported only for the gateway-cached volume architecture.

The response includes disk IDs that are configured as cache, and it includes the amount of cache allocated and used.

fn describe_cachedi_s_c_s_i_volumes(&self, input: &DescribeCachediSCSIVolumesInput) -> Result<DescribeCachediSCSIVolumesOutput, DescribeCachediSCSIVolumesError>

Returns a description of the gateway volumes specified in the request. This operation is supported only for the gateway-cached volume architecture.

The list of gateway volumes in the request must be from one gateway. In the response Amazon Storage Gateway returns volume information sorted by volume Amazon Resource Name (ARN).

fn describe_chap_credentials(&self, input: &DescribeChapCredentialsInput) -> Result<DescribeChapCredentialsOutput, DescribeChapCredentialsError>

Returns an array of Challenge-Handshake Authentication Protocol (CHAP) credentials information for a specified iSCSI target, one for each target-initiator pair.

fn describe_gateway_information(&self, input: &DescribeGatewayInformationInput) -> Result<DescribeGatewayInformationOutput, DescribeGatewayInformationError>

Returns metadata about a gateway such as its name, network interfaces, configured time zone, and the state (whether the gateway is running or not). To specify which gateway to describe, use the Amazon Resource Name (ARN) of the gateway in your request.

fn describe_maintenance_start_time(&self, input: &DescribeMaintenanceStartTimeInput) -> Result<DescribeMaintenanceStartTimeOutput, DescribeMaintenanceStartTimeError>

Returns your gateway's weekly maintenance start time including the day and time of the week. Note that values are in terms of the gateway's time zone.

fn describe_snapshot_schedule(&self, input: &DescribeSnapshotScheduleInput) -> Result<DescribeSnapshotScheduleOutput, DescribeSnapshotScheduleError>

Describes the snapshot schedule for the specified gateway volume. The snapshot schedule information includes intervals at which snapshots are automatically initiated on the volume.

fn describe_storedi_s_c_s_i_volumes(&self, input: &DescribeStorediSCSIVolumesInput) -> Result<DescribeStorediSCSIVolumesOutput, DescribeStorediSCSIVolumesError>

Returns the description of the gateway volumes specified in the request. The list of gateway volumes in the request must be from one gateway. In the response Amazon Storage Gateway returns volume information sorted by volume ARNs.

fn describe_tape_archives(&self, input: &DescribeTapeArchivesInput) -> Result<DescribeTapeArchivesOutput, DescribeTapeArchivesError>

Returns a description of specified virtual tapes in the virtual tape shelf (VTS).

If a specific TapeARN is not specified, AWS Storage Gateway returns a description of all virtual tapes found in the VTS associated with your account.

fn describe_tape_recovery_points(&self, input: &DescribeTapeRecoveryPointsInput) -> Result<DescribeTapeRecoveryPointsOutput, DescribeTapeRecoveryPointsError>

Returns a list of virtual tape recovery points that are available for the specified gateway-VTL.

A recovery point is a point-in-time view of a virtual tape at which all the data on the virtual tape is consistent. If your gateway crashes, virtual tapes that have recovery points can be recovered to a new gateway.

fn describe_tapes(&self, input: &DescribeTapesInput) -> Result<DescribeTapesOutput, DescribeTapesError>

Returns a description of the specified Amazon Resource Name (ARN) of virtual tapes. If a TapeARN is not specified, returns a description of all virtual tapes associated with the specified gateway.

fn describe_upload_buffer(&self, input: &DescribeUploadBufferInput) -> Result<DescribeUploadBufferOutput, DescribeUploadBufferError>

Returns information about the upload buffer of a gateway. This operation is supported for both the gateway-stored and gateway-cached volume architectures.

The response includes disk IDs that are configured as upload buffer space, and it includes the amount of upload buffer space allocated and used.

fn describe_v_t_l_devices(&self, input: &DescribeVTLDevicesInput) -> Result<DescribeVTLDevicesOutput, DescribeVTLDevicesError>

Returns a description of virtual tape library (VTL) devices for the specified gateway. In the response, AWS Storage Gateway returns VTL device information.

The list of VTL devices must be from one gateway.

fn describe_working_storage(&self, input: &DescribeWorkingStorageInput) -> Result<DescribeWorkingStorageOutput, DescribeWorkingStorageError>

Returns information about the working storage of a gateway. This operation is supported only for the gateway-stored volume architecture. This operation is deprecated in cached-volumes API version (20120630). Use DescribeUploadBuffer instead.

Working storage is also referred to as upload buffer. You can also use the DescribeUploadBuffer operation to add upload buffer to a stored-volume gateway.

The response includes disk IDs that are configured as working storage, and it includes the amount of working storage allocated and used.

fn disable_gateway(&self, input: &DisableGatewayInput) -> Result<DisableGatewayOutput, DisableGatewayError>

Disables a gateway when the gateway is no longer functioning. For example, if your gateway VM is damaged, you can disable the gateway so you can recover virtual tapes.

Use this operation for a gateway-VTL that is not reachable or not functioning.

Once a gateway is disabled it cannot be enabled.

fn list_gateways(&self, input: &ListGatewaysInput) -> Result<ListGatewaysOutput, ListGatewaysError>

Lists gateways owned by an AWS account in a region specified in the request. The returned list is ordered by gateway Amazon Resource Name (ARN).

By default, the operation returns a maximum of 100 gateways. This operation supports pagination that allows you to optionally reduce the number of gateways returned in a response.

If you have more gateways than are returned in a response (that is, the response returns only a truncated list of your gateways), the response contains a marker that you can specify in your next request to fetch the next page of gateways.

fn list_local_disks(&self, input: &ListLocalDisksInput) -> Result<ListLocalDisksOutput, ListLocalDisksError>

Returns a list of the gateway's local disks. To specify which gateway to describe, you use the Amazon Resource Name (ARN) of the gateway in the body of the request.

The request returns a list of all disks, specifying which are configured as working storage, cache storage, or stored volume or not configured at all. The response includes a DiskStatus field. This field can have a value of present (the disk is available to use), missing (the disk is no longer connected to the gateway), or mismatch (the disk node is occupied by a disk that has incorrect metadata or the disk content is corrupted).

fn list_tags_for_resource(&self, input: &ListTagsForResourceInput) -> Result<ListTagsForResourceOutput, ListTagsForResourceError>

Lists the tags that have been added to the specified resource.

fn list_tapes(&self, input: &ListTapesInput) -> Result<ListTapesOutput, ListTapesError>

Lists virtual tapes in your virtual tape library (VTL) and your virtual tape shelf (VTS). You specify the tapes to list by specifying one or more tape Amazon Resource Names (ARNs). If you don't specify a tape ARN, the operation lists all virtual tapes in both your VTL and VTS.

This operation supports pagination. By default, the operation returns a maximum of up to 100 tapes. You can optionally specify the Limit parameter in the body to limit the number of tapes in the response. If the number of tapes returned in the response is truncated, the response includes a Marker element that you can use in your subsequent request to retrieve the next set of tapes.

fn list_volume_initiators(&self, input: &ListVolumeInitiatorsInput) -> Result<ListVolumeInitiatorsOutput, ListVolumeInitiatorsError>

Lists iSCSI initiators that are connected to a volume. You can use this operation to determine whether a volume is being used or not.

fn list_volume_recovery_points(&self, input: &ListVolumeRecoveryPointsInput) -> Result<ListVolumeRecoveryPointsOutput, ListVolumeRecoveryPointsError>

Lists the recovery points for a specified gateway. This operation is supported only for the gateway-cached volume architecture.

Each gateway-cached volume has one recovery point. A volume recovery point is a point in time at which all data of the volume is consistent and from which you can create a snapshot. To create a snapshot from a volume recovery point use the CreateSnapshotFromVolumeRecoveryPoint operation.

fn list_volumes(&self, input: &ListVolumesInput) -> Result<ListVolumesOutput, ListVolumesError>

Lists the iSCSI stored volumes of a gateway. Results are sorted by volume ARN. The response includes only the volume ARNs. If you want additional volume information, use the DescribeStorediSCSIVolumes API.

The operation supports pagination. By default, the operation returns a maximum of up to 100 volumes. You can optionally specify the Limit field in the body to limit the number of volumes in the response. If the number of volumes returned in the response is truncated, the response includes a Marker field. You can use this Marker value in your subsequent request to retrieve the next set of volumes.

fn remove_tags_from_resource(&self, input: &RemoveTagsFromResourceInput) -> Result<RemoveTagsFromResourceOutput, RemoveTagsFromResourceError>

Removes one or more tags from the specified resource.

fn reset_cache(&self, input: &ResetCacheInput) -> Result<ResetCacheOutput, ResetCacheError>

Resets all cache disks that have encountered a error and makes the disks available for reconfiguration as cache storage. If your cache disk encounters a error, the gateway prevents read and write operations on virtual tapes in the gateway. For example, an error can occur when a disk is corrupted or removed from the gateway. When a cache is reset, the gateway loses its cache storage. At this point you can reconfigure the disks as cache disks.

If the cache disk you are resetting contains data that has not been uploaded to Amazon S3 yet, that data can be lost. After you reset cache disks, there will be no configured cache disks left in the gateway, so you must configure at least one new cache disk for your gateway to function properly.

fn retrieve_tape_archive(&self, input: &RetrieveTapeArchiveInput) -> Result<RetrieveTapeArchiveOutput, RetrieveTapeArchiveError>

Retrieves an archived virtual tape from the virtual tape shelf (VTS) to a gateway-VTL. Virtual tapes archived in the VTS are not associated with any gateway. However after a tape is retrieved, it is associated with a gateway, even though it is also listed in the VTS.

Once a tape is successfully retrieved to a gateway, it cannot be retrieved again to another gateway. You must archive the tape again before you can retrieve it to another gateway.

fn retrieve_tape_recovery_point(&self, input: &RetrieveTapeRecoveryPointInput) -> Result<RetrieveTapeRecoveryPointOutput, RetrieveTapeRecoveryPointError>

Retrieves the recovery point for the specified virtual tape.

A recovery point is a point in time view of a virtual tape at which all the data on the tape is consistent. If your gateway crashes, virtual tapes that have recovery points can be recovered to a new gateway.

The virtual tape can be retrieved to only one gateway. The retrieved tape is read-only. The virtual tape can be retrieved to only a gateway-VTL. There is no charge for retrieving recovery points.

fn set_local_console_password(&self, input: &SetLocalConsolePasswordInput) -> Result<SetLocalConsolePasswordOutput, SetLocalConsolePasswordError>

Sets the password for your VM local console. When you log in to the local console for the first time, you log in to the VM with the default credentials. We recommend that you set a new password. You don't need to know the default password to set a new password.

fn shutdown_gateway(&self, input: &ShutdownGatewayInput) -> Result<ShutdownGatewayOutput, ShutdownGatewayError>

Shuts down a gateway. To specify which gateway to shut down, use the Amazon Resource Name (ARN) of the gateway in the body of your request.

The operation shuts down the gateway service component running in the storage gateway's virtual machine (VM) and not the VM.

If you want to shut down the VM, it is recommended that you first shut down the gateway component in the VM to avoid unpredictable conditions.

After the gateway is shutdown, you cannot call any other API except StartGateway, DescribeGatewayInformation, and ListGateways. For more information, see ActivateGateway. Your applications cannot read from or write to the gateway's storage volumes, and there are no snapshots taken.

When you make a shutdown request, you will get a 200 OK success response immediately. However, it might take some time for the gateway to shut down. You can call the DescribeGatewayInformation API to check the status. For more information, see ActivateGateway.

If do not intend to use the gateway again, you must delete the gateway (using DeleteGateway) to no longer pay software charges associated with the gateway.

fn start_gateway(&self, input: &StartGatewayInput) -> Result<StartGatewayOutput, StartGatewayError>

Starts a gateway that you previously shut down (see ShutdownGateway). After the gateway starts, you can then make other API calls, your applications can read from or write to the gateway's storage volumes and you will be able to take snapshot backups.

When you make a request, you will get a 200 OK success response immediately. However, it might take some time for the gateway to be ready. You should call DescribeGatewayInformation and check the status before making any additional API calls. For more information, see ActivateGateway.

To specify which gateway to start, use the Amazon Resource Name (ARN) of the gateway in your request.

fn update_bandwidth_rate_limit(&self, input: &UpdateBandwidthRateLimitInput) -> Result<UpdateBandwidthRateLimitOutput, UpdateBandwidthRateLimitError>

Updates the bandwidth rate limits of a gateway. You can update both the upload and download bandwidth rate limit or specify only one of the two. If you don't set a bandwidth rate limit, the existing rate limit remains.

By default, a gateway's bandwidth rate limits are not set. If you don't set any limit, the gateway does not have any limitations on its bandwidth usage and could potentially use the maximum available bandwidth.

To specify which gateway to update, use the Amazon Resource Name (ARN) of the gateway in your request.

fn update_chap_credentials(&self, input: &UpdateChapCredentialsInput) -> Result<UpdateChapCredentialsOutput, UpdateChapCredentialsError>

Updates the Challenge-Handshake Authentication Protocol (CHAP) credentials for a specified iSCSI target. By default, a gateway does not have CHAP enabled; however, for added security, you might use it.

When you update CHAP credentials, all existing connections on the target are closed and initiators must reconnect with the new credentials.

fn update_gateway_information(&self, input: &UpdateGatewayInformationInput) -> Result<UpdateGatewayInformationOutput, UpdateGatewayInformationError>

Updates a gateway's metadata, which includes the gateway's name and time zone. To specify which gateway to update, use the Amazon Resource Name (ARN) of the gateway in your request.

For Gateways activated after September 2, 2015, the gateway's ARN contains the gateway ID rather than the gateway name. However, changing the name of the gateway has no effect on the gateway's ARN.

fn update_gateway_software_now(&self, input: &UpdateGatewaySoftwareNowInput) -> Result<UpdateGatewaySoftwareNowOutput, UpdateGatewaySoftwareNowError>

Updates the gateway virtual machine (VM) software. The request immediately triggers the software update.

When you make this request, you get a 200 OK success response immediately. However, it might take some time for the update to complete. You can call DescribeGatewayInformation to verify the gateway is in the STATE_RUNNING state.

A software update forces a system restart of your gateway. You can minimize the chance of any disruption to your applications by increasing your iSCSI Initiators' timeouts. For more information about increasing iSCSI Initiator timeouts for Windows and Linux, see Customizing Your Windows iSCSI Settings and Customizing Your Linux iSCSI Settings, respectively.

fn update_maintenance_start_time(&self, input: &UpdateMaintenanceStartTimeInput) -> Result<UpdateMaintenanceStartTimeOutput, UpdateMaintenanceStartTimeError>

Updates a gateway's weekly maintenance start time information, including day and time of the week. The maintenance time is the time in your gateway's time zone.

fn update_snapshot_schedule(&self, input: &UpdateSnapshotScheduleInput) -> Result<UpdateSnapshotScheduleOutput, UpdateSnapshotScheduleError>

Updates a snapshot schedule configured for a gateway volume.

The default snapshot schedule for volume is once every 24 hours, starting at the creation time of the volume. You can use this API to change the snapshot schedule configured for the volume.

In the request you must identify the gateway volume whose snapshot schedule you want to update, and the schedule information, including when you want the snapshot to begin on a day and the frequency (in hours) of snapshots.

fn update_v_t_l_device_type(&self, input: &UpdateVTLDeviceTypeInput) -> Result<UpdateVTLDeviceTypeOutput, UpdateVTLDeviceTypeError>

Updates the type of medium changer in a gateway-VTL. When you activate a gateway-VTL, you select a medium changer type for the gateway-VTL. This operation enables you to select a different type of medium changer after a gateway-VTL is activated.