Struct rusoto::ChainProvider

pub struct ChainProvider {
    // some fields omitted
}

Provides AWS credentials from multiple possible sources using a priority order.

The following sources are checked in order for credentials when calling credentials:

Environment variables: AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY

AWS credentials file. Usually located at ~/.aws/credentials.

IAM instance profile. Will only work if running on an EC2 instance with an instance profile/role.

If the sources are exhausted without finding credentials, an error is returned.

Methods

impl ChainProvider

fn new() -> ChainProvider

Create a new ChainProvider using a ProfileProvider with the default settings.

fn with_profile_provider(profile_provider: ProfileProvider) -> ChainProvider

Create a new ChainProvider using the provided ProfileProvider.