Struct rusoto::ProfileProvider

pub struct ProfileProvider {
    // some fields omitted
}

Provides AWS credentials from a profile in a credentials file.

Methods

impl ProfileProvider

fn new() -> Result<ProfileProvider, CredentialsError>

Create a new ProfileProvider for the default credentials file path and profile name.

fn with_configuration<F, P>(file_path: F, profile: P) -> ProfileProvider where F: Into<PathBuf>, P: Into<String>

Create a new ProfileProvider for the credentials file at the given path, using the given profile.

fn file_path(&self) -> &Path

Get a reference to the credentials file path.

fn profile(&self) -> &str

Get a reference to the profile name.

fn set_file_path<F>(&mut self, file_path: F) where F: Into<PathBuf>

Set the credentials file path.

fn set_profile<P>(&mut self, profile: P) where P: Into<String>

Set the profile name.