Struct rusoto::SignedRequest

pub struct SignedRequest<'a> {
    pub method: String,
    pub service: String,
    pub region: Region,
    pub path: String,
    pub headers: BTreeMap<String, Vec<Vec<u8>>>,
    pub params: BTreeMap<String, String>,
    pub hostname: Option<String>,
    pub payload: Option<&'a [u8]>,
    pub content_type: Option<String>,
    pub canonical_query_string: String,
    pub canonical_uri: String,
}

A data structure for all the elements of an HTTP request that are involved in the Amazon Signature Version 4 signing process

Fields
method: String service: String region: Region path: String headers: BTreeMap<String, Vec<Vec<u8>>> params: BTreeMap<String, String> hostname: Option<String> payload: Option<&'a [u8]> content_type: Option<String> canonical_query_string: String canonical_uri: String
Methods

impl<'a> SignedRequest<'a>

fn new(method: &str, service: &str, region: Region, path: &str) -> SignedRequest<'a>

Default constructor

fn set_content_type(&mut self, content_type: String)

fn set_hostname(&mut self, hostname: Option<String>)

fn set_endpoint_prefix(&mut self, endpoint_prefix: String)

fn set_payload(&mut self, payload: Option<&'a [u8]>)

fn method(&self) -> &str

fn path(&self) -> &str

fn canonical_uri(&self) -> &str

fn canonical_query_string(&self) -> &str

fn payload(&self) -> Option<&'a [u8]>

fn headers(&'a self) -> &'a BTreeMap<String, Vec<Vec<u8>>>

fn hostname(&self) -> String

fn remove_header(&mut self, key: &str)

fn add_header(&mut self, key: &str, value: &str)

Add a value to the array of headers for the specified key. Headers are kept sorted by key name for use at signing (BTreeMap)

fn add_param<S>(&mut self, key: S, value: S) where S: Into<String>

fn set_params(&mut self, params: BTreeMap<String, String>)

fn sign(&mut self, creds: &AwsCredentials)