Module rusoto::sts

AWS Security Token Service

If you're using the service, you're probably looking for StsClient.

Examples

Assume role and get session credentials

use rusoto::default_tls_client;
use rusoto::{AwsCredentials, DefaultCredentialsProvider, Region};
use rusoto::sts::{StsClient, NewAwsCredsForStsCreds};
use rusoto::sts::AssumeRoleRequest;

let base_provider = DefaultCredentialsProvider::new().unwrap();
let region = Region::UsEast1;
let sts_client = StsClient::new(default_tls_client().unwrap(), base_provider, region);
let role_arn = "arn:aws:iam::123456789012:role/myRole";
let response = sts_client
        .assume_role(&AssumeRoleRequest{
                        role_arn: role_arn.to_owned(),
                        role_session_name: "my_session".to_owned(),
                        ..Default::default()
                    })
        .map_err(|e| "assume role error")
        .and_then(|response| {
            AwsCredentials::new_for_credentials(response.credentials.unwrap())
                .map_err(|e| "new_for_credentials error")
        });

Assume role and create a credentials provider for use with other Rusoto AWS Api clients.

use rusoto::AutoRefreshingProvider;
use rusoto::default_tls_client;
use rusoto::{DefaultCredentialsProvider, Region};
use rusoto::sts::{StsClient, StsAssumeRoleSessionCredentialsProvider};
use rusoto::sts::AssumeRoleRequest;

let base_provider = DefaultCredentialsProvider::new().unwrap();
let region = Region::UsEast1;
let sts_client = StsClient::new(default_tls_client().unwrap(), base_provider, region);
let role_arn = "arn:aws:iam::123456789012:role/myRole";
let sts_provider = StsAssumeRoleSessionCredentialsProvider::new(
        sts_client,
        role_arn.to_owned(),
        "my_session".to_owned(),
        None, None, None, None);
let auto_sts_provider = AutoRefreshingProvider::with_refcell(sts_provider);

Structs

AssumeRoleRequest

AssumeRoleResponse
Contains the response to a successful AssumeRole request, including temporary AWS credentials that can be used to make AWS requests.

AssumeRoleWithSAMLRequest

AssumeRoleWithSAMLResponse
Contains the response to a successful AssumeRoleWithSAML request, including temporary AWS credentials that can be used to make AWS requests.

AssumeRoleWithWebIdentityRequest

AssumeRoleWithWebIdentityResponse
Contains the response to a successful AssumeRoleWithWebIdentity request, including temporary AWS credentials that can be used to make AWS requests.

AssumedRoleUser
The identifiers for the temporary security credentials that the operation returns.

Credentials
AWS credentials for API authentication.

DecodeAuthorizationMessageRequest

DecodeAuthorizationMessageResponse
A document that contains additional information about the authorization status of a request from an encoded message that is returned in response to an AWS request.

FederatedUser
Identifiers for the federated user that is associated with the credentials.

GetCallerIdentityRequest

GetCallerIdentityResponse
Contains the response to a successful GetCallerIdentity request, including information about the entity making the request.

GetFederationTokenRequest

GetFederationTokenResponse
Contains the response to a successful GetFederationToken request, including temporary AWS credentials that can be used to make AWS requests.

GetSessionTokenRequest

GetSessionTokenResponse
Contains the response to a successful GetSessionToken request, including temporary AWS credentials that can be used to make AWS requests.

StsAssumeRoleSessionCredentialsProvider
AwsCredentials provider that calls AssumeRole using the provided StsClient. To use with MFA, pass in the MFA serial number then set the MFA code. You will need to ensure the provider has a valid code each time you acquire a new STS token.

StsClient
A client for the AWS STS API.

StsSessionCredentialsProvider
AwsCredentials provider that calls GetSessionToken using the provided StsClient. To use with MFA, pass in the MFA serial number then set the MFA code. You will need to ensure the provider has a valid code each time you acquire a new STS token.

StsWebIdentityFederationSessionCredentialsProvider
AwsCredentials provider that calls AssumeRoleWithWebIdentity using the provided StsClient.

Enums

AssumeRoleError
Errors returned by AssumeRole

AssumeRoleWithSAMLError
Errors returned by AssumeRoleWithSAML

AssumeRoleWithWebIdentityError
Errors returned by AssumeRoleWithWebIdentity

DecodeAuthorizationMessageError
Errors returned by DecodeAuthorizationMessage

GetCallerIdentityError
Errors returned by GetCallerIdentity

GetFederationTokenError
Errors returned by GetFederationToken

GetSessionTokenError
Errors returned by GetSessionToken

Traits

NewAwsCredsForStsCreds
Trait for conversions from STS Credentials to AWS Credentials.

Type Definitions

AccessKeyIdType

AccessKeySecretType

AccountType

ArnType

AssumedRoleIdType

Audience

ClientTokenType

DateType

DecodedMessageType

DurationSecondsType

EncodedMessageType

ExpiredIdentityTokenMessage

ExternalIdType

FederatedIdType

IdpCommunicationErrorMessage

IdpRejectedClaimMessage

InvalidAuthorizationMessage

InvalidIdentityTokenMessage

Issuer

MalformedPolicyDocumentMessage

NameQualifier

NonNegativeIntegerType

PackedPolicyTooLargeMessage

RegionDisabledMessage

RoleDurationSecondsType

RoleSessionNameType

SAMLAssertionType

SerialNumberType

SessionPolicyDocumentType

Subject

SubjectType

TokenCodeType

TokenType

UrlType

UserIdType

UserNameType

WebIdentitySubjectType