Struct rusoto::sts::StsSessionCredentialsProvider

pub struct StsSessionCredentialsProvider { /* fields omitted */ }

AwsCredentials provider that calls GetSessionToken using the provided StsClient. To use with MFA, pass in the MFA serial number then set the MFA code. You will need to ensure the provider has a valid code each time you acquire a new STS token.

Methods

impl StsSessionCredentialsProvider

fn new<P, D>(sts_client: StsClient<P, D>,
duration: Option<Duration>,
mfa_serial: Option<String>)
-> StsSessionCredentialsProvider where P: ProvideAwsCredentials + 'static, D: DispatchSignedRequest + 'static

Creates a new StsSessionCredentialsProvider with the given StsClient and session parameters.

sts_client - The StsClient to use to acquire session tokens.

duration - The duration of the session tokens. Default 1 hour.

mfa_serial - Optional MFA hardware device serial number or virtual device ARN. Set the MFA code with set_mfa_code.

fn set_mfa_code<S>(&mut self, code: S) where S: Into<String>

Set the MFA code for use when acquiring session tokens.

fn clear_mfa_code(&mut self)

Clear the MFA code.

fn get_session_token(&self) -> Result<AwsCredentials, CredentialsError>

Calls GetSessionToken to get a session token from the STS Api. Optionally uses MFA if the MFA serial number and code are set.