Struct rusoto::swf::ActivityTaskScheduledEventAttributes

pub struct ActivityTaskScheduledEventAttributes {
    pub activity_id: ActivityId,
    pub activity_type: ActivityType,
    pub control: Option<Data>,
    pub decision_task_completed_event_id: EventId,
    pub heartbeat_timeout: Option<DurationInSecondsOptional>,
    pub input: Option<Data>,
    pub schedule_to_close_timeout: Option<DurationInSecondsOptional>,
    pub schedule_to_start_timeout: Option<DurationInSecondsOptional>,
    pub start_to_close_timeout: Option<DurationInSecondsOptional>,
    pub task_list: TaskList,
    pub task_priority: Option<TaskPriority>,
}

Provides details of the ActivityTaskScheduled event.

Fields

activity_id

The unique ID of the activity task.

activity_type

The type of the activity task.

control

Optional. Data attached to the event that can be used by the decider in subsequent workflow tasks. This data is not sent to the activity.

decision_task_completed_event_id

The ID of the DecisionTaskCompleted event corresponding to the decision that resulted in the scheduling of this activity task. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

heartbeat_timeout

The maximum time before which the worker processing this task must report progress by calling RecordActivityTaskHeartbeat. If the timeout is exceeded, the activity task is automatically timed out. If the worker subsequently attempts to record a heartbeat or return a result, it will be ignored.

input

The input provided to the activity task.

schedule_to_close_timeout

The maximum amount of time for this activity task.

schedule_to_start_timeout

The maximum amount of time the activity task can wait to be assigned to a worker.

start_to_close_timeout

The maximum amount of time a worker may take to process the activity task.

task_list

The task list in which the activity task has been scheduled.

task_priority

Optional. The priority to assign to the scheduled activity task. If set, this will override any default priority value that was assigned when the activity type was registered.

Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon Simple Workflow Developer Guide.