Struct rusoto::swf::PollForDecisionTaskInput

pub struct PollForDecisionTaskInput {
    pub domain: DomainName,
    pub identity: Option<Identity>,
    pub maximum_page_size: Option<PageSize>,
    pub next_page_token: Option<PageToken>,
    pub reverse_order: Option<ReverseOrder>,
    pub task_list: TaskList,
}

Fields
domain: DomainName

The name of the domain containing the task lists to poll.

identity: Option<Identity>

Identity of the decider making the request, which is recorded in the DecisionTaskStarted event in the workflow history. This enables diagnostic tracing when problems arise. The form of this identity is user defined.

maximum_page_size: Option<PageSize>

The maximum number of results that will be returned per call. nextPageToken can be used to obtain futher pages of results. The default is 1000, which is the maximum allowed page size. You can, however, specify a page size smaller than the maximum.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

next_page_token: Option<PageToken>

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.
The nextPageToken returned by this action cannot be used with GetWorkflowExecutionHistory to get the next page. You must call PollForDecisionTask again (with the nextPageToken) to retrieve the next page of history records. Calling PollForDecisionTask with a nextPageToken will not return a new decision task..

reverse_order: Option<ReverseOrder>

When set to true, returns the events in reverse order. By default the results are returned in ascending order of the eventTimestamp of the events.

task_list: TaskList

Specifies the task list to poll for decision tasks.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f - \u009f). Also, it must not contain the literal string quotarnquot.