Struct rusoto::swf::RegisterActivityTypeInput

pub struct RegisterActivityTypeInput {
    pub default_task_heartbeat_timeout: Option<DurationInSecondsOptional>,
    pub default_task_list: Option<TaskList>,
    pub default_task_priority: Option<TaskPriority>,
    pub default_task_schedule_to_close_timeout: Option<DurationInSecondsOptional>,
    pub default_task_schedule_to_start_timeout: Option<DurationInSecondsOptional>,
    pub default_task_start_to_close_timeout: Option<DurationInSecondsOptional>,
    pub description: Option<Description>,
    pub domain: DomainName,
    pub name: Name,
    pub version: Version,
}

Fields
default_task_heartbeat_timeout: Option<DurationInSecondsOptional>

If set, specifies the default maximum time before which a worker processing a task of this type must report progress by calling RecordActivityTaskHeartbeat. If the timeout is exceeded, the activity task is automatically timed out. This default can be overridden when scheduling an activity task using the ScheduleActivityTask decision. If the activity worker subsequently attempts to record a heartbeat or returns a result, the activity worker receives an UnknownResource fault. In this case, Amazon SWF no longer considers the activity task to be valid; the activity worker should clean up the activity task.

The duration is specified in seconds; an integer greater than or equal to 0. The value "NONE" can be used to specify unlimited duration.

default_task_list: Option<TaskList>

If set, specifies the default task list to use for scheduling tasks of this activity type. This default task list is used if a task list is not provided when a task is scheduled through the ScheduleActivityTask decision.

default_task_priority: Option<TaskPriority>

The default task priority to assign to the activity type. If not assigned, then "0" will be used. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon Simple Workflow Developer Guide.

default_task_schedule_to_close_timeout: Option<DurationInSecondsOptional>

If set, specifies the default maximum duration for a task of this activity type. This default can be overridden when scheduling an activity task using the ScheduleActivityTask decision.

The duration is specified in seconds; an integer greater than or equal to 0. The value "NONE" can be used to specify unlimited duration.

default_task_schedule_to_start_timeout: Option<DurationInSecondsOptional>

If set, specifies the default maximum duration that a task of this activity type can wait before being assigned to a worker. This default can be overridden when scheduling an activity task using the ScheduleActivityTask decision.

The duration is specified in seconds; an integer greater than or equal to 0. The value "NONE" can be used to specify unlimited duration.

default_task_start_to_close_timeout: Option<DurationInSecondsOptional>

If set, specifies the default maximum duration that a worker can take to process tasks of this activity type. This default can be overridden when scheduling an activity task using the ScheduleActivityTask decision.

The duration is specified in seconds; an integer greater than or equal to 0. The value "NONE" can be used to specify unlimited duration.

description: Option<Description>

A textual description of the activity type.

domain: DomainName

The name of the domain in which this activity is to be registered.

name: Name

The name of the activity type within the domain.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f - \u009f). Also, it must not contain the literal string quotarnquot.

version: Version

The version of the activity type.
The activity type consists of the name and version, the combination of which must be unique within the domain.
The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f - \u009f). Also, it must not contain the literal string quotarnquot.