Struct rusoto::swf::ScheduleLambdaFunctionFailedEventAttributes

pub struct ScheduleLambdaFunctionFailedEventAttributes {
    pub cause: ScheduleLambdaFunctionFailedCause,
    pub decision_task_completed_event_id: EventId,
    pub id: FunctionId,
    pub name: FunctionName,
}

Provides details for the ScheduleLambdaFunctionFailed event.

Fields
cause: ScheduleLambdaFunctionFailedCause

The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.
If cause is set to OPERATION_NOT_PERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

decision_task_completed_event_id: EventId

The ID of the DecisionTaskCompleted event corresponding to the decision that resulted in the scheduling of this AWS Lambda function. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

id: FunctionId

The unique Amazon SWF ID of the AWS Lambda task.

name: FunctionName

The name of the scheduled AWS Lambda function.