Struct rusoto::swf::WorkflowExecutionDetail

pub struct WorkflowExecutionDetail {
    pub execution_configuration: WorkflowExecutionConfiguration,
    pub execution_info: WorkflowExecutionInfo,
    pub latest_activity_task_timestamp: Option<Timestamp>,
    pub latest_execution_context: Option<Data>,
    pub open_counts: WorkflowExecutionOpenCounts,
}

Contains details about a workflow execution.

Fields
execution_configuration: WorkflowExecutionConfiguration

The configuration settings for this workflow execution including timeout values, tasklist etc.

execution_info: WorkflowExecutionInfo

Information about the workflow execution.

latest_activity_task_timestamp: Option<Timestamp>

The time when the last activity task was scheduled for this workflow execution. You can use this information to determine if the workflow has not made progress for an unusually long period of time and might require a corrective action.

latest_execution_context: Option<Data>

The latest executionContext provided by the decider for this workflow execution. A decider can provide an executionContext (a free-form string) when closing a decision task using RespondDecisionTaskCompleted.

open_counts: WorkflowExecutionOpenCounts

The number of tasks for this workflow execution. This includes open and closed tasks of all types.