Struct rusoto::swf::WorkflowExecutionInfo

pub struct WorkflowExecutionInfo {
    pub cancel_requested: Option<Canceled>,
    pub close_status: Option<CloseStatus>,
    pub close_timestamp: Option<Timestamp>,
    pub execution: WorkflowExecution,
    pub execution_status: ExecutionStatus,
    pub parent: Option<WorkflowExecution>,
    pub start_timestamp: Timestamp,
    pub tag_list: Option<TagList>,
    pub workflow_type: WorkflowType,
}

Contains information about a workflow execution.

Fields
cancel_requested: Option<Canceled>

Set to true if a cancellation is requested for this workflow execution.

close_status: Option<CloseStatus>

If the execution status is closed then this specifies how the execution was closed:

COMPLETED: the execution was successfully completed.

CANCELED: the execution was canceled.Cancellation allows the implementation to gracefully clean up before the execution is closed.

TERMINATED: the execution was force terminated.

FAILED: the execution failed to complete.

TIMED_OUT: the execution did not complete in the alloted time and was automatically timed out.

CONTINUED_AS_NEW: the execution is logically continued. This means the current execution was completed and a new execution was started to carry on the workflow.

close_timestamp: Option<Timestamp>

The time when the workflow execution was closed. Set only if the execution status is CLOSED.

execution: WorkflowExecution

The workflow execution this information is about.

execution_status: ExecutionStatus

The current status of the execution.

parent: Option<WorkflowExecution>

If this workflow execution is a child of another execution then contains the workflow execution that started this execution.

start_timestamp: Timestamp

The time when the execution was started.

tag_list: Option<TagList>

The list of tags associated with the workflow execution. Tags can be used to identify and list workflow executions of interest through the visibility APIs. A workflow execution can have a maximum of 5 tags.

workflow_type: WorkflowType

The type of the workflow execution.