Struct rusoto::swf::WorkflowExecutionSignaledEventAttributes

pub struct WorkflowExecutionSignaledEventAttributes {
    pub external_initiated_event_id: Option<EventId>,
    pub external_workflow_execution: Option<WorkflowExecution>,
    pub input: Option<Data>,
    pub signal_name: SignalName,
}

Provides details of the WorkflowExecutionSignaled event.

Fields
external_initiated_event_id: Option<EventId>

The ID of the SignalExternalWorkflowExecutionInitiated event corresponding to the SignalExternalWorkflow decision to signal this workflow execution.The source event with this ID can be found in the history of the source workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event. This field is set only if the signal was initiated by another workflow execution.

external_workflow_execution: Option<WorkflowExecution>

The workflow execution that sent the signal. This is set only of the signal was sent by another workflow execution.

input: Option<Data>

Inputs provided with the signal (if any). The decider can use the signal name and inputs to determine how to process the signal.

signal_name: SignalName

The name of the signal received. The decider can use the signal name and inputs to determine how to the process the signal.