Struct rusoto::swf::WorkflowExecutionStartedEventAttributes

pub struct WorkflowExecutionStartedEventAttributes {
    pub child_policy: ChildPolicy,
    pub continued_execution_run_id: Option<RunIdOptional>,
    pub execution_start_to_close_timeout: Option<DurationInSecondsOptional>,
    pub input: Option<Data>,
    pub lambda_role: Option<Arn>,
    pub parent_initiated_event_id: Option<EventId>,
    pub parent_workflow_execution: Option<WorkflowExecution>,
    pub tag_list: Option<TagList>,
    pub task_list: TaskList,
    pub task_priority: Option<TaskPriority>,
    pub task_start_to_close_timeout: Option<DurationInSecondsOptional>,
    pub workflow_type: WorkflowType,
}

Provides details of WorkflowExecutionStarted event.

Fields
child_policy: ChildPolicy

The policy to use for the child workflow executions if this workflow execution is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout.

The supported child policies are:

TERMINATE: the child executions will be terminated.

REQUEST_CANCEL: a request to cancel will be attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON: no action will be taken. The child executions will continue to run.

continued_execution_run_id: Option<RunIdOptional>

If this workflow execution was started due to a ContinueAsNewWorkflowExecution decision, then it contains the runId of the previous workflow execution that was closed and continued as this execution.

execution_start_to_close_timeout: Option<DurationInSecondsOptional>

The maximum duration for this workflow execution.

The duration is specified in seconds; an integer greater than or equal to 0. The value "NONE" can be used to specify unlimited duration.

input: Option<Data>

The input provided to the workflow execution (if any).

lambda_role: Option<Arn>

The IAM role attached to this workflow execution to use when invoking AWS Lambda functions.

parent_initiated_event_id: Option<EventId>

The ID of the StartChildWorkflowExecutionInitiated event corresponding to the StartChildWorkflowExecution decision to start this workflow execution. The source event with this ID can be found in the history of the source workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

parent_workflow_execution: Option<WorkflowExecution>

The source workflow execution that started this workflow execution. The member is not set if the workflow execution was not started by a workflow.

tag_list: Option<TagList>

The list of tags associated with this workflow execution. An execution can have up to 5 tags.

task_list: TaskList

The name of the task list for scheduling the decision tasks for this workflow execution.

task_priority: Option<TaskPriority> task_start_to_close_timeout: Option<DurationInSecondsOptional>

The maximum duration of decision tasks for this workflow type.

The duration is specified in seconds; an integer greater than or equal to 0. The value "NONE" can be used to specify unlimited duration.

workflow_type: WorkflowType

The workflow type of this execution.