Struct rusoto::swf::WorkflowExecutionTimedOutEventAttributes

pub struct WorkflowExecutionTimedOutEventAttributes {
    pub child_policy: ChildPolicy,
    pub timeout_type: WorkflowExecutionTimeoutType,
}

Provides details of the WorkflowExecutionTimedOut event.

Fields
child_policy: ChildPolicy

The policy used for the child workflow executions of this workflow execution.

The supported child policies are:

TERMINATE: the child executions will be terminated.

REQUEST_CANCEL: a request to cancel will be attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON: no action will be taken. The child executions will continue to run.

timeout_type: WorkflowExecutionTimeoutType

The type of timeout that caused this event.