Enum rusoto::waf::CreateSqlInjectionMatchSetError

pub enum CreateSqlInjectionMatchSetError {
    WAFDisallowedName(String),
    WAFInternalError(String),
    WAFInvalidAccount(String),
    WAFInvalidParameter(String),
    WAFStaleData(String),
    WAFLimitsExceeded(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Variants

WAFDisallowedName

The name specified is invalid.

WAFInternalError

The operation failed because of a system problem, even though the request was valid. Retry your request.

WAFInvalidAccount

The operation failed because you tried to create, update, or delete an object by using an invalid account identifier.

WAFInvalidParameter

The operation failed because AWS WAF didn't recognize a parameter in the request. For example:

You specified an invalid parameter name.

You specified an invalid value.

You tried to update an object (ByteMatchSet, IPSet, Rule, or WebACL) using an action other than INSERT or DELETE.

You tried to create a WebACL with a DefaultAction Type other than ALLOW, BLOCK, or COUNT.

You tried to update a WebACL with a WafAction Type other than ALLOW, BLOCK, or COUNT.

You tried to update a ByteMatchSet with a FieldToMatch Type other than HEADER, QUERY_STRING, or URI.

You tried to update a ByteMatchSet with a Field of HEADER but no value for Data.

WAFStaleData

The operation failed because you tried to create, update, or delete an object by using a change token that has already been used.

WAFLimitsExceeded

The operation exceeds a resource limit, for example, the maximum number of WebACL objects that you can create for an AWS account. For more information, see Limits in the AWS WAF Developer Guide.

HttpDispatch

An error occurred dispatching the HTTP request

Credentials

An error was encountered with AWS credentials.

Validation

A validation error occurred. Details from AWS are provided.

Unknown

An unknown error occurred. The raw HTTP response is provided.

Methods

impl CreateSqlInjectionMatchSetError

fn from_body(body: &str) -> CreateSqlInjectionMatchSetError