Enum rusoto::waf::UpdateXssMatchSetError

pub enum UpdateXssMatchSetError {
    WAFInternalError(String),
    WAFInvalidAccount(String),
    WAFInvalidOperation(String),
    WAFInvalidParameter(String),
    WAFNonexistentContainer(String),
    WAFNonexistentItem(String),
    WAFStaleData(String),
    WAFLimitsExceeded(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Variants
WAFInternalError(String)

The operation failed because of a system problem, even though the request was valid. Retry your request.

WAFInvalidAccount(String)

The operation failed because you tried to create, update, or delete an object by using an invalid account identifier.

WAFInvalidOperation(String)

The operation failed because there was nothing to do. For example:

You tried to remove a Rule from a WebACL, but the Rule isn't in the specified WebACL.

You tried to remove an IP address from an IPSet, but the IP address isn't in the specified IPSet.

You tried to remove a ByteMatchTuple from a ByteMatchSet, but the ByteMatchTuple isn't in the specified WebACL.

You tried to add a Rule to a WebACL, but the Rule already exists in the specified WebACL.

You tried to add an IP address to an IPSet, but the IP address already exists in the specified IPSet.

You tried to add a ByteMatchTuple to a ByteMatchSet, but the ByteMatchTuple already exists in the specified WebACL.

WAFInvalidParameter(String)

The operation failed because AWS WAF didn't recognize a parameter in the request. For example:

You specified an invalid parameter name.

You specified an invalid value.

You tried to update an object (ByteMatchSet, IPSet, Rule, or WebACL) using an action other than INSERT or DELETE.

You tried to create a WebACL with a DefaultAction Type other than ALLOW, BLOCK, or COUNT.

You tried to update a WebACL with a WafAction Type other than ALLOW, BLOCK, or COUNT.

You tried to update a ByteMatchSet with a FieldToMatch Type other than HEADER, QUERY_STRING, or URI.

You tried to update a ByteMatchSet with a Field of HEADER but no value for Data.

WAFNonexistentContainer(String)

The operation failed because you tried to add an object to or delete an object from another object that doesn't exist. For example:

You tried to add a Rule to or delete a Rule from a WebACL that doesn't exist.

You tried to add a ByteMatchSet to or delete a ByteMatchSet from a Rule that doesn't exist.

You tried to add an IP address to or delete an IP address from an IPSet that doesn't exist.

You tried to add a ByteMatchTuple to or delete a ByteMatchTuple from a ByteMatchSet that doesn't exist.

WAFNonexistentItem(String)

The operation failed because the referenced object doesn't exist.

WAFStaleData(String)

The operation failed because you tried to create, update, or delete an object by using a change token that has already been used.

WAFLimitsExceeded(String)

The operation exceeds a resource limit, for example, the maximum number of WebACL objects that you can create for an AWS account. For more information, see Limits in the AWS WAF Developer Guide.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl UpdateXssMatchSetError

fn from_body(body: &str) -> UpdateXssMatchSetError