Struct rusoto::waf::FieldToMatch

pub struct FieldToMatch {
    pub data: Option<MatchFieldData>,
    pub type_: MatchFieldType,
}

Specifies where in a web request to look for TargetString.

Fields
data: Option<MatchFieldData>

When the value of Type is HEADER, enter the name of the header that you want AWS WAF to search, for example, User-Agent or Referer. If the value of Type is any other value, omit Data.

The name of the header is not case sensitive.

type_: MatchFieldType

The part of the web request that you want AWS WAF to search for a specified string. Parts of a request that you can search include the following:

HEADER: A specified request header, for example, the value of the User-Agent or Referer header. If you choose HEADER for the type, specify the name of the header in Data.

METHOD: The HTTP method, which indicated the type of operation that the request is asking the origin to perform. Amazon CloudFront supports the following methods: DELETE, GET, HEAD, OPTIONS, PATCH, POST, and PUT.

QUERY_STRING: A query string, which is the part of a URL that appears after a ? character, if any.

URI: The part of a web request that identifies a resource, for example, /images/daily-ad.jpg.

BODY: The part of a request that contains any additional data that you want to send to your web server as the HTTP request body, such as data from a form. The request body immediately follows the request headers. Note that only the first 8192 bytes of the request body are forwarded to AWS WAF for inspection. To allow or block requests based on the length of the body, you can create a size constraint set. For more information, see CreateSizeConstraintSet.