Struct rusoto::waf::GetSampledRequestsRequest

pub struct GetSampledRequestsRequest {
    pub max_items: ListMaxItems,
    pub rule_id: ResourceId,
    pub time_window: TimeWindow,
    pub web_acl_id: ResourceId,
}

Fields
max_items: ListMaxItems

The number of requests that you want AWS WAF to return from among the first 5,000 requests that your AWS resource received during the time range. If your resource received fewer requests than the value of MaxItems, GetSampledRequests returns information about all of them.

rule_id: ResourceId

RuleId is one of two values:

The RuleId of the Rule for which you want GetSampledRequests to return a sample of requests.

Default_Action, which causes GetSampledRequests to return a sample of the requests that didn't match any of the rules in the specified WebACL.

time_window: TimeWindow

The start date and time and the end date and time of the range for which you want GetSampledRequests to return a sample of requests. Specify the date and time in Unix time format (in seconds). You can specify any time range in the previous three hours.

web_acl_id: ResourceId

The WebACLId of the WebACL for which you want GetSampledRequests to return a sample of requests.