Struct rusoto::waf::IPSet

pub struct IPSet {
    pub i_p_set_descriptors: IPSetDescriptors,
    pub i_p_set_id: ResourceId,
    pub name: Option<ResourceName>,
}

Contains one or more IP addresses or blocks of IP addresses specified in Classless Inter-Domain Routing (CIDR) notation. To specify an individual IP address, you specify the four-part IP address followed by a /32, for example, 192.0.2.0/31. To block a range of IP addresses, you can specify a /24, a /16, or a /8 CIDR. For more information about CIDR notation, perform an Internet search on cidr notation.

Fields
i_p_set_descriptors: IPSetDescriptors

The IP address type (IPV4) and the IP address range (in CIDR notation) that web requests originate from. If the WebACL is associated with a CloudFront distribution, this is the value of one of the following fields in CloudFront access logs:

c-ip, if the viewer did not use an HTTP proxy or a load balancer to send the request

x-forwarded-for, if the viewer did use an HTTP proxy or a load balancer to send the request

i_p_set_id: ResourceId

The IPSetId for an IPSet. You use IPSetId to get information about an IPSet (see GetIPSet), update an IPSet (see UpdateIPSet), insert an IPSet into a Rule or delete one from a Rule (see UpdateRule), and delete an IPSet from AWS WAF (see DeleteIPSet).

IPSetId is returned by CreateIPSet and by ListIPSets.

name: Option<ResourceName>

A friendly name or description of the IPSet. You can't change the name of an IPSet after you create it.