Struct rusoto::waf::Rule

pub struct Rule {
    pub metric_name: Option<MetricName>,
    pub name: Option<ResourceName>,
    pub predicates: Predicates,
    pub rule_id: ResourceId,
}

A combination of ByteMatchSet, IPSet, and/or SqlInjectionMatchSet objects that identify the web requests that you want to allow, block, or count. For example, you might create a Rule that includes the following predicates:

An IPSet that causes AWS WAF to search for web requests that originate from the IP address 192.0.2.44

A ByteMatchSet that causes AWS WAF to search for web requests for which the value of the User-Agent header is BadBot.

To match the settings in this Rule, a request must originate from 192.0.2.44 AND include a User-Agent header for which the value is BadBot.

Fields
metric_name: Option<MetricName> name: Option<ResourceName>

The friendly name or description for the Rule. You can't change the name of a Rule after you create it.

predicates: Predicates

The Predicates object contains one Predicate element for each ByteMatchSet, IPSet, or SqlInjectionMatchSet object that you want to include in a Rule.

rule_id: ResourceId

A unique identifier for a Rule. You use RuleId to get more information about a Rule (see GetRule), update a Rule (see UpdateRule), insert a Rule into a WebACL or delete a one from a WebACL (see UpdateWebACL), or delete a Rule from AWS WAF (see DeleteRule).

RuleId is returned by CreateRule and by ListRules.