Struct rusoto::waf::UpdateRuleRequest

pub struct UpdateRuleRequest {
    pub change_token: ChangeToken,
    pub rule_id: ResourceId,
    pub updates: RuleUpdates,
}

Fields
change_token: ChangeToken

The value returned by the most recent call to GetChangeToken.

rule_id: ResourceId

The RuleId of the Rule that you want to update. RuleId is returned by CreateRule and by ListRules.

updates: RuleUpdates

An array of RuleUpdate objects that you want to insert into or delete from a Rule. For more information, see the applicable data types:

RuleUpdate: Contains Action and Predicate

Predicate: Contains DataId, Negated, and Type

FieldToMatch: Contains Data and Type