Struct rusoto::waf::UpdateXssMatchSetRequest

pub struct UpdateXssMatchSetRequest {
    pub change_token: ChangeToken,
    pub updates: XssMatchSetUpdates,
    pub xss_match_set_id: ResourceId,
}

A request to update an XssMatchSet.

Fields
change_token: ChangeToken

The value returned by the most recent call to GetChangeToken.

updates: XssMatchSetUpdates

An array of XssMatchSetUpdate objects that you want to insert into or delete from a XssMatchSet. For more information, see the applicable data types:

XssMatchSetUpdate: Contains Action and XssMatchTuple

XssMatchTuple: Contains FieldToMatch and TextTransformation

FieldToMatch: Contains Data and Type

xss_match_set_id: ResourceId

The XssMatchSetId of the XssMatchSet that you want to update. XssMatchSetId is returned by CreateXssMatchSet and by ListXssMatchSets.