Struct rusoto::waf::WebACL

pub struct WebACL {
    pub default_action: WafAction,
    pub metric_name: Option<MetricName>,
    pub name: Option<ResourceName>,
    pub rules: ActivatedRules,
    pub web_a_c_l_id: ResourceId,
}

Contains the Rules that identify the requests that you want to allow, block, or count. In a WebACL, you also specify a default action (ALLOW or BLOCK), and the action for each Rule that you add to a WebACL, for example, block requests from specified IP addresses or block requests from specified referrers. You also associate the WebACL with a CloudFront distribution to identify the requests that you want AWS WAF to filter. If you add more than one Rule to a WebACL, a request needs to match only one of the specifications to be allowed, blocked, or counted. For more information, see UpdateWebACL.

Fields
default_action: WafAction

The action to perform if none of the Rules contained in the WebACL match. The action is specified by the WafAction object.

metric_name: Option<MetricName> name: Option<ResourceName>

A friendly name or description of the WebACL. You can't change the name of a WebACL after you create it.

rules: ActivatedRules

An array that contains the action for each Rule in a WebACL, the priority of the Rule, and the ID of the Rule.

web_a_c_l_id: ResourceId

A unique identifier for a WebACL. You use WebACLId to get information about a WebACL (see GetWebACL), update a WebACL (see UpdateWebACL), and delete a WebACL from AWS WAF (see DeleteWebACL).

WebACLId is returned by CreateWebACL and by ListWebACLs.