Module rusoto::workspaces

Amazon Workspaces

If you're using the service, you're probably looking for WorkspacesClient.

Structs

ComputeType
Contains information about the compute type of a WorkSpace bundle.

CreateTagsRequest
The request of the CreateTags operation.

CreateTagsResult
The result of the CreateTags operation.

CreateWorkspacesRequest
Contains the inputs for the CreateWorkspaces operation.

CreateWorkspacesResult
Contains the result of the CreateWorkspaces operation.

DefaultWorkspaceCreationProperties
Contains default WorkSpace creation information.

DeleteTagsRequest
The request of the DeleteTags operation.

DeleteTagsResult
The result of the DeleteTags operation.

DescribeTagsRequest
The request of the DescribeTags operation.

DescribeTagsResult
The result of the DescribeTags operation.

DescribeWorkspaceBundlesRequest
Contains the inputs for the DescribeWorkspaceBundles operation.

DescribeWorkspaceBundlesResult
Contains the results of the DescribeWorkspaceBundles operation.

DescribeWorkspaceDirectoriesRequest
Contains the inputs for the DescribeWorkspaceDirectories operation.

DescribeWorkspaceDirectoriesResult
Contains the results of the DescribeWorkspaceDirectories operation.

DescribeWorkspacesConnectionStatusRequest

DescribeWorkspacesConnectionStatusResult

DescribeWorkspacesRequest
Contains the inputs for the DescribeWorkspaces operation.

DescribeWorkspacesResult
Contains the results for the DescribeWorkspaces operation.

FailedCreateWorkspaceRequest
Contains information about a WorkSpace that could not be created.

FailedWorkspaceChangeRequest
Contains information about a WorkSpace that could not be rebooted (RebootWorkspaces), rebuilt (RebuildWorkspaces), terminated (TerminateWorkspaces), started (StartWorkspaces), or stopped (StopWorkspaces).

ModifyWorkspacePropertiesRequest

ModifyWorkspacePropertiesResult

RebootRequest
Contains information used with the RebootWorkspaces operation to reboot a WorkSpace.

RebootWorkspacesRequest
Contains the inputs for the RebootWorkspaces operation.

RebootWorkspacesResult
Contains the results of the RebootWorkspaces operation.

RebuildRequest
Contains information used with the RebuildWorkspaces operation to rebuild a WorkSpace.

RebuildWorkspacesRequest
Contains the inputs for the RebuildWorkspaces operation.

RebuildWorkspacesResult
Contains the results of the RebuildWorkspaces operation.

StartRequest
Describes the start request.

StartWorkspacesRequest

StartWorkspacesResult

StopRequest
Describes the stop request.

StopWorkspacesRequest

StopWorkspacesResult

Tag
Describes the tag of the WorkSpace.

TerminateRequest
Contains information used with the TerminateWorkspaces operation to terminate a WorkSpace.

TerminateWorkspacesRequest
Contains the inputs for the TerminateWorkspaces operation.

TerminateWorkspacesResult
Contains the results of the TerminateWorkspaces operation.

UserStorage
Contains information about the user storage for a WorkSpace bundle.

Workspace
Contains information about a WorkSpace.

WorkspaceBundle
Contains information about a WorkSpace bundle.

WorkspaceConnectionStatus
Describes the connection status of a WorkSpace.

WorkspaceDirectory
Contains information about an AWS Directory Service directory for use with Amazon WorkSpaces.

WorkspaceProperties
Describes the properties of a WorkSpace.

WorkspaceRequest
Contains information about a WorkSpace creation request.

WorkspacesClient
A client for the Amazon WorkSpaces API.

Enums

CreateTagsError
Errors returned by CreateTags

CreateWorkspacesError
Errors returned by CreateWorkspaces

DeleteTagsError
Errors returned by DeleteTags

DescribeTagsError
Errors returned by DescribeTags

DescribeWorkspaceBundlesError
Errors returned by DescribeWorkspaceBundles

DescribeWorkspaceDirectoriesError
Errors returned by DescribeWorkspaceDirectories

DescribeWorkspacesConnectionStatusError
Errors returned by DescribeWorkspacesConnectionStatus

DescribeWorkspacesError
Errors returned by DescribeWorkspaces

ModifyWorkspacePropertiesError
Errors returned by ModifyWorkspaceProperties

RebootWorkspacesError
Errors returned by RebootWorkspaces

RebuildWorkspacesError
Errors returned by RebuildWorkspaces

StartWorkspacesError
Errors returned by StartWorkspaces

StopWorkspacesError
Errors returned by StopWorkspaces

TerminateWorkspacesError
Errors returned by TerminateWorkspaces

Type Definitions

ARN

Alias

BooleanObject

BundleId

BundleIdList

BundleList

BundleOwner

Compute

ComputerName

ConnectionState

DefaultOu

Description

DirectoryId

DirectoryIdList

DirectoryList

DirectoryName

DnsIpAddresses

ErrorType

ExceptionMessage

FailedCreateWorkspaceRequests

FailedRebootWorkspaceRequests

FailedRebuildWorkspaceRequests

FailedStartWorkspaceRequests

FailedStopWorkspaceRequests

FailedTerminateWorkspaceRequests

IpAddress

Limit

NonEmptyString

PaginationToken

RebootWorkspaceRequests

RebuildWorkspaceRequests

RegistrationCode

RunningMode

RunningModeAutoStopTimeoutInMinutes

SecurityGroupId

StartWorkspaceRequests

StopWorkspaceRequests

SubnetId

SubnetIds

TagKey

TagKeyList

TagList

TagValue

TerminateWorkspaceRequests

Timestamp

UserName

VolumeEncryptionKey

WorkspaceConnectionStatusList

WorkspaceDirectoryState

WorkspaceDirectoryType

WorkspaceErrorCode

WorkspaceId

WorkspaceIdList

WorkspaceList

WorkspaceRequestList

WorkspaceState