Struct rusoto::workspaces::WorkspaceDirectory

pub struct WorkspaceDirectory {
    pub alias: Option<Alias>,
    pub customer_user_name: Option<UserName>,
    pub directory_id: Option<DirectoryId>,
    pub directory_name: Option<DirectoryName>,
    pub directory_type: Option<WorkspaceDirectoryType>,
    pub dns_ip_addresses: Option<DnsIpAddresses>,
    pub iam_role_id: Option<ARN>,
    pub registration_code: Option<RegistrationCode>,
    pub state: Option<WorkspaceDirectoryState>,
    pub subnet_ids: Option<SubnetIds>,
    pub workspace_creation_properties: Option<DefaultWorkspaceCreationProperties>,
    pub workspace_security_group_id: Option<SecurityGroupId>,
}

Contains information about an AWS Directory Service directory for use with Amazon WorkSpaces.

Fields
alias: Option<Alias>

The directory alias.

customer_user_name: Option<UserName>

The user name for the service account.

directory_id: Option<DirectoryId>

The directory identifier.

directory_name: Option<DirectoryName>

The name of the directory.

directory_type: Option<WorkspaceDirectoryType>

The directory type.

dns_ip_addresses: Option<DnsIpAddresses>

An array of strings that contains the IP addresses of the DNS servers for the directory.

iam_role_id: Option<ARN>

The identifier of the IAM role. This is the role that allows Amazon WorkSpaces to make calls to other services, such as Amazon EC2, on your behalf.

registration_code: Option<RegistrationCode>

The registration code for the directory. This is the code that users enter in their Amazon WorkSpaces client application to connect to the directory.

state: Option<WorkspaceDirectoryState>

The state of the directory's registration with Amazon WorkSpaces

subnet_ids: Option<SubnetIds>

An array of strings that contains the identifiers of the subnets used with the directory.

workspace_creation_properties: Option<DefaultWorkspaceCreationProperties>

A structure that specifies the default creation properties for all WorkSpaces in the directory.

workspace_security_group_id: Option<SecurityGroupId>

The identifier of the security group that is assigned to new WorkSpaces.