Struct rusoto::workspaces::WorkspaceRequest

pub struct WorkspaceRequest {
    pub bundle_id: BundleId,
    pub directory_id: DirectoryId,
    pub root_volume_encryption_enabled: Option<BooleanObject>,
    pub tags: Option<TagList>,
    pub user_name: UserName,
    pub user_volume_encryption_enabled: Option<BooleanObject>,
    pub volume_encryption_key: Option<VolumeEncryptionKey>,
}

Contains information about a WorkSpace creation request.

Fields
bundle_id: BundleId

The identifier of the bundle to create the WorkSpace from. You can use the DescribeWorkspaceBundles operation to obtain a list of the bundles that are available.

directory_id: DirectoryId

The identifier of the AWS Directory Service directory to create the WorkSpace in. You can use the DescribeWorkspaceDirectories operation to obtain a list of the directories that are available.

root_volume_encryption_enabled: Option<BooleanObject>

Specifies whether the data stored on the root volume, or C: drive, is encrypted.

tags: Option<TagList>

The tags of the WorkSpace request.

user_name: UserName

The username that the WorkSpace is assigned to. This username must exist in the AWS Directory Service directory specified by the DirectoryId member.

user_volume_encryption_enabled: Option<BooleanObject>

Specifies whether the data stored on the user volume, or D: drive, is encrypted.

volume_encryption_key: Option<VolumeEncryptionKey>

The KMS key used to encrypt data stored on your WorkSpace.