Module rusoto::xmlutil

Tools for handling XML from AWS with helper functions for testing

Structs

XmlParseError
generic Error for XML parsing

XmlResponseFromAws

XmlResponseFromFile

Traits

Next

Peek

Functions

characters
return some XML Characters

end_element
consume an EndElement with a specific name or throw an XmlParseError

optional_string_field
parse Some(String) if the next tag has the right name, otherwise None

peek_at_name
get the name of the current element in the stack. throw a parse error if it's not a StartElement

start_element
consume a StartElement with a specific name or throw an XmlParseError

string_field
return a string field with the right name or throw a parse error

Type Definitions

XmlStack
syntactic sugar for the XML event stack we pass around