Crate rusoto

Rusoto is an AWS SDK for Rust. A high level overview is available in README.md at https://github.com/rusoto/rusoto.

Modules

dynamodb
Amazon DynamoDB

ec2
Amazon EC2 Container Service

ecs
Amazon EC2 Container Service

ets
Amazon Elastic Transcoder

firehose
Amazon Kinesis Firehose

kinesis
Amazon Kinesis

kms
AWS Key Management Service

s3
The AWS S3 API.

sqs
The AWS SQS API.

Structs

AwsCredentials
AWS API access credentials, including access key, secret key, token (for IAM profiles), and expiration timestamp.

AwsError
An error produced when AWS API calls are unsuccessful.

ChainProvider
Provides AWS credentials from multiple possible sources using a priority order.

EnvironmentProvider
Provides AWS credentials from environment variables.

IamProvider
Provides AWS credentials from a resource's IAM role.

ParseRegionError
An error produced when attempting to convert a str into a Region fails.

ProfileProvider
Provides AWS credentials from a profile in a credentials file.

Enums

Region
An AWS region.

Traits

ProvideAwsCredentials
A trait for types that produce AwsCredentials.

Type Definitions

AwsResult
The result type produced by AWS API calls.