Struct rusoto::acm::CertificateDetail

pub struct CertificateDetail {
    pub certificate_arn: Option<Arn>,
    pub created_at: Option<TStamp>,
    pub domain_name: Option<DomainNameString>,
    pub domain_validation_options: Option<DomainValidationList>,
    pub in_use_by: Option<InUseList>,
    pub issued_at: Option<TStamp>,
    pub issuer: Option<String>,
    pub key_algorithm: Option<KeyAlgorithm>,
    pub not_after: Option<TStamp>,
    pub not_before: Option<TStamp>,
    pub revocation_reason: Option<RevocationReason>,
    pub revoked_at: Option<TStamp>,
    pub serial: Option<String>,
    pub signature_algorithm: Option<String>,
    pub status: Option<CertificateStatus>,
    pub subject: Option<String>,
    pub subject_alternative_names: Option<DomainList>,
}

This structure is returned in the response object of the DescribeCertificate action.

Fields
certificate_arn: Option<Arn>

Amazon Resource Name (ARN) of the certificate. This is of the form:

arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012

For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

created_at: Option<TStamp>

Time at which the certificate was requested.

domain_name: Option<DomainNameString>

Fully qualified domain name (FQDN), such as www.example.com or example.com, for the certificate.

domain_validation_options: Option<DomainValidationList>

References a DomainValidation structure that contains the domain name in the certificate and the email address that can be used for validation.

in_use_by: Option<InUseList>

List that identifies ARNs that are using the certificate. A single ACM Certificate can be used by multiple AWS resources.

issued_at: Option<TStamp>

Time at which the certificate was issued.

issuer: Option<String>

The X.500 distinguished name of the CA that issued and signed the certificate.

key_algorithm: Option<KeyAlgorithm>

Asymmetric algorithm used to generate the public and private key pair. Currently the only supported value is RSA_2048.

not_after: Option<TStamp>

Time after which the certificate is not valid.

not_before: Option<TStamp>

Time before which the certificate is not valid.

revocation_reason: Option<RevocationReason>

A RevocationReason enumeration value that indicates why the certificate was revoked. This value exists only if the certificate has been revoked. This can be one of the following vales:

UNSPECIFIED

KEY_COMPROMISE

CA_COMPROMISE

AFFILIATION_CHANGED

SUPERCEDED

CESSATION_OF_OPERATION

CERTIFICATE_HOLD

REMOVE_FROM_CRL

PRIVILEGE_WITHDRAWN

A_A_COMPROMISE

revoked_at: Option<TStamp>

The time, if any, at which the certificate was revoked. This value exists only if the certificate has been revoked.

serial: Option<String>

String that contains the serial number of the certificate.

signature_algorithm: Option<String>

Algorithm used to generate a signature. Currently the only supported value is SHA256WITHRSA.

status: Option<CertificateStatus>

A CertificateStatus enumeration value that can contain one of the following:

PENDING_VALIDATION

ISSUED

INACTIVE

EXPIRED

REVOKED

FAILED

VALIDATION_TIMED_OUT

subject: Option<String>

The X.500 distinguished name of the entity associated with the public key contained in the certificate.

subject_alternative_names: Option<DomainList>

One or more domain names (subject alternative names) included in the certificate request. After the certificate is issued, this list includes the domain names bound to the public key contained in the certificate. The subject alternative names include the canonical domain name (CN) of the certificate and additional domain names that can be used to connect to the website.