Module rusoto::autoscaling

AWS Auto Scaling

If you're using the service, you're probably looking for AutoscalingClient.

Structs

ActivitiesType
Contains the output of DescribeScalingActivities.

Activity
Describes scaling activity, which is a long-running process that represents a change to your Auto Scaling group, such as changing its size or replacing an instance.

ActivityType
Contains the output of TerminateInstancesInAutoScalingGroup.

AdjustmentType
Describes a policy adjustment type.

For more information, see Dynamic Scaling in the Auto Scaling User Guide.

Alarm
Describes an alarm.

AttachInstancesQuery
Contains the parameters for AttachInstances.

AttachLoadBalancerTargetGroupsResultType

AttachLoadBalancerTargetGroupsType
Contains the parameters for AttachLoadBalancerTargetGroups.

AttachLoadBalancersResultType
Contains the output of AttachLoadBalancers.

AttachLoadBalancersType
Contains the parameters for AttachLoadBalancers.

AutoScalingGroup
Describes an Auto Scaling group.

AutoScalingGroupNamesType
Contains the parameters for DescribeAutoScalingGroups.

AutoScalingGroupsType
Contains the output for DescribeAutoScalingGroups.

AutoScalingInstanceDetails
Describes an EC2 instance associated with an Auto Scaling group.

AutoScalingInstancesType
Contains the output of DescribeAutoScalingInstances.

AutoscalingClient
A client for the Auto Scaling API.

BlockDeviceMapping
Describes a block device mapping.

CompleteLifecycleActionAnswer
Contains the output of CompleteLifecycleAction.

CompleteLifecycleActionType
Contains the parameters for CompleteLifecycleAction.

CreateAutoScalingGroupType
Contains the parameters for CreateAutoScalingGroup.

CreateLaunchConfigurationType
Contains the parameters for CreateLaunchConfiguration.

CreateOrUpdateTagsType
Contains the parameters for CreateOrUpdateTags.

DeleteAutoScalingGroupType
Contains the parameters for DeleteAutoScalingGroup.

DeleteLifecycleHookAnswer
Contains the output of DeleteLifecycleHook.

DeleteLifecycleHookType
Contains the parameters for DeleteLifecycleHook.

DeleteNotificationConfigurationType
Contains the parameters for DeleteNotificationConfiguration.

DeletePolicyType
Contains the parameters for DeletePolicy.

DeleteScheduledActionType
Contains the parameters for DeleteScheduledAction.

DeleteTagsType
Contains the parameters for DeleteTags.

DescribeAccountLimitsAnswer
Contains the parameters for DescribeAccountLimits.

DescribeAdjustmentTypesAnswer
Contains the parameters for DescribeAdjustmentTypes.

DescribeAutoScalingInstancesType
Contains the parameters for DescribeAutoScalingInstances.

DescribeAutoScalingNotificationTypesAnswer
Contains the output of DescribeAutoScalingNotificationTypes.

DescribeLifecycleHookTypesAnswer
Contains the output of DescribeLifecycleHookTypes.

DescribeLifecycleHooksAnswer
Contains the output of DescribeLifecycleHooks.

DescribeLifecycleHooksType
Contains the parameters for DescribeLifecycleHooks.

DescribeLoadBalancerTargetGroupsRequest
Contains the parameters for DescribeLoadBalancerTargetGroups.

DescribeLoadBalancerTargetGroupsResponse
Contains the output of DescribeLoadBalancerTargetGroups.

DescribeLoadBalancersRequest
Contains the parameters for DescribeLoadBalancers.

DescribeLoadBalancersResponse
Contains the output of DescribeLoadBalancers.

DescribeMetricCollectionTypesAnswer
Contains the output of DescribeMetricsCollectionTypes.

DescribeNotificationConfigurationsAnswer
Contains the output from DescribeNotificationConfigurations.

DescribeNotificationConfigurationsType
Contains the parameters for DescribeNotificationConfigurations.

DescribePoliciesType
Contains the parameters for DescribePolicies.

DescribeScalingActivitiesType
Contains the parameters for DescribeScalingActivities.

DescribeScheduledActionsType
Contains the parameters for DescribeScheduledActions.

DescribeTagsType
Contains the parameters for DescribeTags.

DescribeTerminationPolicyTypesAnswer
Contains the output of DescribeTerminationPolicyTypes.

DetachInstancesAnswer
Contains the output of DetachInstances.

DetachInstancesQuery
Contains the parameters for DetachInstances.

DetachLoadBalancerTargetGroupsResultType

DetachLoadBalancerTargetGroupsType

DetachLoadBalancersResultType
Contains the output for DetachLoadBalancers.

DetachLoadBalancersType
Contains the parameters for DetachLoadBalancers.

DisableMetricsCollectionQuery
Contains the parameters for DisableMetricsCollection.

Ebs
Describes an Amazon EBS volume.

EnableMetricsCollectionQuery
Contains the parameters for EnableMetricsCollection.

EnabledMetric
Describes an enabled metric.

EnterStandbyAnswer
Contains the output of EnterStandby.

EnterStandbyQuery
Contains the parameters for EnteStandby.

ExecutePolicyType
Contains the parameters for ExecutePolicy.

ExitStandbyAnswer
Contains the parameters for ExitStandby.

ExitStandbyQuery
Contains the parameters for ExitStandby.

Filter
Describes a filter.

Instance
Describes an EC2 instance.

InstanceMonitoring
Describes whether instance monitoring is enabled.

LaunchConfiguration
Describes a launch configuration.

LaunchConfigurationNameType
Contains the parameters for DeleteLaunchConfiguration.

LaunchConfigurationNamesType
Contains the parameters for DescribeLaunchConfigurations.

LaunchConfigurationsType
Contains the output of DescribeLaunchConfigurations.

LifecycleHook
Describes a lifecycle hook, which tells Auto Scaling that you want to perform an action when an instance launches or terminates. When you have a lifecycle hook in place, the Auto Scaling group will either:

Pause the instance after it launches, but before it is put into service

Pause the instance as it terminates, but before it is fully terminated

For more information, see Auto Scaling Lifecycle in the Auto Scaling User Guide.

LoadBalancerState
Describes the state of a Classic load balancer.

If you specify a load balancer when creating the Auto Scaling group, the state of the load balancer is InService.

If you attach a load balancer to an existing Auto Scaling group, the initial state is Adding. The state transitions to Added after all instances in the group are registered with the load balancer. If ELB health checks are enabled for the load balancer, the state transitions to InService after at least one instance in the group passes the health check. If EC2 health checks are enabled instead, the load balancer remains in the Added state.

LoadBalancerTargetGroupState
Describes the state of a target group.

If you attach a target group to an existing Auto Scaling group, the initial state is Adding. The state transitions to Added after all Auto Scaling instances are registered with the target group. If ELB health checks are enabled, the state transitions to InService after at least one Auto Scaling instance passes the health check. If EC2 health checks are enabled instead, the target group remains in the Added state.

MetricCollectionType
Describes a metric.

MetricGranularityType
Describes a granularity of a metric.

NotificationConfiguration
Describes a notification.

PoliciesType
Contains the output of DescribePolicies.

PolicyARNType
Contains the output of PutScalingPolicy.

ProcessType
Describes a process type.

For more information, see Auto Scaling Processes in the Auto Scaling User Guide.

ProcessesType
Contains the output of DescribeScalingProcessTypes.

PutLifecycleHookAnswer
Contains the output of PutLifecycleHook.

PutLifecycleHookType
Contains the parameters for PutLifecycleHook.

PutNotificationConfigurationType
Contains the parameters for PutNotificationConfiguration.

PutScalingPolicyType
Contains the parameters for PutScalingPolicy.

PutScheduledUpdateGroupActionType
Contains the parameters for PutScheduledUpdateGroupAction.

RecordLifecycleActionHeartbeatAnswer
Contains the output of RecordLifecycleActionHeartBeat.

RecordLifecycleActionHeartbeatType
Contains the parameters for RecordLifecycleActionHeartbeat.

ScalingPolicy
Describes a scaling policy.

ScalingProcessQuery
Contains the parameters for SuspendProcesses and ResumeProcesses.

ScheduledActionsType
Contains the output of DescribeScheduledActions.

ScheduledUpdateGroupAction
Describes a scheduled update to an Auto Scaling group.

SetDesiredCapacityType
Contains the parameters for SetDesiredCapacity.

SetInstanceHealthQuery
Contains the parameters for SetInstanceHealth.

SetInstanceProtectionAnswer
Contains the output of SetInstanceProtection.

SetInstanceProtectionQuery
Contains the parameters for SetInstanceProtection.

StepAdjustment

Describes an adjustment based on the difference between the value of the aggregated CloudWatch metric and the breach threshold that you've defined for the alarm.

For the following examples, suppose that you have an alarm with a breach threshold of 50:

If you want the adjustment to be triggered when the metric is greater than or equal to 50 and less than 60, specify a lower bound of 0 and an upper bound of 10.

If you want the adjustment to be triggered when the metric is greater than 40 and less than or equal to 50, specify a lower bound of -10 and an upper bound of 0.

There are a few rules for the step adjustments for your step policy:

The ranges of your step adjustments can't overlap or have a gap.

At most one step adjustment can have a null lower bound. If one step adjustment has a negative lower bound, then there must be a step adjustment with a null lower bound.

At most one step adjustment can have a null upper bound. If one step adjustment has a positive upper bound, then there must be a step adjustment with a null upper bound.

The upper and lower bound can't be null in the same step adjustment.

SuspendedProcess
Describes an Auto Scaling process that has been suspended. For more information, see ProcessType.

Tag
Describes a tag for an Auto Scaling group.

TagDescription
Describes a tag for an Auto Scaling group.

TagsType
Contains the output of DescribeTags.

TerminateInstanceInAutoScalingGroupType
Contains the parameters for TerminateInstanceInAutoScalingGroup.

UpdateAutoScalingGroupType
Contains the parameters for UpdateAutoScalingGroup.

Enums

AttachInstancesError
Errors returned by AttachInstances

AttachLoadBalancerTargetGroupsError
Errors returned by AttachLoadBalancerTargetGroups

AttachLoadBalancersError
Errors returned by AttachLoadBalancers

CompleteLifecycleActionError
Errors returned by CompleteLifecycleAction

CreateAutoScalingGroupError
Errors returned by CreateAutoScalingGroup

CreateLaunchConfigurationError
Errors returned by CreateLaunchConfiguration

CreateOrUpdateTagsError
Errors returned by CreateOrUpdateTags

DeleteAutoScalingGroupError
Errors returned by DeleteAutoScalingGroup

DeleteLaunchConfigurationError
Errors returned by DeleteLaunchConfiguration

DeleteLifecycleHookError
Errors returned by DeleteLifecycleHook

DeleteNotificationConfigurationError
Errors returned by DeleteNotificationConfiguration

DeletePolicyError
Errors returned by DeletePolicy

DeleteScheduledActionError
Errors returned by DeleteScheduledAction

DeleteTagsError
Errors returned by DeleteTags

DescribeAccountLimitsError
Errors returned by DescribeAccountLimits

DescribeAdjustmentTypesError
Errors returned by DescribeAdjustmentTypes

DescribeAutoScalingGroupsError
Errors returned by DescribeAutoScalingGroups

DescribeAutoScalingInstancesError
Errors returned by DescribeAutoScalingInstances

DescribeAutoScalingNotificationTypesError
Errors returned by DescribeAutoScalingNotificationTypes

DescribeLaunchConfigurationsError
Errors returned by DescribeLaunchConfigurations

DescribeLifecycleHookTypesError
Errors returned by DescribeLifecycleHookTypes

DescribeLifecycleHooksError
Errors returned by DescribeLifecycleHooks

DescribeLoadBalancerTargetGroupsError
Errors returned by DescribeLoadBalancerTargetGroups

DescribeLoadBalancersError
Errors returned by DescribeLoadBalancers

DescribeMetricCollectionTypesError
Errors returned by DescribeMetricCollectionTypes

DescribeNotificationConfigurationsError
Errors returned by DescribeNotificationConfigurations

DescribePoliciesError
Errors returned by DescribePolicies

DescribeScalingActivitiesError
Errors returned by DescribeScalingActivities

DescribeScalingProcessTypesError
Errors returned by DescribeScalingProcessTypes

DescribeScheduledActionsError
Errors returned by DescribeScheduledActions

DescribeTagsError
Errors returned by DescribeTags

DescribeTerminationPolicyTypesError
Errors returned by DescribeTerminationPolicyTypes

DetachInstancesError
Errors returned by DetachInstances

DetachLoadBalancerTargetGroupsError
Errors returned by DetachLoadBalancerTargetGroups

DetachLoadBalancersError
Errors returned by DetachLoadBalancers

DisableMetricsCollectionError
Errors returned by DisableMetricsCollection

EnableMetricsCollectionError
Errors returned by EnableMetricsCollection

EnterStandbyError
Errors returned by EnterStandby

ExecutePolicyError
Errors returned by ExecutePolicy

ExitStandbyError
Errors returned by ExitStandby

PutLifecycleHookError
Errors returned by PutLifecycleHook

PutNotificationConfigurationError
Errors returned by PutNotificationConfiguration

PutScalingPolicyError
Errors returned by PutScalingPolicy

PutScheduledUpdateGroupActionError
Errors returned by PutScheduledUpdateGroupAction

RecordLifecycleActionHeartbeatError
Errors returned by RecordLifecycleActionHeartbeat

ResumeProcessesError
Errors returned by ResumeProcesses

SetDesiredCapacityError
Errors returned by SetDesiredCapacity

SetInstanceHealthError
Errors returned by SetInstanceHealth

SetInstanceProtectionError
Errors returned by SetInstanceProtection

SuspendProcessesError
Errors returned by SuspendProcesses

TerminateInstanceInAutoScalingGroupError
Errors returned by TerminateInstanceInAutoScalingGroup

UpdateAutoScalingGroupError
Errors returned by UpdateAutoScalingGroup

Type Definitions

Activities

ActivityIds

AdjustmentTypes

Alarms

AsciiStringMaxLen255

AssociatePublicIpAddress

AutoScalingGroupDesiredCapacity

AutoScalingGroupMaxSize

AutoScalingGroupMinSize

AutoScalingGroupNames

AutoScalingGroups

AutoScalingInstances

AutoScalingNotificationTypes

AvailabilityZones

BlockDeviceEbsDeleteOnTermination

BlockDeviceEbsEncrypted

BlockDeviceEbsIops

BlockDeviceEbsVolumeSize

BlockDeviceEbsVolumeType

BlockDeviceMappings

ClassicLinkVPCSecurityGroups

Cooldown

EbsOptimized

EnabledMetrics

EstimatedInstanceWarmup

Filters

ForceDelete

GlobalTimeout

HealthCheckGracePeriod

HeartbeatTimeout

HonorCooldown

InstanceIds

InstanceProtected

Instances

LaunchConfigurationNames

LaunchConfigurations

LifecycleActionResult

LifecycleActionToken

LifecycleHookNames

LifecycleHooks

LifecycleState

LifecycleTransition

LoadBalancerNames

LoadBalancerStates

LoadBalancerTargetGroupStates

MaxNumberOfAutoScalingGroups

MaxNumberOfLaunchConfigurations

MaxRecords

MetricCollectionTypes

MetricGranularityTypes

MetricScale

Metrics

MinAdjustmentMagnitude

MinAdjustmentStep

MonitoringEnabled

NoDevice

NotificationConfigurations

NotificationTargetResourceName

NumberOfAutoScalingGroups

NumberOfLaunchConfigurations

PolicyIncrement

PolicyNames

PolicyTypes

ProcessNames

Processes

Progress

PropagateAtLaunch

ProtectedFromScaleIn

ResourceName

ScalingActivityStatusCode

ScalingPolicies

ScheduledActionNames

ScheduledUpdateGroupActions

SecurityGroups

ShouldDecrementDesiredCapacity

ShouldRespectGracePeriod

SpotPrice

StepAdjustments

SuspendedProcesses

TagDescriptionList

TagKey

TagValue

Tags

TargetGroupARNs

TerminationPolicies

TimestampType

Values

XmlString

XmlStringMaxLen1023

XmlStringMaxLen1600

XmlStringMaxLen19

XmlStringMaxLen2047

XmlStringMaxLen255

XmlStringMaxLen32

XmlStringMaxLen511

XmlStringMaxLen64

XmlStringUserData