Struct rusoto::cloudfront::CloudFrontClient

pub struct CloudFrontClient<P, D> where P: ProvideAwsCredentials,
        D: DispatchSignedRequest { /* fields omitted */ }

A client for the CloudFront API.

Methods

impl<P, D> CloudFrontClient<P, D> where P: ProvideAwsCredentials,
D: DispatchSignedRequest

fn new(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn create_cloud_front_origin_access_identity(&self,
input: &CreateCloudFrontOriginAccessIdentityRequest)
-> Result<CreateCloudFrontOriginAccessIdentityResult, CreateCloudFrontOriginAccessIdentityError>

Creates a new origin access identity. If you're using Amazon S3 for your origin, you can use an origin access identity to require users to access your content using a CloudFront URL instead of the Amazon S3 URL. For more information about how to use origin access identities, see Serving Private Content through CloudFront in the Amazon CloudFront Developer Guide.

fn create_distribution(&self,
input: &CreateDistributionRequest)
-> Result<CreateDistributionResult, CreateDistributionError>

Creates a new web distribution. Send a GET request to the /CloudFront API version/distribution/distribution ID resource.

fn create_distribution_with_tags(&self,
input: &CreateDistributionWithTagsRequest)
-> Result<CreateDistributionWithTagsResult, CreateDistributionWithTagsError>

Create a new distribution with tags.

fn create_invalidation(&self,
input: &CreateInvalidationRequest)
-> Result<CreateInvalidationResult, CreateInvalidationError>

Create a new invalidation.

fn create_streaming_distribution(&self,
input: &CreateStreamingDistributionRequest)
-> Result<CreateStreamingDistributionResult, CreateStreamingDistributionError>

Creates a new RMTP distribution. An RTMP distribution is similar to a web distribution, but an RTMP distribution streams media files using the Adobe Real-Time Messaging Protocol (RTMP) instead of serving files using HTTP.

To create a new web distribution, submit a POST request to the CloudFront API version/distribution resource. The request body must include a document with a StreamingDistributionConfig element. The response echoes the StreamingDistributionConfig element and returns other information about the RTMP distribution.

To get the status of your request, use the GET StreamingDistribution API action. When the value of Enabled is true and the value of Status is Deployed, your distribution is ready. A distribution usually deploys in less than 15 minutes.

For more information about web distributions, see Working with RTMP Distributions in the Amazon CloudFront Developer Guide.

Beginning with the 2012-05-05 version of the CloudFront API, we made substantial changes to the format of the XML document that you include in the request body when you create or update a web distribution or an RTMP distribution, and when you invalidate objects. With previous versions of the API, we discovered that it was too easy to accidentally delete one or more values for an element that accepts multiple values, for example, CNAMEs and trusted signers. Our changes for the 2012-05-05 release are intended to prevent these accidental deletions and to notify you when there's a mismatch between the number of values you say you're specifying in the Quantity element and the number of values specified.

fn create_streaming_distribution_with_tags(&self,
input: &CreateStreamingDistributionWithTagsRequest)
-> Result<CreateStreamingDistributionWithTagsResult, CreateStreamingDistributionWithTagsError>

Create a new streaming distribution with tags.

fn delete_cloud_front_origin_access_identity(&self,
input: &DeleteCloudFrontOriginAccessIdentityRequest)
-> Result<(), DeleteCloudFrontOriginAccessIdentityError>

Delete an origin access identity.

fn delete_distribution(&self,
input: &DeleteDistributionRequest)
-> Result<(), DeleteDistributionError>

Delete a distribution.

fn delete_streaming_distribution(&self,
input: &DeleteStreamingDistributionRequest)
-> Result<(), DeleteStreamingDistributionError>

Delete a streaming distribution. To delete an RTMP distribution using the CloudFront API, perform the following steps.

To delete an RTMP distribution using the CloudFront API:

Disable the RTMP distribution.

Submit a GET Streaming Distribution Config request to get the current configuration and the Etag header for the distribution.

Update the XML document that was returned in the response to your GET Streaming Distribution Config request to change the value of Enabled to false.

Submit a PUT Streaming Distribution Config request to update the configuration for your distribution. In the request body, include the XML document that you updated in Step 3. Then set the value of the HTTP If-Match header to the value of the ETag header that CloudFront returned when you submitted the GET Streaming Distribution Config request in Step 2.

Review the response to the PUT Streaming Distribution Config request to confirm that the distribution was successfully disabled.

Submit a GET Streaming Distribution Config request to confirm that your changes have propagated. When propagation is complete, the value of Status is Deployed.

Submit a DELETE Streaming Distribution request. Set the value of the HTTP If-Match header to the value of the ETag header that CloudFront returned when you submitted the GET Streaming Distribution Config request in Step 2.

Review the response to your DELETE Streaming Distribution request to confirm that the distribution was successfully deleted.

For information about deleting a distribution using the CloudFront console, see Deleting a Distribution in the Amazon CloudFront Developer Guide.

fn get_cloud_front_origin_access_identity(&self,
input: &GetCloudFrontOriginAccessIdentityRequest)
-> Result<GetCloudFrontOriginAccessIdentityResult, GetCloudFrontOriginAccessIdentityError>

Get the information about an origin access identity.

fn get_cloud_front_origin_access_identity_config(&self,
input: &GetCloudFrontOriginAccessIdentityConfigRequest)
-> Result<GetCloudFrontOriginAccessIdentityConfigResult, GetCloudFrontOriginAccessIdentityConfigError>

Get the configuration information about an origin access identity.

fn get_distribution(&self,
input: &GetDistributionRequest)
-> Result<GetDistributionResult, GetDistributionError>

Get the information about a distribution.

fn get_distribution_config(&self,
input: &GetDistributionConfigRequest)
-> Result<GetDistributionConfigResult, GetDistributionConfigError>

Get the configuration information about a distribution.

fn get_invalidation(&self,
input: &GetInvalidationRequest)
-> Result<GetInvalidationResult, GetInvalidationError>

Get the information about an invalidation.

fn get_streaming_distribution(&self,
input: &GetStreamingDistributionRequest)
-> Result<GetStreamingDistributionResult, GetStreamingDistributionError>

Gets information about a specified RTMP distribution, including the distribution configuration.

fn get_streaming_distribution_config(&self,
input: &GetStreamingDistributionConfigRequest)
-> Result<GetStreamingDistributionConfigResult, GetStreamingDistributionConfigError>

Get the configuration information about a streaming distribution.

fn list_cloud_front_origin_access_identities(&self,
input: &ListCloudFrontOriginAccessIdentitiesRequest)
-> Result<ListCloudFrontOriginAccessIdentitiesResult, ListCloudFrontOriginAccessIdentitiesError>

Lists origin access identities.

fn list_distributions(&self,
input: &ListDistributionsRequest)
-> Result<ListDistributionsResult, ListDistributionsError>

List distributions.

fn list_distributions_by_web_acl_id(&self,
input: &ListDistributionsByWebACLIdRequest)
-> Result<ListDistributionsByWebACLIdResult, ListDistributionsByWebACLIdError>

List the distributions that are associated with a specified AWS WAF web ACL.

fn list_invalidations(&self,
input: &ListInvalidationsRequest)
-> Result<ListInvalidationsResult, ListInvalidationsError>

Lists invalidation batches.

fn list_streaming_distributions(&self,
input: &ListStreamingDistributionsRequest)
-> Result<ListStreamingDistributionsResult, ListStreamingDistributionsError>

List streaming distributions.

fn list_tags_for_resource(&self,
input: &ListTagsForResourceRequest)
-> Result<ListTagsForResourceResult, ListTagsForResourceError>

List tags for a CloudFront resource.

fn tag_resource(&self,
input: &TagResourceRequest)
-> Result<(), TagResourceError>

Add tags to a CloudFront resource.

fn untag_resource(&self,
input: &UntagResourceRequest)
-> Result<(), UntagResourceError>

Remove tags from a CloudFront resource.

fn update_cloud_front_origin_access_identity(&self,
input: &UpdateCloudFrontOriginAccessIdentityRequest)
-> Result<UpdateCloudFrontOriginAccessIdentityResult, UpdateCloudFrontOriginAccessIdentityError>

Update an origin access identity.

fn update_distribution(&self,
input: &UpdateDistributionRequest)
-> Result<UpdateDistributionResult, UpdateDistributionError>

Update a distribution.

fn update_streaming_distribution(&self,
input: &UpdateStreamingDistributionRequest)
-> Result<UpdateStreamingDistributionResult, UpdateStreamingDistributionError>

Update a streaming distribution.