Struct rusoto::cloudtrail::CloudTrailClient

pub struct CloudTrailClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the CloudTrail API.

Methods

impl<P> CloudTrailClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> CloudTrailClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn add_tags(&self, input: &AddTagsRequest) -> Result<AddTagsResponse, AddTagsError>

Adds one or more tags to a trail, up to a limit of 10. Tags must be unique per trail. Overwrites an existing tag's value when a new value is specified for an existing tag key. If you specify a key without a value, the tag will be created with the specified key and a value of null. You can tag a trail that applies to all regions only from the region in which the trail was created (that is, from its home region).

fn create_trail(&self, input: &CreateTrailRequest) -> Result<CreateTrailResponse, CreateTrailError>

Creates a trail that specifies the settings for delivery of log data to an Amazon S3 bucket. A maximum of five trails can exist in a region, irrespective of the region in which they were created.

fn delete_trail(&self, input: &DeleteTrailRequest) -> Result<DeleteTrailResponse, DeleteTrailError>

Deletes a trail. This operation must be called from the region in which the trail was created. DeleteTrail cannot be called on the shadow trails (replicated trails in other regions) of a trail that is enabled in all regions.

fn describe_trails(&self, input: &DescribeTrailsRequest) -> Result<DescribeTrailsResponse, DescribeTrailsError>

Retrieves settings for the trail associated with the current region for your account.

fn get_trail_status(&self, input: &GetTrailStatusRequest) -> Result<GetTrailStatusResponse, GetTrailStatusError>

Returns a JSON-formatted list of information about the specified trail. Fields include information on delivery errors, Amazon SNS and Amazon S3 errors, and start and stop logging times for each trail. This operation returns trail status from a single region. To return trail status from all regions, you must call the operation on each region.

fn list_public_keys(&self, input: &ListPublicKeysRequest) -> Result<ListPublicKeysResponse, ListPublicKeysError>

Returns all public keys whose private keys were used to sign the digest files within the specified time range. The public key is needed to validate digest files that were signed with its corresponding private key.

CloudTrail uses different private/public key pairs per region. Each digest file is signed with a private key unique to its region. Therefore, when you validate a digest file from a particular region, you must look in the same region for its corresponding public key.

fn list_tags(&self, input: &ListTagsRequest) -> Result<ListTagsResponse, ListTagsError>

Lists the tags for the trail in the current region.

fn lookup_events(&self, input: &LookupEventsRequest) -> Result<LookupEventsResponse, LookupEventsError>

Looks up API activity events captured by CloudTrail that create, update, or delete resources in your account. Events for a region can be looked up for the times in which you had CloudTrail turned on in that region during the last seven days. Lookup supports five different attributes: time range (defined by a start time and end time), user name, event name, resource type, and resource name. All attributes are optional. The maximum number of attributes that can be specified in any one lookup request are time range and one other attribute. The default number of results returned is 10, with a maximum of 50 possible. The response includes a token that you can use to get the next page of results.

The rate of lookup requests is limited to one per second per account. If this limit is exceeded, a throttling error occurs.

Events that occurred during the selected time range will not be available for lookup if CloudTrail logging was not enabled when the events occurred.

fn remove_tags(&self, input: &RemoveTagsRequest) -> Result<RemoveTagsResponse, RemoveTagsError>

Removes the specified tags from a trail.

fn start_logging(&self, input: &StartLoggingRequest) -> Result<StartLoggingResponse, StartLoggingError>

Starts the recording of AWS API calls and log file delivery for a trail. For a trail that is enabled in all regions, this operation must be called from the region in which the trail was created. This operation cannot be called on the shadow trails (replicated trails in other regions) of a trail that is enabled in all regions.

fn stop_logging(&self, input: &StopLoggingRequest) -> Result<StopLoggingResponse, StopLoggingError>

Suspends the recording of AWS API calls and log file delivery for the specified trail. Under most circumstances, there is no need to use this action. You can update a trail without stopping it first. This action is the only way to stop recording. For a trail enabled in all regions, this operation must be called from the region in which the trail was created, or an InvalidHomeRegionException will occur. This operation cannot be called on the shadow trails (replicated trails in other regions) of a trail enabled in all regions.

fn update_trail(&self, input: &UpdateTrailRequest) -> Result<UpdateTrailResponse, UpdateTrailError>

Updates the settings that specify delivery of log files. Changes to a trail do not require stopping the CloudTrail service. Use this action to designate an existing bucket for log delivery. If the existing bucket has previously been a target for CloudTrail log files, an IAM policy exists for the bucket. UpdateTrail must be called from the region in which the trail was created; otherwise, an InvalidHomeRegionException is thrown.