Struct rusoto::cloudtrail::DescribeTrailsRequest

pub struct DescribeTrailsRequest {
    pub include_shadow_trails: Option<Boolean>,
    pub trail_name_list: Option<TrailNameList>,
}

Returns information about the trail.

Fields
include_shadow_trails: Option<Boolean>

Specifies whether to include shadow trails in the response. A shadow trail is the replication in a region of a trail that was created in a different region. The default is true.

trail_name_list: Option<TrailNameList>

Specifies a list of trail names, trail ARNs, or both, of the trails to describe. The format of a trail ARN is:

arn:aws:cloudtrail:us-east-1:123456789012:trail/MyTrail

If an empty list is specified, information for the trail in the current region is returned.

If an empty list is specified and IncludeShadowTrails is false, then information for all trails in the current region is returned.

If an empty list is specified and IncludeShadowTrails is null or true, then information for all trails in the current region and any associated shadow trails in other regions is returned.

If one or more trail names are specified, information is returned only if the names match the names of trails belonging only to the current region. To return information about a trail in another region, you must specify its trail ARN.