Module rusoto::cloudwatch

AWS CloudWatch

Structs

AlarmHistoryItem
The AlarmHistoryItem data type contains descriptive information about the history of a specific alarm. If you call DescribeAlarmHistory, Amazon CloudWatch returns this data type as part of the DescribeAlarmHistoryResult data type.

CloudWatchClient
A client for the CloudWatch API.

Datapoint
The Datapoint data type encapsulates the statistical data that Amazon CloudWatch computes from metric data.

DeleteAlarmsInput
Describes the inputs for DeleteAlarms.

DescribeAlarmHistoryInput
Describes the inputs for DescribeAlarmHistory.

DescribeAlarmHistoryOutput
The output for DescribeAlarmHistory.

DescribeAlarmsForMetricInput
Describes the inputs for DescribeAlarmsForMetric.

DescribeAlarmsForMetricOutput
The output for DescribeAlarmsForMetric.

DescribeAlarmsInput
Describes the inputs for DescribeAlarms.

DescribeAlarmsOutput
The output for DescribeAlarms.

Dimension
The Dimension data type further expands on the identity of a metric using a Name, Value pair.

For examples that use one or more dimensions, see PutMetricData.

DimensionFilter
The DimensionFilter data type is used to filter ListMetrics results.

DisableAlarmActionsInput

EnableAlarmActionsInput
Describes the inputs for EnableAlarmActions.

GetMetricStatisticsInput
Describes the inputs for GetMetricStatistics.

GetMetricStatisticsOutput
The output for GetMetricStatistics.

ListMetricsInput
Describes the inputs for ListMetrics.

ListMetricsOutput
The output for ListMetrics.

Metric
The Metric data type contains information about a specific metric. If you call ListMetrics, Amazon CloudWatch returns information contained by this data type.

The example in the Examples section publishes two metrics named buffers and latency. Both metrics are in the examples namespace. Both metrics have two dimensions, InstanceID and InstanceType.

MetricAlarm
The MetricAlarm data type represents an alarm. You can use PutMetricAlarm to create or update an alarm.

MetricDatum
The MetricDatum data type encapsulates the information sent with PutMetricData to either create a new metric or add new values to be aggregated into an existing metric.

PutMetricAlarmInput
Describes the inputs for PutMetricAlarm.

PutMetricDataInput
Describes the inputs for PutMetricData.

SetAlarmStateInput
Describes the inputs for SetAlarmState.

StatisticSet
The StatisticSet data type describes the StatisticValues component of MetricDatum, and represents a set of statistics that describes a specific metric.

Enums

DeleteAlarmsError

DescribeAlarmHistoryError

DescribeAlarmsError

DescribeAlarmsForMetricError

DisableAlarmActionsError

EnableAlarmActionsError

GetMetricStatisticsError

ListMetricsError

PutMetricAlarmError

PutMetricDataError

SetAlarmStateError

Type Definitions

ActionPrefix

ActionsEnabled

AlarmArn

AlarmDescription

AlarmHistoryItems

AlarmName

AlarmNamePrefix

AlarmNames

AwsQueryErrorMessage

ComparisonOperator

DatapointValue

Datapoints

DimensionFilters

DimensionName

DimensionValue

Dimensions

ErrorMessage

EvaluationPeriods

FaultDescription

HistoryData

HistoryItemType

HistorySummary

MaxRecords

MetricAlarms

MetricData

MetricLabel

MetricName

Metrics

Namespace

NextToken

Period

ResourceList

ResourceName

StandardUnit

StateReason

StateReasonData

StateValue

Statistic

Statistics

Threshold

Timestamp