Module rusoto::cloudwatch

AWS CloudWatch

If you're using the service, you're probably looking for CloudWatchClient.

Structs

AlarmHistoryItem
Represents the history of a specific alarm.

CloudWatchClient
A client for the CloudWatch API.

Datapoint
Encapsulates the statistical data that Amazon CloudWatch computes from metric data.

DeleteAlarmsInput

DescribeAlarmHistoryInput

DescribeAlarmHistoryOutput

DescribeAlarmsForMetricInput

DescribeAlarmsForMetricOutput

DescribeAlarmsInput

DescribeAlarmsOutput

Dimension
Expands the identity of a metric.

DimensionFilter
Represents filters for a dimension.

DisableAlarmActionsInput

EnableAlarmActionsInput

GetMetricStatisticsInput

GetMetricStatisticsOutput

ListMetricsInput

ListMetricsOutput

Metric
Represents a specific metric.

MetricAlarm
Represents an alarm.

MetricDatum
Encapsulates the information sent to either create a metric or add new values to be aggregated into an existing metric.

PutMetricAlarmInput

PutMetricDataInput

SetAlarmStateInput

StatisticSet
Represents a set of statistics that describes a specific metric.

Enums

DeleteAlarmsError
Errors returned by DeleteAlarms

DescribeAlarmHistoryError
Errors returned by DescribeAlarmHistory

DescribeAlarmsError
Errors returned by DescribeAlarms

DescribeAlarmsForMetricError
Errors returned by DescribeAlarmsForMetric

DisableAlarmActionsError
Errors returned by DisableAlarmActions

EnableAlarmActionsError
Errors returned by EnableAlarmActions

GetMetricStatisticsError
Errors returned by GetMetricStatistics

ListMetricsError
Errors returned by ListMetrics

PutMetricAlarmError
Errors returned by PutMetricAlarm

PutMetricDataError
Errors returned by PutMetricData

SetAlarmStateError
Errors returned by SetAlarmState

Type Definitions

ActionPrefix

ActionsEnabled

AlarmArn

AlarmDescription

AlarmHistoryItems

AlarmName

AlarmNamePrefix

AlarmNames

AwsQueryErrorMessage

ComparisonOperator

DatapointValue

DatapointValueMap

Datapoints

DimensionFilters

DimensionName

DimensionValue

Dimensions

ErrorMessage

EvaluationPeriods

ExtendedStatistic

ExtendedStatistics

FaultDescription

HistoryData

HistoryItemType

HistorySummary

MaxRecords

MetricAlarms

MetricData

MetricLabel

MetricName

Metrics

Namespace

NextToken

Period

ResourceList

ResourceName

StandardUnit

StateReason

StateReasonData

StateValue

Statistic

Statistics

Threshold

Timestamp