Module rusoto::config

AWS Config

Structs

Compliance
Indicates whether an AWS resource or AWS Config rule is compliant and provides the number of contributors that affect the compliance.

ComplianceByConfigRule
Indicates whether an AWS Config rule is compliant. A rule is compliant if all of the resources that the rule evaluated comply with it, and it is noncompliant if any of these resources do not comply.

ComplianceByResource
Indicates whether an AWS resource that is evaluated according to one or more AWS Config rules is compliant. A resource is compliant if it complies with all of the rules that evaluate it, and it is noncompliant if it does not comply with one or more of these rules.

ComplianceContributorCount
The number of AWS resources or AWS Config rules responsible for the current compliance of the item, up to a maximum number.

ComplianceSummary
The number of AWS Config rules or AWS resources that are compliant and noncompliant, up to a maximum.

ComplianceSummaryByResourceType
The number of AWS resources of a specific type that are compliant or noncompliant, up to a maximum of 100 for each compliance.

ConfigExportDeliveryInfo
A list that contains the status of the delivery of either the snapshot or the configuration history to the specified Amazon S3 bucket.

ConfigRule
An AWS Lambda function that evaluates configuration items to assess whether your AWS resources comply with your desired configurations. This function can run when AWS Config detects a configuration change to an AWS resource, or when it delivers a configuration snapshot of the resources in the account.

For more information about developing and using AWS Config rules, see Evaluating AWS Resource Configurations with AWS Config in the AWS Config Developer Guide.

ConfigRuleEvaluationStatus
Status information for your AWS managed Config rules. The status includes information such as the last time the rule ran, the last time it failed, and the related error for the last failure.

This action does not return status information about customer managed Config rules.

ConfigServiceClient
A client for the Config Service API.

ConfigSnapshotDeliveryProperties
Options for how AWS Config delivers configuration snapshots to the Amazon S3 bucket in your delivery channel.

ConfigStreamDeliveryInfo
A list that contains the status of the delivery of the configuration stream notification to the Amazon SNS topic.

ConfigurationItem

A list that contains detailed configurations of a specified resource.

Currently, the list does not contain information about non-AWS components (for example, applications on your Amazon EC2 instances).

ConfigurationRecorder
An object that represents the recording of configuration changes of an AWS resource.

ConfigurationRecorderStatus
The current status of the configuration recorder.

DeleteConfigRuleRequest

DeleteDeliveryChannelRequest
The input for the DeleteDeliveryChannel action. The action accepts the following data in JSON format.

DeliverConfigSnapshotRequest
The input for the DeliverConfigSnapshot action.

DeliverConfigSnapshotResponse
The output for the DeliverConfigSnapshot action in JSON format.

DeliveryChannel
A logical container used for storing the configuration changes of an AWS resource.

DeliveryChannelStatus
The status of a specified delivery channel.

Valid values: Success | Failure

DescribeComplianceByConfigRuleRequest

DescribeComplianceByConfigRuleResponse

DescribeComplianceByResourceRequest

DescribeComplianceByResourceResponse

DescribeConfigRuleEvaluationStatusRequest

DescribeConfigRuleEvaluationStatusResponse

DescribeConfigRulesRequest

DescribeConfigRulesResponse

DescribeConfigurationRecorderStatusRequest
The input for the DescribeConfigurationRecorderStatus action.

DescribeConfigurationRecorderStatusResponse
The output for the DescribeConfigurationRecorderStatus action in JSON format.

DescribeConfigurationRecordersRequest
The input for the DescribeConfigurationRecorders action.

DescribeConfigurationRecordersResponse
The output for the DescribeConfigurationRecorders action.

DescribeDeliveryChannelStatusRequest
The input for the DeliveryChannelStatus action.

DescribeDeliveryChannelStatusResponse
The output for the DescribeDeliveryChannelStatus action.

DescribeDeliveryChannelsRequest
The input for the DescribeDeliveryChannels action.

DescribeDeliveryChannelsResponse
The output for the DescribeDeliveryChannels action.

Evaluation
Identifies an AWS resource and indicates whether it complies with the AWS Config rule that it was evaluated against.

EvaluationResult
The details of an AWS Config evaluation. Provides the AWS resource that was evaluated, the compliance of the resource, related timestamps, and supplementary information.

EvaluationResultIdentifier
Uniquely identifies an evaluation result.

EvaluationResultQualifier
Identifies an AWS Config rule that evaluated an AWS resource, and provides the type and ID of the resource that the rule evaluated.

GetComplianceDetailsByConfigRuleRequest

GetComplianceDetailsByConfigRuleResponse

GetComplianceDetailsByResourceRequest

GetComplianceDetailsByResourceResponse

GetComplianceSummaryByConfigRuleResponse

GetComplianceSummaryByResourceTypeRequest

GetComplianceSummaryByResourceTypeResponse

GetResourceConfigHistoryRequest
The input for the GetResourceConfigHistory action.

GetResourceConfigHistoryResponse
The output for the GetResourceConfigHistory action.

ListDiscoveredResourcesRequest

ListDiscoveredResourcesResponse

PutConfigRuleRequest

PutConfigurationRecorderRequest
The input for the PutConfigurationRecorder action.

PutDeliveryChannelRequest
The input for the PutDeliveryChannel action.

PutEvaluationsRequest

PutEvaluationsResponse

RecordingGroup
Specifies the types of AWS resource for which AWS Config records configuration changes.

In the recording group, you specify whether all supported types or specific types of resources are recorded.

By default, AWS Config records configuration changes for all supported types of regional resources that AWS Config discovers in the region in which it is running. Regional resources are tied to a region and can be used only in that region. Examples of regional resources are EC2 instances and EBS volumes.

You can also have AWS Config record configuration changes for supported types of global resources (for example, IAM resources). Global resources are not tied to an individual region and can be used in all regions.
The configuration details for any global resource are the same in all regions. If you customize AWS Config in multiple regions to record global resources, it will create multiple configuration items each time a global resource changes: one configuration item for each region. These configuration items will contain identical data. To prevent duplicate configuration items, you should consider customizing AWS Config in only one region to record global resources, unless you want the configuration items to be available in multiple regions.
If you don't want AWS Config to record all resources, you can specify which types of resources it will record with the resourceTypes parameter.

For a list of supported resource types, see Supported resource types.

For more information, see Selecting Which Resources AWS Config Records.

Relationship
The relationship of the related resource to the main resource.

ResourceIdentifier
The details that identify a resource that is discovered by AWS Config, including the resource type, ID, and (if available) the custom resource name.

Scope
Defines which resources trigger an evaluation for an AWS Config rule. The scope can include one or more resource types, a combination of a tag key and value, or a combination of one resource type and one resource ID. Specify a scope to constrain which resources trigger an evaluation for a rule. Otherwise, evaluations for the rule are triggered when any resource in your recording group changes in configuration.

Source
Provides the AWS Config rule owner (AWS or customer), the rule identifier, and the events that trigger the evaluation of your AWS resources.

SourceDetail
Provides the source and type of the event that triggers AWS Config to evaluate your AWS resources against a rule.

StartConfigurationRecorderRequest
The input for the StartConfigurationRecorder action.

StopConfigurationRecorderRequest

The input for the StopConfigurationRecorder action.

Enums

DeleteConfigRuleError

DeleteDeliveryChannelError

DeliverConfigSnapshotError

DescribeComplianceByConfigRuleError

DescribeComplianceByResourceError

DescribeConfigRuleEvaluationStatusError

DescribeConfigRulesError

DescribeConfigurationRecorderStatusError

DescribeConfigurationRecordersError

DescribeDeliveryChannelStatusError

DescribeDeliveryChannelsError

GetComplianceDetailsByConfigRuleError

GetComplianceDetailsByResourceError

GetComplianceSummaryByConfigRuleError

GetComplianceSummaryByResourceTypeError

GetResourceConfigHistoryError

ListDiscoveredResourcesError

PutConfigRuleError

PutConfigurationRecorderError

PutDeliveryChannelError

PutEvaluationsError

StartConfigurationRecorderError

StopConfigurationRecorderError

Type Definitions

ARN

AccountId

AllSupported

AvailabilityZone

AwsRegion

Boolean

ChannelName

ChronologicalOrder

ComplianceByConfigRules

ComplianceByResources

ComplianceResourceTypes

ComplianceSummariesByResourceType

ComplianceType

ComplianceTypes

ConfigRuleEvaluationStatusList

ConfigRuleNames

ConfigRuleState

ConfigRules

Configuration

ConfigurationItemCaptureTime

ConfigurationItemList

ConfigurationItemMD5Hash

ConfigurationItemStatus

ConfigurationRecorderList

ConfigurationRecorderNameList

ConfigurationRecorderStatusList

ConfigurationStateId

Date

DeliveryChannelList

DeliveryChannelNameList

DeliveryChannelStatusList

DeliveryStatus

EarlierTime

EmptiableStringWithCharLimit256

EvaluationResults

Evaluations

EventSource

IncludeGlobalResourceTypes

Integer

LaterTime

Limit

MaximumExecutionFrequency

MessageType

Name

NextToken

OrderingTimestamp

Owner

RecorderName

RecorderStatus

RelatedEvent

RelatedEventList

RelationshipList

RelationshipName

ResourceCreationTime

ResourceDeletionTime

ResourceId

ResourceIdList

ResourceIdentifierList

ResourceName

ResourceType

ResourceTypeList

ResourceTypes

SourceDetails

StringWithCharLimit128

StringWithCharLimit256

StringWithCharLimit64

Tags

Value

Version