Module rusoto::devicefarm

AWS Device Farm

Structs

AccountSettings
A container for account-level settings within AWS Device Farm.

Artifact
Represents the output of a test. Examples of artifacts include logs and screenshots.

CPU
Represents the amount of CPU that an app is using on a physical device.

Note that this does not represent system-wide CPU usage.

Counters
Represents entity counters.

CreateDevicePoolRequest
Represents a request to the create device pool operation.

CreateDevicePoolResult
Represents the result of a create device pool request.

CreateProjectRequest
Represents a request to the create project operation.

CreateProjectResult
Represents the result of a create project request.

CreateUploadRequest
Represents a request to the create upload operation.

CreateUploadResult
Represents the result of a create upload request.

DeleteDevicePoolRequest
Represents a request to the delete device pool operation.

DeleteDevicePoolResult
Represents the result of a delete device pool request.

DeleteProjectRequest
Represents a request to the delete project operation.

DeleteProjectResult
Represents the result of a delete project request.

DeleteRunRequest
Represents a request to the delete run operation.

DeleteRunResult
Represents the result of a delete run request.

DeleteUploadRequest
Represents a request to the delete upload operation.

DeleteUploadResult
Represents the result of a delete upload request.

Device
Represents a device type that an app is tested against.

DeviceFarmClient
A client for the AWS Device Farm API.

DeviceMinutes
Represents the total (metered or unmetered) minutes used by the resource to run tests. Contains the sum of minutes consumed by all children.

DevicePool
Represents a collection of device types.

DevicePoolCompatibilityResult
Represents a device pool compatibility result.

GetAccountSettingsRequest
Represents the request sent to retrieve the account settings.

GetAccountSettingsResult
Represents the account settings return values from the GetAccountSettings request.

GetDevicePoolCompatibilityRequest
Represents a request to the get device pool compatibility operation.

GetDevicePoolCompatibilityResult
Represents the result of describe device pool compatibility request.

GetDevicePoolRequest
Represents a request to the get device pool operation.

GetDevicePoolResult
Represents the result of a get device pool request.

GetDeviceRequest
Represents a request to the get device request.

GetDeviceResult
Represents the result of a get device request.

GetJobRequest
Represents a request to the get job operation.

GetJobResult
Represents the result of a get job request.

GetOfferingStatusRequest
Represents the request to retrieve the offering status for the specified customer or account.

GetOfferingStatusResult
Returns the status result for a device offering.

GetProjectRequest
Represents a request to the get project operation.

GetProjectResult
Represents the result of a get project request.

GetRunRequest
Represents a request to the get run operation.

GetRunResult
Represents the result of a get run request.

GetSuiteRequest
Represents a request to the get suite operation.

GetSuiteResult
Represents the result of a get suite request.

GetTestRequest
Represents a request to the get test operation.

GetTestResult
Represents the result of a get test request.

GetUploadRequest
Represents a request to the get upload operation.

GetUploadResult
Represents the result of a get upload request.

IncompatibilityMessage
Represents information about incompatibility.

Job
Represents a device.

ListArtifactsRequest
Represents a request to the list artifacts operation.

ListArtifactsResult
Represents the result of a list artifacts operation.

ListDevicePoolsRequest
Represents the result of a list device pools request.

ListDevicePoolsResult
Represents the result of a list device pools request.

ListDevicesRequest
Represents the result of a list devices request.

ListDevicesResult
Represents the result of a list devices operation.

ListJobsRequest
Represents a request to the list jobs operation.

ListJobsResult
Represents the result of a list jobs request.

ListOfferingTransactionsRequest
Represents the request to list the offering transaction history.

ListOfferingTransactionsResult
Returns the transaction log of the specified offerings.

ListOfferingsRequest
Represents the request to list all offerings.

ListOfferingsResult
Represents the return values of the list of offerings.

ListProjectsRequest
Represents a request to the list projects operation.

ListProjectsResult
Represents the result of a list projects request.

ListRunsRequest
Represents a request to the list runs operation.

ListRunsResult
Represents the result of a list runs request.

ListSamplesRequest
Represents a request to the list samples operation.

ListSamplesResult
Represents the result of a list samples request.

ListSuitesRequest
Represents a request to the list suites operation.

ListSuitesResult
Represents the result of a list suites request.

ListTestsRequest
Represents a request to the list tests operation.

ListTestsResult
Represents the result of a list tests request.

ListUniqueProblemsRequest
Represents a request to the list unique problems operation.

ListUniqueProblemsResult
Represents the result of a list unique problems request.

ListUploadsRequest
Represents a request to the list uploads operation.

ListUploadsResult
Represents the result of a list uploads request.

Location
Represents a latitude and longitude pair, expressed in geographic coordinate system degrees (for example 47.6204, -122.3491).

Elevation is currently not supported.

MonetaryAmount
A number representing the monetary amount for an offering or transaction.

Offering
Represents the metadata of a device offering.

OfferingStatus
The status of the offering.

OfferingTransaction
Represents the metadata of an offering transaction.

Problem
Represents a specific warning or failure.

ProblemDetail
Information about a problem detail.

Project
Represents an operating-system neutral workspace for running and managing tests.

PurchaseOfferingRequest
Represents a request for a purchase offering.

PurchaseOfferingResult
The result of the purchase offering (e.g., success or failure).

Radios
Represents the set of radios and their states on a device. Examples of radios include Wi-Fi, GPS, Bluetooth, and NFC.

RecurringCharge
Specifies whether charges for devices will be recurring.

RenewOfferingRequest
A request representing an offering renewal.

RenewOfferingResult
The result of a renewal offering.

Resolution
Represents the screen resolution of a device in height and width, expressed in pixels.

Rule
Represents a condition for a device pool.

Run
Represents an app on a set of devices with a specific test and configuration.

Sample
Represents a sample of performance data.

ScheduleRunConfiguration
Represents the settings for a run. Includes things like location, radio states, auxiliary apps, and network profiles.

ScheduleRunRequest
Represents a request to the schedule run operation.

ScheduleRunResult
Represents the result of a schedule run request.

ScheduleRunTest
Represents additional test settings.

StopRunRequest
Represents the request to stop a specific run.

StopRunResult
Represents the results of your stop run attempt.

Suite
Represents a collection of one or more tests.

Test
Represents a condition that is evaluated.

UniqueProblem
A collection of one or more problems, grouped by their result.

UpdateDevicePoolRequest
Represents a request to the update device pool operation.

UpdateDevicePoolResult
Represents the result of an update device pool request.

UpdateProjectRequest
Represents a request to the update project operation.

UpdateProjectResult
Represents the result of an update project request.

Upload
An app or a set of one or more tests to upload or that have been uploaded.

Enums

CreateDevicePoolError

CreateProjectError

CreateUploadError

DeleteDevicePoolError

DeleteProjectError

DeleteRunError

DeleteUploadError

GetAccountSettingsError

GetDeviceError

GetDevicePoolCompatibilityError

GetDevicePoolError

GetJobError

GetOfferingStatusError

GetProjectError

GetRunError

GetSuiteError

GetTestError

GetUploadError

ListArtifactsError

ListDevicePoolsError

ListDevicesError

ListJobsError

ListOfferingTransactionsError

ListOfferingsError

ListProjectsError

ListRunsError

ListSamplesError

ListSuitesError

ListTestsError

ListUniqueProblemsError

ListUploadsError

PurchaseOfferingError

RenewOfferingError

ScheduleRunError

StopRunError

UpdateDevicePoolError

UpdateProjectError

Type Definitions

AWSAccountNumber

AmazonResourceName

AmazonResourceNames

ArtifactCategory

ArtifactType

Artifacts

BillingMethod

Boolean

ContentType

CurrencyCode

DateTime

DeviceAttribute

DeviceFormFactor

DevicePlatform

DevicePoolCompatibilityResults

DevicePoolType

DevicePools

Devices

Double

ExecutionResult

ExecutionStatus

Filter

IncompatibilityMessages

Integer

Jobs

Long

Message

Metadata

Name

OfferingIdentifier

OfferingStatusMap

OfferingTransactionType

OfferingTransactions

OfferingType

Offerings

PaginationToken

Problems

Projects

PurchasedDevicesMap

RecurringChargeFrequency

RecurringCharges

RuleOperator

Rules

Runs

SampleType

Samples

Suites

TestParameters

TestType

Tests

TransactionIdentifier

URL

UniqueProblems

UniqueProblemsByExecutionResultMap

UploadStatus

UploadType

Uploads