Struct rusoto::directconnect::DirectConnectClient

pub struct DirectConnectClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the AWS Direct Connect API.

Methods

impl<P> DirectConnectClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> DirectConnectClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn allocate_connection_on_interconnect(&self, input: &AllocateConnectionOnInterconnectRequest) -> Result<Connection, AllocateConnectionOnInterconnectError>

Creates a hosted connection on an interconnect.

Allocates a VLAN number and a specified amount of bandwidth for use by a hosted connection on the given interconnect.

This is intended for use by AWS Direct Connect partners only.

fn allocate_private_virtual_interface(&self, input: &AllocatePrivateVirtualInterfaceRequest) -> Result<VirtualInterface, AllocatePrivateVirtualInterfaceError>

Provisions a private virtual interface to be owned by a different customer.

The owner of a connection calls this function to provision a private virtual interface which will be owned by another AWS customer.

Virtual interfaces created using this function must be confirmed by the virtual interface owner by calling ConfirmPrivateVirtualInterface. Until this step has been completed, the virtual interface will be in 'Confirming' state, and will not be available for handling traffic.

fn allocate_public_virtual_interface(&self, input: &AllocatePublicVirtualInterfaceRequest) -> Result<VirtualInterface, AllocatePublicVirtualInterfaceError>

Provisions a public virtual interface to be owned by a different customer.

The owner of a connection calls this function to provision a public virtual interface which will be owned by another AWS customer.

Virtual interfaces created using this function must be confirmed by the virtual interface owner by calling ConfirmPublicVirtualInterface. Until this step has been completed, the virtual interface will be in 'Confirming' state, and will not be available for handling traffic.

fn confirm_connection(&self, input: &ConfirmConnectionRequest) -> Result<ConfirmConnectionResponse, ConfirmConnectionError>

Confirm the creation of a hosted connection on an interconnect.

Upon creation, the hosted connection is initially in the 'Ordering' state, and will remain in this state until the owner calls ConfirmConnection to confirm creation of the hosted connection.

fn confirm_private_virtual_interface(&self, input: &ConfirmPrivateVirtualInterfaceRequest) -> Result<ConfirmPrivateVirtualInterfaceResponse, ConfirmPrivateVirtualInterfaceError>

Accept ownership of a private virtual interface created by another customer.

After the virtual interface owner calls this function, the virtual interface will be created and attached to the given virtual private gateway, and will be available for handling traffic.

fn confirm_public_virtual_interface(&self, input: &ConfirmPublicVirtualInterfaceRequest) -> Result<ConfirmPublicVirtualInterfaceResponse, ConfirmPublicVirtualInterfaceError>

Accept ownership of a public virtual interface created by another customer.

After the virtual interface owner calls this function, the specified virtual interface will be created and made available for handling traffic.

fn create_connection(&self, input: &CreateConnectionRequest) -> Result<Connection, CreateConnectionError>

Creates a new connection between the customer network and a specific AWS Direct Connect location.

A connection links your internal network to an AWS Direct Connect location over a standard 1 gigabit or 10 gigabit Ethernet fiber-optic cable. One end of the cable is connected to your router, the other to an AWS Direct Connect router. An AWS Direct Connect location provides access to Amazon Web Services in the region it is associated with. You can establish connections with AWS Direct Connect locations in multiple regions, but a connection in one region does not provide connectivity to other regions.

fn create_interconnect(&self, input: &CreateInterconnectRequest) -> Result<Interconnect, CreateInterconnectError>

Creates a new interconnect between a AWS Direct Connect partner's network and a specific AWS Direct Connect location.

An interconnect is a connection which is capable of hosting other connections. The AWS Direct Connect partner can use an interconnect to provide sub-1Gbps AWS Direct Connect service to tier 2 customers who do not have their own connections. Like a standard connection, an interconnect links the AWS Direct Connect partner's network to an AWS Direct Connect location over a standard 1 Gbps or 10 Gbps Ethernet fiber-optic cable. One end is connected to the partner's router, the other to an AWS Direct Connect router.

For each end customer, the AWS Direct Connect partner provisions a connection on their interconnect by calling AllocateConnectionOnInterconnect. The end customer can then connect to AWS resources by creating a virtual interface on their connection, using the VLAN assigned to them by the AWS Direct Connect partner.

This is intended for use by AWS Direct Connect partners only.

fn create_private_virtual_interface(&self, input: &CreatePrivateVirtualInterfaceRequest) -> Result<VirtualInterface, CreatePrivateVirtualInterfaceError>

Creates a new private virtual interface. A virtual interface is the VLAN that transports AWS Direct Connect traffic. A private virtual interface supports sending traffic to a single virtual private cloud (VPC).

fn create_public_virtual_interface(&self, input: &CreatePublicVirtualInterfaceRequest) -> Result<VirtualInterface, CreatePublicVirtualInterfaceError>

Creates a new public virtual interface. A virtual interface is the VLAN that transports AWS Direct Connect traffic. A public virtual interface supports sending traffic to public services of AWS such as Amazon Simple Storage Service (Amazon S3).

fn delete_connection(&self, input: &DeleteConnectionRequest) -> Result<Connection, DeleteConnectionError>

Deletes the connection.

Deleting a connection only stops the AWS Direct Connect port hour and data transfer charges. You need to cancel separately with the providers any services or charges for cross-connects or network circuits that connect you to the AWS Direct Connect location.

fn delete_interconnect(&self, input: &DeleteInterconnectRequest) -> Result<DeleteInterconnectResponse, DeleteInterconnectError>

Deletes the specified interconnect.

This is intended for use by AWS Direct Connect partners only.

fn delete_virtual_interface(&self, input: &DeleteVirtualInterfaceRequest) -> Result<DeleteVirtualInterfaceResponse, DeleteVirtualInterfaceError>

Deletes a virtual interface.

fn describe_connections(&self, input: &DescribeConnectionsRequest) -> Result<Connections, DescribeConnectionsError>

Displays all connections in this region.

If a connection ID is provided, the call returns only that particular connection.

fn describe_connections_on_interconnect(&self, input: &DescribeConnectionsOnInterconnectRequest) -> Result<Connections, DescribeConnectionsOnInterconnectError>

Return a list of connections that have been provisioned on the given interconnect.

This is intended for use by AWS Direct Connect partners only.

fn describe_interconnects(&self, input: &DescribeInterconnectsRequest) -> Result<Interconnects, DescribeInterconnectsError>

Returns a list of interconnects owned by the AWS account.

If an interconnect ID is provided, it will only return this particular interconnect.

fn describe_locations(&self) -> Result<Locations, DescribeLocationsError>

Returns the list of AWS Direct Connect locations in the current AWS region. These are the locations that may be selected when calling CreateConnection or CreateInterconnect.

fn describe_virtual_gateways(&self) -> Result<VirtualGateways, DescribeVirtualGatewaysError>

Returns a list of virtual private gateways owned by the AWS account.

You can create one or more AWS Direct Connect private virtual interfaces linking to a virtual private gateway. A virtual private gateway can be managed via Amazon Virtual Private Cloud (VPC) console or the EC2 CreateVpnGateway action.

fn describe_virtual_interfaces(&self, input: &DescribeVirtualInterfacesRequest) -> Result<VirtualInterfaces, DescribeVirtualInterfacesError>

Displays all virtual interfaces for an AWS account. Virtual interfaces deleted fewer than 15 minutes before DescribeVirtualInterfaces is called are also returned. If a connection ID is included then only virtual interfaces associated with this connection will be returned. If a virtual interface ID is included then only a single virtual interface will be returned.

A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer.

If a connection ID is provided, only virtual interfaces provisioned on the specified connection will be returned. If a virtual interface ID is provided, only this particular virtual interface will be returned.