Module rusoto::ds

AWS Directory Service

Structs

Attribute
Represents a named directory attribute.

Computer
Contains information about a computer account in a directory.

ConditionalForwarder
Points to a remote domain with which you are setting up a trust relationship. Conditional forwarders are required in order to set up a trust relationship with another domain.

ConnectDirectoryRequest
Contains the inputs for the ConnectDirectory operation.

ConnectDirectoryResult
Contains the results of the ConnectDirectory operation.

CreateAliasRequest
Contains the inputs for the CreateAlias operation.

CreateAliasResult
Contains the results of the CreateAlias operation.

CreateComputerRequest
Contains the inputs for the CreateComputer operation.

CreateComputerResult
Contains the results for the CreateComputer operation.

CreateConditionalForwarderRequest
Initiates the creation of a conditional forwarder for your AWS Directory Service for Microsoft Active Directory. Conditional forwarders are required in order to set up a trust relationship with another domain.

CreateConditionalForwarderResult

CreateDirectoryRequest
Contains the inputs for the CreateDirectory operation.

CreateDirectoryResult
Contains the results of the CreateDirectory operation.

CreateMicrosoftADRequest
Creates a Microsoft AD in the AWS cloud.

CreateMicrosoftADResult

CreateSnapshotRequest
Contains the inputs for the CreateSnapshot operation.

CreateSnapshotResult
Contains the results of the CreateSnapshot operation.

CreateTrustRequest
AWS Directory Service for Microsoft Active Directory allows you to configure trust relationships. For example, you can establish a trust between your Microsoft AD in the AWS cloud, and your existing on-premises Microsoft Active Directory. This would allow you to provide users and groups access to resources in either domain, with a single set of credentials.

This action initiates the creation of the AWS side of a trust relationship between a Microsoft AD in the AWS cloud and an external domain.

CreateTrustResult

DeleteConditionalForwarderRequest

DeleteConditionalForwarderResult

DeleteDirectoryRequest
Contains the inputs for the DeleteDirectory operation.

DeleteDirectoryResult
Contains the results of the DeleteDirectory operation.

DeleteSnapshotRequest
Contains the inputs for the DeleteSnapshot operation.

DeleteSnapshotResult
Contains the results of the DeleteSnapshot operation.

DeleteTrustRequest
Deletes the local side of an existing trust relationship between the Microsoft AD in the AWS cloud and the external domain.

DeleteTrustResult

DeregisterEventTopicRequest
Removes the specified directory as a publisher to the specified SNS topic.

DeregisterEventTopicResult

DescribeConditionalForwardersRequest

DescribeConditionalForwardersResult

DescribeDirectoriesRequest
Contains the inputs for the DescribeDirectories operation.

DescribeDirectoriesResult
Contains the results of the DescribeDirectories operation.

DescribeEventTopicsRequest

DescribeEventTopicsResult

DescribeSnapshotsRequest
Contains the inputs for the DescribeSnapshots operation.

DescribeSnapshotsResult
Contains the results of the DescribeSnapshots operation.

DescribeTrustsRequest
Describes the trust relationships for a particular Microsoft AD in the AWS cloud. If no input parameters are are provided, such as directory ID or trust ID, this request describes all the trust relationships.

DescribeTrustsResult

DirectoryConnectSettings
Contains information for the ConnectDirectory operation when an AD Connector directory is being created.

DirectoryConnectSettingsDescription
Contains information about an AD Connector directory.

DirectoryDescription
Contains information about an AWS Directory Service directory.

DirectoryLimits
Contains directory limit information for a region.

DirectoryServiceClient
A client for the Directory Service API.

DirectoryVpcSettings
Contains VPC information for the CreateDirectory or CreateMicrosoftAD operation.

DirectoryVpcSettingsDescription
Contains information about the directory.

DisableRadiusRequest
Contains the inputs for the DisableRadius operation.

DisableRadiusResult
Contains the results of the DisableRadius operation.

DisableSsoRequest
Contains the inputs for the DisableSso operation.

DisableSsoResult
Contains the results of the DisableSso operation.

EnableRadiusRequest
Contains the inputs for the EnableRadius operation.

EnableRadiusResult
Contains the results of the EnableRadius operation.

EnableSsoRequest
Contains the inputs for the EnableSso operation.

EnableSsoResult
Contains the results of the EnableSso operation.

EventTopic
Information about SNS topic and AWS Directory Service directory associations.

GetDirectoryLimitsRequest
Contains the inputs for the GetDirectoryLimits operation.

GetDirectoryLimitsResult
Contains the results of the GetDirectoryLimits operation.

GetSnapshotLimitsRequest
Contains the inputs for the GetSnapshotLimits operation.

GetSnapshotLimitsResult
Contains the results of the GetSnapshotLimits operation.

RadiusSettings
Contains information about a Remote Authentication Dial In User Service (RADIUS) server.

RegisterEventTopicRequest

RegisterEventTopicResult

RestoreFromSnapshotRequest
An object representing the inputs for the RestoreFromSnapshot operation.

RestoreFromSnapshotResult
Contains the results of the RestoreFromSnapshot operation.

Snapshot
Describes a directory snapshot.

SnapshotLimits
Contains manual snapshot limit information for a directory.

Trust
Describes a trust relationship between an Microsoft AD in the AWS cloud and an external domain.

UpdateConditionalForwarderRequest

UpdateConditionalForwarderResult

UpdateRadiusRequest
Contains the inputs for the UpdateRadius operation.

UpdateRadiusResult
Contains the results of the UpdateRadius operation.

VerifyTrustRequest
Initiates the verification of an existing trust relationship between a Microsoft AD in the AWS cloud and an external domain.

VerifyTrustResult

Enums

ConnectDirectoryError

CreateAliasError

CreateComputerError

CreateConditionalForwarderError

CreateDirectoryError

CreateMicrosoftADError

CreateSnapshotError

CreateTrustError

DeleteConditionalForwarderError

DeleteDirectoryError

DeleteSnapshotError

DeleteTrustError

DeregisterEventTopicError

DescribeConditionalForwardersError

DescribeDirectoriesError

DescribeEventTopicsError

DescribeSnapshotsError

DescribeTrustsError

DisableRadiusError

DisableSsoError

EnableRadiusError

EnableSsoError

GetDirectoryLimitsError

GetSnapshotLimitsError

RegisterEventTopicError

RestoreFromSnapshotError

UpdateConditionalForwarderError

UpdateRadiusError

VerifyTrustError

Type Definitions

AccessUrl

AliasName

AttributeName

AttributeValue

Attributes

AvailabilityZone

AvailabilityZones

CloudOnlyDirectoriesLimitReached

ComputerName

ComputerPassword

ConditionalForwarders

ConnectPassword

ConnectedDirectoriesLimitReached

CreatedDateTime

DeleteAssociatedConditionalForwarder

Description

DirectoryDescriptions
A list of directory descriptions.

DirectoryId

DirectoryIds
A list of directory identifiers.

DirectoryName

DirectoryShortName

DirectorySize

DirectoryStage

DirectoryType

DnsIpAddrs

EventTopics

ExceptionMessage
The descriptive message for the exception.

IpAddr

IpAddrs

LastUpdatedDateTime

LaunchTime

Limit

ManualSnapshotsLimitReached

NextToken

OrganizationalUnitDN

Password

PortNumber

RadiusAuthenticationProtocol

RadiusDisplayLabel

RadiusRetries

RadiusSharedSecret

RadiusStatus

RadiusTimeout

RemoteDomainName

RemoteDomainNames

ReplicationScope

RequestId
The AWS request identifier.

SID

SecurityGroupId

Server

Servers

SnapshotId

SnapshotIds
A list of directory snapshot identifiers.

SnapshotName

SnapshotStatus

SnapshotType

Snapshots
A list of descriptions of directory snapshots.

SsoEnabled

StageReason

StartTime

StateLastUpdatedDateTime

SubnetId

SubnetIds

TopicArn

TopicName

TopicNames

TopicStatus

TrustDirection

TrustId

TrustIds

TrustPassword

TrustState

TrustStateReason

TrustType

Trusts

UseSameUsername

UserName

VpcId