Struct rusoto::ds::DirectoryServiceClient

pub struct DirectoryServiceClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the Directory Service API.

Methods

impl<P> DirectoryServiceClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> DirectoryServiceClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn connect_directory(&self, input: &ConnectDirectoryRequest) -> Result<ConnectDirectoryResult, ConnectDirectoryError>

Creates an AD Connector to connect to an on-premises directory.

fn create_alias(&self, input: &CreateAliasRequest) -> Result<CreateAliasResult, CreateAliasError>

Creates an alias for a directory and assigns the alias to the directory. The alias is used to construct the access URL for the directory, such as http://<![CDATA[<]]>alias<![CDATA[>]]>.awsapps.com.

After an alias has been created, it cannot be deleted or reused, so this operation should only be used when absolutely necessary.

fn create_computer(&self, input: &CreateComputerRequest) -> Result<CreateComputerResult, CreateComputerError>

Creates a computer account in the specified directory, and joins the computer to the directory.

fn create_conditional_forwarder(&self, input: &CreateConditionalForwarderRequest) -> Result<CreateConditionalForwarderResult, CreateConditionalForwarderError>

Creates a conditional forwarder associated with your AWS directory. Conditional forwarders are required in order to set up a trust relationship with another domain. The conditional forwarder points to the trusted domain.

fn create_directory(&self, input: &CreateDirectoryRequest) -> Result<CreateDirectoryResult, CreateDirectoryError>

Creates a Simple AD directory.

fn create_microsoft_a_d(&self, input: &CreateMicrosoftADRequest) -> Result<CreateMicrosoftADResult, CreateMicrosoftADError>

Creates a Microsoft AD in the AWS cloud.

fn create_snapshot(&self, input: &CreateSnapshotRequest) -> Result<CreateSnapshotResult, CreateSnapshotError>

Creates a snapshot of a Simple AD or Microsoft AD directory in the AWS cloud.

You cannot take snapshots of AD Connector directories.

fn create_trust(&self, input: &CreateTrustRequest) -> Result<CreateTrustResult, CreateTrustError>

AWS Directory Service for Microsoft Active Directory allows you to configure trust relationships. For example, you can establish a trust between your Microsoft AD in the AWS cloud, and your existing on-premises Microsoft Active Directory. This would allow you to provide users and groups access to resources in either domain, with a single set of credentials.

This action initiates the creation of the AWS side of a trust relationship between a Microsoft AD in the AWS cloud and an external domain.

fn delete_conditional_forwarder(&self, input: &DeleteConditionalForwarderRequest) -> Result<DeleteConditionalForwarderResult, DeleteConditionalForwarderError>

Deletes a conditional forwarder that has been set up for your AWS directory.

fn delete_directory(&self, input: &DeleteDirectoryRequest) -> Result<DeleteDirectoryResult, DeleteDirectoryError>

Deletes an AWS Directory Service directory.

fn delete_snapshot(&self, input: &DeleteSnapshotRequest) -> Result<DeleteSnapshotResult, DeleteSnapshotError>

Deletes a directory snapshot.

fn delete_trust(&self, input: &DeleteTrustRequest) -> Result<DeleteTrustResult, DeleteTrustError>

Deletes an existing trust relationship between your Microsoft AD in the AWS cloud and an external domain.

fn deregister_event_topic(&self, input: &DeregisterEventTopicRequest) -> Result<DeregisterEventTopicResult, DeregisterEventTopicError>

Removes the specified directory as a publisher to the specified SNS topic.

fn describe_conditional_forwarders(&self, input: &DescribeConditionalForwardersRequest) -> Result<DescribeConditionalForwardersResult, DescribeConditionalForwardersError>

Obtains information about the conditional forwarders for this account.

If no input parameters are provided for RemoteDomainNames, this request describes all conditional forwarders for the specified directory ID.

fn describe_directories(&self, input: &DescribeDirectoriesRequest) -> Result<DescribeDirectoriesResult, DescribeDirectoriesError>

Obtains information about the directories that belong to this account.

You can retrieve information about specific directories by passing the directory identifiers in the DirectoryIds parameter. Otherwise, all directories that belong to the current account are returned.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the DescribeDirectoriesResult.NextToken member contains a token that you pass in the next call to DescribeDirectories to retrieve the next set of items.

You can also specify a maximum number of return results with the Limit parameter.

fn describe_event_topics(&self, input: &DescribeEventTopicsRequest) -> Result<DescribeEventTopicsResult, DescribeEventTopicsError>

Obtains information about which SNS topics receive status messages from the specified directory.

If no input parameters are provided, such as DirectoryId or TopicName, this request describes all of the associations in the account.

fn describe_snapshots(&self, input: &DescribeSnapshotsRequest) -> Result<DescribeSnapshotsResult, DescribeSnapshotsError>

Obtains information about the directory snapshots that belong to this account.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the DescribeSnapshots.NextToken member contains a token that you pass in the next call to DescribeSnapshots to retrieve the next set of items.

You can also specify a maximum number of return results with the Limit parameter.

fn describe_trusts(&self, input: &DescribeTrustsRequest) -> Result<DescribeTrustsResult, DescribeTrustsError>

Obtains information about the trust relationships for this account.

If no input parameters are provided, such as DirectoryId or TrustIds, this request describes all the trust relationships belonging to the account.

fn disable_radius(&self, input: &DisableRadiusRequest) -> Result<DisableRadiusResult, DisableRadiusError>

Disables multi-factor authentication (MFA) with the Remote Authentication Dial In User Service (RADIUS) server for an AD Connector directory.

fn disable_sso(&self, input: &DisableSsoRequest) -> Result<DisableSsoResult, DisableSsoError>

Disables single-sign on for a directory.

fn enable_radius(&self, input: &EnableRadiusRequest) -> Result<EnableRadiusResult, EnableRadiusError>

Enables multi-factor authentication (MFA) with the Remote Authentication Dial In User Service (RADIUS) server for an AD Connector directory.

fn enable_sso(&self, input: &EnableSsoRequest) -> Result<EnableSsoResult, EnableSsoError>

Enables single-sign on for a directory.

fn get_directory_limits(&self, input: &GetDirectoryLimitsRequest) -> Result<GetDirectoryLimitsResult, GetDirectoryLimitsError>

Obtains directory limit information for the current region.

fn get_snapshot_limits(&self, input: &GetSnapshotLimitsRequest) -> Result<GetSnapshotLimitsResult, GetSnapshotLimitsError>

Obtains the manual snapshot limits for a directory.

fn register_event_topic(&self, input: &RegisterEventTopicRequest) -> Result<RegisterEventTopicResult, RegisterEventTopicError>

Associates a directory with an SNS topic. This establishes the directory as a publisher to the specified SNS topic. You can then receive email or text (SMS) messages when the status of your directory changes. You get notified if your directory goes from an Active status to an Impaired or Inoperable status. You also receive a notification when the directory returns to an Active status.

fn restore_from_snapshot(&self, input: &RestoreFromSnapshotRequest) -> Result<RestoreFromSnapshotResult, RestoreFromSnapshotError>

Restores a directory using an existing directory snapshot.

When you restore a directory from a snapshot, any changes made to the directory after the snapshot date are overwritten.

This action returns as soon as the restore operation is initiated. You can monitor the progress of the restore operation by calling the DescribeDirectories operation with the directory identifier. When the DirectoryDescription.Stage value changes to Active, the restore operation is complete.

fn update_conditional_forwarder(&self, input: &UpdateConditionalForwarderRequest) -> Result<UpdateConditionalForwarderResult, UpdateConditionalForwarderError>

Updates a conditional forwarder that has been set up for your AWS directory.

fn update_radius(&self, input: &UpdateRadiusRequest) -> Result<UpdateRadiusResult, UpdateRadiusError>

Updates the Remote Authentication Dial In User Service (RADIUS) server information for an AD Connector directory.

fn verify_trust(&self, input: &VerifyTrustRequest) -> Result<VerifyTrustResult, VerifyTrustError>

AWS Directory Service for Microsoft Active Directory allows you to configure and verify trust relationships.

This action verifies a trust relationship between your Microsoft AD in the AWS cloud and an external domain.