Struct rusoto::ds::RadiusSettings

pub struct RadiusSettings {
    pub authentication_protocol: Option<RadiusAuthenticationProtocol>,
    pub display_label: Option<RadiusDisplayLabel>,
    pub radius_port: Option<PortNumber>,
    pub radius_retries: Option<RadiusRetries>,
    pub radius_servers: Option<Servers>,
    pub radius_timeout: Option<RadiusTimeout>,
    pub shared_secret: Option<RadiusSharedSecret>,
    pub use_same_username: Option<UseSameUsername>,
}

Contains information about a Remote Authentication Dial In User Service (RADIUS) server.

Fields
authentication_protocol: Option<RadiusAuthenticationProtocol>

The protocol specified for your RADIUS endpoints.

display_label: Option<RadiusDisplayLabel>

Not currently used.

radius_port: Option<PortNumber>

The port that your RADIUS server is using for communications. Your on-premises network must allow inbound traffic over this port from the AWS Directory Service servers.

radius_retries: Option<RadiusRetries>

The maximum number of times that communication with the RADIUS server is attempted.

radius_servers: Option<Servers>

An array of strings that contains the IP addresses of the RADIUS server endpoints, or the IP addresses of your RADIUS server load balancer.

radius_timeout: Option<RadiusTimeout>

The amount of time, in seconds, to wait for the RADIUS server to respond.

shared_secret: Option<RadiusSharedSecret>

The shared secret code that was specified when your RADIUS endpoints were created.

use_same_username: Option<UseSameUsername>

Not currently used.