Struct rusoto::dynamodb::AttributeValue

pub struct AttributeValue {
    pub B: Option<BinaryAttributeValue>,
    pub NULL: Option<NullAttributeValue>,
    pub SS: Option<StringSetAttributeValue>,
    pub M: Option<MapAttributeValue>,
    pub L: Option<ListAttributeValue>,
    pub N: Option<NumberAttributeValue>,
    pub S: Option<StringAttributeValue>,
    pub BOOL: Option<BooleanAttributeValue>,
    pub BS: Option<BinarySetAttributeValue>,
    pub NS: Option<NumberSetAttributeValue>,
}

Represents the data for an attribute. You can set one, and only one, of the elements. Each attribute in an item is a name-value pair. An attribute can be single- valued or multi-valued set. For example, a book item can have title and authors attributes. Each book has one title but can have many authors. The multi-valued attribute is a set; duplicate values are not allowed.

Fields

B

A Binary data type.

NULL

A Null data type.

SS

A String Set data type.

M

A Map of attribute values.

L

A List of attribute values.

N

A Number data type.

S

A String data type.

BOOL

A Boolean data type.

BS

A Binary Set data type.

NS

A Number Set data type.