Struct rusoto::dynamodb::AttributeValue

pub struct AttributeValue {
    pub b: Option<BinaryAttributeValue>,
    pub b_o_o_l: Option<BooleanAttributeValue>,
    pub b_s: Option<BinarySetAttributeValue>,
    pub l: Option<ListAttributeValue>,
    pub m: Option<MapAttributeValue>,
    pub n: Option<NumberAttributeValue>,
    pub n_s: Option<NumberSetAttributeValue>,
    pub n_u_l_l: Option<NullAttributeValue>,
    pub s: Option<StringAttributeValue>,
    pub s_s: Option<StringSetAttributeValue>,
}

Represents the data for an attribute. You can set one, and only one, of the elements.

Each attribute in an item is a name-value pair. An attribute can be single-valued or multi-valued set. For example, a book item can have title and authors attributes. Each book has one title but can have many authors. The multi-valued attribute is a set; duplicate values are not allowed.

Fields
b: Option<BinaryAttributeValue>

A Binary data type.

b_o_o_l: Option<BooleanAttributeValue>

A Boolean data type.

b_s: Option<BinarySetAttributeValue>

A Binary Set data type.

l: Option<ListAttributeValue>

A List of attribute values.

m: Option<MapAttributeValue>

A Map of attribute values.

n: Option<NumberAttributeValue>

A Number data type.

n_s: Option<NumberSetAttributeValue>

A Number Set data type.

n_u_l_l: Option<NullAttributeValue>

A Null data type.

s: Option<StringAttributeValue>

A String data type.

s_s: Option<StringSetAttributeValue>

A String Set data type.