Struct rusoto::dynamodb::CreateTableInput

pub struct CreateTableInput {
    pub GlobalSecondaryIndexes: Option<GlobalSecondaryIndexList>,
    pub AttributeDefinitions: AttributeDefinitions,
    pub LocalSecondaryIndexes: Option<LocalSecondaryIndexList>,
    pub ProvisionedThroughput: ProvisionedThroughput,
    pub TableName: TableName,
    pub StreamSpecification: Option<StreamSpecification>,
    pub KeySchema: KeySchema,
}

Represents the input of a CreateTable operation.

Fields

GlobalSecondaryIndexes

One or more global secondary indexes (the maximum is five) to be created on the table. Each global secondary index in the array includes the following: * IndexName - The name of the global secondary index. Must be unique only for this table. * KeySchema - Specifies the key schema for the global secondary index. * Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of: * ProjectionType - One of the following: * KEYS_ONLY - Only the index and primary keys are projected into the index. * INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes are in NonKeyAttributes. * ALL - All of the table attributes are projected into the index. * NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 20. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total. * ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units.

AttributeDefinitions

An array of attributes that describe the key schema for the table and indexes.

LocalSecondaryIndexes

One or more local secondary indexes (the maximum is five) to be created on the table. Each index is scoped to a given hash key value. There is a 10 GB size limit per hash key; otherwise, the size of a local secondary index is unconstrained. Each local secondary index in the array includes the following: * IndexName - The name of the local secondary index. Must be unique only for this table. * KeySchema - Specifies the key schema for the local secondary index. The key schema must begin with the same hash key attribute as the table. * Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of: * ProjectionType - One of the following: * KEYS_ONLY - Only the index and primary keys are projected into the index. * INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes are in NonKeyAttributes. * ALL - All of the table attributes are projected into the index. * NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 20. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

ProvisionedThroughput

TableName

The name of the table to create.

StreamSpecification

The settings for DynamoDB Streams on the table. These settings consist of: * StreamEnabled - Indicates whether Streams is to be enabled (true) or disabled (false). * StreamViewType - When an item in the table is modified, StreamViewType determines what information is written to the table's stream. Valid values for StreamViewType are: * _KEYS_ONLY_ - Only the key attributes of the modified item are written to the stream. * _NEW_IMAGE_ - The entire item, as it appears after it was modified, is written to the stream. * _OLD_IMAGE_ - The entire item, as it appeared before it was modified, is written to the stream. * _NEW_AND_OLD_IMAGES_ - Both the new and the old item images of the item are written to the stream.

KeySchema

Specifies the attributes that make up the primary key for a table or an index. The attributes in KeySchema must also be defined in the AttributeDefinitions array. For more information, see Data Model in the Amazon DynamoDB Developer Guide. Each KeySchemaElement in the array is composed of: * AttributeName - The name of this key attribute. * KeyType - Determines whether the key attribute is HASH or RANGE. For a primary key that consists of a hash attribute, you must provide exactly one element with a KeyType of HASH. For a primary key that consists of hash and range attributes, you must provide exactly two elements, in this order: The first element must have a KeyType of HASH, and the second element must have a KeyType of RANGE. For more information, see Specifying the Primary Key in the Amazon DynamoDB Developer Guide.

Methods

impl CreateTableInput

fn new() -> CreateTableInput

fn with_name(self, table_name: &str) -> CreateTableInput

fn with_write_capacity(self, write_capacity: PositiveLongObject) -> CreateTableInput

fn with_read_capacity(self, read_capacity: PositiveLongObject) -> CreateTableInput

fn with_attributes(self, attributes: Vec<AttributeDefinition>) -> CreateTableInput

fn with_key_schema(self, key_schema: Vec<KeySchemaElement>) -> CreateTableInput

fn add_attribute<N: Into<String>, T: Into<String>>(self, name: N, attr_type: T) -> CreateTableInput