Struct rusoto::dynamodb::DeleteItemOutput

pub struct DeleteItemOutput {
    pub attributes: AttributeMap,
    pub item_collection_metrics: ItemCollectionMetrics,
    pub consumed_capacity: ConsumedCapacity,
}

Represents the output of a DeleteItem operation.

Fields

attributes

A map of attribute names to AttributeValue objects, representing the item as it appeared before the DeleteItem operation. This map appears in the response only if ReturnValues was specified as ALL_OLD in the request.

item_collection_metrics

Information about item collections, if any, that were affected by the operation. ItemCollectionMetrics is only returned if the request asked for it. If the table does not have any local secondary indexes, this information is not returned in the response. Each ItemCollectionMetrics element consists of: * ItemCollectionKey - The hash key value of the item collection. This is the same as the hash key of the item. * SizeEstimateRange - An estimate of item collection size, in gigabytes. This value is a two-element array containing a lower bound and an upper bound for the estimate. The estimate includes the size of all the items in the table, plus the size of all attributes projected into all of the local secondary indexes on that table. Use this estimate to measure whether a local secondary index is approaching its size limit. The estimate is subject to change over time; therefore, do not rely on the precision or accuracy of the estimate.

consumed_capacity