Struct rusoto::dynamodb::GlobalSecondaryIndexDescription

pub struct GlobalSecondaryIndexDescription {
    pub IndexSizeBytes: Option<Long>,
    pub IndexName: Option<IndexName>,
    pub Projection: Option<Projection>,
    pub ProvisionedThroughput: Option<ProvisionedThroughputDescription>,
    pub IndexStatus: Option<IndexStatus>,
    pub Backfilling: Option<Backfilling>,
    pub KeySchema: Option<KeySchema>,
    pub IndexArn: Option<String>,
    pub ItemCount: Option<Long>,
}

Represents the properties of a global secondary index.

Fields

IndexSizeBytes

The total size of the specified index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

IndexName

The name of the global secondary index.

Projection

ProvisionedThroughput

IndexStatus

The current state of the global secondary index: * CREATING - The index is being created. * UPDATING - The index is being updated. * DELETING - The index is being deleted. * ACTIVE - The index is ready for use.

Backfilling

Indicates whether the index is currently backfilling. Backfilling is the process of reading items from the table and determining whether they can be added to the index. (Not all items will qualify: For example, a hash key attribute cannot have any duplicates.) If an item can be added to the index, DynamoDB will do so. After all items have been processed, the backfilling operation is complete and Backfilling is false. For indexes that were created during a CreateTable operation, the Backfilling attribute does not appear in the DescribeTable output.

KeySchema

The complete key schema for the global secondary index, consisting of one or more pairs of attribute names and key types (HASH or RANGE).

IndexArn

The Amazon Resource Name (ARN) that uniquely identifies the index.

ItemCount

The number of items in the specified index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.