Struct rusoto::dynamodb::ProvisionedThroughputDescription

pub struct ProvisionedThroughputDescription {
    pub number_of_decreases_today: PositiveLongObject,
    pub write_capacity_units: PositiveLongObject,
    pub last_increase_date_time: Date,
    pub read_capacity_units: PositiveLongObject,
    pub last_decrease_date_time: Date,
}

Represents the provisioned throughput settings for the table, consisting of read and write capacity units, along with data about increases and decreases.

Fields

number_of_decreases_today

The number of provisioned throughput decreases for this table during this UTC calendar day. For current maximums on provisioned throughput decreases, see L imits in the Amazon DynamoDB Developer Guide.

write_capacity_units

The maximum number of writes consumed per second before DynamoDB returns a ThrottlingException.

last_increase_date_time

The date and time of the last provisioned throughput increase for this table.

read_capacity_units

The maximum number of strongly consistent reads consumed per second before DynamoDB returns a ThrottlingException. Eventually consistent reads require less effort than strongly consistent reads, so a setting of 50 ReadCapacityUnits per second provides 100 eventually consistent ReadCapacityUnits per second.

last_decrease_date_time

The date and time of the last provisioned throughput decrease for this table.