Struct rusoto::dynamodb::StreamSpecification

pub struct StreamSpecification {
    pub StreamViewType: Option<StreamViewType>,
    pub StreamEnabled: Option<StreamEnabled>,
}

Represents the DynamoDB Streams configuration for a table in DynamoDB.

Fields

StreamViewType

The DynamoDB Streams settings for the table. These settings consist of: * StreamEnabled - Indicates whether DynamoDB Streams is enabled (true) or disabled (false) on the table. * StreamViewType - When an item in the table is modified, StreamViewType determines what information is written to the stream for this table. Valid values for StreamViewType are: * _KEYS_ONLY_ - Only the key attributes of the modified item are written to the stream. * _NEW_IMAGE_ - The entire item, as it appears after it was modified, is written to the stream. * _OLD_IMAGE_ - The entire item, as it appeared before it was modified, is written to the stream. * _NEW_AND_OLD_IMAGES_ - Both the new and the old item images of the item are written to the stream.

StreamEnabled

Indicates whether DynamoDB Streams is enabled (true) or disabled (false) on the table.