Struct rusoto::dynamodb::UpdateTableInput

pub struct UpdateTableInput {
    pub global_secondary_index_updates: Option<GlobalSecondaryIndexUpdateList>,
    pub provisioned_throughput: Option<ProvisionedThroughput>,
    pub stream_specification: Option<StreamSpecification>,
    pub table_name: TableName,
    pub attribute_definitions: Option<AttributeDefinitions>,
}

Represents the input of an UpdateTable operation.

Fields

global_secondary_index_updates

An array of one or more global secondary indexes for the table. For each index in the array, you can request one action: * Create - add a new global secondary index to the table. * Update - modify the provisioned throughput settings of an existing global secondary index. * Delete - remove a global secondary index from the table. For more information, see Managing Global Secondary Indexes in the Amazon DynamoDB Developer Guide.

provisioned_throughput

stream_specification

Represents the DynamoDB Streams configuration for the table. You will receive a ResourceInUseException if you attempt to enable a stream on a table that already has a stream, or if you attempt to disable a stream on a table which does not have a stream.

table_name

The name of the table to be updated.

attribute_definitions

An array of attributes that describe the key schema for the table and indexes. If you are adding a new global secondary index to the table, AttributeDefinitions must include the key element(s) of the new index.