Struct rusoto::ec2::NetworkInterface

pub struct NetworkInterface {
    pub association: Option<NetworkInterfaceAssociation>,
    pub attachment: Option<NetworkInterfaceAttachment>,
    pub availability_zone: Option<String>,
    pub description: Option<String>,
    pub groups: Option<GroupIdentifierList>,
    pub interface_type: Option<NetworkInterfaceType>,
    pub mac_address: Option<String>,
    pub network_interface_id: Option<String>,
    pub owner_id: Option<String>,
    pub private_dns_name: Option<String>,
    pub private_ip_address: Option<String>,
    pub private_ip_addresses: Option<NetworkInterfacePrivateIpAddressList>,
    pub requester_id: Option<String>,
    pub requester_managed: Option<Boolean>,
    pub source_dest_check: Option<Boolean>,
    pub status: Option<NetworkInterfaceStatus>,
    pub subnet_id: Option<String>,
    pub tag_set: Option<TagList>,
    pub vpc_id: Option<String>,
}

Describes a network interface.

Fields
association: Option<NetworkInterfaceAssociation>

The association information for an Elastic IP associated with the network interface.

attachment: Option<NetworkInterfaceAttachment>

The network interface attachment.

availability_zone: Option<String>

The Availability Zone.

description: Option<String>

A description.

groups: Option<GroupIdentifierList>

Any security groups for the network interface.

interface_type: Option<NetworkInterfaceType>

The type of interface.

mac_address: Option<String>

The MAC address.

network_interface_id: Option<String>

The ID of the network interface.

owner_id: Option<String>

The AWS account ID of the owner of the network interface.

private_dns_name: Option<String>

The private DNS name.

private_ip_address: Option<String>

The IP address of the network interface within the subnet.

private_ip_addresses: Option<NetworkInterfacePrivateIpAddressList>

The private IP addresses associated with the network interface.

requester_id: Option<String>

The ID of the entity that launched the instance on your behalf (for example, AWS Management Console or Auto Scaling).

requester_managed: Option<Boolean>

Indicates whether the network interface is being managed by AWS.

source_dest_check: Option<Boolean>

Indicates whether traffic to or from the instance is validated.

status: Option<NetworkInterfaceStatus>

The status of the network interface.

subnet_id: Option<String>

The ID of the subnet.

tag_set: Option<TagList>

Any tags assigned to the network interface.

vpc_id: Option<String>

The ID of the VPC.