Struct rusoto::ecr::EcrClient

pub struct EcrClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the Amazon ECR API.

Methods

impl<P> EcrClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> EcrClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn batch_check_layer_availability(&self, input: &BatchCheckLayerAvailabilityRequest) -> Result<BatchCheckLayerAvailabilityResponse, BatchCheckLayerAvailabilityError>

Check the availability of multiple image layers in a specified registry and repository.

This operation is used by the Amazon ECR proxy, and it is not intended for general use by customers. Use the docker CLI to pull, tag, and push images.

fn batch_delete_image(&self, input: &BatchDeleteImageRequest) -> Result<BatchDeleteImageResponse, BatchDeleteImageError>

Deletes a list of specified images within a specified repository. Images are specified with either imageTag or imageDigest.

fn batch_get_image(&self, input: &BatchGetImageRequest) -> Result<BatchGetImageResponse, BatchGetImageError>

Gets detailed information for specified images within a specified repository. Images are specified with either imageTag or imageDigest.

fn complete_layer_upload(&self, input: &CompleteLayerUploadRequest) -> Result<CompleteLayerUploadResponse, CompleteLayerUploadError>

Inform Amazon ECR that the image layer upload for a specified registry, repository name, and upload ID, has completed. You can optionally provide a sha256 digest of the image layer for data validation purposes.

This operation is used by the Amazon ECR proxy, and it is not intended for general use by customers. Use the docker CLI to pull, tag, and push images.

fn create_repository(&self, input: &CreateRepositoryRequest) -> Result<CreateRepositoryResponse, CreateRepositoryError>

Creates an image repository.

fn delete_repository(&self, input: &DeleteRepositoryRequest) -> Result<DeleteRepositoryResponse, DeleteRepositoryError>

Deletes an existing image repository. If a repository contains images, you must use the force option to delete it.

fn delete_repository_policy(&self, input: &DeleteRepositoryPolicyRequest) -> Result<DeleteRepositoryPolicyResponse, DeleteRepositoryPolicyError>

Deletes the repository policy from a specified repository.

fn describe_repositories(&self, input: &DescribeRepositoriesRequest) -> Result<DescribeRepositoriesResponse, DescribeRepositoriesError>

Describes image repositories in a registry.

fn get_authorization_token(&self, input: &GetAuthorizationTokenRequest) -> Result<GetAuthorizationTokenResponse, GetAuthorizationTokenError>

Retrieves a token that is valid for a specified registry for 12 hours. This command allows you to use the docker CLI to push and pull images with Amazon ECR. If you do not specify a registry, the default registry is assumed.

The authorizationToken returned for each registry specified is a base64 encoded string that can be decoded and used in a docker login command to authenticate to a registry. The AWS CLI offers an aws ecr get-login command that simplifies the login process.

fn get_download_url_for_layer(&self, input: &GetDownloadUrlForLayerRequest) -> Result<GetDownloadUrlForLayerResponse, GetDownloadUrlForLayerError>

Retrieves the pre-signed Amazon S3 download URL corresponding to an image layer. You can only get URLs for image layers that are referenced in an image.

This operation is used by the Amazon ECR proxy, and it is not intended for general use by customers. Use the docker CLI to pull, tag, and push images.

fn get_repository_policy(&self, input: &GetRepositoryPolicyRequest) -> Result<GetRepositoryPolicyResponse, GetRepositoryPolicyError>

Retrieves the repository policy for a specified repository.

fn initiate_layer_upload(&self, input: &InitiateLayerUploadRequest) -> Result<InitiateLayerUploadResponse, InitiateLayerUploadError>

Notify Amazon ECR that you intend to upload an image layer.

This operation is used by the Amazon ECR proxy, and it is not intended for general use by customers. Use the docker CLI to pull, tag, and push images.

fn list_images(&self, input: &ListImagesRequest) -> Result<ListImagesResponse, ListImagesError>

Lists all the image IDs for a given repository.

fn put_image(&self, input: &PutImageRequest) -> Result<PutImageResponse, PutImageError>

Creates or updates the image manifest associated with an image.

This operation is used by the Amazon ECR proxy, and it is not intended for general use by customers. Use the docker CLI to pull, tag, and push images.

fn set_repository_policy(&self, input: &SetRepositoryPolicyRequest) -> Result<SetRepositoryPolicyResponse, SetRepositoryPolicyError>

Applies a repository policy on a specified repository to control access permissions.

fn upload_layer_part(&self, input: &UploadLayerPartRequest) -> Result<UploadLayerPartResponse, UploadLayerPartError>

Uploads an image layer part to Amazon ECR.

This operation is used by the Amazon ECR proxy, and it is not intended for general use by customers. Use the docker CLI to pull, tag, and push images.