Module rusoto::ecs

Amazon EC2 Container Service

Structs

Attribute
The attributes applicable to a container instance when it is registered.

Cluster
A regional grouping of one or more container instances on which you can run task requests. Each account receives a default cluster the first time you use the Amazon ECS service, but you may also create other clusters. Clusters may contain more than one instance type simultaneously.

Container
A Docker container that is part of a task.

ContainerDefinition
Container definitions are used in task definitions to describe the different containers that are launched as part of a task.

ContainerInstance
An EC2 instance that is running the Amazon ECS agent and has been registered with a cluster.

ContainerOverride
The overrides that should be sent to a container.

CreateClusterRequest

CreateClusterResponse

CreateServiceRequest

CreateServiceResponse

DeleteClusterRequest

DeleteClusterResponse

DeleteServiceRequest

DeleteServiceResponse

Deployment
The details of an Amazon ECS service deployment.

DeploymentConfiguration
Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

DeregisterContainerInstanceRequest

DeregisterContainerInstanceResponse

DeregisterTaskDefinitionRequest

DeregisterTaskDefinitionResponse

DescribeClustersRequest

DescribeClustersResponse

DescribeContainerInstancesRequest

DescribeContainerInstancesResponse

DescribeServicesRequest

DescribeServicesResponse

DescribeTaskDefinitionRequest

DescribeTaskDefinitionResponse

DescribeTasksRequest

DescribeTasksResponse

DiscoverPollEndpointRequest

DiscoverPollEndpointResponse

EcsClient
A client for the Amazon ECS API.

Failure
A failed resource.

HostEntry
Hostnames and IP address entries that are added to the /etc/hosts file of a container via the extraHosts parameter of its ContainerDefinition.

HostVolumeProperties
Details on a container instance host volume.

KeyValuePair
A key and value pair object.

ListClustersRequest

ListClustersResponse

ListContainerInstancesRequest

ListContainerInstancesResponse

ListServicesRequest

ListServicesResponse

ListTaskDefinitionFamiliesRequest

ListTaskDefinitionFamiliesResponse

ListTaskDefinitionsRequest

ListTaskDefinitionsResponse

ListTasksRequest

ListTasksResponse

LoadBalancer
Details on a load balancer that is used with a service.

LogConfiguration
Log configuration options to send to a custom log driver for the container.

MountPoint
Details on a volume mount point that is used in a container definition.

NetworkBinding
Details on the network bindings between a container and its host container instance. After a task reaches the RUNNING status, manual and automatic host and container port assignments are visible in the networkBindings section of DescribeTasks API responses.

PortMapping
Port mappings allow containers to access ports on the host container instance to send or receive traffic. Port mappings are specified as part of the container definition. After a task reaches the RUNNING status, manual and automatic host and container port assignments are visible in the networkBindings section of DescribeTasks API responses.

RegisterContainerInstanceRequest

RegisterContainerInstanceResponse

RegisterTaskDefinitionRequest

RegisterTaskDefinitionResponse

Resource
Describes the resources available for a container instance.

RunTaskRequest

RunTaskResponse

Service
Details on a service within a cluster

ServiceEvent
Details on an event associated with a service.

StartTaskRequest

StartTaskResponse

StopTaskRequest

StopTaskResponse

SubmitContainerStateChangeRequest

SubmitContainerStateChangeResponse

SubmitTaskStateChangeRequest

SubmitTaskStateChangeResponse

Task
Details on a task in a cluster.

TaskDefinition
Details of a task definition.

TaskOverride
The overrides associated with a task.

Ulimit
The ulimit settings to pass to the container.

UpdateContainerAgentRequest

UpdateContainerAgentResponse

UpdateServiceRequest

UpdateServiceResponse

VersionInfo
The Docker and Amazon ECS container agent version information about a container instance.

Volume
A data volume used in a task definition.

VolumeFrom
Details on a data volume from another container.

Enums

CreateClusterError

CreateServiceError

DeleteClusterError

DeleteServiceError

DeregisterContainerInstanceError

DeregisterTaskDefinitionError

DescribeClustersError

DescribeContainerInstancesError

DescribeServicesError

DescribeTaskDefinitionError

DescribeTasksError

DiscoverPollEndpointError

ListClustersError

ListContainerInstancesError

ListServicesError

ListTaskDefinitionFamiliesError

ListTaskDefinitionsError

ListTasksError

RegisterContainerInstanceError

RegisterTaskDefinitionError

RunTaskError

StartTaskError

StopTaskError

SubmitContainerStateChangeError

SubmitTaskStateChangeError

UpdateContainerAgentError

UpdateServiceError

Type Definitions

AgentUpdateStatus

Attributes

Boolean

BoxedBoolean

BoxedInteger

Clusters

ContainerDefinitions

ContainerInstances

ContainerOverrides

Containers

Deployments

DesiredStatus

DockerLabelsMap

Double

EnvironmentVariables

Failures

HostEntryList

Integer

LoadBalancers

LogConfigurationOptionsMap

LogDriver

Long

MountPointList

NetworkBindings

PortMappingList

RequiresAttributes

Resources

ServiceEvents

Services

SortOrder

StringList

TaskDefinitionFamilyStatus

TaskDefinitionStatus

Tasks

Timestamp

TransportProtocol

UlimitList

UlimitName

VolumeFromList

VolumeList