Struct rusoto::ecs::Cluster

pub struct Cluster {
    pub active_services_count: Option<Integer>,
    pub cluster_arn: Option<String>,
    pub cluster_name: Option<String>,
    pub pending_tasks_count: Option<Integer>,
    pub registered_container_instances_count: Option<Integer>,
    pub running_tasks_count: Option<Integer>,
    pub status: Option<String>,
}

A regional grouping of one or more container instances on which you can run task requests. Each account receives a default cluster the first time you use the Amazon ECS service, but you may also create other clusters. Clusters may contain more than one instance type simultaneously.

Fields
active_services_count: Option<Integer>

The number of services that are running on the cluster in an ACTIVE state. You can view these services with ListServices.

cluster_arn: Option<String>

The Amazon Resource Name (ARN) that identifies the cluster. The ARN contains the arn:aws:ecs namespace, followed by the region of the cluster, the AWS account ID of the cluster owner, the cluster namespace, and then the cluster name. For example, arn:aws:ecs:region:012345678910:cluster/test ..

cluster_name: Option<String>

A user-generated string that you use to identify your cluster.

pending_tasks_count: Option<Integer>

The number of tasks in the cluster that are in the PENDING state.

registered_container_instances_count: Option<Integer>

The number of container instances registered into the cluster.

running_tasks_count: Option<Integer>

The number of tasks in the cluster that are in the RUNNING state.

status: Option<String>

The status of the cluster. The valid values are ACTIVE or INACTIVE. ACTIVE indicates that you can register container instances with the cluster and the associated instances can accept tasks.