Struct rusoto::ecs::CreateServiceRequest

pub struct CreateServiceRequest {
    pub client_token: Option<String>,
    pub cluster: Option<String>,
    pub deployment_configuration: Option<DeploymentConfiguration>,
    pub desired_count: BoxedInteger,
    pub load_balancers: Option<LoadBalancers>,
    pub role: Option<String>,
    pub service_name: String,
    pub task_definition: String,
}

Fields
client_token: Option<String>

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. Up to 32 ASCII characters are allowed.

cluster: Option<String>

The short name or full Amazon Resource Name (ARN) of the cluster on which to run your service. If you do not specify a cluster, the default cluster is assumed.

deployment_configuration: Option<DeploymentConfiguration>

Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

desired_count: BoxedInteger

The number of instantiations of the specified task definition to place and keep running on your cluster.

load_balancers: Option<LoadBalancers>

A list of load balancer objects, containing the load balancer name, the container name (as it appears in a container definition), and the container port to access from the load balancer.

role: Option<String>

The name or full Amazon Resource Name (ARN) of the IAM role that allows Amazon ECS to make calls to your load balancer on your behalf. This parameter is required if you are using a load balancer with your service. If you specify the role parameter, you must also specify a load balancer object with the loadBalancers parameter.

If your specified role has a path other than /, then you must either specify the full role ARN (this is recommended) or prefix the role name with the path. For example, if a role with the name bar has a path of /foo/ then you would specify /foo/bar as the role name. For more information, see Friendly Names and Paths in the IAM User Guide.

service_name: String

The name of your service. Up to 255 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed. Service names must be unique within a cluster, but you can have similarly named services in multiple clusters within a region or across multiple regions.

task_definition: String

The family and revision (family:revision) or full Amazon Resource Name (ARN) of the task definition to run in your service. If a revision is not specified, the latest ACTIVE revision is used.