Struct rusoto::ecs::Service

pub struct Service {
    pub cluster_arn: Option<String>,
    pub created_at: Option<Timestamp>,
    pub deployment_configuration: Option<DeploymentConfiguration>,
    pub deployments: Option<Deployments>,
    pub desired_count: Option<Integer>,
    pub events: Option<ServiceEvents>,
    pub load_balancers: Option<LoadBalancers>,
    pub pending_count: Option<Integer>,
    pub role_arn: Option<String>,
    pub running_count: Option<Integer>,
    pub service_arn: Option<String>,
    pub service_name: Option<String>,
    pub status: Option<String>,
    pub task_definition: Option<String>,
}

Details on a service within a cluster

Fields
cluster_arn: Option<String>

The Amazon Resource Name (ARN) of the cluster that hosts the service.

created_at: Option<Timestamp>

The Unix time in seconds and milliseconds when the service was created.

deployment_configuration: Option<DeploymentConfiguration>

Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

deployments: Option<Deployments>

The current state of deployments for the service.

desired_count: Option<Integer>

The desired number of instantiations of the task definition to keep running on the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

events: Option<ServiceEvents>

The event stream for your service. A maximum of 100 of the latest events are displayed.

load_balancers: Option<LoadBalancers>

A list of load balancer objects, containing the load balancer name, the container name (as it appears in a container definition), and the container port to access from the load balancer.

pending_count: Option<Integer>

The number of tasks in the cluster that are in the PENDING state.

role_arn: Option<String>

The Amazon Resource Name (ARN) of the IAM role associated with the service that allows the Amazon ECS container agent to register container instances with a load balancer.

running_count: Option<Integer>

The number of tasks in the cluster that are in the RUNNING state.

service_arn: Option<String>

The Amazon Resource Name (ARN) that identifies the service. The ARN contains the arn:aws:ecs namespace, followed by the region of the service, the AWS account ID of the service owner, the service namespace, and then the service name. For example, arn:aws:ecs:region:012345678910:service/my-service .

service_name: Option<String>

The name of your service. Up to 255 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed. Service names must be unique within a cluster, but you can have similarly named services in multiple clusters within a region or across multiple regions.

status: Option<String>

The status of the service. The valid values are ACTIVE, DRAINING, or INACTIVE.

task_definition: Option<String>

The task definition to use for tasks in the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.