Module rusoto::events

Amazon CloudWatch Events

Structs

CloudWatchEventsClient
A client for the Amazon CloudWatch Events API.

DeleteRuleRequest
Container for the parameters to the DeleteRule operation.

DescribeRuleRequest
Container for the parameters to the DescribeRule operation.

DescribeRuleResponse
The result of the DescribeRule operation.

DisableRuleRequest
Container for the parameters to the DisableRule operation.

EnableRuleRequest
Container for the parameters to the EnableRule operation.

ListRuleNamesByTargetRequest
Container for the parameters to the ListRuleNamesByTarget operation.

ListRuleNamesByTargetResponse
The result of the ListRuleNamesByTarget operation.

ListRulesRequest
Container for the parameters to the ListRules operation.

ListRulesResponse
The result of the ListRules operation.

ListTargetsByRuleRequest
Container for the parameters to the ListTargetsByRule operation.

ListTargetsByRuleResponse
The result of the ListTargetsByRule operation.

PutEventsRequest
Container for the parameters to the PutEvents operation.

PutEventsRequestEntry
Contains information about the event to be used in the PutEvents action.

PutEventsResponse
The result of the PutEvents operation.

PutEventsResultEntry
A PutEventsResult contains a list of PutEventsResultEntry.

PutRuleRequest
Container for the parameters to the PutRule operation.

PutRuleResponse
The result of the PutRule operation.

PutTargetsRequest
Container for the parameters to the PutTargets operation.

PutTargetsResponse
The result of the PutTargets operation.

PutTargetsResultEntry
A PutTargetsResult contains a list of PutTargetsResultEntry.

RemoveTargetsRequest
Container for the parameters to the RemoveTargets operation.

RemoveTargetsResponse
The result of the RemoveTargets operation.

RemoveTargetsResultEntry
The ID of the target requested to be removed from the rule by Amazon CloudWatch Events.

Rule
Contains information about a rule in Amazon CloudWatch Events. A ListRulesResult contains a list of Rules.

Target

Targets are the resources that can be invoked when a rule is triggered. For example, AWS Lambda functions, Amazon Kinesis streams, and built-in targets.

Input and InputPath are mutually-exclusive and optional parameters of a target. When a rule is triggered due to a matched event, if for a target:

Neither Input nor InputPath is specified, then the entire event is passed to the target in JSON form.

InputPath is specified in the form of JSONPath (e.g. $.detail), then only the part of the event specified in the path is passed to the target (e.g. only the detail part of the event is passed).

Input is specified in the form of a valid JSON, then the matched event is overridden with this constant.

TestEventPatternRequest
Container for the parameters to the TestEventPattern operation.

TestEventPatternResponse
The result of the TestEventPattern operation.

Enums

DeleteRuleError

DescribeRuleError

DisableRuleError

EnableRuleError

ListRuleNamesByTargetError

ListRulesError

ListTargetsByRuleError

PutEventsError

PutRuleError

PutTargetsError

RemoveTargetsError

TestEventPatternError

Type Definitions

Boolean

ErrorCode

ErrorMessage

EventId

EventPattern

EventResource

EventResourceList

EventTime

Integer

LimitMax100

NextToken

PutEventsRequestEntryList

PutEventsResultEntryList

PutTargetsResultEntryList

RemoveTargetsResultEntryList

RoleArn

RuleArn

RuleDescription

RuleName

RuleNameList

RuleResponseList

RuleState

ScheduleExpression

TargetArn

TargetId

TargetIdList

TargetInput

TargetInputPath

TargetList