Struct rusoto::events::CloudWatchEventsClient

pub struct CloudWatchEventsClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the Amazon CloudWatch Events API.

Methods

impl<P> CloudWatchEventsClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> CloudWatchEventsClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn delete_rule(&self, input: &DeleteRuleRequest) -> Result<(), DeleteRuleError>

Deletes a rule. You must remove all targets from a rule using RemoveTargets before you can delete the rule.

Note: When you make a change with this action, incoming events might still continue to match to the deleted rule. Please allow a short period of time for changes to take effect.

fn describe_rule(&self, input: &DescribeRuleRequest) -> Result<DescribeRuleResponse, DescribeRuleError>

Describes the details of the specified rule.

fn disable_rule(&self, input: &DisableRuleRequest) -> Result<(), DisableRuleError>

Disables a rule. A disabled rule won't match any events, and won't self-trigger if it has a schedule expression.

Note: When you make a change with this action, incoming events might still continue to match to the disabled rule. Please allow a short period of time for changes to take effect.

fn enable_rule(&self, input: &EnableRuleRequest) -> Result<(), EnableRuleError>

Enables a rule. If the rule does not exist, the operation fails.

Note: When you make a change with this action, incoming events might not immediately start matching to a newly enabled rule. Please allow a short period of time for changes to take effect.

fn list_rule_names_by_target(&self, input: &ListRuleNamesByTargetRequest) -> Result<ListRuleNamesByTargetResponse, ListRuleNamesByTargetError>

Lists the names of the rules that the given target is put to. Using this action, you can find out which of the rules in Amazon CloudWatch Events can invoke a specific target in your account. If you have more rules in your account than the given limit, the results will be paginated. In that case, use the next token returned in the response and repeat the ListRulesByTarget action until the NextToken in the response is returned as null.

fn list_rules(&self, input: &ListRulesRequest) -> Result<ListRulesResponse, ListRulesError>

Lists the Amazon CloudWatch Events rules in your account. You can either list all the rules or you can provide a prefix to match to the rule names. If you have more rules in your account than the given limit, the results will be paginated. In that case, use the next token returned in the response and repeat the ListRules action until the NextToken in the response is returned as null.

fn list_targets_by_rule(&self, input: &ListTargetsByRuleRequest) -> Result<ListTargetsByRuleResponse, ListTargetsByRuleError>

Lists of targets assigned to the rule.

fn put_events(&self, input: &PutEventsRequest) -> Result<PutEventsResponse, PutEventsError>

Sends custom events to Amazon CloudWatch Events so that they can be matched to rules.

fn put_rule(&self, input: &PutRuleRequest) -> Result<PutRuleResponse, PutRuleError>

Creates or updates a rule. Rules are enabled by default, or based on value of the State parameter. You can disable a rule using DisableRule.

Note: When you make a change with this action, incoming events might not immediately start matching to new or updated rules. Please allow a short period of time for changes to take effect.

A rule must contain at least an EventPattern or ScheduleExpression. Rules with EventPatterns are triggered when a matching event is observed. Rules with ScheduleExpressions self-trigger based on the given schedule. A rule can have both an EventPattern and a ScheduleExpression, in which case the rule will trigger on matching events as well as on a schedule.

Note: Most services in AWS treat : or / as the same character in Amazon Resource Names (ARNs). However, CloudWatch Events uses an exact match in event patterns and rules. Be sure to use the correct ARN characters when creating event patterns so that they match the ARN syntax in the event you want to match.

fn put_targets(&self, input: &PutTargetsRequest) -> Result<PutTargetsResponse, PutTargetsError>

Adds target(s) to a rule. Updates the target(s) if they are already associated with the role. In other words, if there is already a target with the given target ID, then the target associated with that ID is updated.

Note: When you make a change with this action, when the associated rule triggers, new or updated targets might not be immediately invoked. Please allow a short period of time for changes to take effect.

fn remove_targets(&self, input: &RemoveTargetsRequest) -> Result<RemoveTargetsResponse, RemoveTargetsError>

Removes target(s) from a rule so that when the rule is triggered, those targets will no longer be invoked.

Note: When you make a change with this action, when the associated rule triggers, removed targets might still continue to be invoked. Please allow a short period of time for changes to take effect.

fn test_event_pattern(&self, input: &TestEventPatternRequest) -> Result<TestEventPatternResponse, TestEventPatternError>

Tests whether an event pattern matches the provided event.

Note: Most services in AWS treat : or / as the same character in Amazon Resource Names (ARNs). However, CloudWatch Events uses an exact match in event patterns and rules. Be sure to use the correct ARN characters when creating event patterns so that they match the ARN syntax in the event you want to match.