Module rusoto::firehose

Amazon Kinesis Firehose

Structs

BufferingHints
Describes hints for the buffering to perform before delivering data to the destination. Please note that these options are treated as hints, and therefore Firehose may choose to use different values when it is optimal.

CloudWatchLoggingOptions
Describes CloudWatch logging options for your delivery stream.

CopyCommand
Describes a COPY command for Amazon Redshift.

CreateDeliveryStreamInput
Contains the parameters for CreateDeliveryStream.

CreateDeliveryStreamOutput
Contains the output of CreateDeliveryStream.

DeleteDeliveryStreamInput
Contains the parameters for DeleteDeliveryStream.

DeleteDeliveryStreamOutput
Contains the output of DeleteDeliveryStream.

DeliveryStreamDescription
Contains information about a delivery stream.

DescribeDeliveryStreamInput
Contains the parameters for DescribeDeliveryStream.

DescribeDeliveryStreamOutput
Contains the output of DescribeDeliveryStream.

DestinationDescription
Describes the destination for a delivery stream.

ElasticsearchBufferingHints
Describes the buffering to perform before delivering data to the Amazon ES destination.

ElasticsearchDestinationConfiguration
Describes the configuration of a destination in Amazon ES.

ElasticsearchDestinationDescription
The destination description in Amazon ES.

ElasticsearchDestinationUpdate
Describes an update for a destination in Amazon ES.

ElasticsearchRetryOptions
Configures retry behavior in the event that Firehose is unable to deliver documents to Amazon ES.

EncryptionConfiguration
Describes the encryption for a destination in Amazon S3.

KMSEncryptionConfig
Describes an encryption key for a destination in Amazon S3.

KinesisFirehoseClient
A client for the Firehose API.

ListDeliveryStreamsInput
Contains the parameters for ListDeliveryStreams.

ListDeliveryStreamsOutput
Contains the output of ListDeliveryStreams.

PutRecordBatchInput
Contains the parameters for PutRecordBatch.

PutRecordBatchOutput
Contains the output of PutRecordBatch.

PutRecordBatchResponseEntry
Contains the result for an individual record from a PutRecordBatch request. If the record is successfully added to your delivery stream, it receives a record ID. If the record fails to be added to your delivery stream, the result includes an error code and an error message.

PutRecordInput
Contains the parameters for PutRecord.

PutRecordOutput
Contains the output of PutRecord.

Record
The unit of data in a delivery stream.

RedshiftDestinationConfiguration
Describes the configuration of a destination in Amazon Redshift.

RedshiftDestinationDescription
Describes a destination in Amazon Redshift.

RedshiftDestinationUpdate
Describes an update for a destination in Amazon Redshift.

RedshiftRetryOptions
Configures retry behavior in the event that Firehose is unable to deliver documents to Amazon Redshift.

S3DestinationConfiguration
Describes the configuration of a destination in Amazon S3.

S3DestinationDescription
Describes a destination in Amazon S3.

S3DestinationUpdate
Describes an update for a destination in Amazon S3.

UpdateDestinationInput
Contains the parameters for UpdateDestination.

UpdateDestinationOutput
Contains the output of UpdateDestination.

Enums

CreateDeliveryStreamError

DeleteDeliveryStreamError

DescribeDeliveryStreamError

ListDeliveryStreamsError

PutRecordBatchError

PutRecordError

UpdateDestinationError

Type Definitions

AWSKMSKeyARN

BooleanObject

BucketARN

ClusterJDBCURL

CompressionFormat

CopyOptions

Data

DataTableColumns

DataTableName

DeliveryStreamARN

DeliveryStreamName

DeliveryStreamNameList

DeliveryStreamStatus

DeliveryStreamVersionId

DescribeDeliveryStreamInputLimit

DestinationDescriptionList

DestinationId

ElasticsearchBufferingIntervalInSeconds

ElasticsearchBufferingSizeInMBs

ElasticsearchDomainARN

ElasticsearchIndexName

ElasticsearchIndexRotationPeriod

ElasticsearchRetryDurationInSeconds

ElasticsearchS3BackupMode

ElasticsearchTypeName

ErrorCode

ErrorMessage

IntervalInSeconds

ListDeliveryStreamsInputLimit

LogGroupName

LogStreamName

NoEncryptionConfig

NonNegativeIntegerObject

Password

Prefix

PutRecordBatchRequestEntryList

PutRecordBatchResponseEntryList

PutResponseRecordId

RedshiftRetryDurationInSeconds

RoleARN

SizeInMBs

Timestamp

Username