Struct rusoto::inspector::AssessmentRun

pub struct AssessmentRun {
    pub arn: Arn,
    pub assessment_template_arn: Arn,
    pub completed_at: Option<Timestamp>,
    pub created_at: Timestamp,
    pub data_collected: Bool,
    pub duration_in_seconds: AssessmentRunDuration,
    pub name: AssessmentRunName,
    pub notifications: AssessmentRunNotificationList,
    pub rules_package_arns: AssessmentRulesPackageArnList,
    pub started_at: Option<Timestamp>,
    pub state: AssessmentRunState,
    pub state_changed_at: Timestamp,
    pub state_changes: AssessmentRunStateChangeList,
    pub user_attributes_for_findings: UserAttributeList,
}

A snapshot of an Amazon Inspector assessment run that contains the findings of the assessment run .

Used as the response element in the DescribeAssessmentRuns action.

Fields
arn: Arn

The ARN of the assessment run.

assessment_template_arn: Arn

The ARN of the assessment template that is associated with the assessment run.

completed_at: Option<Timestamp>

The assessment run completion time that corresponds to the rules packages evaluation completion time or failure.

created_at: Timestamp

The time when StartAssessmentRun was called.

data_collected: Bool

A Boolean value (true or false) that specifies whether the process of collecting data from the agents is completed.

duration_in_seconds: AssessmentRunDuration

The duration of the assessment run.

name: AssessmentRunName

The auto-generated name for the assessment run.

notifications: AssessmentRunNotificationList

A list of notifications for the event subscriptions. A notification about a particular generated finding is added to this list only once.

rules_package_arns: AssessmentRulesPackageArnList

The rules packages selected for the assessment run.

started_at: Option<Timestamp>

The time when StartAssessmentRun was called.

state: AssessmentRunState

The state of the assessment run.

state_changed_at: Timestamp

The last time when the assessment run's state changed.

state_changes: AssessmentRunStateChangeList

A list of the assessment run state changes.

user_attributes_for_findings: UserAttributeList

The user-defined attributes that are assigned to every generated finding.