Struct rusoto::iot::IotClient

pub struct IotClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the AWS IoT API.

Methods

impl<P> IotClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> IotClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn accept_certificate_transfer(&self, input: &AcceptCertificateTransferRequest) -> Result<(), AcceptCertificateTransferError>

Accepts a pending certificate transfer. The default state of the certificate is INACTIVE.

To check for pending certificate transfers, call ListCertificates to enumerate your certificates.

fn attach_principal_policy(&self, input: &AttachPrincipalPolicyRequest) -> Result<(), AttachPrincipalPolicyError>

Attaches the specified policy to the specified principal (certificate or other credential).

fn attach_thing_principal(&self, input: &AttachThingPrincipalRequest) -> Result<AttachThingPrincipalResponse, AttachThingPrincipalError>

Attaches the specified principal to the specified thing.

fn cancel_certificate_transfer(&self, input: &CancelCertificateTransferRequest) -> Result<(), CancelCertificateTransferError>

Cancels a pending transfer for the specified certificate.

Note Only the transfer source account can use this operation to cancel a transfer. (Transfer destinations can use RejectCertificateTransfer instead.) After transfer, AWS IoT returns the certificate to the source account in the INACTIVE state. After the destination account has accepted the transfer, the transfer cannot be cancelled.

After a certificate transfer is cancelled, the status of the certificate changes from PENDING_TRANSFER to INACTIVE.

fn create_certificate_from_csr(&self, input: &CreateCertificateFromCsrRequest) -> Result<CreateCertificateFromCsrResponse, CreateCertificateFromCsrError>

Creates an X.509 certificate using the specified certificate signing request.

Note Reusing the same certificate signing request (CSR) results in a distinct certificate.

You can create multiple certificates in a batch by creating a directory, copying multiple .csr files into that directory, and then specifying that directory on the command line. The following commands show how to create a batch of certificates given a batch of CSRs.

Assuming a set of CSRs are located inside of the directory my-csr-directory:
>
On Linux and OS X, the command is:

$ ls my-csr-directory/ | xargs -I {} aws iot create-certificate-from-csr --certificate-signing-request file://my-csr-directory/{}

This command lists all of the CSRs in my-csr-directory and pipes each CSR file name to the aws iot create-certificate-from-csr AWS CLI command to create a certificate for the corresponding CSR.

The aws iot create-certificate-from-csr part of the command can also be run in parallel to speed up the certificate creation process:

$ ls my-csr-directory/ | xargs -P 10 -I {} aws iot create-certificate-from-csr --certificate-signing-request file://my-csr-directory/{}

On Windows PowerShell, the command to create certificates for all CSRs in my-csr-directory is:

> ls -Name my-csr-directory | %{aws iot create-certificate-from-csr --certificate-signing-request file://my-csr-directory/$_}

On a Windows command prompt, the command to create certificates for all CSRs in my-csr-directory is:

> forfiles /p my-csr-directory /c "cmd /c aws iot create-certificate-from-csr --certificate-signing-request file://@path"

fn create_keys_and_certificate(&self, input: &CreateKeysAndCertificateRequest) -> Result<CreateKeysAndCertificateResponse, CreateKeysAndCertificateError>

Creates a 2048-bit RSA key pair and issues an X.509 certificate using the issued public key.

Note This is the only time AWS IoT issues the private key for this certificate, so it is important to keep it in a secure location.

fn create_policy(&self, input: &CreatePolicyRequest) -> Result<CreatePolicyResponse, CreatePolicyError>

Creates an AWS IoT policy.

The created policy is the default version for the policy. This operation creates a policy version with a version identifier of 1 and sets 1 as the policy's default version.

fn create_policy_version(&self, input: &CreatePolicyVersionRequest) -> Result<CreatePolicyVersionResponse, CreatePolicyVersionError>

Creates a new version of the specified AWS IoT policy. To update a policy, create a new policy version. A managed policy can have up to five versions. If the policy has five versions, you must use DeletePolicyVersion to delete an existing version before you create a new one.

Optionally, you can set the new version as the policy's default version. The default version is the operative version (that is, the version that is in effect for the certificates to which the policy is attached).

fn create_thing(&self, input: &CreateThingRequest) -> Result<CreateThingResponse, CreateThingError>

Creates a thing in the Thing Registry.

fn create_topic_rule(&self, input: &CreateTopicRuleRequest) -> Result<(), CreateTopicRuleError>

Creates a rule. Creating rules is an administrator-level action. Any user who has permission to create rules will be able to access data processed by the rule.

fn delete_c_a_certificate(&self, input: &DeleteCACertificateRequest) -> Result<DeleteCACertificateResponse, DeleteCACertificateError>

Deletes a registered CA certificate.

fn delete_certificate(&self, input: &DeleteCertificateRequest) -> Result<(), DeleteCertificateError>

Deletes the specified certificate.

A certificate cannot be deleted if it has a policy attached to it or if its status is set to ACTIVE. To delete a certificate, first use the DetachPrincipalPolicy API to detach all policies. Next, use the UpdateCertificate API to set the certificate to the INACTIVE status.

fn delete_policy(&self, input: &DeletePolicyRequest) -> Result<(), DeletePolicyError>

Deletes the specified policy.

A policy cannot be deleted if it has non-default versions or it is attached to any certificate.

To delete a policy, use the DeletePolicyVersion API to delete all non-default versions of the policy; use the DetachPrincipalPolicy API to detach the policy from any certificate; and then use the DeletePolicy API to delete the policy.

When a policy is deleted using DeletePolicy, its default version is deleted with it.

fn delete_policy_version(&self, input: &DeletePolicyVersionRequest) -> Result<(), DeletePolicyVersionError>

Deletes the specified version of the specified policy. You cannot delete the default version of a policy using this API. To delete the default version of a policy, use DeletePolicy. To find out which version of a policy is marked as the default version, use ListPolicyVersions.

fn delete_registration_code(&self) -> Result<DeleteRegistrationCodeResponse, DeleteRegistrationCodeError>

Deletes a CA certificate registration code.

fn delete_thing(&self, input: &DeleteThingRequest) -> Result<DeleteThingResponse, DeleteThingError>

Deletes the specified thing from the Thing Registry.

fn delete_topic_rule(&self, input: &DeleteTopicRuleRequest) -> Result<(), DeleteTopicRuleError>

Deletes the specified rule.

fn describe_c_a_certificate(&self, input: &DescribeCACertificateRequest) -> Result<DescribeCACertificateResponse, DescribeCACertificateError>

Describes a registered CA certificate.

fn describe_certificate(&self, input: &DescribeCertificateRequest) -> Result<DescribeCertificateResponse, DescribeCertificateError>

Gets information about the specified certificate.

fn describe_endpoint(&self) -> Result<DescribeEndpointResponse, DescribeEndpointError>

Returns a unique endpoint specific to the AWS account making the call. You specify the following URI when updating state information for your thing: https://endpoint/things/thingName/shadow.

fn describe_thing(&self, input: &DescribeThingRequest) -> Result<DescribeThingResponse, DescribeThingError>

Gets information about the specified thing.

fn detach_principal_policy(&self, input: &DetachPrincipalPolicyRequest) -> Result<(), DetachPrincipalPolicyError>

Removes the specified policy from the specified certificate.

fn detach_thing_principal(&self, input: &DetachThingPrincipalRequest) -> Result<DetachThingPrincipalResponse, DetachThingPrincipalError>

Detaches the specified principal from the specified thing.

fn disable_topic_rule(&self, input: &DisableTopicRuleRequest) -> Result<(), DisableTopicRuleError>

Disables the specified rule.

fn enable_topic_rule(&self, input: &EnableTopicRuleRequest) -> Result<(), EnableTopicRuleError>

Enables the specified rule.

fn get_logging_options(&self) -> Result<GetLoggingOptionsResponse, GetLoggingOptionsError>

Gets the logging options.

fn get_policy(&self, input: &GetPolicyRequest) -> Result<GetPolicyResponse, GetPolicyError>

Gets information about the specified policy with the policy document of the default version.

fn get_policy_version(&self, input: &GetPolicyVersionRequest) -> Result<GetPolicyVersionResponse, GetPolicyVersionError>

Gets information about the specified policy version.

fn get_registration_code(&self) -> Result<GetRegistrationCodeResponse, GetRegistrationCodeError>

Gets a registration code used to register a CA certificate with AWS IoT.

fn get_topic_rule(&self, input: &GetTopicRuleRequest) -> Result<GetTopicRuleResponse, GetTopicRuleError>

Gets information about the specified rule.

fn list_c_a_certificates(&self, input: &ListCACertificatesRequest) -> Result<ListCACertificatesResponse, ListCACertificatesError>

Lists the CA certificates registered for your AWS account.

The results are paginated with a default page size of 25. You can use the returned marker to retrieve additional results.

fn list_certificates(&self, input: &ListCertificatesRequest) -> Result<ListCertificatesResponse, ListCertificatesError>

Lists the certificates registered in your AWS account.

The results are paginated with a default page size of 25. You can use the returned marker to retrieve additional results.

fn list_certificates_by_c_a(&self, input: &ListCertificatesByCARequest) -> Result<ListCertificatesByCAResponse, ListCertificatesByCAError>

List the device certificates signed by the specified CA certificate.

fn list_policies(&self, input: &ListPoliciesRequest) -> Result<ListPoliciesResponse, ListPoliciesError>

Lists your policies.

fn list_policy_versions(&self, input: &ListPolicyVersionsRequest) -> Result<ListPolicyVersionsResponse, ListPolicyVersionsError>

Lists the versions of the specified policy, and identifies the default version.

fn list_principal_policies(&self, input: &ListPrincipalPoliciesRequest) -> Result<ListPrincipalPoliciesResponse, ListPrincipalPoliciesError>

Lists the policies attached to the specified principal. If you use an Cognito identity, the ID must be in AmazonCognito Identity format.

fn list_principal_things(&self, input: &ListPrincipalThingsRequest) -> Result<ListPrincipalThingsResponse, ListPrincipalThingsError>

Lists the things associated with the specified principal.

fn list_thing_principals(&self, input: &ListThingPrincipalsRequest) -> Result<ListThingPrincipalsResponse, ListThingPrincipalsError>

Lists the principals associated with the specified thing.

fn list_things(&self, input: &ListThingsRequest) -> Result<ListThingsResponse, ListThingsError>

Lists your things. You can pass an AttributeName or AttributeValue to filter your things (for example, "ListThings where AttributeName=Color and AttributeValue=Red").

fn list_topic_rules(&self, input: &ListTopicRulesRequest) -> Result<ListTopicRulesResponse, ListTopicRulesError>

Lists the rules for the specific topic.

fn register_c_a_certificate(&self, input: &RegisterCACertificateRequest) -> Result<RegisterCACertificateResponse, RegisterCACertificateError>

Registers a CA certificate with AWS IoT. This CA certificate can then be used to sign device certificates, which can be then registered with AWS IoT. You can register up to 10 CA certificates per AWS account that have the same subject field and public key. This enables you to have up to 10 certificate authorities sign your device certificates. If you have more than one CA certificate registered, make sure you pass the CA certificate when you register your device certificates with the RegisterCertificate API.

fn register_certificate(&self, input: &RegisterCertificateRequest) -> Result<RegisterCertificateResponse, RegisterCertificateError>

Registers a device certificate with AWS IoT. If you have more than one CA certificate that has the same subject field, you must specify the CA certificate that was used to sign the device certificate being registered.

fn reject_certificate_transfer(&self, input: &RejectCertificateTransferRequest) -> Result<(), RejectCertificateTransferError>

Rejects a pending certificate transfer. After AWS IoT rejects a certificate transfer, the certificate status changes from PENDING_TRANSFER to INACTIVE.

To check for pending certificate transfers, call ListCertificates to enumerate your certificates.

This operation can only be called by the transfer destination. After it is called, the certificate will be returned to the source's account in the INACTIVE state.

fn replace_topic_rule(&self, input: &ReplaceTopicRuleRequest) -> Result<(), ReplaceTopicRuleError>

Replaces the specified rule. You must specify all parameters for the new rule. Creating rules is an administrator-level action. Any user who has permission to create rules will be able to access data processed by the rule.

fn set_default_policy_version(&self, input: &SetDefaultPolicyVersionRequest) -> Result<(), SetDefaultPolicyVersionError>

Sets the specified version of the specified policy as the policy's default (operative) version. This action affects all certificates to which the policy is attached. To list the principals the policy is attached to, use the ListPrincipalPolicy API.

fn set_logging_options(&self, input: &SetLoggingOptionsRequest) -> Result<(), SetLoggingOptionsError>

Sets the logging options.

fn transfer_certificate(&self, input: &TransferCertificateRequest) -> Result<TransferCertificateResponse, TransferCertificateError>

Transfers the specified certificate to the specified AWS account.

You can cancel the transfer until it is acknowledged by the recipient.

No notification is sent to the transfer destination's account. It is up to the caller to notify the transfer target.

The certificate being transferred must not be in the ACTIVE state. You can use the UpdateCertificate API to deactivate it.

The certificate must not have any policies attached to it. You can use the DetachPrincipalPolicy API to detach them.

fn update_c_a_certificate(&self, input: &UpdateCACertificateRequest) -> Result<(), UpdateCACertificateError>

Updates a registered CA certificate.

fn update_certificate(&self, input: &UpdateCertificateRequest) -> Result<(), UpdateCertificateError>

Updates the status of the specified certificate. This operation is idempotent.

Moving a certificate from the ACTIVE state (including REVOKED) will not disconnect currently connected devices, but these devices will be unable to reconnect.

The ACTIVE state is required to authenticate devices connecting to AWS IoT using a certificate.

fn update_thing(&self, input: &UpdateThingRequest) -> Result<UpdateThingResponse, UpdateThingError>

Updates the data for a thing.