Module rusoto::kinesis

Amazon Kinesis

Structs

AddTagsToStreamInput
Represents the input for AddTagsToStream.

CreateStreamInput
Represents the input for CreateStream.

DecreaseStreamRetentionPeriodInput
Represents the input for DecreaseStreamRetentionPeriod.

DeleteStreamInput
Represents the input for DeleteStream.

DescribeStreamInput
Represents the input for DescribeStream.

DescribeStreamOutput
Represents the output for DescribeStream.

DisableEnhancedMonitoringInput
Represents the input for DisableEnhancedMonitoring.

EnableEnhancedMonitoringInput
Represents the input for EnableEnhancedMonitoring.

EnhancedMetrics
Represents enhanced metrics types.

EnhancedMonitoringOutput
Represents the output for EnableEnhancedMonitoring and DisableEnhancedMonitoring.

GetRecordsInput
Represents the input for GetRecords.

GetRecordsOutput
Represents the output for GetRecords.

GetShardIteratorInput
Represents the input for GetShardIterator.

GetShardIteratorOutput
Represents the output for GetShardIterator.

HashKeyRange
The range of possible hash key values for the shard, which is a set of ordered contiguous positive integers.

IncreaseStreamRetentionPeriodInput
Represents the input for IncreaseStreamRetentionPeriod.

KinesisClient
A client for the Kinesis API.

ListStreamsInput
Represents the input for ListStreams.

ListStreamsOutput
Represents the output for ListStreams.

ListTagsForStreamInput
Represents the input for ListTagsForStream.

ListTagsForStreamOutput
Represents the output for ListTagsForStream.

MergeShardsInput
Represents the input for MergeShards.

PutRecordInput
Represents the input for PutRecord.

PutRecordOutput
Represents the output for PutRecord.

PutRecordsInput
A PutRecords request.

PutRecordsOutput
PutRecords results.

PutRecordsRequestEntry
Represents the output for PutRecords.

PutRecordsResultEntry
Represents the result of an individual record from a PutRecords request. A record that is successfully added to a stream includes SequenceNumber and ShardId in the result. A record that fails to be added to the stream includes ErrorCode and ErrorMessage in the result.

Record
The unit of data of the Amazon Kinesis stream, which is composed of a sequence number, a partition key, and a data blob.

RemoveTagsFromStreamInput
Represents the input for RemoveTagsFromStream.

SequenceNumberRange
The range of possible sequence numbers for the shard.

Shard
A uniquely identified group of data records in an Amazon Kinesis stream.

SplitShardInput
Represents the input for SplitShard.

StreamDescription
Represents the output for DescribeStream.

Tag
Metadata assigned to the stream, consisting of a key-value pair.

Enums

AddTagsToStreamError

CreateStreamError

DecreaseStreamRetentionPeriodError

DeleteStreamError

DescribeStreamError

DisableEnhancedMonitoringError

EnableEnhancedMonitoringError

GetRecordsError

GetShardIteratorError

IncreaseStreamRetentionPeriodError

ListStreamsError

ListTagsForStreamError

MergeShardsError

PutRecordError

PutRecordsError

RemoveTagsFromStreamError

SplitShardError

Type Definitions

ApproximateArrivalTimestamp

BooleanObject

Data

DescribeStreamInputLimit

EnhancedMonitoringList

ErrorCode

ErrorMessage

GetRecordsInputLimit

HashKey

ListStreamsInputLimit

ListTagsForStreamInputLimit

MetricsName

MetricsNameList

MillisBehindLatest

PartitionKey

PositiveIntegerObject

PutRecordsRequestEntryList

PutRecordsResultEntryList

RecordList

RetentionPeriodHours

SequenceNumber

ShardId

ShardIterator

ShardIteratorType

ShardList

StreamARN

StreamName

StreamNameList

StreamStatus

TagKey

TagKeyList

TagList

TagMap

TagValue

Timestamp