Module rusoto::kms

AWS Key Management Service

If you're using the service, you're probably looking for KmsClient.

Structs

AliasListEntry
Contains information about an alias.

CancelKeyDeletionRequest

CancelKeyDeletionResponse

CreateAliasRequest

CreateGrantRequest

CreateGrantResponse

CreateKeyRequest

CreateKeyResponse

DecryptRequest

DecryptResponse

DeleteAliasRequest

DeleteImportedKeyMaterialRequest

DescribeKeyRequest

DescribeKeyResponse

DisableKeyRequest

DisableKeyRotationRequest

EnableKeyRequest

EnableKeyRotationRequest

EncryptRequest

EncryptResponse

GenerateDataKeyRequest

GenerateDataKeyResponse

GenerateDataKeyWithoutPlaintextRequest

GenerateDataKeyWithoutPlaintextResponse

GenerateRandomRequest

GenerateRandomResponse

GetKeyPolicyRequest

GetKeyPolicyResponse

GetKeyRotationStatusRequest

GetKeyRotationStatusResponse

GetParametersForImportRequest

GetParametersForImportResponse

GrantConstraints
A structure for specifying the conditions under which the operations permitted by the grant are allowed.

You can use this structure to allow the operations permitted by the grant only when a specified encryption context is present. For more information about encryption context, see Encryption Context in the AWS Key Management Service Developer Guide.

GrantListEntry
Contains information about an entry in a list of grants.

ImportKeyMaterialRequest

ImportKeyMaterialResponse

KeyListEntry
Contains information about each entry in the key list.

KeyMetadata
Contains metadata about a customer master key (CMK).

This data type is used as a response element for the CreateKey and DescribeKey operations.

KmsClient
A client for the KMS API.

ListAliasesRequest

ListAliasesResponse

ListGrantsRequest

ListGrantsResponse

ListKeyPoliciesRequest

ListKeyPoliciesResponse

ListKeysRequest

ListKeysResponse

ListRetirableGrantsRequest

PutKeyPolicyRequest

ReEncryptRequest

ReEncryptResponse

RetireGrantRequest

RevokeGrantRequest

ScheduleKeyDeletionRequest

ScheduleKeyDeletionResponse

UpdateAliasRequest

UpdateKeyDescriptionRequest

Enums

CancelKeyDeletionError
Errors returned by CancelKeyDeletion

CreateAliasError
Errors returned by CreateAlias

CreateGrantError
Errors returned by CreateGrant

CreateKeyError
Errors returned by CreateKey

DecryptError
Errors returned by Decrypt

DeleteAliasError
Errors returned by DeleteAlias

DeleteImportedKeyMaterialError
Errors returned by DeleteImportedKeyMaterial

DescribeKeyError
Errors returned by DescribeKey

DisableKeyError
Errors returned by DisableKey

DisableKeyRotationError
Errors returned by DisableKeyRotation

EnableKeyError
Errors returned by EnableKey

EnableKeyRotationError
Errors returned by EnableKeyRotation

EncryptError
Errors returned by Encrypt

GenerateDataKeyError
Errors returned by GenerateDataKey

GenerateDataKeyWithoutPlaintextError
Errors returned by GenerateDataKeyWithoutPlaintext

GenerateRandomError
Errors returned by GenerateRandom

GetKeyPolicyError
Errors returned by GetKeyPolicy

GetKeyRotationStatusError
Errors returned by GetKeyRotationStatus

GetParametersForImportError
Errors returned by GetParametersForImport

ImportKeyMaterialError
Errors returned by ImportKeyMaterial

ListAliasesError
Errors returned by ListAliases

ListGrantsError
Errors returned by ListGrants

ListKeyPoliciesError
Errors returned by ListKeyPolicies

ListKeysError
Errors returned by ListKeys

ListRetirableGrantsError
Errors returned by ListRetirableGrants

PutKeyPolicyError
Errors returned by PutKeyPolicy

ReEncryptError
Errors returned by ReEncrypt

RetireGrantError
Errors returned by RetireGrant

RevokeGrantError
Errors returned by RevokeGrant

ScheduleKeyDeletionError
Errors returned by ScheduleKeyDeletion

UpdateAliasError
Errors returned by UpdateAlias

UpdateKeyDescriptionError
Errors returned by UpdateKeyDescription

Type Definitions

AWSAccountIdType

AlgorithmSpec

AliasList

AliasNameType

ArnType

BooleanType

CiphertextType

DataKeySpec

DateType

DescriptionType

EncryptionContextKey

EncryptionContextType

EncryptionContextValue

ErrorMessageType

ExpirationModelType

GrantIdType

GrantList

GrantNameType

GrantOperation

GrantOperationList

GrantTokenList

GrantTokenType

KeyIdType

KeyList

KeyState

KeyUsageType

LimitType

MarkerType

NumberOfBytesType

OriginType

PendingWindowInDaysType

PlaintextType

PolicyNameList

PolicyNameType

PolicyType

PrincipalIdType

WrappingKeySpec