Module rusoto::kms

AWS Key Management Service

Structs

AliasListEntry
Contains information about an alias.

CancelKeyDeletionRequest

CancelKeyDeletionResponse

CreateAliasRequest

CreateGrantRequest

CreateGrantResponse

CreateKeyRequest

CreateKeyResponse

DecryptRequest

DecryptResponse

DeleteAliasRequest

DescribeKeyRequest

DescribeKeyResponse

DisableKeyRequest

DisableKeyRotationRequest

EnableKeyRequest

EnableKeyRotationRequest

EncryptRequest

EncryptResponse

GenerateDataKeyRequest

GenerateDataKeyResponse

GenerateDataKeyWithoutPlaintextRequest

GenerateDataKeyWithoutPlaintextResponse

GenerateRandomRequest

GenerateRandomResponse

GetKeyPolicyRequest

GetKeyPolicyResponse

GetKeyRotationStatusRequest

GetKeyRotationStatusResponse

GrantConstraints
A structure for specifying the conditions under which the operations permitted by the grant are allowed.

You can use this structure to allow the operations permitted by the grant only when a specified encryption context is present. For more information about encryption context, see Encryption Context in the AWS Key Management Service Developer Guide.

GrantListEntry
Contains information about an entry in a list of grants.

KeyListEntry
Contains information about each entry in the key list.

KeyMetadata
Contains metadata about a customer master key (CMK).

This data type is used as a response element for the CreateKey and DescribeKey operations.

KmsClient
A client for the KMS API.

ListAliasesRequest

ListAliasesResponse

ListGrantsRequest

ListGrantsResponse

ListKeyPoliciesRequest

ListKeyPoliciesResponse

ListKeysRequest

ListKeysResponse

ListRetirableGrantsRequest

PutKeyPolicyRequest

ReEncryptRequest

ReEncryptResponse

RetireGrantRequest

RevokeGrantRequest

ScheduleKeyDeletionRequest

ScheduleKeyDeletionResponse

UpdateAliasRequest

UpdateKeyDescriptionRequest

Enums

CancelKeyDeletionError

CreateAliasError

CreateGrantError

CreateKeyError

DecryptError

DeleteAliasError

DescribeKeyError

DisableKeyError

DisableKeyRotationError

EnableKeyError

EnableKeyRotationError

EncryptError

GenerateDataKeyError

GenerateDataKeyWithoutPlaintextError

GenerateRandomError

GetKeyPolicyError

GetKeyRotationStatusError

ListAliasesError

ListGrantsError

ListKeyPoliciesError

ListKeysError

ListRetirableGrantsError

PutKeyPolicyError

ReEncryptError

RetireGrantError

RevokeGrantError

ScheduleKeyDeletionError

UpdateAliasError

UpdateKeyDescriptionError

Type Definitions

AWSAccountIdType

AliasList

AliasNameType

ArnType

BooleanType

CiphertextType

DataKeySpec

DateType

DescriptionType

EncryptionContextKey

EncryptionContextType

EncryptionContextValue

ErrorMessageType

GrantIdType

GrantList

GrantNameType

GrantOperation

GrantOperationList

GrantTokenList

GrantTokenType

KeyIdType

KeyList

KeyState

KeyUsageType

LimitType

MarkerType

NumberOfBytesType

PendingWindowInDaysType

PlaintextType

PolicyNameList

PolicyNameType

PolicyType

PrincipalIdType