Struct rusoto::kms::GrantListEntry

pub struct GrantListEntry {
    pub constraints: Option<GrantConstraints>,
    pub creation_date: Option<DateType>,
    pub grant_id: Option<GrantIdType>,
    pub grantee_principal: Option<PrincipalIdType>,
    pub issuing_account: Option<PrincipalIdType>,
    pub key_id: Option<KeyIdType>,
    pub name: Option<GrantNameType>,
    pub operations: Option<GrantOperationList>,
    pub retiring_principal: Option<PrincipalIdType>,
}

Contains information about an entry in a list of grants.

Fields
constraints: Option<GrantConstraints>

The conditions under which the grant's operations are allowed.

creation_date: Option<DateType>

The date and time when the grant was created.

grant_id: Option<GrantIdType>

The unique identifier for the grant.

grantee_principal: Option<PrincipalIdType>

The principal that receives the grant's permissions.

issuing_account: Option<PrincipalIdType>

The AWS account under which the grant was issued.

key_id: Option<KeyIdType>

The unique identifier for the customer master key (CMK) to which the grant applies.

name: Option<GrantNameType>

The friendly name that identifies the grant. If a name was provided in the CreateGrant request, that name is returned. Otherwise this value is null.

operations: Option<GrantOperationList>

The list of operations permitted by the grant.

retiring_principal: Option<PrincipalIdType>

The principal that can retire the grant.