Module rusoto::lambda

AWS Lambda

If you're using the service, you're probably looking for LambdaClient.

Structs

AccountLimit
Provides limits of code size and concurrency associated with the current account and region.

AccountUsage
Provides code size usage and function count associated with the current account and region.

AddPermissionRequest

AddPermissionResponse

AliasConfiguration
Provides configuration information about a Lambda function version alias.

CreateAliasRequest

CreateEventSourceMappingRequest

CreateFunctionRequest

DeadLetterConfig
The parent object that contains the target Amazon Resource Name (ARN) of an Amazon SQS queue or Amazon SNS topic.

DeleteAliasRequest

DeleteEventSourceMappingRequest

DeleteFunctionRequest

Environment
The parent object that contains your environment's configuration settings.

EnvironmentError
The parent object that contains error information associated with your configuration settings.

EnvironmentResponse
The parent object returned that contains your environment's configuration settings or any error information associated with your configuration settings.

EventSourceMappingConfiguration
Describes mapping between an Amazon Kinesis stream and a Lambda function.

FunctionCode
The code for the Lambda function.

FunctionCodeLocation
The object for the Lambda function location.

FunctionConfiguration
A complex type that describes function metadata.

GetAccountSettingsRequest

GetAccountSettingsResponse

GetAliasRequest

GetEventSourceMappingRequest

GetFunctionConfigurationRequest

GetFunctionRequest

GetFunctionResponse
This response contains the object for the Lambda function location (see FunctionCodeLocation.

GetPolicyRequest

GetPolicyResponse

InvocationRequest

InvocationResponse
Upon success, returns an empty response. Otherwise, throws an exception.

InvokeAsyncRequest

InvokeAsyncResponse
Upon success, it returns empty response. Otherwise, throws an exception.

LambdaClient
A client for the AWS Lambda API.

ListAliasesRequest

ListAliasesResponse

ListEventSourceMappingsRequest

ListEventSourceMappingsResponse
Contains a list of event sources (see EventSourceMappingConfiguration)

ListFunctionsRequest

ListFunctionsResponse
Contains a list of AWS Lambda function configurations (see FunctionConfiguration.

ListVersionsByFunctionRequest

ListVersionsByFunctionResponse

PublishVersionRequest

RemovePermissionRequest

UpdateAliasRequest

UpdateEventSourceMappingRequest

UpdateFunctionCodeRequest

UpdateFunctionConfigurationRequest

VpcConfig
If your Lambda function accesses resources in a VPC, you provide this parameter identifying the list of security group IDs and subnet IDs. These must belong to the same VPC. You must provide at least one security group and one subnet ID.

VpcConfigResponse
VPC configuration associated with your Lambda function.

Enums

AddPermissionError
Errors returned by AddPermission

CreateAliasError
Errors returned by CreateAlias

CreateEventSourceMappingError
Errors returned by CreateEventSourceMapping

CreateFunctionError
Errors returned by CreateFunction

DeleteAliasError
Errors returned by DeleteAlias

DeleteEventSourceMappingError
Errors returned by DeleteEventSourceMapping

DeleteFunctionError
Errors returned by DeleteFunction

GetAccountSettingsError
Errors returned by GetAccountSettings

GetAliasError
Errors returned by GetAlias

GetEventSourceMappingError
Errors returned by GetEventSourceMapping

GetFunctionConfigurationError
Errors returned by GetFunctionConfiguration

GetFunctionError
Errors returned by GetFunction

GetPolicyError
Errors returned by GetPolicy

InvokeAsyncError
Errors returned by InvokeAsync

InvokeError
Errors returned by Invoke

ListAliasesError
Errors returned by ListAliases

ListEventSourceMappingsError
Errors returned by ListEventSourceMappings

ListFunctionsError
Errors returned by ListFunctions

ListVersionsByFunctionError
Errors returned by ListVersionsByFunction

PublishVersionError
Errors returned by PublishVersion

RemovePermissionError
Errors returned by RemovePermission

UpdateAliasError
Errors returned by UpdateAlias

UpdateEventSourceMappingError
Errors returned by UpdateEventSourceMapping

UpdateFunctionCodeError
Errors returned by UpdateFunctionCode

UpdateFunctionConfigurationError
Errors returned by UpdateFunctionConfiguration

Type Definitions

Action

Alias

AliasList

Arn

BatchSize

Blob

BlobStream

Boolean

Date

Description

Enabled

EnvironmentVariableName

EnvironmentVariableValue

EnvironmentVariables

EventSourceMappingsList

EventSourcePosition

EventSourceToken

FunctionArn

FunctionList

FunctionName

Handler

HttpStatus

Integer

InvocationType

KMSKeyArn

LogType

Long

MaxListItems

MemorySize

Principal

Qualifier

ResourceArn

RoleArn

Runtime

S3Bucket

S3Key

S3ObjectVersion

SecurityGroupId

SecurityGroupIds

SensitiveString

SourceOwner

StatementId

SubnetId

SubnetIds

ThrottleReason

Timeout

Timestamp

Version

VpcId