Struct rusoto::logs::CloudWatchLogsClient

pub struct CloudWatchLogsClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the Amazon CloudWatch Logs API.

Methods

impl<P> CloudWatchLogsClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> CloudWatchLogsClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn cancel_export_task(&self, input: &CancelExportTaskRequest) -> Result<(), CancelExportTaskError>

Cancels an export task if it is in PENDING or RUNNING state.

fn create_export_task(&self, input: &CreateExportTaskRequest) -> Result<CreateExportTaskResponse, CreateExportTaskError>

Creates an ExportTask which allows you to efficiently export data from a Log Group to your Amazon S3 bucket.

This is an asynchronous call. If all the required information is provided, this API will initiate an export task and respond with the task Id. Once started, DescribeExportTasks can be used to get the status of an export task. You can only have one active (RUNNING or PENDING) export task at a time, per account.

You can export logs from multiple log groups or multiple time ranges to the same Amazon S3 bucket. To separate out log data for each export task, you can specify a prefix that will be used as the Amazon S3 key prefix for all exported objects.

fn create_log_group(&self, input: &CreateLogGroupRequest) -> Result<(), CreateLogGroupError>

Creates a new log group with the specified name. The name of the log group must be unique within a region for an AWS account. You can create up to 500 log groups per account.

You must use the following guidelines when naming a log group:

Log group names can be between 1 and 512 characters long.

Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (hyphen), '/' (forward slash), and '.' (period).

fn create_log_stream(&self, input: &CreateLogStreamRequest) -> Result<(), CreateLogStreamError>

Creates a new log stream in the specified log group. The name of the log stream must be unique within the log group. There is no limit on the number of log streams that can exist in a log group.

You must use the following guidelines when naming a log stream:

Log stream names can be between 1 and 512 characters long.

The ':' colon character is not allowed.

fn delete_destination(&self, input: &DeleteDestinationRequest) -> Result<(), DeleteDestinationError>

Deletes the destination with the specified name and eventually disables all the subscription filters that publish to it. This will not delete the physical resource encapsulated by the destination.

fn delete_log_group(&self, input: &DeleteLogGroupRequest) -> Result<(), DeleteLogGroupError>

Deletes the log group with the specified name and permanently deletes all the archived log events associated with it.

fn delete_log_stream(&self, input: &DeleteLogStreamRequest) -> Result<(), DeleteLogStreamError>

Deletes a log stream and permanently deletes all the archived log events associated with it.

fn delete_metric_filter(&self, input: &DeleteMetricFilterRequest) -> Result<(), DeleteMetricFilterError>

Deletes a metric filter associated with the specified log group.

fn delete_retention_policy(&self, input: &DeleteRetentionPolicyRequest) -> Result<(), DeleteRetentionPolicyError>

Deletes the retention policy of the specified log group. Log events would not expire if they belong to log groups without a retention policy.

fn delete_subscription_filter(&self, input: &DeleteSubscriptionFilterRequest) -> Result<(), DeleteSubscriptionFilterError>

Deletes a subscription filter associated with the specified log group.

fn describe_destinations(&self, input: &DescribeDestinationsRequest) -> Result<DescribeDestinationsResponse, DescribeDestinationsError>

Returns all the destinations that are associated with the AWS account making the request. The list returned in the response is ASCII-sorted by destination name.

By default, this operation returns up to 50 destinations. If there are more destinations to list, the response would contain a nextToken value in the response body. You can also limit the number of destinations returned in the response by specifying the limit parameter in the request.

fn describe_export_tasks(&self, input: &DescribeExportTasksRequest) -> Result<DescribeExportTasksResponse, DescribeExportTasksError>

Returns all the export tasks that are associated with the AWS account making the request. The export tasks can be filtered based on TaskId or TaskStatus.

By default, this operation returns up to 50 export tasks that satisfy the specified filters. If there are more export tasks to list, the response would contain a nextToken value in the response body. You can also limit the number of export tasks returned in the response by specifying the limit parameter in the request.

fn describe_log_groups(&self, input: &DescribeLogGroupsRequest) -> Result<DescribeLogGroupsResponse, DescribeLogGroupsError>

Returns all the log groups that are associated with the AWS account making the request. The list returned in the response is ASCII-sorted by log group name.

By default, this operation returns up to 50 log groups. If there are more log groups to list, the response would contain a nextToken value in the response body. You can also limit the number of log groups returned in the response by specifying the limit parameter in the request.

fn describe_log_streams(&self, input: &DescribeLogStreamsRequest) -> Result<DescribeLogStreamsResponse, DescribeLogStreamsError>

Returns all the log streams that are associated with the specified log group. The list returned in the response is ASCII-sorted by log stream name.

By default, this operation returns up to 50 log streams. If there are more log streams to list, the response would contain a nextToken value in the response body. You can also limit the number of log streams returned in the response by specifying the limit parameter in the request. This operation has a limit of five transactions per second, after which transactions are throttled.

fn describe_metric_filters(&self, input: &DescribeMetricFiltersRequest) -> Result<DescribeMetricFiltersResponse, DescribeMetricFiltersError>

Returns all the metrics filters associated with the specified log group. The list returned in the response is ASCII-sorted by filter name.

By default, this operation returns up to 50 metric filters. If there are more metric filters to list, the response would contain a nextToken value in the response body. You can also limit the number of metric filters returned in the response by specifying the limit parameter in the request.

fn describe_subscription_filters(&self, input: &DescribeSubscriptionFiltersRequest) -> Result<DescribeSubscriptionFiltersResponse, DescribeSubscriptionFiltersError>

Returns all the subscription filters associated with the specified log group. The list returned in the response is ASCII-sorted by filter name.

By default, this operation returns up to 50 subscription filters. If there are more subscription filters to list, the response would contain a nextToken value in the response body. You can also limit the number of subscription filters returned in the response by specifying the limit parameter in the request.

fn filter_log_events(&self, input: &FilterLogEventsRequest) -> Result<FilterLogEventsResponse, FilterLogEventsError>

Retrieves log events, optionally filtered by a filter pattern from the specified log group. You can provide an optional time range to filter the results on the event timestamp. You can limit the streams searched to an explicit list of logStreamNames.

By default, this operation returns as much matching log events as can fit in a response size of 1MB, up to 10,000 log events, or all the events found within a time-bounded scan window. If the response includes a nextToken, then there is more data to search, and the search can be resumed with a new request providing the nextToken. The response will contain a list of searchedLogStreams that contains information about which streams were searched in the request and whether they have been searched completely or require further pagination. The limit parameter in the request. can be used to specify the maximum number of events to return in a page.

fn get_log_events(&self, input: &GetLogEventsRequest) -> Result<GetLogEventsResponse, GetLogEventsError>

Retrieves log events from the specified log stream. You can provide an optional time range to filter the results on the event timestamp.

By default, this operation returns as much log events as can fit in a response size of 1MB, up to 10,000 log events. The response will always include a nextForwardToken and a nextBackwardToken in the response body. You can use any of these tokens in subsequent GetLogEvents requests to paginate through events in either forward or backward direction. You can also limit the number of log events returned in the response by specifying the limit parameter in the request.

fn put_destination(&self, input: &PutDestinationRequest) -> Result<PutDestinationResponse, PutDestinationError>

Creates or updates a Destination. A destination encapsulates a physical resource (such as a Kinesis stream) and allows you to subscribe to a real-time stream of log events of a different account, ingested through PutLogEvents requests. Currently, the only supported physical resource is a Amazon Kinesis stream belonging to the same account as the destination.

A destination controls what is written to its Amazon Kinesis stream through an access policy. By default, PutDestination does not set any access policy with the destination, which means a cross-account user will not be able to call PutSubscriptionFilter against this destination. To enable that, the destination owner must call PutDestinationPolicy after PutDestination.

fn put_destination_policy(&self, input: &PutDestinationPolicyRequest) -> Result<(), PutDestinationPolicyError>

Creates or updates an access policy associated with an existing Destination. An access policy is an IAM policy document that is used to authorize claims to register a subscription filter against a given destination.

fn put_log_events(&self, input: &PutLogEventsRequest) -> Result<PutLogEventsResponse, PutLogEventsError>

Uploads a batch of log events to the specified log stream.

Every PutLogEvents request must include the sequenceToken obtained from the response of the previous request. An upload in a newly created log stream does not require a sequenceToken.

The batch of events must satisfy the following constraints:

The maximum batch size is 1,048,576 bytes, and this size is calculated as the sum of all event messages in UTF-8, plus 26 bytes for each log event.

None of the log events in the batch can be more than 2 hours in the future.

None of the log events in the batch can be older than 14 days or the retention period of the log group.

The log events in the batch must be in chronological ordered by their timestamp.

The maximum number of log events in a batch is 10,000.

A batch of log events in a single PutLogEvents request cannot span more than 24 hours. Otherwise, the PutLogEvents operation will fail.

fn put_metric_filter(&self, input: &PutMetricFilterRequest) -> Result<(), PutMetricFilterError>

Creates or updates a metric filter and associates it with the specified log group. Metric filters allow you to configure rules to extract metric data from log events ingested through PutLogEvents requests.

The maximum number of metric filters that can be associated with a log group is 100.

fn put_retention_policy(&self, input: &PutRetentionPolicyRequest) -> Result<(), PutRetentionPolicyError>

Sets the retention of the specified log group. A retention policy allows you to configure the number of days you want to retain log events in the specified log group.

fn put_subscription_filter(&self, input: &PutSubscriptionFilterRequest) -> Result<(), PutSubscriptionFilterError>

Creates or updates a subscription filter and associates it with the specified log group. Subscription filters allow you to subscribe to a real-time stream of log events ingested through PutLogEvents requests and have them delivered to a specific destination. Currently, the supported destinations are:

An Amazon Kinesis stream belonging to the same account as the subscription filter, for same-account delivery.

A logical destination (used via an ARN of Destination) belonging to a different account, for cross-account delivery.

An Amazon Kinesis Firehose stream belonging to the same account as the subscription filter, for same-account delivery.

An AWS Lambda function belonging to the same account as the subscription filter, for same-account delivery.

Currently there can only be one subscription filter associated with a log group.

fn test_metric_filter(&self, input: &TestMetricFilterRequest) -> Result<TestMetricFilterResponse, TestMetricFilterError>

Tests the filter pattern of a metric filter against a sample of log event messages. You can use this operation to validate the correctness of a metric filter pattern.