Struct rusoto::logs::DescribeLogStreamsRequest

pub struct DescribeLogStreamsRequest {
    pub descending: Option<Descending>,
    pub limit: Option<DescribeLimit>,
    pub log_group_name: LogGroupName,
    pub log_stream_name_prefix: Option<LogStreamName>,
    pub next_token: Option<NextToken>,
    pub order_by: Option<OrderBy>,
}

Fields
descending: Option<Descending>

If set to true, results are returned in descending order. If you don't specify a value or set it to false, results are returned in ascending order.

limit: Option<DescribeLimit>

The maximum number of items returned in the response. If you don't specify a value, the request would return up to 50 items.

log_group_name: LogGroupName

The log group name for which log streams are to be listed.

log_stream_name_prefix: Option<LogStreamName>

Will only return log streams that match the provided logStreamNamePrefix. If you don't specify a value, no prefix filter is applied.

next_token: Option<NextToken>

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous DescribeLogStreams request.

order_by: Option<OrderBy>

Specifies what to order the returned log streams by. Valid arguments are 'LogStreamName' or 'LastEventTime'. If you don't specify a value, results are ordered by LogStreamName. If 'LastEventTime' is chosen, the request cannot also contain a logStreamNamePrefix.