Struct rusoto::machinelearning::MachineLearningClient

pub struct MachineLearningClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the Amazon Machine Learning API.

Methods

impl<P> MachineLearningClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> MachineLearningClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn create_batch_prediction(&self, input: &CreateBatchPredictionInput) -> Result<CreateBatchPredictionOutput, CreateBatchPredictionError>

Generates predictions for a group of observations. The observations to process exist in one or more data files referenced by a DataSource. This operation creates a new BatchPrediction, and uses an MLModel and the data files referenced by the DataSource as information sources.

CreateBatchPrediction is an asynchronous operation. In response to CreateBatchPrediction, Amazon Machine Learning (Amazon ML) immediately returns and sets the BatchPrediction status to PENDING. After the BatchPrediction completes, Amazon ML sets the status to COMPLETED.

You can poll for status updates by using the GetBatchPrediction operation and checking the Status parameter of the result. After the COMPLETED status appears, the results are available in the location specified by the OutputUri parameter.

fn create_data_source_from_r_d_s(&self, input: &CreateDataSourceFromRDSInput) -> Result<CreateDataSourceFromRDSOutput, CreateDataSourceFromRDSError>

Creates a DataSource object from an Amazon Relational Database Service (Amazon RDS). A DataSource references data that can be used to perform CreateMLModel, CreateEvaluation, or CreateBatchPrediction operations.

CreateDataSourceFromRDS is an asynchronous operation. In response to CreateDataSourceFromRDS, Amazon Machine Learning (Amazon ML) immediately returns and sets the DataSource status to PENDING. After the DataSource is created and ready for use, Amazon ML sets the Status parameter to COMPLETED. DataSource in COMPLETED or PENDING status can only be used to perform CreateMLModel, CreateEvaluation, or CreateBatchPrediction operations.

If Amazon ML cannot accept the input source, it sets the Status parameter to FAILED and includes an error message in the Message attribute of the GetDataSource operation response.

fn create_data_source_from_redshift(&self, input: &CreateDataSourceFromRedshiftInput) -> Result<CreateDataSourceFromRedshiftOutput, CreateDataSourceFromRedshiftError>

Creates a DataSource from Amazon Redshift. A DataSource references data that can be used to perform either CreateMLModel, CreateEvaluation or CreateBatchPrediction operations.

CreateDataSourceFromRedshift is an asynchronous operation. In response to CreateDataSourceFromRedshift, Amazon Machine Learning (Amazon ML) immediately returns and sets the DataSource status to PENDING. After the DataSource is created and ready for use, Amazon ML sets the Status parameter to COMPLETED. DataSource in COMPLETED or PENDING status can only be used to perform CreateMLModel, CreateEvaluation, or CreateBatchPrediction operations.

If Amazon ML cannot accept the input source, it sets the Status parameter to FAILED and includes an error message in the Message attribute of the GetDataSource operation response.

The observations should exist in the database hosted on an Amazon Redshift cluster and should be specified by a SelectSqlQuery. Amazon ML executes Unload command in Amazon Redshift to transfer the result set of SelectSqlQuery to S3StagingLocation.

After the DataSource is created, it's ready for use in evaluations and batch predictions. If you plan to use the DataSource to train an MLModel, the DataSource requires another item -- a recipe. A recipe describes the observation variables that participate in training an MLModel. A recipe describes how each input variable will be used in training. Will the variable be included or excluded from training? Will the variable be manipulated, for example, combined with another variable or split apart into word combinations? The recipe provides answers to these questions. For more information, see the Amazon Machine Learning Developer Guide.

fn create_data_source_from_s_3(&self, input: &CreateDataSourceFromS3Input) -> Result<CreateDataSourceFromS3Output, CreateDataSourceFromS3Error>

Creates a DataSource object. A DataSource references data that can be used to perform CreateMLModel, CreateEvaluation, or CreateBatchPrediction operations.

CreateDataSourceFromS3 is an asynchronous operation. In response to CreateDataSourceFromS3, Amazon Machine Learning (Amazon ML) immediately returns and sets the DataSource status to PENDING. After the DataSource is created and ready for use, Amazon ML sets the Status parameter to COMPLETED. DataSource in COMPLETED or PENDING status can only be used to perform CreateMLModel, CreateEvaluation or CreateBatchPrediction operations.

If Amazon ML cannot accept the input source, it sets the Status parameter to FAILED and includes an error message in the Message attribute of the GetDataSource operation response.

The observation data used in a DataSource should be ready to use; that is, it should have a consistent structure, and missing data values should be kept to a minimum. The observation data must reside in one or more CSV files in an Amazon Simple Storage Service (Amazon S3) bucket, along with a schema that describes the data items by name and type. The same schema must be used for all of the data files referenced by the DataSource.

After the DataSource has been created, it's ready to use in evaluations and batch predictions. If you plan to use the DataSource to train an MLModel, the DataSource requires another item: a recipe. A recipe describes the observation variables that participate in training an MLModel. A recipe describes how each input variable will be used in training. Will the variable be included or excluded from training? Will the variable be manipulated, for example, combined with another variable, or split apart into word combinations? The recipe provides answers to these questions. For more information, see the Amazon Machine Learning Developer Guide.

fn create_evaluation(&self, input: &CreateEvaluationInput) -> Result<CreateEvaluationOutput, CreateEvaluationError>

Creates a new Evaluation of an MLModel. An MLModel is evaluated on a set of observations associated to a DataSource. Like a DataSource for an MLModel, the DataSource for an Evaluation contains values for the Target Variable. The Evaluation compares the predicted result for each observation to the actual outcome and provides a summary so that you know how effective the MLModel functions on the test data. Evaluation generates a relevant performance metric such as BinaryAUC, RegressionRMSE or MulticlassAvgFScore based on the corresponding MLModelType: BINARY, REGRESSION or MULTICLASS.

CreateEvaluation is an asynchronous operation. In response to CreateEvaluation, Amazon Machine Learning (Amazon ML) immediately returns and sets the evaluation status to PENDING. After the Evaluation is created and ready for use, Amazon ML sets the status to COMPLETED.

You can use the GetEvaluation operation to check progress of the evaluation during the creation operation.

fn create_m_l_model(&self, input: &CreateMLModelInput) -> Result<CreateMLModelOutput, CreateMLModelError>

Creates a new MLModel using the data files and the recipe as information sources.

An MLModel is nearly immutable. Users can only update the MLModelName and the ScoreThreshold in an MLModel without creating a new MLModel.

CreateMLModel is an asynchronous operation. In response to CreateMLModel, Amazon Machine Learning (Amazon ML) immediately returns and sets the MLModel status to PENDING. After the MLModel is created and ready for use, Amazon ML sets the status to COMPLETED.

You can use the GetMLModel operation to check progress of the MLModel during the creation operation.

CreateMLModel requires a DataSource with computed statistics, which can be created by setting ComputeStatistics to true in CreateDataSourceFromRDS, CreateDataSourceFromS3, or CreateDataSourceFromRedshift operations.

fn create_realtime_endpoint(&self, input: &CreateRealtimeEndpointInput) -> Result<CreateRealtimeEndpointOutput, CreateRealtimeEndpointError>

Creates a real-time endpoint for the MLModel. The endpoint contains the URI of the MLModel; that is, the location to send real-time prediction requests for the specified MLModel.

fn delete_batch_prediction(&self, input: &DeleteBatchPredictionInput) -> Result<DeleteBatchPredictionOutput, DeleteBatchPredictionError>

Assigns the DELETED status to a BatchPrediction, rendering it unusable.

After using the DeleteBatchPrediction operation, you can use the GetBatchPrediction operation to verify that the status of the BatchPrediction changed to DELETED.

Caution: The result of the DeleteBatchPrediction operation is irreversible.

fn delete_data_source(&self, input: &DeleteDataSourceInput) -> Result<DeleteDataSourceOutput, DeleteDataSourceError>

Assigns the DELETED status to a DataSource, rendering it unusable.

After using the DeleteDataSource operation, you can use the GetDataSource operation to verify that the status of the DataSource changed to DELETED.

Caution: The results of the DeleteDataSource operation are irreversible.

fn delete_evaluation(&self, input: &DeleteEvaluationInput) -> Result<DeleteEvaluationOutput, DeleteEvaluationError>

Assigns the DELETED status to an Evaluation, rendering it unusable.

After invoking the DeleteEvaluation operation, you can use the GetEvaluation operation to verify that the status of the Evaluation changed to DELETED.

Caution: The results of the DeleteEvaluation operation are irreversible.

fn delete_m_l_model(&self, input: &DeleteMLModelInput) -> Result<DeleteMLModelOutput, DeleteMLModelError>

Assigns the DELETED status to an MLModel, rendering it unusable.

After using the DeleteMLModel operation, you can use the GetMLModel operation to verify that the status of the MLModel changed to DELETED.

Caution: The result of the DeleteMLModel operation is irreversible.

fn delete_realtime_endpoint(&self, input: &DeleteRealtimeEndpointInput) -> Result<DeleteRealtimeEndpointOutput, DeleteRealtimeEndpointError>

Deletes a real time endpoint of an MLModel.

fn describe_batch_predictions(&self, input: &DescribeBatchPredictionsInput) -> Result<DescribeBatchPredictionsOutput, DescribeBatchPredictionsError>

Returns a list of BatchPrediction operations that match the search criteria in the request.

fn describe_data_sources(&self, input: &DescribeDataSourcesInput) -> Result<DescribeDataSourcesOutput, DescribeDataSourcesError>

Returns a list of DataSource that match the search criteria in the request.

fn describe_evaluations(&self, input: &DescribeEvaluationsInput) -> Result<DescribeEvaluationsOutput, DescribeEvaluationsError>

Returns a list of DescribeEvaluations that match the search criteria in the request.

fn describe_m_l_models(&self, input: &DescribeMLModelsInput) -> Result<DescribeMLModelsOutput, DescribeMLModelsError>

Returns a list of MLModel that match the search criteria in the request.

fn get_batch_prediction(&self, input: &GetBatchPredictionInput) -> Result<GetBatchPredictionOutput, GetBatchPredictionError>

Returns a BatchPrediction that includes detailed metadata, status, and data file information for a Batch Prediction request.

fn get_data_source(&self, input: &GetDataSourceInput) -> Result<GetDataSourceOutput, GetDataSourceError>

Returns a DataSource that includes metadata and data file information, as well as the current status of the DataSource.

GetDataSource provides results in normal or verbose format. The verbose format adds the schema description and the list of files pointed to by the DataSource to the normal format.

fn get_evaluation(&self, input: &GetEvaluationInput) -> Result<GetEvaluationOutput, GetEvaluationError>

Returns an Evaluation that includes metadata as well as the current status of the Evaluation.

fn get_m_l_model(&self, input: &GetMLModelInput) -> Result<GetMLModelOutput, GetMLModelError>

Returns an MLModel that includes detailed metadata, and data source information as well as the current status of the MLModel.

GetMLModel provides results in normal or verbose format.

fn predict(&self, input: &PredictInput) -> Result<PredictOutput, PredictError>

Generates a prediction for the observation using the specified ML Model.
Note
Not all response parameters will be populated. Whether a response parameter is populated depends on the type of model requested.

fn update_batch_prediction(&self, input: &UpdateBatchPredictionInput) -> Result<UpdateBatchPredictionOutput, UpdateBatchPredictionError>

Updates the BatchPredictionName of a BatchPrediction.

You can use the GetBatchPrediction operation to view the contents of the updated data element.

fn update_data_source(&self, input: &UpdateDataSourceInput) -> Result<UpdateDataSourceOutput, UpdateDataSourceError>

Updates the DataSourceName of a DataSource.

You can use the GetDataSource operation to view the contents of the updated data element.

fn update_evaluation(&self, input: &UpdateEvaluationInput) -> Result<UpdateEvaluationOutput, UpdateEvaluationError>

Updates the EvaluationName of an Evaluation.

You can use the GetEvaluation operation to view the contents of the updated data element.

fn update_m_l_model(&self, input: &UpdateMLModelInput) -> Result<UpdateMLModelOutput, UpdateMLModelError>

Updates the MLModelName and the ScoreThreshold of an MLModel.

You can use the GetMLModel operation to view the contents of the updated data element.