Module rusoto::opsworks

AWS OpsWorks

Structs

AgentVersion
Describes an agent version.

App
A description of the app.

AssignInstanceRequest

AssignVolumeRequest

AssociateElasticIpRequest

AttachElasticLoadBalancerRequest

AutoScalingThresholds
Describes a load-based auto scaling upscaling or downscaling threshold configuration, which specifies when AWS OpsWorks starts or stops load-based instances.

BlockDeviceMapping
Describes a block device mapping. This data type maps directly to the Amazon EC2 BlockDeviceMapping data type.

ChefConfiguration
Describes the Chef configuration.

CloneStackRequest

CloneStackResult
Contains the response to a CloneStack request.

Command
Describes a command.

CreateAppRequest

CreateAppResult
Contains the response to a CreateApp request.

CreateDeploymentRequest

CreateDeploymentResult
Contains the response to a CreateDeployment request.

CreateInstanceRequest

CreateInstanceResult
Contains the response to a CreateInstance request.

CreateLayerRequest

CreateLayerResult
Contains the response to a CreateLayer request.

CreateStackRequest

CreateStackResult
Contains the response to a CreateStack request.

CreateUserProfileRequest

CreateUserProfileResult
Contains the response to a CreateUserProfile request.

DataSource
Describes an app's data source.

DeleteAppRequest

DeleteInstanceRequest

DeleteLayerRequest

DeleteStackRequest

DeleteUserProfileRequest

Deployment
Describes a deployment of a stack or app.

DeploymentCommand
Used to specify a stack or deployment command.

DeregisterEcsClusterRequest

DeregisterElasticIpRequest

DeregisterInstanceRequest

DeregisterRdsDbInstanceRequest

DeregisterVolumeRequest

DescribeAgentVersionsRequest

DescribeAgentVersionsResult
Contains the response to a DescribeAgentVersions request.

DescribeAppsRequest

DescribeAppsResult
Contains the response to a DescribeApps request.

DescribeCommandsRequest

DescribeCommandsResult
Contains the response to a DescribeCommands request.

DescribeDeploymentsRequest

DescribeDeploymentsResult
Contains the response to a DescribeDeployments request.

DescribeEcsClustersRequest

DescribeEcsClustersResult
Contains the response to a DescribeEcsClusters request.

DescribeElasticIpsRequest

DescribeElasticIpsResult
Contains the response to a DescribeElasticIps request.

DescribeElasticLoadBalancersRequest

DescribeElasticLoadBalancersResult
Contains the response to a DescribeElasticLoadBalancers request.

DescribeInstancesRequest

DescribeInstancesResult
Contains the response to a DescribeInstances request.

DescribeLayersRequest

DescribeLayersResult
Contains the response to a DescribeLayers request.

DescribeLoadBasedAutoScalingRequest

DescribeLoadBasedAutoScalingResult
Contains the response to a DescribeLoadBasedAutoScaling request.

DescribeMyUserProfileResult
Contains the response to a DescribeMyUserProfile request.

DescribePermissionsRequest

DescribePermissionsResult
Contains the response to a DescribePermissions request.

DescribeRaidArraysRequest

DescribeRaidArraysResult
Contains the response to a DescribeRaidArrays request.

DescribeRdsDbInstancesRequest

DescribeRdsDbInstancesResult
Contains the response to a DescribeRdsDbInstances request.

DescribeServiceErrorsRequest

DescribeServiceErrorsResult
Contains the response to a DescribeServiceErrors request.

DescribeStackProvisioningParametersRequest

DescribeStackProvisioningParametersResult
Contains the response to a DescribeStackProvisioningParameters request.

DescribeStackSummaryRequest

DescribeStackSummaryResult
Contains the response to a DescribeStackSummary request.

DescribeStacksRequest

DescribeStacksResult
Contains the response to a DescribeStacks request.

DescribeTimeBasedAutoScalingRequest

DescribeTimeBasedAutoScalingResult
Contains the response to a DescribeTimeBasedAutoScaling request.

DescribeUserProfilesRequest

DescribeUserProfilesResult
Contains the response to a DescribeUserProfiles request.

DescribeVolumesRequest

DescribeVolumesResult
Contains the response to a DescribeVolumes request.

DetachElasticLoadBalancerRequest

DisassociateElasticIpRequest

EbsBlockDevice
Describes an Amazon EBS volume. This data type maps directly to the Amazon EC2 EbsBlockDevice data type.

EcsCluster
Describes a registered Amazon ECS cluster.

ElasticIp
Describes an Elastic IP address.

ElasticLoadBalancer
Describes an Elastic Load Balancing instance.

EnvironmentVariable
Represents an app's environment variable.

GetHostnameSuggestionRequest

GetHostnameSuggestionResult
Contains the response to a GetHostnameSuggestion request.

GrantAccessRequest

GrantAccessResult
Contains the response to a GrantAccess request.

Instance
Describes an instance.

InstanceIdentity
Contains a description of an Amazon EC2 instance from the Amazon EC2 metadata service. For more information, see Instance Metadata and User Data.

InstancesCount
Describes how many instances a stack has for each status.

Layer
Describes a layer.

LifecycleEventConfiguration
Specifies the lifecycle event configuration

LoadBasedAutoScalingConfiguration
Describes a layer's load-based auto scaling configuration.

OpsWorksClient
A client for the AWS OpsWorks API.

Permission
Describes stack or user permissions.

RaidArray
Describes an instance's RAID array.

RdsDbInstance
Describes an Amazon RDS instance.

RebootInstanceRequest

Recipes
AWS OpsWorks supports five lifecycle events: setup, configuration, deploy, undeploy, and shutdown. For each layer, AWS OpsWorks runs a set of standard recipes for each event. In addition, you can provide custom recipes for any or all layers and events. AWS OpsWorks runs custom event recipes after the standard recipes. LayerCustomRecipes specifies the custom recipes for a particular layer to be run in response to each of the five events.

To specify a recipe, use the cookbook's directory name in the repository followed by two colons and the recipe name, which is the recipe's file name without the .rb extension. For example: phpapp2::dbsetup specifies the dbsetup.rb recipe in the repository's phpapp2 folder.

RegisterEcsClusterRequest

RegisterEcsClusterResult
Contains the response to a RegisterEcsCluster request.

RegisterElasticIpRequest

RegisterElasticIpResult
Contains the response to a RegisterElasticIp request.

RegisterInstanceRequest

RegisterInstanceResult
Contains the response to a RegisterInstanceResult request.

RegisterRdsDbInstanceRequest

RegisterVolumeRequest

RegisterVolumeResult
Contains the response to a RegisterVolume request.

ReportedOs
A registered instance's reported operating system.

SelfUserProfile
Describes a user's SSH information.

ServiceError
Describes an AWS OpsWorks service error.

SetLoadBasedAutoScalingRequest

SetPermissionRequest

SetTimeBasedAutoScalingRequest

ShutdownEventConfiguration
The Shutdown event configuration.

Source
Contains the information required to retrieve an app or cookbook from a repository. For more information, see Creating Apps or Custom Recipes and Cookbooks.

SslConfiguration
Describes an app's SSL configuration.

Stack
Describes a stack.

StackConfigurationManager
Describes the configuration manager.

StackSummary
Summarizes the number of layers, instances, and apps in a stack.

StartInstanceRequest

StartStackRequest

StopInstanceRequest

StopStackRequest

TemporaryCredential
Contains the data needed by RDP clients such as the Microsoft Remote Desktop Connection to log in to the instance.

TimeBasedAutoScalingConfiguration
Describes an instance's time-based auto scaling configuration.

UnassignInstanceRequest

UnassignVolumeRequest

UpdateAppRequest

UpdateElasticIpRequest

UpdateInstanceRequest

UpdateLayerRequest

UpdateMyUserProfileRequest

UpdateRdsDbInstanceRequest

UpdateStackRequest

UpdateUserProfileRequest

UpdateVolumeRequest

UserProfile
Describes a user's SSH information.

Volume
Describes an instance's Amazon EBS volume.

VolumeConfiguration
Describes an Amazon EBS volume configuration.

WeeklyAutoScalingSchedule
Describes a time-based instance's auto scaling schedule. The schedule consists of a set of key-value pairs.

The key is the time period (a UTC hour) and must be an integer from 0 - 23.

The value indicates whether the instance should be online or offline for the specified period, and must be set to "on" or "off"

The default setting for all time periods is off, so you use the following parameters primarily to specify the online periods. You don't have to explicitly specify offline periods unless you want to change an online period to an offline period.

The following example specifies that the instance should be online for four hours, from UTC 1200 - 1600. It will be off for the remainder of the day.

{ "12":"on", "13":"on", "14":"on", "15":"on" }

Enums

AssignInstanceError

AssignVolumeError

AssociateElasticIpError

AttachElasticLoadBalancerError

CloneStackError

CreateAppError

CreateDeploymentError

CreateInstanceError

CreateLayerError

CreateStackError

CreateUserProfileError

DeleteAppError

DeleteInstanceError

DeleteLayerError

DeleteStackError

DeleteUserProfileError

DeregisterEcsClusterError

DeregisterElasticIpError

DeregisterInstanceError

DeregisterRdsDbInstanceError

DeregisterVolumeError

DescribeAgentVersionsError

DescribeAppsError

DescribeCommandsError

DescribeDeploymentsError

DescribeEcsClustersError

DescribeElasticIpsError

DescribeElasticLoadBalancersError

DescribeInstancesError

DescribeLayersError

DescribeLoadBasedAutoScalingError

DescribeMyUserProfileError

DescribePermissionsError

DescribeRaidArraysError

DescribeRdsDbInstancesError

DescribeServiceErrorsError

DescribeStackProvisioningParametersError

DescribeStackSummaryError

DescribeStacksError

DescribeTimeBasedAutoScalingError

DescribeUserProfilesError

DescribeVolumesError

DetachElasticLoadBalancerError

DisassociateElasticIpError

GetHostnameSuggestionError

GrantAccessError

RebootInstanceError

RegisterEcsClusterError

RegisterElasticIpError

RegisterInstanceError

RegisterRdsDbInstanceError

RegisterVolumeError

SetLoadBasedAutoScalingError

SetPermissionError

SetTimeBasedAutoScalingError

StartInstanceError

StartStackError

StopInstanceError

StopStackError

UnassignInstanceError

UnassignVolumeError

UpdateAppError

UpdateElasticIpError

UpdateInstanceError

UpdateLayerError

UpdateMyUserProfileError

UpdateRdsDbInstanceError

UpdateStackError

UpdateUserProfileError

UpdateVolumeError

Type Definitions

AgentVersions

AppAttributes

AppAttributesKeys

AppType

Apps

Architecture

AutoScalingType

BlockDeviceMappings

Boolean

Commands

DailyAutoScalingSchedule

DataSources

DateTime

DeploymentCommandArgs

DeploymentCommandName

Deployments

Double

EcsClusters

ElasticIps

ElasticLoadBalancers

EnvironmentVariables

Hour

Instances

Integer

LayerAttributes

LayerAttributesKeys

LayerType

Layers

LoadBasedAutoScalingConfigurations

Minute

Parameters

Permissions

RaidArrays

RdsDbInstances

RootDeviceType

ServiceErrors

SourceType

StackAttributes

StackAttributesKeys

Stacks

Strings

Switch

TimeBasedAutoScalingConfigurations

UserProfiles

ValidForInMinutes

VirtualizationType

VolumeConfigurations

VolumeType

Volumes