Struct rusoto::opsworks::Instance

pub struct Instance {
    pub agent_version: Option<String>,
    pub ami_id: Option<String>,
    pub architecture: Option<Architecture>,
    pub auto_scaling_type: Option<AutoScalingType>,
    pub availability_zone: Option<String>,
    pub block_device_mappings: Option<BlockDeviceMappings>,
    pub created_at: Option<DateTime>,
    pub ebs_optimized: Option<Boolean>,
    pub ec_2_instance_id: Option<String>,
    pub ecs_cluster_arn: Option<String>,
    pub ecs_container_instance_arn: Option<String>,
    pub elastic_ip: Option<String>,
    pub hostname: Option<String>,
    pub infrastructure_class: Option<String>,
    pub install_updates_on_boot: Option<Boolean>,
    pub instance_id: Option<String>,
    pub instance_profile_arn: Option<String>,
    pub instance_type: Option<String>,
    pub last_service_error_id: Option<String>,
    pub layer_ids: Option<Strings>,
    pub os: Option<String>,
    pub platform: Option<String>,
    pub private_dns: Option<String>,
    pub private_ip: Option<String>,
    pub public_dns: Option<String>,
    pub public_ip: Option<String>,
    pub registered_by: Option<String>,
    pub reported_agent_version: Option<String>,
    pub reported_os: Option<ReportedOs>,
    pub root_device_type: Option<RootDeviceType>,
    pub root_device_volume_id: Option<String>,
    pub security_group_ids: Option<Strings>,
    pub ssh_host_dsa_key_fingerprint: Option<String>,
    pub ssh_host_rsa_key_fingerprint: Option<String>,
    pub ssh_key_name: Option<String>,
    pub stack_id: Option<String>,
    pub status: Option<String>,
    pub subnet_id: Option<String>,
    pub tenancy: Option<String>,
    pub virtualization_type: Option<VirtualizationType>,
}

Describes an instance.

Fields
agent_version: Option<String>

The agent version. This parameter is set to INHERIT if the instance inherits the default stack setting or to a a version number for a fixed agent version.

ami_id: Option<String>

A custom AMI ID to be used to create the instance. For more information, see Instances

architecture: Option<Architecture>

The instance architecture: "i386" or "x86_64".

auto_scaling_type: Option<AutoScalingType>

For load-based or time-based instances, the type.

availability_zone: Option<String>

The instance Availability Zone. For more information, see Regions and Endpoints.

block_device_mappings: Option<BlockDeviceMappings>

An array of BlockDeviceMapping objects that specify the instance's block device mappings.

created_at: Option<DateTime>

The time that the instance was created.

ebs_optimized: Option<Boolean>

Whether this is an Amazon EBS-optimized instance.

ec_2_instance_id: Option<String>

The ID of the associated Amazon EC2 instance.

ecs_cluster_arn: Option<String>

For container instances, the Amazon ECS cluster's ARN.

ecs_container_instance_arn: Option<String>

For container instances, the instance's ARN.

elastic_ip: Option<String>

The instance Elastic IP address .

hostname: Option<String>

The instance host name.

infrastructure_class: Option<String>

For registered instances, the infrastructure class: ec2 or on-premises.

install_updates_on_boot: Option<Boolean>

Whether to install operating system and package updates when the instance boots. The default value is true. If this value is set to false, you must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

instance_id: Option<String>

The instance ID.

instance_profile_arn: Option<String>

The ARN of the instance's IAM profile. For more information about IAM ARNs, see Using Identifiers.

instance_type: Option<String>

The instance type, such as t2.micro.

last_service_error_id: Option<String>

The ID of the last service error. For more information, call DescribeServiceErrors.

layer_ids: Option<Strings>

An array containing the instance layer IDs.

os: Option<String>

The instance's operating system.

platform: Option<String>

The instance's platform.

private_dns: Option<String>

The The instance's private DNS name.

private_ip: Option<String>

The instance's private IP address.

public_dns: Option<String>

The instance public DNS name.

public_ip: Option<String>

The instance public IP address.

registered_by: Option<String>

For registered instances, who performed the registration.

reported_agent_version: Option<String>

The instance's reported AWS OpsWorks agent version.

reported_os: Option<ReportedOs>

For registered instances, the reported operating system.

root_device_type: Option<RootDeviceType>

The instance's root device type. For more information, see Storage for the Root Device.

root_device_volume_id: Option<String>

The root device volume ID.

security_group_ids: Option<Strings>

An array containing the instance security group IDs.

ssh_host_dsa_key_fingerprint: Option<String>

The SSH key's Deep Security Agent (DSA) fingerprint.

ssh_host_rsa_key_fingerprint: Option<String>

The SSH key's RSA fingerprint.

ssh_key_name: Option<String>

The instance's Amazon EC2 key-pair name.

stack_id: Option<String>

The stack ID.

status: Option<String>

The instance status:

booting

connection_lost

online

pending

rebooting

requested

running_setup

setup_failed

shutting_down

start_failed

stopped

stopping

terminated

terminating

subnet_id: Option<String>

The instance's subnet ID; applicable only if the stack is running in a VPC.

tenancy: Option<String>

The instance's tenancy option, such as dedicated or host.

virtualization_type: Option<VirtualizationType>

The instance's virtualization type: paravirtual or hvm.