Struct rusoto::opsworks::OpsWorksClient

pub struct OpsWorksClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the AWS OpsWorks API.

Methods

impl<P> OpsWorksClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> OpsWorksClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn assign_instance(&self, input: &AssignInstanceRequest) -> Result<(), AssignInstanceError>

Assign a registered instance to a layer.

You can assign registered on-premises instances to any layer type.

You can assign registered Amazon EC2 instances only to custom layers.

You cannot use this action with instances that were created with AWS OpsWorks.

Required Permissions: To use this action, an AWS Identity and Access Management (IAM) user must have a Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn assign_volume(&self, input: &AssignVolumeRequest) -> Result<(), AssignVolumeError>

Assigns one of the stack's registered Amazon EBS volumes to a specified instance. The volume must first be registered with the stack by calling RegisterVolume. After you register the volume, you must call UpdateVolume to specify a mount point before calling AssignVolume. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn associate_elastic_ip(&self, input: &AssociateElasticIpRequest) -> Result<(), AssociateElasticIpError>

Associates one of the stack's registered Elastic IP addresses with a specified instance. The address must first be registered with the stack by calling RegisterElasticIp. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn attach_elastic_load_balancer(&self, input: &AttachElasticLoadBalancerRequest) -> Result<(), AttachElasticLoadBalancerError>

Attaches an Elastic Load Balancing load balancer to a specified layer. For more information, see Elastic Load Balancing.

You must create the Elastic Load Balancing instance separately, by using the Elastic Load Balancing console, API, or CLI. For more information, see Elastic Load Balancing Developer Guide.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn clone_stack(&self, input: &CloneStackRequest) -> Result<CloneStackResult, CloneStackError>

Creates a clone of a specified stack. For more information, see Clone a Stack. By default, all parameters are set to the values used by the parent stack.

Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn create_app(&self, input: &CreateAppRequest) -> Result<CreateAppResult, CreateAppError>

Creates an app for a specified stack. For more information, see Creating Apps.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn create_deployment(&self, input: &CreateDeploymentRequest) -> Result<CreateDeploymentResult, CreateDeploymentError>

Runs deployment or stack commands. For more information, see Deploying Apps and Run Stack Commands.

Required Permissions: To use this action, an IAM user must have a Deploy or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn create_instance(&self, input: &CreateInstanceRequest) -> Result<CreateInstanceResult, CreateInstanceError>

Creates an instance in a specified stack. For more information, see Adding an Instance to a Layer.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn create_layer(&self, input: &CreateLayerRequest) -> Result<CreateLayerResult, CreateLayerError>

Creates a layer. For more information, see How to Create a Layer.

You should use CreateLayer for noncustom layer types such as PHP App Server only if the stack does not have an existing layer of that type. A stack can have at most one instance of each noncustom layer; if you attempt to create a second instance, CreateLayer fails. A stack can have an arbitrary number of custom layers, so you can call CreateLayer as many times as you like for that layer type.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn create_stack(&self, input: &CreateStackRequest) -> Result<CreateStackResult, CreateStackError>

Creates a new stack. For more information, see Create a New Stack.

Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn create_user_profile(&self, input: &CreateUserProfileRequest) -> Result<CreateUserProfileResult, CreateUserProfileError>

Creates a new user profile.

Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn delete_app(&self, input: &DeleteAppRequest) -> Result<(), DeleteAppError>

Deletes a specified app.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn delete_instance(&self, input: &DeleteInstanceRequest) -> Result<(), DeleteInstanceError>

Deletes a specified instance, which terminates the associated Amazon EC2 instance. You must stop an instance before you can delete it.

For more information, see Deleting Instances.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn delete_layer(&self, input: &DeleteLayerRequest) -> Result<(), DeleteLayerError>

Deletes a specified layer. You must first stop and then delete all associated instances or unassign registered instances. For more information, see How to Delete a Layer.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn delete_stack(&self, input: &DeleteStackRequest) -> Result<(), DeleteStackError>

Deletes a specified stack. You must first delete all instances, layers, and apps or deregister registered instances. For more information, see Shut Down a Stack.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn delete_user_profile(&self, input: &DeleteUserProfileRequest) -> Result<(), DeleteUserProfileError>

Deletes a user profile.

Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn deregister_ecs_cluster(&self, input: &DeregisterEcsClusterRequest) -> Result<(), DeregisterEcsClusterError>

Deregisters a specified Amazon ECS cluster from a stack. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see .

fn deregister_elastic_ip(&self, input: &DeregisterElasticIpRequest) -> Result<(), DeregisterElasticIpError>

Deregisters a specified Elastic IP address. The address can then be registered by another stack. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn deregister_instance(&self, input: &DeregisterInstanceRequest) -> Result<(), DeregisterInstanceError>

Deregister a registered Amazon EC2 or on-premises instance. This action removes the instance from the stack and returns it to your control. This action can not be used with instances that were created with AWS OpsWorks.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn deregister_rds_db_instance(&self, input: &DeregisterRdsDbInstanceRequest) -> Result<(), DeregisterRdsDbInstanceError>

Deregisters an Amazon RDS instance.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn deregister_volume(&self, input: &DeregisterVolumeRequest) -> Result<(), DeregisterVolumeError>

Deregisters an Amazon EBS volume. The volume can then be registered by another stack. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_agent_versions(&self, input: &DescribeAgentVersionsRequest) -> Result<DescribeAgentVersionsResult, DescribeAgentVersionsError>

Describes the available AWS OpsWorks agent versions. You must specify a stack ID or a configuration manager. DescribeAgentVersions returns a list of available agent versions for the specified stack or configuration manager.

fn describe_apps(&self, input: &DescribeAppsRequest) -> Result<DescribeAppsResult, DescribeAppsError>

Requests a description of a specified set of apps.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_commands(&self, input: &DescribeCommandsRequest) -> Result<DescribeCommandsResult, DescribeCommandsError>

Describes the results of specified commands.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_deployments(&self, input: &DescribeDeploymentsRequest) -> Result<DescribeDeploymentsResult, DescribeDeploymentsError>

Requests a description of a specified set of deployments.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_ecs_clusters(&self, input: &DescribeEcsClustersRequest) -> Result<DescribeEcsClustersResult, DescribeEcsClustersError>

Describes Amazon ECS clusters that are registered with a stack. If you specify only a stack ID, you can use the MaxResults and NextToken parameters to paginate the response. However, AWS OpsWorks currently supports only one cluster per layer, so the result set has a maximum of one element.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack or an attached policy that explicitly grants permission. For more information on user permissions, see Managing User Permissions.

fn describe_elastic_ips(&self, input: &DescribeElasticIpsRequest) -> Result<DescribeElasticIpsResult, DescribeElasticIpsError>

Describes Elastic IP addresses.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_elastic_load_balancers(&self, input: &DescribeElasticLoadBalancersRequest) -> Result<DescribeElasticLoadBalancersResult, DescribeElasticLoadBalancersError>

Describes a stack's Elastic Load Balancing instances.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_instances(&self, input: &DescribeInstancesRequest) -> Result<DescribeInstancesResult, DescribeInstancesError>

Requests a description of a set of instances.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_layers(&self, input: &DescribeLayersRequest) -> Result<DescribeLayersResult, DescribeLayersError>

Requests a description of one or more layers in a specified stack.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_load_based_auto_scaling(&self, input: &DescribeLoadBasedAutoScalingRequest) -> Result<DescribeLoadBasedAutoScalingResult, DescribeLoadBasedAutoScalingError>

Describes load-based auto scaling configurations for specified layers.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_my_user_profile(&self) -> Result<DescribeMyUserProfileResult, DescribeMyUserProfileError>

Describes a user's SSH information.

Required Permissions: To use this action, an IAM user must have self-management enabled or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_permissions(&self, input: &DescribePermissionsRequest) -> Result<DescribePermissionsResult, DescribePermissionsError>

Describes the permissions for a specified stack.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_raid_arrays(&self, input: &DescribeRaidArraysRequest) -> Result<DescribeRaidArraysResult, DescribeRaidArraysError>

Describe an instance's RAID arrays.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_rds_db_instances(&self, input: &DescribeRdsDbInstancesRequest) -> Result<DescribeRdsDbInstancesResult, DescribeRdsDbInstancesError>

Describes Amazon RDS instances.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_service_errors(&self, input: &DescribeServiceErrorsRequest) -> Result<DescribeServiceErrorsResult, DescribeServiceErrorsError>

Describes AWS OpsWorks service errors.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_stack_provisioning_parameters(&self, input: &DescribeStackProvisioningParametersRequest) -> Result<DescribeStackProvisioningParametersResult, DescribeStackProvisioningParametersError>

Requests a description of a stack's provisioning parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_stack_summary(&self, input: &DescribeStackSummaryRequest) -> Result<DescribeStackSummaryResult, DescribeStackSummaryError>

Describes the number of layers and apps in a specified stack, and the number of instances in each state, such as running_setup or online.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_stacks(&self, input: &DescribeStacksRequest) -> Result<DescribeStacksResult, DescribeStacksError>

Requests a description of one or more stacks.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_time_based_auto_scaling(&self, input: &DescribeTimeBasedAutoScalingRequest) -> Result<DescribeTimeBasedAutoScalingResult, DescribeTimeBasedAutoScalingError>

Describes time-based auto scaling configurations for specified instances.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_user_profiles(&self, input: &DescribeUserProfilesRequest) -> Result<DescribeUserProfilesResult, DescribeUserProfilesError>

Describe specified users.

Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn describe_volumes(&self, input: &DescribeVolumesRequest) -> Result<DescribeVolumesResult, DescribeVolumesError>

Describes an instance's Amazon EBS volumes.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn detach_elastic_load_balancer(&self, input: &DetachElasticLoadBalancerRequest) -> Result<(), DetachElasticLoadBalancerError>

Detaches a specified Elastic Load Balancing instance from its layer.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn disassociate_elastic_ip(&self, input: &DisassociateElasticIpRequest) -> Result<(), DisassociateElasticIpError>

Disassociates an Elastic IP address from its instance. The address remains registered with the stack. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn get_hostname_suggestion(&self, input: &GetHostnameSuggestionRequest) -> Result<GetHostnameSuggestionResult, GetHostnameSuggestionError>

Gets a generated host name for the specified layer, based on the current host name theme.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn grant_access(&self, input: &GrantAccessRequest) -> Result<GrantAccessResult, GrantAccessError>

This action can be used only with Windows stacks.
Grants RDP access to a Windows instance for a specified time period.

fn reboot_instance(&self, input: &RebootInstanceRequest) -> Result<(), RebootInstanceError>

Reboots a specified instance. For more information, see Starting, Stopping, and Rebooting Instances.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn register_ecs_cluster(&self, input: &RegisterEcsClusterRequest) -> Result<RegisterEcsClusterResult, RegisterEcsClusterError>

Registers a specified Amazon ECS cluster with a stack. You can register only one cluster with a stack. A cluster can be registered with only one stack. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn register_elastic_ip(&self, input: &RegisterElasticIpRequest) -> Result<RegisterElasticIpResult, RegisterElasticIpError>

Registers an Elastic IP address with a specified stack. An address can be registered with only one stack at a time. If the address is already registered, you must first deregister it by calling DeregisterElasticIp. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn register_instance(&self, input: &RegisterInstanceRequest) -> Result<RegisterInstanceResult, RegisterInstanceError>

Registers instances with a specified stack that were created outside of AWS OpsWorks.
We do not recommend using this action to register instances. The complete registration operation has two primary steps, installing the AWS OpsWorks agent on the instance and registering the instance with the stack. RegisterInstance handles only the second step. You should instead use the AWS CLI register command, which performs the entire registration operation. For more information, see Registering an Instance with an AWS OpsWorks Stack.
Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn register_rds_db_instance(&self, input: &RegisterRdsDbInstanceRequest) -> Result<(), RegisterRdsDbInstanceError>

Registers an Amazon RDS instance with a stack.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn register_volume(&self, input: &RegisterVolumeRequest) -> Result<RegisterVolumeResult, RegisterVolumeError>

Registers an Amazon EBS volume with a specified stack. A volume can be registered with only one stack at a time. If the volume is already registered, you must first deregister it by calling DeregisterVolume. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn set_load_based_auto_scaling(&self, input: &SetLoadBasedAutoScalingRequest) -> Result<(), SetLoadBasedAutoScalingError>

Specify the load-based auto scaling configuration for a specified layer. For more information, see Managing Load with Time-based and Load-based Instances.

To use load-based auto scaling, you must create a set of load-based auto scaling instances. Load-based auto scaling operates only on the instances from that set, so you must ensure that you have created enough instances to handle the maximum anticipated load.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn set_permission(&self, input: &SetPermissionRequest) -> Result<(), SetPermissionError>

Specifies a user's permissions. For more information, see Security and Permissions.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn set_time_based_auto_scaling(&self, input: &SetTimeBasedAutoScalingRequest) -> Result<(), SetTimeBasedAutoScalingError>

Specify the time-based auto scaling configuration for a specified instance. For more information, see Managing Load with Time-based and Load-based Instances.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn start_instance(&self, input: &StartInstanceRequest) -> Result<(), StartInstanceError>

Starts a specified instance. For more information, see Starting, Stopping, and Rebooting Instances.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn start_stack(&self, input: &StartStackRequest) -> Result<(), StartStackError>

Starts a stack's instances.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn stop_instance(&self, input: &StopInstanceRequest) -> Result<(), StopInstanceError>

Stops a specified instance. When you stop a standard instance, the data disappears and must be reinstalled when you restart the instance. You can stop an Amazon EBS-backed instance without losing data. For more information, see Starting, Stopping, and Rebooting Instances.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn stop_stack(&self, input: &StopStackRequest) -> Result<(), StopStackError>

Stops a specified stack.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn unassign_instance(&self, input: &UnassignInstanceRequest) -> Result<(), UnassignInstanceError>

Unassigns a registered instance from all of it's layers. The instance remains in the stack as an unassigned instance and can be assigned to another layer, as needed. You cannot use this action with instances that were created with AWS OpsWorks.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn unassign_volume(&self, input: &UnassignVolumeRequest) -> Result<(), UnassignVolumeError>

Unassigns an assigned Amazon EBS volume. The volume remains registered with the stack. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn update_app(&self, input: &UpdateAppRequest) -> Result<(), UpdateAppError>

Updates a specified app.

Required Permissions: To use this action, an IAM user must have a Deploy or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn update_elastic_ip(&self, input: &UpdateElasticIpRequest) -> Result<(), UpdateElasticIpError>

Updates a registered Elastic IP address's name. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn update_instance(&self, input: &UpdateInstanceRequest) -> Result<(), UpdateInstanceError>

Updates a specified instance.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn update_layer(&self, input: &UpdateLayerRequest) -> Result<(), UpdateLayerError>

Updates a specified layer.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn update_my_user_profile(&self, input: &UpdateMyUserProfileRequest) -> Result<(), UpdateMyUserProfileError>

Updates a user's SSH public key.

Required Permissions: To use this action, an IAM user must have self-management enabled or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn update_rds_db_instance(&self, input: &UpdateRdsDbInstanceRequest) -> Result<(), UpdateRdsDbInstanceError>

Updates an Amazon RDS instance.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn update_stack(&self, input: &UpdateStackRequest) -> Result<(), UpdateStackError>

Updates a specified stack.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn update_user_profile(&self, input: &UpdateUserProfileRequest) -> Result<(), UpdateUserProfileError>

Updates a specified user profile.

Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

fn update_volume(&self, input: &UpdateVolumeRequest) -> Result<(), UpdateVolumeError>

Updates an Amazon EBS volume's name or mount point. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.