Struct rusoto::s3::UploadPartRequest

pub struct UploadPartRequest<'a> {
    pub body: Option<&'a [u8]>,
    pub sse_customer_algorithm: Option<SSECustomerAlgorithm>,
    pub request_payer: Option<RequestPayer>,
    pub content_length: Option<ContentLength>,
    pub content_md5: Option<ContentMD5>,
    pub bucket: BucketName,
    pub sse_customer_key: Option<SSECustomerKey>,
    pub upload_id: MultipartUploadId,
    pub key: ObjectKey,
    pub sse_customer_key_md5: Option<SSECustomerKeyMD5>,
    pub part_number: PartNumber,
}

Fields
body: Option<&'a [u8]> sse_customer_algorithm: Option<SSECustomerAlgorithm>

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

request_payer: Option<RequestPayer> content_length: Option<ContentLength>

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically.

content_md5: Option<ContentMD5> bucket: BucketName sse_customer_key: Option<SSECustomerKey>

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side- encryption-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

upload_id: MultipartUploadId

Upload ID identifying the multipart upload whose part is being uploaded.

key: ObjectKey sse_customer_key_md5: Option<SSECustomerKeyMD5>

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

part_number: PartNumber

Part number of part being uploaded. This is a positive integer between 1 and 10,000.