Module rusoto::sqs

The AWS SQS API.

Structs

AddPermissionRequest

BatchResultErrorEntry
This is used in the responses of batch API to give a detailed description of the result of an action on each entry in the request.

ChangeMessageVisibilityBatchRequest

ChangeMessageVisibilityBatchRequestEntry
Encloses a receipt handle and an entry id for each message in ChangeMessageVisibilityBatch.

All of the following parameters are list parameters that must be prefixed with ChangeMessageVisibilityBatchRequestEntry.n, where n is an integer value starting with 1. For example, a parameter list for this action might look like this:

ChangeMessageVisibilityBatchResult
For each message in the batch, the response contains a ChangeMessageVisibilityBatchResultEntry tag if the message succeeds or a BatchResultErrorEntry tag if the message fails.

ChangeMessageVisibilityBatchResultEntry
Encloses the id of an entry in ChangeMessageVisibilityBatch.

ChangeMessageVisibilityRequest

CreateQueueRequest

CreateQueueResult
Returns the QueueUrl element of the created queue.

DeleteMessageBatchRequest

DeleteMessageBatchRequestEntry
Encloses a receipt handle and an identifier for it.

DeleteMessageBatchResult
For each message in the batch, the response contains a DeleteMessageBatchResultEntry tag if the message is deleted or a BatchResultErrorEntry tag if the message cannot be deleted.

DeleteMessageBatchResultEntry
Encloses the id an entry in DeleteMessageBatch.

DeleteMessageRequest

DeleteQueueRequest

GetQueueAttributesRequest

GetQueueAttributesResult
A list of returned queue attributes.

GetQueueUrlRequest

GetQueueUrlResult
For more information, see Responses in the Amazon SQS Developer Guide.

ListDeadLetterSourceQueuesRequest

ListDeadLetterSourceQueuesResult
A list of your dead letter source queues.

ListQueuesRequest

ListQueuesResult
A list of your queues.

Message
An Amazon SQS message.

MessageAttributeValue
The user-specified message attribute value. For string data types, the value attribute has the same restrictions on the content as the message body. For more information, see SendMessage.

Name, type, and value must not be empty or null. In addition, the message body should not be empty or null. All parts of the message attribute, including name, type, and value, are included in the message size restriction, which is currently 256 KB (262,144 bytes).

PurgeQueueRequest

ReceiveMessageRequest

ReceiveMessageResult
A list of received messages.

RemovePermissionRequest

SendMessageBatchRequest

SendMessageBatchRequestEntry
Contains the details of a single Amazon SQS message along with a Id.

SendMessageBatchResult
For each message in the batch, the response contains a SendMessageBatchResultEntry tag if the message succeeds or a BatchResultErrorEntry tag if the message fails.

SendMessageBatchResultEntry
Encloses a message ID for successfully enqueued message of a SendMessageBatch.

SendMessageRequest

SendMessageResult
The MD5OfMessageBody and MessageId elements.

SetQueueAttributesRequest

SqsClient
A client for the Amazon SQS API.

Enums

AddPermissionError

ChangeMessageVisibilityBatchError

ChangeMessageVisibilityError

CreateQueueError

DeleteMessageBatchError

DeleteMessageError

DeleteQueueError

GetQueueAttributesError

GetQueueUrlError

ListDeadLetterSourceQueuesError

ListQueuesError

PurgeQueueError

ReceiveMessageError

RemovePermissionError

SendMessageBatchError

SendMessageError

SetQueueAttributesError

Type Definitions

AWSAccountIdList

ActionNameList

AttributeMap

AttributeNameList

BatchResultErrorEntryList

Binary

BinaryList

Boolean

ChangeMessageVisibilityBatchRequestEntryList

ChangeMessageVisibilityBatchResultEntryList

DeleteMessageBatchRequestEntryList

DeleteMessageBatchResultEntryList

Integer

MessageAttributeMap

MessageAttributeName

MessageAttributeNameList

MessageList

QueueAttributeName

QueueUrlList

SendMessageBatchRequestEntryList

SendMessageBatchResultEntryList

StringList