Module rusoto::ssm

Amazon EC2 Simple Systems Manager

Structs

Association
Describes an association of an SSM document and an instance.

AssociationDescription
Describes the parameters for a document.

AssociationFilter
Describes a filter.

AssociationStatus
Describes an association status.

CancelCommandRequest

CancelCommandResult
Whether or not the command was successfully canceled. There is no guarantee that a request can be canceled.

Command
Describes a command request.

CommandFilter
Describes a command filter.

CommandInvocation
An invocation is copy of a command sent to a specific instance. A command can apply to one or more instances. A command invocation applies to one instance. For example, if a user executes SendCommand against three instances, then a command invocation is created for each requested instance ID. A command invocation returns status and detail information about a command you executed.

CommandPlugin
Describes plugin details.

CreateAssociationBatchRequest

CreateAssociationBatchRequestEntry
Describes the association of an SSM document and an instance.

CreateAssociationBatchResult

CreateAssociationRequest

CreateAssociationResult

CreateDocumentRequest

CreateDocumentResult

DeleteAssociationRequest

DeleteAssociationResult

DeleteDocumentRequest

DeleteDocumentResult

DescribeAssociationRequest

DescribeAssociationResult

DescribeDocumentPermissionRequest

DescribeDocumentPermissionResponse

DescribeDocumentRequest

DescribeDocumentResult

DescribeInstanceInformationRequest

DescribeInstanceInformationResult

DocumentDescription
Describes an SSM document.

DocumentFilter
Describes a filter.

DocumentIdentifier
Describes the name of an SSM document.

DocumentParameter
Parameters specified in the SSM document that execute on the server when the command is run.

FailedCreateAssociation
Describes a failed association.

GetDocumentRequest

GetDocumentResult

InstanceInformation
Describes a filter for a specific list of instances.

InstanceInformationFilter
Describes a filter for a specific list of instances.

ListAssociationsRequest

ListAssociationsResult

ListCommandInvocationsRequest

ListCommandInvocationsResult

ListCommandsRequest

ListCommandsResult

ListDocumentsRequest

ListDocumentsResult

ModifyDocumentPermissionRequest

ModifyDocumentPermissionResponse

SendCommandRequest

SendCommandResult

SsmClient
A client for the Amazon SSM API.

UpdateAssociationStatusRequest

UpdateAssociationStatusResult

Enums

CancelCommandError

CreateAssociationBatchError

CreateAssociationError

CreateDocumentError

DeleteAssociationError

DeleteDocumentError

DescribeAssociationError

DescribeDocumentError

DescribeDocumentPermissionError

DescribeInstanceInformationError

GetDocumentError

ListAssociationsError

ListCommandInvocationsError

ListCommandsError

ListDocumentsError

ModifyDocumentPermissionError

SendCommandError

UpdateAssociationStatusError

Type Definitions

AccountId

AccountIdList

AssociationDescriptionList

AssociationFilterKey

AssociationFilterList

AssociationFilterValue

AssociationList

AssociationStatusName

BatchErrorMessage

Boolean

CommandFilterKey

CommandFilterList

CommandFilterValue

CommandId

CommandInvocationList

CommandInvocationStatus

CommandList

CommandMaxResults

CommandPluginList

CommandPluginName

CommandPluginOutput

CommandPluginStatus

CommandStatus

Comment

CreateAssociationBatchRequestEntries

DateTime

DescriptionInDocument

DocumentARN

DocumentContent

DocumentFilterKey

DocumentFilterList

DocumentFilterValue

DocumentHash

DocumentHashType

DocumentIdentifierList

DocumentName

DocumentOwner

DocumentParameterDefaultValue

DocumentParameterDescrption

DocumentParameterList

DocumentParameterName

DocumentParameterType

DocumentPermissionType

DocumentSha1

DocumentStatus

FailedCreateAssociationList

Fault

InstanceId

InstanceIdList

InstanceInformationFilterKey

InstanceInformationFilterList

InstanceInformationFilterValue

InstanceInformationFilterValueSet

InstanceInformationList

InvocationTraceOutput

ListDocumentsMaxResults

MaxResults

MaxResultsEC2Compatible

NextToken

ParameterName

ParameterValue

ParameterValueList

Parameters

PingStatus

PlatformType

PlatformTypeList

ResponseCode

S3BucketName

S3KeyPrefix

StatusAdditionalInfo

StatusMessage

TimeoutSeconds

Version