Struct rusoto::ssm::SendCommandRequest

pub struct SendCommandRequest {
    pub comment: Option<Comment>,
    pub document_hash: Option<DocumentHash>,
    pub document_hash_type: Option<DocumentHashType>,
    pub document_name: DocumentARN,
    pub instance_ids: Option<InstanceIdList>,
    pub max_concurrency: Option<VelocityConstraint>,
    pub max_errors: Option<VelocityConstraint>,
    pub notification_config: Option<NotificationConfig>,
    pub output_s3_bucket_name: Option<S3BucketName>,
    pub output_s3_key_prefix: Option<S3KeyPrefix>,
    pub output_s3_region: Option<S3Region>,
    pub parameters: Option<Parameters>,
    pub service_role_arn: Option<ServiceRole>,
    pub targets: Option<Targets>,
    pub timeout_seconds: Option<TimeoutSeconds>,
}

Fields
comment: Option<Comment>

User-specified information about the command, such as a brief description of what the command should do.

document_hash: Option<DocumentHash>

The Sha256 or Sha1 hash created by the system when the document was created.

Sha1 hashes have been deprecated.

document_hash_type: Option<DocumentHashType>

Sha256 or Sha1.

Sha1 hashes have been deprecated.

document_name: DocumentARN

Required. The name of the SSM document to execute. This can be an SSM public document or a custom document.

instance_ids: Option<InstanceIdList>

Required. The instance IDs where the command should execute. You can specify a maximum of 50 IDs.

max_concurrency: Option<VelocityConstraint>

(Optional) The maximum number of instances that are allowed to execute the command at the same time. You can specify a number such as “10” or a percentage such as “10%”. The default value is 50. For more information about how to use MaxConcurrency, see Executing a Command Using Amazon EC2 Run Command (Linux) or Executing a Command Using Amazon EC2 Run Command (Windows).

max_errors: Option<VelocityConstraint>

The maximum number of errors allowed without the command failing. When the command fails one more time beyond the value of MaxErrors, the systems stops sending the command to additional targets. You can specify a number like “10” or a percentage like “10%”. The default value is 50. For more information about how to use MaxErrors, see Executing a Command Using Amazon EC2 Run Command (Linux) or Executing a Command Using Amazon EC2 Run Command (Windows).

notification_config: Option<NotificationConfig>

Configurations for sending notifications.

output_s3_bucket_name: Option<S3BucketName>

The name of the S3 bucket where command execution responses should be stored.

output_s3_key_prefix: Option<S3KeyPrefix>

The directory structure within the S3 bucket where the responses should be stored.

output_s3_region: Option<S3Region>

(Optional) The region where the Amazon Simple Storage Service (Amazon S3) output bucket is located. The default value is the region where Run Command is being called.

parameters: Option<Parameters>

The required and optional parameters specified in the SSM document being executed.

service_role_arn: Option<ServiceRole>

The IAM role that Systems Manager uses to send notifications.

targets: Option<Targets>

(Optional) An array of search criteria that targets instances using a Key;Value combination that you specify. Targets is required if you don't provide one or more instance IDs in the call. For more information about how to use Targets, see Executing a Command Using Amazon EC2 Run Command (Linux) or Executing a Command Using Amazon EC2 Run Command (Windows).

timeout_seconds: Option<TimeoutSeconds>

If this time is reached and the command has not already started executing, it will not execute.