Struct rusoto::ssm::SsmClient

pub struct SsmClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the Amazon SSM API.

Methods

impl<P> SsmClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> SsmClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn cancel_command(&self, input: &CancelCommandRequest) -> Result<CancelCommandResult, CancelCommandError>

Attempts to cancel the command specified by the Command ID. There is no guarantee that the command will be terminated and the underlying process stopped.

fn create_association(&self, input: &CreateAssociationRequest) -> Result<CreateAssociationResult, CreateAssociationError>

Associates the specified SSM document with the specified instance.

When you associate an SSM document with an instance, the configuration agent on the instance processes the document and configures the instance as specified.

If you associate a document with an instance that already has an associated document, the system throws the AssociationAlreadyExists exception.

fn create_association_batch(&self, input: &CreateAssociationBatchRequest) -> Result<CreateAssociationBatchResult, CreateAssociationBatchError>

Associates the specified SSM document with the specified instances.

When you associate an SSM document with an instance, the configuration agent on the instance processes the document and configures the instance as specified.

If you associate a document with an instance that already has an associated document, the system throws the AssociationAlreadyExists exception.

fn create_document(&self, input: &CreateDocumentRequest) -> Result<CreateDocumentResult, CreateDocumentError>

Creates an SSM document.

After you create an SSM document, you can use CreateAssociation to associate it with one or more running instances.

fn delete_association(&self, input: &DeleteAssociationRequest) -> Result<DeleteAssociationResult, DeleteAssociationError>

Disassociates the specified SSM document from the specified instance.

When you disassociate an SSM document from an instance, it does not change the configuration of the instance. To change the configuration state of an instance after you disassociate a document, you must create a new document with the desired configuration and associate it with the instance.

fn delete_document(&self, input: &DeleteDocumentRequest) -> Result<DeleteDocumentResult, DeleteDocumentError>

Deletes the SSM document and all instance associations to the document.

Before you delete the SSM document, we recommend that you use DeleteAssociation to disassociate all instances that are associated with the document.

fn describe_association(&self, input: &DescribeAssociationRequest) -> Result<DescribeAssociationResult, DescribeAssociationError>

Describes the associations for the specified SSM document or instance.

fn describe_document(&self, input: &DescribeDocumentRequest) -> Result<DescribeDocumentResult, DescribeDocumentError>

Describes the specified SSM document.

fn describe_document_permission(&self, input: &DescribeDocumentPermissionRequest) -> Result<DescribeDocumentPermissionResponse, DescribeDocumentPermissionError>

Describes the permissions for an SSM document. If you created the document, you are the owner. If a document is shared, it can either be shared privately (by specifying a user’s AWS account ID) or publicly (All).

fn describe_instance_information(&self, input: &DescribeInstanceInformationRequest) -> Result<DescribeInstanceInformationResult, DescribeInstanceInformationError>

Describes one or more of your instances. You can use this to get information about instances like the operating system platform, the SSM agent version, status etc. If you specify one or more instance IDs, it returns information for those instances. If you do not specify instance IDs, it returns information for all your instances. If you specify an instance ID that is not valid or an instance that you do not own, you receive an error.

fn get_document(&self, input: &GetDocumentRequest) -> Result<GetDocumentResult, GetDocumentError>

Gets the contents of the specified SSM document.

fn list_associations(&self, input: &ListAssociationsRequest) -> Result<ListAssociationsResult, ListAssociationsError>

Lists the associations for the specified SSM document or instance.

fn list_command_invocations(&self, input: &ListCommandInvocationsRequest) -> Result<ListCommandInvocationsResult, ListCommandInvocationsError>

An invocation is copy of a command sent to a specific instance. A command can apply to one or more instances. A command invocation applies to one instance. For example, if a user executes SendCommand against three instances, then a command invocation is created for each requested instance ID. ListCommandInvocations provide status about command execution.

fn list_commands(&self, input: &ListCommandsRequest) -> Result<ListCommandsResult, ListCommandsError>

Lists the commands requested by users of the AWS account.

fn list_documents(&self, input: &ListDocumentsRequest) -> Result<ListDocumentsResult, ListDocumentsError>

Describes one or more of your SSM documents.

fn modify_document_permission(&self, input: &ModifyDocumentPermissionRequest) -> Result<ModifyDocumentPermissionResponse, ModifyDocumentPermissionError>

Share a document publicly or privately. If you share a document privately, you must specify the AWS user account IDs for those people who can use the document. If you share a document publicly, you must specify All as the account ID.

fn send_command(&self, input: &SendCommandRequest) -> Result<SendCommandResult, SendCommandError>

Executes commands on one or more remote instances.

fn update_association_status(&self, input: &UpdateAssociationStatusRequest) -> Result<UpdateAssociationStatusResult, UpdateAssociationStatusError>

Updates the status of the SSM document associated with the specified instance.