Module rusoto::storagegateway

Amazon Storage Gateway

Structs

ActivateGatewayInput

A JSON object containing one or more of the following fields:

ActivateGatewayInput$ActivationKey

ActivateGatewayInput$GatewayName

ActivateGatewayInput$GatewayRegion

ActivateGatewayInput$GatewayTimezone

ActivateGatewayInput$GatewayType

ActivateGatewayInput$TapeDriveType

ActivateGatewayInput$MediumChangerType

ActivateGatewayOutput

AWS Storage Gateway returns the Amazon Resource Name (ARN) of the activated gateway. It is a string made of information such as your account, gateway name, and region. This ARN is used to reference the gateway in other API operations as well as resource-based authorization.

For gateways activated prior to September 02, 2015 the gateway ARN contains the gateway name rather than the gateway id. Changing the name of the gateway has no effect on the gateway ARN.

AddCacheInput

AddCacheOutput

AddTagsToResourceInput
AddTagsToResourceInput

AddTagsToResourceOutput
AddTagsToResourceOutput

AddUploadBufferInput

AddUploadBufferOutput

AddWorkingStorageInput

A JSON object containing one or more of the following fields:

AddWorkingStorageInput$DiskIds

AddWorkingStorageOutput
A JSON object containing the of the gateway for which working storage was configured.

CachediSCSIVolume

CancelArchivalInput
CancelArchivalInput

CancelArchivalOutput
CancelArchivalOutput

CancelRetrievalInput
CancelRetrievalInput

CancelRetrievalOutput
CancelRetrievalOutput

ChapInfo
Describes Challenge-Handshake Authentication Protocol (CHAP) information that supports authentication between your gateway and iSCSI initiators.

CreateCachediSCSIVolumeInput

CreateCachediSCSIVolumeOutput

CreateSnapshotFromVolumeRecoveryPointInput

CreateSnapshotFromVolumeRecoveryPointOutput

CreateSnapshotInput

A JSON object containing one or more of the following fields:

CreateSnapshotInput$SnapshotDescription

CreateSnapshotInput$VolumeARN

CreateSnapshotOutput
A JSON object containing the following fields:

CreateStorediSCSIVolumeInput

A JSON object containing one or more of the following fields:

CreateStorediSCSIVolumeInput$DiskId

CreateStorediSCSIVolumeInput$NetworkInterfaceId

CreateStorediSCSIVolumeInput$PreserveExistingData

CreateStorediSCSIVolumeInput$SnapshotId

CreateStorediSCSIVolumeInput$TargetName

CreateStorediSCSIVolumeOutput
A JSON object containing the following fields:

CreateTapeWithBarcodeInput
CreateTapeWithBarcodeInput

CreateTapeWithBarcodeOutput
CreateTapeOutput

CreateTapesInput
CreateTapesInput

CreateTapesOutput
CreateTapeOutput

DeleteBandwidthRateLimitInput

DeleteBandwidthRateLimitOutput
A JSON object containing the of the gateway whose bandwidth rate information was deleted.

DeleteChapCredentialsInput

A JSON object containing one or more of the following fields:

DeleteChapCredentialsInput$InitiatorName

DeleteChapCredentialsInput$TargetARN

DeleteChapCredentialsOutput
A JSON object containing the following fields:

DeleteGatewayInput
A JSON object containing the id of the gateway to delete.

DeleteGatewayOutput
A JSON object containing the id of the deleted gateway.

DeleteSnapshotScheduleInput

DeleteSnapshotScheduleOutput

DeleteTapeArchiveInput
DeleteTapeArchiveInput

DeleteTapeArchiveOutput
DeleteTapeArchiveOutput

DeleteTapeInput
DeleteTapeInput

DeleteTapeOutput
DeleteTapeOutput

DeleteVolumeInput
A JSON object containing the DeleteVolumeInput$VolumeARN to delete.

DeleteVolumeOutput
A JSON object containing the of the storage volume that was deleted

DescribeBandwidthRateLimitInput
A JSON object containing the of the gateway.

DescribeBandwidthRateLimitOutput
A JSON object containing the following fields:

DescribeCacheInput

DescribeCacheOutput

DescribeCachediSCSIVolumesInput

DescribeCachediSCSIVolumesOutput
A JSON object containing the following fields:

DescribeChapCredentialsInput
A JSON object containing the Amazon Resource Name (ARN) of the iSCSI volume target.

DescribeChapCredentialsOutput
A JSON object containing a .

DescribeGatewayInformationInput
A JSON object containing the id of the gateway.

DescribeGatewayInformationOutput
A JSON object containing the following fields:

DescribeMaintenanceStartTimeInput
A JSON object containing the of the gateway.

DescribeMaintenanceStartTimeOutput

DescribeSnapshotScheduleInput
A JSON object containing the DescribeSnapshotScheduleInput$VolumeARN of the volume.

DescribeSnapshotScheduleOutput

DescribeStorediSCSIVolumesInput
A JSON object containing a list of DescribeStorediSCSIVolumesInput$VolumeARNs.

DescribeStorediSCSIVolumesOutput

DescribeTapeArchivesInput
DescribeTapeArchivesInput

DescribeTapeArchivesOutput
DescribeTapeArchivesOutput

DescribeTapeRecoveryPointsInput
DescribeTapeRecoveryPointsInput

DescribeTapeRecoveryPointsOutput
DescribeTapeRecoveryPointsOutput

DescribeTapesInput
DescribeTapesInput

DescribeTapesOutput
DescribeTapesOutput

DescribeUploadBufferInput

DescribeUploadBufferOutput

DescribeVTLDevicesInput
DescribeVTLDevicesInput

DescribeVTLDevicesOutput
DescribeVTLDevicesOutput

DescribeWorkingStorageInput
A JSON object containing the of the gateway.

DescribeWorkingStorageOutput
A JSON object containing the following fields:

DeviceiSCSIAttributes
Lists iSCSI information about a VTL device.

DisableGatewayInput
DisableGatewayInput

DisableGatewayOutput
DisableGatewayOutput

Disk

GatewayInfo
Describes a gateway object.

ListGatewaysInput

A JSON object containing zero or more of the following fields:

ListGatewaysInput$Limit

ListGatewaysInput$Marker

ListGatewaysOutput

ListLocalDisksInput
A JSON object containing the of the gateway.

ListLocalDisksOutput

ListTagsForResourceInput
ListTagsForResourceInput

ListTagsForResourceOutput
ListTagsForResourceOutput

ListTapesInput

A JSON object that contains one or more of the following fields:

ListTapesInput$Limit

ListTapesInput$Marker

ListTapesInput$TapeARNs

ListTapesOutput

A JSON object containing the following fields:

ListTapesOutput$Marker

ListTapesOutput$VolumeInfos

ListVolumeInitiatorsInput
ListVolumeInitiatorsInput

ListVolumeInitiatorsOutput
ListVolumeInitiatorsOutput

ListVolumeRecoveryPointsInput

ListVolumeRecoveryPointsOutput

ListVolumesInput

A JSON object that contains one or more of the following fields:

ListVolumesInput$Limit

ListVolumesInput$Marker

ListVolumesOutput

NetworkInterface
Describes a gateway's network interface.

RemoveTagsFromResourceInput
RemoveTagsFromResourceInput

RemoveTagsFromResourceOutput
RemoveTagsFromResourceOutput

ResetCacheInput

ResetCacheOutput

RetrieveTapeArchiveInput
RetrieveTapeArchiveInput

RetrieveTapeArchiveOutput
RetrieveTapeArchiveOutput

RetrieveTapeRecoveryPointInput
RetrieveTapeRecoveryPointInput

RetrieveTapeRecoveryPointOutput
RetrieveTapeRecoveryPointOutput

SetLocalConsolePasswordInput
SetLocalConsolePasswordInput

SetLocalConsolePasswordOutput

ShutdownGatewayInput
A JSON object containing the of the gateway to shut down.

ShutdownGatewayOutput
A JSON object containing the of the gateway that was shut down.

StartGatewayInput
A JSON object containing the of the gateway to start.

StartGatewayOutput
A JSON object containing the of the gateway that was restarted.

StorageGatewayClient
A client for the AWS Storage Gateway API.

StorageGatewayError
Provides additional information about an error that was returned by the service as an or. See the errorCode and errorDetails members for more information about the error.

StorediSCSIVolume

Tag

Tape
Describes a virtual tape object.

TapeArchive
Represents a virtual tape that is archived in the virtual tape shelf (VTS).

TapeInfo
Describes a virtual tape.

TapeRecoveryPointInfo
Describes a recovery point.

UpdateBandwidthRateLimitInput

A JSON object containing one or more of the following fields:

UpdateBandwidthRateLimitInput$AverageDownloadRateLimitInBitsPerSec

UpdateBandwidthRateLimitInput$AverageUploadRateLimitInBitsPerSec

UpdateBandwidthRateLimitOutput
A JSON object containing the of the gateway whose throttle information was updated.

UpdateChapCredentialsInput

A JSON object containing one or more of the following fields:

UpdateChapCredentialsInput$InitiatorName

UpdateChapCredentialsInput$SecretToAuthenticateInitiator

UpdateChapCredentialsInput$SecretToAuthenticateTarget

UpdateChapCredentialsInput$TargetARN

UpdateChapCredentialsOutput
A JSON object containing the following fields:

UpdateGatewayInformationInput

UpdateGatewayInformationOutput
A JSON object containing the ARN of the gateway that was updated.

UpdateGatewaySoftwareNowInput
A JSON object containing the of the gateway to update.

UpdateGatewaySoftwareNowOutput
A JSON object containing the of the gateway that was updated.

UpdateMaintenanceStartTimeInput

A JSON object containing the following fields:

UpdateMaintenanceStartTimeInput$DayOfWeek

UpdateMaintenanceStartTimeInput$HourOfDay

UpdateMaintenanceStartTimeInput$MinuteOfHour

UpdateMaintenanceStartTimeOutput
A JSON object containing the of the gateway whose maintenance start time is updated.

UpdateSnapshotScheduleInput

A JSON object containing one or more of the following fields:

UpdateSnapshotScheduleInput$Description

UpdateSnapshotScheduleInput$RecurrenceInHours

UpdateSnapshotScheduleInput$StartAt

UpdateSnapshotScheduleInput$VolumeARN

UpdateSnapshotScheduleOutput
A JSON object containing the of the updated storage volume.

UpdateVTLDeviceTypeInput

UpdateVTLDeviceTypeOutput
UpdateVTLDeviceTypeOutput

VTLDevice
Represents a device object associated with a gateway-VTL.

VolumeInfo
Describes a storage volume object.

VolumeRecoveryPointInfo

VolumeiSCSIAttributes
Lists iSCSI information about a volume.

Enums

ActivateGatewayError

AddCacheError

AddTagsToResourceError

AddUploadBufferError

AddWorkingStorageError

CancelArchivalError

CancelRetrievalError

CreateCachediSCSIVolumeError

CreateSnapshotError

CreateSnapshotFromVolumeRecoveryPointError

CreateStorediSCSIVolumeError

CreateTapeWithBarcodeError

CreateTapesError

DeleteBandwidthRateLimitError

DeleteChapCredentialsError

DeleteGatewayError

DeleteSnapshotScheduleError

DeleteTapeArchiveError

DeleteTapeError

DeleteVolumeError

DescribeBandwidthRateLimitError

DescribeCacheError

DescribeCachediSCSIVolumesError

DescribeChapCredentialsError

DescribeGatewayInformationError

DescribeMaintenanceStartTimeError

DescribeSnapshotScheduleError

DescribeStorediSCSIVolumesError

DescribeTapeArchivesError

DescribeTapeRecoveryPointsError

DescribeTapesError

DescribeUploadBufferError

DescribeVTLDevicesError

DescribeWorkingStorageError

DisableGatewayError

ListGatewaysError

ListLocalDisksError

ListTagsForResourceError

ListTapesError

ListVolumeInitiatorsError

ListVolumeRecoveryPointsError

ListVolumesError

RemoveTagsFromResourceError

ResetCacheError

RetrieveTapeArchiveError

RetrieveTapeRecoveryPointError

SetLocalConsolePasswordError

ShutdownGatewayError

StartGatewayError

UpdateBandwidthRateLimitError

UpdateChapCredentialsError

UpdateGatewayInformationError

UpdateGatewaySoftwareNowError

UpdateMaintenanceStartTimeError

UpdateSnapshotScheduleError

UpdateVTLDeviceTypeError

Type Definitions

ActivationKey

BandwidthDownloadRateLimit

BandwidthType

BandwidthUploadRateLimit

Boolean

CachediSCSIVolumes

ChapCredentials

ChapSecret

ClientToken

DayOfWeek

Description

DeviceType

DiskAllocationType

DiskId

DiskIds

Disks

Double

DoubleObject

ErrorCode

ErrorDetails

GatewayARN
The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

GatewayId

GatewayName
The name you configured for your gateway.

GatewayNetworkInterfaces

GatewayOperationalState

GatewayState

GatewayTimezone

GatewayType

Gateways

HourOfDay

Initiator

Initiators

Integer

IqnName

LastSoftwareUpdate

LocalConsolePassword

Long

Marker

MediumChangerType

MinuteOfHour

NetworkInterfaceId

NextUpdateAvailabilityDate

NumTapesToCreate

PositiveIntObject

RecurrenceInHours

RegionId

ResourceARN

SnapshotDescription

SnapshotId

StorediSCSIVolumes

TagKey

TagKeys

TagValue

Tags

TapeARN

TapeARNs
The Amazon Resource Name (ARN) of each of the tapes you want to list. If you don't specify a tape ARN, the response lists all tapes in both your VTL and VTS.

TapeArchiveStatus

TapeArchives

TapeBarcode

TapeBarcodePrefix

TapeDriveType

TapeInfos
An array of TapeInfo objects, where each object describes an a single tape. If there not tapes in the tape library or VTS, then the TapeInfos is an empty array.

TapeRecoveryPointInfos

TapeRecoveryPointStatus

TapeSize

TapeStatus

Tapes

TargetARN

TargetName

Time

VTLDeviceARN

VTLDeviceARNs

VTLDeviceProductIdentifier

VTLDeviceType

VTLDeviceVendor

VTLDevices

VolumeARN

VolumeARNs

VolumeId

VolumeInfos

VolumeRecoveryPointInfos

VolumeStatus

VolumeType