Module rusoto::workspaces

Amazon Workspaces

Structs

ComputeType
Contains information about the compute type of a WorkSpace bundle.

CreateTagsRequest
The request of the create tags action.

CreateTagsResult
The result of the create tags action.

CreateWorkspacesRequest
Contains the inputs for the CreateWorkspaces operation.

CreateWorkspacesResult
Contains the result of the CreateWorkspaces operation.

DefaultWorkspaceCreationProperties
Contains default WorkSpace creation information.

DeleteTagsRequest
The request of the delete tags action.

DeleteTagsResult
The result of the delete tags action.

DescribeTagsRequest
The request of the describe tags action.

DescribeTagsResult
The result of the describe tags action.

DescribeWorkspaceBundlesRequest
Contains the inputs for the DescribeWorkspaceBundles operation.

DescribeWorkspaceBundlesResult
Contains the results of the DescribeWorkspaceBundles operation.

DescribeWorkspaceDirectoriesRequest
Contains the inputs for the DescribeWorkspaceDirectories operation.

DescribeWorkspaceDirectoriesResult
Contains the results of the DescribeWorkspaceDirectories operation.

DescribeWorkspacesRequest
Contains the inputs for the DescribeWorkspaces operation.

DescribeWorkspacesResult
Contains the results for the DescribeWorkspaces operation.

FailedCreateWorkspaceRequest
Contains information about a WorkSpace that could not be created.

FailedWorkspaceChangeRequest
Contains information about a WorkSpace that could not be rebooted (RebootWorkspaces), rebuilt (RebuildWorkspaces), or terminated (TerminateWorkspaces).

RebootRequest
Contains information used with the RebootWorkspaces operation to reboot a WorkSpace.

RebootWorkspacesRequest
Contains the inputs for the RebootWorkspaces operation.

RebootWorkspacesResult
Contains the results of the RebootWorkspaces operation.

RebuildRequest
Contains information used with the RebuildWorkspaces operation to rebuild a WorkSpace.

RebuildWorkspacesRequest
Contains the inputs for the RebuildWorkspaces operation.

RebuildWorkspacesResult
Contains the results of the RebuildWorkspaces operation.

Tag
Describes the tag of the WorkSpace.

TerminateRequest
Contains information used with the TerminateWorkspaces operation to terminate a WorkSpace.

TerminateWorkspacesRequest
Contains the inputs for the TerminateWorkspaces operation.

TerminateWorkspacesResult
Contains the results of the TerminateWorkspaces operation.

UserStorage
Contains information about the user storage for a WorkSpace bundle.

Workspace
Contains information about a WorkSpace.

WorkspaceBundle
Contains information about a WorkSpace bundle.

WorkspaceDirectory
Contains information about an AWS Directory Service directory for use with Amazon WorkSpaces.

WorkspaceRequest
Contains information about a WorkSpace creation request.

WorkspacesClient
A client for the Amazon WorkSpaces API.

Enums

CreateTagsError

CreateWorkspacesError

DeleteTagsError

DescribeTagsError

DescribeWorkspaceBundlesError

DescribeWorkspaceDirectoriesError

DescribeWorkspacesError

RebootWorkspacesError

RebuildWorkspacesError

TerminateWorkspacesError

Type Definitions

ARN

Alias

BooleanObject

BundleId

BundleIdList

BundleList

BundleOwner

Compute

ComputerName

DefaultOu

Description

DirectoryId

DirectoryIdList

DirectoryList

DirectoryName

DnsIpAddresses

ErrorType

ExceptionMessage

FailedCreateWorkspaceRequests

FailedRebootWorkspaceRequests

FailedRebuildWorkspaceRequests

FailedTerminateWorkspaceRequests

IpAddress

Limit

NonEmptyString

PaginationToken

RebootWorkspaceRequests

RebuildWorkspaceRequests

RegistrationCode

SecurityGroupId

SubnetId

SubnetIds

TagKey

TagKeyList

TagList

TagValue

TerminateWorkspaceRequests

UserName

VolumeEncryptionKey

WorkspaceDirectoryState

WorkspaceDirectoryType

WorkspaceErrorCode

WorkspaceId

WorkspaceIdList

WorkspaceList

WorkspaceRequestList

WorkspaceState