Struct rusoto::workspaces::WorkspacesClient

pub struct WorkspacesClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest {
    // some fields omitted
}

A client for the Amazon WorkSpaces API.

Methods

impl<P> WorkspacesClient<P, Client> where P: ProvideAwsCredentials

fn new(credentials_provider: P, region: Region) -> Self

impl<P, D> WorkspacesClient<P, D> where P: ProvideAwsCredentials, D: DispatchSignedRequest

fn with_request_dispatcher(request_dispatcher: D, credentials_provider: P, region: Region) -> Self

fn create_tags(&self, input: &CreateTagsRequest) -> Result<CreateTagsResult, CreateTagsError>

Creates tags for a WorkSpace.

fn create_workspaces(&self, input: &CreateWorkspacesRequest) -> Result<CreateWorkspacesResult, CreateWorkspacesError>

Creates one or more WorkSpaces.

This operation is asynchronous and returns before the WorkSpaces are created.

fn delete_tags(&self, input: &DeleteTagsRequest) -> Result<DeleteTagsResult, DeleteTagsError>

Deletes tags from a WorkSpace.

fn describe_tags(&self, input: &DescribeTagsRequest) -> Result<DescribeTagsResult, DescribeTagsError>

Describes tags for a WorkSpace.

fn describe_workspace_bundles(&self, input: &DescribeWorkspaceBundlesRequest) -> Result<DescribeWorkspaceBundlesResult, DescribeWorkspaceBundlesError>

Obtains information about the WorkSpace bundles that are available to your account in the specified region.

You can filter the results with either the BundleIds parameter, or the Owner parameter, but not both.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the NextToken response member contains a token that you pass in the next call to this operation to retrieve the next set of items.

fn describe_workspace_directories(&self, input: &DescribeWorkspaceDirectoriesRequest) -> Result<DescribeWorkspaceDirectoriesResult, DescribeWorkspaceDirectoriesError>

Retrieves information about the AWS Directory Service directories in the region that are registered with Amazon WorkSpaces and are available to your account.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the NextToken response member contains a token that you pass in the next call to this operation to retrieve the next set of items.

fn describe_workspaces(&self, input: &DescribeWorkspacesRequest) -> Result<DescribeWorkspacesResult, DescribeWorkspacesError>

Obtains information about the specified WorkSpaces.

Only one of the filter parameters, such as BundleId, DirectoryId, or WorkspaceIds, can be specified at a time.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the NextToken response member contains a token that you pass in the next call to this operation to retrieve the next set of items.

fn reboot_workspaces(&self, input: &RebootWorkspacesRequest) -> Result<RebootWorkspacesResult, RebootWorkspacesError>

Reboots the specified WorkSpaces.

To be able to reboot a WorkSpace, the WorkSpace must have a State of AVAILABLE, IMPAIRED, or INOPERABLE.

This operation is asynchronous and will return before the WorkSpaces have rebooted.

fn rebuild_workspaces(&self, input: &RebuildWorkspacesRequest) -> Result<RebuildWorkspacesResult, RebuildWorkspacesError>

Rebuilds the specified WorkSpaces.

Rebuilding a WorkSpace is a potentially destructive action that can result in the loss of data. Rebuilding a WorkSpace causes the following to occur:

The system is restored to the image of the bundle that the WorkSpace is created from. Any applications that have been installed, or system settings that have been made since the WorkSpace was created will be lost.

The data drive (D drive) is re-created from the last automatic snapshot taken of the data drive. The current contents of the data drive are overwritten. Automatic snapshots of the data drive are taken every 12 hours, so the snapshot can be as much as 12 hours old.

To be able to rebuild a WorkSpace, the WorkSpace must have a State of AVAILABLE or ERROR.

This operation is asynchronous and will return before the WorkSpaces have been completely rebuilt.

fn terminate_workspaces(&self, input: &TerminateWorkspacesRequest) -> Result<TerminateWorkspacesResult, TerminateWorkspacesError>

Terminates the specified WorkSpaces.

Terminating a WorkSpace is a permanent action and cannot be undone. The user's data is not maintained and will be destroyed. If you need to archive any user data, contact Amazon Web Services before terminating the WorkSpace.

You can terminate a WorkSpace that is in any state except SUSPENDED.

This operation is asynchronous and will return before the WorkSpaces have been completely terminated.